% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{add_binary_states}
\alias{add_binary_states}
\title{Add binary state information to the scd.}
\usage{
add_binary_states(scd, column = NULL)
}
\arguments{
\item{scd}{Seurat single cell dataset.}

\item{column}{Get identities from this metadata column.}
}
\description{
I am adding these only so that it is easier to visualize the cells
compared when performing FindAllMarkers(); e.g. it compares each
identity to all others; so I imagine it would be nice to see a
dimplot or something of each state vs. all others as a binary pair
rather than as n separate groups.
}
