% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{add_clonotype_annotations}
\alias{add_clonotype_annotations}
\title{Add VDJ information using some code I found.}
\usage{
add_clonotype_annotations(obj, start_path, type = "t")
}
\arguments{
\item{obj}{Seurat object to which we will add some information.}

\item{start_path}{root of the 10x data in which the vdj information should reside.}

\item{type}{The type of VDJ we expect, heavy(B) or light(T).}
}
\value{
The Seurat object with some new information.
}
\description{
The original implementation of this idea resides at:
https://ucdavis-bioinformatics-training.github.io/2020-Advanced_Single_Cell_RNA_Seq/data_analysis/VDJ_Analysis_fixed
}
