% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{adjuster_counts}
\alias{adjuster_counts}
\title{Apply the suite of destructive batch estimate tools}
\usage{
adjuster_counts(
  input,
  model_fstring = "~ 0 + condition",
  null_fstring = NULL,
  estimate_type = "limma",
  batch1 = "batch",
  batch2 = NULL,
  num_surrogates = "be",
  low_to_zero = FALSE,
  cpus = 4,
  na_to_zero = TRUE,
  confounders = NULL,
  adjust_method = "ruv",
  filter = "raw",
  thresh = 1,
  noscale = FALSE,
  prior_plots = FALSE,
  control_type = "norm"
)
}
\arguments{
\item{input}{Input data structure.}

\item{model_fstring}{Formula string describing the relationship of interest}

\item{null_fstring}{Model describing the null hypothesis}

\item{estimate_type}{The method used to modify the data.}

\item{batch1}{Column containing the primary known batch factor}

\item{batch2}{Column containing a secondary batch factor.}

\item{num_surrogates}{Method to guess the number of surrogate variables,
not likely actually useful for this function.}

\item{low_to_zero}{Push all values which are less than zero to
zero.}

\item{cpus}{dedicate a specific number of cpus}

\item{na_to_zero}{Push all NA values to zero?}

\item{confounders}{Expected confounding factors -- I think only for
smartsva}

\item{adjust_method}{Also not likely used by this.}

\item{filter}{Filter the data before applying the method of choice?}

\item{thresh}{Threshold when filtering}

\item{noscale}{Used by combat, scale the data?}

\item{prior_plots}{Perform the various plots provided by combat}

\item{control_type}{I do not remember.}
}
\value{
Adjusted data
}
\description{
Combat, limma, RUV and friends provide modified counts but not model estimates.
}
