% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{adjuster_expt_svs}
\alias{adjuster_expt_svs}
\title{Search for SVs in an expressionset-like object.}
\usage{
adjuster_expt_svs(
  input,
  model_fstring = "~ 0 + condition",
  null_fstring = NULL,
  model_svs = "sva",
  batch1 = "batch",
  batch2 = NULL,
  num_surrogates = "be",
  low_to_zero = FALSE,
  cpus = NULL,
  na_to_zero = TRUE,
  confounders = NULL,
  filter = "raw",
  thresh = 1,
  noscale = FALSE,
  prior_plots = FALSE,
  control_type = "norm"
)
}
\description{
Search for SVs in an expressionset-like object.
}
