% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{all_adjusters}
\alias{all_adjusters}
\title{Combine all surrogate estimators and batch correctors into one function.}
\usage{
all_adjusters(
  input,
  design = NULL,
  estimate_type = "sva",
  batch1 = "batch",
  batch2 = NULL,
  surrogates = "be",
  low_to_zero = FALSE,
  cpus = 4,
  na_to_zero = TRUE,
  expt_state = NULL,
  confounders = NULL,
  chosen_surrogates = NULL,
  adjust_method = "ruv",
  ...
)
}
\arguments{
\item{input}{Dataframe or expt or whatever as the data to analyze/modify.}

\item{design}{If the data is not an expt, then put the design here.}

\item{estimate_type}{Name of the estimator.}

\item{batch1}{Column in the experimental design for the first known batch.}

\item{batch2}{Only used by the limma method, a second batch column.}

\item{surrogates}{Either a number of surrogates or a method to search for them.}

\item{expt_state}{If this is not an expt, provide the state of the data here.}

\item{confounders}{List of confounded factors for smartSVA/iSVA.}

\item{...}{Extra arguments passed along to other methods.}
}
\value{
List containing surrogate estimates, new counts, the models, and
  some plots, as available.
}
\description{
For a long time, I have mostly kept my surrogate estimators and batch correctors
separate.  However, that separation was not complete, and it really did not make
sense.  This function brings them together. This now contains all the logic from
the freshly deprecated get_model_adjust().
}
\details{
This applies the methodologies very nicely explained by Jeff Leek at
https://github.com/jtleek/svaseq/blob/master/recount.Rmd
and attempts to use them to acquire estimates which may be applied to an
experimental model by either EdgeR, DESeq2, or limma.  In addition, it
modifies the count tables using these estimates so that one may play with the
modified counts and view the changes (with PCA or heatmaps or whatever).
Finally, it prints a couple of the plots shown by Leek in his document. In
other words, this is entirely derivative of someone much smarter than me.
}
\seealso{
[all_adjuster()] [isva] [sva] [limma::removeBatchEffect()]
 [corpcor] [edgeR] [RUVSeq] [SmartSVA] [variancePartition] [counts_from_surrogates()]
}
