% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{all_cprofiler}
\alias{all_cprofiler}
\title{Run simple_clusterprofiler on every table from extract_significant_genes()}
\usage{
all_cprofiler(
  sig,
  tables,
  according_to = "deseq",
  together = FALSE,
  todo = NULL,
  orgdb = "org.Hs.eg.db",
  orgdb_from = NULL,
  orgdb_to = "ENTREZID",
  go_level = 3,
  pcutoff = 0.05,
  qcutoff = 0.1,
  fc_column = "logFC",
  second_fc_column = "deseq_logfc",
  internal = FALSE,
  updown = "up",
  permutations = 1000,
  min_groupsize = 5,
  max_groupsize = 500,
  kegg_prefix = NULL,
  kegg_organism = NULL,
  organism = "human",
  categories = 12,
  padj_type = "BH",
  mesh_category = "C",
  mesh_dbname = "gendoo",
  msigdb_category = "C2",
  msig_db = NULL,
  kegg_universe = NULL,
  reactome_organism = NULL,
  mesh_db = NULL,
  ah_data = NULL,
  signature_data = NULL,
  signature_df = NULL,
  de_table_namedf = NULL,
  sig_genes_namedf = NULL,
  excel = "excel/all_cp.xlsx",
  ...
)
}
\arguments{
\item{sig}{Result from extract_significant_genes.}

\item{tables}{Result from combine_de_tables.}

\item{according_to}{Use this result type for the clusterprofiler searches.}

\item{together}{Concatenate the up/down genes into one set?}

\item{todo}{List of methods to try enrichment/gsea}

\item{orgdb}{Name of the DBI containing the annotations for this organism.}

\item{orgdb_from}{Column from which to convert the gene IDs.}

\item{orgdb_to}{Column to which to convert, this must match the ontology IDS.}

\item{go_level}{Ignore categories above this level in the ontology tree.}

\item{pcutoff}{p-value significance cutoff.}

\item{qcutoff}{FDR adjusted significance cutoff.}

\item{fc_column}{Column containing the fold-change values.}

\item{second_fc_column}{A fallback column for FC values.}

\item{internal}{This changes the output data structure, but I forget how.}

\item{updown}{Seek out categories with increased enrichment,  or decreased.}

\item{permutations}{Run x permutations in clusterProfiler. (deprecated by cp authors)}

\item{min_groupsize}{Ignore groups with less than x genes.}

\item{max_groupsize}{Ignore groups with more than x genes.}

\item{kegg_prefix}{Prefix of this kegg organism ID.}

\item{kegg_organism}{Full name of this organism when querying KEGG.}

\item{organism}{Organism for things like DOSE etc, TODO: double check that this is still needed}

\item{categories}{Plot this number of categories by default.}

\item{padj_type}{FDR correction method.}

\item{mesh_category}{Mesh category to query}

\item{mesh_dbname}{Database in mesh to query}

\item{msigdb_category}{Category of mSigDB to query}

\item{msig_db}{Dataframe of mSigDB gene sets.}

\item{kegg_universe}{Manually set universe of kegg genes/IDs.}

\item{reactome_organism}{I think not needed any longer.}

\item{mesh_db}{Location of mesh database/package.}

\item{ah_data}{Which annotationhub dataset to use with mesh}

\item{signature_data}{Manually provided list of gene sets.}

\item{signature_df}{Manually provided df of signatures.}

\item{de_table_namedf}{Alternate set of gene names.}

\item{sig_genes_namedf}{Alternate set of gene names for the significant genes.}

\item{excel}{Output xlsx filename.}

\item{...}{Arguments to pass to simple_clusterprofiler().}
}
\description{
Most of the options are taken from simple_cprofiler and passed directly down.
}
