% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{all_enricher}
\alias{all_enricher}
\title{Use clusterProfiler enricher() to query an arbitrary set of genes.}
\usage{
all_enricher(
  sig,
  gsc,
  according_to = "deseq",
  together = FALSE,
  plot_type = "dotplot",
  excel = "excel/all_en.xlsx",
  orgdb = "org.Hs.eg.db",
  from = "ENSEMBL",
  to = "SYMBOL"
)
}
\arguments{
\item{sig}{Set of results from extract_significant_genes()}

\item{gsc}{Gene Set Collection; currently only taken from msigdb.}

\item{according_to}{Method to trust}

\item{together}{Do up and down enrichment together?}

\item{plot_type}{Provide these plots -- this should be removed in favor of my arbitrary plotter.}

\item{excel}{Output xlsx filename}

\item{orgdb}{Orgdb used to extract IDs}

\item{from}{ID type from which to convert.}

\item{to}{ID type to which to convert.}
}
\description{
This has only been used with mSigDB at this time.
}
