% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.R
\name{all_gprofiler}
\alias{all_gprofiler}
\title{Run simple_gprofiler on every table from extract_significant_genes()}
\usage{
all_gprofiler(
  sig,
  according_to = "deseq",
  together = FALSE,
  sleep = 7,
  plot_type = "dotplot",
  excel = "excel/all_gp.xlsx",
  ...
)
}
\arguments{
\item{sig}{Result from extract_significant_genes}

\item{according_to}{Use this result type for the gprofiler searches.}

\item{together}{Concatenate the up/down genes into one set?}

\item{sleep}{Give the gProfiler servers a break between queries.}

\item{plot_type}{Choose a plot method as the default.}

\item{excel}{Output xlsx filename.}

\item{...}{Arguments to pass to simple_gprofiler().}
}
\description{
Run simple_gprofiler on every table from extract_significant_genes()
}
