% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{all_ontology_searches}
\alias{all_ontology_searches}
\title{Perform ontology searches of the output from limma.}
\usage{
all_ontology_searches(de_out, gene_lengths = NULL, goids = NULL, n = NULL,
  z = NULL, fc = NULL, p = NULL, overwrite = FALSE,
  goid_map = "reference/go/id2go.map", gff_file = NULL, gff_type = "gene",
  goids_df = NULL, do_goseq = TRUE, do_cluster = TRUE, do_topgo = TRUE,
  do_gostats = TRUE, do_trees = FALSE)
}
\arguments{
\item{de_out}{a list of topTables comprising limma/deseq/edger outputs.}

\item{gene_lengths}{a data frame of gene lengths for goseq.}

\item{goids}{a data frame of goids and genes.}

\item{n}{a number of genes at the top/bottom to search.}

\item{z}{a number of standard deviations to search. (if this and n are null, it assumes 1z)}

\item{fc}{a number of standard deviations to search. (if this and n are null, it assumes 1z)}

\item{p}{a maximum pvalue}

\item{overwrite}{overwrite the excel file}

\item{goid_map}{a map file used by topGO, if it does not exist then provide goids_df to make it.}

\item{gff_file}{a gff file containing the annotations used by gff2genetable from clusterprofiler, which I hacked to make faster.}

\item{gff_type}{column to use from the gff file}

\item{goids_df}{FIXME! a dataframe of genes and goids which I am relatively certain is no longer needed and superseded by goids.}

\item{do_goseq}{perform simple_goseq()?}

\item{do_cluster}{perform simple_clusterprofiler()?}

\item{do_topgo}{perform simple_topgo()?}

\item{do_gostats}{perform simple_gostats()?}

\item{do_trees}{make topGO trees from the data?}
}
\value{
a list of up/down ontology results from goseq/clusterprofiler/topgo/gostats, and associated trees, all optionally.
}
\description{
This passes a set of limma results to (optionally) goseq, clusterprofiler, topgo, and gostats,
collects the outputs, and provides them as a list.  This function needs a species argument,
as I recently made the simple_() functions able to automatically use the various supported organisms.
}
\examples{
\dontrun{
 many_comparisons = limma_pairwise(expt=an_expt)
 tables = many_comparisons$limma
 this_takes_forever = limma_ontology(tables, gene_lengths=lengthdb, goids=goids_df, z=1.5, gff_file='length_db.gff')
}
}

