% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{annotate_network}
\alias{annotate_network}
\title{Use grep to add a vector of annotations/colors to a network.}
\usage{
annotate_network(
  network,
  names,
  color = NULL,
  default = NULL,
  annot_name = "type",
  annot_value = "high"
)
}
\arguments{
\item{network}{Input network}

\item{names}{set of node-names to which to add annotations.}

\item{color}{Color to attach to the added annotation.}

\item{default}{Set a default annotation for this name to all nodes.}

\item{annot_name}{Annotation name to attach to the nodes.}

\item{annot_value}{and the associated value.}
}
\value{
a new network!
}
\description{
The igraph syntaxes are a little clunky, but the set_attr()
functions mostly make sense.
}
