% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{annotate_network_df}
\alias{annotate_network_df}
\title{A version of annotate_network, but which uses a dataframe as input.}
\usage{
annotate_network_df(network, df, default = NULL)
}
\arguments{
\item{network}{input network.}

\item{df}{input dataframe, columns are the new metadata, rows are
the node-strings to search on.}

\item{default}{Set a default?}
}
\description{
The annotate_network() function uses a vector of values, this
extends that logic to add every column of a dataframe.  I would
like to make this function a little more fun vis a vis abilities to
add colors and such.
}
