% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{annotation,expt-method}
\alias{annotation,expt-method}
\title{A getter for the annotation databased used to create an expt/se.}
\usage{
\S4method{annotation}{expt}(object)
}
\arguments{
\item{object}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}
}
\description{
A getter for the annotation databased used to create an expt/se.
}
