% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{annotation<-,expt,ANY-method}
\alias{annotation<-,expt,ANY-method}
\title{A setter for the annotation databased used to create an expt/se.}
\usage{
\S4method{annotation}{expt,ANY}(object) <- value
}
\arguments{
\item{object}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}
}
\description{
A setter for the annotation databased used to create an expt/se.
}
