% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{annotation<-,SummarizedExperiment,character-method}
\alias{annotation<-,SummarizedExperiment,character-method}
\title{Ensure we have an annotation setter for SEs}
\usage{
\S4method{annotation}{SummarizedExperiment,character}(object, ...) <- value
}
\description{
Ensure we have an annotation setter for SEs
}
