% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{annotation<-}
\alias{annotation<-}
\title{Generic set_annotations function calling to BiocGenerics.}
\usage{
annotation(object, ...) <- value
}
\arguments{
\item{object}{Input data structure.}

\item{...}{Other options.}

\item{value}{New annotation identifier.}
}
\description{
I do not think I should need to define this, I have an importFrom
in 01_hpgltools.R and this file includes that.
}
