% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{annotations}
\alias{annotations}
\title{Streptococcus pyogenes 5005 annotation data taken from microbesonline.org}
\format{a \code{data.frame} instance, 1 row per annotation.}
\source{
microbesonline.org
}
\usage{
annotations
}
\description{
This data set results from an import.gff call
}
\author{
atb
}

