% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{as.data.frame,hpgltools::load_annotations_biomart-method}
\alias{as.data.frame,hpgltools::load_annotations_biomart-method}
\title{Pull the dataframe from load_biomart_annotations()}
\usage{
\S4method{as.data.frame}{hpgltools::load_annotations_biomart}(x, row.names = NULL, optional = FALSE, ...)
}
\value{
List containing: a data frame of the found annotations, a copy of
 The mart instance to help with finding problems, the hostname queried, the
 name of the mart queried, a vector of rows queried, vector of the available
 attributes, and the ensembl dataset queried.
}
\description{
Biomart is an amazing resource of information, but using it is a bit
annoying.  This function hopes to alleviate some common headaches.
}
\details{
Tested in test_40ann_biomart.R
This goes to some lengths to find the relevant tables in biomart.  But
biomart is incredibly complex and one should carefully inspect the output if
it fails to see if there are more appropriate marts, datasets, and columns to
download.
}
\examples{
## Provide examples of the biomart annotation functions.

## Search for a biomart instance which will accept queries.
archive_mart <- try(find_working_mart(archive = TRUE, year = 2023, month = 02))
class(archive_mart)
archive_mart

## I should probably change this function to accept the mart from
## find_working_mart()
hs_annot <- try(load_biomart_annotations(year = 2023, month = 02))
class(hs_annot)

## I added a S3 print method to try to avoid being made sad if I
## print the result from load_annotations.
hs_annot

## If one wishes to check out all of the possible data types for a
## single gene, try this:
## However, this function is painfully slow and likely annoying to the
## ensembl servers.
## single_gene <- get_biomart_example_gene()

## Grab gene ontology data for a species:
hs_go <- try(load_biomart_go(year = 2023, month = 02))
hs_go

## The following example is fragile I think because multiple
## ensembl instances must be responding properly to work
## If it fails; check that the examples for biomaRt::getLDS are
## functioning -- most likely they are failing due to an upstream
## error.
hs_mm_ortho <- try(load_biomart_orthologs(gene_ids = "ENSMUSG00000000001",
                                          year = 2023, month = 02,
                                          first_species = "mmusculus",
                                          second_species = "scerevisiae"))

## Create salmon-like ID.version strings using ensembl annotations.
tx_gene_map <- try(make_tx_gene_map(as.data.frame(hs_annot)))
head(tx_gene_map)
}
\seealso{
[biomaRt::listDatasets()] [biomaRt::getBM()] [find_working_mart()]
}
