% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{as.data.frame.annotations_biomart}
\alias{as.data.frame.annotations_biomart}
\title{Pull the dataframe from load_biomart_annotations()}
\usage{
as.data.frame.annotations_biomart(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Result from load_biomart_annotations()}

\item{row.names}{Not currently used}

\item{optional}{I am not sure}

\item{...}{Unused optional parameters.}
}
\description{
Pull the dataframe from load_biomart_annotations()
}
