% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{assay,ExpressionSet,ANY-method}
\alias{assay,ExpressionSet,ANY-method}
\title{A getter to pull the assay data from an ExpressionSet.}
\usage{
\S4method{assay}{ExpressionSet,ANY}(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{withDimnames}{I do not know.}

\item{...}{Extra args!}
}
\description{
A getter to pull the assay data from an ExpressionSet.
}
