% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{assay,expt,ANY-method}
\alias{assay,expt,ANY-method}
\title{A getter to pull the assay data from an expt.}
\usage{
\S4method{assay}{expt,ANY}(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}
}
\description{
A getter to pull the assay data from an expt.
}
