% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{assay<-,ExpressionSet-method}
\alias{assay<-,ExpressionSet-method}
\title{A setter to put the assay data into an ExpressionSet.}
\usage{
\S4method{assay}{ExpressionSet}(x, i, withDimnames = TRUE, ...) <- value
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{i}{Subset to replace.}

\item{withDimnames}{I do not know, I need to look this up.}

\item{...}{Extra args.}

\item{value}{New values for the expressionset.}
}
\description{
A setter to put the assay data into an ExpressionSet.
}
