% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{assay<-}
\alias{assay<-}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
assay(x, i, withDimnames = TRUE, ...) <- value
}
\arguments{
\item{x}{The SummarizedExperiment input}

\item{i}{undef}

\item{withDimnames}{undef}

\item{...}{extra args.}

\item{value}{New value.}
}
\description{
message("I am from SummarizedExperiment and am explicitly imported, wtf.")
}
