% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{assay}
\alias{assay}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
assay(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{The SummarizedExperiment input}

\item{i}{undef}

\item{withDimnames}{undef}

\item{...}{extra args.}
}
\description{
message("I am from SummarizedExperiment and am explicitly imported, wtf.")
}
