% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoloads.r
\name{autoloads_all}
\alias{autoloads_all}
\title{Automatic installation of stuff I use.}
\usage{
autoloads_all(update = FALSE)
}
\arguments{
\item{update}{Update installed packages?}
}
\value{
The number of installed packages
}
\description{
hpgltools uses too many packages, ~65 at last count, I use autoloads_all() to make sure they are
all installed.
}
\seealso{
\link[BiocInstaller]{biocLite} \link{install.packages}
}

