% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{backup_expression_data,ExpressionSet-method}
\alias{backup_expression_data,ExpressionSet-method}
\title{Backup the state of an expressionSet.}
\usage{
\S4method{backup_expression_data}{ExpressionSet}(expt)
}
\arguments{
\item{expt}{An ExpressionSet.}
}
\description{
Backup the state of an expressionSet.
}
