% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{backup_expression_data,SummarizedExperiment-method}
\alias{backup_expression_data,SummarizedExperiment-method}
\title{Backup the state of an SummarizedExperiment.}
\usage{
\S4method{backup_expression_data}{SummarizedExperiment}(expt)
}
\arguments{
\item{expt}{A SummarizedExperiment.}
}
\description{
Backup the state of an SummarizedExperiment.
}
