% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{backup_expression_data}
\alias{backup_expression_data}
\title{Keep a copy of the original state of an expressionset in case of shenanigans.}
\usage{
backup_expression_data(expt)
}
\arguments{
\item{expt}{The expt before messing with it.}
}
\value{
The expt with a new slot 'original_expressionset', but only
 if it did not already exist.
}
\description{
Keep a copy of the original state of an expressionset in case of shenanigans.
}
