% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpgltools.r
\docType{data}
\name{base_size}
\alias{base_size}
\title{What happens if I set a base ggplot2 text size here?}
\format{An object of class \code{numeric} of length 1.}
\usage{
base_size
}
\description{
What happens if I set a base ggplot2 text size here?
}
\keyword{datasets}
