% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_basic.r
\name{basic_coefficient_scatter}
\alias{basic_coefficient_scatter}
\title{Plot two coefficients with respect to one another from basic.}
\usage{
basic_coefficient_scatter(output, toptable = NULL, x = 1, y = 2,
  gvis_filename = NULL, gvis_trendline = TRUE, z = 1.5,
  tooltip_data = NULL, base_url = NULL, color_low = "#DD0000",
  color_high = "#7B9F35", ...)
}
\arguments{
\item{output}{Set of pairwise comparisons provided by basic_pairwise().}

\item{toptable}{The table to use for extracting the logfc values.}

\item{x}{Name or number of the x-axis coefficient column to extract.}

\item{y}{Name or number of the y-axis coefficient column to extract.}

\item{gvis_filename}{Filename for plotting gvis interactive graphs of the data.}

\item{gvis_trendline}{Add a trendline to the gvis plot?}

\item{z}{Make pretty colors for genes this number of z-scores from the median.}

\item{tooltip_data}{Dataframe of gene annotations to be used in the gvis plot.}

\item{base_url}{Add a linkout to gvis plots to this base url.}

\item{color_low}{Color to use for low-logfc values.}

\item{color_high}{Color to use for high-logfc values.}

\item{...}{A few options may be added outside this scope and are left in the arglist, notably
qlimit, fc_column, p_column.  I need to make a consistent decision about how to handle these
not-always needed parameters, either always define them in the function body, or always put
them in arglist(...), doing a little of both is stupid.}
}
\value{
Ggplot2 plot showing the relationship between the two coefficients.
}
\description{
It can be nice to see a plot of two coefficients from a basic comparison with respect to one another
This hopefully makes that easy.
}
\examples{
\dontrun{
 pretty = coefficient_scatter(limma_data, x="wt", y="mut")
}
}
\seealso{
\link{plot_linear_scatter} \link{basic_pairwise}
}

