% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_basic.r
\name{basic_ma}
\alias{basic_ma}
\title{Make a MA plot of some limma output with pretty colors and shapes}
\usage{
basic_ma(output, table = NULL, fc_col = "logFC", p_col = "p",
  expr_col = "numerator_median", fc = 1, pval_cutoff = 0.05)
}
\arguments{
\item{output}{The result from all_pairwise(), which should be changed to handle other invocations too.}

\item{table}{Result from basic to use, left alone it chooses the first.}

\item{fc_col}{Column for logFC data.}

\item{p_col}{Column to use for p-value data.}

\item{expr_col}{Column for the average data.}
}
\value{
a plot!
}
\description{
Yay pretty colors and shapes!
}
\examples{
 \dontrun{
  prettyplot <- basic_ma(all_aprwise) ## [sic, I'm witty! and can speel]
}
}
\seealso{
\link{plot_ma_de}
}

