% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{batch_counts}
\alias{batch_counts}
\title{Perform different batch corrections using limma, sva, ruvg, and cbcbSEQ.}
\usage{
batch_counts(count_table, design, batch = TRUE, batch1 = "batch",
  batch2 = NULL, noscale = TRUE, ...)
}
\arguments{
\item{count_table}{a matrix of (pseudo)counts.}

\item{design}{a model matrix defining the experimental conditions/batches/etc}

\item{batch}{a string describing the method to try to remove the batch effect (or FALSE to leave it alone, TRUE uses limma)}

\item{batch1}{the column in the design table describing the presumed covariant to remove.}

\item{batch2}{the column in the design table describing the second covariant to remove (only used by limma at the moment).}

\item{noscale}{used for combatmod, when true it removes the scaling parameter from the invocation of the modified combat.}

\item{...}{more options for you!}
}
\value{
The 'batch corrected' count table and new library size.  Please remember that the library size which comes out of this
may not be what you want for voom/limma and would therefore lead to spurious differential expression values.
}
\description{
Perform different batch corrections using limma, sva, ruvg, and cbcbSEQ.
}
\examples{
\dontrun{
limma_batch <- batch_counts(table, design, batch1='batch', batch2='strain')
sva_batch <- batch_counts(table, design, batch='sva')
}
}
\seealso{
\pkg{limma} \pkg{edgeR} \pkg{RUVSeq} \pkg{sva} \pkg{cbcbSEQ}
}

