% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{batches<-}
\alias{batches<-}
\title{Add a batch column to a se.}
\usage{
batches(se) <- value
}
\arguments{
\item{se}{Summarized Experiment to modify.}

\item{value}{vector of batches.}
}
\description{
Add a batch column to a se.
}
