% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{batches}
\alias{batches}
\title{Generic get_batches function with a reminder if I pass it something new.}
\usage{
batches(exp)
}
\arguments{
\item{exp}{Input data structure.}
}
\description{
Generic get_batches function with a reminder if I pass it something new.
}
