% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{binary_pairwise}
\alias{binary_pairwise}
\title{Perform all_pairwise only using deseq/edger.}
\usage{
binary_pairwise(...)
}
\arguments{
\item{...}{Args usually passed to all_pairwise()}
}
\description{
The thing I want to do which I presume will be of use to Zhezhen is
to have a variant of this which takes the list of interesting
contrasts and only performs them rather than my default of doing
all possible pairwise contrasts.  I think that will only require a
little logic in make_contrasts to skip contrasts not in the list of
interest.
}
