% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{binary_paiwise}
\alias{binary_paiwise}
\title{Perform all_pairwise only using deseq/edger.}
\usage{
binary_paiwise(...)
}
\arguments{
\item{...}{Args usually passed to all_pairwise()}

\item{do_limma}{(Don't) do limma.}

\item{do_edger}{Do edger.}

\item{do_basic}{(Don't) do basic.}

\item{do_deseq}{Do deseq.}

\item{do_ebseq}{(Don't) do ebseq.}

\item{do_noiseq}{(Don't) do noiseq.}

\item{do_dream}{(Don't) do dream.}
}
\description{
The thing I want to do which I presume will be of use to Zhezhen is
to have a variant of this which takes the list of interesting
contrasts and only performs them rather than my default of doing
all possible pairwise contrasts.  I think that will only require a
little logic in make_contrasts to skip contrasts not in the list of
interest.
}
