% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.r
\name{biomart_orthologs}
\alias{biomart_orthologs}
\title{Use biomart to get orthologs between supported species.}
\usage{
biomart_orthologs(gene_ids, first_species = "hsapiens",
  second_species = "mmusculus", host = "dec2015.archive.ensembl.org",
  trymart = "ENSEMBL_MART_ENSEMBL")
}
\arguments{
\item{gene_ids}{List of gene IDs to translate.}

\item{first_species}{Linnean species name for one species.}

\item{second_species}{Linnean species name for the second species.}

\item{host}{Ensembl server to query.}

\item{trymart}{Assumed mart name to use.}
}
\value{
Df of orthologs.
}
\description{
Biomart's function getLDS is incredibly powerful, but it makes me think very polite people are
going to start knocking on my door, and it fails weirdly pretty much always. This function
attempts to alleviate some of that frustration.
}

