% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{cbcb_batch}
\alias{cbcb_batch}
\title{A function suggested by Hector Corrada Bravo and Kwame Okrah for batch
removal.}
\usage{
cbcb_batch(normalized_counts, model, batch1 = "batch",
  condition = "condition", matrix_scale = "linear",
  return_scale = "linear", method = "subtract")
}
\arguments{
\item{normalized_counts}{Data frame of log2cpm counts.}

\item{model}{Balanced experimental model containing condition and batch
factors.}
}
\value{
Dataframe of residuals after subtracting batch from the model.
}
\description{
During a lab meeting, the following function was suggested as a quick and
dirty batch removal tool.  It takes data and a model including a 'batch'
factor, invokes limma on them, removes the batch factor, does a cross
product of the fitted data and modified model and uses that with residuals to
get a new data set.
}
\examples{
\dontrun{
 newdata <- cbcb_batch_effect(counts, expt_model)
}
}
\seealso{
\pkg{limma}
 \code{\link[limma]{voom}} \code{\link[limma]{lmFit}}
}
