% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{cbcb_batch_effect}
\alias{cbcb_batch_effect}
\title{A function suggested by Hector Corrada Bravo and Kwame Okrah for batch removal}
\usage{
cbcb_batch_effect(normalized_counts, model)
}
\arguments{
\item{normalized_counts}{Data frame of log2cpm counts.}

\item{model}{Balanced experimental model containing condition and batch factors.}
}
\value{
Dataframe of residuals after subtracting batch from the model.
}
\description{
During a lab meeting, the following function was suggested as a quick and dirty batch removal tool
}
\examples{
\dontrun{
newdata <- cbcb_batch_effect(counts, expt_model)
}
}
\seealso{
\link[limma]{voom} \link[limma]{lmFit}
}

