% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085_normalize_batch.r
\name{cbcb_combat}
\alias{cbcb_combat}
\title{A modified version of comBatMod.}
\usage{
cbcb_combat(dat, batch, mod, noscale = TRUE, prior.plots = FALSE, ...)
}
\arguments{
\item{dat}{Df to modify.}

\item{batch}{Factor of batches.}

\item{mod}{Factor of conditions.}

\item{noscale}{The normal 'scale' option squishes the data too much, so this
defaults to TRUE.}

\item{prior.plots}{Print out prior plots?}

\item{...}{Extra options are passed to arglist}
}
\value{
Df of batch corrected data
}
\description{
This is a hack of Kwame Okrah's combatMod to make it not fail on corner-cases.
This was mostly copy/pasted from
https://github.com/kokrah/cbcbSEQ/blob/master/R/transform.R
}
\examples{
\dontrun{
 df_new = cbcb_combat(df, batches, model)
}
}
\seealso{
\pkg{sva}
 \code{\link[sva]{ComBat}}
}
