% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{cdm_expt_rda}
\alias{cdm_expt_rda}
\title{Data file containing an expressionset of some GAS 5005 RNASeq data.}
\format{
An object of class \code{character} of length 1.
}
\usage{
cdm_expt_rda
}
\description{
This rda file contains an expt which comprises a subset of a RNASeq
experiment performed with the McIver lab.  The goal was to examine changes in
transcription across two media types (the very permissive THY and restrictive
CDM).  In addition, this used two strains, one which is more and less
pathogenic.
}
\keyword{datasets}
