% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{cheap_tm}
\alias{cheap_tm}
\title{Simplified TM calculator}
\usage{
cheap_tm(sequence)
}
\arguments{
\item{sequence}{String of atgc letters (not smart enough to do RNA).}
}
\description{
A quick and dirty TM calculator, taken from:
Taken from: https://www.biostars.org/p/58437/
}
