% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_methods_circos.R
\name{check_circos}
\alias{check_circos}
\title{Validation function when creating a circos class.}
\usage{
check_circos(object)
}
\arguments{
\item{object}{The object to check for validity.}
}
\value{
TRUE or FALSE
}
\description{
This is the one of the first steps taken to make the circos plot
builder into an object oriented set of functions.  Thank you,
Theresa!
}
