% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{check_clusterprofiler}
\alias{check_clusterprofiler}
\title{Make sure that clusterProfiler is ready to run}
\usage{
check_clusterprofiler(gff = "test.gff", gomap = NULL)
}
\arguments{
\item{gff}{The gff file containing annotation data (gene lengths)}

\item{gomap}{a data frame of gene IDs and GO ontologies 1:1, other columns are ignored.}
}
\value{
the GO2EG data structure created, probably don't save this, its big
}
\description{
Make sure that clusterProfiler is ready to run
}
\examples{
\dontrun{
 go2eg <- check_clusterprofiler(gff, goids)
 rm(go2eg)
}
}

