% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{check_fstring}
\alias{check_fstring}
\title{Perform some checks on a formula string}
\usage{
check_fstring(fstring)
}
\arguments{
\item{fstring}{Input fstring}
}
\description{
It might be nice to be able to add some checks of the fstring against
the experimental design as well.  For now, I will just query the fstring for
a ~ character.
}
