% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{check_includes}
\alias{check_includes}
\title{Convert a vector of yes/no by DE method to a list.}
\usage{
check_includes(includes = NULL, apr = NULL, methods = NULL, ...)
}
\arguments{
\item{includes}{List of methods desired}

\item{apr}{Input pairwise result, not always needed.}

\item{methods}{In a few place I use a list called methods instead of includes,
this should be removed.}

\item{...}{Extra args, currently ignored.}
}
\value{
List containing TRUE/FALSE for each method desired,
 depending on if we actually have the relevant data.
}
\description{
This compiles the set of possible methods to include in an
all_pairwise() from a series of booleans into a simpler list and
checks that the elements have some data that may be used.
}
