% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{check_metadata_year}
\alias{check_metadata_year}
\title{Figure out when mappings were performed by their timestamp}
\usage{
check_metadata_year(metadata = NULL, column = NULL)
}
\arguments{
\item{metadata}{File containing the metadata for this experiment.
If none is provided, this function will just give the current
year, which is only what you want if this is brand new data.}

\item{column}{Sanitized column name in the metadata containing the
count tables of interest.  If this is not provided, it will return
the month/year of the timestamp for the metadata.  This has a
reasonable chance of giving correct information.}
}
\description{
I got bit in the butt by mismatching ensembl IDs from some older
count tables and newer annotations.  Happily my biomart annotation
gatherer is smart enough to collect from the archive servers, so it
should not be difficult for me to ensure that they match in the
future.
}
\details{
With that in mind, provide this function with the filename of some
metadata and the file column in it, and it will look at the first
file and return the year and month it was created.  Therefore, you
may ask ensembl for the appropriately dated gene annotations.
}
