% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.R
\name{check_plot_scale}
\alias{check_plot_scale}
\title{Look at the range of the data for a plot and use it to suggest if a plot
should be on log scale.}
\usage{
check_plot_scale(data, scale = NULL, max_data = 10000, min_data = 10)
}
\arguments{
\item{data}{Data to plot.}

\item{scale}{If known, this will be used to define what (if any) values to
change.}

\item{max_data}{Define the upper limit for the heuristic.}

\item{min_data}{Define the lower limit for the heuristic.}
}
\description{
There are a bunch of plots which often-but-not-always benefit from being
displayed on a log scale rather than base 10.  This is a quick and dirty
heuristic which suggests the appropriate scale.  If the data 'should' be on
the log scale and it has 0s, then they are moved to 1 so that when logged
they will return to 0.  Similarly, if there are negative numbers and the
intended scale is log, then this will set values less than 0 to zero to avoid
imaginary numbers.
}
