% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{check_xlsx_worksheet}
\alias{check_xlsx_worksheet}
\title{Create the named worksheet in a workbook, this function was not well named.}
\usage{
check_xlsx_worksheet(wb, sheet)
}
\arguments{
\item{wb}{Workbook to modify}

\item{sheet}{Sheet to check/create.}
}
\value{
The workbook object hopefully with a new worksheet.
}
\description{
This tries to make sure that some of the problems of creating new worksheets
do not occur.  E.g. Names must be less than something and must be unique.
}
\seealso{
[openxlsx::addWorksheet()]
}
