% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_basic.R
\name{choose_basic_dataset}
\alias{choose_basic_dataset}
\title{Attempt to ensure that input data to basic_pairwise() is suitable.}
\usage{
choose_basic_dataset(input, force = FALSE, ...)
}
\arguments{
\item{input}{An expressionset containing expt to test and/or modify.}

\item{force}{If we want to try out other distributed data sets, force it in using me.}

\item{...}{future options, I think currently unused.}
}
\value{
data ready for basic_pairwise()
}
\description{
basic_pairwise() assumes log2 data as input, use this to ensure that is true.
}
\examples{
\dontrun{
 ready <- choose_basic_dataset(expt)
}
}
\seealso{
[Biobase] [choose_dataset()] [normalize_expt()]
}
