% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{choose_dataset}
\alias{choose_dataset}
\title{Choose a suitable data set for Edger/DESeq}
\usage{
choose_dataset(input)
}
\arguments{
\item{input}{Expt input.}
}
\value{
List the data, conditions, and batches in the data.
}
\description{
The _pairwise family of functions all demand data in specific formats.
This tries to make that consistent.
}

