% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{choose_limma_dataset}
\alias{choose_limma_dataset}
\title{A sanity check that a given set of data is suitable for analysis by limma.}
\usage{
choose_limma_dataset(input, force = FALSE, which_voom = "limma", ...)
}
\arguments{
\item{input}{Expressionset containing expt object.}

\item{force}{Ingore warnings and use the provided data asis.}

\item{which_voom}{Choose between limma'svoom, voomWithQualityWeights, or the hpgl equivalents.}

\item{...}{Extra arguments passed to arglist.}
}
\value{
dataset suitable for limma analysis
}
\description{
Take an expt and poke at it to ensure that it will not result in troubled limma results.
}

