% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{choose_model}
\alias{choose_model}
\title{Try out a few experimental models and return a likely working option.}
\usage{
choose_model(input, conditions, batches, model_batch = TRUE,
  model_cond = TRUE, model_intercept = TRUE, alt_model = NULL,
  alt_string = NULL, intercept = 0, reverse = FALSE, surrogates = "be",
  ...)
}
\arguments{
\item{input}{Input data used to make the model.}

\item{conditions}{Factor of conditions in the putative model.}

\item{batches}{Factor of batches in the putative model.}

\item{model_batch}{Try to include batch in the model?}

\item{model_cond}{Try to include condition in the model? (Yes!)}

\item{model_intercept}{Use an intercept model instead of cell-means?}

\item{alt_model}{Use your own model.}

\item{alt_string}{String describing an alternate model.}

\item{intercept}{Choose an intercept for the model as opposed to 0.}

\item{reverse}{Reverse condition/batch in the model?  This shouldn't/doesn't matter but I wanted
to test.}

\item{surrogates}{Number of or method used to choose the number of surrogate variables.}

\item{...}{Further options are passed to arglist.}
}
\value{
List including a model matrix and strings describing cell-means and intercept models.
}
\description{
The _pairwise family of functions all demand an experimental model.  This tries to choose a
consistent and useful model for all for them.  This does not try to do multi-factor, interacting,
nor dependent variable models, if you want those do them yourself and pass them off as alt_model.
}
\details{
Invoked by the _pairwise() functions.
}
\seealso{
\pkg{stats}
 \code{\link[stats]{model.matrix}}
}
