% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{choose_sequence_regions}
\alias{choose_sequence_regions}
\title{Given a named vector of fun regions, make a dataframe which
includes putative primers and the spec strings for expected variants.}
\usage{
choose_sequence_regions(
  vector,
  max_primer_length = 45,
  topn = 200,
  bin_size = 600
)
}
\description{
Given a named vector of fun regions, make a dataframe which
includes putative primers and the spec strings for expected variants.
}
