% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{choose_txdb}
\alias{choose_txdb}
\title{Load the appropriate TxDb environment for a given species.}
\usage{
choose_txdb(species = "saccharomyces_cerevisiae")
}
\arguments{
\item{species}{Human readable species name}
}
\value{
orgdb object for the relevant species, or an error if I don't have a mapping for it.
}
\description{
Ok, so these are a bit more complex than I realized.  The heirarchy as I now understand it
(probably wrong) is that orgdb objects provide ID mappings among the various DBs.  txdb objects
provide the actual annotation information, and organismdbs acquire both (but only exist for a few
species). Let's face it, I will never remember that the yeast orgdb is 'org.Sc.sgd.something'.  This
function is intended to make that process easier.  Feed it a species name which makes sense:
'homo_sapiens' and it will assume you mean orgdb.whatever and load that into your
environment. This should also make a reasonable attempt at installing the appropriate orgdb if it
is not already in your R library tree.
}
\examples{
 object <- choose_txdb("homo_sapiens")
}

