% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{circos_arc}
\alias{circos_arc}
\title{circos_arc()  Write arcs between chromosomes in circos.}
\usage{
circos_arc(df, cfgout = "circos/conf/default.conf", first_col = "chr1",
  second_col = "chr2", color = "blue", radius = 0.75, thickness = 3)
}
\arguments{
\item{df}{a dataframe with starts/ends and the floating point information}

\item{cfgout}{default='circos/conf/default.conf'  The master
configuration file to write.}

\item{first_col}{default='chr1'  The name of the first chromosome}

\item{second_col}{default='chr2'  The name of the second chromosome}

\item{color}{default='blue'  the color of the histogram}

\item{radius}{default=0.75  the radius at which to add the arcs}

\item{thickness}{default=3  integer thickness of the arcs}
}
\value{
undef
}
\description{
Ok, so when I said I only do 1 chromosome images, I lied.
This function tries to make writing arcs between chromosomes easier.
It too works in 3 stages,
It writes out a data file using cfgout as a basename and the data
from df in the circos arc format into
circos/data/bob_arc.txt
It then writes out a configuration plot stanza in
circos/conf/bob_arc.conf
and finally adds an include to circos/bob.conf
}
\details{
In its current implementation, this only understands two
chromosomes.  A minimal amount of logic and data organization will
address this weakness.
}

