% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{circos_etc}
\alias{circos_etc}
\title{Extra fonts and useful bits and bobs for working with circos.}
\format{
An object of class \code{character} of length 1.
}
\usage{
circos_etc
}
\description{
This is a tarball of the circos etc/ directory from my Debian linux
installation.  I discovered to my annoyance that other systems were missing
the fonts required to make circos plots work properly along with something
else.  In a fit of pique I tarred them up and left them here.
}
\keyword{datasets}
