% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_heatmap}
\alias{circos_heatmap}
\title{Write tiles of arbitrary heat-mappable data in circos.}
\usage{
circos_heatmap(df, annot_df, cfgout = "circos/conf/default.conf",
  colname = "logFC", chr = "chr1", basename = "", colors = NULL,
  color_choice = "spectral-9-div", scale_log_base = 1, outer = 0.9,
  width = 0.08, spacing = 0.02)
}
\arguments{
\item{df}{Dataframe with starts/ends and the floating point information.}

\item{annot_df}{Annotation data frame with starts/ends.}

\item{cfgout}{Master configuration file to write.}

\item{colname}{Name of the column with the data of interest.}

\item{chr}{Name of the chromosome (This currently assumes a bacterial chromosome).}

\item{basename}{Make sure the written configuration files get different names with this.}

\item{colors}{Colors of the heat map.}

\item{color_choice}{Name of the heatmap to use, I forget how this interacts with color...}

\item{scale_log_base}{Defines how the range of colors will be ranged with
respect to the values in the data.}

\item{outer}{Floating point radius of the circle into which to place the heatmap.}

\item{width}{Width of each tile in the heatmap.}

\item{spacing}{Radial distance between outer, inner, and inner to whatever follows.}
}
\value{
Radius after adding the histogram and the spacing.
}
\description{
This function tries to make the writing circos heatmaps easier.  Like
circos_plus_minus() and circos_hist() it works in 3 stages,
It writes out a data file using cfgout as a basename and the data
from df in the circos histogram format into
circos/data/bob_heatmap.txt
It then writes out a configuration plot stanza in
circos/conf/bob_heatmap.conf
and finally adds an include to circos/bob.conf
}
