% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circos.R
\name{circos_karyotype}
\alias{circos_karyotype}
\title{circos_karyotype()  Create the description of (a)chromosome(s) for circos.}
\usage{
circos_karyotype(outfile = "circos/conf/karyotypes/default.txt",
  length = 1838554, name = "chr1", segments = 6, color = "white",
  chr_num = 1)
}
\arguments{
\item{outfile}{default='circos/conf/karyotypes/default.txt'  a file to write}

\item{length}{default=1838554  the default length of the chromosome (That is mgas5005)}

\item{name}{default='chr1'  the name of the chromosome (This currently assumes a bacterial chromosome)}

\item{segments}{default=6  how many segments to cut it into}

\item{color}{default='white'  how to colors the chromosomal arc. (circos images are cluttered enough)}

\item{chr_num}{default=1  the number to record (This and name above should change for multi-chromosomal species)}
}
\value{
undef
}
\description{
This function tries to save me from having to get the lengths of arcs for bacterial chromosomes
manually correct, and writes them as a circos compatible karyotype file.
The outfile parameter was chosen to match the configuration directive
outlined in circos_prefix(), however that will need to be changed in order for this to work
in variable conditions.  Next time I make one of these graphs I will do that I suspect.
In addition, this currently only understands how to write bacterial chromosomes, that will
likely be fixed when I am asked to write out a L.major karyotype.
}
\details{
These defaults were chosen because I have a chromosome of this length that is correct.
}

