% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circos.R
\name{circos_makefile}
\alias{circos_makefile}
\title{circos_makefile()  Write a simple makefile for circos.}
\usage{
circos_makefile(output = "circos/Makefile", circos = "/usr/bin/circos")
}
\arguments{
\item{output}{default='circos/Makefile' the makefile}

\item{circos}{default='/usr/bin/circos'  the location of circos. (I
    have a copy in home/bin/circos and use that sometimes.}
}
\value{
a kitten
}
\description{
I regenerate all my circos pictures with make(1).  This is my
makefile.
}

