% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_plus_minus}
\alias{circos_plus_minus}
\title{Write tiles of bacterial ontology groups using the categories from
microbesonline.org.}
\usage{
circos_plus_minus(
  cfg,
  outer = 1,
  width = 0.08,
  thickness = 95,
  spacing = 0,
  padding = 1,
  margin = 0,
  plus_orientation = "out",
  minus_orientation = "in",
  layers = 1,
  layers_overflow = "hide",
  acol = "orange",
  bcol = "reds-9-seq",
  ccol = "yellow",
  dcol = "vlpurple",
  ecol = "vlgreen",
  fcol = "dpblue",
  gcol = "vlgreen",
  hcol = "vlpblue",
  icol = "vvdpgreen",
  jcol = "dpred",
  kcol = "orange",
  lcol = "vvlorange",
  mcol = "dpgreen",
  ncol = "vvlpblue",
  ocol = "vvlgreen",
  pcol = "vvdpred",
  qcol = "ylgn-3-seq",
  rcol = "vlgrey",
  scol = "grey",
  tcol = "vlpurple",
  ucol = "greens-3-seq",
  vcol = "vlred",
  wcol = "vvdppurple",
  xcol = "black",
  ycol = "lred",
  zcol = "vlpblue",
  max = NULL,
  label_column = NULL,
  url_string = ""
)
}
\arguments{
\item{cfg}{Result from circos_prefix().}

\item{outer}{Floating point radius of the circle into which to place the
plus-strand data.}

\item{width}{Radial width of each tile.}

\item{thickness}{How wide to make the bars.}

\item{spacing}{Radial distance between outer, inner, and inner to whatever follows.}

\item{padding}{How much space between them.}

\item{margin}{Margin between elements.}

\item{plus_orientation}{Orientation of the plus pieces.}

\item{minus_orientation}{Orientation of the minus pieces.}

\item{layers}{How many layers to use}

\item{layers_overflow}{How to handle too many layers.}

\item{acol}{A color: RNA processing and modification. (orange)}

\item{bcol}{B color: Chromatin structure and dynamics. (red-9)}

\item{ccol}{C color: Energy production conversion. (yellow)}

\item{dcol}{D color: Cell cycle control, mitosis and meiosis. (very light purple)}

\item{ecol}{E color: Amino acid transport metabolism. (very light green)}

\item{fcol}{F color: Nucleotide transport and metabolism. (deep blue)}

\item{gcol}{G color: Carbohydrate transport and metabolism. (very light green)}

\item{hcol}{H color: Coenzyme transport and metabolism. (very light purple blue)}

\item{icol}{I color: Lipid transport and metabolism. (very very deep green)}

\item{jcol}{J color: Translation, ribosome structure and biogenesis. (deep red)}

\item{kcol}{K color: Transcription. (orange)}

\item{lcol}{L color: Replication, recombination, and repair. (very very light orange)}

\item{mcol}{M color: Cell wall/membrane biogenesis. (deep green)}

\item{ncol}{N color: Cell motility (very very light purple blue)}

\item{ocol}{O color: Posttranslational modification, protein turnover, chaperones. (very very light green)}

\item{pcol}{P color: Inorganic ion transport and metabolism. (very very deep red)}

\item{qcol}{Q color: Secondary metabolite biosynthesis, transport, and catabolism. (very light green 3)}

\item{rcol}{R color: General function prediction only. (very light grey)}

\item{scol}{S color: Function unknown. (grey)}

\item{tcol}{T color: Signal transduction mechanisms. (very light purple)}

\item{ucol}{U color: Intracellular trafficking(sp?) and secretion. (green 3)}

\item{vcol}{V color: Defense mechanisms. (very light red)}

\item{wcol}{W color: Extracellular structures. (very very deep purple)}

\item{xcol}{X color: Not in COG. (black)}

\item{ycol}{Y color: Nuclear structure. (light red)}

\item{zcol}{Z color: Cytoskeleton. (very light purple blue)}

\item{max}{Maximum length for chromosomal lengths}
}
\value{
Radius after adding the plus/minus information and the spacing between them.
}
\description{
This function tries to save me from writing out ontology
definitions and likely making mistakes.  It uses the start/ends
from the gff annotation along with the 1 letter GO-like categories
from microbesonline.org.  It then writes two data files
circos/data/bob_plus_go.txt, circos/data/bob_minus_go.txt along
with two configuration files circos/conf/bob_minus_go.conf and
circos/conf/bob_plus_go.conf and finally adds an include to
circos/bob.conf
}
