% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_prefix}
\alias{circos_prefix}
\title{Write the beginning of a circos configuration file.}
\usage{
circos_prefix(
  annotation,
  name = "mgas",
  base_dir = "circos",
  chr_column = "seqnames",
  cog_column = "COGFun",
  start_column = "start",
  stop_column = "end",
  strand_column = "strand",
  id_column = NULL,
  cog_map = NULL,
  radius = 1800,
  chr_units = 1000,
  band_url = NULL,
  ...
)
}
\arguments{
\item{annotation}{Annotation data frame.}

\item{name}{Name of the map, called with 'make name'.}

\item{base_dir}{Base directory for writing the data.}

\item{chr_column}{Name of the column containing the chromosome names in the annotations.}

\item{cog_column}{Name of the column containing the COG groups in the annotations.}

\item{start_column}{Name of the column containing the starts in the annotations.}

\item{stop_column}{Name of the column containing the stops in the annotations.}

\item{strand_column}{Name of the column containing the strand information.}

\item{id_column}{Where do the gene IDs live? NULL means rownames.}

\item{cog_map}{Not yet used, but used to provide an alternate map of groups/colors.}

\item{radius}{Size of the image.}

\item{chr_units}{How often to print chromosome in 'prefix' units.}

\item{band_url}{Place to imagemap link.}

\item{...}{Extra arguments passed to the tick/karyotype makers.}
}
\value{
The master configuration file name.
}
\description{
A few parameters need to be set when starting circos.  This sets
some of them and gets ready for plot stanzas.
}
\details{
In its current implementation, this really assumes that there will
be no highlight stanzas and at most 1 link stanza.
chromosomes.  A minimal amount of logic and data organization will
address these weaknesses.
}
