% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{circos_suffix}
\alias{circos_suffix}
\title{Write the end of a circos master configuration.}
\usage{
circos_suffix(cfgout = "circos/conf/default.conf")
}
\arguments{
\item{cfgout}{The master configuration file to write.}
}
\value{
undef
}
\description{
circos configuration files need an ending.  This writes it.
}

