% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{circos_tile}
\alias{circos_tile}
\title{circos_tile()  Write tiles of arbitrary categorical point data in
circos.}
\usage{
circos_tile(df, cfgout = "circos/conf/default.conf", colname = "datum",
  chr = "chr1", colors = NULL, outer = 0.9, width = 0.08, spacing = 0)
}
\arguments{
\item{df}{a dataframe with starts/ends and the floating point information}

\item{cfgout}{default='circos/conf/default.conf'  The master
configuration file to write.}

\item{colname}{default='datum'  The name of the column with the
data of interest.}

\item{chr}{default='chr1'  the name of the chromosome (This
currently assumes a bacterial chromosome)}

\item{outer}{default=1.0  the floating point radius of the circle
into which to place the plus-strand data}

\item{width}{default=0.08  the radial width of each tile}

\item{spacing}{default=0.0  the radial distance between outer,inner
and inner,whatever follows.}

\item{color}{default='blue'  the color of the histogram}

\item{fill_color}{default='blue'  guess}
}
\value{
the radius after adding the histogram and the spacing.
}
\description{
This function tries to make the writing circos tiles easier.  Like
circos_plus_minus() and circos_hist() it works in 3 stages,
It writes out a data file using cfgout as a basename and the data
from df in the circos histogram format into
circos/data/bob_tile.txt
It then writes out a configuration plot stanza in
circos/conf/bob_tile.conf
and finally adds an include to circos/bob.conf
}

