% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_from_model}
\alias{classify_from_model}
\title{Given an existing model, invoke predict on a new dataset.}
\usage{
classify_from_model(
  exp,
  model_file,
  filter = TRUE,
  convert = "cpm",
  norm = "quant"
)
}
\arguments{
\item{exp}{Input dataset}

\item{model_file}{existing model produced by the various caret
child-utilities.}

\item{filter}{Data filter used by the training set.}

\item{convert}{Data conversion used by the training set.}

\item{norm}{Data normalization used by the training set.}
}
\description{
This makes a few important assumptions: The first column of the
model's learn slot must by definition be the thing we are
classifying.  This function goes to some lengths to ensure that the
elements of the model and elements of the dataset are identical so
that predict can move forward; but one must be aware of weird
things like changing gene names over time (LpaL13.01 vs LpaL13-01
for example).
}
