% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_n_times}
\alias{classify_n_times}
\title{Rerun a model generator and classifier on a training/testing set multiple times.}
\usage{
classify_n_times(
  full_df,
  interesting_meta,
  outcome_column = "finaloutcome",
  p = 0.4,
  list = FALSE,
  formula_string = NULL,
  run_times = 10,
  method = "xgbTree",
  sampler = "cv",
  sample_number = 10,
  tuner = NULL,
  state = NULL,
  ...
)
}
\arguments{
\item{full_df}{The matrix of preProcessed() data.}

\item{interesting_meta}{dataframe of metadata of potential interest.}

\item{outcome_column}{metadata column of interest.}

\item{p}{The proportion of training/testing}

\item{list}{How to return the partitions.}

\item{formula_string}{Optional formula string, otherwise genrated on thee fly.}

\item{run_times}{How many times to repeat this process}

\item{method}{Modelling method to employ with caret.}

\item{sampler}{Sampler to employ, bootstrap or cv right now.}

\item{sample_number}{How many times to use the sampler}

\item{tuner}{Tuning arguments for the method above.}

\item{...}{Others, currently unused I think}
}
\description{
Rerun a model generator and classifier on a training/testing set multiple times.
}
