% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{classify_variants}
\alias{classify_variants}
\title{Given a pile of variants from freebayes and friends, make a table of what changed.}
\usage{
classify_variants(
  metadata,
  coverage_column = "bedtoolscoveragefile",
  variants_column = "freebayesvariantsbygene",
  min_missing = 100
)
}
\arguments{
\item{metadata}{Usually the result of gather_preprocessing_metadata(), but
whatever it is, it should have a column containing the observed
coverage and observed variants as a table.}

\item{coverage_column}{Metadata column name containing coverage information
from bedtools in a tabular format.}

\item{variants_column}{Metadata column name containing the variants/gene.}

\item{min_missing}{Bin size above which to call a region missing
 from one or more samples when looking for large-scale deletions
using coverage information.}
}
\value{
List containing some fun stuff.
}
\description{
My post-processor of the results from mpileup/freebayes provides
some hopefully fun output files.  This function seeks to leverage
them into tables which might be fun to look at.
}
