% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{clbr_count_tables}
\alias{clbr_count_tables}
\title{Archive of the count tables from the CL-Brener experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
clbr_count_tables
}
\description{
These tables were created via TopHat2 mapping of the CL-Br and CL-14 samples
using the CL-Brener genome.
}
\keyword{datasets}
