% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{clbr_gff}
\alias{clbr_gff}
\title{CL-Brener GFF file containing all haplotypes.}
\format{
An object of class \code{character} of length 1.
}
\usage{
clbr_gff
}
\description{
One of the interesting and bizarre things about CL-Brener: it is a
multi-haplotype strain, containing bits and pieces from two lineages, named
'Esmeraldo' and 'Non-Esmeraldo'.  In addition, there is a large portion which
has not been characterized and is therefore called 'Unassigned'.  Thus, when
mapping the data, we create a concatenated genome with all three haplotypes.
}
\keyword{datasets}
