% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{clbr_sample_sheet}
\alias{clbr_sample_sheet}
\title{Sample sheet from a portion of an RNASeq experiment of Trypanosoma cruzi CL-Brener.}
\format{
An object of class \code{character} of length 1.
}
\usage{
clbr_sample_sheet
}
\description{
This contains a portion of an experiment performed with Santuza in which we
were comparing two closely related T.cruzi strains: CL-14 and CL-Brener,
which are super-similar but vastly different in terms of their
pathogenicity.  (I would much rather be infected with CL-14!).
}
\keyword{datasets}
