% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{clbr_vcf_output}
\alias{clbr_vcf_output}
\title{CL-Brener/CL-14 output from vcfutils.}
\format{
An object of class \code{character} of length 1.
}
\usage{
clbr_vcf_output
}
\description{
One thing we did not include in the paper (because I didn't think of it until
later), was an analysis of the single nucleotide variants between the two
strains.  RNASeq data is of course not an ideal format for performing these
analyses, but I think I figured out a reasonable method to extract mostly
robust differences (in snp.r).
}
\keyword{datasets}
