% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{clean_pkg}
\alias{clean_pkg}
\title{Cleans up illegal characters in packages generated by make_organismdbi(), make_orgdb(),
and make_txdb(). This attempts to fix some of the common problems therein.}
\usage{
clean_pkg(path, removal = "-like", replace = "", sqlite = TRUE)
}
\arguments{
\item{path}{Location for the original Db/Dbi instance.}

\item{removal}{String to remove from the instance.}

\item{replace}{What to replace removal with, when necessary.}

\item{sqlite}{Also modify the sqlite database?}
}
\value{
A new OrgDb/TxDb/OrganismDbi
}
\description{
OrganismDbi instances are pretty neat, they pull together OrgDb and TxDb.  With any luck, this
function provides the ability to pull together all the data from the TriTrypDb, GO.db, and
KEGGREST in order to accomplish these peculiar tasks.
}
\examples{
\dontrun{
 crazytown <- make_organismdbi()  ## wait a loong time
}
}
