% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{clear_session}
\alias{clear_session}
\title{Clear an R session, this is probably unwise given what I have read about R.}
\usage{
clear_session(keepers = NULL, depth = 10)
}
\arguments{
\item{keepers}{List of namespaces to leave alone (unimplemented).}

\item{depth}{Cheesy forloop of attempts to remove packages stops after this
many tries.}
}
\value{
A spring-fresh R session, hopefully.
}
\description{
Clear an R session, this is probably unwise given what I have read about R.
}
\seealso{
[R.utils]
}
