% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proteomics.r
\name{cleavage_histogram}
\alias{cleavage_histogram}
\title{Make a histogram of how many peptides are expected at every integer dalton
from a given start to end size for a given enzyme digestion.}
\usage{
cleavage_histogram(pep_sequences, enzyme = "trypsin", start = 600,
  end = 1500, color = "black")
}
\arguments{
\item{pep_sequences}{Protein sequences as per plot_cleaved().}

\item{enzyme}{Compatible enzyme name from cleaver.}

\item{start}{Print histogram from here}

\item{end}{to here.}

\item{color}{Make the bars this color.}
}
\value{
List containing the plot and size distribution.
}
\description{
This is very similar to plot_cleaved() above, but tries to be a little bit smarter.
}
