% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology.R
\name{cluster_trees}
\alias{cluster_trees}
\title{Make fun trees a la topgo from goseq data.}
\usage{
cluster_trees(de_genes, cpdata, goid_map = "reference/go/id2go.map",
  goids_df = NULL, score_limit = 0.1, overwrite = FALSE,
  selector = "topDiffGenes", pval_column = "adj.P.Value")
}
\arguments{
\item{de_genes}{some differentially expressed genes}

\item{godata}{data from cluster Profiler}

\item{goids}{a mapping of IDs to GO in the Ramigo expected format}

\item{sigforall}{Print significance on all nodes?}
}
\value{
a plot!
}
\description{
Make fun trees a la topgo from goseq data.
}
\seealso{
\code{\link{Ramigo}}
}

