% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{colData,expt-method}
\alias{colData,expt-method}
\title{A getter to pull the sample data from an expt.}
\usage{
\S4method{colData}{expt}(x, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{withDimnames}{Again, haven't looked it up yet.}

\item{...}{Extra args.}
}
\description{
A getter to pull the sample data from an expt.
}
