% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colData<-}
\alias{colData<-}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
colData(x, ...) <- value
}
\description{
message("I am from SummarizedExperiment and am explicitly imported, wtf.")
}
