% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{color_hash}
\alias{color_hash}
\title{colors provided to hpgltools for plotting}
\format{a \code{hash} instance}
\source{
the hash package
}
\usage{
color_hash
}
\description{
A hash of color to sample names
}
\author{
atb
}

