% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.R
\name{color_hex}
\alias{color_hex}
\title{Given separate RGB values, make a hex string.}
\usage{
color_hex(red, green, blue)
}
\arguments{
\item{red}{Red value.}

\item{green}{Green value.}

\item{blue}{Blue value!}
}
\description{
Given separate RGB values, make a hex string.
}
