% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.R
\name{color_int}
\alias{color_int}
\title{Translate the hexadecimal color codes to three decimal numbers.}
\usage{
color_int(rgb)
}
\arguments{
\item{rgb}{hexadecimal color input.}
}
\description{
Translate the hexadecimal color codes to three decimal numbers.
}
