% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{colors,expt-method}
\alias{colors,expt-method}
\title{A getter to pull the colors from an expt.}
\usage{
\S4method{colors}{expt}(expt)
}
\arguments{
\item{expt}{An expt.}
}
\description{
A getter to pull the colors from an expt.
}
