% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{colors<-,expt-method}
\alias{colors<-,expt-method}
\title{A setter to put the colors into an expt.}
\usage{
\S4method{colors}{expt}(expt) <- value
}
\arguments{
\item{expt}{An expt.}

\item{value}{List of new colors.}
}
\description{
A setter to put the colors into an expt.
}
