% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{colors<-,SummarizedExperiment-method}
\alias{colors<-,SummarizedExperiment-method}
\title{A setter to put the colors into a SummarizedExperiment.}
\usage{
\S4method{colors}{SummarizedExperiment}(x, ...) <- value
}
\arguments{
\item{x}{A SummarizedExperiment.}

\item{...}{extra args.}

\item{value}{List of new colors.}
}
\description{
A setter to put the colors into a SummarizedExperiment.
}
