% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colors<-}
\alias{colors<-}
\title{Add colors to a dataset}
\usage{
colors(exp, ...) <- value
}
\arguments{
\item{exp}{Object to modify}

\item{...}{extra arguments.}

\item{value}{vector of colors}
}
\description{
Add colors to a dataset
}
