% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{colors,SummarizedExperiment-method}
\alias{colors,SummarizedExperiment-method}
\title{A getter to pull the colors from a SummarizedExperiment.}
\usage{
\S4method{colors}{SummarizedExperiment}(expt)
}
\arguments{
\item{expt}{An expt.}
}
\description{
A getter to pull the colors from a SummarizedExperiment.
}
