% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colors}
\alias{colors}
\title{Alias function to get_colors()}
\usage{
colors(input, ...)
}
\arguments{
\item{input}{Input datastructure}

\item{...}{Arguments passed along.}
}
\description{
Alias function to get_colors()
}
