% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.r
\name{combine_de_table}
\alias{combine_de_table}
\title{combine_de_table()  Given a limma, edger, and deseq table, combine them}
\usage{
combine_de_table(li, ed, de, table, annot_df = NULL, inverse = FALSE)
}
\arguments{
\item{li}{a limma output}

\item{ed}{a edger output}

\item{de}{a deseq output}
}
\description{
combine_de_table()  Given a limma, edger, and deseq table, combine them
}

