% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{combine_de_tables}
\alias{combine_de_tables}
\title{combine_de_tables()  Combine portions of deseq/limma/edger table output}
\usage{
combine_de_tables(all_pairwise_result, annot_df = NULL, excel = NULL,
  excel_title = "Table SXXX: Combined Differential Expression of YYY",
  excel_sheet = "combined_DE", keepers = "all", include_basic = TRUE,
  add_plots = TRUE, plot_dim = 3)
}
\arguments{
\item{all_pairwise_result}{the output from all_pairwise()}

\item{annot_df}{default=NULL  add some annotation information}

\item{excel}{default=NULL  print the excel workbook}

\item{excel_title}{default='Table SXXX: Combined Differential Expression of YYY' a title}

\item{excel_sheet}{default='combined_DE'  name the sheet}

\item{keepers}{default='all'  a list of reformatted table names to explicitly keep
certain contrasts in specific orders}

\item{include_basic}{default=TRUE  Include my stupid basic logFC tables}

\item{add_plots}{default=FALSE  add plots to the end of the sheets}

\item{plot_dim}{default=4  number of inches squared for the plot if added}
}
\value{
a table combinine limma/edger/deseq outputs.
}
\description{
This hopefully makes it easy to compare the outputs from limma/DESeq2/EdgeR on a table-by-table basis.
}
\examples{
\dontrun{
pretty = combine_de_tables(big_result, table='t12_vs_t0')
}
}
\seealso{
\code{\link{all_pairwise}}
}

