% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{combine_expts}
\alias{combine_expts}
\title{combine summarized experiments when using deprecated function names.}
\usage{
combine_expts(...)
}
\arguments{
\item{...}{Arguments passed to combine_se()}
}
\description{
combine summarized experiments when using deprecated function names.
}
