% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{combine_extracted_plots}
\alias{combine_extracted_plots}
\title{Gather data required to make MA/Volcano plots for pairwise comparisons.}
\usage{
combine_extracted_plots(
  pairwise,
  name,
  combined,
  denominator,
  numerator,
  basic,
  deseq,
  edger,
  limma,
  ebseq,
  include_basic = TRUE,
  include_deseq = TRUE,
  include_edger = TRUE,
  include_limma = TRUE,
  include_ebseq = FALSE,
  loess = FALSE,
  do_inverse = FALSE,
  found_table = NULL
)
}
\arguments{
\item{pairwise}{A pairwise result without modification.}

\item{name}{Name of the table to plot.}

\item{combined}{Modified pairwise result, containing the various DE methods.}

\item{denominator}{Name of the denominator coefficient.}

\item{numerator}{Name of the numerator coefficient.}

\item{basic}{Tables from basic_pairiwse().}

\item{deseq}{Tables from deseq_pairwise().}

\item{edger}{Tables from edger_pairwise().}

\item{limma}{Tables from limma_pairwise().}

\item{ebseq}{Tables from ebseq_pairwise().}

\item{include_basic}{Add basic data?}

\item{include_deseq}{Add deseq data?}

\item{include_edger}{Add edger data?}

\item{include_limma}{Add limma data?}

\item{include_ebseq}{Add ebseq data?}

\item{loess}{Add a loess estimation?}

\item{do_inverse}{Flip the numerator/denominator?}

\item{found_table}{The table name actually used.}
}
\description{
It should be possible to significantly simplify the arguments passed to this
function, but I have thus far focused only on getting it to work with the
newly split-apart combine_de_tables() functions.
}
