% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{combine_se}
\alias{combine_se}
\title{Bring together two Summarized Experiments into a larger one.}
\usage{
combine_se(
  se1,
  se2,
  condition = "condition",
  all_x = TRUE,
  all_y = TRUE,
  batch = "batch",
  merge_meta = TRUE,
  how = "cbind"
)
}
\arguments{
\item{se1}{The first SE}

\item{se2}{The second SE}

\item{condition}{Metadata factor used as the primary condition of interest.}

\item{all_x}{The all.x parameter for merge()}

\item{all_y}{The all.y parameter for merge()}

\item{batch}{Factor used as the batch for the new SE.}

\item{merge_meta}{Merge the metadata?}
}
\description{
Bring together two Summarized Experiments into a larger one.
}
