% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{combine_single_de_table}
\alias{combine_single_de_table}
\title{Given a limma, edger, and deseq table, combine them into one.}
\usage{
combine_single_de_table(li = NULL, ed = NULL, eb = NULL, de = NULL,
  ba = NULL, table_name = "", annot_df = NULL, inverse = FALSE,
  adjp = TRUE, padj_type = "fdr", include_deseq = TRUE,
  include_edger = TRUE, include_ebseq = TRUE, include_limma = TRUE,
  include_basic = TRUE, lfc_cutoff = 1, p_cutoff = 0.05,
  excludes = NULL)
}
\arguments{
\item{li}{Limma output table.}

\item{ed}{Edger output table.}

\item{eb}{EBSeq output table}

\item{de}{DESeq2 output table.}

\item{ba}{Basic output table.}

\item{table_name}{Name of the table to merge.}

\item{annot_df}{Add some annotation information?}

\item{inverse}{Invert the fold changes?}

\item{adjp}{Use adjusted p-values?}

\item{padj_type}{Add this consistent p-adjustment.}

\item{include_deseq}{Include tables from deseq?}

\item{include_edger}{Include tables from edger?}

\item{include_limma}{Include tables from limma?}

\item{include_basic}{Include the basic table?}

\item{lfc_cutoff}{Preferred logfoldchange cutoff.}

\item{p_cutoff}{Preferred pvalue cutoff.}

\item{excludes}{Set of genes to exclude from the output.}
}
\value{
List containing a) Dataframe containing the merged
 limma/edger/deseq/basic tables, and b) A summary of how many
 genes were observed as up/down by output table.
}
\description{
This combines the outputs from the various differential expression
tools and formalizes some column names to make them a little more
consistent.
}
\seealso{
\pkg{data.table} \pkg{openxlsx}
}
