% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{compare_batches}
\alias{compare_batches}
\title{Attempt to compare the results from the various batch/sv methods.}
\usage{
compare_batches(expt = NULL, methods = NULL)
}
\arguments{
\item{expt}{Input expressionset}

\item{methods}{Set of methods to try out.}
}
\description{
Given an expressionset and list of methods, try to find out how
well the various methods agree via correlation.
}
