% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{compare_de_results}
\alias{compare_de_results}
\title{Compare the results of separate all_pairwise() invocations.}
\usage{
compare_de_results(
  first,
  second,
  cor_method = "pearson",
  try_methods = c("limma", "deseq", "edger")
)
}
\arguments{
\item{first}{One invocation of combine_de_tables to examine.}

\item{second}{A second invocation of combine_de_tables to examine.}

\item{cor_method}{Method to use for cor.test().}

\item{try_methods}{List of methods to attempt comparing.}
}
\value{
A list of compared columns, tables, and methods.
}
\description{
Where compare_led_tables looks for changes between limma and friends, this
function looks for differences/similarities across the models/surrogates/etc
across invocations of limma/deseq/edger.
}
\details{
Tested in 29de_shared.R
}
\examples{
\dontrun{
 first <- all_pairwise(expt, model_batch = FALSE, excel = "first.xlsx")
 second <- all_pairwise(expt, model_batch = "svaseq", excel = "second.xlsx")
 comparison <- compare_de_results(first$combined, second$combined)
}
}
\seealso{
[all_pairwise()]
}
