% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{compare_de_tables}
\alias{compare_de_tables}
\title{Use plot_linear_scatter to compare to de tables.}
\usage{
compare_de_tables(
  first,
  second,
  fcx = "deseq_logfc",
  px = "deseq_adjp",
  fcy = "deseq_logfc",
  py = "deseq_adjp",
  first_table = NULL,
  second_table = NULL
)
}
\arguments{
\item{first}{First table to compare.}

\item{second}{Second table to compare.}

\item{fcx}{Column for the x-axis fold-change.}

\item{px}{Column for the x-axis p-value.}

\item{fcy}{Column containing the y-axis fold-change.}

\item{py}{Column containing the y-axis p-value.}

\item{first_table}{If the input data are actually of type de_table, then find the table(s) inside them.}

\item{second_table}{Ibid.}
}
\description{
Use plot_linear_scatter to compare to de tables.
}
