% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{compare_go_searches}
\alias{compare_go_searches}
\title{Compare the results from different ontology tools}
\usage{
compare_go_searches(goseq = NULL, cluster = NULL, topgo = NULL, gostats = NULL)
}
\arguments{
\item{goseq}{Result from simple_goseq()}

\item{cluster}{Result from simple_clusterprofiler()}

\item{topgo}{Result from topGO}

\item{gostats}{Result from GOstats}
}
\value{
Summary of the similarities of ontology searches
}
\description{
Combine the results from goseq, cluster profiler, topgo, and gostats; poke at
them with a stick and see what happens. The general idea is to pull the
p-value data from each tool and contrast that to the set of all possibile
ontologies.  This allows one to do a correlation coefficient between them.
In addition, take the 1-pvalue for each ontology for each tool. Thus for
strong p-values the score will be near 1 and so we can sum the scores for all
the tools.  Since topgo has 4 tools, the total possible is 7 if everything
has a p-value equal to 0.
}
\seealso{
[goseq] [clusterProfiler] [topGO] [goStats]
}
