% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{compare_led_tables}
\alias{compare_led_tables}
\title{See how similar are results from limma/deseq/edger.}
\usage{
compare_led_tables(limma = NULL, deseq = NULL, edger = NULL,
  basic = NULL, include_basic = TRUE, annot_df = NULL, ...)
}
\arguments{
\item{limma}{Data from limma_pairwise().}

\item{deseq}{Data from deseq2_pairwise().}

\item{edger}{Data from edger_pairwise().}

\item{basic}{Data from basic_pairwise().}

\item{include_basic}{include the basic data?}

\item{annot_df}{Include annotation data?}

\item{...}{More options!}
}
\value{
Heatmap showing how similar they are along with some
 correlations betwee the three players.
}
\description{
limma, DEseq2, and EdgeR all make somewhat different assumptions.
and choices about what makes a meaningful set of differentially.
expressed genes.  This seeks to provide a quick and dirty metric
describing the degree to which they (dis)agree.
}
\details{
Invoked by all_pairwise().
}
\examples{
\dontrun{
 l = limma_pairwise(expt)
 d = deseq_pairwise(expt)
 e = edger_pairwise(expt)
 fun = compare_tables(limma=l, deseq=d, edger=e)
}
}
\seealso{
\code{\link{limma_pairwise}} \code{\link{edger_pairwise}} \code{\link{deseq2_pairwise}}
}
