% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{compare_logfc_plots}
\alias{compare_logfc_plots}
\title{Compare logFC values from limma and friends}
\usage{
compare_logfc_plots(combined_tables)
}
\arguments{
\item{combined_tables}{The combined tables from limma et al.}
}
\value{
Some plots
}
\description{
There are some peculiar discrepencies among these tools, what is up with that?
}
\details{
Invoked by combine_de_tables() in order to compare the results.
}
\examples{
\dontrun{
 limma_vs_deseq_vs_edger <- compare_logfc_plots(combined)
 ## Get a list of plots of logFC by contrast of LvD, LvE, DvE
 ## It provides comparisons against the basic analysis, but who cares about that.
}
}
\seealso{
[plot_linear_scatter()]
}
