% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{compare_pc_sv}
\alias{compare_pc_sv}
\title{Incomplete function to compare PCs and SVs.}
\usage{
compare_pc_sv(
  expt,
  norm = NULL,
  transform = "log2",
  convert = "cpm",
  filter = TRUE,
  batch = "svaseq"
)
}
\arguments{
\item{expt}{Input expressionset.}

\item{norm}{Normalization performed.}

\item{transform}{Assuming using PCA and so log2 the data.}

\item{convert}{Scale the data, presumably with cpm().}

\item{filter}{Low-count filter the data?}

\item{batch}{Method which provides SVs to apply.}
}
\value{
Currently just a plot of the SVs.
}
\description{
This function is the beginning of a method to get a sense of what
happens to data when performing things like SVA.
}
