% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{compare_results_de}
\alias{compare_results_de}
\title{Compare the results of separate all_pairwise() invocations.}
\usage{
compare_results_de(first, second)
}
\arguments{
\item{first}{One invocation of combine_de_tables to examine.}

\item{second}{A second invocation of combine_de_tables to examine.}
}
\value{
A list of compared columns, tables, and methods.
}
\description{
Where compare_tables looks for changes between limma and friends, this
function looks for differences/similarities across the models/surrogates/etc
across invocations of limma/deseq/edger.
}
\details{
Tested in 29de_shared.R
}
