% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.r
\name{compare_tables}
\alias{compare_tables}
\title{compare_tables()  See how similar are results from limma/deseq/edger.}
\usage{
compare_tables(limma = NULL, deseq = NULL, edger = NULL, basic = NULL,
  include_basic = TRUE, annot_df = annot_df, ...)
}
\arguments{
\item{limma}{default=NULL  limma data from limma_pairwise()}

\item{deseq}{default=NULL  deseq data from deseq2_pairwise()}

\item{edger}{default=NULL  edger data from edger_pairwise()}
}
\value{
a heatmap showing how similar they are along with some
correlations betwee the three players.
}
\description{
limma, DEseq2, and EdgeR all make somewhat different assumptions
and choices about what makes a meaningful set of differentially
expressed genes.  This seeks to provide a quick and dirty metric
describing the degree to which they (dis)agree.
}
\examples{
## l = limma_pairwise(expt)
## d = deseq_pairwise(expt)
## e = edger_pairwise(expt)
fun = compare_tables(limma=l, deseq=d, edger=e)
}
\seealso{
\code{\link{limma_pairwise}} \code{\link{edger_pairwise}} \code{\link{deseq2_pairwise}}
}

