% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{concatenate_runs,SummarizedExperiment,character-method}
\alias{concatenate_runs,SummarizedExperiment,character-method}
\title{Sum the reads/gene for multiple sequencing runs of a single condition/batch.}
\usage{
\S4method{concatenate_runs}{SummarizedExperiment,character}(expt, column = "replicate")
}
\arguments{
\item{expt}{Experiment class containing the requisite metadata and count tables.}

\item{column}{Column of the design matrix used to specify which samples are replicates.}
}
\value{
Expt with the concatenated counts, new design matrix, batches, conditions, etc.
}
\description{
On occasion we have multiple technical replicates of a sequencing run.  This
can use a column in the experimental design to identify those replicates and
sum the counts into a single column in the count tables.
}
\details{
Untested as of 2016-12-01, but used in a couple of projects where sequencing
runs got repeated.
}
\examples{
\dontrun{
 compressed <- concatenate_runs(expt)
}
}
\seealso{
[Biobase] [exprs()] [fData()] [pData()] [create_expt()]
}
