% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{conditions,expt-method}
\alias{conditions,expt-method}
\title{A getter to pull the conditions from an expt.}
\usage{
\S4method{conditions}{expt}(object, ...)
}
\arguments{
\item{object}{Input expt}

\item{...}{extra args}
}
\description{
A getter to pull the conditions from an expt.
}
