% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{conditions<-,SummarizedExperiment-method}
\alias{conditions<-,SummarizedExperiment-method}
\title{A setter to put the conditions into a SummarizedExperiment.}
\usage{
\S4method{conditions}{SummarizedExperiment}(object, ...) <- value
}
\arguments{
\item{object}{A SummarizedExperiment.}

\item{...}{arbitrary arguments}

\item{value}{List of new conditions.}
}
\description{
A setter to put the conditions into a SummarizedExperiment.
}
