% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{conditions<-}
\alias{conditions<-}
\title{The infix version of set_conditions.}
\usage{
conditions(exp, ...) <- value
}
\arguments{
\item{exp}{Input data structure}

\item{...}{Other args passed along to set_conditions()}
}
\description{
The infix version of set_conditions.
}
