% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{conditions,SummarizedExperiment-method}
\alias{conditions,SummarizedExperiment-method}
\title{Get the experimental conditions from a SE}
\usage{
\S4method{conditions}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Input SE}

\item{...}{extra args.}
}
\description{
Get the experimental conditions from a SE
}
