% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{conditions}
\alias{conditions}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
conditions(object, ...)
}
\arguments{
\item{object}{Input object}

\item{...}{extra args}
}
\description{
message("I am from BiocGenerics and am explicitly imported, wtf.")
}
