% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{convert_gsc_ids}
\alias{convert_gsc_ids}
\title{Use AnnotationDbi to translate geneIDs from type x to type y.}
\usage{
convert_gsc_ids(
  gsc,
  orgdb = "org.Hs.eg.db",
  from_type = NULL,
  to_type = "ENTREZID"
)
}
\arguments{
\item{gsc}{geneSetCollection with IDs of a type one wishes to change.}

\item{orgdb}{Annotation object containing the various IDs.}

\item{from_type}{Name of the ID which your gsc is using.  This can probably
be automagically detected...}

\item{to_type}{Name of the ID you wish to use.}
}
\value{
Fresh gene set collection replete with new names.
}
\description{
This is intended to convert all the IDs in a geneSet from one ID type to
another and giving back the geneSet with the new IDs.
FIXME: This should use convert_ids() to simplify itself
}
\details{
One caveat: this will collapse redundant IDs via unique().
}
\seealso{
[AnnotationDbi] [guess_orgdb_keytypes()] [convert_ids()] [GSEABase]
}
