% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{convert_ids}
\alias{convert_ids}
\title{GSVA expects Entrez IDs, but we usually have Ensembl IDs.  Use this to
convert them.}
\usage{
convert_ids(ids, from = "ENSEMBL", to = "ENTREZID", orgdb = "org.Hs.eg.db")
}
\arguments{
\item{ids}{Set of gene IDs in the 'from' format.}

\item{from}{Name of the ID type as found in the 'orgdb'.}

\item{to}{Format required by GSVA (or other tool/method if so desired).}

\item{orgdb}{AnnotationDBI package containing the various IDs.}
}
\value{
Spit out a list of new IDs and a message telling how successful we were.
}
\description{
There is a non-zero chance that I have at least one other similar
implementation hiding in this package.  This should be made more consistent,
but this particular function provides a quick and easy way to make sure that
GSVA has the gene IDs that it wants.
}
