% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{cordist}
\alias{cordist}
\title{Similarity measure which combines elements from Pearson correlation and
Euclidean distance.}
\usage{
cordist(
  data,
  cor_method = "pearson",
  dist_method = "euclidean",
  cor_weight = 0.5,
  ...
)
}
\arguments{
\item{data}{Matrix of data}

\item{cor_method}{Which correlation method to use?}

\item{dist_method}{Which distance method to use?}

\item{cor_weight}{0-1 weight of the correlation, the distance weight
will be 1-cor_weight.}

\item{...}{extra arguments for cor/dist}
}
\value{
Matrix of the correlation-modified distances of the original matrix.
}
\description{
Here is Keith's summary:
Where the cor returns the Pearson correlation matrix for the input matrix,
and the dist function returns the Euclidean distance matrix for the input
matrix. The LHS of the equation is simply the sign of the correlation
function, which serves to preserve the sign of the interaction. The RHS
combines the Pearson correlation and the log inverse Euclidean distance with
equal weights. The result is a number in the range from -1 to 1 where values
close to -1 indicate a strong negative correlation and values close to 1
indicate a strong positive corelation.  While the Pearson correlation and
Euclidean distance each contribute equally in the above equation, one could
also assign tuning parameters to each of the metrics to allow for unequal
contributions.
}
\author{
Keigth Hughitt
}
