% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{correlate_de_tables}
\alias{correlate_de_tables}
\title{See how similar are results from limma/deseq/edger/ebseq.}
\usage{
correlate_de_tables(results, annot_df = NULL, extra_contrasts = NULL)
}
\arguments{
\item{results}{Data from do_pairwise()}

\item{annot_df}{Include annotation data?}

\item{extra_contrasts}{include some extra contrasts when comparing results.}
}
\value{
Heatmap showing how similar they are along with some
 correlations betwee the three players.
}
\description{
limma, DEseq2, and EdgeR all make somewhat different assumptions.
and choices about what makes a meaningful set of differentially.
expressed genes.  This seeks to provide a quick and dirty metric
describing the degree to which they (dis)agree.
}
\details{
Invoked by all_pairwise().
}
\examples{
\dontrun{
 l = limma_pairwise(expt)
 d = deseq_pairwise(expt)
 e = edger_pairwise(expt)
 fun = compare_led_tables(limma = l, deseq = d, edger = e)
}
}
\seealso{
[limma_pairwise()] [edger_pairwise()] [deseq_pairwise()]
}
