% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{count_expt_snps}
\alias{count_expt_snps}
\title{Gather snp information for an expt}
\usage{
count_expt_snps(expt, type = "counts",
  input_dir = "preprocessing/outputs", tolower = TRUE)
}
\arguments{
\item{expt}{an expressionset from which to extract information.}

\item{type}{Use counts / samples or ratios?}

\item{input_dir}{Directory to scan for snps output files.}

\item{tolower}{Lowercase stuff like 'HPGL'?}
}
\value{
A new expt object
}
\description{
This function attempts to gather a set of variant positions using an extant
expressionset.  This therefore seeks to keep the sample metadata consistent
with the original data.  In its current iteration, it therefore makes some
potentially bad assumptions about the naming conventions for its input
files.  It furthermore assumes inputs from the variant calling methods in
cyoa.
}
