% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{count_nmer,character-method}
\alias{count_nmer,character-method}
\title{Count nmers given a filename instead of genome object.}
\usage{
\S4method{count_nmer}{character}(genome, pattern = "ATG", mismatch = 0)
}
\arguments{
\item{genome}{filename of the genome in question}

\item{pattern}{Pattern for which to search.}

\item{mismatch}{Number of mismatches allowed.}
}
\description{
Count nmers given a filename instead of genome object.
}
