% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/170_sequence.r
\name{count_nmer}
\alias{count_nmer}
\title{Count n-mers in a given data set using Biostrings}
\usage{
count_nmer(genome, pattern = "ATG", mismatch = 0)
}
\arguments{
\item{genome}{Sequence database, genome in this case.}

\item{pattern}{Count off this string.}

\item{mismatch}{How many mismatches are acceptable?}
}
\value{
Set of counts by sequence.
}
\description{
This just calls PDict() and vcountPDict() on a sequence database given a
pattern and number of mismatches.  This may be used by divide_seq()
normalization.
}
