% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{counts_from_surrogates}
\alias{counts_from_surrogates}
\title{A single place to extract count tables from a set of surrogate variables.}
\usage{
counts_from_surrogates(data, adjust, design = NULL)
}
\arguments{
\item{data}{Original count table, may be an expt/expressionset or df/matrix.}
}
\description{
Given an initial set of counts and a series of surrogates, what would the resulting count table
    look like? Hopefully this function answers that question.
}

