% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{cp_options}
\alias{cp_options}
\title{Set up appropriate option sets for clusterProfiler}
\usage{
cp_options(species)
}
\arguments{
\item{species}{Currently it only works for humans and fruit flies.}
}
\description{
This hard-sets some defaults for orgdb/kegg databases when using
clusterProfiler.
}
