% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_experiment}
\alias{create_experiment}
\title{Wrap bioconductor's expressionset to include some other extraneous
information.}
\usage{
create_experiment(file, color_hash)
}
\arguments{
\item{file}{a comma separated file describing the samples with
information like condition,batch,count_filename,etc}

\item{color_hash}{a hash which describes how to color the samples}
}
\value{
experiment an expressionset
}
\description{
Wrap bioconductor's expressionset to include some other extraneous
information.
}
\examples{
## new_experiment = create_experiment("some_csv_file.csv", color_hash)
}
\seealso{
\code{\link{pData}}, \code{\link{fData}},
\code{\link{exprs}}, \code{\link{my_read_files}},
\code{\link{as.list.hash}}
}

