% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{create_expt}
\alias{create_expt}
\title{Create a summarized experiment when using deprecated function names.}
\usage{
create_expt(...)
}
\description{
Create a summarized experiment when using deprecated function names.
}
