% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{create_partitions}
\alias{create_partitions}
\title{Use createDataPartition to create test/train sets and massage them a little.}
\usage{
create_partitions(
  full_df,
  interesting_meta,
  outcome_factor = "condition",
  p = 0.6,
  list = FALSE,
  times = 5,
  outcome_column = "outcome_factor"
)
}
\description{
This will also do some massaging of the data to make it easier to work with for
downstream tasks.  Most notably, since I am mostly evaluating classifiers of
clinical data to see how well they agree with extant annotations, I want to make sure
the relevant columns are renamed in the testing sets.
}
