% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{create_scd}
\alias{create_scd}
\title{Create a combined seurat object from a sample sheet.}
\usage{
create_scd(
  metadata,
  expression_column = "gexfile",
  vdj_t_column = "vdjtcells",
  vdj_b_column = "vdjbcells",
  prefix = NULL,
  separate = FALSE,
  types = "gex",
  mito_pattern = "^mt-",
  ribo_pattern = "^Rp[sl]"
)
}
\arguments{
\item{metadata}{Sample sheet.}

\item{expression_column}{Metadata column containing the base
location of the cellranger outputs.}

\item{vdj_t_column}{Column, which if filled in, tells this to look
for VDJ information specific to light chains.}

\item{vdj_b_column}{Column, which if filled in, tells this to look
for VDJ information specific to heavy chains.}

\item{prefix}{Arbitrary prefix for the location information,
included because I am messing with cellranger and have multiple
output directories and want to be able to switch between them.}

\item{separate}{When true, this function should return a list
comprised of the individual sample objects.}

\item{types}{Types of data to add to the scd.}

\item{mito_pattern}{Pattern used to find mitochondrial genes.}

\item{ribo_pattern}{Pattern used to find ribosomal proteins.}
}
\value{
Either a list or merged seurat object(s).
}
\description{
I would like to have a simpler function for creating seurat data
structures similar to my create_expt().  This will try to do so.
}
