% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{create_se}
\alias{create_se}
\title{Create a SummarizedExperiment given some metadata}
\usage{
create_se(
  metadata = NULL,
  gene_info = NULL,
  count_dataframe = NULL,
  sanitize_rownames = FALSE,
  sample_colors = NULL,
  title = NULL,
  notes = NULL,
  countdir = NULL,
  include_type = "all",
  include_gff = NULL,
  file_column = "file",
  id_column = NULL,
  savefile = NULL,
  low_files = FALSE,
  annotation = "org.Hs.eg.db",
  palette = "Dark2",
  round = FALSE,
  tx_gene_map = NULL,
  ...
)
}
\arguments{
\item{metadata}{Filename or table of metadata about the samples of interest.}

\item{gene_info}{Annotations for the genes in the count data.}

\item{count_dataframe}{Optional table of counts.}

\item{sanitize_rownames}{Clean up unruly gene IDs?}

\item{sample_colors}{Specify the colors for the samples?}

\item{title}{Provide a title for the experiment.}

\item{notes}{Provide arbitrary notes.}

\item{countdir}{(deprecated) Directory containing count tables.}

\item{include_type}{Used to specify types of genes/annotations to use.}

\item{include_gff}{Keep a copy of the gff with the data?}

\item{file_column}{Metadata column containing the counts for each sample.}

\item{id_column}{Non-default column containing the sample IDs.}

\item{savefile}{Filename to which to save a rda file of the data structure.}

\item{low_files}{I don't remember this, I bet it is deprecated.}

\item{annotation}{orgDB associated with this, primarily used with gsva-like tools.}

\item{palette}{Color palette when auto-choosing colors for the samples.}

\item{round}{Round the data if/when it is not integer?}

\item{tx_gene_map}{When using tximport, use this to convert from transcripts to genes.}

\item{...}{Extra options.}
}
\description{
This function was taken from create_expt() and repurposed to create SummarizedExperiments.
}
