% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.R
\name{darker_color}
\alias{darker_color}
\title{Given a hex color, make it x percent darker.}
\usage{
darker_color(color, percent = 20)
}
\arguments{
\item{color}{Existing hex color string.}

\item{percent}{How much darker to make it.}
}
\value{
a new hex string.
}
\description{
Given a hex color, make it x percent darker.
}
