% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.r
\name{de_venn}
\alias{de_venn}
\title{Create venn diagrams describing how well deseq/limma/edger agree.}
\usage{
de_venn(table, adjp = FALSE, euler = FALSE, p = 0.05, ...)
}
\arguments{
\item{table}{Which table to query?}

\item{adjp}{Use adjusted p-values}

\item{euler}{Perform a euler plot}

\item{p}{p-value cutoff, I forget what for right now.}

\item{...}{More arguments are passed to arglist.}
}
\description{
The sets of genes provided by limma and friends would ideally always agree,  but they do not.
Use this to see out how much the (dis)agree.
}

