% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.r
\name{default_norm}
\alias{default_norm}
\title{Perform a default normalization of some data}
\usage{
default_norm(expt, ...)
}
\arguments{
\item{expt}{An expressionset containing expt object}

\item{...}{More options to pass to normalize_expt()}
}
\value{
The normalized expt
}
\description{
This just calls normalize expt with the most common arguments except log2 transformation, but
that may be appended with 'transform=log2', so I don't feel bad.  Indeed, it will allow you to
overwrite any arguments if you wish.  In our work, the most common normalization is:
quantile(cpm(low-filter(data))).
}
\seealso{
\code{\link{normalize_expt}}
}
