% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{define_expt_colors}
\alias{define_expt_colors}
\title{Get a named vector of colors by condition.}
\usage{
define_expt_colors(expt, fact = "condition", levels = NULL)
}
\arguments{
\item{expt}{Expression from which to gather colors.}

\item{fact}{Use this metadata column to set the colors.}

\item{levels}{When not null, colors may be set to arbitrary samples.}
}
\value{
List of colors by condition.
}
\description{
Usually we give a vector of all samples by colors.  This just
simplifies that to one element each.  Currently only used in
combine_de_tables() but I think it will have use elsewhere.
}
