% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{deparse_go_value}
\alias{deparse_go_value}
\title{Extract more easily readable information from a GOTERM datum.}
\usage{
deparse_go_value(value)
}
\arguments{
\item{value}{the result of try(as.character(somefunction(GOTERM[id])), silent=TRUE)
somefunction would be 'Synonym' 'Secondary' 'Ontology', etc...}
}
\value{
something more sane (hopefully)
}
\description{
The output from the GOTERM/GO.db functions is inconsistent, to put it nicely.
This attempts to extract from that heterogeneous datatype something easily readable.
Example:  Synonym() might return any of the following:
NA, NULL, "NA", "NULL", c("NA",NA,"GO:00001"), "GO:00002", c("Some text",NA, NULL, "GO:00003")
This function will boil that down to 'not found', '', 'GO:00004', or "GO:0001,  some text, GO:00004"
}
\examples{
\dontrun{
 goterms = GOTERM[ids]
 sane_goterms = deparse_go_value(goterms)
}
}

