% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_deseq.r
\name{deseq_coefficient_scatter}
\alias{deseq_coefficient_scatter}
\title{Plot out 2 coefficients with respect to one another from limma}
\usage{
deseq_coefficient_scatter(output, x = 1, y = 2, gvis_filename = NULL,
  gvis_trendline = TRUE, tooltip_data = NULL, flip = FALSE,
  base_url = NULL)
}
\arguments{
\item{output}{the set of pairwise comparisons provided by limma_pairwise()}

\item{x}{the name or number of the first coefficient column to extract, this will be the x-axis of the plot}

\item{y}{the name or number of the second coefficient column to extract, this will be the y-axis of the plot}

\item{gvis_filename}{A filename for plotting gvis interactive graphs of the data.}

\item{gvis_trendline}{add a trendline to the gvis plot?}

\item{tooltip_data}{a dataframe of gene annotations to be used in the gvis plot}

\item{flip}{flip the axes}

\item{base_url}{for gvis plots}
}
\value{
a ggplot2 plot showing the relationship between the two coefficients
}
\description{
It can be nice to see a plot of two coefficients from a limma comparison with respect to one another
This hopefully makes that easy.
}
\examples{
\dontrun{
 pretty = coefficient_scatter(limma_data, x="wt", y="mut")
}
}
\seealso{
\link{hpgl_linear_scatter} \link{limma_pairwise}
}

