% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_deseq.r
\name{deseq_pairwise}
\alias{deseq_pairwise}
\title{deseq_pairwise()  Because I can't be trusted to remember '2'.}
\usage{
deseq_pairwise(...)
}
\arguments{
\item{...}{I like cats.}
}
\value{
stuff deseq2_pairwise results.
}
\description{
This calls deseq2_pairwise(...) because I am determined to forget typing deseq2.
}
\seealso{
\code{\link{deseq2_pairwise}}
}
