% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/068_de_deseq.r
\name{deseq_try_sv}
\alias{deseq_try_sv}
\title{Given a set of surrogate variables from sva and friends, try adding them to a DESeqDataSet.}
\usage{
deseq_try_sv(data, summarized, svs, num_sv = NULL)
}
\arguments{
\item{data}{DESeqDataSet to test out.}

\item{summarized}{Existing DESeq metadata to append svs.}

\item{svs}{Surrogates from sva and friends to test out.}

\item{num_sv}{Optionally, provide the number of SVs, primarily used if
recursing in the hunt for a valid number of surrogates.}
}
\value{
DESeqDataSet with at least some of the SVs appended to the model.
}
\description{
Sometimes sva returns a set of surrogate variable estimates which lead to
models which are invalid according to DESeq2.  This function will try before
buying and tell the user if the sva model additions are valid according to
DESeq.
}
