% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.R
\name{df_to_mappings}
\alias{df_to_mappings}
\title{Coerce a df of gene IDs and GO categories to the topGO list}
\usage{
df_to_mappings(go_db)
}
\arguments{
\item{go_db}{The df used by goseq and most commonly by me.}
}
\description{
The read_mappings function in topGO produces a list of x element
where x is named by the gene IDs and the value of each element
is a character vector comprised of the categories associated with it.
}
