% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{dia_counts}
\alias{dia_counts}
\title{A few scored tsv files from Dr. Briken's DIA-SWATH experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
dia_counts
}
\description{
I used a fairly exhaustive set of open source tools to interpret Dr. Briken's
data.  These files comprise the endpoint of the preprocessing and the inputs
for the R package 'SWATH2stats'.  This file is excessively large, but the
smallest by far of the various inputs I wanted to include to test my various
proteomics functions.
}
\keyword{datasets}
