% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{dia_samples}
\alias{dia_samples}
\title{Portion of a sample sheet used in a DIA-SWATH proteomics experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
dia_samples
}
\description{
In this experiment, Dr. Briken sought to learn about proteins which are
exported by Mycobacterium tuberculosis.  He therefore performed a DIA SWATH
experiment using two strains and collected the supernatant fraction (to get
the exported proteins) and the intracellular fraction.  This file contains
the metadata for a portion of that experiment.
}
\keyword{datasets}
