% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dirty_go}
\alias{dirty_go}
\title{Perform a quick and dirty gene ontology enrichment search}
\usage{
dirty_go(de_genes, lengths = NULL, goids = NULL, adjust = 0.05)
}
\arguments{
\item{de_genes}{a dataframe of gene IDs which have been deemed
'differential' by whatever metric you choose.}

\item{lengths}{a dataframe of gene id and lengths, null by default.}

\item{goids}{a dataframe of gene to goid mappings}

\item{adjust}{minimum p value for (bh by default) adjustment}
}
\value{
a list including all the goseq data, subsets deemed
interesting by the pvalue, and associated goterms.
}
\description{
Perform a quick and dirty gene ontology enrichment search
}
\examples{
## dirty_godata = dirty_go(de_genes, lengths=length_df, goids=genetogo_df)
}
\seealso{
\code{\link{nullp}}, \code{\link{goseq}},
}

