% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{disjunct_pvalues}
\alias{disjunct_pvalues}
\title{Test for infected/control/beads -- a placebo effect?}
\usage{
disjunct_pvalues(contrast_fit, cellmeans_fit, conj_contrasts, disj_contrast)
}
\arguments{
\item{contrast_fit}{Result of lmFit.}

\item{cellmeans_fit}{Result of a cellmeans fit.}

\item{conj_contrasts}{Result from the makeContrasts of the first set.}

\item{disj_contrast}{Result of the makeContrasts of the second set.}
}
\description{
This was a function I copied out of Keith/Hector/Laura/Cecilia's paper in
which they sought to discriminate the effect of inert beads on macrophages
vs. the effect of parasites.  The simpler way of expressing it is: take the
worst p-value observed for the pair of contrasts, infected/uninfected and
beads/uninfected.
}
\details{
The goal is therefore to find responses different than beads
The null hypothesis is (H0): (infected == uninfected) | (infected == beads)
The alt hypothesis is (HA): (infected != uninfected) & (infected != beads)
}
