% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{disjunct_tab}
\alias{disjunct_tab}
\title{Test for infected/control/beads -- a placebo effect?
The goal is therefore to find responses different than beads
The null hypothesis is (H0): (infected == uninfected) || (infected == beads)
The alt hypothesis is (HA): (infected != uninfected) && (infected != beads)}
\usage{
disjunct_tab(contrast_fit, coef1, coef2, ...)
}
\description{
Test for infected/control/beads -- a placebo effect?
The goal is therefore to find responses different than beads
The null hypothesis is (H0): (infected == uninfected) || (infected == beads)
The alt hypothesis is (HA): (infected != uninfected) && (infected != beads)
}

