% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_count_lines}
\alias{dispatch_count_lines}
\title{Count the number of lines in an input file spec and add it to the metadata.}
\usage{
dispatch_count_lines(
  meta,
  search,
  input_file_spec,
  verbose = verbose,
  species = "*",
  basedir = "preprocessing",
  type = "genome",
  inverse = FALSE
)
}
\arguments{
\item{meta}{Input metadata}

\item{search}{Pattern to count}

\item{input_file_spec}{Input file specification to hunt down the
file of interest.}

\item{verbose}{Print diagnostic information while running?}

\item{species}{Specify a species to search for.}

\item{basedir}{Root directory containing the files/logs of metadata.}

\item{type}{Add columns for only the genome mapping and/or rRNA by default.}

\item{inverse}{Count the lines that do _not_ match the pattern.}
}
\description{
Sometimes the number of lines of a file is a good proxy for some
aspect of a sample. For example, jellyfish provides 1 line for
every kmer observed in a sample.  This function extracts that
number and puts it into each cell of a sample sheet.
}
