% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{dispatch_csv_search}
\alias{dispatch_csv_search}
\title{Pull some information from a csv/tsv file.}
\usage{
dispatch_csv_search(
  meta,
  column,
  input_file_spec,
  type = "csv",
  which = "first",
  verbose = FALSE,
  ...
)
}
\description{
Pull some information from a csv/tsv file.
}
