% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{dispatch_fasta_lengths}
\alias{dispatch_fasta_lengths}
\title{Get the lengths of sequences from a fasta file.}
\usage{
dispatch_fasta_lengths(meta, input_file_spec, verbose = verbose)
}
\description{
Get the lengths of sequences from a fasta file.
}
