% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{dispatch_filename_search}
\alias{dispatch_filename_search}
\title{Pull out the filename matching an input spec}
\usage{
dispatch_filename_search(meta, input_file_spec, verbose = verbose)
}
\description{
This is useful for putting the count table name into a metadata file.
}
