% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_gc}
\alias{dispatch_gc}
\title{Pull GC content into the metadata sheet.}
\usage{
dispatch_gc(meta, input_file_spec, verbose = FALSE, basedir = "preprocessing")
}
\arguments{
\item{meta}{Input metadata}

\item{input_file_spec}{Input file specification to hunt down the
file of interest.}

\item{verbose}{Print diagnostic information while running?}

\item{basedir}{Root directory containing the files/logs of metadata.}
}
\description{
As the name suggests, this only works for fasta files.
}
