% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_md5}
\alias{dispatch_md5}
\title{Pull raw MD5 sums into the metadata sheet.}
\usage{
dispatch_md5(
  meta,
  entry_type,
  input_file_spec,
  specification,
  basedir = "preprocessing",
  species = "*",
  type = "genome",
  subtype = "gene",
  tag = "ID",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{meta}{Input metadata}

\item{entry_type}{Perhaps not used, I copy/pasted the function declaration.}

\item{input_file_spec}{Input file specification to hunt down the
file of interest.}

\item{specification}{Full specification list}

\item{basedir}{Used to find the path to the input to md5sum}

\item{species}{Probably not needed}

\item{type}{Almost definitely not needed.}

\item{subtype}{Clean this too}

\item{tag}{And this}

\item{verbose}{Print diagnostic information while running?}

\item{...}{Extra arguments passed through, not currently used.}
}
\description{
Pull raw MD5 sums into the metadata sheet.
}
