% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{dispatch_metadata_extract}
\alias{dispatch_metadata_extract}
\title{This is basically just a switch and set of regexes for finding the
numbers of interest in the various log files.}
\usage{
dispatch_metadata_extract(
  meta,
  entry_type,
  input_file_spec,
  specification,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{meta}{Starting metadata}

\item{entry_type}{String which defines the type of log entry to
hunt down.  If the specification does not include a column, this
will be used as the column name to write to the metadata.}

\item{input_file_spec}{Glue specification defining the log file for
each sample to hunt down.}

\item{specification}{This is the reason I am thinking having this
as a separate function might be stupid.  I added it to make it
easier to calculate ratios of column_x/column_y; but it is a
def-facto argument to either get rid of input_file_spec as an arg
or to just get rid of this function.}

\item{verbose}{used for testing regexes.}

\item{...}{passed to glue to add more variables to the file spec.}
}
\value{
Vector of entries which will be used to populate the new
 column in the metadata.
}
\description{
When I initially wrote this, it made sense to me to have it
separate from the top-level function.  I am not sure that is true
now, having slept on it.
}
