% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_regex_search}
\alias{dispatch_regex_search}
\title{Generic dispatcher to hunt down useful information from logs.}
\usage{
dispatch_regex_search(
  meta,
  search,
  replace,
  input_file_spec,
  basedir = "preprocessing",
  extraction = "\\\\1",
  which = "first",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{meta}{Input metadata.}

\item{search}{regex used to go hunting for the line of interest.}

\item{replace}{probably the same regex with parentheses in place
for gsub().}

\item{input_file_spec}{filename extractor expression.}

\item{extraction}{the replacement portion of gsub(). I am thinking
to make it possible to have this function return more interesting
outputs if this changes, but for the moment I am sort of assuming
\\1 will always suffice.}

\item{which}{Usually 'first', which means grab the first match and get out.}

\item{verbose}{For testing regexes.}

\item{...}{Used to pass extra variables to glue for finding files.}
}
\description{
Given the metadata, a couple of regular expressions, and a filename
specification, this should be able to pull out the interesting
number(s) from one logfile per sample from the metadata.
}
