% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_sum_column}
\alias{dispatch_sum_column}
\title{Given an arbitrary input column, get the sum of its values.}
\usage{
dispatch_sum_column(
  meta,
  input_file_spec,
  verbose = verbose,
  species = "*",
  basedir = "preprocessing",
  type = "genome",
  subtype = "gene",
  new_spec = NULL,
  tag = "ID"
)
}
\arguments{
\item{meta}{Input metadata}

\item{input_file_spec}{Filename specification to seek out the input data.}

\item{verbose}{Be verbose about what is happening?}

\item{species}{Choose a specific species, or all.}

\item{basedir}{Start searching here.}

\item{type}{Analysis type to search.}

\item{subtype}{Specific annotation type to search.}

\item{new_spec}{New specification resulting from this invocation.}

\item{tag}{GFF tag used to extract IDs.}
}
\description{
Note: This is not taking some important things into account yet.
}
