% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{dm_orgdb}
\alias{dm_orgdb}
\title{Name of the Drosphila melanogaster orgdb.}
\format{
An object of class \code{character} of length 1.
}
\usage{
dm_orgdb
}
\description{
Some of my examples use this and I forgot the make it explicit...
}
\keyword{datasets}
