% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{do_eupath_table}
\alias{do_eupath_table}
\title{Invoke the appropriate get_eupath_ function.}
\usage{
do_eupath_table(type, granges = NULL, provider = NULL,
  genus_species = NULL, species = NULL, flatten = FALSE,
  abbreviation = NULL, dir = ".")
}
\arguments{
\item{type}{One of: gff, genes, go, pathways, etc. (The type of table to download.)}

\item{granges}{When downloading a gff, use this.}

\item{provider}{Generally only tritrypdb or eupathdb}

\item{genus_species}{A string identifying the genus and species}

\item{species}{As opposed to just the species}

\item{flatten}{Flatten the resulting table (if it is a table of tables)}

\item{abbreviation}{Used for kegg downloads.}

\item{dir}{Where to dump the data.}
}
\description{
This small function provides a single entrypoint for multiple get_eupath funcitons.
}
