% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{do_topgo}
\alias{do_topgo}
\title{An attempt to make topgo invocations a bit more standard.}
\usage{
do_topgo(
  type,
  go_map = NULL,
  fisher_genes = NULL,
  ks_genes = NULL,
  selector = "topDiffGenes",
  sigforall = TRUE,
  numchar = 300,
  pval_column = "adj.P.Val",
  overwrite = FALSE,
  cutoff = 0.05,
  densities = FALSE,
  pval_plots = TRUE
)
}
\arguments{
\item{type}{Type of topgo search to perform: fisher, KS, EL, or weight.}

\item{go_map}{Mappings of gene and GO IDs.}

\item{fisher_genes}{List of genes used for fisher analyses.}

\item{ks_genes}{List of genes used for KS analyses.}

\item{selector}{Function to use when selecting genes.}

\item{sigforall}{Provide significance metrics for all ontologies observed,
not only the ones deemed statistically significant.}

\item{numchar}{A limit on characters printed when printing topgo tables
(used?)}

\item{pval_column}{Column from which to extract DE p-values.}

\item{overwrite}{Overwrite an existing gene ID/GO mapping?}

\item{cutoff}{Define 'significant'?}

\item{densities}{Perform gene density plots by ontology?}

\item{pval_plots}{Print p-values plots as per clusterProfiler?}
}
\value{
A list of results from the various tests in topGO.
}
\description{
My function 'simple_topgo()' was excessively long and a morass of copy/pasted
  fragments.  This attempts to simplify that and converge on a single piece
  of code for all the methodologies provided by topgo.
}
