% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\name{.data}
\alias{.data}
\title{Taken from Nathan Eastwood to help using mutate and friends.}
\description{
It is fairly common for me to get annoyed with R CMD check due to NSE,
thus the previous declarations and the rando NULL assignments in this package.
https://nathaneastwood.github.io/2019/08/18/no-visible-binding-for-global-variable/
}
