% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{download_eupathdb_metadata}
\alias{download_eupathdb_metadata}
\title{Returns metadata for all eupathdb organisms.}
\usage{
download_eupathdb_metadata(overwrite = FALSE, webservice = "eupathdb",
  dir = "eupathdb", use_savefile = TRUE)
}
\arguments{
\item{overwrite}{Overwrite existing data?}

\item{webservice}{Optional alternative webservice for hard-to-find species.}

\item{dir}{Where to put the json.}

\item{use_savefile}{Make a savefile of the data for future reference.}
}
\value{
Dataframe with lots of rows for the various species in eupathdb.
}
\description{
Returns metadata for all eupathdb organisms.
}
\author{
Keith Hughitt
}
