% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.R
\name{download_microbesonline_files}
\alias{download_microbesonline_files}
\title{Download the various file formats from microbesoline.}
\usage{
download_microbesonline_files(id = "160490", type = NULL)
}
\arguments{
\item{id}{Species ID to query.}

\item{type}{File type(s) to download, if left null it will grab the genbank,
tab, protein fasta, transcript fasta, and genome.}
}
\value{
List describing the files downloaded and their locations.
}
\description{
Microbesonline provides an interesting set of file formats to download.  Each
format proves useful under one condition or another, ergo this defaults to
iterating through them all and getting every file.
}
