% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_tritrypdb.r
\name{download_tritrypdb_text}
\alias{download_tritrypdb_text}
\title{Download the various data files from http://tritrypdb.org/}
\usage{
download_tritrypdb_text(version = "27", species = "lmajor",
  strain = "friedlin", dl_dir = "organdb/tritryp", quiet = TRUE)
}
\arguments{
\item{version}{What version of the tritrypdb to use?}

\item{species}{Human readable species to use.}

\item{strain}{Strain of the given species to download.}

\item{dl_dir}{Directory into which to download the various files.}

\item{quiet}{Print download progress?}
}
\value{
List of downloaded files.
}
\description{
The tritrypdb nicely makes their downloads standardized!
}
\examples{
\dontrun{
 filenames <- tritryp_downloads(species="lmajor", strain="friedlin", version="28")
}
}
