% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_varpart.R
\name{dream_pairwise}
\alias{dream_pairwise}
\title{Set up a model matrix and set of contrasts for pairwise comparisons using
voom/limma via their modified functions from variancePartition.}
\usage{
dream_pairwise(
  input = NULL,
  model_fstring = "~ 0 + condition + batch",
  null_fstring = "~",
  model_svs = NULL,
  extra_contrasts = NULL,
  annot_df = NULL,
  libsize = NULL,
  filter = TRUE,
  num_surrogates = "be",
  limma_method = "ls",
  limma_robust = FALSE,
  voom_norm = "none",
  limma_trend = FALSE,
  force = FALSE,
  keepers = NULL,
  keep_underscore = TRUE,
  adjust = "BH",
  ...
)
}
\arguments{
\item{input}{Dataframe/vector or expt class containing count tables,
normalization state, etc.}

\item{model_fstring}{Formula string describing the statistical
model of interest.}

\item{null_fstring}{Formula string describing the null model.}

\item{model_svs}{Matrix of surrogates or character describing how
to get them.}

\item{extra_contrasts}{Some extra contrasts to add to the list.
This can be pretty neat, lets say one has conditions A,B,C,D,E
and wants to do (C/B)/A and (E/D)/A or (E/D)/(C/B) then use this
with a string like: "c_vs_b_ctrla = (C-B)-A, e_vs_d_ctrla = (E-D)-A,
de_vs_cb = (E-D)-(C-B),"}

\item{annot_df}{Data frame for annotations.}

\item{libsize}{I've recently figured out that libsize is far more important
than I previously realized.  Play with it here.}

\item{filter}{Filter the data before seeking SVs?}

\item{num_surrogates}{Number of SVs or way to guesstimate them.}

\item{limma_method}{Choose one of limma's lm methods.}

\item{limma_robust}{Make the significance estimation robust?}

\item{voom_norm}{Use this method to normalize the voom inputs.}

\item{limma_trend}{Add trend lines to limma's voom plot?}

\item{force}{Force data which may not be appropriate for limma into it?}

\item{keepers}{Perform an explicit set of contrasts instead of all.}

\item{keep_underscore}{Sanitize underscores from the model matrix?}

\item{adjust}{Use this p-value adjustment.}

\item{...}{Use the elipsis parameter to feed options to write_limma().}
}
\value{
List including the following information:
 macb = the mashing together of condition/batch so you can look at it
 macb_model = The result of calling model.matrix(~0 + macb)
 macb_fit = The result of calling lmFit(data, macb_model)
 voom_result = The result from voom()
 voom_design = The design from voom (redundant from voom_result, but convenient)
 macb_table = A table of the number of times each condition/batch pairing happens
 cond_table = A table of the number of times each condition appears (the
  denominator for the identities)
 batch_table = How many times each batch appears
 identities = The list of strings defining each condition by itself
 all_pairwise = The list of strings defining all the pairwise contrasts
 contrast_string = The string making up the makeContrasts() call
 pairwise_fits = The result from calling contrasts.fit()
 pairwise_comparisons = The result from eBayes()
 limma_result = The result from calling write_limma()
}
\description{
Creates the set of all possible contrasts and performs them using voom/limma.
}
\examples{
\dontrun{
 pretend <- dream_pairwise(expt)
}
}
\seealso{
[limma] [Biobase] [deseq_pairwise()] [edger_pairwise()] [basic_pairwise()]
 DOI:10.1101/2023.03.17.533005
}
