% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{early_late}
\alias{early_late}
\title{an expt superclass of early vs. late 5005 samples in thy media}
\format{a \code{expt} instance}
\source{
the create_expt()
}
\usage{
summary(early_late)
}
\description{
Generated by running create_expt() with a sample sheet
and extracting only the early vs. late samples.
}
\author{
atb
}

