% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_ebseq.r
\name{ebseq_few}
\alias{ebseq_few}
\title{Invoke EBMultiTest() when we do not have too many conditions to deal with.}
\usage{
ebseq_few(data, conditions, patterns = NULL, ng_vector = NULL,
  rounds = 10, target_fdr = 0.05, norm = "median")
}
\arguments{
\item{data}{Expressionset/matrix}

\item{conditions}{Factor of conditions in the data to compare.}

\item{patterns}{Set of patterns as described in the ebseq documentation to query.}

\item{ng_vector}{Passed along to ebmultitest().}

\item{rounds}{Passed to ebseq.}

\item{target_fdr}{Passed to ebseq.}

\item{norm}{Normalization method to apply to the data.}
}
\description{
Starting at approximately 5 conditions, ebseq becomes too unwieldy to use
effectively. But, its results until then are pretty neat.
}
