% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_ebseq.R
\name{ebseq_pairwise}
\alias{ebseq_pairwise}
\title{Set up model matrices contrasts and do pairwise comparisons of all conditions
using EBSeq.}
\usage{
ebseq_pairwise(
  input = NULL,
  patterns = NULL,
  model_fstring = "~ 0 + condition + batch",
  null_fstring = "~",
  model_svs = NULL,
  keepers = NULL,
  ng_vector = NULL,
  rounds = 20,
  target_fdr = 0.05,
  method = "pairwise_subset",
  norm = "median",
  force = FALSE,
  keep_underscore = TRUE,
  ...
)
}
\arguments{
\item{input}{Dataframe/vector or expt class containing data, normalization
state, etc.}

\item{patterns}{Set of expression patterns to query.}

\item{model_fstring}{Formula string describing the model of interest.}

\item{null_fstring}{Formula string describing the null model.}

\item{model_svs}{Matrix of SVs or character describing how to find them.}

\item{keepers}{Perform a specific set of contrasts instead of all?}

\item{ng_vector}{I think this is for isoform quantification, but am not yet
certain.}

\item{rounds}{Number of iterations for doing the multi-test}

\item{target_fdr}{Definition of 'significant'}

\item{method}{The default ebseq methodology is to create the set of all
possible 'patterns' in the data; for data sets which are more than
trivially complex, this is not tenable, so this defaults to subsetting the
data into pairs of conditions.}

\item{norm}{Normalization method to use.}

\item{force}{Force ebseq to accept bad data (notably NA containing
stuff from proteomics.}

\item{keep_underscore}{Sanitize away underscores?}

\item{...}{Extra arguments currently unused.}
}
\value{
List containing tables from ebseq, the conditions tested, and the
 ebseq table of conditions.
}
\description{
Invoking EBSeq is confusing, this should help.
}
\examples{
 \dontrun{
  expt <- create_expt(metadata = "sample_sheet.xlsx", gene_info = annotations)
  ebseq_de <- ebseq_pairwise(input = expt)
}
}
\seealso{
[limma_pairwise()] [deseq_pairwise()] [edger_pairwise()] [basic_pairwise()]
}
