% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_ebseq.r
\name{ebseq_pairwise_subset}
\alias{ebseq_pairwise_subset}
\title{Perform pairwise comparisons with ebseq, one at a time.}
\usage{
ebseq_pairwise_subset(input, ng_vector = NULL, rounds = 10,
  target_fdr = 0.05, model_batch = FALSE, model_cond = TRUE,
  model_intercept = FALSE, alt_model = NULL, conditions = NULL,
  norm = "median", force = FALSE, ...)
}
\arguments{
\item{input}{Expressionset/expt to perform de upon.}

\item{ng_vector}{Passed on to ebseq, I forget what this does.}

\item{rounds}{Passed on to ebseq, I think it defines how many iterations to
perform before return the de estimates}

\item{target_fdr}{If we reach this fdr before iterating rounds times, return.}

\item{model_batch}{Provided by all_pairwise()  I do not think a Bayesian
analysis really care about models, but if one wished to try to add a batch
factor, do it here.  It is currently ignored though.}

\item{model_cond}{Provided by all_pairwise(), ibid.}

\item{model_intercept}{Ibid.}

\item{alt_model}{Ibid.}

\item{conditions}{Factor of conditions in the data, used to define the
contrasts.}

\item{norm}{EBseq normalization method to apply to the data.}

\item{force}{Flag used to force inappropriate data into the various methods.}

\item{...}{Extra arguments passed downstream, noably to choose_model()}
}
\value{
A pairwise comparison of the various conditions in the data.
}
\description{
This uses the same logic as in the various *_pairwise functions to invoke
the 'normal' ebseq pairwise comparison for each pair of conditions in an
expressionset.  It therefore avoids the strange logic inherent in the ebseq
multitest function.
}
