% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_ebseq.R
\name{ebseq_size_factors}
\alias{ebseq_size_factors}
\title{Choose the ebseq normalization method to apply to the data.}
\usage{
ebseq_size_factors(data_mtrx, norm = NULL)
}
\arguments{
\item{data_mtrx}{This is exprs(expressionset)}

\item{norm}{The method to pass along.}
}
\value{
a new matrix using the ebseq specific method of choice.
}
\description{
EBSeq provides three normaliation methods.  Median, Quantile, and Rank.
Choose among them here.
}
\seealso{
[EBSeq]
}
