% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_ebseq.r
\name{ebseq_two}
\alias{ebseq_two}
\title{The primary function used in my EBSeq implementation.}
\usage{
ebseq_two(pair_data, conditions, numerator = 2, denominator = 1,
  ng_vector = NULL, rounds = 10, target_fdr = 0.05,
  norm = "median", force = FALSE)
}
\arguments{
\item{pair_data}{Matrix containing the samples comprising two experimental
factors of interest.}

\item{conditions}{Factor of conditions in the data.}

\item{numerator}{Which factor has the numerator in the data.}

\item{denominator}{Which factor has the denominator in the data.}

\item{ng_vector}{Passed to ebseq.}

\item{rounds}{Passed to ebseq.}

\item{target_fdr}{Passed to ebseq.}

\item{norm}{Normalization method of ebseq to apply.}

\item{force}{Force inappropriate data into ebseq?}
}
\value{
EBSeq result table with some extra formatting.
}
\description{
Most of the time, my invocation of ebseq will fall into this function.
}
