% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_edger.r
\name{edger_coefficient_scatter}
\alias{edger_coefficient_scatter}
\title{Plot two coefficients with respect to one another from edgeR.}
\usage{
edger_coefficient_scatter(output, x = 1, y = 2, gvis_filename = NULL,
  gvis_trendline = TRUE, tooltip_data = NULL, base_url = NULL)
}
\arguments{
\item{output}{Set of pairwise comparisons provided by edger_pairwise().}

\item{x}{Name or number of the x-axis coefficient column to extract.}

\item{y}{Name or number of the y-axis coefficient column to extract.}

\item{gvis_filename}{Filename for plotting gvis interactive graphs of the data.}

\item{gvis_trendline}{Add a trendline to the gvis plot?}

\item{tooltip_data}{Dataframe of gene annotations to be used in the gvis plot.}

\item{base_url}{Add a linkout to gvis plots to this base url.}
}
\value{
Ggplot2 plot showing the relationship between the two coefficients.
}
\description{
It can be nice to see a plot of two coefficients from a edger comparison with respect to one another
This hopefully makes that easy.
}
\examples{
\dontrun{
 pretty = coefficient_scatter(limma_data, x="wt", y="mut")
}
}
\seealso{
\link{hpgl_linear_scatter} \link{edger_pairwise}
}

