% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_edger.r
\name{edger_ma}
\alias{edger_ma}
\title{Make a MA plot of some limma output with pretty colors and shapes}
\usage{
edger_ma(output, table = NULL, fc_col = "logFC", p_col = "qvalue",
  expr_col = "logCPM", fc = 1, pval_cutoff = 0.05)
}
\arguments{
\item{output}{The result from all_pairwise(), which should be changed to handle other invocations too.}

\item{table}{Result from edger to use, left alone it chooses the first.}

\item{fc_col}{Column for logFC data.}

\item{p_col}{Column to use for p-value data.}

\item{expr_col}{Column for the average data.}

\item{fc}{Fold change cutoff on the up and down defining significant.}

\item{pval_cutoff}{And the p-value cutoff.}
}
\value{
a plot!
}
\description{
Yay pretty colors and shapes!
}
\examples{
 \dontrun{
  prettyplot <- edger_ma(all_aprwise) ## [sic, I'm witty! and can speel]
}
}
\seealso{
\link{plot_ma_de}
}

