% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\name{enrichResult-class}
\alias{enrichResult-class}
\title{The enrichResult class.}
\description{
I create enrichResult objects in each of the xxx2enrich().
}
\details{
I am not completely certain how to properly use roxygen to make
available classes from another package.  It looks like I should
just need to do 'importClassesFrom package class', but I thought I
already did that? I have a series of functions which coerce various
enrichment results to DOSE's enrichResult.  I thought this class
was actually in a package named soemthing like 'enrich' but I think
that was just one of my fever dreams.  In any event, I am going to
mess around here and try to stop the error:
'## Error in getClass(Class, where = topenv(parent.frame())):
 ## "enrichResult" is not a defined class'
from making me sad.

One note, this seems only to be a problem in my containerized
version of hpgltools, opening the possibility that this is
dependency mismanagement.
}
