% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{example_fasta}
\alias{example_fasta}
\title{Group A Streptococcus pyogenes strain 5005 genome.}
\format{
An object of class \code{character} of length 1.
}
\usage{
example_fasta
}
\description{
Yeah, what the heck, I am not likely to try to include this in CRAN, I can
waste some space.
}
\keyword{datasets}
