% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{example_gff}
\alias{example_gff}
\title{Group A Streptococcus pyogenes strain 5005 GFF file.}
\format{
An object of class \code{character} of length 1.
}
\usage{
example_gff
}
\description{
This gff contains the annotations which correspond to NC_007297.
I am not sure if I want to include the fasta file.
}
\keyword{datasets}
