% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{exclude_genes}
\alias{exclude_genes}
\title{A temporary alias to subset_genes}
\usage{
exclude_genes(...)
}
\arguments{
\item{...}{Parameters passed to subset_genes().}
}
\description{
A temporary alias to subset_genes
}
