% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{exclude_genes_expt}
\alias{exclude_genes_expt}
\title{A temporary alias to subset_genes}
\usage{
exclude_genes_expt(...)
}
\arguments{
\item{...}{Parameters passed to subset_genes().}
}
\description{
A temporary alias to subset_genes
}
