% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{exclude_genes_from_combined_list}
\alias{exclude_genes_from_combined_list}
\title{Remove specific genes from the result of combine_de_tables and similar ilk.}
\usage{
exclude_genes_from_combined_list(comb, excludes)
}
\arguments{
\item{comb}{Result from combine_de_tables()}

\item{excludes}{Set of genes to remove.}
}
\description{
Remove specific genes from the result of combine_de_tables and similar ilk.
}
