% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{exprs<-,ExpressionSet,data.frame-method}
\alias{exprs<-,ExpressionSet,data.frame-method}
\title{A setter to put the expression data into an expt.}
\usage{
\S4method{exprs}{ExpressionSet,data.frame}(object) <- value
}
\arguments{
\item{object}{ExpressionSet to modify.}

\item{value}{New expression data.}
}
\description{
A setter to put the expression data into an expt.
}
