% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{exprs<-,expt,ANY-method}
\alias{exprs<-,expt,ANY-method}
\title{A setter to put the expression data into an expt.}
\usage{
\S4method{exprs}{expt,ANY}(object) <- value
}
\arguments{
\item{expt}{An expt.}
}
\description{
A setter to put the expression data into an expt.
}
