% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{exprs<-,SummarizedExperiment,ANY-method}
\alias{exprs<-,SummarizedExperiment,ANY-method}
\title{A setter to put the expression data to a SummarizedExperiment.}
\usage{
\S4method{exprs}{SummarizedExperiment,ANY}(object) <- value
}
\arguments{
\item{object}{A SummarizedExperiment.}
}
\description{
A setter to put the expression data to a SummarizedExperiment.
}
