% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{exprs<-}
\alias{exprs<-}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
exprs(object) <- value
}
\arguments{
\item{object}{Input object}

\item{value}{new value}
}
\description{
message("I am from Biobase and am explicitly imported, wtf.")
}
