% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{exprs,SummarizedExperiment-method}
\alias{exprs,SummarizedExperiment-method}
\title{A getter to pull the expression data from a SummarizedExperiment.}
\usage{
\S4method{exprs}{SummarizedExperiment}(object)
}
\arguments{
\item{expt}{A SummarizedExperiment.}
}
\description{
A getter to pull the expression data from a SummarizedExperiment.
}
