% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{expt}
\alias{expt}
\title{An expt is an ExpressionSet superclass with a shorter name}
\usage{
expt(...)
}
\arguments{
\item{...}{Parameters for create_expt()}
}
\description{
It is also a simple list so that one may summarize it more simply,
provides colors and some slots to make one's life easier.
It is created via the function create_expt() which perhaps should be changed.
}
\section{Slots}{

\describe{
\item{\code{colors}}{Colors for the expt.}
}}

