% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{expt_snp}
\alias{expt_snp}
\title{Gather snp information for an expt}
\usage{
expt_snp(expt, input_dir = "preprocessing/outputs",
  file_suffix = "_parsed_ratio.txt", bam_suffix = "_accepted_paired.bam",
  tolower = TRUE)
}
\arguments{
\item{expt}{an expressionset from which to extract information.}

\item{input_dir}{Directory to scan for snps output files.}

\item{file_suffix}{What to add on the end of the files for the resulting output.}

\item{bam_suffix}{How do we find the bam files?}

\item{tolower}{Lowercase stuff like 'HPGL'?}
}
\value{
some stuff
}
\description{
I have some initial code for working with snps, but it seems that it will be getting more use, so
make it testable etc.
}

