% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.r
\name{extract_de_ma}
\alias{extract_de_ma}
\title{Make a MA plot of some limma output with pretty colors and shapes}
\usage{
extract_de_ma(pairwise, type = "edger", table = NULL, fc = 1,
  pval_cutoff = 0.05, ...)
}
\arguments{
\item{table}{Result from edger to use, left alone it chooses the first.}

\item{fc}{Fold change cutoff on the up and down defining significant.}

\item{pval_cutoff}{And the p-value cutoff.}

\item{output}{The result from all_pairwise(), which should be changed to handle other invocations too.}

\item{fc_col}{Column for logFC data.}

\item{p_col}{Column to use for p-value data.}

\item{expr_col}{Column for the average data.}
}
\value{
a plot!
}
\description{
Yay pretty colors and shapes!
}
\examples{
 \dontrun{
  prettyplot <- edger_ma(all_aprwise) ## [sic, I'm witty! and can speel]
}
}
\seealso{
\link{plot_ma_de}
}

