% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{extract_eupath_orthologs}
\alias{extract_eupath_orthologs}
\title{Given 2 species names from the eupathdb, make orthology tables betwixt them.}
\usage{
extract_eupath_orthologs(first, master = "GID",
  query_species = c("L. braziliensis MHOM/BR/75/M2903",
  "L. mexicana MHOM/GT/2001/U1103"), group_columns = c("Ortholog_ID",
  "Ortholog_Group", "Organism"), abbrev = NULL, xref = TRUE, ...)
}
\arguments{
\item{first}{Species name from one eupath database.}

\item{abbrev}{It is likely you do not care about the reciprocals and such,
in that case the second species is irrelvant and just a species
abbreviation is sufficient.}

\item{xref}{Cross reference the stuff in the sqlite vs. the abbreviation?
  If no, then this will just return the entire ortholog table.

The xref parameter reminds me, this function does make a peculiar attempt to
  handle a lot of corner cases without failure, but instead to try to always
  provide some useful information, even if that information is not precisely
  what you want, but the only thing I can find.
With that in mind, one could easily argue that this function is too long and
  it should be split into a couple pieces, but hopefully it is sufficiently
  clear that this is not necessary -- or I will do it later.}

\item{second}{Another species name from the same eupath database.}
}
\value{
List containing the information available given the species provided
  and/or the abbreviation/xref.
}
\description{
The eupathdb provides such a tremendous wealth of information.  For me
though, it is difficult sometimes to boil it down into just the bits of
comparison I want for 1 species or between 2 species.  A singularly common
question I am asked is: "What are the most similar genes between species x
and y among these two arbitrary parasites?"  There are lots of ways to poke
at this question: run BLAST/fasta36, use biomart, query the ortholog tables
from the eupathdb, etc.  However, in all these cases, it is not trivial to
ask the next question:  What about: a:b and b:a?
This function attempts to address that for the case of two eupath species
from the same domain. (tritrypdb/fungidb/etc.)  It does however assume that
the sqlite package has been installed locally, if not it suggests you run the
make_organismdbi function in order to do that.
}
