% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{extract_gene_locations}
\alias{extract_gene_locations}
\title{Clean up the gene location field from eupathdb derived gene location data.}
\usage{
extract_gene_locations(annot_df, column = "annot_gene_location_text")
}
\arguments{
\item{annot_df}{Data frame resulting from load_orgdb_annotations()}

\item{column}{Name of the column to extract the start/end/length/etc from.}
}
\value{
Somewhat nicer data frame.
}
\description{
Clean up the gene location field from eupathdb derived gene location data.
}
