% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_go}
\alias{extract_go}
\title{Make a table of gene ontology information}
\usage{
extract_go(df, file = NULL)
}
\arguments{
\item{df}{a dataframe of ontology information.  This is intended to
be the output from goseq including information like
numbers/category, GOids, etc.}

\item{file}{a csv file to which to write the table}
}
\value{
the ontology table with annotation information included
}
\description{
Make a table of gene ontology information
}
\examples{
## annotated_go = extract_go(go_ids)
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

