% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.R
\name{extract_interesting_goseq}
\alias{extract_interesting_goseq}
\title{Filter a goseq significance search}
\usage{
extract_interesting_goseq(
  godata,
  expand_categories = TRUE,
  pvalue = 0.05,
  minimum_interesting = 1,
  adjust = 0.05,
  padjust_method = "BH"
)
}
\arguments{
\item{godata}{goseq result}

\item{expand_categories}{Extract GO terms from GO.db and add them
to the table}

\item{pvalue}{Significance filter.}

\item{minimum_interesting}{The category should have more than this
number of elements.}

\item{adjust}{Adjusted p-value filter.}

\item{padjust_method}{Method for adjusting the p-values.}
}
\description{
Given a goseq result, use some simple filters to pull out the
categories of likely interest.
}
