% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{extract_keepers,ANY,character-method}
\alias{extract_keepers,ANY,character-method}
\title{Pull the set of keepers from a character vector.}
\usage{
\S4method{extract_keepers}{ANY,character}(
  extracted,
  keepers,
  table_names,
  all_coefficients,
  limma,
  edger,
  ebseq,
  deseq,
  basic,
  adjp,
  annot_df,
  include_deseq,
  include_edger,
  include_ebseq,
  include_limma,
  include_basic,
  excludes,
  padj_type,
  fancy = FALSE,
  loess = FALSE,
  lfc_cutoff = 1,
  p_cutoff = 0.05,
  sheet_prefix = NULL,
  sheet_number = NULL,
  format_sig = 4,
  plot_colors = plot_colors,
  z = 1.5,
  alpha = 0.4,
  z_lines = FALSE,
  label = 10,
  label_column = "hgncsymbol"
)
}
\arguments{
\item{keepers}{Character vector of keepers.}
}
\description{
Instead of a list of numerators/denominators, one might feed combine_de_tables a vector of
things like: 'a_vs_b'.
}
