% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R, R/zzz_methods.R
\name{extract_keepers}
\alias{extract_keepers}
\alias{extract_keepers,list,character-method}
\title{When a list of 'keeper' contrasts is specified, extract it from the data.}
\usage{
extract_keepers(
  extracted,
  keepers,
  table_names,
  all_coefficients,
  apr,
  adjp,
  annot_df,
  includes,
  excludes,
  padj_type,
  fancy = FALSE,
  loess = FALSE,
  lfc_cutoff = 1,
  p_cutoff = 0.05,
  format_sig = 4,
  plot_colors = plot_colors,
  z = 1.5,
  alpha = 0.4,
  z_lines = FALSE,
  label = 10,
  label_column = "hgncsymbol"
)

\S4method{extract_keepers}{list,character}(
  extracted,
  keepers,
  table_names,
  all_coefficients,
  apr,
  adjp,
  annot_df,
  includes,
  excludes,
  padj_type,
  fancy = FALSE,
  loess = FALSE,
  lfc_cutoff = 1,
  p_cutoff = 0.05,
  format_sig = 4,
  plot_colors = plot_colors,
  z = 1.5,
  alpha = 0.4,
  z_lines = FALSE,
  label = 10,
  label_column = "hgncsymbol"
)
}
\arguments{
\item{extracted}{Tables extracted from the all_pairwise data.}

\item{keepers}{In this case, one may assume either NULL or 'all'.}

\item{table_names}{The set of tables produced by all_pairwise().}

\item{all_coefficients}{The set of all experimental conditions in the
experimental metadata.}

\item{apr}{The result from all_pairwise(), containing the limma/edger/deseq/etc data.}

\item{adjp}{Pull out the adjusted p-values from the data?}

\item{annot_df}{What annotations should be added to the table?}

\item{includes}{List of predicates by method.}

\item{excludes}{Set of genes to exclude.}

\item{padj_type}{Choose a specific p adjustment.}

\item{fancy}{Include larger pdf/svg plots with the xlsx output?}

\item{loess}{Add a loess to plots?}

\item{lfc_cutoff}{Passed for volcano/MA plots and defining 'significant'}

\item{p_cutoff}{Passed for volcano/MA plots and defining 'significant'}

\item{format_sig}{Number of significant digits for stuff like
pvalues.}

\item{plot_colors}{Define what colors should be used for
'up'/'down'}

\item{z}{Define significantly away from the identity line in a
coefficient plot.}

\item{alpha}{Use this alpha transparency for plots.}

\item{z_lines}{Include lines denoting significant z-scores?}

\item{label}{When not NULL, label this many genes.}

\item{label_column}{Try using this column for labeling genes.}
}
\value{
The extracted, but with more stuff at the end!
}
\description{
This is the most interesting of the extract_keeper functions.  It must check
that the numerators and denominators match the desired contrast and flip the
signs in the logFCs when appropriate.
}
\section{Methods (by class)}{
\itemize{
\item \code{extract_keepers(extracted = list, keepers = character)}: Use a character vector instead of a list.

}}
