% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{extract_keepers_all}
\alias{extract_keepers_all}
\title{When no set of 'keeper' contrasts is specified, grab them all.}
\usage{
extract_keepers_all(
  apr,
  extracted,
  keepers,
  table_names,
  all_coefficients,
  limma,
  edger,
  ebseq,
  deseq,
  basic,
  adjp,
  annot_df,
  include_deseq,
  include_edger,
  include_ebseq,
  include_limma,
  include_basic,
  excludes,
  padj_type,
  loess = FALSE,
  lfc_cutoff = 1,
  p_cutoff = 0.05
)
}
\arguments{
\item{apr}{Result from all_pairwise().}

\item{extracted}{Table of extracted data.}

\item{keepers}{In this case, one may assume either NULL or 'all'.}

\item{table_names}{The set of tables produced by all_pairwise().}

\item{all_coefficients}{The set of all experimental conditions in the
experimental metadata.}

\item{limma}{The limma data from all_pairwise().}

\item{edger}{The edger data from all_pairwise().}

\item{ebseq}{The ebseq data from all_pairwise().}

\item{deseq}{The deseq data from all_pairwise().}

\item{basic}{The basic data from all_pairwise().}

\item{adjp}{Pull out the adjusted p-values from the data?}

\item{annot_df}{What annotations should be added to the table?}

\item{include_deseq}{Whether or not to include the deseq data.}

\item{include_edger}{Whether or not to include the edger data.}

\item{include_ebseq}{Whether or not to include the ebseq data.}

\item{include_limma}{Whether or not to include the limma data.}

\item{include_basic}{Whether or not to include the basic data.}

\item{excludes}{Set of genes to exclude.}

\item{padj_type}{Choose a specific p adjustment.}

\item{loess}{Include a loess estimator in the plots?}

\item{lfc_cutoff}{Passed for plotting volcano/MA plots.}

\item{p_cutoff}{Passed for volcano/MA plots.}
}
\description{
This has a couple of cousin functions, extract_keepers_list and _single.
These handle extracting one or more contrasts out of the various tables
produced by all_pairwise().
}
