% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{extract_logistic_regression}
\alias{extract_logistic_regression}
\title{Invoke what I think is an appropriate logistical regression model.}
\usage{
extract_logistic_regression(
  design,
  query = "condition",
  multivariable = TRUE,
  factors = NULL,
  family = "binomial",
  conf = 0.95,
  excel = NULL
)
}
\arguments{
\item{design}{Experimental design, I need to change this it is not a
matrix.}

\item{query}{Response variable.}

\item{multivariable}{When not true, this will iterate over every factor individually.}

\item{factors}{set of factors to query against the query.}

\item{family}{The family passed to glm.}

\item{conf}{Confidence interval chosen for plotting.}

\item{excel}{Output xlsx file to which we print the f values etc.}
}
\description{
The current implementation uses lm() and assumes everything is a
linear model, this will attempt to invoke an appropriate logistic
model via glm() and provide similar/identical tables/plots.
}
\details{
A reference to myself regarding families:
  gaussian: identity, log, inverse
  binomial: logit, probit, cauchit
  Gamma: inverse, identity, log ??  ooo Gamma.log etc
  quasi: logit, probit, cloglog, identity, inverse, log 1/mu^2,
  sqrt

 For the purposes of a 'normal' logistic regression, I think
'binomial' is sufficient.
}
