% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{extract_mayu_pps_fdr}
\alias{extract_mayu_pps_fdr}
\title{Read output from mayu to get the IP/PP number corresponding to a given FDR value.}
\usage{
extract_mayu_pps_fdr(file, fdr = 0.01)
}
\arguments{
\item{file}{Mayu output file.}

\item{fdr}{Chosen fdr value to acquire.}
}
\value{
List of two elements: the full mayu table sorted by fdr and the number
 corresponding to the chosen fdr value.
}
\description{
Read output from mayu to get the IP/PP number corresponding to a given FDR value.
}
