% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{extract_metadata}
\alias{extract_metadata}
\title{Pull metadata from a table (xlsx/xls/csv/whatever)}
\usage{
extract_metadata(metadata, ...)
}
\arguments{
\item{metadata}{file or df of metadata}

\item{...}{Arguments to pass to the child functions.}
}
\value{
Metadata dataframe hopefully cleaned up to not be obnoxious.
}
\description{
Pull metadata from a table (xlsx/xls/csv/whatever)
}
