% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.R
\name{extract_msraw_data}
\alias{extract_msraw_data}
\title{Read a bunch of mzXML files to acquire their metadata.}
\usage{
extract_msraw_data(
  metadata,
  write_windows = TRUE,
  id_column = "sampleid",
  file_column = "raw_file",
  allow_window_overlap = FALSE,
  start_add = 0,
  format = "mzXML",
  parallel = TRUE,
  savefile = NULL,
  ...
)
}
\arguments{
\item{metadata}{Data frame describing the samples, including the mzXML
filenames.}

\item{write_windows}{Write out SWATH window frames.}

\item{id_column}{What column in the sample sheet provides the ID for the samples?}

\item{file_column}{Which column in the sample sheet provides the filenames?}

\item{allow_window_overlap}{What it says on the tin, some tools do not like
DIA windows to overlap, if TRUE, this will make sure each annotated window
starts at the end of the previous window if they overlap.}

\item{start_add}{Another strategy is to just add a static amount to each
window.}

\item{format}{Currently this handles mzXML or mzML files.}

\item{parallel}{Perform operations using an R foreach cluster?}

\item{savefile}{If not null, save the resulting data structure to an rda file.}

\item{...}{Extra arguments, presumably color palettes and column names and
stuff like that.}
}
\value{
List of data extracted from every sample in the MS run (DIA or DDA).
}
\description{
I have had difficulties getting the full set of correct parameters for a
DDA/DIA experiment.  After some poking, I eventually found most of these
required prameters in the mzXML raw files.  Ergo, this function uses them.
20190310: I had forgotten about the mzR library.  I think much (all?) of this
is redundant with respect to it and perhaps should be removed in deference to
the more complete and fast implementation included in mzR.
}
