% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.R
\name{extract_mzML_scans}
\alias{extract_mzML_scans}
\title{Parse a mzML file and return the relevant data.}
\usage{
extract_mzML_scans(
  file,
  id = NULL,
  write_acquisitions = TRUE,
  allow_window_overlap = FALSE,
  start_add = 0
)
}
\arguments{
\item{file}{Input mzML file to parse.}

\item{id}{Chosen ID for the given file.}

\item{write_acquisitions}{Write acquisition windows.}

\item{allow_window_overlap}{Some downstream tools cannot deal with
overlapping windows. Toggle that here.}

\item{start_add}{Other downstream tools appear to expect some padding at the
beginning of each window.  Add that here.}
}
\value{
The list of metadata, scan data, etc from the mzXML file.
}
\description{
This does the actual work for extract_scan_data().  This levers mzR to
provide the data and goes a step further to pull out the windows acquired in
the MS/MS scan and print them in formats acceptable to TPP/OpenMS (eg. with
and without headers).
}
