% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{extract_mzxml_data}
\alias{extract_mzxml_data}
\title{Read a bunch of mzXML files to acquire their metadata.}
\usage{
extract_mzxml_data(metadata, write_windows = TRUE,
  id_column = "sampleid", parallel = TRUE, savefile = NULL, ...)
}
\arguments{
\item{metadata}{Data frame describing the samples, including the mzXML
filenames.}

\item{write_windows}{Write out SWATH window frames.}

\item{id_column}{What column in the sample sheet provides the ID for the samples?}

\item{parallel}{Perform operations using an R foreach cluster?}

\item{savefile}{If not null, save the resulting data structure to an rda file.}

\item{...}{Extra arguments, presumably color palettes and column names and
stuff like that.}
}
\value{
A list of data extracted from every sample in the MS run (DIA or DDA).
}
\description{
I have had difficulties getting the full set of correct parameters for a
DDA/DIA experiment.  After some poking, I eventually found most of these
required prameters in the mzXML raw files.  Ergo, this function uses them.
}
