% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{extract_mzXML_scans}
\alias{extract_mzXML_scans}
\title{Parse a mzXML file and return the relevant data.}
\usage{
extract_mzXML_scans(file, id = NULL, write_acquisitions = TRUE,
  allow_window_overlap = FALSE, start_add = 0)
}
\arguments{
\item{file}{Input mzXML file to parse.}

\item{id}{Chosen ID for the given file.}

\item{write_acquisitions}{Write acquisition windows.}

\item{allow_window_overlap}{Some downstream tools cannot deal with
overlapping windows. Toggle that here.}

\item{start_add}{Other downstream tools appear to expect some padding at the
beginning of each window.  Add that here.}
}
\value{
The list of metadata, scan data, etc from the mzXML file.
}
\description{
This does the actual work for extract_scan_data().  When I wrote this
function, I had forgotten about the mzR library; with that in mind, this
seems to give a bit more information and be a bit faster than my short tests
with mzR (note however that my tests were to compare mzR parsing mzML files
vs. this function with mzXML, which is a classic apples to oranges).
}
\details{
This goes a step further to pull out the windows acquired in the MS/MS scan
and print them in formats acceptable to TPP/OpenMS (eg. with and without
headers).
}
