% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{extract_pyprophet_data}
\alias{extract_pyprophet_data}
\title{Read a bunch of scored swath outputs from pyprophet to acquire their metrics.}
\usage{
extract_pyprophet_data(metadata, pyprophet_column = "diascored",
  savefile = NULL, ...)
}
\arguments{
\item{metadata}{Data frame describing the samples, including the mzXML
filenames.}

\item{savefile}{If not null, save the data from this to the given filename.}

\item{...}{Extra arguments, presumably color palettes and column names and
stuff like that.}

\item{scored_column}{Which column from the metadata provides the requisite filenames?}
}
\value{
A list of data from each sample in the pyprophet scored DIA run.
}
\description{
This function is mostly cribbed from the other extract_ functions in this file.
With it, I hope to be able to provide some metrics of a set of openswath runs, thus
potentially opening the door to being able to objectively compare the same
run with different options and/or different runs.
}
