% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{extract_scan_data}
\alias{extract_scan_data}
\title{Read a mzML/mzXML file and extract from it some important metadata.}
\usage{
extract_scan_data(file, id = NULL, write_acquisitions = TRUE,
  format = "mzXML", allow_window_overlap = FALSE, start_add = 0)
}
\arguments{
\item{file}{Filename to read.}

\item{id}{An id to give the result.}

\item{write_acquisitions}{If a filename is provided, write a tab separated
table of windows.}

\item{start_add}{Add a minute to the start of the windows to avoid overlaps?}
}
\value{
List containing a table of scan and precursor data.
}
\description{
When working with swath data, it is fundamentally important to know the
correct values for a bunch of the input variables.  These are not trivial
to acquire.  This function attempts to make this easier (but slow) by reading
the mzXML file and parsing out helpful data.
}
