% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{extract_siggenes}
\alias{extract_siggenes}
\title{Alias for extract_significant_genes because I am dumb.}
\usage{
extract_siggenes(...)
}
\arguments{
\item{...}{The parameters for extract_significant_genes()}
}
\value{
It should return a reminder for me to remember my function names or change them to
    something not stupid.
}
\description{
Alias for extract_significant_genes because I am dumb.
}
