% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\docType{methods}
\name{fData}
\alias{fData}
\alias{fData,}
\alias{fData-methods}
\title{Extend Biobase::fData to handle expt objects.}
\arguments{
\item{object}{An expt from which to extract the expressionset.}
}
\description{
Extend Biobase::fData to handle expt objects.
}
