% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{fData<-,expt-method}
\alias{fData<-,expt-method}
\title{A setter to put the gene annotation data into an expt.}
\usage{
\S4method{fData}{expt}(object) <- value
}
\arguments{
\item{object}{An expt.}

\item{value}{New annotations for the expressionset.}
}
\description{
A setter to put the gene annotation data into an expt.
}
