% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{fData,SummarizedExperiment-method}
\alias{fData,SummarizedExperiment-method}
\title{A getter to pull the gene annotation data from a SummarizedExperiment.}
\usage{
\S4method{fData}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{A SummarizedExperiment.}
}
\description{
A getter to pull the gene annotation data from a SummarizedExperiment.
}
