% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{fData}
\alias{fData}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
fData(object)
}
\arguments{
\item{object}{Input object}
}
\description{
message("I am from Biobase and am explicitly imported, wtf.")
}
