% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{features_greater_than}
\alias{features_greater_than}
\title{Count the number of features(genes) greater than x in a data set.}
\usage{
features_greater_than(data, cutoff = 1, hard = TRUE, inverse = FALSE)
}
\arguments{
\item{data}{Dataframe/exprs/matrix/whatever of counts.}

\item{cutoff}{Minimum number of counts.}

\item{hard}{Greater-than is hard, greater-than-equals is not.}

\item{inverse}{when inverted, this provides features less than the cutoff.}
}
\value{
A list of two elements, the first comprised of the number of genes
  greater than the cutoff, the second with the identities of said genes.
}
\description{
Sometimes I am asked how many genes have >= x counts.  Well, here you go.
}
\details{
Untested as of 2016-12-01 but used with Lucia.  I think it would be interesting to iterate
this function from small to large cutoffs and plot how the number of kept genes decreases.
}
\examples{
\dontrun{
 features <- features_greater_than(expt)
 fewer <- features_greater_than(expt, cutoff = 100)
}
}
\seealso{
[Biobase]
}
