% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{features_less_than}
\alias{features_less_than}
\title{Do features_greater_than() inverted!}
\usage{
features_less_than(...)
}
\arguments{
\item{...}{Arguments passed to features_greather_than()}
}
\value{
The set of features less than whatever you would have done with
  features_greater_than().
}
\description{
Do features_greater_than() inverted!
}
\seealso{
\code{\link{features_greater_than}}
}
