% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter low-count genes from a data set.}
\usage{
filter_counts(counts, thresh = 2, min_samples = 2)
}
\arguments{
\item{thresh}{lower threshold of counts (4 by default)}

\item{df}{input data frame of counts by sample}

\item{minSamples}{minimum number of samples}
}
\value{
dataframe of counts without the low-count genes
}
\description{
Filter low-count genes from a data set.
}
\examples{
## filtered_table = filter_counts(count_table)
}
\seealso{
\code{\link{log2CPM}} which this uses to decide what to keep
}

