% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{filter_scd}
\alias{filter_scd}
\title{Perform a series of filters on a single-cell dataset.}
\usage{
filter_scd(
  scd,
  min_num_rna = 200,
  max_num_rna = NULL,
  min_pct_ribo = 5,
  max_pct_ribo = NULL,
  max_pct_mito = 15,
  min_pct_mito = NULL,
  mito_pattern = "^mt-",
  ribo_pattern = "^Rp[sl]",
  min_gene_counts = 3,
  verbose = FALSE
)
}
\arguments{
\item{scd}{Single Cell Dataset to filter.}

\item{min_num_rna}{Drop cells with fewer than this number of
observed RNA species.}

\item{min_pct_ribo}{Drop cells with less than this percentage of
ribosomal protein RNAs observed.}

\item{max_pct_ribo}{Drop cells with more than this percentage of
ribosomal protein RNAs observed.}

\item{max_pct_mito}{Drop cells with more than this percentage of
mitochondrial RNA observed.}

\item{min_pct_mito}{Drop cells with less than this percentage of
mitochondrial RNA observed.}

\item{mito_pattern}{Regex pattern to search RNA symbols for
mitochondrial species.}

\item{ribo_pattern}{Regex pattern to search RNA symbols for
ribosomal protein species.}

\item{min_gene_counts}{Drop genes across cells which are observed
less than this number of times, I don't expect many of these.}
}
\value{
Filtered scd
}
\description{
This function should perform a series of relatively consistent
filters on a single-cell dataset, with options to play with the
various filters and their parameters.
}
