% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{find_subseq_target_temp}
\alias{find_subseq_target_temp}
\title{Find a subsequence with a target PCR temperature.}
\usage{
find_subseq_target_temp(
  sequence,
  target = 53,
  direction = "forward",
  verbose = FALSE
)
}
\arguments{
\item{sequence}{Starting sequence.}

\item{target}{Desire TM of the final sequence.}

\item{direction}{What strand is expected for annealing this primer?}

\item{verbose}{Be chatty?}
}
\description{
Given a relatively large sequence, this function will iteratively
remove a single nucleotide and recalulate the TM until the TM falls
to the target temperature.
}
