% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{find_working_dataset}
\alias{find_working_dataset}
\title{Search a mart for a usable dataset.}
\usage{
find_working_dataset(mart, trydataset, species)
}
\description{
Search a mart for a usable dataset.
}
