% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{find_working_mart}
\alias{find_working_mart}
\title{Find a functional biomart instance.}
\usage{
find_working_mart(
  default_hosts = c("useast.ensembl.org", "uswest.ensembl.org", "www.ensembl.org",
    "asia.ensembl.org"),
  trymart = "ENSEMBL_MART_ENSEMBL",
  archive = FALSE,
  year = NULL,
  month = NULL
)
}
\arguments{
\item{default_hosts}{List of biomart mirrors to try.}

\item{trymart}{Specific mart to query.}

\item{archive}{Try an archive server instead of a mirror?  If this is a
character, it will assume it is a specific archive hostname.}

\item{year}{Choose specific year(s) for the archive servers?}

\item{month}{Choose specific month(s) for the archive servers?}
}
\value{
Either a mart instance or NULL if no love was forthcoming.
}
\description{
In my experience, the various biomart mirrors are not varyingly likely to be
functional at any given time.  In addition, I often find it useful to use an
archive instance rather than the most recent ensembl instance.  This function
therefore iterates over the various mirrors; or if archive = TRUE it will try a
series of archive servers from 1, 2, and 3 years ago.
}
\seealso{
[biomaRt::useMart()] [biomaRt::listMarts()]
}
