% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.r
\name{flanking_sequence}
\alias{flanking_sequence}
\title{Extract sequence flanking a set of annotations (generally coding sequences)}
\usage{
flanking_sequence(bsgenome, annotation, distance = 200, type = "gene",
  prefix = "")
}
\arguments{
\item{bsgenome}{Genome sequence}

\item{annotation}{Set of annotations}

\item{distance}{How far from each annotation is desired?}

\item{type}{What type of annotation is desired?}

\item{prefix}{Provide a prefix to the names to distinguish them from the existing annotations.}
}
\value{
A list of sequences before and after each sequence.
}
\description{
Given a set of annotations and genome, one might want to get the set of adjacent sequences.
}

