% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.R
\name{gather_cp_genes}
\alias{gather_cp_genes}
\title{Collect gene IDs from a table and make them readable.}
\usage{
gather_cp_genes(table, mappings, new = "ORF", primary_key = 1)
}
\arguments{
\item{table}{Gene table from (initially) clusterProfiler.}

\item{mappings}{Table of mapped gene IDs.}

\item{new}{String used to disambiguate mappings when it is not provided by the table.}

\item{primary_key}{Column name to use when extracting IDs.}
}
\description{
Collect gene IDs from a table and make them readable.
}
