% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{gather_eupath_utrs_padding}
\alias{gather_eupath_utrs_padding}
\title{Given an eupathdb species lacking UTR boundaries, extract an arbitrary region
before/after each gene.}
\usage{
gather_eupath_utrs_padding(
  species_name = "Leishmania major",
  entry = NULL,
  webservice = "tritrypdb",
  padding = 200,
  ...
)
}
\arguments{
\item{species_name}{Species name for which to query the eupathdb.}

\item{entry}{EuPathDB metadatum entry.}

\item{webservice}{If specified, makes the query faster, I always used
tritrypdb.org.}

\item{...}{Extra arguments for the various EuPathDB functions.}
}
\value{
Set of padding UTR sequences/coordinates.
}
\description{
This is a very domain-specific function.
}
