% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{gather_genes}
\alias{gather_genes}
\title{Given a set of goseq data from simple_goseq(),
make a list of genes represented in each ontology.}
\usage{
gather_genes(goseq_data, ontology = "MF", pval = 0.05,
  include_all = FALSE)
}
\arguments{
\item{goseq_data}{a list of goseq specific results as generated by simple_goseq()}

\item{ontology}{an ontology to search}

\item{pval}{a maximum accepted pvalue to include in the list of categories to cross reference.}

\item{include_all}{include all genes in the ontology search}
}
\value{
a data frame of categories/genes.
}
\description{
This function uses the GO2ALLEG data structure to reverse map ontology
categories to a list of genes represented. It therefore assumes that the
GO2ALLEG.rda data structure has been deposited in pwd().  This in turn
may be generated by clusterProfilers buildGOmap() function if it doesn't
exist.  For some species it may also be auto-generated.
With little work this can be made much more generic, and it probably should.
}
\examples{
\dontrun{
 data = simple_goseq(de_genes=limma_output, lengths=annotation_df, goids=goids_df)
 genes_in_cats = gather_genes(data, ont='BP')
}
}
\seealso{
\link{simple_goseq} \code{\link[clusterProfiler]{buildGOmap}},
}

