% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{gather_genes_orgdb}
\alias{gather_genes_orgdb}
\title{Use the orgdb instances from clusterProfiler to gather annotation data for GO.}
\usage{
gather_genes_orgdb(goseq_data, orgdb_go, orgdb_ensembl)
}
\arguments{
\item{goseq_data}{Some data from goseq and friends.}

\item{orgdb_go}{The orgDb instance with GO data.}

\item{orgdb_ensembl}{The orgDb instance with ensembl data.}
}
\value{
a go mapping
}
\description{
Since clusterprofiler no longer builds gomaps, I need to start understanding
how to properly get information from orgDBs.
}
\seealso{
\pkg{clusterProfiler}
}
