% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.R
\name{gather_masses}
\alias{gather_masses}
\title{Use BRAIN to find the peptide mass from a sequence.}
\usage{
gather_masses(sequence)
}
\arguments{
\item{sequence}{Sequence to count.}
}
\value{
Rounded average mass.
}
\description{
This rounds the avgMass from BRAIN to deal with isotopes, maybe this should be changed.
}
