% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/108_ontology_xlsx.r
\name{gather_ontology_genes}
\alias{gather_ontology_genes}
\title{Given a set of goseq data from simple_goseq(),
make a list of genes represented in each ontology.}
\usage{
gather_ontology_genes(
  result,
  ontology = NULL,
  column = "over_represented_pvalue",
  pval = 0.1,
  include_all = FALSE,
  ...
)
}
\arguments{
\item{result}{List of results as generated by simple_*().}

\item{ontology}{Ontology to search (MF/BP/CC).}

\item{column}{Which column to use for extracting ontologies?}

\item{pval}{Maximum accepted pvalue to include in the list of categories to
cross reference.}

\item{include_all}{Include all genes in the ontology search?}

\item{...}{Extra options without a purpose just yet.}
}
\value{
Data frame of categories/genes.
}
\description{
This function uses the GO2ALLEG data structure to reverse map ontology
categories to a list of genes represented. It therefore assumes that the
GO2ALLEG.rda data structure has been deposited in pwd().  This in turn
may be generated by clusterProfilers buildGOmap() function if it doesn't
exist.  For some species it may also be auto-generated.
With little work this can be made much more generic, and it probably should.
}
\examples{
\dontrun{
 data <- simple_goseq(sig_genes=limma_output, lengths=annotation_df, goids=goids_df)
 genes_in_cats <- gather_genes(data, ont='BP')
}
}
\seealso{
\pkg{goseq} \pkg{clusterProfiler}
 \code{\link{simple_goseq}}
}
