% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_splicing.R
\name{gather_rmats_files}
\alias{gather_rmats_files}
\title{Collect the outputs from rMATS.}
\usage{
gather_rmats_files(file_prefix = NULL, numerator = "t4h", denominator = "no")
}
\arguments{
\item{file_prefix}{Where do we expect to find the files?}

\item{numerator}{The numerator and denominator are prefixed to the
output files.}

\item{denominator}{Ibid.}
}
\description{
I invoked rMATS in a pretty specific way and the various outputs
get specific home directories and suffixes.  This is intended
partially to document that for myself and make it easier to find
them.
}
