% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_splicing.R
\name{gather_suppa_files}
\alias{gather_suppa_files}
\title{Collect the outputs from suppa.}
\usage{
gather_suppa_files(file_prefix = NULL, numerator = "d15", denominator = "d10")
}
\arguments{
\item{file_prefix}{Where do we expect to find the files?}

\item{numerator}{The numerator and denominator are prefixed to the
output files.}

\item{denominator}{Ibid.}
}
\description{
I invoked suppa in a pretty specific way and the various outputs
get specific home directories and suffixes.  This is intended
partially to document that for myself and make it easier to find
them.
}
