% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{gather_trimomatic_metadata}
\alias{gather_trimomatic_metadata}
\title{Gather metadata for trimomatic.}
\usage{
gather_trimomatic_metadata(
  metadata,
  file_spec = "preprocessing/{meta[['sampleid']]}/outputs/*-trimomatic.out",
  columns = NULL,
  new_metadata = NULL,
  ...
)
}
\arguments{
\item{metadata}{Starting metadata}

\item{file_spec}{Filename containing the information of interest.}

\item{columns}{Arbitrarily set the column names}

\item{new_metadata}{Filename for the new metadata.}

\item{...}{extra stuff for glue.}
}
\description{
Simplify the specification for extracting trimomatic data.
}
