% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{gather_utrs_txdb}
\alias{gather_utrs_txdb}
\title{Get UTR sequences using information provided by TxDb and fiveUTRsByTranscript}
\usage{
gather_utrs_txdb(bsgenome, fivep_utr = NULL, threep_utr = NULL,
  start_column = "start", end_column = "end", strand_column = "strand",
  chr_column = "seqnames", name_column = "group_name", ...)
}
\description{
For species like Mus musculus, load_orgdb_annotations(Mus.musculus)
should return a list including the requisite GRanges for the 5'/3' UTRs.
}
