% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.r
\name{gbk2txdb}
\alias{gbk2txdb}
\title{Given a genbank accession, make a txDb object along with sequences, etc.}
\usage{
gbk2txdb(accession = "AE009949")
}
\arguments{
\item{accession}{Accession to download and import}
}
\value{
List containing a txDb, sequences, and some other stuff which I haven't yet finalized.
}
\description{
Let us admit it, sometimes biomart is a pain.  It also does not have easily accessible data for
microbes.  Genbank does!
}

