% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.r
\name{gbk_annotations}
\alias{gbk_annotations}
\title{Extract some useful information from a gbk imported as a txDb.}
\usage{
gbk_annotations(gbr)
}
\arguments{
\item{gbr}{TxDb object to poke at.}
}
\value{
Granges data
}
\description{
Maybe this should get pulled into the previous function?
}

