% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{gbs_counts}
\alias{gbs_counts}
\title{Count tables of the mapped GBS TNSeq DNA fragments.}
\format{
An object of class \code{character} of length 1.
}
\usage{
gbs_counts
}
\description{
One thing I like to do with TNSeq data is to treat it similarly to RNASeq
data in order to get a sense of the changing 'fitness' of each gene.  The
data has all the same distribution attributes of a RNASeq dataset, after all;
so why not use the same plots and tests to see if it is valid?
}
\keyword{datasets}
