% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{gbs_essentiality}
\alias{gbs_essentiality}
\title{Outputs from the DeJesus Essentiality package from the GBS TNSeq experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
gbs_essentiality
}
\description{
The Essentiality package from the DeJesus lab uses a Bayesian framework to
look for genes which are essential in a TNSeq experiment.  In my pipeline, I
invoke this tool with multiple parameters in an attempt to find the
parameters which provide the most likely 'true' result.  This archive
contains those results for our GBS TNSeq experiment.
}
\keyword{datasets}
