% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{gbs_fasta}
\alias{gbs_fasta}
\title{Genome for the GBS TNSeq experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
gbs_fasta
}
\description{
If you read this far, you know what this is.
}
\keyword{datasets}
