% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{gbs_gff}
\alias{gbs_gff}
\title{GFF annotations used for the GBS TNSeq experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
gbs_gff
}
\description{
At the time of the experiment, there was not a very good genome for this
strain.  We therefore chose to use strain A909 as the reference.  Since then,
Lindsey's lab made a complete genome, though the annotations remain a bit
sparse.
}
\keyword{datasets}
