% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{gbs_sample_sheet}
\alias{gbs_sample_sheet}
\title{Sample sheet used for a portion of a Group B Streptococcal TNSeq experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
gbs_sample_sheet
}
\description{
TNSeq is sort of the inverse of RNASeq, one is instead looking for the genes
_not_ represented in the dataset.  This sample sheet lays out the
experimental design for an in vitro TNSeq experiment from Streptococcus
agalactiae strain CJB111.
}
\keyword{datasets}
