% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{gbs_wig}
\alias{gbs_wig}
\title{Wig files used as input for the DeJesus Essentiality package.}
\format{
An object of class \code{character} of length 1.
}
\usage{
gbs_wig
}
\description{
My preprocessing pipeline uses the bam alignments from bowtie to extract all
reads which start/end on a mariner insertion site (TA) and count how many
occured at every position of the genome.  These files are the result of that
process, thus each line is the position of a 'T' in the 'TA' followed by the
number of reads which start/end with it.
}
\keyword{datasets}
