% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.r
\name{genefilter_kofa_counts}
\alias{genefilter_kofa_counts}
\title{\code{genefilter_kofa_counts()}  Filter low-count genes from a data set using genefilter's kOverA()}
\usage{
genefilter_kofa_counts(count_table, k = 1, A = 1, verbose = FALSE)
}
\arguments{
\item{k}{a minimum number of samples to have >A counts}

\item{A}{the minimum number of counts for each gene's sample in kOverA()}

\item{verbose}{If set to true, prints number of genes removed / remaining}

\item{counts}{input data frame of counts by sample}
}
\value{
dataframe of counts without the low-count genes
}
\description{
\code{genefilter_kofa_counts()}  Filter low-count genes from a data set using genefilter's kOverA()
}
\examples{
\dontrun{
 filtered_table = genefilter_kofa_counts(count_table)
}
}
\seealso{
\pkg{genefilter} \code{\link[genefilter]{kOverA}} which this uses to decide what to keep
}

