% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{genefilter_pofa_counts}
\alias{genefilter_pofa_counts}
\title{Filter low-count genes from a data set using genefilter's pOverA()}
\usage{
genefilter_pofa_counts(count_table, p = 0.01, A = 100, verbose = TRUE)
}
\arguments{
\item{count_table}{input data frame of counts by sample}

\item{p}{a minimum proportion of each gene's counts/sample to be greater than a minimum(A)}

\item{A}{the minimum number of counts in the above proportion}

\item{verbose}{If set to true, prints number of genes removed / remaining}
}
\value{
dataframe of counts without the low-count genes
}
\description{
I keep thinking this function is pofa... oh well.
}
\examples{
\dontrun{
 filtered_table = genefilter_pofa_counts(count_table)
}
}
\seealso{
\pkg{genefilter} \code{\link[genefilter]{pOverA}} which this uses to decide what to keep
}

