% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{generate_expt_colors}
\alias{generate_expt_colors}
\title{Set up default colors for a data structure containing usable metadata}
\usage{
generate_expt_colors(sample_definitions, cond_column = "condition",
  by = "sampleid", ...)
}
\arguments{
\item{sample_definitions}{Metadata, presumably containing a 'condition'
column.}

\item{cond_column}{Which column in the sample data provides the set of
'conditions' used to define the colors?}

\item{by}{Name the factor of colors according to this column.}

\item{...}{Other arguments like a color palette, etc.}
}
\value{
Colors!
}
\description{
In theory this function should be useful in any context when one has a blob
of metadata and wants to have a set of colors.  Since my taste is utterly
terrible, I rely entirely upon RColorBrewer, but also allow one to choose
his/her own colors.
}
