% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{generate_gene_kegg_mapping}
\alias{generate_gene_kegg_mapping}
\title{Generate GENE/KEGG mapping}
\usage{
generate_gene_kegg_mapping(pathways, org_abbreviation, verbose = FALSE)
}
\arguments{
\item{pathways}{Vector of KEGG pathway IDs returned from call to keggLink()
e.g. "path:mmu05134"}

\item{org_abbreviation}{KEGG identifier for the species of interest (e.g.
"hsa" for Homo sapiens)}
}
\description{
Generate GENE/KEGG mapping
}

