% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{generate_kegg_pathway_mapping}
\alias{generate_kegg_pathway_mapping}
\title{Generate a KEGG PATHWAY / description mapping}
\usage{
generate_kegg_pathway_mapping(pathways, verbose = FALSE)
}
\arguments{
\item{pathways}{Vector of KEGG pathway identifiers}

\item{verbose}{talk talk?}
}
\value{
a data frame describing some kegg pathways
}
\description{
Generate a KEGG PATHWAY / description mapping
}

