% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{genomic_kmer_dist}
\alias{genomic_kmer_dist}
\title{Use ape to generate a distance based nj tree from fasta files.}
\usage{
genomic_kmer_dist(directory = "tree", root = NULL)
}
\arguments{
\item{directory}{Directory of fasta genomes.}

\item{root}{Species ID to place at the root of the tree.}
}
\value{
List containing the phylogeny and some other stuff.
}
\description{
I was thinking that a standardized version of this might be useful
for Theresa's recent exploration of variants in her data.
}
