% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genplot.R
\name{genoplot_chromosome}
\alias{genoplot_chromosome}
\title{Try plotting a chromosome (region)}
\usage{
genoplot_chromosome(
  accession = "AE009949",
  start = NULL,
  end = NULL,
  plot_title = "Genome plot"
)
}
\arguments{
\item{accession}{An accession to plot, this will download it.}

\item{start}{First segment to plot (doesn't quite work yet).}

\item{end}{Final segment to plot (doesn't quite work yet).}

\item{plot_title}{Put a title on the resulting plot.}
}
\value{
Hopefully a pretty plot of a genome
}
\description{
genoplotr is cool, I don't yet understand it though
}
\seealso{
[genoPlotR]
}
