% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{get_abundant_genes}
\alias{get_abundant_genes}
\title{Find the set of most/least abundant genes according to limma and friends
following a differential expression analysis.}
\usage{
get_abundant_genes(datum, type = "limma", n = NULL, z = NULL, unique = FALSE)
}
\arguments{
\item{datum}{Output from the _pairwise() functions.}

\item{type}{Extract abundant genes according to what?}

\item{n}{Perhaps take just the top/bottom n genes.}

\item{z}{Or take genes past a given z-score.}

\item{unique}{Unimplemented: take only the genes unique among the conditions surveyed.}
}
\value{
List of data frames containing the genes of interest.
}
\description{
Given a data set provided by limma, deseq, edger, etc; one might want to know
what are the most and least abundant genes, much like get_sig_genes() does to
find the most significantly different genes for each contrast.
}
\examples{
\dontrun{
 abundant <- get_abundant_genes(all_pairwise_output, type = "deseq", n = 100)
 ## Top 100 most abundant genes from deseq
 least <- get_abundant_genes(all_pairwise_output, type = "deseq", n = 100, least = TRUE)
 ## Top 100 least abundant genes from deseq
 abundant <- get_abundant_genes(all_pairwise_output, type = "edger", z = 1.5)
 ## Get the genes more than 1.5 standard deviations from the mean.
}
}
\seealso{
[get_sig_genes()]
}
