% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{get_backup_expression_data,ExpressionSet-method}
\alias{get_backup_expression_data,ExpressionSet-method}
\title{Get the backup data from an ExpressionSet.}
\usage{
\S4method{get_backup_expression_data}{ExpressionSet}(expt)
}
\arguments{
\item{expt}{An ExpressionSet does not contain backup data.}
}
\description{
Get the backup data from an ExpressionSet.
}
