% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{get_backup_expression_data,SummarizedExperiment-method}
\alias{get_backup_expression_data,SummarizedExperiment-method}
\title{Get the backup data from a SummarizedExperiment.}
\usage{
\S4method{get_backup_expression_data}{SummarizedExperiment}(expt)
}
\arguments{
\item{expt}{An expt containing the backup data.}
}
\description{
Get the backup data from a SummarizedExperiment.
}
