% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{get_backup_expression_data}
\alias{get_backup_expression_data}
\title{Extract the backup copy of an expressionset from an expt.}
\usage{
get_backup_expression_data(expt)
}
\arguments{
\item{expt}{Expt which should contain the backup.}
}
\description{
Extract the backup copy of an expressionset from an expt.
}
