% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.r
\name{get_biomart_annotations}
\alias{get_biomart_annotations}
\title{Extract annotation information from biomart.}
\usage{
get_biomart_annotations(species = "hsapiens", overwrite = FALSE,
  do_save = TRUE, host = "dec2015.archive.ensembl.org",
  trymart = "ENSEMBL_MART_ENSEMBL", include_lengths = TRUE)
}
\arguments{
\item{species}{Choose a species.}

\item{overwrite}{Overwite an existing save file?}

\item{do_save}{Create a savefile of annotations for future runs?}

\item{host}{Ensembl hostname to use.}

\item{trymart}{Biomart has become a circular dependency, this makes me sad, now to list the
marts, you need to have a mart loaded...}

\item{include_lengths}{Also perform a search on structural elements in the genome?}
}
\value{
Df of some (by default) human annotations.
}
\description{
Biomart is an amazing resource of information, but using it is a bit annoying.  This function
hopes to alleviate some common headaches.
}
\examples{
\dontrun{
 tt = get_biomart_annotations()
}
}

