% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{get_biomart_example_gene}
\alias{get_biomart_example_gene}
\title{Extract a single gene from biomart in order to check out the various annotations.}
\usage{
get_biomart_example_gene(
  species = "mmusculus",
  attributes = "feature_page",
  host = NULL,
  trymart = "ENSEMBL_MART_ENSEMBL",
  archive = TRUE,
  default_hosts = c("useast.ensembl.org", "uswest.ensembl.org", "www.ensembl.org",
    "asia.ensembl.org"),
  gene_requests = c("ensembl_gene_id", "version")
)
}
\arguments{
\item{species}{Using this species}

\item{attributes}{and these attributes}

\item{host}{auto choose a host, or specify one.}

\item{trymart}{The default mart}

\item{archive}{Use an archive server}

\item{default_hosts}{A set of likely host candidates.}

\item{gene_requests}{The two most common columns to query.}
}
\description{
Extract a single gene from biomart in order to check out the various annotations.
}
