% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.r
\name{get_biomart_ontologies}
\alias{get_biomart_ontologies}
\title{Extract gene ontology information from biomart.}
\usage{
get_biomart_ontologies(species = "hsapiens", overwrite = FALSE,
  do_save = TRUE, host = "dec2015.archive.ensembl.org",
  trymart = "ENSEMBL_MART_ENSEMBL", secondtry = "_gene",
  dl_rows = c("ensembl_gene_id", "go_accession"),
  dl_rowsv2 = c("ensembl_gene_id", "go_id"))
}
\arguments{
\item{species}{Species to query.}

\item{overwrite}{Overwrite existing savefile?}

\item{do_save}{Create a savefile of the annotations? (if not false, then a filename.)}

\item{host}{Ensembl hostname to use.}

\item{trymart}{Default mart to try, newer marts use a different notation.}

\item{secondtry}{The newer mart name.}

\item{dl_rows}{List of rows from the final biomart object to download.}
}
\value{
Df of geneIDs and GOIDs.
}
\description{
I perceive that every time I go to acquire annotation data from biomart, they have changed
something important and made it more difficult for me to find what I want. I recently found the
*.archive.ensembl.org, and so this function uses that to try to keep things predictable, if not
consistent.
}
\examples{
\dontrun{
 tt = get_biomart_ontologies()
}
}
\seealso{
\link[biomaRt]{getBM}
}

