% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\name{get_circos_data}
\alias{get_circos_data}
\title{Extra fonts and useful bits and bobs for working with circos.}
\usage{
get_circos_data()
}
\description{
This is a tarball of the circos etc/ directory from my Debian linux
installation.  I discovered to my annoyance that other systems were missing
the fonts required to make circos plots work properly along with something
else.  In a fit of pique I tarred them up and left them here.
}
