% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{get_colors,ANY-method}
\alias{get_colors,ANY-method}
\title{Get the colors from any arbitrary object, I am still trying to figure out
where my S4 dispatch is going wrong; this resides in BiocGenerics.}
\usage{
\S4method{get_colors}{ANY}(x)
}
\arguments{
\item{x}{Object from which to get the colors.}
}
\description{
Get the colors from any arbitrary object, I am still trying to figure out
where my S4 dispatch is going wrong; this resides in BiocGenerics.
}
