% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_colors,data.frame-method}
\alias{get_colors,data.frame-method}
\title{Get colors from a dataframe.}
\usage{
\S4method{get_colors}{data.frame}(exp, ...)
}
\arguments{
\item{exp}{Input}

\item{...}{Other arguments}
}
\description{
Get colors from a dataframe.
}
