% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_colors,DFrame-method}
\alias{get_colors,DFrame-method}
\title{Get colors from the weirdo S4 dataframe}
\usage{
\S4method{get_colors}{DFrame}(exp, ...)
}
\arguments{
\item{exp}{Input}

\item{...}{Other arguments}
}
\description{
Get colors from the weirdo S4 dataframe
}
