% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_colors,expt-method}
\alias{get_colors,expt-method}
\title{A getter to pull the colors from an expt.}
\usage{
\S4method{get_colors}{expt}(exp)
}
\arguments{
\item{x}{An x.}
}
\description{
A getter to pull the colors from an expt.
}
