% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{get_colors,SummarizedExperiment-method}
\alias{get_colors,SummarizedExperiment-method}
\title{A getter to pull the colors from a SummarizedExperiment.}
\usage{
\S4method{get_colors}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{An x.}
}
\description{
A getter to pull the colors from a SummarizedExperiment.
}
