% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_colors}
\alias{get_colors}
\title{While I am struggling with S4 dispatch, here is a garbage generic.}
\usage{
get_colors(exp, ...)
}
\arguments{
\item{exp}{Input}

\item{...}{Other arguments}
}
\description{
While I am struggling with S4 dispatch, here is a garbage generic.
}
