% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_colors_by_condition,data.frame-method}
\alias{get_colors_by_condition,data.frame-method}
\title{Figure out likely colors from a dataframe.}
\usage{
\S4method{get_colors_by_condition}{data.frame}(exp, fact = "condition", levels = NULL)
}
\arguments{
\item{exp}{Expression from which to gather colors.}

\item{fact}{Use this metadata column to set the colors.}

\item{levels}{When not null, colors may be set to arbitrary samples.}
}
\value{
List of colors by condition.
}
\description{
Usually we give a vector of all samples by colors.  This just
simplifies that to one element each.  Currently only used in
combine_de_tables() but I think it will have use elsewhere.
}
