% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{get_degrees}
\alias{get_degrees}
\title{Use a vector of factors and design to count up the levels.}
\usage{
get_degrees(design, fctrs)
}
\arguments{
\item{design}{experimental design.}

\item{fctrs}{vector of factors to count up.}
}
\description{
This implements a quick and dirty degrees of freedom counter.
}
