% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb_organdb.r
\name{get_eupath_config}
\alias{get_eupath_config}
\title{Grab some configuration data collated and used to make OrganismDbi/OrgDb/TxDb objects.}
\usage{
get_eupath_config(cfg = NULL)
}
\arguments{
\item{cfg}{Optional data frame}
}
\value{
Dataframe of configuration data, a few columns are required, run it with no args to see
    which ones.
}
\description{
This function uses some data copied into inst/ to decide some parameters used for generating the
various packages generated here.
}

