% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupath_webservices.r
\name{get_eupath_fields}
\alias{get_eupath_fields}
\title{Extract query-able fields from the EupathDb.}
\usage{
get_eupath_fields(webservice)
}
\arguments{
\item{webservice}{Eupathdb, tritrypdb, fungidb, etc...}
}
\value{
List of parameters.
}
\description{
This parses the result of a query to Eupath's webservice:
'GenesByMolecularWeight' and uses it to get a list of fields which are
acquireable elsewhere.
}
