% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{get_eupath_gff_table}
\alias{get_eupath_gff_table}
\title{Extract gene information from a GFF file}
\usage{
get_eupath_gff_table(granges = NULL)
}
\arguments{
\item{granges}{GenomicRanges instance as returned by `import.gff3`}
}
\value{
data.frame containing basic gene information (ID, description,
etc.)
}
\description{
Extract gene information from a GFF file
}
\author{
Keith Hughitt
}
