% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{get_eupath_go_term_table}
\alias{get_eupath_go_term_table}
\title{Returns a mapping of gene ID to gene GO categories for a given organism.}
\usage{
get_eupath_go_term_table(provider = "TriTrypDB",
  species = "Leishmania major", overwrite = FALSE, dir = ".")
}
\arguments{
\item{provider}{Name of data provider to query.}

\item{species}{Full name of the organism as per the eupathdb api.}

\item{overwrite}{Overwrite the existing table?}

\item{dir}{Where to dump the data.}
}
\description{
Returns a mapping of gene ID to gene GO categories for a given organism.
}
