% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{get_eupath_interpro_table}
\alias{get_eupath_interpro_table}
\title{Returns a mapping of gene ID to InterPro domains for a specified organism}
\usage{
get_eupath_interpro_table(data_provider, organism)
}
\arguments{
\item{data_provider}{Name of data provider to query (e.g. 'TriTrypDB')}

\item{organism}{Full name of organism, as used by EuPathDB APIs}
}
\value{
Dataframe with ....
}
\description{
Returns a mapping of gene ID to InterPro domains for a specified organism
}
\author{
Keith Hughitt
}
