% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{get_eupath_pathway_table}
\alias{get_eupath_pathway_table}
\title{2017/07/08: IN DEVELOPMENT}
\usage{
get_eupath_pathway_table(provider = "TriTrypDB",
  species = "Leishmania major", overwrite = FALSE, dir = ".")
}
\arguments{
\item{species}{Full name of organism, as used by EuPathDB APIs}

\item{overwrite}{Overwrite already existing data?}

\item{dir}{Where to place the data.}

\item{data_provider}{Name of data provider to query (e.g. 'TriTrypDB')}
}
\value{
Dataframe with gene/pathway mapping
}
\description{
Returns a mapping of gene ID to metabolic pathways (KEGG, LeishCyc, etc.)
}
\author{
Keith Hughitt
}
