% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{get_eupath_text}
\alias{get_eupath_text}
\title{Gather most(all?) of the data which used to be in the .txt files from the eupathdb.}
\usage{
get_eupath_text(species = NULL, entry = NULL, dir = "eupathdb",
  metadata = NULL, question_type = "GeneQuestions",
  question = "GenesByTaxon")
}
\arguments{
\item{species}{Species to query.}

\item{entry}{Metadata entry to use.}

\item{dir}{Directory in which to put intermediate data.}

\item{metadata}{Existing dataframe of metadata.}

\item{question_type}{Usually GeneQuestions, I am not sure what other types are in use.}

\item{question}{There are a bunch of possible query types, GenesByTaxon is
the most comprehensive.}
}
\value{
data table of fun annotation data.
}
\description{
Starting with revision 33, the eupathdb discontinued maintaining the
downloadable text tables. For much of the eupath data, this was not a
significant concern, Keith wrote the get_eupath_ family of functions which
provided a reasonable facsimile of that information.  Some pieces remained
missing, this function attempts to address that.
}
