% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{get_expt_colors}
\alias{get_expt_colors}
\title{Get a named vector of colors by condition.}
\usage{
get_expt_colors(expt)
}
\arguments{
\item{expt}{Expression from which to gather colors.}
}
\value{
List of colors by condition.
}
\description{
Usually we give a vector of all samples by colors.  This just
simplifies that to one element each.  Currently only used in
combine_de_tables() but I think it will have use elsewhere.
}
