% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{get_formula_factors,formula-method}
\alias{get_formula_factors,formula-method}
\title{Set method for formula input to get_formula_factors}
\usage{
\S4method{get_formula_factors}{formula}(formula_string = NULL)
}
\arguments{
\item{formula_string}{What should have been a string, but is a formula.}
}
\value{
List describing the formula
}
\description{
Set method for formula input to get_formula_factors
}
