% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{get_formula_factors}
\alias{get_formula_factors}
\title{Simplify a model formula string to a simple vector of factors.}
\usage{
get_formula_factors(formula_string = NULL)
}
\arguments{
\item{formula_string}{Formula describing the formula of interest.}
}
\value{
Ordered vector of factors in the formula ignoring
interactions etc.
}
\description{
I would like to be able to make my various pairwise methods
generalizable, therefore I need to no longer rely on an expected
pair of factors: 'condition' and 'batch', instead I want to observe
the first factor and anything which follows.  Thus, I will string
split the formula string by punctuation and yank out the names of
interest.
}
