% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_genelengths}
\alias{get_genelengths}
\title{Grab gene lengths from a gff file}
\usage{
get_genelengths(gff, type = "gene", key = "ID")
}
\arguments{
\item{gff}{a gff file with (hopefully) IDs and widths}

\item{type}{The annotation type to use (gene)}
}
\value{
a data frame of gene IDs and widths.
}
\description{
Grab gene lengths from a gff file
}
\examples{
## tt = hpgltools:::get_genelengths('reference/fun.gff.gz')
## head(tt)
##          ID width
##1   YAL069W   312
##2   YAL069W   315
##3   YAL069W     3
##4 YAL068W-A   252
##5 YAL068W-A   255
##6 YAL068W-A     3
}
\seealso{
\code{\link{import.gff3}}, \code{\link{import.gff}}, \code{\link{import.gff2}}
}

