% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_shared.r
\name{get_genesizes}
\alias{get_genesizes}
\title{Grab gene length/width/size from an annotation database.}
\usage{
get_genesizes(
  annotation = NULL,
  type = "gff",
  gene_type = "gene",
  type_column = "type",
  key = NULL,
  length_names = NULL,
  ...
)
}
\arguments{
\item{annotation}{There are a few likely data sources when getting gene
sizes, choose one with this.}

\item{type}{What type of annotation data are we using?}

\item{gene_type}{Annotation type to use (3rd column of a gff file).}

\item{type_column}{Type identifier (10th column of a gff file).}

\item{key}{What column has ID information?}

\item{length_names}{Provide some column names which give gene length information?}

\item{...}{Extra arguments likely for load_annotations()}
}
\value{
Data frame of gene IDs and widths.
}
\description{
This function tries to gather an appropriate gene length column from
whatever annotation data source is provided.
}
\examples{
 pa_genesizes <- get_genesizes(gff = pa_gff)
 head(pa_genesizes)
}
\seealso{
[rtracklayer] [load_gff_annotations()]
}
