% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{get_git_commit}
\alias{get_git_commit}
\title{Get the current git commit for hpgltools}
\usage{
get_git_commit(gitdir = "/home/trey/hpgltools")
}
\arguments{
\item{gitdir}{Directory containing the git repository.}
}
\description{
Get the current git commit for hpgltools
}
