% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{get_group_gsva_means}
\alias{get_group_gsva_means}
\title{Create dataframe which gets the maximum within group mean gsva score for each gene set}
\usage{
get_group_gsva_means(gsva_scores, groups)
}
\arguments{
\item{gsva_result}{}
}
\value{
dataframe containing max_gsva_score, and within group means for gsva scores
}
\description{
Create dataframe which gets the maximum within group mean gsva score for each gene set
}
