% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\name{get_hsapiens_data}
\alias{get_hsapiens_data}
\title{Subset of a human RNASeq expressionset.}
\usage{
get_hsapiens_data()
}
\description{
This is a portion of an expressionset used to examine changes caused by
infection with Leishmania panamensis.
}
