% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{get_identifier}
\alias{get_identifier}
\title{Get an Identifier function from a shorthand name.}
\usage{
get_identifier(type)
}
\arguments{
\item{type}{String name for the identifier in question.}
}
\description{
I am hoping to write one for EuPathDB and some other source, thus the switch.
}
