% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{get_individual_snps}
\alias{get_individual_snps}
\title{Extract the observed snps unique to individual categories in a snp set.}
\usage{
get_individual_snps(retlist)
}
\arguments{
\item{retlist}{The result from get_snp_sets().}
}
\description{
The result of get_snp_sets provides sets of snps for all possible
categories.  This is cool and all, but most of the time we just want the
results of a single group in that rather large set (2^number of categories)
}
