% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{get_input_colors,expt-method}
\alias{get_input_colors,expt-method}
\title{Extract colors from an expt.}
\usage{
\S4method{get_input_colors}{expt}(input)
}
\arguments{
\item{input}{input expt.}
}
\description{
Extract colors from an expt.
}
