% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{get_input_colors,SummarizedExperiment-method}
\alias{get_input_colors,SummarizedExperiment-method}
\title{Extract colors from a SE.}
\usage{
\S4method{get_input_colors}{SummarizedExperiment}(input)
}
\arguments{
\item{input}{input SE.}
}
\description{
Extract colors from a SE.
}
