% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{get_input_colors}
\alias{get_input_colors}
\title{I keep messing with my S4 dispatch of object attributes.}
\usage{
get_input_colors(input)
}
\arguments{
\item{input}{dimensional object with color information.}
}
\description{
I keep messing with my S4 dispatch of object attributes.
}
