% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_inter_txdb,character-method}
\alias{get_inter_txdb,character-method}
\title{Given an input gff, create intron/intergenic gff files.}
\usage{
\S4method{get_inter_txdb}{character}(input)
}
\arguments{
\item{input}{gff file containing the annotations.}
}
\description{
I was poking around for ways to mix and match gff files and found this:
https://support.bioconductor.org/p/66003/
}
