% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_kegg.R
\name{get_kegg_compounds}
\alias{get_kegg_compounds}
\title{Gather all Compounds from all pathways for a given species.}
\usage{
get_kegg_compounds(
  pathway = "all",
  abbreviation = NULL,
  species = "leishmania major",
  savefile = NULL
)
}
\arguments{
\item{pathway}{One or more pathways, all does what it says on the tin.}

\item{abbreviation}{Approximately 3 character KEGG abbreviation.}

\item{species}{If you do not have the abbreviation, this will try
to find it.}

\item{savefile}{Currently unused I think, but eventually should
make a savefile of the results.}
}
\description{
This function attempts to iterate over every pathway for a given
abbreviation/species and extract from them the set of compounds.
This was mostly copy/pasted from get_kegg_genes.
}
