% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{get_kegg_genes}
\alias{get_kegg_genes}
\title{Extract the set of geneIDs matching pathways for a given species.}
\usage{
get_kegg_genes(pathway = "all", abbreviation = NULL,
  species = "leishmania major", savefile = NULL)
}
\arguments{
\item{pathway}{Either a single pathway kegg id or 'all'.}

\item{abbreviation}{Optional 3 letter species kegg id.}

\item{species}{Stringified species name used to extract the 3 letter abbreviation.}

\item{savefile}{Filename to which to save the relevant data.}
}
\value{
Dataframe of the various kegg data for each pathway, 1 row/gene.
}
\description{
This uses KEGGREST to extract the mappings for all genes for a species and pathway or 'all'.
Because downloading them takes a while, it will save the results to kegg_species.rda.  When run
interactively, it will give some information regarding the number of genes observed in each
pathway.
}
\examples{
 \dontrun{
  kegg_info <- get_kegg_genes(species="Canis familiaris")
 }
}

