% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{get_kegg_sub}
\alias{get_kegg_sub}
\title{Provide a set of simple substitutions to convert geneIDs from KEGG->TriTryDB}
\usage{
get_kegg_sub(species = "lma")
}
\arguments{
\item{species}{3 letter abbreviation for a given kegg type}
}
\value{
2 character lists containing the patterns and replace arguments for gsub(), order
    matters!
}
\description{
This function should provide 2 character lists which, when applied sequentially, will result in a
hopefully coherent set of mapped gene IDs matching the TriTypDB/KEGG specifications.
}

