% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\name{get_lmajor_data}
\alias{get_lmajor_data}
\title{The UTR regions of every highly translated L.major gene.}
\usage{
get_lmajor_data()
}
\description{
Once upon a time I performed a ribosome profiling experiment in Leishmania
major.  Sadly, we still have not published it.  This file contains the UTRs
of every highly translated gene in procyclic promastigotes.  I used it in
some motif analyses for fun.
}
