% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.r
\name{get_loci_go}
\alias{get_loci_go}
\title{Extract the set of GO categories by microbesonline locus}
\usage{
get_loci_go(taxonid = "160490")
}
\arguments{
\item{taxonid}{Which species to query.}
}
\description{
The microbesonline is such a fantastic resource, it is a bit of a shame that it is such a pain to query.
}

