% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.r
\name{get_microbesonline_annotation}
\alias{get_microbesonline_annotation}
\title{Skip the db and download all the text annotations for a given species.}
\usage{
get_microbesonline_annotation(ids = "160490", species = NULL)
}
\arguments{
\item{ids}{List of ids to query.}

\item{species}{Species name(s) to use instead.}
}
\value{
List of dataframes with the annotation information.
}
\description{
Like I said, the microbesonline mysqldb is rather more complex than I prefer.  This shortcuts
that process and just grabs a tsv copy of everything and loads it into a dataframe.
}
\details{
Tested in test_70expt_spyogenes.R
There is so much awesome information in microbesonline, but damn is it annoying to download.
This function makes that rather easier, or so I hope at least.
}
\examples{
\dontrun{
 annotations <- get_microbesonline_annotation(ids=c("160490","160491"))
}
}
\seealso{
\pkg{RCurl}
 \code{\link[RCurl]{getURL}}
}
