% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbes.r
\name{get_microbesonline_ids}
\alias{get_microbesonline_ids}
\title{Use the publicly available microbesonline mysql instance to get species ids.}
\usage{
get_microbesonline_ids(name, exact = FALSE)
}
\arguments{
\item{name}{Text string containing some part of the species name of interest.}

\item{exact}{Use an exact species name?}
}
\value{
Dataframe of ids and names.
}
\description{
The microbesonline mysql instance is more complex than I like.  Their id system is reminiscent of
KEGG's and similarly annoying.  Though I haven't figured out how the tables interact, a query to
get ids is simple enough.
}

