% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_annotation_microbesonline.r
\name{get_microbesonline_taxid}
\alias{get_microbesonline_taxid}
\title{Extract microbesonline taxon IDs without having to click on the weird boxes
at the top of the website.}
\usage{
get_microbesonline_taxid(species = "Acyrthosiphon pisum virus")
}
\arguments{
\item{species}{String to search the set of microbesonline taxa.}
}
\value{
NULL or 1 or more taxon ids.
}
\description{
This should simplify getting material from microbesonline.
}
\examples{
 coli_taxids <- get_microbesonline_taxid(species="coli S88")
 head(coli_taxids)
}
