% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate_estimators.r
\name{get_model_adjust}
\alias{get_model_adjust}
\title{Extract some surrogate estimations from a raw data set using sva, ruv, and/or pca.}
\usage{
get_model_adjust(expt, estimate_type = "sva_supervised", surrogates = NULL,
  ...)
}
\arguments{
\item{estimate_type}{one of sva_supervised, sva_unsupervised, ruv_empirical, ruv_supervised, ruv_residuals, or pca}

\item{...}{parameters fed to arglist}

\item{raw_expt}{a raw experiment object}
}
\value{
a list including the adjustments for a model matrix, a modified count table, and 3 plots of the known batch, surrogates, and batch/surrogate.
}
\description{
This applies the methodologies very nicely explained by Jeff Leek at
https://github.com/jtleek/svaseq/blob/master/recount.Rmd
and attempts to use them to acquire estimates which may be applied to an experimental model
by either EdgeR, DESeq2, or limma.  In addition, it modifies the count tables using these
estimates so that one may play with the modified counts and view the changes (with PCA or heatmaps
or whatever).  Finally, it prints a couple of the plots shown by Leek in his document.
In other words, this is entirely derivative of someone much smarter than me.
}

