% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{get_msigdb_metadata}
\alias{get_msigdb_metadata}
\title{Create a metadata dataframe of msigdb data, this hopefully will be usable to
fill the fData slot of a gsva returned expressionset.}
\usage{
get_msigdb_metadata(
  msig_db = "msigdb_v6.2.xml",
  wanted_meta = c("ORGANISM", "DESCRIPTION_BRIEF", "AUTHORS", "PMID")
)
}
\arguments{
\item{msig_db}{Filename containing the MSigDB metadata.}

\item{wanted_meta}{Choose metadata columns of interest.}
}
\value{
list containing 2 data frames: all metadata from broad, and the set
 matching the sig_data GeneSets.
}
\description{
Create a metadata dataframe of msigdb data, this hopefully will be usable to
fill the fData slot of a gsva returned expressionset.
}
\seealso{
[xml2] [rvest]
}
