% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\name{get_mtuberculosis_data}
\alias{get_mtuberculosis_data}
\title{Portion of a sample sheet used in a DIA-SWATH proteomics experiment.}
\usage{
get_mtuberculosis_data()
}
\description{
In this experiment, Dr. Briken sought to learn about proteins which are
exported by Mycobacterium tuberculosis.  He therefore performed a DIA SWATH
experiment using two strains and collected the supernatant fraction (to get
the exported proteins) and the intracellular fraction.  This file contains
the metadata for a portion of that experiment.  I used a fairly exhaustive
set of open source tools to interpret Dr. Briken's data.  These files
comprise the endpoint of the preprocessing and the inputs for the R package
'SWATH2stats'.  This file is excessively large, but the smallest by far of
the various inputs I wanted to include to test my various proteomics functions.
In a separate series of experiments, we sought to look at the effect of
infection on splicing in the host.  Thus I performed rmats and suppa and
attempted to compare the results to see how reliable they are.  (spoiler: not very).
}
