% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb_organdb.r
\name{get_ncbi_taxonid}
\alias{get_ncbi_taxonid}
\title{Use taxize to get ncbi taxon IDs}
\usage{
get_ncbi_taxonid(species = "Leishmania major")
}
\arguments{
\item{species}{Human readable species name}
}
\value{
potential NCBI taxon IDs
}
\description{
taxize looks like it might be awesome, but it is also pretty annoying
}

