% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupath_webservices.r
\name{get_orthologs_all_genes}
\alias{get_orthologs_all_genes}
\title{Query ortholog tables from the eupathdb one gene at a time.}
\usage{
get_orthologs_all_genes(species = "Leishmania major", dir = "eupathdb",
  entry = NULL, metadata = NULL, ...)
}
\arguments{
\item{species}{What species to query}

\item{dir}{Directory to which to save intermediate data (currently unused)}

\item{entry}{An entry from the eupathdb metadata to use for other parameters.}

\item{metadata}{The set of eupathdb metadata from which to query.}

\item{...}{Extra parameters for downloading eupathdb metadata.}
}
\description{
Querying the full ortholog table at eupathdb.org fails mysteriously.
This is a horrible brute-force approach to get around this.
}
