% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{get_pairwise_gene_abundances}
\alias{get_pairwise_gene_abundances}
\title{A companion function for get_abundant_genes()}
\usage{
get_pairwise_gene_abundances(datum, type = "limma", excel = NULL)
}
\arguments{
\item{datum}{Output from _pairwise() functions.}

\item{type}{According to deseq/limma/ed ger/basic?}

\item{excel}{Print this to an excel file?}
}
\value{
List containing the expression values and some metrics of
 variance/error.
}
\description{
Instead of pulling to top/bottom abundant genes, get all abundances and
variances or stderr.
}
\examples{
\dontrun{
 abundance_excel <- get_pairwise_gene_abundances(combined, excel = "abundances.xlsx")
 ## This should provide a set of abundances after voom by condition.
}
}
\seealso{
[get_abundant_genes()]
}
