% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{get_plot_columns}
\alias{get_plot_columns}
\title{A small rat's nest of if statements intended to figure out what columns
are wanted to plot a MA/Volcano from any one of a diverse set of possible
input types.}
\usage{
get_plot_columns(pairwise, type, found_table = NULL, p_type = "adj")
}
\description{
I split this function away from the main body of extract_de_plots()
so that I can come back to it and strip it down to something a bit
more legible.  One idea is to make use of the fact that I gave
class assignments to all of the outputs from xxx_pairwise()
}
