% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/115_dimension_reduction.r
\name{get_res}
\alias{get_res}
\title{Attempt to get residuals from tsne data}
\usage{
get_res(
  svd_result,
  design,
  factors = c("condition", "batch"),
  res_slot = "v",
  var_slot = "d"
)
}
\arguments{
\item{svd_result}{The set of results from one of the many potential svd-ish methods.}

\item{design}{Experimental design from which to get experimental factors.}

\item{factors}{Set of experimental factors for which to calculate rsquared values.}

\item{res_slot}{Where is the res data in the svd result?}

\item{var_slot}{Where is the var data in the svd result?}
}
\description{
I strongly suspect that this is not correct, but it is a start.
}
