% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\name{get_sagalactiae_data}
\alias{get_sagalactiae_data}
\title{Sample sheet used for a portion of a Group B Streptococcal TNSeq experiment.}
\usage{
get_sagalactiae_data()
}
\description{
TNSeq is sort of the inverse of RNASeq, one is instead looking for the genes
_not_ represented in the dataset.  This sample sheet lays out the
experimental design for an in vitro TNSeq experiment from Streptococcus
agalactiae strain CJB111.  At the time of the experiment, there was not a
very good genome for this strain.  We therefore chose to use strain A909 as
the reference.  Since then, Lindsey's lab made a complete genome, though the
annotations remain a bit sparse.  One thing I like to do with TNSeq data
is to treat it similarly to RNASeq data in order to get a sense of the
changing 'fitness' of each gene.  The data has all the same distribution
attributes of a RNASeq dataset, after all; so why not use the same plots and
tests to see if it is valid?  The Essentiality package from the DeJesus lab
uses a Bayesian framework to look for genes which are essential in a TNSeq
experiment.  In my pipeline, I invoke this tool with multiple parameters in
an attempt to find the parameters which provide the most likely 'true'
result.  This archive contains those results for our GBS TNSeq experiment.
My preprocessing pipeline uses the bam alignments from bowtie to extract all
reads which start/end on a mariner insertion site (TA) and count how many
occured at every position of the genome.  These files are the result of that
process, thus each line is the position of a 'T' in the 'TA' followed by the
number of reads which start/end with it.
}
