% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\name{get_sbetaceum_data}
\alias{get_sbetaceum_data}
\title{Portion of the RNASeq results from Solanum betaceum.}
\usage{
get_sbetaceum_data()
}
\description{
I had the opportunity to work the Sandra Correia, she was awesome.  She was
seeking to learn about differences among embryogenic cells in the Tree
tomato.  I therefore got to learn first-hand a tiny portion of what is meant
when one says 'plant genetics are hard.'  I had it far easier than Sandra.  I
just used Trinity to make some de-novo transcriptomes and attempt to provide
some metrics about which ones are real and really different across conditions
in her experiment.  Her work was many thousands of times more difficult.  The
full S.betaceum trinotate annotation is quite large, so I just pulled a
portion as an example for this package.
}
