% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_se_colors}
\alias{get_se_colors}
\title{Get the colors from a summarized experiment.}
\usage{
get_se_colors(se, keep_underscore = TRUE)
}
\arguments{
\item{se}{Input se}

\item{keep_underscore}{Sanitize the columns for underscores?}
}
\description{
Get the colors from a summarized experiment.
}
\examples{
## Use this for examples of SE attribute functions.

if (is.null(pombe_se)) {
  source(system.file("examples", "create_se.R", package = "hpgltools"))
}

## Use this data structure to change things in these examples.
modify_se <- pombe_se

## Pull the 'batch' column from the sample metadata.
observed_batches <- batches(modify_se)
head(observed_batches)

new_batches <- rep("hi", length(observed_batches))
batches(modify_se) <- new_batches
batches(modify_se)

## I have been messing with my various color setter/getters
current_colors <- get_se_colors(modify_se)
head(current_colors)

current_colors <- get_colors(modify_se)
new_colors <- rep("red", length(current_colors))
colors(modify_se) <- new_colors
get_colors(modify_se)

## I broke something in this: FIXME
##current_libsize <- libsize(modify_se)
##current_libsize
##play_libsize <- current_libsize / 2
##libsize(modify_se) <- play_libsize
##current_libsize <- libsize(modify_se)
##current_libsize
}
