% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{get_sig_genes}
\alias{get_sig_genes}
\title{Get a set of up/down genes using the top/bottom n or >/< z scores away from the median.}
\usage{
get_sig_genes(table, n = NULL, z = NULL, fc = NULL, p = NULL,
  column = "logFC", fold = "plusminus", p_column = "adj.P.Val")
}
\arguments{
\item{table}{a table from limma/edger/deseq.}

\item{n}{a rank-order top/bottom number of genes to take.}

\item{z}{a number of z-scores >/< the median to take.}

\item{fc}{a number of fold-changes to take}

\item{p}{a p-value cutoff}

\item{column}{a column to use to distinguish top/bottom}

\item{fold}{an identifier reminding how to get the bottom portion of a fold-change (plusminus says to get the negative of the positive, otherwise 1/positive is taken).}

\item{p_column}{a column containing (adjusted or not)p-values}
}
\value{
a list of up/down genes
}
\description{
Get a set of up/down genes using the top/bottom n or >/< z scores away from the median.
}

