% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{get_sig_gsva_categories}
\alias{get_sig_gsva_categories}
\title{Attempt to score the results from simple_gsva()}
\usage{
get_sig_gsva_categories(
  gsva_result,
  cutoff = 0.95,
  excel = "excel/gsva_subset.xlsx",
  model_batch = FALSE,
  factor_column = "condition",
  factor = NULL,
  label_size = NULL,
  col_margin = 6,
  row_margin = 12,
  type = "mean"
)
}
\arguments{
\item{gsva_result}{Result from simple_gsva()}

\item{cutoff}{Significance cutoff}

\item{excel}{Excel file to write the results.}

\item{model_batch}{Add batch to limma's model.}

\item{factor_column}{When extracting significance information, use this
metadata factor.}

\item{factor}{Use this metadata factor as the reference.}

\item{label_size}{Used to make the category names easier to read at the expense
of dropping some.}

\item{col_margin}{Attempt to make heatmaps fit better on the screen with this and...}

\item{row_margin}{this parameter}
}
\value{
List containing the gsva results, limma results, scores, some plots, etc.
}
\description{
This function uses a couple of methods to try to get an idea of whether the
results from gsva are actually interesting.  It does so via the following
methods:
  1.  Use limma on the expressionset returned by simple_gsva(), this might
provide an idea of if there are changing signatures among the sample types.
  2.  Perform a simplified likelihood estimate to get a sense of the
significant categories.
}
\seealso{
[score_gsva_likelihoods()] [get_group_gsva_means()] [limma_pairwise()]
 [simple_gsva()]
}
