% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{get_snp_sets}
\alias{get_snp_sets}
\title{Collect variants associated with specific conditions.}
\usage{
get_snp_sets(
  snp_expt,
  factor = "pathogenstrain",
  stringency = NULL,
  do_save = FALSE,
  savefile = "variants.rda",
  proportion = 0.9
)
}
\arguments{
\item{snp_expt}{variant collection.}

\item{factor}{metadata factor}

\item{stringency}{method to determin 'real' variants.}

\item{do_save}{Save the result?}

\item{savefile}{outptu savefile.}

\item{proportion}{Used with stringency.}
}
\description{
Collect variants associated with specific conditions.
}
