% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\name{get_spyogenes_data}
\alias{get_spyogenes_data}
\title{Get the filenames of the Streptococcus pyogenes strain 5005 data.}
\usage{
get_spyogenes_data()
}
\description{
This should return the Group A Streptococcus pyogenes strain 5005 GFF file,
fasta genome, and some count tables. The gff contains the annotations which
correspond to NC_007297.  The rda file contains an expt which comprises a
subset of a RNASeq experiment performed with the McIver lab.  The goal was
to examine changes in transcription across two media types (the very
permissive THY and restrictive CDM).  In addition, this used two strains,
one which is more and less pathogenic.
}
