% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\name{get_tcruzi_data}
\alias{get_tcruzi_data}
\title{Sample sheet from a portion of an RNASeq experiment of Trypanosoma cruzi CL-Brener.}
\usage{
get_tcruzi_data()
}
\description{
This contains a portion of an experiment performed with Santuza in which we
were comparing two closely related T.cruzi strains: CL-14 and CL-Brener,
which are super-similar but vastly different in terms of their
pathogenicity.  (I would much rather be infected with CL-14!).  The count
tables were created via TopHat2 mapping of the CL-Br and CL-14 samples
using the CL-Brener genome.  One of the interesting and bizarre things about
CL-Brener: it is a multi-haplotype strain, containing bits and pieces from
two lineages, named 'Esmeraldo' and 'Non-Esmeraldo'.  In addition, there is a
large portion which has not been characterized and is therefore called
'Unassigned'.  Thus, when mapping the data, we create a concatenated genome
with all three haplotypes.  One thing we did not include in the paper
(because I didn't think of it until later), was an analysis of the single
nucleotide variants between the two strains.  RNASeq data is of course not an
ideal format for performing these analyses, but I think I figured out a
reasonable method to extract mostly robust differences (in snp.r).
}
