% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{get_vennerable_rows}
\alias{get_vennerable_rows}
\title{Extract rows of interesting information from vennerable.}
\usage{
get_vennerable_rows(tables, intersections)
}
\arguments{
\item{tables}{The set of tables associated with the intersections.}

\item{intersections}{The intersections themselves.}
}
\description{
Extract rows of interesting information from vennerable.
}
