% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{get_yyyymm_commit}
\alias{get_yyyymm_commit}
\title{Find the git commit closest to the given yyyymmdd.}
\usage{
get_yyyymm_commit(
  gitdir = "~/hpgltools",
  version = NULL,
  year = NULL,
  month = NULL,
  day = NULL
)
}
\arguments{
\item{gitdir}{Location of the git repository, I assume hpgltools.}

\item{version}{String containing all yyyymmdd.}

\item{year}{Chosen year which will be coerced to yyyy.}

\item{month}{Chosen month coerced to mm.}

\item{day}{Chosen day coerced to dd.}
}
\description{
Find the git commit closest to the given yyyymmdd.
}
