% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{getEdgeWeights}
\alias{getEdgeWeights}
\title{getEdgeWeights() Plot the ontology DAG}
\usage{
getEdgeWeights(graph)
}
\arguments{
\item{graph}{A graph from topGO

This function was stolen from topgo in order to figure out where it was failing}
}
\value{
weights
}
\description{
getEdgeWeights() Plot the ontology DAG
}

