% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gff2df}
\alias{gff2df}
\title{gff2df()  Try to make import.gff a little more robust}
\usage{
gff2df(gff)
}
\arguments{
\item{gff}{a gff filename

This function wraps import.gff/import.gff3/import.gff2 calls in try()
Because sometimes those functions fail in unpredictable ways.}
}
\value{
a df!
}
\description{
gff2df()  Try to make import.gff a little more robust
}

