% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grange.R
\name{ggcoverage_prepare}
\alias{ggcoverage_prepare}
\title{Copied the prepare function from ggcoverage and made some small changes.
Note, that extend is in nucleotides while padding is in genes.}
\usage{
ggcoverage_prepare(
  gr,
  region_string = NULL,
  gene_name = "HNRNPC",
  wanted_type = "gene",
  name_column = "gene_name",
  extend = 2000
)
}
\description{
Copied the prepare function from ggcoverage and made some small changes.
Note, that extend is in nucleotides while padding is in genes.
}
