% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.r
\name{ggplt}
\alias{ggplt}
\title{Simplify plotly ggplot conversion so that there are no shenanigans.}
\usage{
ggplt(
  gg,
  filename = "ggplot.html",
  selfcontained = TRUE,
  libdir = NULL,
  background = "white",
  title = class(gg)[[1]],
  knitrOptions = list(),
  ...
)
}
\arguments{
\item{gg}{Plot from ggplot2.}

\item{filename}{Output filename.}

\item{selfcontained}{htmlwidgets: Return the plot as a self-contained file
with images re-encoded base64.}

\item{libdir}{htmlwidgets: Directory into which to put dependencies.}

\item{background}{htmlwidgets: String for the background of the image.}

\item{title}{htmlwidgets: Title of the page!}

\item{knitrOptions}{htmlwidgets: I am not a fan of camelCase, but
nonetheless, options from knitr for htmlwidgets.}

\item{...}{Any remaining elipsis options are passed to ggplotly.}
}
\value{
The final output filename
}
\description{
I am a fan of ggplotly, but its conversion to an html file is not perfect.
This hopefully will get around the most likely/worst problems.
}
