% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{ggsignif_paired_genes}
\alias{ggsignif_paired_genes}
\title{Extract groups of samples from a SE and make a boxplot of their expression}
\usage{
ggsignif_paired_genes(
  exp,
  conditions = NULL,
  genes = NULL,
  norm = "raw",
  convert = "cpm",
  filter = TRUE,
  transform = "log2",
  batch = "svaseq",
  name_column = "hgnc_symbol",
  condition_column = "condition",
  test = "wilcox.test",
  excel = NULL
)
}
\arguments{
\item{exp}{Input dataset}

\item{conditions}{Pair of conditions to compare (hopefully more than a pair soon)}

\item{genes}{Vector of genes to plot}

\item{norm}{Normalization to apply}

\item{convert}{Conversion to apply}

\item{filter}{Filter to apply}

\item{transform}{Transformation}

\item{batch}{Batch method to apply}

\item{name_column}{Gene information column from which to get gene names.}

\item{condition_column}{Metadata column containing the conditions to compare.}

\item{test}{Pairwise test to invoke.}
}
\description{
Extract groups of samples from a SE and make a boxplot of their expression
}
