% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{godef}
\alias{godef}
\title{Get a go long-form definition from an id.}
\usage{
godef(go = "GO:0032432")
}
\arguments{
\item{go}{GO ID, this may be a character or list (assuming the elements are
goids).}
}
\value{
Some text providing the long definition of each provided GO id.
}
\description{
Sometimes it is nice to be able to read the full definition of some GO terms.
}
\examples{
\dontrun{
 godef("GO:0032432")
 ## > GO:0032432
 ## > "An assembly of actin filaments that are on the same axis but may be
 ## > same or opposite polarities and may be packed with different levels of tightness."
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}
