% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{golevel}
\alias{golevel}
\title{Get a go level approximation from a set of IDs.
This just wraps golev() in mapply.}
\usage{
golevel(go)
}
\arguments{
\item{go}{a character list of IDs.}
}
\value{
Some text
}
\description{
Get a go level approximation from a set of IDs.
This just wraps golev() in mapply.
}
\examples{
\dontrun{
 golevel(c("GO:0032559", "GO:0000001")
## > 3 4
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}

