% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{goont}
\alias{goont}
\title{Get a go ontology name from an ID}
\usage{
goont(go)
}
\arguments{
\item{id}{A go ID -- this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Get a go ontology name from an ID
}
\examples{
## goont(c("GO:0032432", "GO:0032433"))
## > GO:0032432 GO:0032433
## > "CC" "CC"
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

