% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology.R
\name{gosc}
\alias{gosc}
\title{gosc does the real work for gosec()}
\usage{
gosc(go)
}
\arguments{
\item{go}{A go id}
}
\value{
One of the following:
  "Not found" : for when the goID does not exist
  "" : when there is no secondary id
  or a character list of secondary IDs
}
\description{
gosc does the real work for gosec()
}

