% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_ontology_shared.r
\name{gosec}
\alias{gosec}
\title{Get a GO secondary ID from an id.}
\usage{
gosec(go = "GO:0032432")
}
\arguments{
\item{go}{GO ID, this may be a character or list(assuming the elements, not
names, are goids).}
}
\value{
Some text comprising the secondary GO id(s).
}
\description{
Unfortunately, GOTERM's returns for secondary IDs are not consistent, so this
function has to have a whole bunch of logic to handle the various outputs.
}
\examples{
\dontrun{
 gosec("GO:0032432")
 ## > GO:0032432
 ## > "GO:0000141" "GO:0030482"
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}
