% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{goseq2enrich}
\alias{goseq2enrich}
\title{Create a clusterProfiler compatible enrichResult data structure from a goseq result.}
\usage{
goseq2enrich(
  retlist,
  ontology = "MF",
  cutoff = 1,
  cutoff_column = "over_represented_pvalue",
  organism = NULL,
  padjust_method = "BH"
)
}
\arguments{
\item{retlist}{Result from simple_goseq().}

\item{ontology}{Ontology sub-tree of interest.}

\item{cutoff}{(adjusted)p cutoff.}

\item{organism}{Currently unused.}
}
\value{
enrichResult object ready to pass to things like dotplot.
}
\description{
The metrics and visualization methods in clusterProfiler are the
best.  It is not always trivial to get non-model organisms working
well with clusterProfiler.  Therefore I still like using tools like
topgo/goseq/gostats/gprofiler.  This function and its companions
seek to make them cross-compatible.  Ideally, they will lead me to
being able to rip out a lot of superfluous material.
}
