% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.r
\name{goseq_msigdb}
\alias{goseq_msigdb}
\title{Pass MSigDB categorical data to goseq and run it.}
\usage{
goseq_msigdb(
  sig_genes,
  signatures = "c2BroadSets",
  data_pkg = "GSVAdata",
  signature_category = "c2",
  current_id = "ENSEMBL",
  required_id = "ENTREZID",
  length_db = NULL,
  doplot = TRUE,
  adjust = 0.1,
  pvalue = 0.1,
  length_keytype = "transcripts",
  go_keytype = "entrezid",
  goseq_method = "Wallenius",
  padjust_method = "BH",
  bioc_length_db = "ensGene",
  excel = NULL,
  orgdb = "org.Hs.eg.db"
)
}
\arguments{
\item{sig_genes}{Character list of genes deemed significant.  I think in the
current implementation this must be just a list of IDs as opposed to the
full dataframe of interesting genes because we likely need to convert IDs.}

\item{signatures}{Used by load_gmt_signatures(), the signature file or set.}

\item{data_pkg}{Used by load_gmt_signatures().}

\item{signature_category}{Ibid, but the name of the signatures group.}

\item{current_id}{Used by convert_msig_ids(), when converting IDs, the name
of the existing type.}

\item{required_id}{What type to convert to in convert_msig_ids().}

\item{length_db}{Dataframe of lengths.  It is worth noting that goseq
explicitly states that one might wish to use other potentially confounding
factors here, but they only examine lenghts in their paper. Starting with
this parameter, everything is just passed directly to simple_goseq()}

\item{doplot}{Print the prior plot?}

\item{adjust}{passed to simple_goseq()}

\item{pvalue}{passed to simple_goseq()}

\item{length_keytype}{passed to simple_goseq()}

\item{go_keytype}{passed to simple_goseq()}

\item{goseq_method}{passed to simple_goseq()}

\item{padjust_method}{passed to simple_goseq()}

\item{bioc_length_db}{passed to simple_goseq()}

\item{excel}{passed to simple_goseq()}

\item{orgdb}{Ideally used to help goseq collect lengths.}
}
\value{
Some goseq data!
}
\description{
goseq is probably the easiest method to push varying data types into.  Thus
it was the first thing I thought of when looking to push MSigDB data into a
GSEA method.
}
\seealso{
[gsva] [goseq]
}
