% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{goseq_table}
\alias{goseq_table}
\title{Enhance the goseq table of gene ontology information.}
\usage{
goseq_table(df, file = NULL)
}
\arguments{
\item{df}{a dataframe of ontology information.  This is intended to
be the output from goseq including information like
numbers/category, GOids, etc.  It requires a column 'category' which contains: GO:000001 and such.}

\item{file}{a csv file to which to write the table}
}
\value{
the ontology table with annotation information included
}
\description{
Enhance the goseq table of gene ontology information.
}
\examples{
## annotated_go = goseq_table(go_ids)
## head(annotated_go, n=1)
## >        category numDEInCat numInCat over_represented_pvalue
## > 571  GO:0006364          9       26            4.655108e-08
## >      under_represented_pvalue       qvalue ontology
## > 571                 1.0000000 6.731286e-05       BP
## >                                term
## > 571                 rRNA processing
## >                               synonym
## > 571        "35S primary transcript processing, GO:0006365"
## >        secondary    definition
## > 571    GO:0006365   Any process involved in the conversion of a primary ribosomal RNA (rRNA) transcript into one or more mature rRNA molecules.
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

