% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq-test.r, R/ontology_plots.r
\name{goseq_trees}
\alias{goseq_trees}
\alias{goseq_trees}
\title{Make fun trees a la topgo from goseq data.}
\usage{
goseq_trees(goseq, goid_map = "id2go.map", score_limit = 0.01,
  overwrite = FALSE, selector = "topDiffGenes", pval_column = "adj.P.Val")

goseq_trees(goseq, goid_map = "id2go.map", score_limit = 0.01,
  overwrite = FALSE, selector = "topDiffGenes", pval_column = "adj.P.Val")
}
\arguments{
\item{goseq}{Data from goseq.}

\item{goid_map}{File to save go id mapping.}

\item{score_limit}{Score limit for the coloring.}

\item{overwrite}{Overwrite the trees?}

\item{selector}{Function for choosing genes.}

\item{pval_column}{Column to acquire pvalues.}

\item{de_genes}{Some differentially expressed genes.}

\item{godata}{Data from goseq.}

\item{goids_df}{Mapping of IDs to GO in the Ramigo expected format.}

\item{goid_map}{File to save go id mapping.}

\item{score_limit}{Score limit for the coloring.}

\item{overwrite}{Overwrite the trees?}

\item{selector}{Function for choosing genes.}

\item{pval_column}{Column to acquire pvalues.}
}
\value{
A plot!

A plot!
}
\description{
This seeks to force goseq data into a format suitable for topGO and then use its tree plotting
function to make it possible to see significantly increased ontology trees.

This seeks to force goseq data into a format suitable for topGO and then use its tree plotting
function to make it possible to see significantly increased ontology trees.
}
\seealso{
\pkg{Ramigo}

\pkg{Ramigo}
}
