% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{gosn}
\alias{gosn}
\title{gosn()  Does the actual go synonym extraction for gosyn()}
\usage{
gosn(go)
}
\arguments{
\item{go}{a go ID.}
}
\value{
some text
}
\description{
This merely calls as.character(Synonym(GOTERM[go]))
}

