% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{gostats2enrich}
\alias{gostats2enrich}
\title{Create a clusterProfiler compatible enrichResult data structure from a gostats result.}
\usage{
gostats2enrich(
  retlist,
  ontology = "MF",
  cutoff = 0.1,
  cutoff_column = "qvalue",
  organism = NULL,
  padjust_method = "BH"
)
}
\arguments{
\item{retlist}{Result from simple_gostats().}

\item{ontology}{Ontology sub-tree of interest.}

\item{cutoff}{(adjusted)p cutoff.}

\item{cutoff_column}{Choose a column of p-values.}

\item{organism}{Currently unused.}

\item{padjust_method}{Define the desired p.adjust method.}
}
\value{
enrichResult object ready to pass to things like dotplot.
}
\description{
The metrics and visualization methods in clusterProfiler are the
best.  It is not always trivial to get non-model organisms working
well with clusterProfiler.  Therefore I still like using tools like
topgo/goseq/gostats/gprofiler.  This function and its companions
seek to make them cross-compatible.  Ideally, they will lead me to
being able to rip out a lot of superfluous material.
}
