% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gostats.r
\name{gostats_trees}
\alias{gostats_trees}
\title{Make fun trees a la topgo from goseq data.}
\usage{
gostats_trees(de_genes, mf_over, bp_over, cc_over, mf_under, bp_under, cc_under,
  goid_map = "reference/go/id2go.map", score_limit = 0.01,
  goids_df = NULL, overwrite = FALSE, selector = "topDiffGenes",
  pval_column = "adj.P.Val")
}
\arguments{
\item{de_genes}{some differentially expressed genes}

\item{mf_over}{mfover data}

\item{bp_over}{bpover data}

\item{cc_over}{ccover data}

\item{mf_under}{mfunder data}

\item{bp_under}{bpunder data}

\item{cc_under}{ccunder expression data}

\item{goid_map}{a mapping of IDs to GO in the Ramigo expected format}

\item{score_limit}{maximum score to include as 'significant'}

\item{goids_df}{a dataframe of available goids (used to generate goid_map)}

\item{overwrite}{overwrite the goid_map?}

\item{selector}{a function to choose differentially expressed genes in the data}

\item{pval_column}{a column in the data to be used to extract pvalue scores}
}
\value{
plots! Trees! oh my!
}
\description{
Make fun trees a la topgo from goseq data.
}
\seealso{
\pkg{topGO}
}

