% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{gosyn}
\alias{gosyn}
\title{Get a go synonym from an ID
I think I will need to do similar parsing of the output for this function as per gosec()
In some cases this also returns stuff like c("some text", "GO:someID")
versus "some other text"  versus NULL versus NA}
\usage{
gosyn(go)
}
\arguments{
\item{id}{A go ID -- this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Get a go synonym from an ID
I think I will need to do similar parsing of the output for this function as per gosec()
In some cases this also returns stuff like c("some text", "GO:someID")
versus "some other text"  versus NULL versus NA
}
\examples{
## text =  gosyn("GO:0000001")
## text
## > GO:000001
## > "mitochondrial inheritance"
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

