% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{goterm}
\alias{goterm}
\title{Get a go term from ID.}
\usage{
goterm(go = "GO:0032559")
}
\arguments{
\item{go}{a go ID or list thereof
this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Get a go term from ID.
}
\examples{
\dontrun{
 goterm("GO:0032559")
## > GO:0032559
## > "adenyl ribonucleotide binding"
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}

