% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{gotest}
\alias{gotest}
\title{gotest()  Test GO ids to see if they are useful.
This just wraps gotst in mapply.}
\usage{
gotest(go)
}
\arguments{
\item{go}{go IDs as characters.}
}
\value{
Some text
}
\description{
gotest()  Test GO ids to see if they are useful.
This just wraps gotst in mapply.
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

