% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.R
\name{gprofiler22enrich}
\alias{gprofiler22enrich}
\title{Recast gProfiler data to the output class produced by clusterProfiler.}
\usage{
gprofiler22enrich(
  retlst,
  ontology = "MF",
  cutoff = 1,
  organism = NULL,
  padjust_method = "BH"
)
}
\arguments{
\item{ontology}{Category type to extract, currently only GO?}

\item{cutoff}{Use a p-value cutoff to get only the significant
categories?}

\item{organism}{Set the orgdb organism name?}

\item{padjust_method}{what it says on the tin.}

\item{retlist}{Output from simple_gprofiler()}
}
\description{
I would like to use the various clusterProfiler plots more easily.
Therefore I figured it would be advantageous to coerce the various
outputs from gprofiler and friends into the datastructure produced by
clusterProfiler.
}
