% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grange.R
\name{gr_from_to}
\alias{gr_from_to}
\title{Extract a grange from gene x to gene y}
\usage{
gr_from_to(
  gr,
  from,
  to,
  column = "gene_id",
  type_column = "type",
  type = NULL,
  padding_start = 1,
  padding_end = 1
)
}
\arguments{
\item{gr}{Input grange.}

\item{from}{Starting point.}

\item{to}{Ending point.}

\item{column}{What mcol column to use to extract IDs.}

\item{type_column}{Use this feature type.}

\item{type}{Type of entries from which to create the granges.}

\item{padding_start}{Add this many genes as padding to the grange before the start.}

\item{padding_end}{Add this many genes as padding to the grange after the end.}
}
\description{
Extract a grange from gene x to gene y
}
