% Generated by roxygen2 (4.0.1): do not edit by hand
\name{graph_metrics}
\alias{graph_metrics}
\title{Make a bunch of graphs describing the state of an experiment
before/after normalization.}
\usage{
graph_metrics(expt, norm_type = "quant", filter = "log2",
  out_type = "cpm", filter_low = TRUE, cormethod = "pearson",
  distmethod = "euclidean")
}
\arguments{
\item{expt}{an expt experiment containing the data, design, and colors}

\item{norm_type}{normalization strategy for the data.  Defaults to
quantile.}

\item{filter}{whether to log2/10 filter the data.  Defaults to
log2.}

\item{out_type}{whether to cpm/rpkm the data.  Defaults to cpm.}

\item{filter_low}{whether to low-count filter the data.  Defaults
to TRUE.}

\item{cormethod}{define the correlation test for heatmaps.
Defaults to pearson (Available: pearson, spearman, kendal, robust)}

\item{distmethod}{define the distance metric for heatmaps.
Defaults to euclidean (Lots are available, I don't understand them.)}
}
\value{
a list of plots.  This is a mix of ggplots and replayed built-ins.
}
\description{
Make a bunch of graphs describing the state of an experiment
before/after normalization.
}
\examples{
## toomany_plots = graph_metrics(expt)
## testnorm = graph_metrics(expt, norm_type="tmm", filter="log2", out_type="rpkm", cormethod="robust")
}
\seealso{
\code{\link{exprs}}, \code{\link{my_norm}},
\code{\link{graph_nonzero}}, \code{\link{my_libsize}},
\code{\link{my_boxplot}}, \code{\link{my_corheat}},
\code{\link{my_smc}}, \code{\link{my_disheat}},
\code{\link{my_smd}}, \code{\link{my_pca}},
\code{\link{replayPlot}}, \code{\link{recordPlot}}
}

