% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_shared.R
\name{group_mean_cds_length}
\alias{group_mean_cds_length}
\title{Average the cds length over known transcripts for a single gene.}
\usage{
group_mean_cds_length(
  annot,
  gene_column = "ensembl_gene_id",
  tx_column = "ensembl_transcript_id",
  length_column = "cds_length"
)
}
\arguments{
\item{annot}{Dataframe of annotations.}

\item{gene_column}{Column containing the gene IDs.}

\item{tx_column}{Column containing the transcript IDs.}

\item{length_column}{Column containing the cds lengths.}
}
\value{
The annotations with a new column 'mean_cdS_len' at the end.
}
\description{
Average the cds length over known transcripts for a single gene.
}
