% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{gsva_likelihoods}
\alias{gsva_likelihoods}
\title{Score the results from gsva().}
\usage{
gsva_likelihoods(gsva_result, score = NULL, category = NULL,
  factor = NULL, sample = NULL, factor_column = "condition",
  method = "mean")
}
\value{
The scores according to the provided category, factor, sample, or
score(s).
}
\description{
Yeah, this is a bit meta, but the scores from gsva seem a bit meaningless to
me, so I decided to look at the distribution of observed scores in some of my
data; I quickly realized that they follow a nicely normal distribution.
Therefore, I thought to calculate some scores of gsva() using that
information.
}
\details{
The nicest thing in this, I think, is that it provides its scoring metric(s)
according to a few different possibilities, including:
\itemize{
\item the mean of samples found in an experimental factor
\item All provided scores against the distribution of observed scores as
z-scores.
\item A single score against all scores.
\item Rows (gene sets) against the set of all gene sets.
}
}
