% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{guess_bitr_keytype}
\alias{guess_bitr_keytype}
\title{Use a heuristic to guess the appropriate keytype when using clusterProfiler::enricher().}
\usage{
guess_bitr_keytype(
  org,
  from,
  sig_genes = NULL,
  to = "ENTREZ",
  possible_keys = NULL,
  universe = NULL
)
}
\arguments{
\item{org}{orgdb containing the potential keys}

\item{from}{starting keytype}

\item{sig_genes}{Input gene set}

\item{to}{new keytype}

\item{possible_keys}{Set of keys to test}

\item{universe}{Universe of possible genes.}
}
\description{
Use a heuristic to guess the appropriate keytype when using clusterProfiler::enricher().
}
