% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{guess_factors}
\alias{guess_factors}
\title{A silly function to guesstimate factor columns in metadata.}
\usage{
guess_factors(meta_df, ratio = 3)
}
\description{
Use the heuristic that any column with a number of different
elements which is <= (samples / ratio) has a reasonable chance of
being usable as a categorical.
}
