% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{guess_num_surrogates}
\alias{guess_num_surrogates}
\title{Use methods from sva/RUVseq to guesstimate the number of surrogates.}
\usage{
guess_num_surrogates(
  design,
  linear_mtrx,
  log2_mtrx,
  model_svs = "svaseq",
  conditional_fstring = "~ condition",
  guess_type = "be"
)
}
\description{
Use methods from sva/RUVseq to guesstimate the number of surrogates.
}
