% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{heatmap.3}
\alias{heatmap.3}
\title{a minor change to heatmap.2 makes heatmap.3}
\usage{
heatmap.3(
  x,
  Rowv = TRUE,
  Colv = if (symm) "Rowv" else TRUE,
  distfun = dist,
  hclustfun = fastcluster::hclust,
  dendrogram = c("both", "row", "column", "none"),
  reorderfun = function(d, w) reorder(d, w),
  symm = FALSE,
  scale = c("none", "row", "column"),
  na.rm = TRUE,
  revC = identical(Colv, "Rowv"),
  add.expr,
  breaks,
  symbreaks = min(x < 0, na.rm = TRUE) || scale != "none",
  col = "heat.colors",
  colsep,
  rowsep,
  sepcolor = "white",
  sepwidth = c(0.05, 0.05),
  cellnote,
  notecex = 1,
  notecol = "cyan",
  na.color = par("bg"),
  trace = c("column", "row", "both", "none"),
  tracecol = "cyan",
  hline = median(breaks),
  vline = median(breaks),
  linecol = tracecol,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  cexRow = 0.2 + 1/log10(nr),
  cexCol = 0.2 + 1/log10(nc),
  labRow = NULL,
  labCol = NULL,
  srtRow = NULL,
  srtCol = NULL,
  adjRow = c(0, NA),
  adjCol = c(NA, 0),
  offsetRow = 0.5,
  offsetCol = 0.5,
  key = TRUE,
  keysize = 1.5,
  density.info = c("histogram", "density", "none"),
  denscol = tracecol,
  symkey = min(x < 0, na.rm = TRUE) || symbreaks,
  densadj = 0.25,
  key.title = NULL,
  key.xlab = NULL,
  key.ylab = NULL,
  key.xtickfun = NULL,
  key.ytickfun = NULL,
  key.par = list(),
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  lmat = NULL,
  lhei = NULL,
  lwid = NULL,
  extrafun = NULL,
  linewidth = 1,
  ...
)
}
\arguments{
\item{x}{data}

\item{Rowv}{add rows?}

\item{Colv}{add columns?}

\item{distfun}{distance function to use}

\item{hclustfun}{clustering function to use}

\item{dendrogram}{which axes to put trees on}

\item{reorderfun}{reorder the rows/columns?}

\item{symm}{symmetrical?}

\item{scale}{add the scale?}

\item{na.rm}{remove nas from the data?}

\item{revC}{reverse the columns?}

\item{add.expr}{no clue}

\item{breaks}{also no clue}

\item{symbreaks}{still no clue}

\item{col}{colors!}

\item{colsep}{column separator}

\item{rowsep}{row separator}

\item{sepcolor}{color to put between columns/rows}

\item{sepwidth}{how much to separate}

\item{cellnote}{mur?}

\item{notecex}{size of the notes}

\item{notecol}{color of the notes}

\item{na.color}{a parameter call to bg}

\item{trace}{do a trace for rows/columns?}

\item{tracecol}{color of the trace}

\item{hline}{the hline}

\item{vline}{the vline}

\item{linecol}{the line color}

\item{margins}{margins are good}

\item{ColSideColors}{colors for the columns as annotation}

\item{RowSideColors}{colors for the rows as annotation}

\item{cexRow}{row size}

\item{cexCol}{column size}

\item{labRow}{hmmmm}

\item{labCol}{still dont know}

\item{srtRow}{srt the row?}

\item{srtCol}{srt the column?}

\item{adjRow}{adj the row?}

\item{adjCol}{adj the column?}

\item{offsetRow}{how far to place the text from the row}

\item{offsetCol}{how far to place the text from the column}

\item{key}{add a key?}

\item{keysize}{if so, how big?}

\item{density.info}{for the key, what information to add}

\item{denscol}{tracecol hmm ok}

\item{symkey}{I like keys}

\item{densadj}{adj the dens?}

\item{key.title}{title for the key}

\item{key.xlab}{text for the x axis of the key}

\item{key.ylab}{text for the y axis of the key}

\item{key.xtickfun}{add text to the ticks of the key x axis}

\item{key.ytickfun}{add text to the ticks of the key y axis}

\item{key.par}{parameters for the key}

\item{main}{the main title of the plot}

\item{xlab}{main x label}

\item{ylab}{main y label}

\item{lmat}{the lmat}

\item{lhei}{the lhei}

\item{lwid}{the lwid}

\item{extrafun}{I do enjoy me some extra fun}

\item{linewidth}{the width of lines}

\item{...}{because this function did not already have enough options}
}
\value{
a heatmap!
}
\description{
heatmap.2 is the devil.
}
\seealso{
[heatmap.2()]
}
