% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{hpgl_arescore}
\alias{hpgl_arescore}
\title{hpgl_arescore()  Implement the arescan function in R}
\usage{
hpgl_arescore(x, basal = 1, overlapping = 1.5, d1.3 = 0.75, d4.6 = 0.4,
  d7.9 = 0.2, within.AU = 0.3, aub.min.length = 10,
  aub.p.to.start = 0.8, aub.p.to.end = 0.55)
}
\arguments{
\item{basal}{default=1  I dunno.}

\item{overlapping}{default=1.5}

\item{d1.3}{default=0.75  These parameter names are so stupid, lets be realistic}

\item{d4.6}{default=0.4}

\item{d7.9}{default=0.2}

\item{within.AU}{default=0.3}

\item{aub.min.length}{default=10}

\item{aub.p.to.start}{default=0.8}

\item{aub.p.to.end}{default=0.55}

\item{stringset}{A DNA/RNA StringSet containing the UTR sequences of interest}
}
\value{
a DataFrame of scores
}
\description{
This function was taken almost verbatim from AREScore() in SeqTools
Available at: https://github.com/lianos/seqtools.git
At least on my computer I could not make that implementation work
So I rewrapped its apply() calls and am now hoping to extend its logic
a little to make it more sensitive and get rid of some of the spurious
parameters or at least make them more transparent.
}

