% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_bcv_plot}
\alias{hpgl_bcv_plot}
\title{Steal EdgeR's plotBCV()}
\usage{
hpgl_bcv_plot(data)
}
\arguments{
\item{expt}{}
}
\value{
a plot! of the BCV a la ggplot2.
}
\description{
Steal EdgeR's plotBCV()
}

