% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_boxplot}
\alias{hpgl_boxplot}
\title{Make a ggplot boxplot of a set of samples.}
\usage{
hpgl_boxplot(df = NULL, colors_fill = NULL, names = NULL, expt = NULL,
  title = NULL, scale = "raw", ...)
}
\arguments{
\item{df}{alternately a data frame which must be accompanied by}

\item{names}{a nicer version of the sample names}

\item{expt}{an expt set of samples}

\item{scale}{whether to log scale the y-axis}

\item{design}{a design matrix and}

\item{colors}{a color scheme}
}
\value{
a ggplot2 boxplot of the samples.  Each boxplot
contains the following information: a centered line describing the
median value of counts of all genes in the sample, a box around the
line describing the inner-quartiles around the median (quartiles 2
and 3 for those who are counting), a vertical line above/below the
box which shows 1.5x the inner quartile range (a common metric of
the non-outliers), and single dots for each gene which is outside
that range.  A single dot is transparent.
}
\description{
Make a ggplot boxplot of a set of samples.
}
\examples{
## a_boxplot = hpgl_boxplot(expt=expt)
## a_boxplot  ## ooo pretty boxplot look at the lines
}
\seealso{
\code{\link{geom_boxplot}}, \code{\link{melt}},
\code{\link{scale_x_discrete}}
}

