% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{hpgl_cor}
\alias{hpgl_cor}
\title{Wrap cor() to include robust correlations.}
\usage{
hpgl_cor(df, method = "pearson", ...)
}
\arguments{
\item{df}{a data frame to test.}

\item{method}{correlation method to use. Includes pearson, spearman, kendal, robust.}

\item{...}{other options to pass to stats::cor()}
}
\value{
correlation some fun correlation statistics
}
\description{
Wrap cor() to include robust correlations.
}
\examples{
\dontrun{
hpgl_cor(df=df)
hpgl_cor(df=df, method="robust")
}
}
\seealso{
\pkg{robust} \link{cor} \link{cov} \link[robust]{covRob}
}

