% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{hpgl_corheat}
\alias{hpgl_corheat}
\title{Make a heatmap.3 description of the correlation between samples.}
\usage{
hpgl_corheat(data, colors = NULL, design = NULL, method = "pearson",
  names = NULL, title = NULL, ...)
}
\arguments{
\item{data}{Dataframe, expt, or expressionset to work with.}

\item{colors}{Color scheme for the samples, not needed if this is an expt.}

\item{design}{Design matrix describing the experiment, not needed if this is an expt.}

\item{method}{Correlation statistic to use. (pearson, spearman, kendall, robust).}

\item{names}{Alternate names to use for the samples.}

\item{title}{Title for the plot.}

\item{...}{More options are wonderful!}
}
\value{
Gplots heatmap describing describing how the samples are clustering vis a vis pairwise correlation.
}
\description{
Given a set of count tables and design, this will calculate the pairwise correlations and plot
them as a heatmap.  It attempts to standardize the inputs and eventual output.
}
\examples{
## corheat_plot = hpgl_corheat(expt=expt, method="robust")
## corheat_plot
}
\seealso{
\link{hpgl_cor} \link[RColorBrewer]{brewer.pal}
\link[grDevices]{recordPlot}
}

