% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_density}
\alias{hpgl_density}
\title{hpgl_density()  Density plots!}
\usage{
hpgl_density(data, colors = NULL, names = NULL, position = "identity",
  fill = NULL, title = NULL, log = FALSE)
}
\arguments{
\item{data}{an expt, expressionset, or data frame.}

\item{colors}{default=NULL  a color scheme to use.}

\item{names}{default=NULL  names of the samples.}

\item{position}{default='identity'  how to place the lines, either let them overlap (identity), or stack them.}

\item{fill}{default=NULL  fill the distributions?  This might make the plot unreasonably colorful.}

\item{title}{default=NULL  a title for the plot.}

\item{log}{default=FALSE  plot on the log scale?}
}
\value{
a density plot!
}
\description{
hpgl_density()  Density plots!
}

