% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_helpers_misc.r
\name{hpgl_dist}
\alias{hpgl_dist}
\title{Because I am not smart enough to remember t()}
\usage{
hpgl_dist(df, method = "euclidean", ...)
}
\arguments{
\item{df}{data frame from which to calculate distances.}

\item{method}{Which distance calculation to use?}

\item{...}{Extra arguments for dist.}
}
\description{
It seems to me there should be a function as easy for distances are there is for correlations.
}
