% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_enrich.internal}
\alias{hpgl_enrich.internal}
\title{A minor hack in the clusterProfiler function 'enrich.internal'}
\usage{
hpgl_enrich.internal(gene, organism, pvalueCutoff = 1, pAdjustMethod = "BH",
  ont, minGSSize = 2, qvalueCutoff = 0.2, readable = FALSE,
  universe = NULL)
}
\arguments{
\item{gene}{some differentially expressed genes}

\item{organism}{by default 'human'}

\item{pvalueCutoff}{default=1  a pvalue cutoff}

\item{pAdjustMethod}{default='BH' p adjust method}

\item{ont}{by default 'MF'}

\item{minGSSize}{default=2  a minimum gs size}

\item{qvalueCutoff}{default=0.2  maximum q value}

\item{readable}{default=FALSE  set the readable flag for dose}

\item{universe}{default=NULL a universe to use}
}
\value{
some clusterProfiler data
}
\description{
A minor hack in the clusterProfiler function 'enrich.internal'
}
\seealso{
\pkg{clusterProfiler}
}

