% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_enrichGO}
\alias{hpgl_enrichGO}
\title{A minor hack in the clusterProfiler function 'enrichGO'.}
\usage{
hpgl_enrichGO(gene, organism = "human", ont = "MF", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", universe, qvalueCutoff = 0.2, minGSSize = 2,
  readable = FALSE)
}
\arguments{
\item{gene}{Some differentially expressed genes.}

\item{organism}{if used will cause this to pull the ensG annotations.}

\item{ont}{Molecular function, Biological process, or Cellular component?}

\item{pvalueCutoff}{P-value cutoff.}

\item{pAdjustMethod}{P-value adjustment.}

\item{universe}{Gene universe to use.}

\item{qvalueCutoff}{Maximum qvalue before adding.}

\item{minGSSize}{Smallest ontology group size allowed.}

\item{readable}{Set the readable tag on the returned object?}
}
\value{
Some clusterProfiler data.
}
\description{
I do not remember any longer why, but enrichGO errors out in ways which do not always make sense,
this was written to alleviate that problem.  I believe I sent a diff to the clusterProfiler
author but did not hear back and so added this function.
}
\seealso{
\pkg{clusterProfiler}
}

