% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{hpgl_filter_counts}
\alias{hpgl_filter_counts}
\title{Filter low-count genes from a data set using cpm data and a threshold.}
\usage{
hpgl_filter_counts(count_table, threshold = 2, min_samples = 2,
  libsize = NULL, ...)
}
\arguments{
\item{count_table}{Data frame of (pseudo)counts by sample.}

\item{threshold}{Lower threshold of counts for each gene.}

\item{min_samples}{Minimum number of samples.}
}
\value{
Dataframe of counts without the low-count genes.
}
\description{
This is identical to cbcb_filter_counts except it does not do the somewhat tortured
log2CPM() but instead just uses a 4 cpm non-log threshold.  It should therefore give
basically the same result, but without the shenanigans.
}
\examples{
\dontrun{
 filtered_table <- cbcb_filter_counts(count_table)
}
}
\seealso{
\pkg{edgeR}
}
