% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_Gff2GeneTable}
\alias{hpgl_Gff2GeneTable}
\title{A copy and paste of clusterProfiler's readGff}
\usage{
hpgl_Gff2GeneTable(gffFile, compress = TRUE, split = "=")
}
\arguments{
\item{gffFile}{a gff file}

\item{compress}{default=TRUE compress them}

\item{split}{default='='  the splitter when reading gff files}
}
\description{
A copy and paste of clusterProfiler's readGff
}

