% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_topgo.R
\name{hpgl_GroupDensity}
\alias{hpgl_GroupDensity}
\title{hpgl_GroupDensity()  A hack of topGO's groupDensity()}
\usage{
hpgl_GroupDensity(object, whichGO, ranks = TRUE, rm.one = FALSE)
}
\arguments{
\item{object}{a topGO enrichment object}

\item{whichGO}{an individual ontology group to compare with}

\item{ranks}{default=TRUE  rank order the set of ontologies}

\item{rm.one}{default=FALSE  remove pvalue=1 groups}
}
\description{
This just adds a couple wrappers to avoid errors in groupDensity.
}

