% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_libsize}
\alias{hpgl_libsize}
\title{Make a ggplot graph of library sizes.}
\usage{
hpgl_libsize(data, colors = NULL, names = NULL, text = TRUE,
  title = NULL, yscale = NULL, ...)
}
\arguments{
\item{data}{an expt, dataframe, or expressionset of samples.}

\item{colors}{a color scheme.}

\item{names}{alternate names for the x-axis.}

\item{text}{add the numeric values inside the top of the bars of the plot?}

\item{title}{a title for the plot.}

\item{yscale}{whether or not to log10 the y-axis.}

\item{...}{more parameters for your good time}
}
\value{
a ggplot2 bar plot of every sample's size
}
\description{
Make a ggplot graph of library sizes.
}
\examples{
\dontrun{
 libsize_plot = hpgl_libsize(expt=expt)
 libsize_plot  ## ooo pretty bargraph
}
}
\seealso{
\link[ggplot2]{geom_bar} \link[ggplot2]{geom_text}
\link{prettyNum} \link[ggplot2]{scale_y_log10}
}

