% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{hpgl_ma_plot}
\alias{hpgl_ma_plot}
\title{Make a pretty MA plot from the output of voom/limma/eBayes/toptable.}
\usage{
hpgl_ma_plot(counts, de_genes, pval_cutoff = 0.05, alpha = 0.4,
  logfc_cutoff = 1, pval = "adjpval", size = 2, tooltip_data = NULL,
  gvis_filename = NULL, ...)
}
\arguments{
\item{counts}{Df of linear-modelling, normalized counts by sample-type,
which is to say the output from voom/voomMod/hpgl_voom().}

\item{de_genes}{Df from toptable or its friends containing p-values.}

\item{pval_cutoff}{Cutoff defining significant from not.}

\item{alpha}{How transparent to make the dots.}

\item{logfc_cutoff}{Fold change cutoff.}

\item{pval}{Name of the pvalue column to use for cutoffs.}

\item{size}{How big are the dots?}

\item{tooltip_data}{Df of tooltip information for gvis.}

\item{gvis_filename}{Filename to write a fancy html graph.}

\item{...}{More options for you!}
}
\value{
Ggplot2 MA scatter plot.  This is defined as the rowmeans of the normalized counts by
    type across all sample types on the x-axis, and the log fold change between conditions on the
    y-axis. Dots are colored depending on if they are 'significant.'  This will make a fun clicky
    googleVis graph if requested.
}
\description{
Make a pretty MA plot from the output of voom/limma/eBayes/toptable.
}
\examples{
\dontrun{
hpgl_ma_plot(voomed_data, toptable_data, gvis_filename="html/fun_ma_plot.html")
## Currently this assumes that a variant of toptable was used which
## gives adjusted p-values.  This is not always the case and I should
## check for that, but I have not yet.
}
}
\seealso{
\link{hpgl_gvis_ma_plot} \link[limma]{toptable}
\link[limma]{voom} \link{hpgl_voom}
\link[limma]{lmFit} \link[limma]{makeContrasts}
\link[limma]{contrasts.fit}
}

