% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_multihistogram}
\alias{hpgl_multihistogram}
\title{Make a pretty histogram of multiple datasets.}
\usage{
hpgl_multihistogram(data, log = FALSE, binwidth = NULL, bins = NULL,
  verbose = FALSE)
}
\arguments{
\item{data}{a dataframe of lots of pretty numbers, this also accepts lists.}

\item{log}{plot the data on the log scale?}

\item{binwidth}{set a static bin width with an unknown # of bins?  If neither of these are provided, then bins is set to 500, if both are provided, then bins wins.}

\item{bins}{set a static # of bins of an unknown width?}

\item{verbose}{be verbose?}
}
\value{
a ggplot histogram comparing multiple data sets
Along the way this generates pairwise t tests of the columns of
data.
}
\description{
Make a pretty histogram of multiple datasets.
}
\examples{
\dontrun{
 kittytime = hpgl_multihistogram(df)
}
}
\seealso{
\link[stats]{pairwise.t.test} \link[plyr]{ddply}
}

