% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{hpgl_nonzero}
\alias{hpgl_nonzero}
\title{Make a ggplot graph of the number of non-zero genes by sample.}
\usage{
hpgl_nonzero(data, design = NULL, colors = NULL, labels = NULL,
  title = NULL, ...)
}
\arguments{
\item{data}{Expt, expressionset, or dataframe.}

\item{design}{Eesign matrix.}

\item{colors}{Ccolor scheme.}

\item{labels}{How do you want to label the graph? 'fancy' will use directlabels() to try to match
the labels with the positions without overlapping anything else will just stick them on a 45'
offset next to the graphed point}

\item{title}{Add a title?}

\item{...}{rawr!}
}
\value{
a ggplot2 plot of the number of non-zero genes with respect to each library's CPM
}
\description{
Made by Ramzi Temanni <temanni at umd dot edu>
}
\examples{
\dontrun{
 nonzero_plot = hpgl_nonzero(expt=expt)
 nonzero_plot  ## ooo pretty
}
}
\seealso{
\link[ggplot2]{geom_point} \link[directlabels]{geom_dl}
}

