% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{hpgl_padjust}
\alias{hpgl_padjust}
\title{Wrap p.adjust to add IHW adjustments as an option.}
\usage{
hpgl_padjust(
  data,
  pvalue_column = "pvalue",
  mean_column = "base_mean",
  method = "fdr",
  significance = 0.05,
  type = NULL
)
}
\arguments{
\item{data}{Column or table containing values to adjust.}

\item{pvalue_column}{Name of the column in a table containing the p-values.}

\item{mean_column}{Name of the column in a table containing the mean count
values to weight.}

\item{method}{p adjustment method to apply.}

\item{significance}{Passed to IHW}

\item{type}{Assuming a DE table, what type of DE is this?}
}
\value{
Newly adjusted p-values using either p.adjust() or IHW.
}
\description{
IHW and apeglm are the two new toys I found, this adds the former as a way to
adjust p-values.
}
\seealso{
[IHW]
}
