% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_pairwise_ma}
\alias{hpgl_pairwise_ma}
\title{Plot all pairwise MA plots in an experiment.}
\usage{
hpgl_pairwise_ma(data, log = NULL, ...)
}
\arguments{
\item{data}{an expt expressionset or data frame}

\item{log}{is the data in log format?}

\item{...}{more options are good}
}
\value{
a list of affy::maplots
}
\description{
Use affy's ma.plot() on every pair of columns in a data set to help
diagnose problematic samples.
}
\examples{
\dontrun{
 ma_plots = hpgl_pairwise_ma(expt=some_expt)
}
}
\seealso{
\link[affy]{ma.plot}
}

