% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{hpgl_pca}
\alias{hpgl_pca}
\title{hpgl_pca()  Make a ggplot PCA plot describing the samples' clustering.}
\usage{
hpgl_pca(data, design = NULL, plot_colors = NULL, plot_labels = NULL,
  plot_title = NULL, plot_size = 5, ...)
}
\arguments{
\item{data}{an expt set of samples.}

\item{design}{default=NULL  a design matrix and.}

\item{plot_colors}{default=NULL  a color scheme.}

\item{plot_labels}{default=NULL  add labels?  Also, what type?  FALSE, "default", or "fancy".}

\item{plot_title}{default=NULL  a title for the plot.}

\item{plot_size}{default=5  size for the glyphs on the plot.}
}
\value{
a list containing the following:
  pca = the result of fast.svd()
  plot = ggplot2 pca_plot describing the principle component analysis of the samples.
  table = a table of the PCA plot data
  res = a table of the PCA res data
  variance = a table of the PCA plot variance
This makes use of cbcbSEQ and prints the table of variance by component.
}
\description{
hpgl_pca()  Make a ggplot PCA plot describing the samples' clustering.
}
\examples{
## pca_plot = hpgl_pca(expt=expt)
## pca_plot
}
\seealso{
\code{\link{makeSVD}}, \code{\link{pcRes}},
\code{\link{geom_dl}}
}

