% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{hpgl_qq_plot}
\alias{hpgl_qq_plot}
\title{Perform a qqplot between two columns of a matrix.}
\usage{
hpgl_qq_plot(data, x = 1, y = 2, labels = TRUE)
}
\arguments{
\item{data}{Data frame/expt/expressionset.}

\item{x}{First column to compare.}

\item{y}{Second column to compare.}

\item{labels}{Include the lables?}
}
\value{
a list of the logs, ratios, and mean between the plots as ggplots.
}
\description{
Given two columns of data, how well do the distributions match one another?  The answer to that
question may be visualized through a qq plot!
}

