% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_tables.R
\name{hpgl_read_files}
\alias{hpgl_read_files}
\title{Read a bunch of count tables and create a usable data frame from
them.}
\usage{
hpgl_read_files(ids, files, header = FALSE, include_summary_rows = FALSE,
  suffix = NULL, ...)
}
\arguments{
\item{ids}{a list of experimental ids}

\item{files}{a list of files to read}

\item{header}{whether or not the count tables include a header row
(default: FALSE)}

\item{include_summary_rows}{whether HTSeq summary rows should be included
       (default: FALSE)}
}
\value{
count_table a data frame of count tables
}
\description{
Read a bunch of count tables and create a usable data frame from
them.
}
\examples{
## count_tables = hpgl_read_files(as.character(sample_ids), as.character(count_filenames))
}
\seealso{
\code{\link{create_experiment}}
}

