% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_convert.R
\name{hpgl_rpkm}
\alias{hpgl_rpkm}
\title{Reads/(kilobase(gene) * million reads)}
\usage{
hpgl_rpkm(
  count_table,
  annotations,
  length_column = "length",
  length_na = 1000,
  ...
)
}
\arguments{
\item{count_table}{Data frame of counts, alternately an edgeR DGEList.}

\item{annotations}{dataframe of annotation information}

\item{length_column}{Column in the annotations with gene lengths.}

\item{length_na}{Set the length of NA to a specific value, 1000 to leave it un-normalized
or Inf to force its value to 0}

\item{...}{extra options including annotations for defining gene lengths.}
}
\value{
Data frame of counts expressed as rpkm.
}
\description{
Express a data frame of counts as reads per kilobase(gene) per
million(library). This function wraps EdgeR's rpkm in an attempt to make sure
that the required gene lengths get sent along.
}
\examples{
\dontrun{
 rpkm_df = hpgl_rpkm(df, annotations = gene_annotations)
}
}
\seealso{
[edgeR::rpkm()]
}
