% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{hpgl_sample_heatmap}
\alias{hpgl_sample_heatmap}
\title{Make a heatmap.3 description of the similarity of the genes among samples.}
\usage{
hpgl_sample_heatmap(data, colors = NULL, design = NULL, names = NULL,
  title = NULL, Rowv = FALSE, ...)
}
\arguments{
\item{data}{an expt/expressionset/dataframe set of samples}

\item{colors}{a color scheme}

\item{design}{a design matrix}

\item{names}{add names?}

\item{title}{title of the plot.}

\item{Rowv}{include the row names}

\item{...}{more parameters for a good time}
}
\value{
a recordPlot() heatmap describing the samples.
}
\description{
Make a heatmap.3 description of the similarity of the genes among samples.
}
\seealso{
\link[RColorBrewer]{brewer.pal} \link[grDevices]{recordPlot}
}

