% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_scatter}
\alias{hpgl_scatter}
\title{hpgl_scatter()  Make a pretty scatter plot between two sets of numbers.}
\usage{
hpgl_scatter(df, tooltip_data = NULL, color = "black",
  gvis_filename = NULL, size = 2)
}
\arguments{
\item{df}{a dataframe likely containing two columns}

\item{tooltip_data}{default=NULL  a df of tooltip information for gvis}

\item{color}{default='black'  color of the dots on the graph.}

\item{gvis_filename}{default=NULL  a filename to write a fancy html graph.}

\item{size}{default=3  the size of the dots on the graph.}
}
\value{
a ggplot2 scatter plot.
}
\description{
hpgl_scatter()  Make a pretty scatter plot between two sets of numbers.
}
\examples{
## hpgl_scatter(lotsofnumbers_intwo_columns, tooltip_data=tooltip_dataframe, gvis_filename="html/fun_scatterplot.html")
}
\seealso{
\code{\link{hpgl_gvis_scatter}}, \code{\link{geom_scatter}},
\code{\link{hpgl_linear_scatter}}
}

