% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_smc}
\alias{hpgl_smc}
\title{Make an R plot of the standard median correlation among samples}
\usage{
hpgl_smc(df = NULL, colors = NULL, expt = NULL, method = "pearson",
  names = NULL, title = NULL, ...)
}
\arguments{
\item{df}{alternately a data frame which must be accompanied by}

\item{colors}{a color scheme}

\item{expt}{an expt set of samples}

\item{method}{a correlation method to use.  Defaults to pearson.}

\item{names}{use pretty names for the samples?}

\item{design}{a design matrix and}
}
\value{
a recordPlot() of the standard median pairwise correlation
among the samples.  This will also write to an
open device.  The resulting plot measures the median correlation of
each sample among its peers.  It notes 1.5* the interquartile range
among the samples and makes a horizontal line at that correlation
coefficient.  Any sample which falls below this line is considered
for removal because it is much less similar to all of its peers.
}
\description{
Make an R plot of the standard median correlation among samples
}
\examples{
## smc_plot = hpgl_smc(expt=expt)
}
\seealso{
\code{\link{hpgl_cor}}, \code{\link{matrixStats::rowMedians}},
\code{\link{quantile}}, \code{\link{diff}}, \code{\link{recordPlot}}
}

