% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_smd}
\alias{hpgl_smd}
\title{Make an R plot of the standard median distance among samples}
\usage{
hpgl_smd(data, colors = NULL, names = NULL, method = "euclidean",
  title = NULL, ...)
}
\arguments{
\item{colors}{a color scheme}

\item{names}{use pretty names for the samples?}

\item{method}{a distance metric to use.  Defaults to euclidean.}

\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}
}
\value{
smd_plot a recordPlot of plot.  This will also write to an
open device.  This plot takes the median distance of each sample
with all of its peers.  It then calculates 1.5* the interquartile
range of distances.  Any sample which has a median distance greater
than this is considered for removal.
}
\description{
Make an R plot of the standard median distance among samples
}
\examples{
## smd_plot = hpgl_smd(expt=expt)
}
\seealso{
\code{\link{dist}}, \code{\link{quantile}},
\code{\link{diff}}, \code{\link{recordPlot}}
}

