% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\docType{class}
\name{hpgltools::all_cprofiler-class}
\alias{hpgltools::all_cprofiler-class}
\title{I think this is wrong.  all_cprofiler returns a list named by
the tablesnd their clusterProfiler results; I should therefore
clarify how they are organized in the body of the function
and therefore in the representation of the class.}
\description{
I think this is wrong.  all_cprofiler returns a list named by
the tablesnd their clusterProfiler results; I should therefore
clarify how they are organized in the body of the function
and therefore in the representation of the class.
}
