% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\docType{class}
\name{hpgltools::all_pairwise-class}
\alias{hpgltools::all_pairwise-class}
\title{I am working to properly define my various classes
and using a combination of https://rguides.dev/guides/r-s4-classes/
the book 'Software for Data Analysis', and
https://adv-r.hadley.nz/s4.html in order to learn the process.}
\description{
For the moment, I am going to define the representation of my
classes as the important elements of their return list.  My assumption
is that it will be immediately useful to have a getter for each
of those elements.
}
\details{
For now, the elements like basic/deseq/etc will just be 'list'
but as I fill in S4 heirarchy for hpgltools, that should change
to things like 'hpgltools::de_basic', and the de_basic class
should have its own proper representation, validation, etc...

But, until I complete that process, I think just using the base
class (list) for these elements should get me a resonable start.
}
