% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\docType{class}
\name{hpgltools::combine_de_tables-class}
\alias{hpgltools::combine_de_tables-class}
\title{I am adding a new verb to my understanding of the S4 language}
\description{
Well, let us be honest, understanding is a bit of an overstatement.
But if I properly understand what is going on, I should start
making proper class declarations for the various things I am doing
}
\details{
When using setClass, one may define the slots with a c() vector,
or use representation with a list(); presumably using the list
form allows one to nest stuff and do weird things I do not yet get.
}
