% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\docType{package}
\name{hpgltools}
\alias{hpgltools-package}
\alias{hpgltools}
\title{hpgltools: a suite of tools to make our analyses easier}
\description{
This provides a series of helpers for working with sequencing data
}
\details{
It falls under a few main topics

\itemize{
 \item Data exploration, look for trends in sequencing data and identify batch
       effects or skewed distributions.
 \item Differential expression analyses, use DESeq2/limma/EdgeR in a hopefully
       robust and flexible fashion.
 \item Ontology analyses, use goseq/clusterProfiler/topGO/GOStats/gProfiler in
       hopefully robust ways.
 \item Perform some simple TnSeq analyses.
}

To see examples of this in action, check out the hpgltool2 vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abelew/hpgltools}
}

}
\author{
\strong{Maintainer}: Ashton Belew \email{abelew@umd.edu}

Authors:
\itemize{
  \item April Hussey \email{ahussey@umd.edu}
  \item Keith Hughitt \email{khughitt@umd.edu}
  \item Jerry Zhao \email{zhaojerry123@gmail.com}
}

}
