% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{hs_expt_rda}
\alias{hs_expt_rda}
\title{Subset of a human RNASeq expressionset.}
\format{
An object of class \code{character} of length 1.
}
\usage{
hs_expt_rda
}
\description{
This is a portion of an expressionset used to examine changes caused by
infection with Leishmania panamensis.
}
\keyword{datasets}
