% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_methods_other.R
\name{iDA,matrix-method}
\alias{iDA,matrix-method}
\title{Set method for matrix to input data to iDA}
\usage{
\S4method{iDA}{matrix}(object, ...)
}
\arguments{
\item{object}{The object to run iDA on}

\item{...}{Additonal arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights
}
\description{
Set method for matrix to input data to iDA
}
