% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_methods.R
\name{iDA}
\alias{iDA}
\title{Generic method to input data to iDA}
\usage{
iDA(object, ...)
}
\arguments{
\item{object}{The object to run iDA on}

\item{...}{Additonal arguments passed to object constructors}
}
\value{
iDA output with clustering, gene weights, and cell weights
}
\description{
Generic method to input data to iDA
}
