% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{ignorant_pairwise}
\alias{ignorant_pairwise}
\title{ignorant_pairwise()  Perform a pairwise comparison among conditions which takes
nothing into account.  It _only_ takes the conditions, a mean value/variance among
them, divides by condition, and returns the result.  No fancy nomalizations, no
statistical models, no nothing.  It should be the very worst method possible.
But, it should also provide a baseline to compare the other tools against, they should
all do better than this, always.}
\usage{
ignorant_pairwise(input, design)
}
\arguments{
\item{conditions}{}
}
\value{
I am not sure yet
}
\description{
ignorant_pairwise()  Perform a pairwise comparison among conditions which takes
nothing into account.  It _only_ takes the conditions, a mean value/variance among
them, divides by condition, and returns the result.  No fancy nomalizations, no
statistical models, no nothing.  It should be the very worst method possible.
But, it should also provide a baseline to compare the other tools against, they should
all do better than this, always.
}
\seealso{
\code{\link{limma}} \code{\link{deseq2}} \code{\link{edger}}
}

