% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{ihw_adjust}
\alias{ihw_adjust}
\title{Make sure the outputs from limma and friends are in a format suitable for IHW.}
\usage{
ihw_adjust(
  de_result,
  pvalue_column = "pvalue",
  type = NULL,
  mean_column = "baseMean",
  significance = 0.05
)
}
\arguments{
\item{de_result}{Table which should have the 2 types of requisite columns:
mean value of counts and p-value.}

\item{pvalue_column}{Name of the column of p-values.}

\item{type}{If specified, this will explicitly perform the calculation for
the given type of differential expression analysis: limma, edger, deseq,
etc.}

\item{mean_column}{Name of the column of mean values.}

\item{significance}{IHW uses this parameter, I don't know why.}
}
\value{
weight adjusted p-values.
}
\description{
IHW seems like an excellent way to improve the confidence in the p-values
provided by the various DE methods.  It expects inputs fairly specific to
DESeq2, however, it is trivial to convert other methods to this, ergo this
function.
}
\details{
https://bioconductor.org/packages/release/bioc/vignettes/IHW/inst/doc/introduction_to_ihw.html
}
