% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_deseq.r
\name{import_deseq}
\alias{import_deseq}
\title{Try to add data to DESeq in a flexible fashion.  This currently only handles
matrices, htseq data, and tximport data.}
\usage{
import_deseq(data, column_data, model_string, tximport = NULL)
}
\arguments{
\item{data}{Counts from htseq/mtrx/tximport/etc}

\item{column_data}{I think this is the sample names, I forget.}

\item{model_string}{Model describing the data by sample names.}

\item{tximport}{Where is this data coming from?}
}
\description{
This will hopefully make adding counts to a DESeq data set easier, as it
tries to handle the various arguments with minimal fuss.
}
