% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_edger.r
\name{import_edger}
\alias{import_edger}
\title{Import tximport information into edgeR.}
\usage{
import_edger(data, conditions, tximport = NULL)
}
\arguments{
\item{data}{data to be coerced into edgeR.}

\item{conditions}{Set of conditions used to make the DGEList.}

\item{tximport}{Tell this if the data is actually coming from tximport.}
}
\value{
Hopefully valid DGEList for edgeR.
}
\description{
This was taken from the tximport manual with minor modifications.
}
\seealso{
[import_deseq()]
}
