% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r, R/testfile.r
\name{intersect_signatures}
\alias{intersect_signatures}
\title{Take a result from simple_gsva(), a list of gene IDs, and intersect them.}
\usage{
intersect_signatures(
  gsva_result,
  lst,
  freq_cutoff = 2,
  sig_weights = TRUE,
  gene_weights = TRUE
)

intersect_signatures(
  gsva_result,
  lst,
  freq_cutoff = 2,
  sig_weights = TRUE,
  gene_weights = TRUE
)
}
\arguments{
\item{gsva_result}{Result from simple_gsva().}

\item{lst}{List of genes of interest.}

\item{freq_cutoff}{Minimum number of observations to be counted.}

\item{sig_weights}{When making venn diagrams, weight them?}

\item{gene_weights}{When venning genes, weight them?}
}
\value{
List containing some venns, lists, and such.

List containing some venns, lists, and such.
}
\description{
Najib is curious about the relationship of genes in sets, the sets, and the
genes that comprise those sets.  This is pushing gsva towards a oroborous-ish
state.

Najib is curious about the relationship of genes in sets, the sets, and the
genes that comprise those sets.  This is pushing gsva towards a oroborous-ish
state.
}
