% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{intersect_significant}
\alias{intersect_significant}
\title{Find the sets of intersecting significant genes}
\usage{
intersect_significant(combined, fc = 1, p = 0.05, z = NULL,
  p_type = "adj", excel = "excel/intersect_significant.xlsx")
}
\arguments{
\item{combined}{A result from combine_de_tables().}

\item{fc}{Define significant via fold-change.}

\item{p}{Or p-value.}

\item{z}{Or z-score.}

\item{p_type}{Use normal or adjusted p-values.}

\item{excel}{An optional excel workbook to which to write.}
}
\description{
Use extract_significant_genes() to find the points of agreement between limma/deseq/edger.
}
