% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{iterate_logistic_regression}
\alias{iterate_logistic_regression}
\title{Perform a series of single regression analyses and tabulate/plot the results.}
\usage{
iterate_logistic_regression(
  design,
  query = "condition",
  factors = NULL,
  family = "binomial",
  conf = 0.95,
  excel = NULL
)
}
\arguments{
\item{design}{Experimental design.}

\item{query}{Factor of primary interest.}

\item{factors}{Set of factors to query against (if not set, then query will be
the first design column, and these will be 2:end)}

\item{family}{model family to use.}

\item{conf}{Choose the confidence interval.}

\item{excel}{Write the results to this file.}
}
\description{
Perform a series of single regression analyses and tabulate/plot the results.
}
