% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{kegg_get_orgn}
\alias{kegg_get_orgn}
\title{Search the kegg identifier for a given species}
\usage{
kegg_get_orgn(species = "Leishmania", short = TRUE)
}
\arguments{
\item{species}{default='Leishmania'  A search string (Something like 'Homo sapiens')}

\item{short}{default=TRUE  only pull the orgid}
}
\value{
a data frame of possible KEGG identifier codes,
genome ID numbers, species, and phylogenetic classifications.
}
\description{
Search the kegg identifier for a given species
}
\examples{
## fun = kegg_get_orgn('Canis')
## >     Tid     orgid      species                   phylogeny
## >  17 T01007   cfa Canis familiaris (dog) Eukaryotes;Animals;Vertebrates;Mammals
}
\seealso{
\pkg{RCurl}
}

