% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.r
\name{kmeans_testing}
\alias{kmeans_testing}
\title{This is for the moment just a code dump of some arbitrarily chosen kmeans clustering stuff}
\usage{
kmeans_testing(gene_ids = get0("gene_ids"))
}
\arguments{
\item{gene_ids}{A set of gene IDs to query.}
}
\description{
Fill this in asap with real code for Ginger's search of gene sets which have similar profiles over time.
}
