% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{libsize,expt-method}
\alias{libsize,expt-method}
\title{A getter to pull the library sizes from an expt.}
\usage{
\S4method{libsize}{expt}(x)
}
\arguments{
\item{x}{An expt.}
}
\description{
A getter to pull the library sizes from an expt.
}
