% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{libsize<-,expt-method}
\alias{libsize<-,expt-method}
\title{Setter for library sizes in an expt.}
\usage{
\S4method{libsize}{expt}(x, ...) <- value
}
\arguments{
\item{x}{expt to add library sizes}

\item{...}{extra args}

\item{value}{new library sizes}
}
\description{
Setter for library sizes in an expt.
}
