% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{libsize<-,SummarizedExperiment-method}
\alias{libsize<-,SummarizedExperiment-method}
\title{Setter for library sizes in a se.}
\usage{
\S4method{libsize}{SummarizedExperiment}(x, ...) <- value
}
\arguments{
\item{x}{se to add library sizes}

\item{...}{extra args}

\item{value}{new library sizes}
}
\description{
Setter for library sizes in a se.
}
