% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{libsize<-}
\alias{libsize<-}
\title{Setter for library sizes in an expt.}
\usage{
libsize(x, ...) <- value
}
\arguments{
\item{x}{Starting data}

\item{...}{extra args.}

\item{value}{New library sizes}
}
\description{
Setter for library sizes in an expt.
}
