% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{libsize,SummarizedExperiment-method}
\alias{libsize,SummarizedExperiment-method}
\title{Get the library sizes of a summarized experiment.}
\usage{
\S4method{libsize}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{a summarized experiment.}
}
\description{
Get the library sizes of a summarized experiment.
}
