% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{limma_pairwise}
\alias{limma_pairwise}
\title{limma_pairwise()  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions using voom/limma.}
\usage{
limma_pairwise(input, conditions = NULL, batches = NULL,
  model_cond = TRUE, model_batch = FALSE, model_intercept = FALSE,
  extra_contrasts = NULL, alt_model = NULL, libsize = NULL)
}
\arguments{
\item{input}{a dataframe/vector or expt class containing count tables, normalization state, etc.}

\item{conditions}{default=NULL  a factor of conditions in the experiment}

\item{batches}{default=NULL  a factor of batches in the experiment}

\item{model_cond}{default=TRUE  include condition in the model?}

\item{model_batch}{default=FALSE  include batch in the model? This is hopefully TRUE.}

\item{model_intercept}{default=FALSE  perform a cell-means or intercept model?  A little more difficult for me to understand.  I have tested and get the same answer either way.}

\item{extra_contrasts}{default=NULL  some extra contrasts to add to the list
This can be pretty neat, lets say one has conditions A,B,C,D,E
and wants to do (C/B)/A and (E/D)/A or (E/D)/(C/B) then use this
with a string like: "c_minus_b_ctrla = (C-B)-A, e_minus_d_ctrla = (E-D)-A,
de_minus_cb = (E-D)-(C-B),"}

\item{alt_model}{default=NULL  a separate model matrix instead of the normal condition/batch.}

\item{libsize}{default=NULL  I've recently figured out that libsize is far more important than I previously realized.  Play with it here.}

\item{...}{The elipsis parameter is fed to write_limma() at the end.}
}
\value{
A list including the following information:
  macb = the mashing together of condition/batch so you can look at it
  macb_model = The result of calling model.matrix(~0 + macb)
  macb_fit =  The result of calling lmFit(data, macb_model)
  voom_result = The result from voom()
  voom_design = The design from voom (redundant from voom_result, but convenient)
  macb_table = A table of the number of times each condition/batch pairing happens
  cond_table = A table of the number of times each condition appears (the denominator for the identities)
  batch_table = How many times each batch appears
  identities = The list of strings defining each condition by itself
  all_pairwise = The list of strings defining all the pairwise contrasts
  contrast_string = The string making up the makeContrasts() call
  pairwise_fits = The result from calling contrasts.fit()
  pairwise_comparisons = The result from eBayes()
  limma_result = The result from calling write_limma()
}
\description{
limma_pairwise()  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions using voom/limma.
}
\examples{
## pretend = balanced_pairwise(data, conditions, batches)
}
\seealso{
\code{\link{write_limma}}
}

