% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_limma.r
\name{limma_subset}
\alias{limma_subset}
\title{A quick and dirty way to pull the top/bottom genes from toptable().}
\usage{
limma_subset(table, n = NULL, z = NULL)
}
\arguments{
\item{table}{Original data from limma.}

\item{n}{Number of genes to keep.}

\item{z}{Number of z-scores from the mean.}
}
\value{
Dataframe subset from toptable.
}
\description{
If neither n nor z is provided, it assumes you want 1.5 z-scores from the median.
}
\examples{
\dontrun{
 subset = limma_subset(df, n=400)
 subset = limma_subset(df, z=1.5)
}
}
\seealso{
\pkg{limma}
}

