% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{lm_high_utr}
\alias{lm_high_utr}
\title{The UTR regions of every highly translated L.major gene.}
\format{
An object of class \code{character} of length 1.
}
\usage{
lm_high_utr
}
\description{
Once upon a time I performed a ribosome profiling experiment in Leishmania
major.  Sadly, we still have not published it.  This file contains the UTRs
of every highly translated gene in procyclic promastigotes.  I used it in
some motif analyses for fun.
}
\keyword{datasets}
