% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{load_annotations}
\alias{load_annotations}
\title{Load organism annotation data (parasite).}
\usage{
load_annotations(orgdb, gene_ids = NULL, include_go = FALSE,
  keytype = "ENSEMBL", fields = NULL, sum_exons = FALSE)
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Gene identifiers for retrieving annotations.}

\item{include_go}{Ask the Dbi for gene ontology information?}

\item{keytype}{mmm the key type used?}

\item{fields}{Columns included in the output.}

\item{sum_exons}{Perform a sum of the exons in the data set?}
}
\value{
Table of geneids, chromosomes, descriptions, strands, types, and lengths.
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the OrganismDbi interface.
}
\details{
Tested in test_45ann_organdb.R
This defaults to a few fields which I have found most useful, but the brave can pass it 'all'.
}
\examples{
\dontrun{
 one_gene <- load_annotations(org, c("LmJF.01.0010"))
}
}
\seealso{
\pkg{AnnotationDbi} \pkg{GenomicFeatures} \pkg{BiocGenerics}
 \code{\link[AnnotationDbi]{columns}} \code{\link[AnnotationDbi]{keytypes}}
 \code{\link[AnnotationDbi]{select}} \code{\link[GenomicFeatures]{exonsBy}}
}

