% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{load_biomart_annotations}
\alias{load_biomart_annotations}
\title{Extract annotation information from biomart.}
\usage{
load_biomart_annotations(
  species = "hsapiens",
  overwrite = FALSE,
  do_save = TRUE,
  host = NULL,
  trymart = "ENSEMBL_MART_ENSEMBL",
  archive = FALSE,
  default_hosts = c("useast.ensembl.org", "uswest.ensembl.org", "www.ensembl.org",
    "asia.ensembl.org"),
  year = NULL,
  month = NULL,
  drop_haplotypes = FALSE,
  trydataset = NULL,
  gene_requests = c("ensembl_gene_id", "version", "ensembl_transcript_id",
    "transcript_version", "description", "gene_biotype"),
  length_requests = c("ensembl_transcript_id", "cds_length", "chromosome_name", "strand",
    "start_position", "end_position"),
  gene_tx_map = TRUE,
  gene_id_column = "ensembl_gene_id",
  gene_version_column = "version",
  tx_id_column = "ensembl_transcript_id",
  tx_version_column = "transcript_version",
  symbol_columns = NULL,
  include_lengths = TRUE,
  do_load = TRUE,
  savefile = NULL
)
}
\arguments{
\item{species}{Choose a species.}

\item{overwrite}{Overwite an existing save file?}

\item{do_save}{Create a savefile of annotations for future runs?}

\item{host}{Ensembl hostname to use.}

\item{trymart}{Biomart has become a circular dependency, this makes me sad,
now to list the marts, you need to have a mart loaded.}

\item{archive}{Try an archive server instead of a mirror?  If this is a
character, it will assume it is a specific archive hostname.}

\item{default_hosts}{List of biomart mirrors to try.}

\item{year}{Choose specific year(s) for the archive servers?}

\item{month}{Choose specific month(s) for the archive server?}

\item{drop_haplotypes}{Some chromosomes have stupid names because they are
from non-standard haplotypes and they should go away.  Setting this to
false stops that.}

\item{trydataset}{Choose the biomart dataset from which to query.}

\item{gene_requests}{Set of columns to query for description-ish annotations.}

\item{length_requests}{Set of columns to query for location-ish annotations.}

\item{gene_tx_map}{Provide a gene2tx map for things like salmon (perhaps rename this to tx_gene_map?)}

\item{gene_id_column}{Column containing the gene ID.}

\item{gene_version_column}{Column containing the ensembl gene version.}

\item{tx_id_column}{Column containing the transcript ID.}

\item{tx_version_column}{Columns containing the ensembl transcript version.}

\item{symbol_columns}{Vector of columns containing the gene symbols.}

\item{include_lengths}{Also perform a search on structural elements in the genome?}

\item{do_load}{Load the data?}

\item{savefile}{Use this savefile.}
}
\value{
List containing: a data frame of the found annotations, a copy of
 The mart instance to help with finding problems, the hostname queried, the
 name of the mart queried, a vector of rows queried, vector of the available
 attributes, and the ensembl dataset queried.
}
\description{
Biomart is an amazing resource of information, but using it is a bit
annoying.  This function hopes to alleviate some common headaches.
}
\details{
Tested in test_40ann_biomart.R
This goes to some lengths to find the relevant tables in biomart.  But
biomart is incredibly complex and one should carefully inspect the output if
it fails to see if there are more appropriate marts, datasets, and columns to
download.
}
\examples{
## Provide examples of the biomart annotation functions.

## Search for a biomart instance which will accept queries.
archive_mart <- try(find_working_mart(archive = TRUE, year = 2023, month = 02))
class(archive_mart)
archive_mart

## I should probably change this function to accept the mart from
## find_working_mart()
hs_annot <- try(load_biomart_annotations(year = 2023, month = 02))
class(hs_annot)

## I added a S3 print method to try to avoid being made sad if I
## print the result from load_annotations.
hs_annot

## If one wishes to check out all of the possible data types for a
## single gene, try this:
## However, this function is painfully slow and likely annoying to the
## ensembl servers.
## single_gene <- get_biomart_example_gene()

## Grab gene ontology data for a species:
hs_go <- try(load_biomart_go(year = 2023, month = 02))
hs_go

## The following example is fragile I think because multiple
## ensembl instances must be responding properly to work
## If it fails; check that the examples for biomaRt::getLDS are
## functioning -- most likely they are failing due to an upstream
## error.
hs_mm_ortho <- try(load_biomart_orthologs(gene_ids = "ENSMUSG00000000001",
                                          year = 2023, month = 02,
                                          first_species = "mmusculus",
                                          second_species = "scerevisiae"))

## Create salmon-like ID.version strings using ensembl annotations.
tx_gene_map <- try(make_tx_gene_map(as.data.frame(hs_annot)))
head(tx_gene_map)
}
\seealso{
[biomaRt::listDatasets()] [biomaRt::getBM()] [find_working_mart()]
}
