% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{load_biomart_go}
\alias{load_biomart_go}
\title{Extract gene ontology information from biomart.}
\usage{
load_biomart_go(
  species = "hsapiens",
  overwrite = FALSE,
  do_save = TRUE,
  host = NULL,
  trymart = "ENSEMBL_MART_ENSEMBL",
  archive = FALSE,
  default_hosts = c("useast.ensembl.org", "uswest.ensembl.org", "www.ensembl.org",
    "asia.ensembl.org"),
  year = NULL,
  month = NULL,
  trydataset = NULL,
  dl_rows = c("ensembl_gene_id", "go_id"),
  dl_rowsv2 = c("ensembl_gene_id", "go_id")
)
}
\arguments{
\item{species}{Species to query.}

\item{overwrite}{Overwrite existing savefile?}

\item{do_save}{Create a savefile of the annotations? (if not false, then a filename.)}

\item{host}{Ensembl hostname to use.}

\item{trymart}{Biomart has become a circular dependency, this makes me sad,
now to list the marts, you need to have a mart loaded.}

\item{archive}{Try an archive server instead of a mirror?  If this is a
character, it will assume it is a specific archive hostname.}

\item{default_hosts}{List of biomart mirrors to try.}

\item{year}{Choose specific year(s) for the archive servers?}

\item{month}{Choose specific month(s) for the archive servers?}

\item{trydataset}{Define a dataset to which to attempt connecting.}

\item{dl_rows}{List of rows from the final biomart object to download.}

\item{dl_rowsv2}{A second list of potential rows.}
}
\value{
List containing the following:  data frame of ontology data, a copy
 of the biomart instance for further querying, the host queried, the biomart
 queried, a vector providing the attributes queried, and the ensembl dataset
 queried.
}
\description{
I perceive that every time I go to acquire annotation data from biomart, they
have changed something important and made it more difficult for me to find
what I want. I recently found the *.archive.ensembl.org, and so this function
uses that to try to keep things predictable, if not consistent.
}
\details{
This function makes a couple of attempts to pick up the correct tables from
biomart.  It is worth noting that it uses the archive.ensembl host(s) because
of changes in table organization after December 2015 as well as an attempt to
keep the annotation sets relatively consistent.
}
\examples{
## Provide examples of the biomart annotation functions.

## Search for a biomart instance which will accept queries.
archive_mart <- try(find_working_mart(archive = TRUE, year = 2023, month = 02))
class(archive_mart)
archive_mart

## I should probably change this function to accept the mart from
## find_working_mart()
hs_annot <- try(load_biomart_annotations(year = 2023, month = 02))
class(hs_annot)

## I added a S3 print method to try to avoid being made sad if I
## print the result from load_annotations.
hs_annot

## If one wishes to check out all of the possible data types for a
## single gene, try this:
## However, this function is painfully slow and likely annoying to the
## ensembl servers.
## single_gene <- get_biomart_example_gene()

## Grab gene ontology data for a species:
hs_go <- try(load_biomart_go(year = 2023, month = 02))
hs_go

## The following example is fragile I think because multiple
## ensembl instances must be responding properly to work
## If it fails; check that the examples for biomaRt::getLDS are
## functioning -- most likely they are failing due to an upstream
## error.
hs_mm_ortho <- try(load_biomart_orthologs(gene_ids = "ENSMUSG00000000001",
                                          year = 2023, month = 02,
                                          first_species = "mmusculus",
                                          second_species = "scerevisiae"))

## Create salmon-like ID.version strings using ensembl annotations.
tx_gene_map <- try(make_tx_gene_map(as.data.frame(hs_annot)))
head(tx_gene_map)
}
\seealso{
[biomaRt::listMarts()] [biomaRt::useDatasets()] [biomaRt::getBM()]
}
