% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{load_biomart_orthologs}
\alias{load_biomart_orthologs}
\title{Use biomart to get orthologs between supported species.}
\usage{
load_biomart_orthologs(
  gene_ids = NULL,
  first_species = "hsapiens",
  second_species = "mmusculus",
  host = NULL,
  trymart = "ENSEMBL_MART_ENSEMBL",
  archive = TRUE,
  default_hosts = c("useast.ensembl.org", "uswest.ensembl.org", "www.ensembl.org",
    "asia.ensembl.org"),
  year = NULL,
  month = NULL,
  trydataset = NULL,
  attributes = "ensembl_gene_id"
)
}
\arguments{
\item{gene_ids}{List of gene IDs to translate.}

\item{first_species}{Linnean species name for one species.}

\item{second_species}{Linnean species name for the second species.}

\item{host}{Ensembl server to query.}

\item{trymart}{Assumed mart name to use.}

\item{archive}{Use an archive server?}

\item{default_hosts}{Set of default hosts to query.}

\item{year}{When using an archive server, use this year (otherwise it will choose last year).}

\item{month}{When using an archive server, use this month (otherwise, this month).}

\item{trydataset}{Choose a dataset to query.}

\item{attributes}{Key to query}
}
\value{
list of 4 elements:  The first is the set of all ids, as getLDS seems
 to always send them all; the second is the subset corresponding to the
 actual ids of interest, and the 3rd/4th are other, optional ids from other datasets.
}
\description{
Biomart's function getLDS is incredibly powerful, but it makes me think very
polite people are going to start knocking on my door, and it fails weirdly
pretty much always. This function attempts to alleviate some of that frustration.
}
\details{
Tested in test_40ann_biomart.R
As with my other biomart functions, this one grew out of frustrations when
attempting to work with the incredibly unforgiving biomart service.  It does
not attempt to guarantee a useful biomart connection, but will hopefully
point out potentially correct marts and attributes to use for a successful
query.  I can say with confidence that it works well between mice and
humans.
}
\examples{
## Provide examples of the biomart annotation functions.

## Search for a biomart instance which will accept queries.
archive_mart <- try(find_working_mart(archive = TRUE, year = 2023, month = 02))
class(archive_mart)
archive_mart

## I should probably change this function to accept the mart from
## find_working_mart()
hs_annot <- try(load_biomart_annotations(year = 2023, month = 02))
class(hs_annot)

## I added a S3 print method to try to avoid being made sad if I
## print the result from load_annotations.
hs_annot

## If one wishes to check out all of the possible data types for a
## single gene, try this:
## However, this function is painfully slow and likely annoying to the
## ensembl servers.
## single_gene <- get_biomart_example_gene()

## Grab gene ontology data for a species:
hs_go <- try(load_biomart_go(year = 2023, month = 02))
hs_go

## The following example is fragile I think because multiple
## ensembl instances must be responding properly to work
## If it fails; check that the examples for biomaRt::getLDS are
## functioning -- most likely they are failing due to an upstream
## error.
hs_mm_ortho <- try(load_biomart_orthologs(gene_ids = "ENSMUSG00000000001",
                                          year = 2023, month = 02,
                                          first_species = "mmusculus",
                                          second_species = "scerevisiae"))

## Create salmon-like ID.version strings using ensembl annotations.
tx_gene_map <- try(make_tx_gene_map(as.data.frame(hs_annot)))
head(tx_gene_map)
}
\seealso{
[biomaRt::getLDS()]
}
