% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_sql.R
\name{load_biomart_sql_annotations}
\alias{load_biomart_sql_annotations}
\title{Load biomart data from a MariaDB copy.}
\usage{
load_biomart_sql_annotations(
  species = "hsapiens",
  overwrite = FALSE,
  do_save = TRUE,
  db_name = "ensembl_mart",
  db_host = "localhost",
  db_user = "anonymous",
  db_port = 3306,
  db_type = "MariaDB",
  db_password = NULL,
  wanted = NULL,
  savefile = NULL
)
}
\arguments{
\item{species}{The tables are prefixed by species, so pick one.}

\item{overwrite}{This can write out a local rda file of the
results, overwrite it?}

\item{do_save}{Save the output to a local rda file?}

\item{db_name}{Name of the local database.}

\item{db_user}{User at the database.}

\item{db_port}{This assumes MariaDB's 3306, I was thinking to use
postgres but did not know if their schemas had mysql-specific
stuff.  I may try messing with this in the future.}

\item{db_type}{Type of DBI connection to create.}

\item{db_password}{If you want to mess with the data, change the
user and this.}

\item{wanted}{List of tables and columns of interest.}

\item{savefile}{Filename to which to save the results.}

\item{dh_host}{Where to send queries.}
}
\value{
List with the same organization as 'wanted' containing
 dataframes with the columns specified in 'wanted'
}
\description{
Given some tables and columns, select out all rows in the
appropriate biomart tables.  I use this to create species
annotation tables for RNASeq and similar tasks; as a result I
pretty much only ever want everything.  Once I need subsets I will
fill in the 'WHERE' clauses.
}
