% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_genbank.R
\name{load_genbank_annotations}
\alias{load_genbank_annotations}
\title{Given a genbank accession, make a txDb object along with sequences, etc.}
\usage{
load_genbank_annotations(
  accession = "AE009949",
  database = "nuccore",
  return_type = "gbwithparts",
  save_gb = TRUE,
  restez_db = "/sw/local/genbank/current"
)
}
\arguments{
\item{accession}{Accession to download and import.}

\item{database}{Entrez Database to query}

\item{return_type}{Type of file to download from ncbi.}

\item{save_gb}{Save a copy of the genbank file locally.}

\item{restez_db}{Location of local genbank used by restez.}
}
\value{
List containing a txDb, sequences, and some other stuff which I
 haven't yet finalized.
}
\description{
Let us admit it, sometimes biomart is a pain.  It also does not have easily
accessible data for microbes.  Genbank does!
}
\details{
This switched to using restez, which does not provide the helpful
data structures that genbankr did.  I am not sure if I can still
create orgdb/txdb entries therefore unless I yank that code out of
an old version of genbankr, or if I make use of txdbmaker etc.
}
\examples{
## 202601 eutils seems to be having some trouble.

test_accession <- "AE009948"
gbk_file <- try(download_gbk(accessions = test_accession))
if (! "try-error" \%in\% class(gbk_file)) {
  removed <- file.remove(gbk_file[["written_file"]])
}

sagalacticae_genbank_annot <- load_genbank_annotations(accession = test_accession)
if (! "try-error" \%in\% class(sagalacticae_genbank_annot)) {
  dim(sagalacticae_genbank_annot)
}
}
\seealso{
[Biostrings] [GenomicFeatures] [genbankr::import()] [genbankr::readGenBank()]
}
