% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020_annotation_gff.r
\name{load_gff_annotations}
\alias{load_gff_annotations}
\title{Extract annotation information from a gff file into a df}
\usage{
load_gff_annotations(
  gff,
  type = NULL,
  id_col = "ID",
  ret_type = "data.frame",
  second_id_col = "locus_tag",
  try = NULL,
  row.names = NULL
)
}
\arguments{
\item{gff}{Gff filename.}

\item{type}{Subset the gff file for entries of a specific type.}

\item{id_col}{Column in a successful import containing the IDs of interest.}

\item{ret_type}{Return a data.frame or something else?}

\item{second_id_col}{Second column to check.}

\item{try}{Give your own function call to use for importing.}

\item{row.names}{Choose another column for setting the rownames of the data frame.}
}
\value{
Dataframe of the annotation information found in the gff file.
}
\description{
Try to make import.gff a little more robust; I acquire (hopefully) valid gff
files from various sources: yeastgenome.org, microbesonline, tritrypdb, ucsc,
ncbi. To my eyes, they all look like reasonably good gff3 files, but some of
them must be loaded with import.gff2, import.gff3, etc. That is super
annoying. Also, I pretty much always just do as.data.frame() when I get
something valid from rtracklayer, so this does that for me, I have another
function which returns the iranges etc.  This function wraps
import.gff/import.gff3/import.gff2 calls in try() because sometimes those
functions fail in unpredictable ways.
}
\examples{
 gas_gff_annot <- load_gff_annotations(gff_file)
 dim(gas_gff_annot)
}
\seealso{
\pkg{rtracklayer} \pkg{GenomicRanges}
 \code{\link[rtracklayer]{import.gff}}
}
