% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_gff.R
\name{load_gff_go}
\alias{load_gff_go}
\title{Take the df from load_gff_annotations and pull the GO information from it.}
\usage{
load_gff_go(df, go_column = "GO", id_column = "row.names", separator = ",")
}
\arguments{
\item{df}{Result from load_gff_annotations()}

\item{go_column}{Column containing the GO list}

\item{id_column}{Column containing the gene IDs}

\item{separator}{Character separating the GO categories.}
}
\description{
Take the df from load_gff_annotations and pull the GO information from it.
}
\seealso{
[load_gff_annotations()]
}
