% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{load_gmt_signatures}
\alias{load_gmt_signatures}
\title{Extract particular gene sets from either gmt files or GSVA data.}
\usage{
load_gmt_signatures(
  signatures = "c2BroadSets",
  data_pkg = "GSVAdata",
  signature_category = "c2"
)
}
\arguments{
\item{signatures}{Potentially a filename or variable name.  If it is a
filename, it should be a locally downloaded gmt file, otherwise it should be
a variable from 'data_pkg'.}

\item{data_pkg}{Most likely GSVAdata, but not necessarily.}

\item{signature_category}{Choose a specific category from the
gmt/xml/database.}
}
\value{
gene set collection!
}
\description{
This could be extended to pull in the XML files from MSigDB as well.  In its
current form it seeks to fill in the likely corners when looking to extract
the signatures from either GSVA or manually downloaded gmt files from the
MSigDB.
}
