% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{load_go_terms}
\alias{load_go_terms}
\title{Retrieve GO terms associated with a set of genes.}
\usage{
load_go_terms(orgdb, gene_ids, keytype = "ENSEMBL")
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Identifiers of the genes to retrieve annotations.}

\item{keytype}{the mysterious keytype returns yet again to haunt my dreams}
}
\value{
Data frame of gene IDs, go terms, and names.
}
\description{
AnnotationDbi provides a reasonably complete set of GO mappings between gene ID and
ontologies.  This will extract that table for a given set of gene IDs.
}
\details{
Tested in test_45ann_organdb.R
This is a nice way to extract GO data primarily because the Orgdb data sets are extremely fast
and flexible, thus by changing the keytype argument, one may use a lot of different ID types
and still score some useful ontology data.
}
\examples{
\dontrun{
 go_terms <- load_go_terms(org, c("a","b"))
}
}
\seealso{
\pkg{AnnotationDbi} \pkg{GO.db} \pkg{magrittr}
}

