% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{load_host_annotations}
\alias{load_host_annotations}
\title{Load organism annotation data (mouse/human).}
\usage{
load_host_annotations(orgdb = NULL, gene_ids = NULL, keytype = "ensembl",
  chromosome_column = "txchrom", strand_column = "txstrand",
  start_column = "txstart", end_column = "txend",
  description_column = "genename", fields = c("geneid"),
  biomart_dataset = NULL)
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Gene identifiers for retrieving annotations.}

\item{keytype}{a, umm keytype? I need to properly read this code.}

\item{chromosome_column}{Which column contains the chromosome information?}

\item{strand_column}{Ibid' but for the gene's strand.}

\item{start_column}{Ibid' but to find where each gene starts.}

\item{end_column}{Ibid' but to find where each gene ends.}

\item{description_column}{Ibid' but to find gene descriptions.}

\item{fields}{Columns to include in the output.}

\item{biomart_dataset}{Name of the biomaRt dataset to query for gene type.}
}
\value{
a table of gene information
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the OrganismDbi interface.
}
\examples{
\dontrun{
 host <- load_host_annotations(org, c("a","b"))
}
}
\seealso{
\pkg{AnnotationDbi} \pkg{dplyr} \pkg{biomaRt}
 \code{\link[AnnotationDbi]{select}} \code{\link[AnnotationDbi]{keytypes}}
}
