% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_kegg.R
\name{load_kegg_annotations}
\alias{load_kegg_annotations}
\title{Create a data frame of pathways to gene IDs from KEGGREST}
\usage{
load_kegg_annotations(species = "coli", abbreviation = NULL, flatten = TRUE)
}
\arguments{
\item{species}{String to use to query KEGG abbreviation.}

\item{abbreviation}{If you already know the abbreviation, use it.}

\item{flatten}{Flatten nested tables?}
}
\value{
dataframe with rows of KEGG gene IDs and columns of NCBI gene IDs
 and KEGG paths.
}
\description{
This seeks to take the peculiar format from KEGGREST for pathway<->genes and
make it easier to deal with.  Sadly, this only works for a subset of species now.
}
\examples{
library(hpgltools)
## This function takes forever, let us only run it once in a while
## Perhaps I should have it create a savefile
if (interactive()) {
  pathfindr_kegg <- make_kegg_df("hsa")
  ## Load a pile of KEGGREST annotations.
  coli_kegg <- load_kegg_annotations()
  head(coli_kegg)
}
}
\seealso{
[KEGGREST]
}
