% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{load_kegg_mapping}
\alias{load_kegg_mapping}
\title{Creates a gene/KEGG mapping dataframe}
\usage{
load_kegg_mapping(orgdb, gene_ids, keytype = "ENSEMBL")
}
\arguments{
\item{orgdb}{An OrganismDb instance}

\item{gene_ids}{Identifiers of the genes to retrieve annotations for}
}
\description{
Creates a gene/KEGG mapping dataframe
}

