% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{load_kegg_pathways}
\alias{load_kegg_pathways}
\title{Creates a KEGG pathway/description mapping dataframe}
\usage{
load_kegg_pathways(orgdb, gene_ids, keytype = "ENSEMBL")
}
\arguments{
\item{orgdb}{An OrganismDb instance}

\item{gene_ids}{Identifiers of the genes to retrieve annotations for}
}
\description{
Creates a KEGG pathway/description mapping dataframe
}

