% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.R
\name{load_microbesonline_annotations}
\alias{load_microbesonline_annotations}
\title{Skip the db and download all the text annotations for a given species.}
\usage{
load_microbesonline_annotations(species = NULL, id = NULL)
}
\arguments{
\item{species}{Microbesonline species.}

\item{id}{Microbesonline ID to query.}
}
\value{
Dataframe containing the annotation information.
}
\description{
The microbesonline publicly available mysqldb is rather more complex than I
prefer.  This skips that process and just grabs a tsv copy of everything and
loads it into a dataframe.  I have not yet figured out how to so-easily query
microbesonline for species IDs, thus one will have to manually query the
database to find species of interest.
}
\details{
Tested in test_70expt_spyogenes.R
There is so much awesome information in microbesonline, but damn is it
annoying to download. This function makes that rather easier, or so I hope at
least.
}
\examples{
## On those relatively rare occasions when the internet is being
## cranky, these tests will most certainly fail.

## Get the numeric ID from microbesonline for a species of interest.
## When microbesonline is under maintenance/etc this fails.
ecoli_id <- try(get_microbesonline_taxid(species = "S88"))
ecoli_id

if (! "try-error" \%in\% class(ecoli_id)) {
  ## Download a genbank/fasta/etc file for that species.
  ecoli_files <- try(download_microbesonline_files(id = ecoli_id, type = "gbk"))
  ecoli_files
  if (! "try-error" \%in\% class(ecoli_files)) {
    removed <- file.remove(ecoli_files[["gbk"]])
  }

  ## Download a species annotations to a dataframe
  ecoli_annot <- try(load_microbesonline_annotations(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_annot)) {
    dim(ecoli_annot)
    colnames(ecoli_annot)
  }

  ## Get gene ontology data from microbesonline
  ecoli_go <- try(load_microbesonline_go(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_go)) {
    dim(ecoli_go)
  }
}
}
\seealso{
[rvest] [xml2] [readr]
}
