% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.R
\name{load_microbesonline_go}
\alias{load_microbesonline_go}
\title{Extract the set of GO categories by microbesonline locus}
\usage{
load_microbesonline_go(
  id = NULL,
  species = NULL,
  table_df = NULL,
  id_column = "name",
  data_column = "GO",
  name = NULL
)
}
\arguments{
\item{id}{Which species to query.}

\item{species}{Microbesonline species.}

\item{table_df}{Pre-existing data frame of annotations containing GO stuff.}

\item{id_column}{This no longer uses MySQL, so which column from the html
table to pull?}

\item{data_column}{Similar to above, there are lots of places from which one might
extract the data.}

\item{name}{Allowing for non-specific searches by species name.}
}
\value{
data frame of GO terms from www.microbesonline.org
}
\description{
The microbesonline is such a fantastic resource, it is a bit of a shame that
it is such a pain to query.
}
\details{
Tested in test_42ann_microbes.R
I am not 100% certain that this is giving me the full correct set of gene
ontology accessions. At the very least, it does return a large number of
them, which is a start.
}
\examples{
## On those relatively rare occasions when the internet is being
## cranky, these tests will most certainly fail.

## Get the numeric ID from microbesonline for a species of interest.
## When microbesonline is under maintenance/etc this fails.
ecoli_id <- try(get_microbesonline_taxid(species = "S88"))
ecoli_id

if (! "try-error" \%in\% class(ecoli_id)) {
  ## Download a genbank/fasta/etc file for that species.
  ecoli_files <- try(download_microbesonline_files(id = ecoli_id, type = "gbk"))
  ecoli_files
  if (! "try-error" \%in\% class(ecoli_files)) {
    removed <- file.remove(ecoli_files[["gbk"]])
  }

  ## Download a species annotations to a dataframe
  ecoli_annot <- try(load_microbesonline_annotations(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_annot)) {
    dim(ecoli_annot)
    colnames(ecoli_annot)
  }

  ## Get gene ontology data from microbesonline
  ecoli_go <- try(load_microbesonline_go(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_go)) {
    dim(ecoli_go)
  }
}
}
\seealso{
[tidyr]
}
