% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{load_msig_metadata}
\alias{load_msig_metadata}
\title{Extract the fun stuff from the mSigDB.}
\usage{
load_msig_metadata(db = "reference/msigdb_v2024.1.Hs.db")
}
\arguments{
\item{db}{Database filename}
}
\description{
The new mSigDB releases provide the data as a sqlite file, use this
to find fun information from them.
}
