% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.R
\name{load_orgdb_go}
\alias{load_orgdb_go}
\title{Retrieve GO terms associated with a set of genes.}
\usage{
load_orgdb_go(
  orgdb = NULL,
  gene_ids = NULL,
  keytype = "ensembl",
  columns = c("go", "goall", "goid"),
  guess_columns = FALSE
)
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Identifiers of the genes to retrieve annotations.}

\item{keytype}{The mysterious keytype returns yet again to haunt my dreams.}

\item{columns}{The set of columns to request.}

\item{guess_columns}{Instead of a set of specific columns, use grep to find anything with 'go'}
}
\value{
Data frame of gene IDs, go terms, and names.
}
\description{
AnnotationDbi provides a reasonably complete set of GO mappings between gene
ID and ontologies.  This will extract that table for a given set of gene
IDs.
}
\details{
This is a nice way to extract GO data primarily because the Orgdb data sets
are extremely fast and flexible, thus by changing the keytype argument, one
may use a lot of different ID types and still score some useful ontology data.
}
\examples{
hs_annot <- load_orgdb_annotations("org.Hs.eg.db")
dim(hs_annot[["genes"]])

## I changed this fairly significantly to simplify the select statements.
hs_go <- load_orgdb_go("org.Mm.eg.db")
summary(hs_go)

## extract_eupath_orthologs should be moved to EuPathDB

## This is similar to clusterProfiler's bitr I think.
dm_symbols <- map_orgdb_ids("org.Dm.eg.db", keytype = "entrezid", mapto = "symbol")
head(dm_symbols)

## Guess the most appropriate keytype given some IDs
## Essentially: what column has this?
wanted_ids <- c("G9a", "cin", "ewg")
keytype_guess <- guess_orgdb_keytype(ids = wanted_ids, "org.Dm.eg.db")
keytype_guess

## Extract The actual data of interest from AnnotationHub
if (interactive()) {
  wanted <- orgdb_from_ah(type = "OrgDb", title = "Gorilla")
  class(wanted)
}
}
\seealso{
[AnnotationDbi] [GO.db]
}
\author{
I think Keith provided the initial implementation of this, but atb
 messed with it pretty extensively.
}
