% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{load_orgdb_kegg}
\alias{load_orgdb_kegg}
\title{Creates a gene/KEGG mapping dataframe.}
\usage{
load_orgdb_kegg(orgdb, gene_ids = NULL, keytype = "ENSEMBL",
  columns = c("KEGG_PATH"))
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Identifiers of the genes to retrieve annotations.}

\item{keytype}{The keytype, eg. the primary key used to query the orgdb.}

\item{columns}{Columns to extract.}
}
\value{
Df of kegg mappings
}
\description{
In much the same way AnnotationDbi provides GO data, it also provides KEGG data.
}
\details{
Tested in test_45ann_organdb.R
Perhaps this function should be merged with the GO above?
}
\examples{
\dontrun{
 kegg_data <- load_kegg_mapping(org, c("a","b"))
}
}
\seealso{
\pkg{AnnotationDbi} \pkg{dplyr}
 \code{\link[AnnotationDbi]{select}} \code{\link[dplyr]{tbl_df}}
}
