% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdb.r
\name{load_parasite_annotations}
\alias{load_parasite_annotations}
\title{Load organism annotation data (parasite).}
\usage{
load_parasite_annotations(orgdb, gene_ids = NULL, keytype = "ENSEMBL",
  fields = NULL, sum_exons = FALSE)
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Gene identifiers for retrieving annotations.}

\item{keytype}{mmm the key type used?}

\item{fields}{Columns included in the output.}

\item{sum_exons}{Perform a sum of the exons in the data set?}
}
\value{
Table of geneids, chromosomes, descriptions, strands, types, and lengths.
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the OrganismDbi interface.
}
\examples{
\dontrun{
one_gene <- load_parasite_annotations(org, c("LmJF.01.0010"))
}
}
\seealso{
\link[AnnotationDbi]{select}
}

