% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_mygene.r
\name{load_querymany_annotations}
\alias{load_querymany_annotations}
\title{Use mygene's queryMany to translate gene ID types}
\usage{
load_querymany_annotations(queries, email = "abelew@gmail.com",
  fields = "all", species = "all")
}
\arguments{
\item{queries}{Gene IDs to translate.}

\item{email}{E-mail address to send with the querymany query.}

\item{fields}{Set of fields to request, pass null for all.}

\item{species}{Human readable species for translation (Eg. 'human' instead of 'hsapiens'.)}
}
\value{
Df of translated IDs/accessions
}
\description{
Juggling between entrez, ensembl, etc can be quite a hassel.  This hopes to make it easier.
}
\details{
Tested in test_40ann_biomart.R
This function really just sets a couple of hopefully helpful defaults.  When I first attempted
to use queryMany, it seemed to need much more intervention than it does now.  But at the least
this function should provide a reminder of this relatively fast and useful ID translation service.
}
\examples{
\dontrun{
 data <- translate_ids_querymany(genes)
}
}
\seealso{
\pkg{mygene}
 \code{\link[mygene]{queryMany}}
}
