% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grange.R
\name{load_se_tracks}
\alias{load_se_tracks}
\title{Stealing the LoadTrackFile function from ggcoverage and making it more helpful.}
\usage{
load_se_tracks(
  se,
  track_column = "deeptools_coverage",
  region_string = NULL,
  name_column = "Parent",
  gene_name = NULL,
  padding = 100,
  bin_size = 10,
  transform = NULL,
  convert = "cpm",
  norm = NULL,
  type_column = "media",
  group_column = "media_type",
  cores = NULL,
  extend = NULL
)
}
\arguments{
\item{se}{Input summarized experiment.}

\item{track_column}{Column name with track files.}
}
\description{
Stealing the LoadTrackFile function from ggcoverage and making it more helpful.
}
