% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_txt.R
\name{load_trinotate_annotations}
\alias{load_trinotate_annotations}
\title{Read a csv file from trinotate and make an annotation data frame.}
\usage{
load_trinotate_annotations(
  trinotate = "reference/trinotate.csv",
  collapse = FALSE
)
}
\arguments{
\item{trinotate}{CSV of trinotate annotation data.}

\item{collapse}{Collapse isoforms to genes?}
}
\value{
Dataframe of fun data.
}
\description{
Trinotate performs some neat sequence searches in order to seek out likely
annotations for the trinity contigs.  The resulting csv file is encoded in a
peculiar fashion, so this function attempts to make it easier to read and put
them into a format usable in an expressionset.
}
\examples{
library(hpgltools)
example_txt <- system.file("share", "sb", "trinotate_head.csv.xz", package = "hpgldata")

## Create a dataframe from the peculiarly formatted tsv/csv from trinotate.
trinotate_df <- load_trinotate_annotations(example_txt)
head(trinotate_df)

## Extract the gene ontology information from trinotate.
trinotate_go <- load_trinotate_go(example_txt,
                                  blast2go_column = "gene_ontology_blast")
head(trinotate_go[["go_table"]])
}
\seealso{
[tidyr] [readr()]
}
