% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_uniprot.R
\name{load_uniprot_go}
\alias{load_uniprot_go}
\title{Extract ontology information from a uniprot dataframe.}
\usage{
load_uniprot_go(...)
}
\arguments{
\item{...}{Whatever args are required for load_uniprot_annotations()}
}
\value{
Ontology dataframe
}
\description{
Extract ontology information from a uniprot dataframe.
}
\examples{
uniprot_sc <- load_uniprot_annotations(species = "Saccharomyces cerevisiae S288c")
head(uniprot_sc)

uniprot_go <- load_uniprot_go(species = "Saccharomyces cerevisiae S288c")
head(uniprot_go)
}
\seealso{
[load_uniprot_annotations()] [stringr] [tidyr]
}
