% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_uniprot.r
\name{load_uniprotws_annotations}
\alias{load_uniprotws_annotations}
\title{Extract annotation data from the uniprot webservices.}
\usage{
load_uniprotws_annotations(id = NULL,
  species = "Mycobacterium tuberculosis", keytype = "GI_NUMBER*",
  chosen_columns = NULL)
}
\arguments{
\item{id}{Species ID, if not provided, then this will try to find it using
the species}

\item{species}{Assuming no ID, use this to find one.}

\item{keytype}{The primary keytype when doing the final select statement.}

\item{chosen_columns}{What columns are desired from the webservices data?
If not provided, this will attempt to choose useful ones.}
}
\value{
Data frame from selecting the hopefully appropriate columns with
AnnotationDbi.
}
\description{
I keep thinking that this is in fact querying NCBI, but I think that is incorrect.
This is because all of the examples are using ENTREZ_GENE as the primary key
I bet.  In any event, this function seeks to simplify getting useful
annotation from UniProt.ws by filling in some of the arguments and hopefully
telling the user when things do not go according to plan.
}
