% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{loadme}
\alias{loadme}
\title{saveme()  Load a backup rdata file}
\usage{
loadme(dir = "savefiles")
}
\arguments{
\item{dir}{default='savefiles'  the directory containing the RData.rda.xz file.}
}
\description{
saveme()  Load a backup rdata file
}

