% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{local_get_value}
\alias{local_get_value}
\title{Perform a get_value for delimited files}
\usage{
local_get_value(x, delimiter = ": ")
}
\arguments{
\item{x}{Some stuff to split}

\item{delimiter}{The tritrypdb uses ': ' ergo the default.}
}
\value{
A value!
}
\description{
Keith wrote this as .get_value() but functions which start with . trouble me.
}

