% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{lowfilter_counts}
\alias{lowfilter_counts}
\title{A caller for different low-count filters}
\usage{
lowfilter_counts(count_table, type = "cbcb", p = 0.01, A = 1, k = 1,
  cv_min = 0.01, cv_max = 1000, thresh = 2, min_samples = 2)
}
\arguments{
\item{count_table}{some counts to filter}

\item{type}{Filtering method to apply (cbcb, pofa, kofa, cv right now)}

\item{p}{For pofa()}

\item{A}{For pofa()}

\item{k}{For kofa()}

\item{cv_min}{For cv()}

\item{cv_max}{For cv()}

\item{thresh}{Minimum threshold across samples for cbcb}

\item{min_samples}{Minimum number of samples for cbcb}
}
\value{
a data frame of lowfiltered counts
}
\description{
A caller for different low-count filters
}
\examples{
\dontrun{
new <- lowfilter_counts(old)
}
}
\seealso{
\pkg{genefilter}
}

