% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_sql.R
\name{make_biomart_wanted}
\alias{make_biomart_wanted}
\title{Translate 'normal' column IDs to whatever the heck is going on in
the biomart schema.}
\usage{
make_biomart_wanted(wanted = NULL, species = "mmusculus", dbh = NULL)
}
\arguments{
\item{wanted}{List of desired tables and columns}

\item{species}{Species to query}

\item{dbh}{Existing database handle to the local (or not I guess)
db.}
}
\value{
A new list with the weirdo table and column names.
}
\description{
There are a lot of biomart tables (53,100), almost all of them are
full of alignment information used to infer gene homology
relationships.  The remaining tables are the fun stuff and fall
into a few categories: gene, transcript, translation.
}
\details{
The tables also have a peculiar naming scheme which I have not yet
figured out; the end result is that if one wants the gene_id column
for mmusculus, one does not invoke
>'SELECT gene_id from mmusculus_gene'
but instead
'SELECT gene_id_1020_key from mmusculus_gene_ensembl__gene__main'
I am sure there are good reasons for this, but I do not know them.
So, this function exists to take 'normal' names, invoke describe
table, and use some regex shenanigans to find the weirdo names.
}
