% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{make_bsgenome_from_fasta}
\alias{make_bsgenome_from_fasta}
\title{Create a BSgenome from a fasta file.}
\usage{
make_bsgenome_from_fasta(
  fasta,
  pkgname,
  title,
  organism,
  common_name,
  provider,
  build_dir = "build",
  genome_prefix = NULL,
  author = "Ashton Trey Belew <abelew@gmail.com>",
  URL = "https://github.com/abelew/hpgltools",
  installp = TRUE
)
}
\arguments{
\item{fasta}{Input fasta file.}

\item{pkgname}{Name for the new BSgenome.}

\item{title}{Its title.}

\item{organism}{The organism name for this package.}

\item{common_name}{The corresponding common name}

\item{provider}{Where it was sequenced.}

\item{build_dir}{Location to put intermediate files.}

\item{genome_prefix}{Prefix to put on the intermediary files.}

\item{author}{hmm}

\item{installp}{Install the created package?}

\item{url}{Location for the publicly available package.}
}
\description{
Create a BSgenome from a fasta file.
}
