% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_dnaseq_spec}
\alias{make_dnaseq_spec}
\title{Generate a DNASeq specification for use by gather_preprocessing_metadata()}
\usage{
make_dnaseq_spec()
}
\description{
This currently assumes the set of tools used by one doing RNASeq to be trimomatic,
fastqc, hisat2, htseq, freebayes, and my variant post-processor.
}
