% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_organismdbi}
\alias{make_eupath_organismdbi}
\title{Create an organismDbi instance for an eupathdb organism.}
\usage{
make_eupath_organismdbi(species = "Leishmania major strain Friedlin",
  dir = ".", overwrite = FALSE, kegg_abbreviation = NULL,
  metadata = NULL, ...)
}
\arguments{
\item{species}{A species in the eupathDb metadata.}

\item{dir}{Directory in which to build the packages.}

\item{overwrite}{Overwrite existing data files?}

\item{kegg_abbreviation}{For when we cannot automagically find the kegg species id.}

\item{metadata}{Use a pre-existing metadata table, or download a new one.}
}
\value{
The result of attempting to install the organismDbi package.
}
\description{
Create an organismDbi instance for an eupathdb organism.
}
\author{
Keith Hughitt
}
