% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_pkgnames}
\alias{make_eupath_pkgnames}
\title{Generate standardized package names for the various eupathdb species.}
\usage{
make_eupath_pkgnames(species = "Coprinosis.cinerea.okayama7#130",
  metadata = NULL, ...)
}
\arguments{
\item{species}{Guess.}

\item{metadata}{Eupathdb metadata.}

\item{...}{Further arguments to pass to download_eupathdb_metadata()}
}
\value{
List of package names and some booleans to see if they have already been installed.
}
\description{
In my test directory, I have a little for loop which randomly chooses a couple
of eupathdb species for which to try and generate genome/annotation packages.
I am on my 4th or 5th iteration of passing that loop and in all of them
I have found some new and exciting exception to how a strain should be named.
The default argument for this function shows the funniest one so far.
With that in mind, this function should provide consistent, valid package names.
}
