% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_txdb}
\alias{make_eupath_txdb}
\title{Generate TxDb for EuPathDB organism}
\usage{
make_eupath_txdb(species = NULL, entry = NULL, dir = "eupathdb",
  version = NULL, reinstall = FALSE, metadata = NULL, ...)
}
\arguments{
\item{species}{Species name from the eupathdb metadata.}

\item{entry}{One row from the organism metadata.}

\item{dir}{Base directory for building the package.}

\item{version}{Which version of the eupathdb to use for creating this package?}

\item{reinstall}{Overwrite an existing installed package?}

\item{metadata}{dataframe of eupathdb metadata.}

\item{...}{Extra arguments for getting metadata.}
}
\value{
TxDb instance name.
}
\description{
Generate TxDb for EuPathDB organism
}
\author{
Keith Hughitt with significant modifications by atb.
}
