% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{make_exampledata}
\alias{make_exampledata}
\title{make_exampledata()  A small hack of limma's exampleData()
function to allow for arbitrary data set sizes.}
\usage{
make_exampledata(ngenes = 1000, columns = 5)
}
\arguments{
\item{ngenes}{default=1000  how many genes in the fictional data set.}

\item{columns}{default=5  how many samples in this data set.}
}
\value{
a matrix of pretend counts
}
\description{
make_exampledata()  A small hack of limma's exampleData()
function to allow for arbitrary data set sizes.
}
\examples{
## pretend = make_exampledata()
}
\seealso{
\code{\link{makeExampleData}}
}

