% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{make_gsc_from_abundant}
\alias{make_gsc_from_abundant}
\title{Given a pairwise result, make a gene set collection.}
\usage{
make_gsc_from_abundant(pairwise, according_to = "deseq",
  orgdb = "org.Hs.eg.db", pair_names = c("ups", "downs"),
  category_name = "infection", phenotype_name = "parasite",
  set_name = "elsayed_macrophage", color = TRUE,
  current_id = "ENSEMBL", required_id = "ENTREZID", ...)
}
\arguments{
\item{pairwise}{A pairwise result, or combined de result, or extracted genes.}

\item{according_to}{When getting significant genes, use this method.}

\item{orgdb}{Annotation dataset.}

\item{pair_names}{What set of names will be used: up vs. down, high vs. low, etc.}

\item{category_name}{What category does this GSC describe?}

\item{phenotype_name}{What phenotype does this GSC describe?}

\item{set_name}{A name for the created gene set.}

\item{color}{Make a colorSet?}

\item{current_id}{Usually we use ensembl IDs, but that does not \emph{need} to be the case.}

\item{required_id}{gsva uses entrezids by default.}

\item{...}{Extra arguments for extract_abundant_genes().}
}
\value{
List containing 3 GSCs, one containing both the ups/downs called
'colored', one of the ups, and one of the downs.
}
\description{
If I want to play with gsva and friends, then I need GeneSetCollections!
Much like make_gsc_from_significant(), this function extract the genes deemed
'abundant' and generates gene sets accordingly.
}
