% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{make_gsc_from_ids}
\alias{make_gsc_from_ids}
\title{Create a gene set collection from a set of arbitrary IDs.}
\usage{
make_gsc_from_ids(first_ids, second_ids = NULL, orgdb = "org.Hs.eg.db",
  pair_names = "up", category_name = "infection",
  phenotype_name = NULL, set_name = "elsayed_macrophage",
  current_id = "ENSEMBL", required_id = "ENTREZID")
}
\arguments{
\item{first_ids}{The required IDs for a single set.}

\item{second_ids}{Potentially null optionally used for a second, presumably contrasting set.}

\item{orgdb}{Orgdb annotation, used to translate IDs to the required type.}

\item{pair_names}{The suffix of the generated set(s).}

\item{category_name}{Primarily used for colored gene sets, but describes the
category of data examined in this set.}

\item{phenotype_name}{Also used for colored gene sets.}

\item{set_name}{Primary identifier for the set.}

\item{current_id}{What type of ID is the data currently using?}

\item{required_id}{What type of ID should the use?}
}
\description{
Create a gene set collection from a set of arbitrary IDs.
}
