% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{make_gsc_from_pairwise}
\alias{make_gsc_from_pairwise}
\title{Given a pairwise result, make a gene set collection.}
\usage{
make_gsc_from_pairwise(pairwise, according_to = "deseq",
  orgdb = "org.Hs.eg.db", pair_names = c("ups", "downs"),
  color = TRUE, current_id = "ENSEMBL",
  set_name = "elsayed_macrophage", phenotype_name = "infection",
  required_id = "ENTREZID", ...)
}
\arguments{
\item{pairwise}{A pairwise result, or combined de result, or extracted genes.}

\item{according_to}{When getting significant genes, use this method.}

\item{orgdb}{Annotation dataset.}

\item{color}{Make a colorSet?}

\item{current_id}{Usually we use ensembl IDs, but that does not _need_ to be the case.}

\item{set_name}{A name for the created gene set.}

\item{phenotype_name}{When making color sets, use this phenotype name.}

\item{required_id}{gsva uses entrezids by default.}

\item{...}{Extra arguments for extract_significant_genes().}
}
\description{
If I want to play with gsva and friends, then I need GeneSetCollections!
}
