% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{make_gsc_from_significant}
\alias{make_gsc_from_significant}
\title{Create a gene set category from the results of extract_significant_genes().}
\usage{
make_gsc_from_significant(
  significant,
  according_to = "deseq",
  annotation_name = "org.Hs.eg.db",
  category_name = "infection",
  phenotype_name = "parasite",
  set_name = "elsayed_macrophage",
  color = TRUE,
  current_id = "ENSEMBL",
  required_id = "ENTREZID",
  min_gmt_genes = 10,
  ...
)
}
\arguments{
\item{significant}{extract_significant_genes() result.}

\item{according_to}{Default method defining the set of significant genes.}

\item{annotation_name}{orgdb name from which to collect annotations.}

\item{category_name}{Name of the new gsc category.}

\item{phenotype_name}{Name of the new gsc.}

\item{set_name}{Name of the new set.}

\item{color}{Define if this is both up+down or just one.}

\item{current_id}{Most of the GSC tools we have on hand want entrez IDs, use this to convert.}

\item{required_id}{Most of the GSC tools we have on hand want entrez IDs, use this to convert.}

\item{min_gmt_genes}{If groups have less than this number of genes, do not bother making a gsc.}
}
\description{
Create a gene set category from the results of extract_significant_genes().
}
