% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{make_id2gomap}
\alias{make_id2gomap}
\title{Make a go mapping from IDs in a format suitable for topGO}
\usage{
make_id2gomap(goid_map = "reference/go/id2go.map", goids_df = NULL,
  overwrite = FALSE)
}
\arguments{
\item{goid_map}{A topGO mapping file}

\item{goids_df}{If there is no goid_map, create it with this}

\item{overwrite}{A boolean, if it already exists, rewrite the mapping file?}
}
\value{
a summary of the new goid table
}
\description{
Make a go mapping from IDs in a format suitable for topGO
}

