% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{make_organ}
\alias{make_organ}
\title{Create an organismDbi object by joining a txdb and orgdb together.}
\usage{
make_organ(txdb, keytype = NA, orgdb = NA)
}
\arguments{
\item{txdb}{Txdb input to merge}

\item{keytype}{When merging to an orgdb, what key to use?}

\item{orgdb}{The orgdb to help create the OrganismDbi instance.}
}
\value{
An OrganismDb instance
}
\description{
This function is a bit more fragile than I would like.
}
\examples{
\dontrun{
 orgdbi <- make_organ(Tcruzi_txdb, orgdb=Tcruzi_orgdb)
}
}
\seealso{
\pkg{S4Vectors} \pkg{GenomicFeatures} \pkg{AnnotationDbi} \pkg{OrganismDbi}
 \code{\link[GenomicFeatures]{makePackageName}}
}
