% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb_organdb.r
\name{make_organismdbi}
\alias{make_organismdbi}
\title{Create an OrganismDbi for a species at the TriTrypDb}
\usage{
make_organismdbi(id = "lmajor_friedlin", cfg = NULL,
  output_dir = "organdb", ...)
}
\arguments{
\item{id}{Unique tritrypdb identifier.}

\item{cfg}{A configuration dataframe, when null it will be replaced by reading a csv file in
inst/extdata.}

\item{output_dir}{The directory into which to put the various intermediate files, including
downloads from the TriTrypdb, the created OrgDb and TxDb instances, and the final
OrganismDbi.}

\item{...}{Extra arguments including a boolean for whether to include kegg.}
}
\value{
A path, some data files, and a kitty!
}
\description{
OrganismDbi instances are pretty neat, they pull together OrgDb and TxDb.  With any luck, this
function provides the ability to pull together all the data from the TriTrypDb, GO.db, and
KEGGREST in order to accomplish these peculiar tasks.
}
\examples{
\dontrun{
 crazytown <- make_organismdbi()  ## wait a loong time
}
}
\seealso{
\pkg{AnnotationForge}
}

