% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_orgdb}
\alias{make_orgdb}
\title{Make an orgDb object from some information provided by make_orgdb_info()}
\usage{
make_orgdb(orgdb_info, id = "lmajor_friedlin", cfg = NULL, kegg = TRUE,
  output_dir = "organismdbi", ...)
}
\arguments{
\item{orgdb_info}{List of data frames generated by make_orgdb_info()}

\item{id}{Human readable species identifier, keys off the cfg data frame.}

\item{cfg}{Configuration data extracted either from inst/eupath_configuration.csv or provided by
the user.}

\item{kegg}{Attempt adding kegg data?}

\item{output_dir}{Base output directory for the resulting packages.}
}
\value{
List of the resulting package name(s) and whether they installed.
}
\description{
An orgDb object should provide some useful annotation data including fun stuff like gene
ontology, kegg, etc.  In the case of the species at the TriTrypDb, much of this information
is available in the species .txt file.  This function takes that data and collates it into the
final orgDb objects using AnnotationForge.  It then makes some attempts to ensure that the
resulting material created in the filesystem conforms to specifications which allow one to have
multiple strains, etc.  Finally, if everything goes according to plan, it calls
devtools::install() and installs the resulting package.
}

