% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{make_orgdb_from_ah}
\alias{make_orgdb_from_ah}
\title{Create an orgdb from an taxonID}
\usage{
make_orgdb_from_ah(taxid)
}
\arguments{
\item{taxid}{TaxonID from AnnotationHub}
}
\value{
An Orgdb instance
}
\description{
This function is a bit more fragile than I would like.  I am not completely sold
on AnnotationHub yet.
}
\examples{
\dontrun{
 orgdbi <- mytaxIdToOrgDb(taxid)
}
}
\seealso{
\pkg{AnnotationHub} \pkg{S4Vectors}
}
