% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{make_pairwise_contrasts}
\alias{make_pairwise_contrasts}
\title{make_pairwise_contrasts()  Run makeContrasts() with all pairwise comparisons.}
\usage{
make_pairwise_contrasts(model, conditions, do_identities = TRUE,
  do_pairwise = TRUE, extra_contrasts = NULL)
}
\arguments{
\item{model}{a model describing the conditions/batches/etc in the experiment}

\item{conditions}{a factor of conditions in the experiment}

\item{do_identities}{default=TRUE  whether or not to include all the identity strings.
Limma can handle this, edgeR cannot.}

\item{do_pairwise}{default=TRUE  whether or not to include all the pairwise strings.
This shouldn't need to be set to FALSE, but just in case.}

\item{extra_contrasts}{default=NULL  an optional string of extra contrasts to include.}
}
\value{
A list including the following information:
  all_pairwise_contrasts = the result from makeContrasts(...)
  identities = the string identifying each condition alone
  all_pairwise = the string identifying each pairwise comparison alone
  contrast_string = the string passed to R to call makeContrasts(...)
  names = the names given to the identities/contrasts
}
\description{
make_pairwise_contrasts()  Run makeContrasts() with all pairwise comparisons.
}
\examples{
## pretend = make_pairwise_contrasts(model, conditions)
}
\seealso{
\code{\link{makeContrasts}}
}

