% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{make_pombe_expt}
\alias{make_pombe_expt}
\title{Create the pombe summarized experiment when using deprecated function names.}
\usage{
make_pombe_expt(...)
}
\description{
Create the pombe summarized experiment when using deprecated function names.
}
