% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{make_pombe_se}
\alias{make_pombe_se}
\title{Create a default summarized experiment using the publicly available fission dataset.}
\usage{
make_pombe_se(
  annotation = TRUE,
  host = "sep2025-fungi.ensembl.org",
  backup_annotation = "org.Spombe.972h.v46.eg.db"
)
}
\arguments{
\item{annotation}{Include annotations?}

\item{host}{ensembl host to query}

\item{backup_annotation}{Use this annotation source if ensembl is down.}
}
\description{
Create a default summarized experiment using the publicly available fission dataset.
}
