% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{make_quartile_factor}
\alias{make_quartile_factor}
\title{Make a quartile factor out of a numeric vector}
\usage{
make_quartile_factor(numeric_vector)
}
\arguments{
\item{numeric_vector}{Vector of numbers!}
}
\value{
Factor with levels from q1 to q4.
}
\description{
Make a quartile factor out of a numeric vector
}
