% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{make_report}
\alias{make_report}
\title{Make a knitr report with some defaults set a priori.}
\usage{
make_report(name = "report", type = "pdf")
}
\arguments{
\item{name}{Name the document!}

\item{type}{Html or pdf reports?}
}
\value{
Dated report file.
}
\description{
I keep forgetting to set appropriate options for knitr.  This tries to set them.
}
\seealso{
\pkg{knitr} \pkg{rmarkdown} \pkg{knitrBootstrap}
}

