% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_rnaseq_spec}
\alias{make_rnaseq_spec}
\title{Generate a RNASeq specification for use by gather_preprocessing_metadata()}
\usage{
make_rnaseq_spec(species_list = NULL)
}
\description{
This currently assumes the set of tools used by one doing RNASeq to be trimomatic,
fastqc, hisat2, and htseq.
}
