% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{make_summary_df}
\alias{make_summary_df}
\title{Create a df out of the results of summary().}
\usage{
make_summary_df(df, iqr_multiplier = 1.5)
}
\arguments{
\item{df}{Input dataframe.}

\item{iqr_multiplier}{Range for outlier detection.}
}
\value{
a dataframe!
}
\description{
Create a df out of the results of summary().
}
