% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_taxon_names}
\alias{make_taxon_names}
\title{Iterate through the various ways of representing taxon names}
\usage{
make_taxon_names(entry)
}
\arguments{
\item{entry}{An entry of the eupathdb metadata.}
}
\description{
Iterate through the various ways of representing taxon names
}
