% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{make_tooltips}
\alias{make_tooltips}
\title{make_tooltips()  Create a simple df from gff which contains tooltip usable information for gVis graphs.}
\usage{
make_tooltips(annotations = NULL, gff = NULL)
}
\arguments{
\item{gff}{or annotations: Either a gff file or annotation data frame (which likely came from a gff file.)}
}
\value{
a df of tooltip information
}
\description{
make_tooltips()  Create a simple df from gff which contains tooltip usable information for gVis graphs.
}

