% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{make_tx_gene_map}
\alias{make_tx_gene_map}
\title{I keep messing up the creation of the salmon trancript to gene map.}
\usage{
make_tx_gene_map(
  annotations,
  gene_column = "ensembl_gene_id",
  transcript_column = "ensembl_transcript_id",
  tx_version_column = "transcript_version",
  new_column = "salmon_transcript"
)
}
\description{
Maybe this will help.  I have a smarter but much slower method in
the tmrc3 data which first creates an expressionset without
annotations then cross references the rownames against combinations
of columns in the annotations to figure out the correct pairing.
This helps when I have a combined transcriptome and get confused.
}
