% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_tritrypdb.r
\name{make_txdb}
\alias{make_txdb}
\title{Create a TxDb object given data provided by make_orgdb_info()}
\usage{
make_txdb(orgdb_info, cfg_line, gff = NULL, from_gff = FALSE,
  output_dir = "organismdbi", ...)
}
\arguments{
\item{orgdb_info}{List of data frames generated by make_orgdb_info().}

\item{cfg_line}{Configuration data frame as per make_orgdb.}

\item{gff}{File to read}

\item{output_dir}{Place to put rda intermediates.}

\item{...}{Extra arguments to pass through.}
}
\value{
List of the resulting txDb package and whether it installed.
}
\description{
Much like make_orgdb() above, this uses the same data to generate a TxDb object.
}

