% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pca.r
\name{makeSVD}
\alias{makeSVD}
\title{this a function scabbed from Hector and Kwame's cbcbSEQ}
\usage{
makeSVD(data)
}
\arguments{
\item{data}{A data frame to decompose}
}
\value{
a list containing the s,v,u from fast.svd
}
\description{
It just does fast.svd of a matrix comprised of the matrix - rowMeans(matrix)
}
\examples{
\dontrun{
 svd = makeSVD(data)
}
}
\seealso{
\pkg{corpcor} \code{\link[corpcor]{fast.svd}}
}

