% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{map_keepers}
\alias{map_keepers}
\title{Find the correct tables given a set of definitions of desired tables, numerators/denominators.}
\usage{
map_keepers(keepers, table_names, datum)
}
\arguments{
\item{keepers}{List/scalar representation of desired tables.}

\item{table_names}{The actual list of results produced by the various methods employed.}

\item{datum}{The full dataset.}
}
\description{
This is responsible for hunting down tables which correspond to the various ways one may
represent them.
}
