% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_kegg.r
\name{map_kegg_dbs}
\alias{map_kegg_dbs}
\title{Maps KEGG identifiers to ENSEMBL gene ids.}
\usage{
map_kegg_dbs(kegg_ids)
}
\arguments{
\item{kegg_ids}{List of KEGG identifiers to be mapped.}
}
\value{
Ensembl IDs as a character list.
}
\description{
Takes a list of KEGG gene identifiers and returns a list of ENSEMBL
ids corresponding to those genes.
}
\examples{
\dontrun{
 kegg_df <- load_kegg_annotations(species="coli")
 kegg_ids <- kegg_df[["kegg_geneid"]]
 ensembl_list <- kegg_to_ensembl(kegg_ids)
}
}
\seealso{
\pkg{KEGGREST}
 \code{\link[KEGGREST]{keggGet}}
}
