% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.R
\name{map_species_orgdb}
\alias{map_species_orgdb}
\title{Guess the orgdb from a genusspecies.}
\usage{
map_species_orgdb(species, genus = NULL)
}
\arguments{
\item{species}{Input species}

\item{genus}{and genus.}
}
\description{
Given a name like 'mmusculus', guess the orgdb package name.
}
