% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{mean_by_bioreplicate}
\alias{mean_by_bioreplicate}
\title{An attempt to address a troubling question when working with DIA data.}
\usage{
mean_by_bioreplicate(expt, fact = "bioreplicate", fun = "mean")
}
\arguments{
\item{expt}{Starting expressionset to mangle.}

\item{fact}{Metadata factor to use when taking the mean of biological
replicates.}

\item{fun}{Assumed to be mean, but one might want median.}
}
\value{
new expressionset
}
\description{
My biggest concern when treating DIA data in a RNASeqish manner is the fact
that if a given peptide is not identified, that is not the same thing as
stating that it was not translated.  It is somewhat reminiscent of the often
mocked and repeated Donald Rumsfeld statement regarding known unknowns
vs. unknown unknowns.  Thus, in an RNASeq experiment, if one sees a zero, one
may assume that transcript was not transcribed, it may be assumed to be a
known zero(unknown).  In contrast, if the same thing happens in a DIA data
set, that represents an unknown unknown.  Perhaps it was not translated, and
perhaps it was not identified.
}
\details{
This function therefore does the following:
  1.  Backfill all 0s in the matrix to NA.
  2.  Performs a mean across all samples which are known technical replicates
  of the same biological replicate.  This mean is performed using
  na.rm = TRUE.  Thus the entries which used to be 0 should no longer affect
  the result.
  3.  Recreate the expressionset with the modified set of samples.
}
