% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{mean_by_factor}
\alias{mean_by_factor}
\title{Runs median_by_factor with fun set to 'mean'.}
\usage{
mean_by_factor(data, fact = "condition")
}
\arguments{
\item{data}{Input expt}

\item{fact}{Metadata factor over which to perform mean().}
}
\description{
Runs median_by_factor with fun set to 'mean'.
}
