% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_tables.r
\name{median_by_factor}
\alias{median_by_factor}
\title{\code{median_by_factor()}  Create a data frame of the medians of rows by a given factor in the data}
\usage{
median_by_factor(data, fact)
}
\arguments{
\item{data}{a data frame, presumably of counts.}

\item{fact}{a factor describing the columns in the data.}
}
\value{
a data frame of the medians
}
\description{
This assumes of course that (like expressionsets) there are separate columns for each replicate
of the conditions.  This will just iterate through the levels of a factor describing the columns,
extract them, calculate the median, and add that as a new column in a separate data frame.
}
\examples{
\dontrun{
 compressed = hpgltools:::median_by_factor(data, experiment$condition)
}
}

