% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_sql.R
\name{merge_biomart_tables}
\alias{merge_biomart_tables}
\title{I guess if I were smart I would do a join in SQL; but I am still
learning the structure of these tables.}
\usage{
merge_biomart_tables(sql_annotations, primary_key = "gene_id")
}
\description{
I guess if I were smart I would do a join in SQL; but I am still
learning the structure of these tables.
}
