% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{merge_counts_annotations,GRanges-method}
\alias{merge_counts_annotations,GRanges-method}
\title{Use grange information to merge counts and annotations.}
\usage{
\S4method{merge_counts_annotations}{GRanges}(gene_info, all_count_tables, tx_gene_map = NULL)
}
\arguments{
\item{gene_info}{Table of gene annotations.}

\item{all_count_tables}{Matrix of expression data.}

\item{tx_gene_map}{Map of transcripts to genes.}
}
\description{
Use grange information to merge counts and annotations.
}
