% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{merge_counts_annotations}
\alias{merge_counts_annotations}
\title{Merge gene annotations and count tables when the gene_info is a data table/dataframe}
\usage{
merge_counts_annotations(gene_info, all_count_tables, tx_gene_map = NULL)
}
\description{
Merge gene annotations and count tables when the gene_info is a data table/dataframe
}
