% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_gff.R
\name{merge_gff_children}
\alias{merge_gff_children}
\title{Merge child annotations into a parental gff file.}
\usage{
merge_gff_children(
  gff,
  grandparent = "gene",
  parent = "mRNA",
  parent_tag = "Parent",
  id_tag = "ID",
  children = c("CDS", "exon", "three_prime_UTR", "five_prime_UTR")
)
}
\arguments{
\item{gff}{Starting gff file.}

\item{grandparent}{Top-level heirarchy of the gff file (excepting chromosome)}

\item{parent}{Second-level in the gff heirarchy}

\item{parent_tag}{What gff tag to use to trace the lineage of a gene.}

\item{id_tag}{How to extract gene IDs given the above.}

\item{children}{New annotation types to create.}
}
\description{
On occasion we create gff files with specific features of interest.
Most notably, when working with our parasites, we sometimes create
new 5' and 3' UTR entries using spliced leader/polyA junction
reads.  This function attempts to merge together these annotations.
}
\seealso{
[import.gff3()]
}
