% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.R
\name{mesg}
\alias{mesg}
\title{message() but with a verbose flag.}
\usage{
mesg(..., verbosity = NULL, warn = FALSE)
}
\arguments{
\item{...}{parameters for message()}

\item{verbosity}{actually print the message?}

\item{warn}{Also print a warning?}
}
\description{
message() but with a verbose flag.
}
