% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{getColors}
\alias{getColors}
\title{Generic method to get colors from expression data.}
\usage{
getColors(expt)
}
\arguments{
\item{object}{The object from which to gather colors.}

\item{...}{Additional arguments passed to function constructors.}
}
\value{
colors!
}
\description{
Generic method to get colors from expression data.
}
