% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.r
\name{model_test}
\alias{model_test}
\title{Make sure a given experimental factor and design will play together.}
\usage{
model_test(design, goal = "condition", factors = NULL, ...)
}
\arguments{
\item{design}{Dataframe describing the design of the experiment.}

\item{goal}{Experimental factor you actually want to learn about.}

\item{factors}{Experimental factors you rather wish would just go away.}

\item{...}{I might decide to add more options from other functions.}
}
\value{
List of booleans telling if the factors + goal will work.
}
\description{
Have you ever wanted to set up a differential expression analysis and after minutes of the
computer churning away it errors out with some weird error about rank?  Then this is the function
for you!
}
\seealso{
\code{\link{model.matrix}} \code{\link{qr}}
}
