% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{mtb_rmats}
\alias{mtb_rmats}
\title{The Rmats alternative splicing results for a Mycobacterium tuberculosis experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
mtb_rmats
}
\description{
This is the comparison of the splicing isoforms observed in the host
following infection with a few Mtb strains.  Our goal was to learn how
reliable the various alternative splicing quantification tools are. (spoiler:
not very).
}
\keyword{datasets}
