% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_boxplot}
\alias{my_boxplot}
\title{Make a ggplot boxplot of a set of samples.}
\usage{
my_boxplot(df = NULL, colors = NULL, expt = NULL, names = NULL,
  scale = "raw")
}
\arguments{
\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{names}{a nicer version of the sample names}

\item{scale}{whether to log scale the y-axis}
}
\value{
box_plot a ggplot2 boxplot of the samples.  Each boxplot
contains the following information: a centered line describing the
median value of counts of all genes in the sample, a box around the
line describing the inner-quartiles around the median (quartiles 2
and 3 for those who are counting), a vertical line above/below the
box which shows 1.5x the inner quartile range (a common metric of
the non-outliers), and single dots for each gene which is outside
that range.  A single dot is transparent.
}
\description{
Make a ggplot boxplot of a set of samples.
}
\examples{
## a_boxplot = my_boxplot(expt=expt)
## a_boxplot  ## ooo pretty boxplot look at the lines
}
\seealso{
\code{\link{geom_boxplot}}, \code{\link{melt}},
\code{\link{scale_x_discrete}}
}

