% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_cor}
\alias{my_cor}
\title{Wrap cor() to include robust correlations}
\usage{
my_cor(df = NULL, method = "pearson")
}
\arguments{
\item{df}{a data frame to test}

\item{method}{Correlation method to use.  Defaults to pearson.
Includes pearson, spearman, kendal, robust.}
}
\value{
correlation some fun correlation statistics
}
\description{
Wrap cor() to include robust correlations
}
\examples{
## my_cor(df=df)
## my_cor(df=df, method="robust")
}
\seealso{
\code{\link{cor}}, \code{\link{cov}}, \code{\link{covRob}}
}

