% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_corheat}
\alias{my_corheat}
\title{Make a heatmap2 description of the correlation between samples.}
\usage{
my_corheat(df = NULL, colors = NULL, design = NULL, expt = NULL,
  method = "pearson")
}
\arguments{
\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{method}{correlation statistic to use.  Defaults to pearson.}
}
\value{
corheat_plot a gplots heatmap describing how the samples
pairwise correlate with one another.
}
\description{
Make a heatmap2 description of the correlation between samples.
}
\examples{
## corheat_plot = my_corheat(expt=expt, method="robust")
## corheat_plot
}
\seealso{
\code{\link{my_cor}}, \code{\link{brewer.pal}},
\code{\link{heatmap.2}}, \code{\link{recordPlot}}
}

