% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_disheat}
\alias{my_disheat}
\title{Make a heatmap2 description of the distance between samples.}
\usage{
my_disheat(df = NULL, colors = NULL, design = NULL, expt = NULL,
  method = "euclidean")
}
\arguments{
\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{method}{distance metric to use.  Defaults to euclidean.
Available: euclidean, manhattan, maximum, canberra, binary,
minkowski, I might add more}
}
\value{
disheat_plot a gplots heatmap describing the distance
between each pair of samples.
}
\description{
Make a heatmap2 description of the distance between samples.
}
\examples{
## disheat_plot = my_disheat(expt=expt)
## disheat_plot  ## ooo blue
}
\seealso{
\code{\link{dist}}, \code{\link{brewer.pal}},
\code{\link{heatmap.2}}, \code{\link{recordPlot}}
}

