% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{my_doisva}
\alias{my_doisva}
\title{There are some funky scoping problems in isva::DoISVA().}
\usage{
my_doisva(
  data.m,
  pheno.v,
  cf.m = NULL,
  factor.log = FALSE,
  pvthCF = 0.01,
  th = 0.05,
  ncomp = NULL,
  icamethod = "fastICA"
)
}
\description{
Thus I copy/pasted the function and attempted to address them here.
}
