% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_gvis_ma_plot}
\alias{my_gvis_ma_plot}
\title{Make an html version of an MA plot.}
\usage{
my_gvis_ma_plot(counts, degenes, tooltip_data = NULL,
  filename = "html/gvis_ma_plot.html", base_url = "")
}
\arguments{
\item{counts}{df of linear-modelling, normalized counts by sample-type,
which is to say the output from voom/voomMod/my_voom().}

\item{de_genes}{df from toptable or its friends containing p-values.}

\item{adjpval_cutoff}{a cutoff defining significant from not.
Defaults to 0.05.}

\item{gvis_filename}{a filename to write a fancy html graph.
Defaults to NULL in which case the following parameter isn't needed.}

\item{tooltip_data}{a df of tooltip information for gvis
graphs. NULL by default.}
}
\value{
NULL, but along the way an html file is generated which
contains a googleVis MA plot.  See my_ma_plot() for details.
}
\description{
Make an html version of an MA plot.
}
\examples{
## my_gvis_ma_plot(voomed_data, toptable_data, filename="html/fun_ma_plot.html", base_url="http://yeastgenome.org/accession?")
}
\seealso{
\code{\link{my_ma_plot}}
}

