% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_gvis_scatter}
\alias{my_gvis_scatter}
\title{Make an html version of a scatter plot.}
\usage{
my_gvis_scatter(df, tooltip_data = NULL,
  filename = "html/gvis_scatter.html", base_url = "")
}
\arguments{
\item{counts}{df of two columns to compare}

\item{filename}{a filename to write a fancy html graph.
Defaults to NULL in which case the following parameter isn't needed.}

\item{tooltip_data}{a df of tooltip information for gvis
graphs. NULL by default.}

\item{base_url}{a url to send click events which will be suffixed
with the gene name}
}
\value{
NULL, but along the way an html file is generated which
contains a googleVis scatter plot.  See my_scatter_plot() for details.
}
\description{
Make an html version of a scatter plot.
}
\examples{
## my_gvis_scatter(a_dataframe_twocolumns, filename="html/fun_scatter_plot.html", base_url="http://yeastgenome.org/accession?")
}
\seealso{
\code{\link{gvisScatterChart}}
}

