% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_histogram}
\alias{my_histogram}
\title{Make a pretty histogram of something}
\usage{
my_histogram(df)
}
\arguments{
\item{df}{a dataframe of lots of pretty numbers}
}
\value{
a ggplot histogram
}
\description{
Make a pretty histogram of something
}
\examples{
## kittytime = my_histogram(df)
}
\seealso{
\code{\link{geom_histogram}}, \code{\link{geom_density}},
}

