% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{my_isva}
\alias{my_isva}
\title{There are some funky scoping problems in isva::DoISVA().}
\usage{
my_isva(
  data.m,
  pheno.v,
  cf.m = NULL,
  factor.log = FALSE,
  pvthCF = 0.01,
  th = 0.05,
  ncomp = NULL,
  icamethod = "fastICA"
)
}
\arguments{
\item{data.m}{Input matrix.}

\item{pheno.v}{Vector of conditions of interest in the data.}

\item{cf.m}{Matrix of confounded conditions in the data.}

\item{factor.log}{I forget.}

\item{pvthCF}{Minimal p-value for considering.}

\item{th}{threshold for inclusion.}

\item{ncomp}{Number of SVA components to estimate.}

\item{icamethod}{Which ICA implementation to use?}
}
\description{
Thus I copy/pasted the function and attempted to address them here.
}
\seealso{
[isva]
}
