% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_libsize}
\alias{my_libsize}
\title{Make a ggplot graph of library sizes.}
\usage{
my_libsize(df = NULL, design = NULL, colors = NULL, expt = NULL,
  scale = TRUE, names = NULL)
}
\arguments{
\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{scale}{whether or not to log10 the y-axis}
}
\value{
libsize_plot a ggplot2 plot of each library's size
}
\description{
Make a ggplot graph of library sizes.
}
\examples{
## libsize_plot = my_libsize(expt=expt)
## libsize_plot  ## ooo pretty bargraph
}
\seealso{
\code{\link{geom_bar}}, \code{\link{geom_text}},
\code{\link{prettyNum}}, \code{\link{scale_y_log10}}
}

