% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_linear_scatter}
\alias{my_linear_scatter}
\title{Make a pretty scatter plot between two sets of numbers with a
linear model superimposed and some supporting statistics.}
\usage{
my_linear_scatter(df, tooltip_data = NULL, gvis_filename = NULL,
  cormethod = "pearson")
}
\arguments{
\item{df}{a dataframe likely containing two columns}

\item{gvis_filename}{a filename to write a fancy html graph.
Defaults to NULL in which case the following parameter isn't needed.}

\item{tooltip_data}{a df of tooltip information for gvis
graphs. NULL by default.}

\item{cormethod}{what type of correlation to check?  Defaults to
'pearson'}
}
\value{
a list including a ggplot2 scatter plot and some
histograms.  This plot provides a "bird's eye"
view of two data sets.  This plot assumes a (potential) linear
correlation between the data, so it calculates the correlation
between them.  It then calculates and plots a robust linear model
of the data using an 'SMDM' estimator (which I don't remember how
to describe, just that the document I was reading said it is good).
The median/mad of each axis is calculated and plotted as well.  The
distance from the linear model is finally used to color the dots on
the plot.  Histograms of each axis are plotted separately and then
together under a single cdf to allow tests of distribution
similarity.  This will make a fun clicky googleVis graph if
requested.
}
\description{
Make a pretty scatter plot between two sets of numbers with a
linear model superimposed and some supporting statistics.
}
\examples{
## my_linear_scatter(lotsofnumbers_intwo_columns, tooltip_data=tooltip_dataframe, gvis_filename="html/fun_scatterplot.html")
}
\seealso{
\code{\link{lmrob}}, \code{\link{weights}},
\code{\link{hsv}}, \code{\link{mad}}, \code{\link{my_histogram}}
}

