% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_multihistogram}
\alias{my_multihistogram}
\title{Make a pretty histogram of multiple datasets}
\usage{
my_multihistogram(df)
}
\arguments{
\item{df}{a dataframe of lots of pretty numbers}
}
\value{
a ggplot histogram comparing multiple data sets
Along the way this generates pairwise t tests of the columns of
data.
}
\description{
Make a pretty histogram of multiple datasets
}
\examples{
## kittytime = my_multihistogram(df)
}
\seealso{
\code{\link{pairwise.t.test}}, \code{\link{ddply}},
\code{\link{rbind}}
}

