% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_pca}
\alias{my_pca}
\title{Make a ggplot PCA plot describing the samples' clustering}
\usage{
my_pca(df = NULL, colors = NULL, design = NULL, expt = NULL,
  names = NULL)
}
\arguments{
\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{method}{a correlation method to use.  Defaults to pearson.}

\item{names}{use pretty names for the samples?}
}
\value{
pca_plot a ggplot2 describing the principle component
analysis of the samples.  This makes use of cbcbSEQ and prints the
table of variance by component.
}
\description{
Make a ggplot PCA plot describing the samples' clustering
}
\examples{
## pca_plot = my_pca(expt=expt)
## pca_plot
}
\seealso{
\code{\link{makeSVD}}, \code{\link{pcRes}},
\code{\link{geom_dl}}
}

