% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_read_files}
\alias{my_read_files}
\title{Read a bunch of count tables and create a usable data frame from
them.}
\usage{
my_read_files(ids, files)
}
\arguments{
\item{ids}{a list of experimental ids}

\item{files}{a list of files to read}
}
\value{
initial_count a data frame of count tables
}
\description{
Read a bunch of count tables and create a usable data frame from
them.
}
\examples{
## count_tables = my_read_files(as.character(sample_ids), as.character(count_filenames))
}
\seealso{
\code{\link{create_experiment}}
}

