% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_rpkm}
\alias{my_rpkm}
\title{Express a data frame of counts as reads per killobase(gene) per
million(library).}
\usage{
my_rpkm(df, annotations = gene_annotations)
}
\arguments{
\item{df}{a data frame of counts, alternately an edgeR DGEList}

\item{annotations}{containing gene lengths, defaulting to
'gene_annotations'}
}
\value{
rpkm_df a data frame of counts expressed as rpkm
}
\description{
Express a data frame of counts as reads per killobase(gene) per
million(library).
}
\examples{
## rpkm_df = my_rpkm(df, annotations=gene_annotations)
}
\seealso{
\code{\link{edgeR}} and \code{\link{cpm}}
}

