% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_smc}
\alias{my_smc}
\title{Make an R plot of the standard median correlation among samples}
\usage{
my_smc(df = NULL, colors = NULL, expt = NULL, method = "pearson")
}
\arguments{
\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{method}{a correlation method to use.  Defaults to pearson.}

\item{names}{use pretty names for the samples?}
}
\value{
smc_plot a recordPlot of plot.  This will also write to an
open device.  The resulting plot measures the median correlation of
each sample among its peers.  It notes 1.5* the interquartile range
among the samples and makes a horizontal line at that correlation
coefficient.  Any sample which falls below this line is considered
for removal because it is much less similar to all of its peers.
}
\description{
Make an R plot of the standard median correlation among samples
}
\examples{
## smc_plot = my_smc(expt=expt)
}
\seealso{
\code{\link{my_cor}}, \code{\link{matrixStats::rowMedians}},
\code{\link{quantile}}, \code{\link{diff}}, \code{\link{recordPlot}}
}

