% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_smd}
\alias{my_smd}
\title{Make an R plot of the standard median distance among samples}
\usage{
my_smd(expt = NULL, df = NULL, colors = NULL, method = "euclidean")
}
\arguments{
\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{method}{a distance metric to use.  Defaults to euclidean.}

\item{names}{use pretty names for the samples?}
}
\value{
smd_plot a recordPlot of plot.  This will also write to an
open device.  This plot takes the median distance of each sample
with all of its peers.  It then calculates 1.5* the interquartile
range of distances.  Any sample which has a median distance greater
than this is considered for removal.
}
\description{
Make an R plot of the standard median distance among samples
}
\examples{
## smd_plot = my_smd(expt=expt)
}
\seealso{
\code{\link{dist}}, \code{\link{quantile}},
\code{\link{diff}}, \code{\link{recordPlot}}
}

