% Generated by roxygen2 (4.0.1): do not edit by hand
\name{my_voom}
\alias{my_voom}
\title{Estimate mean-variance relationship between samples and generate
'observational-level weights' in preparation for linear modelling
RNAseq data.  This particular implementation was primarily scabbed
from cbcbSEQ, but changes the mean-variance plot slightly and
attempts to handle corner cases where the sample design is
confounded by setting the coefficient to 1 for those samples rather
than throwing an unhelpful error.  Also, the Elist output gets a
'plot' slot which contains the plot rather than just printing it.}
\usage{
my_voom(dataframe, model, libsize = NULL)
}
\arguments{
\item{dataframe}{a dataframe of sample counts which have been
normalized and log transformed}

\item{model}{an experimental model defining batches/conditions/etc}

\item{libsize}{the size of the libraries (usually provided by
edgeR).  By default this is NULL.}
}
\value{
an EList containing the modified data, weights, design,
libsize, and mean-variance plot.
}
\description{
Estimate mean-variance relationship between samples and generate
'observational-level weights' in preparation for linear modelling
RNAseq data.  This particular implementation was primarily scabbed
from cbcbSEQ, but changes the mean-variance plot slightly and
attempts to handle corner cases where the sample design is
confounded by setting the coefficient to 1 for those samples rather
than throwing an unhelpful error.  Also, the Elist output gets a
'plot' slot which contains the plot rather than just printing it.
}
\examples{
## funkytown = my_voom(samples, model)
}
\seealso{
\code{\link{voom}}, \code{\link{voomMod}}, \code{\link{lmFit}}
}

