% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{mymakeContrasts}
\alias{mymakeContrasts}
\title{A copy of limma::makeContrasts() with special sauce.}
\usage{
mymakeContrasts(..., contrasts = NULL, levels)
}
\arguments{
\item{...}{Conditions used to make the contrasts.}

\item{contrasts}{Actual contrast names.}

\item{levels}{contrast levels used.}
}
\value{
Same contrasts as used in makeContrasts, but with unique names.
}
\description{
This is a copy of limma::makeContrasts without the test of make.names()
Because I want to be able to use it with interaction models potentially
and if a model has first:second, make.names() turns the ':' to a '.'
and then the equivalence test fails, causing makeContrasts() to error
spuriously (I think).
}
