% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_kegg.r
\name{myretrieveKGML}
\alias{myretrieveKGML}
\title{A couple functions from KEGGgraph that have broken}
\usage{
myretrieveKGML(pathwayid, organism, destfile, method = "wget",
  hostname = "http://www.kegg.jp", ...)
}
\description{
A couple functions from KEGGgraph that have broken
}
