% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_from_matrix}
\alias{network_from_matrix}
\title{Given a matrix of scores (bit score, e-value, etc), create an adjacency graph.}
\usage{
network_from_matrix(
  scores,
  metadata = NULL,
  type = "distcor",
  simplify = TRUE,
  mode = "undirected",
  weighted = TRUE,
  diag = FALSE
)
}
\arguments{
\item{metadata}{Currently unused, but intended to provide a
starting point for annotating the resulting adjacency network.
When implemented, it should make use of the annotate_network()
functions which follow.}

\item{type}{Currently I only know of networks which use
correlation, distance, and distcor matrices of the original
scores; but I suspect a cursory glance at the WGCNA documentation
will teach me that there are many more possibilities.}

\item{simplify}{Return a simplified matrix without loops and
redundancies?}

\item{mode}{Network type to create, I don't yet understand the implications
of changing this.}

\item{weighted}{Add weights to the nodes?  I also don't yet
understand what happens when you mess with this.}

\item{diag}{Include the matrix-diagonal nodes?  I do not know when
one would want these.}

\item{score_file}{tsv or matrix of scores with column and row names containing IDs.}
}
\value{
igraph adjacency network.
}
\description{
I am hoping to use this as the starting point for a generic network generator.
In its current form it takes a matrix of pairwise scores and
generates an adjacency graph of those scores.
}
