% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_noiseq.R
\name{noiseq_pairwise}
\alias{noiseq_pairwise}
\title{Perform pairwise comparisons using noiseq.}
\usage{
noiseq_pairwise(
  input = NULL,
  model_fstring = "~ 0 + condition + batch",
  null_fstring = "~",
  model_svs = NULL,
  extra_contrasts = NULL,
  annot_df = NULL,
  force = FALSE,
  keepers = NULL,
  batch = FALSE,
  logtransf = FALSE,
  k = 0.5,
  norm = "tmm",
  factor = NULL,
  lc = 1,
  r = 20,
  adj = 1.5,
  a0per = 0.9,
  filter = 1,
  keep_underscore = TRUE,
  ...
)
}
\arguments{
\item{input}{Expressionset to compare.}

\item{model_fstring}{Formula string describing the model of interest.}

\item{null_fstring}{Formula string describing the null model.}

\item{model_svs}{Surrogate matrix or how to find them.}

\item{extra_contrasts}{Extra contrasts beyond the default set to seek.}

\item{annot_df}{Extra annotations.}

\item{force}{Force the data even if it violates the method's assumptions.}

\item{keepers}{Perform the comparison only over these specific
contrasts instead of all.}

\item{batch}{Noiseq batch factor.}

\item{logtransf}{noiseq log transformer.}

\item{k}{Taken from the noiseq docs.}

\item{norm}{Normalization method (noiseq oddly defaults to rpkm).}

\item{factor}{Metadata factor over which to iterate.}

\item{lc}{taken from the noiseq docs.}

\item{r}{taken from the noiseq docs.}

\item{adj}{taken from the noiseq docs.}

\item{a0per}{taken from the noiseq docs.}

\item{filter}{Filter the data?}

\item{keep_underscore}{Sanitize out underscores?}

\item{...}{Extra arguments.}
}
\value{
List similar to deseq_pairwise/edger_pairwise/etc.
}
\description{
Perform pairwise comparisons using noiseq.
}
\seealso{
DOI:10.1093/nar/gkv711
}
