% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{normalize,SummarizedExperiment-method}
\alias{normalize,SummarizedExperiment-method}
\title{Normalization of se}
\usage{
\S4method{normalize}{SummarizedExperiment}(input, ...)
}
\description{
Normalization of se
}
