% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{normalize}
\alias{normalize}
\title{I really do not think this should be necessary.}
\usage{
normalize(object, ...)
}
\arguments{
\item{object}{Data structure to normalize.}

\item{...}{Other options.}
}
\description{
It seems to me that if I have a judicious importFrom, then
the following setMethod should dispatch correctly.
}
