% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_norm.r
\name{normalize_counts}
\alias{normalize_counts}
\title{Perform a simple normalization of a count table.}
\usage{
normalize_counts(data, design = NULL, method = "raw", ...)
}
\arguments{
\item{data}{Matrix of count data.}

\item{design}{Dataframe describing the experimental
design. (conditions/batches/etc)}

\item{...}{More arguments might be necessary.}

\item{norm}{Normalization to perform:
'sf|quant|qsmooth|tmm|upperquartile|tmm|rle' I keep wishy-washing on
whether design is a required argument.}
}
\value{
Dataframe of normalized(counts)
}
\description{
This provides shortcut interfaces for normalization functions from
deseq2/edger and friends.
}
\examples{
\dontrun{
 norm_table = normalize_counts(count_table, design = design, norm='qsmooth')
}
}
\seealso{
\pkg{edgeR} \pkg{limma} \pkg{DESeq2}
}
