% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{normalize_todo}
\alias{normalize_todo}
\title{Simplified and ideally improved normalization function}
\usage{
normalize_todo(exp, todo = list())
}
\arguments{
\item{exp}{Input data}

\item{todo}{List of tasks to perform.}
}
\description{
This function is ideally should provide a simpler and more capable
version of normalize. I also want to move everything to using
summarizedExperiments and this simpler method provides an
opportunity.
}
