% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{notes,expt-method}
\alias{notes,expt-method}
\title{A getter to pull the notes an expt.}
\usage{
\S4method{notes}{expt}(object)
}
\arguments{
\item{object}{An expt.}
}
\description{
A getter to pull the notes an expt.
}
