% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\docType{methods}
\name{notes}
\alias{notes}
\alias{notes,}
\alias{notes-methods}
\title{Extend Biobase::notes to handle expt objects.}
\usage{
\S4method{notes}{expt}(object)
}
\arguments{
\item{expt}{The expt object from which to extract the expressionset.}
}
\description{
Extend Biobase::notes to handle expt objects.
}
