% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{orgdb_from_ah}
\alias{orgdb_from_ah}
\title{Get an orgdb from an AnnotationHub taxonID.}
\usage{
orgdb_from_ah(ahid = NULL, title = NULL, species = NULL,
  type = "OrgDb")
}
\arguments{
\item{ahid}{TaxonID from AnnotationHub}

\item{title}{Title for the annotation hub instance}

\item{species}{Species to download}

\item{type}{Datatype to download}
}
\value{
An Orgdb instance
}
\description{
Ideally, annotationhub will one day provide a one-stop shopping source for a
tremendous wealth of curated annotation databases, sort of like a
non-obnoxious biomart.  But for the moment, this function is more
fragile than I would like.
}
\examples{
\dontrun{
 org <- mytaxIdToOrgDb(species="Leishmania", type="TxDb")
}
}
\seealso{
\pkg{AnnotationHub} \pkg{S4Vectors}
}
