% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb_organdb.R
\name{orgdb_gene_types}
\alias{orgdb_gene_types}
\title{EuPathDB gene information table gene type parser}
\usage{
orgdb_gene_types(filepath)
}
\arguments{
\item{filepath}{Location of TriTrypDB gene information table.}
}
\value{
Returns a dataframe mapping gene ids to gene types
}
\description{
EuPathDB gene information table gene type parser
}
\author{
Keith Hughitt
}

