% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{orgdb_idmap}
\alias{orgdb_idmap}
\title{Load organism annotation data (mouse/human).}
\usage{
orgdb_idmap(orgdb, gene_ids = NULL, mapto = c("ENSEMBL"))
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Gene identifiers for retrieving annotations.}

\item{keytype}{a, umm keytype? I need to properly read this code.}

\item{fields}{Columns to include in the output.}

\item{biomart_dataset}{Name of the biomaRt dataset to query for gene type.}
}
\value{
a table of gene information
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the OrganismDbi interface.
}
\examples{
\dontrun{
host <- load_host_annotations(org, c("a","b"))
}
}
\seealso{
\link[AnnotationDbi]{select}
}

