% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{orgdb_info_from_gff}
\alias{orgdb_info_from_gff}
\title{Extract gene information from a GFF file}
\usage{
orgdb_info_from_gff(gff)
}
\arguments{
\item{gff}{GenomicRanges instance as returned by `import.gff3`}

\item{return}{data.frame containing basic gene information (ID, description,
etc.)}
}
\description{
Extract gene information from a GFF file
}
\author{
Keith Hughitt
}
