% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{orgdb_match_keytypes}
\alias{orgdb_match_keytypes}
\title{Generate a set of joins suitable for the creation of an organismdbi package}
\usage{
orgdb_match_keytypes(first_name, second_name, starting = 1, exclude = NULL)
}
\arguments{
\item{first_name}{Name of the first package to search}

\item{second_name}{Name of the second package to search}

\item{starting}{What number join to start from}
}
\value{
A list named join# where the number is the nth join discovered and the elements
  are non-zero matches between the sqlite packages described by first_name and second_name.
}
\description{
The graph data required in an organismdbi is pretty specific, this function creates it!
It does so by iterating through all keytype pairs between the two packages and looking for
matching keys, whichever keys have the most matches win.  It is therefore rather slow.
}
