% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{overlap_geneids}
\alias{overlap_geneids}
\title{Mostly as a reminder of how to get the gene IDs from a specific group in an upset plot.}
\usage{
overlap_geneids(overlapping_groups, group)
}
\arguments{
\item{overlapping_groups}{Result from overlap_groups, which just makes an indexed
version of the genes by venn/upset group.}

\item{group}{Name of the subset of interest, something like 'a:b' for the union of a:b.}
}
\description{
Given a set of groups from upsetr, extract the elements from one of them.
}
