% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{overlap_groups}
\alias{overlap_groups}
\title{Extract overlapping groups from an upset}
\usage{
overlap_groups(input, sort = TRUE)
}
\arguments{
\item{sort}{Sort the result?}

\item{lst}{upset data structure.}
}
\description{
Taken from: https://github.com/hms-dbmi/UpSetR/issues/85
and lightly modified to match my style and so I could more
easily understand what it is doing.
}
