% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{pa_counts}
\alias{pa_counts}
\title{The count directory from the Pseudomonas experiment subset.}
\format{
An object of class \code{character} of length 1.
}
\usage{
pa_counts
}
\description{
In other examples I include an archive file of the counts, here are the raw
tables.  These were created via bowtie2 -> samtools -> htseq-count.
}
\keyword{datasets}
