% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{pa_fasta}
\alias{pa_fasta}
\title{PA14 genome}
\format{
An object of class \code{character} of length 1.
}
\usage{
pa_fasta
}
\description{
This is the genome which corresponds to pa_gff.
}
\keyword{datasets}
