% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{pa_gb}
\alias{pa_gb}
\title{PA14 genbank flat file}
\format{
An object of class \code{character} of length 1.
}
\usage{
pa_gb
}
\description{
This may be used instead to extract genomic information and annotations via
load_genbank_annotations().
}
\keyword{datasets}
