% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{pa_gff}
\alias{pa_gff}
\title{PA14 annotations in a GFF file}
\format{
An object of class \code{character} of length 1.
}
\usage{
pa_gff
}
\description{
Along with the sample sheet, this provides everything required to make an
expressionsheet via create_expt() for our PA 14 experiment.
}
\keyword{datasets}
