% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb_organdb.r
\name{parse_go_terms}
\alias{parse_go_terms}
\title{EuPathDB gene information table GO term parser}
\usage{
parse_go_terms(filepath)
}
\arguments{
\item{filepath}{Location of TriTrypDB gene information table.}

\item{verbose}{Whether or not to enable verbose output.}
}
\value{
Returns a dataframe where each line includes a gene/GO terms pair
        along with some addition information about the GO term. Note that
        because each gene may have multiple GO terms, a single gene ID may
        appear on multiple lines.
}
\description{
Note: EuPathDB currently includes some GO annotations corresponding to
obsolete terms. For example, the L. major gene LmjF.19.1390
(http://tritrypdb.org/tritrypdb/showRecord.do?name=GeneRecordClasses.GeneRecordClass&source_id=LmjF.19.1390&project_id=TriTrypDB)
includes the term "GO:0003702" on the website and txt data file. The
term has been deprecated, and does not have a category associated with it
on the website. These will not be included in the final database.
}
\author{
Keith Hughitt
}

