% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_tritrypdb.r
\name{parse_interpro_domains}
\alias{parse_interpro_domains}
\title{EuPathDB gene information table InterPro domain parser}
\usage{
parse_interpro_domains(filepath)
}
\arguments{
\item{filepath}{Location of TriTrypDB gene information table.}
}
\value{
Returns a dataframe where each line includes a gene/domain pairs.
}
\description{
EuPathDB gene information table InterPro domain parser
}
\author{
Keith Hughitt
}
