% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{parse_msigdb}
\alias{parse_msigdb}
\title{Parse either xml or sqlite data from MSigDB.  I think I will likely remove the
xml version as I think the msigdb xml files are poorly formatted.}
\usage{
parse_msigdb(filename)
}
\arguments{
\item{filename}{Input file}
}
\description{
Parse either xml or sqlite data from MSigDB.  I think I will likely remove the
xml version as I think the msigdb xml files are poorly formatted.
}
