% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{parse_msigdb_sqlite}
\alias{parse_msigdb_sqlite}
\title{Extract metadata from the msigdb sqlite data.}
\usage{
parse_msigdb_sqlite(filename)
}
\arguments{
\item{filename}{db file.}
}
\description{
Extract metadata from the msigdb sqlite data.
}
