% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{parse_msigdb_xml}
\alias{parse_msigdb_xml}
\title{Extract fun experimental metadata from a MSigDB xml file.}
\usage{
parse_msigdb_xml(filename)
}
\arguments{
\item{filename}{input file.}
}
\description{
Extract fun experimental metadata from a MSigDB xml file.
}
