% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_gff.r
\name{pattern_count_genome}
\alias{pattern_count_genome}
\title{Find how many times a given pattern occurs in every gene of a genome.}
\usage{
pattern_count_genome(fasta, gff = NULL, pattern = "TA",
  type = "gene", key = NULL)
}
\arguments{
\item{fasta}{Genome sequence.}

\item{gff}{Gff of annotation information from which to acquire CDS (if not
provided it will just query the entire genome).}

\item{pattern}{What to search for? This was used for tnseq and TA is the
mariner insertion point.}

\item{type}{Column to use in the gff file.}

\item{key}{What type of entry of the gff file to key from?}
}
\value{
Data frame of gene names and number of times the pattern appears/gene.
}
\description{
There are times when knowing how many times a given string appears in a
genome/CDS is helpful. This function provides that information and is
primarily used by cp_seq_m().
}
\details{
This is once again a place where mcols() usage might improve the overall
quality of life.
}
\examples{
\dontrun{
 num_pattern <- pattern_count_genome('mgas_5005.fasta', 'mgas_5005.gff')
}
}
\seealso{
\pkg{Biostrings} \pkg{Rsamtools} \pkg{Rsamtools}
 \code{\link[Rsamtools]{FaFile}} \code{\link[Biostrings]{getSeq}}
 \code{\link[Biostrings]{PDict}} \code{\link[Biostrings]{vcountPDict}}
}
\author{
atb
}
