% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{pc_fstatistics}
\alias{pc_fstatistics}
\title{A sister function to sv_fstatistics()}
\usage{
pc_fstatistics(
  expt,
  pc_df = NULL,
  num_pcs = 10,
  queries = c("typeofcells", "visitnumber", "donor"),
  ...
)
}
\arguments{
\item{expt}{Input expressionset, redo everything to use SE, stupid.}

\item{pc_df}{dataframe of PCs}

\item{num_pcs}{How many PCs to query?}

\item{queries}{List of metadata factors to query.}

\item{...}{Parameters to plot_pca.}
}
\description{
TODO: Use this to yank a bunch of code out of pca_information and simplify.
}
\details{
Calculate f-statistics between metadata factors and principal components.

Random note to self: principle or principal?  I can _NEVER_
remember; also I am just pasting my docstring from the
sv_fstatistics for the moment.  I think the key observation: this
might actually be Hector's idea to both Theresa and I in different
contexts.

This is taken directly from Theresa's TMRC work and is her idea.  I
mainly want to be able to use it on a few datasets without risking
typeographical or logical errors.  In addition, I would like to be
able to play with things like the number of surrogates and/or other
methods of estimating them.  In addition, I have some f-statistics
of PCs vs metadata in the function 'pca_information().' which I
think is likely complementary to her work (which makes sense,
Hector was her professor before she joined us, and Hector suggested
the PC idea to me).
}
