% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{pc_vs_factor}
\alias{pc_vs_factor}
\title{Plot n principle components sequentially vs. factors in the experiment}
\usage{
pc_vs_factor()
}
\arguments{
\item{expt}{an experiment}

\item{n}{number of principle components to use}

\item{factor}{name(s) of experimental factor(s) to compare against}
}
\value{
a bunch of plots and some sort of summary.
}
\description{
The goal of this should be to try to identify principle components which
change with respect to some other known (biological) factor
}
\examples{
\dontrun{
 bob = pc_vs_factor(stuff)
}
}
\seealso{
\code{\link[stats]{princomp}}
}

