% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_highscores}
\alias{pca_highscores}
\title{pca_highscores(): Get the highest/lowest scoring genes for every principle component}
\usage{
pca_highscores(df = NULL, conditions = NULL, batches = NULL, n = 20)
}
\arguments{
\item{df}{A dataframe of (pseudo)counts}

\item{conditions}{A factor or character of conditions in the
experiment}

\item{batches}{a factor or character of batches in the experiment}

\item{n}{(20) the number of genes to extract}
}
\value{
a list including the princomp biplot, histogram, and tables
of top/bottom n scored genes with their scores by component.
}
\description{
This function uses princomp to acquire a principle component biplot
for some data and extracts a dataframe of the top n genes for each
component by score.
}
\examples{
## information = pca_highscores(df=df, conditions=cond, batches=bat)
## information$pca_bitplot  ## oo pretty
}
\seealso{
\code{\link{princomp}}
}

