% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{pca_information,data.frame-method}
\alias{pca_information,data.frame-method}
\title{If the pieces used for pca_information are provided as separate
data; dispatch accordingly.}
\usage{
\S4method{pca_information}{data.frame}(
  exp,
  factors = c("condition", "batch"),
  num_components = NULL,
  plot_pcas = FALSE,
  ...
)
}
\arguments{
\item{num_components}{Number of principle components to compare the design
factors against. If left null, it will query the same number of components
as factors asked for.}

\item{plot_pcas}{Plot the set of PCA plots for every pair of PCs queried.}

\item{...}{Extra arguments for the pca plotter}
}
\description{
If the pieces used for pca_information are provided as separate
data; dispatch accordingly.
}
