% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{pca_information,ExpressionSet-method}
\alias{pca_information,ExpressionSet-method}
\title{If pca_information is run on an expressionset, everything except
perhaps colors should already be there.}
\usage{
\S4method{pca_information}{ExpressionSet}(
  input,
  input_design = NULL,
  input_factors = c("condition", "batch"),
  colors_chosen = NULL,
  num_components = NULL,
  plot_pcas = FALSE,
  ...
)
}
\arguments{
\item{input}{Data to analyze (usually exprs(somedataset)).}

\item{input_design}{Dataframe describing the experimental design, containing
columns with useful information like the conditions, batches, number of
cells, whatever...}

\item{input_factors}{Character list of experimental conditions to query for
R^2 against the fast.svd of the data.}

\item{colors_chosen}{Colors to use when comparing fstats etc.}

\item{num_components}{Number of principle components to compare the design
factors against. If left null, it will query the same number of components
as factors asked for.}

\item{plot_pcas}{Plot the set of PCA plots for every pair of PCs queried.}

\item{...}{Extra arguments for the pca plotter}
}
\description{
If pca_information is run on an expressionset, everything except
perhaps colors should already be there.
}
