% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{pca_plot_smallbatch}
\alias{pca_plot_smallbatch}
\title{pca_plot_smallbatch()  ggplot2 plots of PCA data with <= 5 batches.}
\usage{
pca_plot_smallbatch(df, size = 5, first = "PC1", second = "PC2")
}
\arguments{
\item{df}{A dataframe of PC1/PC2 and other arbitrary data.}

\item{size}{default=5  The size of glyphs in the plot.}
}
\value{
a ggplot2 plot of principle components 1 and 2.
}
\description{
This uses hard-coded scale_shape_manual values 21-25 to have solid shapes in the plot.
}

