% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{pct_kegg_diff}
\alias{pct_kegg_diff}
\title{Extract the percent differentially expressed genes in a given KEGG pathway.}
\usage{
pct_kegg_diff(all_ids, sig_ids, pathway = "00500", organism = "dme",
  pathdir = "kegg_pathways", ...)
}
\arguments{
\item{all_ids}{Set of all gene IDs in a given analysis.}

\item{sig_ids}{Set of significant gene IDs.}

\item{pathway}{Numeric pathway identifier.}

\item{organism}{KEGG organism identifier.}

\item{pathdir}{Directory into which to copy downloaded pathway files.}

\item{...}{Options I might pass from other functions are dropped into arglist.}
}
\value{
Percent genes/pathway deemed significant.
}
\description{
KEGGgraph provides some interesting functionality for mapping KEGGids and examining the
pieces. This attempts to use that in order to evaluate how many 'significant' genes are in a
given pathway.
}
\seealso{
\pkg{KEGGgraph} \pkg{KEGGREST}
}

