% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{pData,expt-method}
\alias{pData,expt-method}
\title{A getter to pull the experimental metadata from an expt.}
\usage{
\S4method{pData}{expt}(object)
}
\arguments{
\item{object}{An expt.}
}
\description{
A getter to pull the experimental metadata from an expt.
}
