% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\docType{methods}
\name{pData}
\alias{pData}
\title{Extend Biobase::pData to handle expt objects.}
\usage{
\S4method{pData}{expt}(object)
}
\arguments{
\item{object}{The expt object from which to extract the expressionset.}
}
\description{
Extend Biobase::pData to handle expt objects.
}
