% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{pData<-,SummarizedExperiment,ANY-method}
\alias{pData<-,SummarizedExperiment,ANY-method}
\title{A setter to put the experimental metadata into a SummarizedExperiment.}
\usage{
\S4method{pData}{SummarizedExperiment,ANY}(object) <- value
}
\arguments{
\item{object}{An expt.}
}
\description{
This is essentially synonymous with colData, except I cannot seem
to remember that function when I am working; so I just added
another signature to pData.
}
