% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{pData<-,SummarizedExperiment-method}
\alias{pData<-,SummarizedExperiment-method}
\title{A setter to put the experimental metadata into a SummarizedExperiment.}
\usage{
\S4method{pData}{SummarizedExperiment}(object) <- value
}
\arguments{
\item{object}{An expt.}

\item{value}{New metadata.}
}
\description{
This is essentially synonymous with colData, except I cannot seem
to remember that function when I am working; so I just added
another signature to pData.
}
