% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{pData,SummarizedExperiment-method}
\alias{pData,SummarizedExperiment-method}
\title{A getter to pull the experimental metadata from a SummarizedExperiment.}
\usage{
\S4method{pData}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{An expt.}
}
\description{
This is essentially synonymous with colData, except I cannot seem
to remember that function when I am working; so I just added
another signature to pData.
}
