% Generated by roxygen2 (4.0.1): do not edit by hand
\name{perform_comparison}
\alias{perform_comparison}
\title{Perform a quick and dirty differential expression comparison}
\usage{
perform_comparison(expt_subset, model, time = "LL", media = "THY",
  new_media = NULL, mga = "wt", basename = "testing")
}
\arguments{
\item{expt_subset}{an experiment subset including normalized
samples and an experimental design}

\item{model}{a model matrix defining what to compare}
}
\value{
a list including the voomed data, fitted data by lmFit,
contrasts from limma, some histograms of the data, a table of the
differentially expressed data from toptable.  Along the way this
attempts to print some hopefully useful diagnostic information and
write the relevant tables to an excel file.  In its current
implementation it assumes a specific experiment, which is bad.  In
addition it must perforce assume that the first provided condition is a
control and the second is an experimental.  It changes the columns
accordingly.
}
\description{
Perform a quick and dirty differential expression comparison
}
\examples{
## de_information = perform_comparison(expt_subset, model)
}
\seealso{
\code{\link{voom}}, \code{\link{lmFit}},
\code{\link{makeContrasts}}, \code{\link{contrasts.fit}},
\code{\link{eBayes}}, \code{\link{topTable}},
\code{\link{my_linear_scatter}}, \code{\link{write_xls}},
\code{\link{my_ma_plot}}
}

