% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{pkg_cleaner}
\alias{pkg_cleaner}
\title{Packages generated by make_organismdbi(), make_orgdb(), and make_txdb() are entirely too
fragile.  This attempts to fix some of the common problems therein.}
\usage{
pkg_cleaner(path, removal = "-like", replace = "")
}
\arguments{
\item{path}{Location for the original Db/Dbi instance.}

\item{removal}{String to remove from the instance.}

\item{replace}{What to replace removal with, when necessary.}
}
\value{
A new OrgDb/TxDb/OrganismDbi
}
\description{
OrganismDbi instances are pretty neat, they pull together OrgDb and TxDb.  With any luck, this
function provides the ability to pull together all the data from the TriTrypDb, GO.db, and
KEGGREST in order to accomplish these peculiar tasks.
}
\examples{
\dontrun{
 crazytown <- make_organismdbi()  ## wait a loong time
}
}

