% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.r
\name{plot_3d_pca}
\alias{plot_3d_pca}
\title{Something silly for Najib.}
\usage{
plot_3d_pca(pc_result, components = c(1, 2, 3), file = "3dpca.html")
}
\arguments{
\item{pc_result}{The result from plot_pca()}

\item{components}{List of three axes by component.}

\item{file}{File to write the created plotly object.}
}
\value{
List containing the plotly data and filename for the html widget.
}
\description{
This will make him very happy, but I remain skeptical.
}
\seealso{
[plotly] [htmlwidgets]
}
