% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_assay_by_chromosome}
\alias{plot_assay_by_chromosome}
\title{Create a plot showing relative expression with respect to each chromosome/contig.}
\usage{
plot_assay_by_chromosome(
  exp,
  chromosome_column = "chromosome",
  scaffolds = TRUE,
  min_genes = 10
)
}
\arguments{
\item{exp}{Input expressionset.}

\item{chromosome_column}{Annotation column containing the chromosome ID.}

\item{scaffolds}{Include scaffolds in addition to the actual chromosomes.}

\item{min_genes}{The minimum number of genes which should be on the 'chromosome' before it
is considered worth considering.}
}
\description{
Create a plot showing relative expression with respect to each chromosome/contig.
}
