% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.r
\name{plot_batchsv}
\alias{plot_batchsv}
\title{Make a dotplot of known batches vs. SVs.}
\usage{
plot_batchsv(expt, svs, batch_column = "batch", factor_type = "factor")
}
\arguments{
\item{expt}{Experiment from which to acquire the design, counts, etc.}

\item{svs}{Set of surrogate variable estimations from sva/svg
or batch estimates.}

\item{batch_column}{Which experimental design column to use?}

\item{factor_type}{This may be a factor or range, it is intended to plot
a scatterplot if it is a range, a dotplot if a factor.}
}
\value{
Plot of batch vs surrogate variables as per Leek's work.
}
\description{
This should make a quick df of the factors and surrogates and plot them.
Maybe it should be folded into plot_svfactor?  Hmm, I think first I will
write this and see if it is better.
}
\examples{
\dontrun{
 estimate_vs_snps <- plot_batchsv(start, surrogate_estimate, "snpcategory")
}
}
\seealso{
\pkg{sva} \pkg{ggplot2}
}
