% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_boxplot,SummarizedExperiment-method}
\alias{plot_boxplot,SummarizedExperiment-method}
\title{Make a ggplot boxplot of a set of samples.}
\usage{
\S4method{plot_boxplot}{SummarizedExperiment}(
  data,
  colors = NULL,
  plot_title = NULL,
  order = NULL,
  violin = FALSE,
  scale = NULL,
  sample_names = NULL,
  label_chars = 10,
  ...
)
}
\arguments{
\item{data}{Exp or data frame set of samples.}

\item{colors}{Color scheme, if not provided will make its own.}

\item{plot_title}{A title!}

\item{order}{Set the order of boxen.}

\item{violin}{Print this as a violin rather than a just box/whiskers?}

\item{scale}{Whether to log scale the y-axis.}

\item{sample_names}{Another version of the sample names for printing.}

\item{label_chars}{Maximum number of characters for abbreviating sample names.}

\item{...}{More parameters are more fun!}
}
\description{
Boxplots and density plots provide complementary views of data distributions.
The general idea is that if the box for one sample is significantly shifted
from the others, then it is likely an outlier in the same way a density plot
shifted is an outlier.
}
