% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Generic plot_boxplot dispatch}
\usage{
plot_boxplot(data, ...)
}
\arguments{
\item{data}{data of indeterminate type}

\item{...}{Other arguments}
}
\description{
Generic plot_boxplot dispatch
}
