% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{plot_corheat}
\alias{plot_corheat}
\title{Make a heatmap.3 description of the correlation between samples.}
\usage{
plot_corheat(
  expt_data,
  expt_colors = NULL,
  expt_design = NULL,
  method = "pearson",
  expt_names = NULL,
  batch_row = "batch",
  title = NULL,
  label_chars = 10,
  ...
)
}
\arguments{
\item{expt_data}{Dataframe, expt, or expressionset to work with.}

\item{expt_colors}{Color scheme for the samples, not needed if this is an expt.}

\item{expt_design}{Design matrix describing the experiment, not needed if this is an expt.}

\item{method}{Correlation statistic to use. (pearson, spearman, kendall, robust).}

\item{expt_names}{Alternate names to use for the samples.}

\item{batch_row}{Name of the design row used for 'batch' column colors.}

\item{title}{Title for the plot.}

\item{label_chars}{Limit on the number of label characters.}

\item{...}{More options are wonderful!}
}
\value{
Gplots heatmap describing describing how the samples are clustering
  vis a vis pairwise correlation.
}
\description{
Given a set of count tables and design, this will calculate the pairwise
correlations and plot them as a heatmap.  It attempts to standardize the
inputs and eventual output.
}
\examples{
\dontrun{
 corheat_plot <- hpgl_corheat(expt=expt, method="robust")
}
}
\seealso{
\pkg{grDevice}
 \code{\link{hpgl_cor}} \code{\link[RColorBrewer]{brewer.pal}}
  \code{\link[grDevices]{recordPlot}}
}
