% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.r
\name{plot_de_pvals}
\alias{plot_de_pvals}
\title{Given a DE table with p-values, plot them.}
\usage{
plot_de_pvals(combined, type = "limma", columns = NULL)
}
\arguments{
\item{combined}{Table to extract the values from.}

\item{type}{If provided, extract the {type}_p and {type}_adjp columns.}

\item{columns}{Otherwise, extract whatever columns are provided.}
}
\value{
Multihistogram of the result.
}
\description{
Plot a multi-histogram containing (adjusted)p-values.
}
