% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_density}
\alias{plot_density}
\title{Generic dispatch of plot_density when the input is unknown.}
\usage{
plot_density(data, ...)
}
\arguments{
\item{data}{Arbitrary data passed to plot_density.}

\item{...}{Other arguments}
}
\description{
Generic dispatch of plot_density when the input is unknown.
}
