% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{plot_disheat}
\alias{plot_disheat}
\title{Make a heatmap.3 description of the distances (euclidean by default) between samples.}
\usage{
plot_disheat(expt_data, expt_colors = NULL, expt_design = NULL,
  method = "euclidean", expt_names = NULL, batch_row = "batch",
  title = NULL, ...)
}
\arguments{
\item{expt_data}{Dataframe, expt, or expressionset to work with.}

\item{expt_colors}{Color scheme (not needed if an expt is provided).}

\item{expt_design}{Design matrix (not needed if an expt is provided).}

\item{method}{Distance metric to use.}

\item{expt_names}{Alternate names to use for the samples.}

\item{batch_row}{Name of the design row used for 'batch' column colors.}

\item{title}{Title for the plot.}

\item{...}{More parameters!}
}
\value{
a recordPlot() heatmap describing the distance between samples.
}
\description{
Given a set of count tables and design, this will calculate the pairwise distances and plot
them as a heatmap.  It attempts to standardize the inputs and eventual output.
}
\examples{
\dontrun{
 disheat_plot = plot_disheat(expt=expt, method="euclidean")
 disheat_plot
}
}
\seealso{
\link[RColorBrewer]{brewer.pal} \link[gplots]{heatmap.2} \link[grDevices]{recordPlot}
}

