% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.R
\name{plot_enrichresult}
\alias{plot_enrichresult}
\title{Invoke ther various fun plots created by Guangchuang Yu.}
\usage{
plot_enrichresult(enrichresult)
}
\arguments{
\item{enrichresult}{S4 object of type enrichResult.}
}
\description{
I would like to replace all of my bad ontology plotting functions
with the nicer versions from enrichplot.  I therefore have a series
of functions which recast my ontology results to enrichResults,
which is suitable for those plots.
}
\details{
For the moment this is just a skeleton with reminders to me for the
various plots available.  Also, when I looked up these plots it
appears that clusterProfiler has some new functionality to make it
easier to send results to it.
}
