% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnseq.R
\name{plot_essentiality}
\alias{plot_essentiality}
\title{plot_essentiality()  Plot the essentiality of a library as per DeJesus et al.}
\usage{
plot_essentiality(file)
}
\arguments{
\item{file}{a file created using the perl script 'essentiality_tas.pl'}
}
\value{
A couple of plots
}
\description{
plot_essentiality()  Plot the essentiality of a library as per DeJesus et al.
}

