% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{plot_forest_from_regression}
\alias{plot_forest_from_regression}
\title{Given the result from one of the regression testers, plot it!}
\usage{
plot_forest_from_regression(
  df,
  percent = 95,
  type = "logistic",
  iterative = TRUE,
  family = "binomial"
)
}
\arguments{
\item{df}{The primary dataframe from one of the sister regression functions above.}

\item{percent}{Confidence interval chosen.}

\item{type}{Either linear or logistic.}

\item{iterative}{Was this a series of single-variable regressions, or all in one?}

\item{family}{Only currently used for logistic.}
}
\description{
Given the result from one of the regression testers, plot it!
}
