% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.r
\name{plot_fun_venn}
\alias{plot_fun_venn}
\title{A quick wrapper around venneuler to help label stuff}
\usage{
plot_fun_venn(ones = c(), twos = c(), threes = c(), fours = c(),
  fives = c(), factor = 0.9)
}
\arguments{
\item{ones}{Character list of singletone categories}

\item{twos}{Character list of doubletone categories}

\item{threes}{Character list of tripletone categories}

\item{fours}{Character list of quad categories}

\item{fives}{Character list of quint categories}

\item{factor}{Currently unused, but intended to change the radial distance to the label from the center of each circle.}
}
\value{
Two element list containing the venneuler data and the plot.
}
\description{
venneuler makes pretty venn diagrams, but no labels!
}

