% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{plot_ggbio_heatmap}
\alias{plot_ggbio_heatmap}
\title{Use ggbio to create a heatmap of a grange dataset.}
\usage{
plot_ggbio_heatmap(grange_data, tbl, tilesize = 1000, scaffolds = FALSE)
}
\arguments{
\item{grange_data}{Input grange with starts/stops/etc}

\item{tbl}{Table of locations of interest}

\item{tilesize}{Bin the genome to this length}

\item{scaffolds}{Include scaffolds?}
}
\description{
Use ggbio to create a heatmap of a grange dataset.
}
