% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{plot_ggcoverage}
\alias{plot_ggcoverage}
\title{Plot coverage from a grange using ggcoverage}
\usage{
plot_ggcoverage(
  gr,
  from,
  to,
  id_column,
  coverage_files,
  padding = 100,
  span = 0.03,
  smoothed = TRUE,
  ribbon = FALSE,
  cov_color = "black",
  y_max = NULL
)
}
\arguments{
\item{gr}{input granges}

\item{from}{starting point for the plot}

\item{to}{ending point}

\item{id_column}{mcols() portion of the granges containing the gene names}

\item{coverage_files}{Set of bigwig file used to define the coverage.}

\item{padding}{not currently used, but add some padding to the plot}

\item{span}{degree of smoothing}

\item{smoothed}{Smooth the coverage plot?}

\item{ribbon}{fill in the smoothed data?}

\item{cov_color}{Color for the coverage region}

\item{y_max}{Explicitly set the maximum of the y axis.}
}
\description{
ggcoverage makes a bunch of assumptions about how the input gff is formatted.
}
