% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{plot_ggcoverage_se}
\alias{plot_ggcoverage_se}
\title{Given a SE, plot a coverage map with some helpful defaults and workarounds.}
\usage{
plot_ggcoverage_se(
  se,
  from = 1,
  to = 10,
  id_column = "gene_id",
  coverage_column = "deeptools_coverage",
  convert = "cpm",
  norm = NULL,
  transform = NULL,
  padding = c(1, 2),
  span = 0.03,
  smoothed = TRUE,
  ribbon = FALSE,
  cov_color = NULL,
  y_max = NULL,
  bin_size = 10,
  hide_bars = FALSE,
  feature_type_column = "type",
  feature_type = "gene",
  meta_type_column = "media",
  feature_name_column = "Parent",
  gene_name = NULL,
  meta_group_column = "condition",
  overlap_type = "tight",
  overlap_gap = 0.1,
  arrow_type = "closed",
  arrow_gap = NULL,
  label_column = NULL,
  bg_color = "white",
  gene_color_by = "strand",
  gene_colors = NULL,
  cores = NULL,
  facet = TRUE,
  y_scale = NULL
)
}
\arguments{
\item{se}{Input summarized experiment, it should have a full genome
GRanges and colors}

\item{from}{Starting gene region's name or number.}

\item{to}{Ending gene region's name or number.}

\item{id_column}{Column of the rowData containing the gene IDs.}

\item{coverage_column}{Column of the colData containing the
location of the coverage files in bigwig format.}

\item{convert}{Perform a cpmish conversion of the data?}

\item{norm}{Normalize the data?}

\item{transform}{log transform the data?}

\item{padding}{Add this amount of sequence before/after the
granges.}

\item{span}{When smoothing, adjust the loess transformation.}

\item{smoothed}{Smooth the coverage data? (currently only loess)}

\item{ribbon}{Add a ribbon to the smoothing (currently disabled)}

\item{cov_color}{Specify colors for the coverage lines}

\item{y_max}{Specify a y-maximum}

\item{bin_size}{Specify the coverage bin size}

\item{hide_bars}{Hide the coverage bars when smoothing the data?}

\item{feature_type_column}{Use this rowData column to extract the
feature types.}

\item{feature_type}{Use this type for creating the GR}

\item{meta_type_column}{Use this colData column to get the
ggcoverage Type column}

\item{feature_name_column}{Use this rowData column to fill in the
gene names below the gene arrows.}

\item{gene_name}{Choose a gene?}

\item{meta_group_column}{Use this colData column to group samples.}

\item{overlap_type}{Define the behavior of the arrows/row}

\item{overlap_gap}{Modify the arrow overlaps with this.}

\item{arrow_type}{Either closed or open}

\item{arrow_gap}{Not sure what this does.}

\item{label_column}{FIXME: perhaps redundant with feature_name_column?}

\item{bg_color}{Define the plot background color}

\item{gene_color_by}{Play with the gene arrow colors (I want to
modify ggcoverage to allow me to color by FC status)}

\item{gene_colors}{specify factor of gene colors.}
}
\description{
ggcoverage is really cool, but has some hard-coded values which are
not useful. This function seeks to overcome some of these
limitations and fill in some default values.
}
