% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/107_ontology_plots.r
\name{plot_gostats_pval}
\alias{plot_gostats_pval}
\title{Make a pvalue plot similar to that from clusterprofiler from gostats data.}
\usage{
plot_gostats_pval(
  gs_result,
  wrapped_width = 20,
  cutoff = 0.1,
  n = 30,
  group_minsize = 5
)
}
\arguments{
\item{gs_result}{Ontology search results.}

\item{wrapped_width}{Make the text large enough to read.}

\item{cutoff}{What is the maximum pvalue allowed?}

\item{n}{How many groups to include in the plot?}

\item{group_minsize}{Minimum group size before inclusion.}
}
\value{
Plots!
}
\description{
clusterprofiler provides beautiful plots describing significantly
overrepresented categories. This function attempts to expand the repetoire of
data available to them to include data from gostats. The pval_plot function
upon which this is based now has a bunch of new helpers now that I understand
how the ontology trees work better, this should take advantage of that, but
currently does not.
}
\seealso{
\pkg{clusterProfiler}
 \code{\link{plot_ontpval}}
}
