% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.r
\name{plot_gprofiler_pval}
\alias{plot_gprofiler_pval}
\title{Make a pvalue plot from gprofiler data.}
\usage{
plot_gprofiler_pval(gp_result, wrapped_width = 20, cutoff = 0.1, n = 12,
  group_minsize = 5)
}
\arguments{
\item{wrapped_width}{Maximum width of the text names.}

\item{cutoff}{P-value cutoff for the plots.}

\item{n}{Maximum number of ontologies to include.}

\item{topgo}{Some data from topgo!}

\item{type}{Type of score to use.}
}
\value{
List of MF/BP/CC pvalue plots.
}
\description{
The p-value plots from clusterProfiler are pretty, this sets the gprofiler data into a format
suitable for plotting in that fashion and returns the resulting plots of significant ontologies.
}
\seealso{
\pkg{topgo} \code{clusterProfiler}
}

