% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gvis.r
\name{plot_gvis_scatter}
\alias{plot_gvis_scatter}
\title{Make an html version of a scatter plot.}
\usage{
plot_gvis_scatter(df, tooltip_data = NULL,
  filename = "html/gvis_scatter.html", base_url = "",
  trendline = NULL)
}
\arguments{
\item{df}{Df of two columns to compare.}

\item{tooltip_data}{Df of tooltip information for gvis graphs.}

\item{filename}{Filename to write a fancy html graph.}

\item{base_url}{Url to send click events which will be suffixed with the gene name.}

\item{trendline}{Add a trendline?}
}
\value{
NULL, but along the way an html file is generated which contains a
  googleVis scatter plot.  See plot_scatter() for details.
}
\description{
Given an arbitrary scatter plot, we can make it pretty and javascript-tacular
using this function.
}
\examples{
\dontrun{
 gvis_scatter(a_dataframe_twocolumns, filename="html/fun_scatter_plot.html",
              base_url="http://yeastgenome.org/accession?")
}
}
\seealso{
\pkg{googleVis}
 \code{\link[googleVis]{gvisScatterChart}}
}
