% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gvis.r
\name{plot_gvis_volcano}
\alias{plot_gvis_volcano}
\title{Make an html version of an volcano plot.}
\usage{
plot_gvis_volcano(toptable_data, fc_cutoff = 0.8, p_cutoff = 0.05,
  tooltip_data = NULL, filename = "html/gvis_vol_plot.html",
  base_url = "", ...)
}
\arguments{
\item{toptable_data}{Df of toptable() data.}

\item{fc_cutoff}{Fold change cutoff.}

\item{p_cutoff}{Maximum p value to allow.}

\item{tooltip_data}{Df of tooltip information.}

\item{filename}{Filename to write a fancy html graph.}

\item{base_url}{String with a basename used for generating URLs for clicking dots on the graph.}

\item{...}{more options}
}
\value{
NULL, but along the way an html file is generated which contains a googleVis volcano
 plot.
}
\description{
Volcano plots provide some visual clues regarding the success of a given contrast.  For our data,
it has the -log10(pvalue) on the y-axis and fold-change on the x.  Here is a neat snippet from
wikipedia describing them generally: "The concept of volcano plot can be generalized to other
applications, where the x-axis is related to a measure of the strength of a statistical signal,
and y-axis is related to a measure of the statistical significance of the signal."
}
\examples{
\dontrun{
 plot_gvis_volcano(voomed_data, toptable_data, filename="html/fun_ma_plot.html",
                    base_url="http://yeastgenome.org/accession?")
}
}
\seealso{
\pkg{googleVis}
 \code{\link{plot_de_volcano}}
}
