% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_heatmap,SummarizedExperiment-method}
\alias{plot_heatmap,SummarizedExperiment-method}
\title{Run plot_heatmap with a SummarizedExperiment as input.}
\usage{
\S4method{plot_heatmap}{SummarizedExperiment}(
  expt_data,
  expt_colors = NULL,
  expt_design = NULL,
  method = "pearson",
  expt_names = NULL,
  type = "correlation",
  batch_row = "batch",
  plot_title = NULL,
  label_chars = 10,
  ...
)
}
\description{
Run plot_heatmap with a SummarizedExperiment as input.
}
