% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Make a pretty histogram of something.}
\usage{
plot_histogram(
  df,
  binwidth = NULL,
  log = FALSE,
  bins = 500,
  adjust = 1,
  fillcolor = "darkgrey",
  color = "black"
)
}
\arguments{
\item{df}{Dataframe of lots of pretty numbers.}

\item{binwidth}{Width of the bins for the histogram.}

\item{log}{Replot on the log scale?}

\item{bins}{Number of bins for the histogram.}

\item{adjust}{The prettification parameter in the ggplot2 density.}

\item{fillcolor}{Change the fill colors of the plotted elements?}

\item{color}{Change the color of the lines of the plotted elements?}
}
\value{
Ggplot histogram.
}
\description{
A shortcut to make a ggplot2 histogram which makes an attempt to set
reasonable bin widths and set the scale to log if that seems a good idea.
}
\examples{
\dontrun{
 kittytime = plot_histogram(df)
}
}
\seealso{
[ggplot2]
}
