% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.r
\name{plot_hypotrochoid}
\alias{plot_hypotrochoid}
\title{Make hypotrochoid plots!}
\usage{
plot_hypotrochoid(radius_a = 7, radius_b = 1, dist_b = 5,
  revolutions = 7, increments = 6480)
}
\description{
3,7,1 should give the classic 7 leaf clover
}

