% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.r
\name{plot_legend}
\alias{plot_legend}
\title{Scab the legend from a PCA plot and print it alone}
\usage{
plot_legend(stuff)
}
\arguments{
\item{stuff}{This can take either a ggplot2 pca plot or some data from which to make one.}
}
\value{
A legend!
}
\description{
This way I can have a legend object to move about.
}
