% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_libsize,data.frame,factor,character-method}
\alias{plot_libsize,data.frame,factor,character-method}
\title{Run plot_libsize() with a dataframe as input.}
\usage{
\S4method{plot_libsize}{data.frame,factor,character}(
  data,
  condition = NULL,
  colors = NULL,
  text = TRUE,
  order = NULL,
  plot_title = NULL,
  yscale = NULL,
  expt_names = NULL,
  label_chars = 10,
  ...
)
}
\description{
Run plot_libsize() with a dataframe as input.
}
