% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_libsize,expt,ANY,ANY-method}
\alias{plot_libsize,expt,ANY,ANY-method}
\title{Run plot_libsize() with an expt as input.}
\usage{
\S4method{plot_libsize}{expt,ANY,ANY}(
  data,
  condition = NULL,
  colors = NULL,
  text = TRUE,
  order = NULL,
  plot_title = NULL,
  yscale = NULL,
  expt_names = NULL,
  label_chars = 10,
  ...
)
}
\description{
Run plot_libsize() with an expt as input.
}
