% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_libsize}
\alias{plot_libsize}
\title{I changed this to 'plot_quantreads', but at least for a while cannot
be trusted to remember that.}
\usage{
plot_libsize(...)
}
\arguments{
\item{...}{Passed to plot_quantreads}
}
\description{
I changed this to 'plot_quantreads', but at least for a while cannot
be trusted to remember that.
}
