% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_linear_scatter}
\alias{plot_linear_scatter}
\title{Make a scatter plot between two groups with a linear model superimposed and
some supporting statistics.}
\usage{
plot_linear_scatter(
  df,
  cormethod = "pearson",
  size = 2,
  loess = FALSE,
  xcol = NULL,
  ycol = NULL,
  text_col = NULL,
  logfc = 2,
  identity = FALSE,
  z = 1.5,
  z_lines = FALSE,
  first = NULL,
  second = NULL,
  base_url = NULL,
  color_weights = TRUE,
  xlab = NULL,
  ylab = NULL,
  model_type = "robust",
  add_equation = TRUE,
  add_rsq = TRUE,
  add_cor = TRUE,
  label_prefix = "Expression of",
  color_high = NULL,
  color_low = NULL,
  alpha = 0.4,
  ...
)
}
\arguments{
\item{df}{Dataframe likely containing two columns.}

\item{cormethod}{What type of correlation to check?}

\item{size}{Size of the dots on the plot.}

\item{loess}{Add a loess estimation?}

\item{xcol}{Column name of x-values}

\item{ycol}{Column name of y-values#'}

\item{text_col}{Column containing text annotations.}

\item{logfc}{Point out genes with a specific logfc.}

\item{identity}{Add the identity line?}

\item{z}{Use this z-score cutoff.}

\item{z_lines}{Include lines defining the z-score boundaries.}

\item{first}{First column to plot.}

\item{second}{Second column to plot.}

\item{base_url}{Base url to add to the plot.}

\item{color_weights}{Apply colors by the weights of the points vs
the linear model.}

\item{xlab}{Alternate x-axis label.}

\item{ylab}{Alternate x-axis label.}

\item{model_type}{Type of lm to use.}

\item{add_equation}{Add the lm equation to the plot.}

\item{add_rsq}{Add the rsquared value to the plot.}

\item{add_cor}{Add the correlation coefficient to the plot.}

\item{label_prefix}{Use this prefix for the axis labels.}

\item{color_high}{Chosen color for points significantly above the mean.}

\item{color_low}{Chosen color for points significantly below the mean.}

\item{alpha}{Choose an alpha channel to define how see-through the dots are.}

\item{...}{Extra args likely used for choosing significant genes.}
}
\value{
List including a ggplot2 scatter plot and some histograms.  This plot
 provides a "bird's eye" view of two data sets.  This plot assumes a
 (potential) linear correlation between the data, so it calculates the
 correlation between them.  It then calculates and plots a robust linear
 model of the data using an 'SMDM' estimator (which I don't remember how to
 describe, just that the document I was reading said it is good).  The
 median/mad of each axis is calculated and plotted as well.  The distance
 from the linear model is finally used to color the dots on the plot.
 Histograms of each axis are plotted separately and then together under a
 single cdf to allow tests of distribution similarity.  This will make a fun
 clicky googleVis graph if requested.
}
\description{
Make a scatter plot between two groups with a linear model superimposed and
some supporting statistics.
}
\examples{
if (is.null(pombe_se)) {
  source(system.file("examples", "create_pombe.R", package = "hpgltools"))
}

## Biological Coefficient of variation plots are most helpful for raw data.
bcv <- plot_bcv(pombe_se)
bcv[["plot"]]

if (is.null(pombe_norm)) {
  source(system.file("examples", "normalize_pombe.R", package = "hpgltools"))
}

## Since I have the data loaded, I will plot two samples
## of my pombe data against each other...
test_df <- as.data.frame(assay(pombe_norm[, c(1, 2)]))

test_scatter <- plot_dist_scatter(test_df)
test_scatter

boring_scatter <- plot_scatter(test_df)
boring_scatter

test_scatter <- plot_linear_scatter(test_df)
summary(test_scatter)
test_scatter[["scatter"]]
## Hey, check it, they are very similar.

## See how many nonzero genes there are per sample vs. coverage
nonzero <- plot_nonzero(pombe_se)
nonzero

## MA samples against each other
## In this case I am just comparing the first two samples
## If I did not do this subset, it would do all pairwise comparisons.
sample_ma <- plot_pairwise_ma(pombe_se[, c(1, 2)])
sample_ma
}
\seealso{
[robust] [stats] [ggplot2] [robust::lmRob] [stats::weights] [plot_histogram()]
}
