% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{plot_ma_condition_de}
\alias{plot_ma_condition_de}
\title{Create a MA plot with colors from the original expressionset.}
\usage{
plot_ma_condition_de(
  input,
  table_name,
  expr_col = "logCPM",
  fc_col = "logFC",
  p_col = "qvalue",
  color_high = "red",
  color_low = "blue",
  pval = 0.05,
  alpha = 0.4,
  logfc = 1,
  label_numbers = TRUE,
  size = 2,
  shapes = TRUE,
  invert = FALSE,
  label = 10,
  label_column = "hgncsymbol",
  ...
)
}
\arguments{
\item{input}{Result from all_pairwise() and friends.}

\item{table_name}{Name the table!}

\item{expr_col}{Column name from the input containing expression data.}

\item{fc_col}{Ibid but the fold change column.}

\item{p_col}{Ibid but the p-value.}

\item{color_high}{Color for the values above the identity line.}

\item{color_low}{and the low side.}

\item{pval}{Significance cutoff.}

\item{alpha}{Degree of see-through-ness.}

\item{logfc}{Fold-change cutoff.}

\item{label_numbers}{Add a legend containing counts by significance.}

\item{size}{Relative size of the dots.}

\item{shapes}{Use fun shapes for categories?}

\item{invert}{Invert the plot?}

\item{label}{Add labels for this number of genes.}

\item{label_column}{Use this column for the labels.}

\item{...}{Arbitrary passthrough.}
}
\description{
The logic for this is directly from its volcano plot sister, but I think that function
is more complete.
}
