% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{plot_ma_de}
\alias{plot_ma_de}
\title{Make a pretty MA plot from one of limma, deseq, edger, or basic.}
\usage{
plot_ma_de(table, expr_col = "logCPM", fc_col = "logFC", p_col = "qvalue",
  pval_cutoff = 0.05, alpha = 0.4, logfc_cutoff = 1,
  label_numbers = TRUE, size = 2, tooltip_data = NULL,
  gvis_filename = NULL, ...)
}
\arguments{
\item{table}{Df of linear-modelling, normalized counts by sample-type,}

\item{expr_col}{Column showing the average expression across genes.}

\item{fc_col}{Column showing the logFC for each gene.}

\item{p_col}{Column containing the relevant p-values.}

\item{pval_cutoff}{Name of the pvalue column to use for cutoffs.}

\item{alpha}{How transparent to make the dots.}

\item{logfc_cutoff}{Fold change cutoff.}

\item{label_numbers}{Show how many genes were 'significant', 'up', and 'down'?}

\item{size}{How big are the dots?}

\item{tooltip_data}{Df of tooltip information for gvis.}

\item{gvis_filename}{Filename to write a fancy html graph.}

\item{...}{More options for you!}
}
\value{
Ggplot2 MA scatter plot.  This is defined as the rowmeans of the normalized counts by
    type across all sample types on the x-axis, and the log fold change between conditions on the
    y-axis. Dots are colored depending on if they are 'significant.'  This will make a fun clicky
    googleVis graph if requested.
}
\description{
Because I can never remember, the following from wikipedia: "An MA plot is an application of a
Bland–Altman plot for visual representation of two channel DNA microarray gene expression data
which has been transformed onto the M (log ratios) and A (mean average) scale."
}
\examples{
\dontrun{
plot_ma(voomed_data, toptable_data, gvis_filename="html/fun_ma_plot.html")
## Currently this assumes that a variant of toptable was used which
## gives adjusted p-values.  This is not always the case and I should
## check for that, but I have not yet.
}
}
\seealso{
\link{plot_gvis_ma} \link[limma]{toptable}
\link[limma]{voom} \link{hpgl_voom}
\link[limma]{lmFit} \link[limma]{makeContrasts}
\link[limma]{contrasts.fit}
}

