% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_meta_sankey,expt-method}
\alias{plot_meta_sankey,expt-method}
\title{Feed an expt to a sankey plot.}
\usage{
\S4method{plot_meta_sankey}{expt}(
  design,
  factors = c("condition", "batch"),
  color_choices = NULL,
  html = "sankey.html"
)
}
\description{
Feed an expt to a sankey plot.
}
