% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.R
\name{plot_meta_sankey}
\alias{plot_meta_sankey}
\title{Plot metadata factors as a sankey diagram.}
\usage{
plot_meta_sankey(
  design,
  factors = c("condition", "batch"),
  fill = "node",
  font_size = 18,
  node_width = 30,
  color_choices = NULL,
  drill_down = TRUE
)
}
\arguments{
\item{design}{Metadata from which to extract the categories/numbers.}

\item{factors}{Factors/columns in the metadata to count and plot.}

\item{fill}{Use either the current or next node for coloring the transitions.}

\item{color_choices}{Either a named vector of states and colors, or NULL
(in which case it will use viridis.)}

\item{drill_down}{When true, this will end in the product of the
 factor levels number of final states. (e.g. if there are 2 sexes,
 3 visits, and 4 genotypes, there will be 2, 6, 24 states going
from right to left).  If FALSE, there will be 2,3,4 states going
from right to left.}

\item{html}{Write the interactive plot to this file.}
}
\value{
List containing a couple of plots, one interactive, one gg.
}
\description{
This provides two implementations of a sankey plot, one interactive and one using ggplot2.
}
