% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{plot_metadata_factors}
\alias{plot_metadata_factors}
\title{Produce plots of metadata factor(s) of interest.}
\usage{
plot_metadata_factors(
  expt,
  column = "hisatsinglemapped",
  second_column = NULL,
  norm_column = NULL,
  type = NULL,
  scale = "base10"
)
}
\arguments{
\item{expt}{Input expressionset.}

\item{column}{Currently a single, but soon multiple column(s) of metadata.}

\item{second_column}{Or perhaps put other columns here.}

\item{norm_column}{Normalize the data?}

\item{type}{Assume a vioiln unless otherwise specified.}

\item{scale}{Rescale the data?}
}
\value{
ggplot and maybe some form of useful table.
}
\description{
Produce plots of metadata factor(s) of interest.
}
