% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.r
\name{plot_multiplot}
\alias{plot_multiplot}
\title{Make a grid of plots.}
\usage{
plot_multiplot(plots, file, cols = NULL, layout = NULL)
}
\arguments{
\item{plots}{a list of plots}

\item{file}{a file to write to}

\item{cols}{the number of columns in the grid}

\item{layout}{set the layout specifically}
}
\value{
a multiplot!
}
\description{
Make a grid of plots.
}

