% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proteomics.r
\name{plot_mzxml_boxplot}
\alias{plot_mzxml_boxplot}
\title{Make a boxplot out of some of the various data available in the mzxml data.}
\usage{
plot_mzxml_boxplot(
  mzxml_data,
  table = "precursors",
  column = "precursorintensity",
  violin = FALSE,
  names = NULL,
  title = NULL,
  scale = NULL,
  ...
)
}
\arguments{
\item{mzxml_data}{Provide a list of mzxml data, one element for each sample.}

\item{table}{One of precursors or scans}

\item{column}{One of the columns from the table; if 'scans' is chosen, then
likely choices include: 'peakscount', 'basepeakmz', 'basepeakintensity'; if
'precursors' is chosen, then the only likely choice for the moment is
'precursorintensity'.}

\item{violin}{Print the samples as violins rather than only box/whiskers?}

\item{names}{Names for the x-axis of the plot.}

\item{title}{Title the plot?}

\item{scale}{Put the data on a specific scale?}

\item{...}{Further arguments, presumably for colors or some such.}
}
\value{
Boxplot describing the requested column of data in the set of mzXML files.
}
\description{
There are a few data within the mzXML raw data files which are likely
  candidates for simple summary via a boxplot/densityplot/whatever.  For the
  moment I am just doing boxplots of a few of them.  Since my metadata
  extractor dumps a couple of tables, one must choose a desired table and
  column from it to plot.
}
