% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{plot_nifty_heatmap}
\alias{plot_nifty_heatmap}
\title{Adapt the nifty msstats heatmap to an expressionset}
\usage{
plot_nifty_heatmap(expt_data, expt_colors = NULL, expt_design = NULL,
  first_type = "correlation", first_method = "pearson",
  second_type = "distance", second_method = "euclidean",
  expt_names = NULL, batch_row = "batch", title = NULL, ...)
}
\description{
I have made fun of the code quality in msstats, but their heatmap function is really nice.
I want a version of it for other analyses.  Oh, you know what nevermind, this
is actually from SWATH2stats, which is quite nicely written.
}
