% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_nonzero,expt-method}
\alias{plot_nonzero,expt-method}
\title{Make a nonzero plot given an expt.}
\usage{
\S4method{plot_nonzero}{expt}(
  data,
  design = NULL,
  colors = NULL,
  plot_labels = "repel",
  expt_names = NULL,
  max_overlaps = 5,
  label_chars = 10,
  plot_legend = FALSE,
  plot_title = NULL,
  cutoff = 0.65,
  ...
)
}
\description{
Make a nonzero plot given an expt.
}
