% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_nonzero}
\alias{plot_nonzero}
\title{Make a ggplot graph of the number of non-zero genes by sample.}
\usage{
plot_nonzero(
  data,
  design = NULL,
  colors = NULL,
  plot_labels = "repel",
  exp_names = NULL,
  max_overlaps = 5,
  label_chars = 10,
  plot_legend = FALSE,
  plot_title = NULL,
  cutoff = 0.65,
  y_intercept = NULL,
  rug = TRUE,
  loess = FALSE,
  x_intercept = FALSE,
  ...
)
}
\arguments{
\item{data}{Exp, expressionset, or dataframe.}

\item{design}{Eesign matrix.}

\item{colors}{Color scheme.}

\item{plot_labels}{How do you want to label the graph? 'fancy' will use
directlabels() to try to match the labels with the positions without
overlapping anything else will just stick them on a 45' offset next to the
graphed point.}

\item{exp_names}{Column or character list of preferred sample names.}

\item{max_overlaps}{Permit this many labels to overlap before dropping some.}

\item{label_chars}{How many characters for sample names before abbreviation.}

\item{plot_legend}{Print a legend for this plot?}

\item{plot_title}{Add a title?}

\item{cutoff}{Minimum proportion (or number) of genes below which samples might be in trouble.}

\item{y_intercept}{Add a y-intercept to define 'good' coverage.}

\item{rug}{Add a rug.}

\item{loess}{Add a loess estimation.}

\item{x_intercept}{Perhaps show the mean expression via a x-intercept.}

\item{...}{rawr!}
}
\value{
a ggplot2 plot of the number of non-zero genes with respect to each
 library's CPM.
}
\description{
This puts the number of genes with > 0 hits on the y-axis and CPM on the
x-axis. Made by Ramzi Temanni <temanni at umd dot edu>.
}
\examples{
library(hpgltools)
pombe_se <- get0("pombe_se")
if (is.null(pombe_se)) {
  source(system.file("examples", "create_pombe.R", package = "hpgltools"))
}

## Biological Coefficient of variation plots are most helpful for raw data.
bcv <- plot_bcv(pombe_se)
bcv[["plot"]]

pombe_norm <- get0("pombe_norm")
if (is.null(pombe_norm)) {
  source(system.file("examples", "normalize_pombe.R", package = "hpgltools"))
}

## Since I have the data loaded, I will plot two samples
## of my pombe data against each other...
test_df <- as.data.frame(assay(pombe_norm[, c(1, 2)]))

test_scatter <- plot_dist_scatter(test_df)
test_scatter

boring_scatter <- plot_scatter(test_df)
boring_scatter

test_scatter <- plot_linear_scatter(test_df)
summary(test_scatter)
test_scatter[["scatter"]]
## Hey, check it, they are very similar.

## See how many nonzero genes there are per sample vs. coverage
nonzero <- plot_nonzero(pombe_se)
nonzero

## MA samples against each other
## In this case I am just comparing the first two samples
## If I did not do this subset, it would do all pairwise comparisons.
sample_ma <- plot_pairwise_ma(pombe_se[, c(1, 2)])
sample_ma
}
\seealso{
[ggplot2]
}
