% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{plot_num_siggenes}
\alias{plot_num_siggenes}
\title{Given a DE table with fold changes and p-values, show how 'significant'
changes with changing cutoffs.}
\usage{
plot_num_siggenes(
  table,
  methods = c("limma", "edger", "deseq", "ebseq"),
  bins = 100,
  constant_p = 0.05,
  constant_fc = 0
)
}
\arguments{
\item{table}{DE table to examine.}

\item{methods}{List of methods to use when plotting.}

\item{bins}{Number of incremental changes in p-value/FC to examine.}

\item{constant_p}{When plotting changing FC, where should the p-value be held?}

\item{constant_fc}{When plotting changing p, where should the FC be held?}
}
\value{
Plots and dataframes describing the changing definition of 'significant.'
}
\description{
Sometimes one might want to know how many genes are deemed significant while
shifting the bars which define significant.  This provides that metrics as a
set of tables of numbers of significant up/down genes when p-value is held
constant, as well as number when fold-change is held constant.
}
\examples{
\dontrun{
 pairwise_result <- all_pairwise(expt)
 crazy_sigplots <- plot_num_siggenes(pairwise_result)
}
}
\seealso{
[ggplot2]
}
