% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.R
\name{plot_ontpval}
\alias{plot_ontpval}
\title{Make a pvalue plot from a df of IDs, scores, and p-values.}
\usage{
plot_ontpval(
  df,
  ontology = "MF",
  fontsize = 14,
  plot_title = NULL,
  text_location = "right",
  text_color = "black",
  x_column = "score",
  numerator = NULL,
  denominator = NULL
)
}
\arguments{
\item{df}{Some data from topgo/goseq/clusterprofiler.}

\item{ontology}{Ontology to plot (MF,BP,CC).}

\item{fontsize}{Fiddling with the font size may make some plots more readable.}

\item{plot_title}{Set an explicit plot title.}

\item{text_location}{Choose where to put the text describing the number of genes in the category.}

\item{text_color}{Choose the text color, I have a fun function for this now...}

\item{x_column}{Use this column to arrange the x-axis.}

\item{numerator}{Column used for printing a ratio of genes/category.}

\item{denominator}{Column used for printing a ratio of genes/category.}
}
\value{
Ggplot2 plot of pvalues vs. ontology.
}
\description{
This function seeks to make generating pretty pvalue plots as shown by
clusterprofiler easier.
}
\seealso{
[ggplot2]
}
