% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_overrepresented_species}
\alias{plot_overrepresented_species}
\title{Given results from kraken, create a bar plot of the most over represented species observed.}
\usage{
plot_overrepresented_species(exp, factor = 1.5, prop_vs = "Mean")
}
\arguments{
\item{exp}{Input se}

\item{factor}{Define overrepresented as this factor multiplied by 'prop_vs', which
in turn may be a portion of the quartile or mean/median.}

\item{prop_vs}{The portion of the data distribution used to define an outlier.}
}
\description{
Given results from kraken, create a bar plot of the most over represented species observed.
}
