% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_pairwise_ma,ExpressionSet-method}
\alias{plot_pairwise_ma,ExpressionSet-method}
\title{Plot all pairwise MA plots in an experiment.}
\usage{
\S4method{plot_pairwise_ma}{ExpressionSet}(data, colors = NULL, design = NULL, log = NULL, ...)
}
\arguments{
\item{data}{Expt expressionset or data frame.}

\item{colors}{Explicit colors to use.}

\item{design}{df of experimental design.}

\item{log}{Is the data in log format?}

\item{...}{Options are good and passed to arglist().}
}
\value{
List of affy::maplots
}
\description{
Use affy's ma.plot() on every pair of columns in a data set to help diagnose
problematic samples.
}
