% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_pairwise_ma}
\alias{plot_pairwise_ma}
\title{Generic for plot_pairwise_ma for unknown input types.}
\usage{
plot_pairwise_ma(data, ...)
}
\arguments{
\item{data}{Arbitrary input data.}

\item{...}{Other arguments intended for plot_pairwise_ma()}
}
\description{
Generic for plot_pairwise_ma for unknown input types.
}
