% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{plot_pca,ExpressionSet-method}
\alias{plot_pca,ExpressionSet-method}
\title{Using plot_pca with an expressionset may require some
color-specific arguments.}
\usage{
\S4method{plot_pca}{ExpressionSet}(
  data,
  design = NULL,
  state = NULL,
  plot_colors = NULL,
  plot_title = TRUE,
  plot_size = 5,
  plot_alpha = NULL,
  plot_labels = FALSE,
  size_column = NULL,
  pc_method = "fast_svd",
  x_pc = 1,
  y_pc = 2,
  max_overlaps = 20,
  num_pc = NULL,
  expt_names = NULL,
  label_chars = 10,
  cond_column = "condition",
  batch_column = "batch",
  ...
)
}
\arguments{
\item{data}{Input data type.}

\item{design}{a design matrix and.}

\item{state}{State of the data.}

\item{plot_colors}{a color scheme.}

\item{plot_title}{a title for the plot.}

\item{plot_size}{size for the glyphs on the plot.}

\item{plot_alpha}{Add an alpha channel to the dots?}

\item{plot_labels}{add labels?  Also, what type?  FALSE, "default", or "fancy".}

\item{size_column}{use an experimental factor to size the glyphs of the plot}

\item{pc_method}{how to extract the components? (svd}

\item{x_pc}{Component to put on the x axis.}

\item{y_pc}{Component to put on the y axis.}

\item{max_overlaps}{Passed to ggrepel.}

\item{num_pc}{How many components to calculate, default to the number of
rows in the metadata.}

\item{expt_names}{Column or character list of preferred sample names.}

\item{label_chars}{Maximum number of characters before abbreviating sample names.}

\item{cond_column}{Column containing the color information.}

\item{batch_column}{Column containing the shape information.}

\item{...}{Arguments passed through to the pca implementations and plotter.}
}
\description{
Using plot_pca with an expressionset may require some
color-specific arguments.
}
