% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pca.r
\name{plot_pca}
\alias{plot_pca}
\title{Make a ggplot PCA plot describing the samples' clustering.}
\usage{
plot_pca(data, design = NULL, plot_colors = NULL, plot_labels = NULL,
  plot_title = NULL, plot_size = 5, size_column = NULL, ...)
}
\arguments{
\item{data}{an expt set of samples.}

\item{design}{a design matrix and.}

\item{plot_colors}{a color scheme.}

\item{plot_labels}{add labels?  Also, what type?  FALSE, "default", or "fancy".}

\item{plot_title}{a title for the plot.}

\item{plot_size}{size for the glyphs on the plot.}

\item{size_column}{use an experimental factor to size the glyphs of the plot}

\item{...}{arglist from elipsis!}
}
\value{
a list containing the following:
 \enumerate{
  \item  pca = the result of fast.svd()
  \item  plot = ggplot2 pca_plot describing the principle component analysis of the samples.
  \item  table = a table of the PCA plot data
  \item  res = a table of the PCA res data
  \item  variance = a table of the PCA plot variance
 }
}
\description{
Make a ggplot PCA plot describing the samples' clustering.
}
\examples{
\dontrun{
 pca_plot <- plot_pca(expt=expt)
 pca_plot
}
}
\seealso{
\code{\link{makeSVD}},
\code{\link[directlabels]{geom_dl}} \code{\link{plot_pcs}}
}

