% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.r
\name{plot_pca_genes}
\alias{plot_pca_genes}
\title{Make a PC plot describing the gene' clustering.}
\usage{
plot_pca_genes(
  data,
  design = NULL,
  plot_colors = NULL,
  plot_title = NULL,
  plot_size = 2,
  plot_alpha = 0.4,
  plot_labels = FALSE,
  size_column = NULL,
  pc_method = "fast_svd",
  x_pc = 1,
  y_pc = 2,
  label_column = "description",
  num_pc = 2,
  expt_names = NULL,
  label_chars = 10,
  ...
)
}
\arguments{
\item{data}{an expt set of samples.}

\item{design}{a design matrix and.}

\item{plot_colors}{a color scheme.}

\item{plot_title}{a title for the plot.}

\item{plot_size}{size for the glyphs on the plot.}

\item{plot_alpha}{Add an alpha channel to the dots?}

\item{plot_labels}{add labels?  Also, what type?  FALSE, "default", or "fancy".}

\item{size_column}{use an experimental factor to size the glyphs of the plot}

\item{pc_method}{how to extract the components? (svd}

\item{x_pc}{Component to put on the x axis.}

\item{y_pc}{Component to put on the y axis.}

\item{label_column}{Which metadata column to use for labels.}

\item{num_pc}{How many components to calculate, default to the number of
rows in the metadata.}

\item{expt_names}{Column or character list of preferred sample names.}

\item{label_chars}{Maximum number of characters before abbreviating sample names.}

\item{...}{Arguments passed through to the pca implementations and plotter.}
}
\value{
a list containing the following (this is currently wrong)
\enumerate{
 \item  pca = the result of fast.svd()
 \item  plot = ggplot2 pca_plot describing the principle component analysis of the samples.
 \item  table = a table of the PCA plot data
 \item  res = a table of the PCA res data
 \item  variance = a table of the PCA plot variance
}
}
\description{
Make a PC plot describing the gene' clustering.
}
\examples{
\dontrun{
 pca_plot <- plot_pca(expt=expt)
 pca_plot
}
}
\seealso{
\pkg{directlabels}
 \code{\link[directlabels]{geom_dl}} \code{\link{plot_pcs}}
}
