% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.r
\name{plot_pcfactor}
\alias{plot_pcfactor}
\title{make a dotplot of some categorised factors and a set of principle components.}
\usage{
plot_pcfactor(pc_df, exp_factpr)
}
\arguments{
\item{pc_df}{Df of principle components.}

\item{exp_factor}{Experimental factor to compare against.}
}
\description{
This should make a quick df of the factors and PCs and plot them.
}
\examples{
\dontrun{
estimate_vs_pcs <- plot_pcfactor(pcs, times)
}
}

