% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{plot_pcload}
\alias{plot_pcload}
\title{Print a plot of the top-n most PC loaded genes.}
\usage{
plot_pcload(expt, genes = 40, desired_pc = 1, which_scores = "high", ...)
}
\arguments{
\item{expt}{Input expressionset.}

\item{genes}{How many genes to observe?}

\item{desired_pc}{Which component to examine?}

\item{which_scores}{Perhaps one wishes to see the least-important genes, if
so set this to low.}

\item{...}{Extra arguments passed, currently to nothing.}
}
\value{
List containing an expressionset of the subset and a plot of their
 expression.
}
\description{
Sometimes it is nice to know what is happening with the genes which have the
greatest effect on a given principal component.  This function provides that.
}
\seealso{
[plot_sample_heatmap()]
}
