% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.r
\name{plot_pct_kept}
\alias{plot_pct_kept}
\title{Make a ggplot graph of the percentage/number of reads kept/removed.}
\usage{
plot_pct_kept(
  data,
  row = "pct_kept",
  condition = NULL,
  colors = NULL,
  names = NULL,
  text = TRUE,
  title = NULL,
  yscale = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe of the material remaining, usually expt$summary_table}

\item{row}{Row name to plot.}

\item{condition}{Vector of sample condition names.}

\item{colors}{Color scheme if the data is not an expt.}

\item{names}{Alternate names for the x-axis.}

\item{text}{Add the numeric values inside the top of the bars of the plot?}

\item{title}{Title for the plot.}

\item{yscale}{Whether or not to log10 the y-axis.}

\item{...}{More parameters for your good time!}
}
\value{
a ggplot2 bar plot of every sample's size
}
\description{
The function expt_exclude_genes() removes some portion of the original reads.
This function will make it possible to see what is left.
}
\examples{
\dontrun{
 kept_plot <- plot_pct_kept(expt_removed)
 kept_plot  ## ooo pretty bargraph
}
}
\seealso{
[plot_sample_bars()]
}
