% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{plot_pyprophet_data}
\alias{plot_pyprophet_data}
\title{Plot some data from the result of extract_peprophet_data()}
\usage{
plot_pyprophet_data(pyprophet_data, xaxis = "mass", xscale = NULL,
  yaxis = "leftwidth", yscale = NULL, alpha = 0.4,
  size_column = "mscore", ...)
}
\arguments{
\item{xaxis}{Column to plot on the x-axis}

\item{xscale}{Change the scale of the x-axis?}

\item{yaxis}{guess!}

\item{yscale}{Change the scale of the y-axis?}

\item{size_column}{Use a column for scaling the sizes of dots in the plot?}

\item{...}{extra options which may be used for plotting.}

\item{table}{Big honking data table from extract_peprophet_data()}
}
\value{
a plot!
}
\description{
extract_pyprophet_data() provides a ridiculously large data table of a scored
openswath data after processing by pyprophet.
}
