% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proteomics.R
\name{plot_pyprophet_points}
\alias{plot_pyprophet_points}
\title{Plot some data from the result of extract_pyprophet_data()}
\usage{
plot_pyprophet_points(
  pyprophet_data,
  xaxis = "mass",
  xscale = NULL,
  sample = NULL,
  yaxis = "leftwidth",
  yscale = NULL,
  alpha = 0.4,
  color_by = "sample",
  legend = TRUE,
  size_column = "mscore",
  rug = TRUE,
  ...
)
}
\arguments{
\item{pyprophet_data}{List of pyprophet data, one element for each sample,
taken from extract_peprophet_data()}

\item{xaxis}{Column to plot on the x-axis}

\item{xscale}{Change the scale of the x-axis?}

\item{sample}{Which sample(s) to include?}

\item{yaxis}{guess!}

\item{yscale}{Change the scale of the y-axis?}

\item{alpha}{How see-through to make the dots?}

\item{color_by}{Change the colors of the points either by sample or condition?}

\item{legend}{Include a legend of samples?}

\item{size_column}{Use a column for scaling the sizes of dots in the plot?}

\item{rug}{Add a distribution rug to the axes?}

\item{...}{extra options which may be used for plotting.}
}
\value{
a plot!
}
\description{
extract_pyprophet_data() provides a ridiculously large data table of a scored
openswath data after processing by pyprophet.
}
