% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proteomics.r
\name{plot_pyprophet_protein}
\alias{plot_pyprophet_protein}
\title{Read data from pyprophet and plot columns from it.}
\usage{
plot_pyprophet_protein(
  pyprophet_data,
  column = "intensity",
  keep_real = TRUE,
  keep_decoys = FALSE,
  expt_names = NULL,
  label_chars = 10,
  protein = NULL,
  title = NULL,
  scale = NULL,
  legend = NULL,
  order_by = "condition",
  show_all = TRUE,
  ...
)
}
\arguments{
\item{pyprophet_data}{Data from extract_pyprophet_data()}

\item{column}{Chosen column to plot.}

\item{keep_real}{FIXME: This should be changed to something like 'data_type'
here and in plot_pyprophet_distribution.}

\item{keep_decoys}{Do we keep the decoys when plotting the data?}

\item{expt_names}{Names for the x-axis of the plot.}

\item{label_chars}{Maximum number of characters before abbreviating sample
names.}

\item{protein}{chosen protein(s) to plot.}

\item{title}{Title the plot?}

\item{scale}{Put the data on a specific scale?}

\item{legend}{Include the legend?}

\item{order_by}{Reorder the samples by some factor, presumably condition.}

\item{show_all}{Skip samples for which no observations were made.}

\item{...}{Further arguments, presumably for colors or some such.}
}
\value{
Boxplot describing the desired column from the data.
}
\description{
More proteomics diagnostics!  Now that I am looking more closely, I think
this should be folded into plot_pyprophet_distribution().
}
