% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proteomics.R
\name{plot_pyprophet_xy}
\alias{plot_pyprophet_xy}
\title{Invoked plot_pyprophet_counts() twice, once for the x-axis, and once for the y.}
\usage{
plot_pyprophet_xy(
  pyprophet_data,
  keep_real = TRUE,
  size = 6,
  label_size = 4,
  keep_decoys = TRUE,
  expt_names = NULL,
  label_chars = 10,
  x_type = "count",
  y_type = "intensity",
  title = NULL,
  scale = NULL,
  ...
)
}
\arguments{
\item{pyprophet_data}{List of pyprophet matrices by sample.}

\item{keep_real}{Use the real identifications (as opposed to the decoys)?}

\item{size}{Size of the glyphs used in the plot.}

\item{label_size}{Set the label sizes.}

\item{keep_decoys}{Use the decoy identifications (vs. the real)?}

\item{expt_names}{Manually change the labels to some other column than sample.}

\item{label_chars}{Maximum number of characters in the label before
shortening.}

\item{x_type}{Column in the data to put on the x-axis.}

\item{y_type}{Column in the data to put on the y-axis.}

\item{title}{Plot title.}

\item{scale}{Put the data onto the log scale?}

\item{...}{Extra arguments passed along.}
}
\description{
Then plot the result, hopefully adding some new insights into the state of
the post-pyprophet results.  By default, this puts the number of
identifications (number of rows) on the x-axis for each sample, and the sum
of intensities on the y.  Currently missing is the ability to change this
from sum to mean/median/etc.  That should trivially be possible via the
addition of arguments for the various functions of interest.
}
