% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_qq_all,ExpressionSet-method}
\alias{plot_qq_all,ExpressionSet-method}
\title{Invoke plot_qq_all using an ExpressionSet}
\usage{
\S4method{plot_qq_all}{ExpressionSet}(data, design = NULL, colors = NULL, labels = "short", ...)
}
\arguments{
\item{data}{Input assay}

\item{design}{provided by the assay}

\item{colors}{also from the assay}

\item{labels}{shorten the sample IDs?}

\item{...}{passed along}
}
\description{
Invoke plot_qq_all using an ExpressionSet
}
