% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_qq_all,expt-method}
\alias{plot_qq_all,expt-method}
\title{Invoke plot_qq_all using an expt}
\usage{
\S4method{plot_qq_all}{expt}(data, design = NULL, colors = NULL, labels = "short", ...)
}
\arguments{
\item{data}{Input expt}

\item{design}{provided by the expt}

\item{colors}{also from the expt}

\item{labels}{shorten the sample IDs?}

\item{...}{passed along}
}
\description{
Invoke plot_qq_all using an expt
}
