% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_qq_all,SummarizedExperiment-method}
\alias{plot_qq_all,SummarizedExperiment-method}
\title{Invoke plot_qq_all using a SE}
\usage{
\S4method{plot_qq_all}{SummarizedExperiment}(data, design = NULL, colors = NULL, labels = "short", ...)
}
\arguments{
\item{data}{Input SE}

\item{design}{provided by the SE}

\item{colors}{also from the SE}

\item{labels}{shorten the sample IDs?}

\item{...}{passed along}
}
\description{
Invoke plot_qq_all using a SE
}
