% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.r
\name{plot_rpm}
\alias{plot_rpm}
\title{Make relatively pretty bar plots of coverage in a genome.}
\usage{
plot_rpm(input, workdir = "images", output = "01.svg",
  name = "LmjF.01.0010", start = 1000, end = 2000, strand = 1,
  padding = 100)
}
\arguments{
\item{input}{Coverage / position filename.}

\item{workdir}{Where to put the resulting images.}

\item{output}{Output image filename.}

\item{name}{Gene name to print at the bottom of the plot.}

\item{start}{Relative to 0, where is the gene's start codon.}

\item{end}{Relative to 0, where is the gene's stop codon.}

\item{strand}{Is this on the + or - strand? (+1/-1)}

\item{padding}{How much space to provide on the sides?}
}
\value{
coverage plot surrounging the ORF of interest
}
\description{
This was written for ribosome profiling coverage / gene.
It should however, work for any data with little or no modification, it was
also written when I was first learning R and when I look at it now I see a
few obvious places which can use improvement.
}
\seealso{
\pkg{ggplot2}
}
