% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{plot_sample_cvheatmap}
\alias{plot_sample_cvheatmap}
\title{An experiment to see if I can visualize the genes with the highest variance.}
\usage{
plot_sample_cvheatmap(
  expt,
  fun = "mean",
  fact = "condition",
  row_label = NA,
  title = NULL,
  Rowv = TRUE,
  Colv = TRUE,
  label_chars = 10,
  dendrogram = "column",
  min_delta = 0.5,
  x_factor = 1,
  y_factor = 2,
  min_cvsd = NULL,
  cv_min = 1,
  cv_max = Inf,
  remove_equal = TRUE
)
}
\arguments{
\item{expt}{ExpressionSet}

\item{fun}{mean or median}

\item{fact}{Which factor to slice/dice the data?}

\item{row_label}{Label the rows?}

\item{title}{Title for the plot}

\item{Rowv}{Row vs (yeah I forgot what this does.)}

\item{Colv}{Col vs}

\item{label_chars}{Maximum number of characters in the sample IDs.}

\item{dendrogram}{Make a tree of the samples?}

\item{min_delta}{Minimum delta value for filtering}

\item{x_factor}{When plotting two factors against each other, which is x?}

\item{y_factor}{When plotting two factors against each other, which is y?}

\item{min_cvsd}{Include only those with a minimal CV?}

\item{cv_min}{Minimum cv to examine (I think this should be slightly lower)}

\item{cv_max}{Maximum cV to examine (I think this should be limited to ~ 0.7?)}

\item{remove_equal}{Filter uninteresting genes.}
}
\description{
An experiment to see if I can visualize the genes with the highest variance.
}
