% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{plot_sample_heatmap}
\alias{plot_sample_heatmap}
\title{Make a heatmap.3 description of the similarity of the genes among samples.}
\usage{
plot_sample_heatmap(data, colors = NULL, design = NULL, names = NULL,
  title = NULL, Rowv = TRUE, ...)
}
\arguments{
\item{data}{Expt/expressionset/dataframe set of samples.}

\item{colors}{Color scheme of the samples (not needed if input is an expt).}

\item{design}{Design matrix describing the experiment (gotten for free if an expt).}

\item{names}{Alternate samples names.}

\item{title}{Title of the plot!}

\item{Rowv}{Reorder the rows by expression?}

\item{...}{More parameters for a good time!}
}
\value{
a recordPlot() heatmap describing the samples.
}
\description{
Sometimes you just want to see how the genes of an experiment are related to each other.  This
can handle that.  These heatmap functions should probably be replaced with neatmaps or heatplus
or whatever it is, as the annotation dataframes in them are pretty awesome.
}
\seealso{
\link[RColorBrewer]{brewer.pal} \link[grDevices]{recordPlot}
}

