% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{plot_scatter}
\alias{plot_scatter}
\title{Make a pretty scatter plot between two sets of numbers.}
\usage{
plot_scatter(df, tooltip_data = NULL, color = "black",
  gvis_filename = NULL, size = 2)
}
\arguments{
\item{df}{Dataframe likely containing two columns.}

\item{tooltip_data}{Df of tooltip information for gvis.}

\item{color}{Color of the dots on the graph.}

\item{gvis_filename}{Filename to write a fancy html graph.}

\item{size}{Size of the dots on the graph.}
}
\value{
Ggplot2 scatter plot.
}
\description{
This function tries to supplement a normal scatterplot with some information describing the
relationship between the columns of data plotted.
}
\examples{
\dontrun{
 plot_scatter(lotsofnumbers_intwo_columns, tooltip_data=tooltip_dataframe,
              gvis_filename="html/fun_scatterplot.html")
}
}
\seealso{
\pkg{ggplot2} \pkg{googleVis}
\code{\link{plot_gvis_scatter}} \code{\link[ggplot2]{geom_point}}
\code{\link{plot_linear_scatter}}
}
