% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{plot_seurat_scatter}
\alias{plot_seurat_scatter}
\title{Make a few of the likely scatterplots provided by FeatureScatter.}
\usage{
plot_seurat_scatter(scd, set = NULL)
}
\arguments{
\item{scd}{SCD to plot.}

\item{set}{List of plots, use my favorites when NULL.}
}
\value{
List of plots.
}
\description{
It seems I have used the same couple of scatter plots more often than others.
}
