% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.r
\name{plot_significant_bar}
\alias{plot_significant_bar}
\title{Make a bar plot of the numbers of significant genes by contrast.
These plots are quite difficult to describe.}
\usage{
plot_significant_bar(ups, downs, maximum = NULL, text = TRUE,
  color_list = c("lightcyan", "lightskyblue", "dodgerblue", "plum1", "orchid",
  "purple4"), color_names = c("a_up_inner", "b_up_middle", "c_up_outer",
  "a_down_inner", "b_down_middle", "c_down_outer"))
}
\arguments{
\item{ups}{Set of up-regulated genes.}

\item{downs}{Set of down-regulated genes.}

\item{maximum}{Maximum/minimum number of genes to display.}

\item{text}{Add text at the ends of the bars describing the number of genes >/< 0 fc.}

\item{color_list}{Set of colors to use for the bars.}

\item{color_names}{Categories associated with aforementioned colors.}
}
\description{
Make a bar plot of the numbers of significant genes by contrast.
These plots are quite difficult to describe.
}

