% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{plot_single_qq}
\alias{plot_single_qq}
\title{Perform a qqplot between two columns of a matrix.}
\usage{
plot_single_qq(data, x = 1, y = 2, labels = TRUE)
}
\arguments{
\item{data}{Data frame/expt/expressionset.}

\item{x}{First column to compare.}

\item{y}{Second column to compare.}

\item{labels}{Include the lables?}
}
\value{
a list of the logs, ratios, and mean between the plots as ggplots.
}
\description{
Given two columns of data, how well do the distributions match one another?  The answer to that
question may be visualized through a qq plot!
}
\seealso{
\pkg{Biobase}
}
