% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_sm,SummarizedExperiment-method}
\alias{plot_sm,SummarizedExperiment-method}
\title{Plot the standard median pairwise values of a SummarizedExperiment.}
\usage{
\S4method{plot_sm}{SummarizedExperiment}(
  data,
  design = NULL,
  colors = NULL,
  method = "pearson",
  plot_legend = FALSE,
  expt_names = NULL,
  label_chars = 10,
  plot_title = NULL,
  dot_size = 5,
  ...
)
}
\description{
Plot the standard median pairwise values of a SummarizedExperiment.
}
