% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.r
\name{plot_sm}
\alias{plot_sm}
\title{Make an R plot of the standard median correlation or distance among samples.}
\usage{
plot_sm(data, colors = NULL, method = "pearson", names = NULL,
  title = NULL, ...)
}
\arguments{
\item{data}{Expt, expressionset, or data frame.}

\item{colors}{Color scheme if data is not an expt.}

\item{method}{Correlation or distance method to use.}

\item{names}{Use pretty names for the samples?}

\item{title}{Title for the graph.}

\item{...}{More parameters to make you happy!}
}
\value{
ggplot of the standard median something
among the samples.  This will also write to an
open device.  The resulting plot measures the median correlation of
each sample among its peers.  It notes 1.5* the interquartile range
among the samples and makes a horizontal line at that correlation
coefficient.  Any sample which falls below this line is considered
for removal because it is much less similar to all of its peers.
}
\description{
This was written by a mix of Kwame Okrah <kokrah at gmail dot com>, Laura
Dillon <dillonl at umd dot edu>, and Hector Corrada Bravo <hcorrada at umd dot edu>
I reimplemented it using ggplot2 and tried to make it a little more flexible.
The general idea is to take the pairwise correlations/distances of the samples, then take the
medians, and plot them.
}
\examples{
\dontrun{
 smc_plot = hpgl_smc(expt=expt)
}
}
\seealso{
\link{hpgl_cor} \link[matrixStats]{rowMedians}
\link[stats]{quantile} \link{diff} \link[grDevices]{recordPlot}
}

