% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.r
\name{plot_spirograph}
\alias{plot_spirograph}
\title{Make spirographs!}
\usage{
plot_spirograph(radius_a = 1, radius_b = -4, dist_bc = -2,
  revolutions = 158, increments = 3160, center_a = list(x = 0, y = 0))
}
\arguments{
\item{radius_a}{The radius of the primary circle.}

\item{radius_b}{The radius of the circle travelling around a.}

\item{dist_bc}{A point relative to the center of 'b' which rotates with the turning of 'b'.}

\item{revolutions}{How many revolutions to perform in the plot}

\item{increments}{The number of radial increments to be calculated per revolution}

\item{center_a}{The position of the center of 'a'.}
}
\value{
something which I don't yet know.
}
\description{
Taken (with modifications) from: http://menugget.blogspot.com/2012/12/spirograph-with-r.html#more
A positive value for 'B' will result in a epitrochoid, while a negative value will result in a hypotrochoid.
}

