% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_splicing.R
\name{plot_suppa}
\alias{plot_suppa}
\title{Given some psi and tpm data, make a pretty plot!}
\usage{
plot_suppa(
  file_prefix,
  file_list = NULL,
  type = "type",
  annot = NULL,
  annot_column = NULL,
  sig_threshold = 0.05,
  label_type = NULL,
  alpha = 0.3,
  numerator = "infected",
  denominator = "uninfected"
)
}
\arguments{
\item{file_prefix}{Directory containing the various requisite input files.}

\item{file_list}{Vector of filenames.}

\item{type}{Either transcript or 'type' referring to the type of DPSI analysis performed.}

\item{annot}{Dataframe of annotations.}

\item{annot_column}{Column in annot with the IDs to cross reference against.}

\item{sig_threshold}{Use this significance threshold.}

\item{label_type}{Choose a type of event to label.}

\item{alpha}{How see-through should the points be in the plot?}

\item{numerator}{Name of the desired comparison's numerator.}

\item{denominator}{Name of the desired comparison's denominator.}
}
\value{
List containing the plot and some of the requisite data.
}
\description{
This should take either a dataframe or filename for the psi data from suppa,
along with the same for the average log tpm data (acquired from suppa
diffSplice with --save_tpm_events)
}
\examples{
 \dontrun{
 suppa_plot <- plot_suppa(dpsi_file, tmp_file)
}
}
\seealso{
[plot_rmats()]
}
