% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{plot_sv_meta}
\alias{plot_sv_meta}
\title{Compare a series of SVs to a metadata factor of interest.}
\usage{
plot_sv_meta(sv_meta, meta_column = "typeofcells", sv = 1, alpha = 0.75)
}
\arguments{
\item{sv_meta}{Dataframe containing SVs and metadata}

\item{meta_column}{Metadatum of interest}

\item{sv}{SV column number of interest}

\item{alpha}{DEgree of transparency}
}
\description{
Compare a series of SVs to a metadata factor of interest.
}
