% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate_estimators.r
\name{plot_svfactor}
\alias{plot_svfactor}
\title{make a dotplot of some categorised factors and a set of SVs (for other factors)}
\usage{
plot_svfactor(expt, svest, chosen_factor = "snpcategory",
  factor_type = "factor")
}
\arguments{
\item{factors}{some portion of the experimental design}

\item{sv}{a set of surrogate variable estimations from sva/svg or batch estimates}
}
\description{
This should make a quick df of the factors and surrogates and plot them.
}
\examples{
\dontrun{
estimate_vs_snps <- plot_svfactor(start, surrogate_estimate, "snpcategory")
}
}

