% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{plot_topgo_densities}
\alias{plot_topgo_densities}
\title{plot_topgo_densities()  Plot the density of categories vs. the possibilities}
\usage{
plot_topgo_densities(godata, table)
}
\arguments{
\item{godata}{the result from topgo}

\item{table}{a table of genes}
}
\description{
This can make a large number of plots
}

