% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.r
\name{plot_topgo_pval}
\alias{plot_topgo_pval}
\title{Make a pvalue plot from topgo data.}
\usage{
plot_topgo_pval(topgo, wrapped_width = 20, cutoff = 0.1, n = 30,
  type = "fisher", ...)
}
\arguments{
\item{topgo}{Some data from topgo!}

\item{wrapped_width}{Maximum width of the text names.}

\item{cutoff}{P-value cutoff for the plots.}

\item{n}{Maximum number of ontologies to include.}

\item{type}{Type of score to use.}

\item{...}{arguments passed through presumably from simple_topgo()}
}
\value{
List of MF/BP/CC pvalue plots.
}
\description{
The p-value plots from clusterProfiler are pretty, this sets the topgo data
into a format suitable for plotting in that fashion and returns the resulting
plots of significant ontologies.
}
\seealso{
\pkg{topgo} \pkg{clusterProfiler}
}
