% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_topn}
\alias{plot_topn}
\title{Plot the representation of the top-n genes in the total counts / sample.}
\usage{
plot_topn(
  data,
  plot_title = NULL,
  num = 100,
  sample_names = NULL,
  plot_labels = NULL,
  label_chars = 10,
  plot_legend = FALSE,
  ...
)
}
\arguments{
\item{data}{Dataframe/matrix/whatever for performing topn-plot.}

\item{plot_title}{A title for the plot.}

\item{num}{The N in top-n genes, if null, do them all.}

\item{sample_names}{Column or character list of sample names.}

\item{plot_labels}{Method for labelling the lines.}

\item{label_chars}{Maximum number of characters before abbreviating samples.}

\item{plot_legend}{Add a legend to the plot?}

\item{...}{Extra arguments, currently unused.}
}
\value{
List containing the ggplot2
}
\description{
One question we might ask is: how much do the most abundant genes in a
samples comprise the entire sample?  This plot attempts to provide a visual
hint toward answering this question.  It does so by rank-ordering all the
genes in every sample and dividing their counts by the total number of reads
in that sample.  It then smooths the points to provide the resulting trend.
The steeper the resulting line, the more over-represented these top-n genes
are.  I suspect, but haven't tried yet, that the inflection point of the
resulting curve is also a useful diagnostic in this question.
}
\examples{
library(hpgltools)
pombe_se <- get0("pombe_se")
if (is.null(pombe_se)) {
  source(system.file("examples", "create_pombe.R", package = "hpgltools"))
}

## Most distribution plots are particularly informative with the raw data.

## Make a boxplot of the SE
a_boxplot <- plot_boxplot(pombe_se)
a_boxplot

## The analagous density plot of the samples
a_density <- plot_density(pombe_se)
a_density

pombe_subset <- pombe_se[, c(1, 2, 3, 4)]
## QQ plots comparing samples to each other
qq <- plot_qq_all(pombe_subset)
qq[["logs"]]

## Show the percentage of all reads comprised by the top-n genes.
## This might be smarter if moved to the log scale?
topn_genes <- plot_topn(pombe_subset)
topn_genes

## Create a violin of the coefficient of variance on a per-condition basis.
varcoef <- plot_variance_coefficients(pombe_se)
varcoef
}
