% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.R
\name{plot_topn_gsea}
\alias{plot_topn_gsea}
\title{Iterate gseaplot2 over an arbitrary number of enrichments.}
\usage{
plot_topn_gsea(gse, topn = 20, id = NULL, add_score = TRUE)
}
\arguments{
\item{gse}{clusterProfiler GSEA result.}

\item{topn}{Number of enrichments to plot.}

\item{id}{Focus on a specific category.}

\item{add_score}{Add the score to the plot.}
}
\value{
List of plots.
}
\description{
Iterate gseaplot2 over an arbitrary number of enrichments.
}
