% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tsne.r
\name{plot_tsne_genes}
\alias{plot_tsne_genes}
\title{Plot tnse data for the genes in a data set.}
\usage{
plot_tsne_genes(data, design = NULL, plot_colors = NULL, seed = 1,
  chosen_features = NULL, number_features = NULL, perplexity = NULL,
  min_variance = 0.01, plot_title = NULL, components = 2,
  iterations = 1000, theta = 0.3, pca = TRUE, component_x = 1,
  component_y = 2, ...)
}
\arguments{
\item{data}{Some data!}

\item{design}{a design!}

\item{plot_colors}{Some colors!}

\item{seed}{for tsne}

\item{chosen_features}{Use these genes for labeling}

\item{number_features}{Somethingsomething}

\item{perplexity}{for tsne}

\item{min_variance}{Only include genes with more variance than this.}

\item{plot_title}{A title!}

\item{components}{How many components to plot.}

\item{iterations}{Use x tsne iterations.}

\item{theta}{Yay greek!}

\item{pca}{Seed this with an initial pca plot?}

\item{component_x}{Put which component on the x-axis?}

\item{component_y}{And which component on the y-axis?}

\item{...}{Arglist arguments.}
}
\description{
Plot tnse data for the genes in a data set.
}
