% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_variance_coefficients,ExpressionSet-method}
\alias{plot_variance_coefficients,ExpressionSet-method}
\title{Plot the coefficient of variance values of an ExpressionSet.}
\usage{
\S4method{plot_variance_coefficients}{ExpressionSet}(
  data,
  design = NULL,
  x_axis = "condition",
  colors = NULL,
  plot_title = NULL,
  ...
)
}
\arguments{
\item{data}{Expressionset/epxt to poke at.}

\item{design}{Specify metadata if necessary.}

\item{x_axis}{Factor in the experimental design we may use to group the data
and calculate the dispersion metrics.}

\item{colors}{Set of colors to use when making the violins}

\item{plot_title}{Optional title to include with the plot.}

\item{...}{Extra arguments to pass along.}
}
\description{
Plot the coefficient of variance values of an ExpressionSet.
}
