% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_variance_coefficients,expt-method}
\alias{plot_variance_coefficients,expt-method}
\title{Plot the coefficient of variance values of a SummarizedExperiment.}
\usage{
\S4method{plot_variance_coefficients}{expt}(
  data,
  design = NULL,
  x_axis = "condition",
  colors = NULL,
  plot_title = NULL,
  ...
)
}
\description{
Plot the coefficient of variance values of a SummarizedExperiment.
}
