% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{plot_volcano_de}
\alias{plot_volcano_de}
\title{Make a pretty Volcano plot!}
\usage{
plot_volcano_de(table, alpha = 0.6, color_by = "p",
  color_list = c(`FALSE` = "darkred", `TRUE` = "darkblue"),
  fc_col = "logFC", fc_name = "log2 fold change",
  gvis_filename = NULL, line_color = "black",
  line_position = "bottom", logfc_cutoff = 1, p_col = "adj.P.Val",
  p_name = "-log10 p-value", pval_cutoff = 0.05,
  shapes_by_state = TRUE, size = 2, tooltip_data = NULL, ...)
}
\arguments{
\item{table}{Dataframe from limma's toptable which includes log(fold change) and an
adjusted p-value.}

\item{alpha}{How transparent to make the dots.}

\item{color_by}{By p-value something else?}

\item{color_list}{A list of colors for significance.}

\item{fc_col}{Which column contains the fc data?}

\item{fc_name}{Name of the fold-change to put on the plot.}

\item{gvis_filename}{Filename to write a fancy html graph.}

\item{line_color}{What color for the significance lines?}

\item{line_position}{Put the significance lines above or below the dots?}

\item{logfc_cutoff}{Cutoff defining the minimum/maximum fold change for interesting.  This is log,
so I went with +/- 0.8 mostly arbitrarily as the default.}

\item{p_col}{Which column contains the p-value data?}

\item{p_name}{Name of the p-value to put on the plot.}

\item{pval_cutoff}{Cutoff defining significant from not.}

\item{shapes_by_state}{Add fun shapes for the various significance states?}

\item{size}{How big are the dots?}

\item{tooltip_data}{Df of tooltip information for gvis.}

\item{...}{I love parameters!}
}
\value{
Ggplot2 volcano scatter plot.  This is defined as the -log10(p-value) with respect to
log(fold change).  The cutoff values are delineated with lines and mark the boundaries
between 'significant' and not.  This will make a fun clicky googleVis graph if requested.
}
\description{
Volcano plots and MA plots provide quick an easy methods to view the set of (in)significantly
differentially expressed genes.  In the case of a volcano plot, it places the -log10 of the
p-value estimate on the y-axis and the fold-change between conditions on the x-axis.  Here is a
neat snippet from wikipedia: "The concept of volcano plot can be generalized to other
applications, where the x-axis is related to a measure of the strength of a statistical signal,
and y-axis is related to a measure of the statistical significance of the signal."
}
\examples{
\dontrun{
 plot_volcano_de(table, gvis_filename="html/fun_ma_plot.html")
 ## Currently this assumes that a variant of toptable was used which
 ## gives adjusted p-values.  This is not always the case and I should
 ## check for that, but I have not yet.
}
}
\seealso{
\pkg{limma}
\code{\link{plot_gvis_ma}} \code{\link[limma]{toptable}}
\code{\link[limma]{voom}} \code{\link{hpgl_voom}} \code{\link[limma]{lmFit}}
\code{\link[limma]{makeContrasts}} \code{\link[limma]{contrasts.fit}}
}
