% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.r
\name{plotly_pca}
\alias{plotly_pca}
\title{Plot a PC plot with options suitable for ggplotly.}
\usage{
plotly_pca(
  data,
  design = NULL,
  plot_colors = NULL,
  plot_title = NULL,
  plot_size = 5,
  plot_alpha = NULL,
  plot_labels = NULL,
  size_column = NULL,
  pc_method = "fast_svd",
  x_pc = 1,
  y_pc = 2,
  outlines = FALSE,
  num_pc = NULL,
  expt_names = NULL,
  label_chars = 10,
  tooltip = c("shape", "fill", "sampleid"),
  ...
)
}
\arguments{
\item{data}{an expt set of samples.}

\item{design}{a design matrix and.}

\item{plot_colors}{a color scheme.}

\item{plot_title}{a title for the plot.}

\item{plot_size}{size for the glyphs on the plot.}

\item{plot_alpha}{Add an alpha channel to the dots?}

\item{plot_labels}{add labels?  Also, what type?  FALSE, "default", or "fancy".}

\item{size_column}{use an experimental factor to size the glyphs of the plot}

\item{pc_method}{how to extract the components? (svd}

\item{x_pc}{Component to put on the x axis.}

\item{y_pc}{Component to put on the y axis.}

\item{outlines}{Include black outlines around glyphs?}

\item{num_pc}{How many components to calculate, default to the number of
rows in the metadata.}

\item{expt_names}{Column or character list of preferred sample names.}

\item{label_chars}{Maximum number of characters before abbreviating sample names.}

\item{tooltip}{Which columns to include in the tooltip.}

\item{...}{Arguments passed through to the pca implementations and plotter.}
}
\value{
This passes directly to plot_pca(), so its returns should be
  applicable along with the result from ggplotly.
}
\description{
Plot a PC plot with options suitable for ggplotly.
}
