% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{post_eupath_annotations}
\alias{post_eupath_annotations}
\title{Gather all available annotation data for a given eupathdb species.}
\usage{
post_eupath_annotations(species = "Leishmania major", entry = NULL,
  metadata = NULL, dir = "eupathdb")
}
\arguments{
\item{species}{guess.}

\item{entry}{The full annotation entry.}

\item{metadata}{A metadata table from which to get some annotation data.}

\item{dir}{FIXME: I want to write some intermediate data to dir in case of transient error.}
}
\value{
A big honking table.
}
\description{
This function fills in the parameters to post_eupath_raw() so that one can download
all the available data for a given parasite into one massive table.  It should also
provide some constraints to the data rather than leaving it all as characters.
}
