% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupath_webservices.r
\name{post_eupath_interpro_table}
\alias{post_eupath_interpro_table}
\title{Use the post interface to get interpro data.}
\usage{
post_eupath_interpro_table(species = "Leishmania major strain Friedlin",
  entry = NULL, metadata = NULL, dir = "eupathdb", ...)
}
\arguments{
\item{species}{guess.}

\item{entry}{The full annotation entry.}

\item{metadata}{A metadata table from which to get some annotation data.}

\item{dir}{FIXME: I want to write some intermediate data to dir in case of transient error.}

\item{...}{Extra options when downloading metadata.}
}
\value{
A big honking table.
}
\description{
Use the post interface to get interpro data.
}
