% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{post_eupath_ortholog_table}
\alias{post_eupath_ortholog_table}
\title{Use the post interface to get ortholog data.}
\usage{
post_eupath_ortholog_table(species = "Leishmania major", entry = NULL,
  metadata = NULL, dir = "eupathdb")
}
\arguments{
\item{species}{guess.}

\item{entry}{The full annotation entry.}

\item{metadata}{A metadata table from which to get some annotation data.}

\item{dir}{FIXME: I want to write some intermediate data to dir in case of transient error.}
}
\value{
A big honking table.
}
\description{
Use the post interface to get ortholog data.
}
