% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupath_webservices.r
\name{post_eupath_raw}
\alias{post_eupath_raw}
\title{The new eupath system provides 3 output types for downloading data.  This
uses the raw one.}
\usage{
post_eupath_raw(entry, question = "GeneQuestions.GenesByMolecularWeight",
  table_name = NULL, parameters = NULL, columns = "primary_key",
  minutes = 40)
}
\arguments{
\item{entry}{Annotation entry for a given species}

\item{question}{Which query to try?  Molecular weight is the easiest, as it was their example.}

\item{table_name}{Used to make sure all columns are unique by prefixing them with the table name.}

\item{parameters}{Query parameters when posting}

\item{columns}{Columns for which to ask.}

\item{minutes}{How long to wait until giving up and throwing an error.}
}
\value{
A hopefully huge table of eupath data.
}
\description{
For the life of me, I could not figure out how to query the big text tables as the
tabular format.  Every query I sent came back telling me I gave it incorrect parameter
despite the fact that I was copy/pasting the example given me by the eupathdb maintainers.
So, I got mad and asked it for the raw format, and so this function was born.
}
