% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupath_webservices.r
\name{post_eupath_table}
\alias{post_eupath_table}
\title{Queries one of the EuPathDB APIs using a POST request and returns a
dataframe representation of the result.
Note: As of 2017/07/13, POST requests are not yet supported on EuPathDB.
Note: 2017/07/13 POST queries can only use the new API}
\usage{
post_eupath_table(query_body, species = NULL, entry = NULL,
  metadata = NULL, table_name = NULL, minutes = 30, ...)
}
\arguments{
\item{query_body}{String of additional query arguments}

\item{species}{Species name if missing an entry}

\item{entry}{The single metadatum containing the base url of the provider, species, etc.}

\item{metadata}{If no entry is provided, then it may be retrieved given a species and this.}

\item{table_name}{The name of the table to extract, this is provided to make for prettier labeling.}

\item{minutes}{A timeout when querying the eupathdb.}

\item{...}{Extra arguments for stuff like download_metadtata()}
}
\value{
list containing response from API request.More information
\enumerate{
\item https://tritrypdb.org/tritrypdb/serviceList.jsp
}
}
\description{
Queries one of the EuPathDB APIs using a POST request and returns a
dataframe representation of the result.
Note: As of 2017/07/13, POST requests are not yet supported on EuPathDB.
Note: 2017/07/13 POST queries can only use the new API
}
\author{
Keith Hughitt
}
