% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.R
\name{pp}
\alias{pp}
\title{Plot a picture, with hopefully useful options for most(any) format.}
\usage{
pp(
  file,
  image = NA,
  width = 9,
  height = 9,
  res = 180,
  crop = FALSE,
  old = FALSE,
  passthrough = TRUE,
  ...
)
}
\arguments{
\item{file}{Filename to write}

\item{image}{Optionally, add the image you wish to plot and this will both
print it to file and screen.}

\item{width}{Image width in inches.}

\item{height}{Image height in inches.}

\item{res}{The chosen resolution.}

\item{crop}{Crop off the edges of the plot?}

\item{old}{A couple of plotters have new and old implementations (svg),
use the old one?}

\item{passthrough}{Pass the plot through to the default plotter after printint it?}

\item{...}{Arguments passed to the image plotters.}
}
\value{
a png/svg/eps/ps/pdf with height = width=9 inches and a high resolution
}
\description{
This calls svg/png/postscript/etc according to the filename provided.
}
\seealso{
[png()] [svg()] [postscript()] [cairo_ps()] [cairo_pdf()] [tiff()] [devEMF::emf()]
 [jpg()] [bmp()]
}
