% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{primer_qc}
\alias{primer_qc}
\title{Perform a series of tests of a putative primer.}
\usage{
primer_qc(
  entry,
  genome,
  variant_gr,
  target_temp = 60,
  direction = "fwd",
  run_pattern = "AAAA",
  min_gc_prop = 0.3,
  seq_object = NULL
)
}
\arguments{
\item{entry}{Single row of the table of potential primers.}

\item{genome}{bsgenome used to search against.}

\item{variant_gr}{GRanges of variants to xref against.}

\item{target_temp}{Desired Tm.}

\item{direction}{Either fwd or rev.}

\item{run_pattern}{Regex to look for bad runs of a single nt.}

\item{min_gc_prop}{Minimum proportion of GC content.}

\item{seq_object}{Used to hunt for multi-hit primers.}
}
\description{
This function should probably replace the morass of code found in
snp_density_primers().  It is current used by snp_cds_primers().
}
