% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.abundant_genes}
\alias{print.abundant_genes}
\title{Print a summary of a set of abundant genes.}
\usage{
\method{print}{abundant_genes}(x)
}
\arguments{
\item{x}{Abundant gene list.}
}
\description{
Print a summary of a set of abundant genes.
}
