% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.all_gprofiler}
\alias{print.all_gprofiler}
\title{Print the result of all_gprofiler()}
\usage{
\method{print}{all_gprofiler}(x)
}
\arguments{
\item{x}{The set of results produced by all_gprofiler().}
}
\description{
Print the result of all_gprofiler()
}
