% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.all_pairwise}
\alias{print.all_pairwise}
\title{Print function for a pairwise differential expression result.}
\usage{
\method{print}{all_pairwise}(x)
}
\arguments{
\item{x}{pairwise comparisons.}
}
\description{
Print function for a pairwise differential expression result.
}
