% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.annotations_biomart}
\alias{print.annotations_biomart}
\title{Print function for a set of annotations downloaded from biomart.}
\usage{
\method{print}{annotations_biomart}(x, ...)
}
\arguments{
\item{x}{List containing the relevant information gathered from ensembl's biomart.}

\item{...}{Other args to match the generic.}
}
\description{
Print function for a set of annotations downloaded from biomart.
}
