% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.aucc_info}
\alias{print.aucc_info}
\title{Print the result from calculate_aucc().}
\usage{
\method{print}{aucc_info}(x)
}
\arguments{
\item{x}{List containing the AUCC, correlation, and a plot
describing the AUCC.}
}
\description{
Print the result from calculate_aucc().
}
