% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.cds_variant_primers}
\alias{print.cds_variant_primers}
\title{Print the results of snp_cds_primers().}
\usage{
\method{print}{cds_variant_primers}(x, ...)
}
\arguments{
\item{x}{List currently only containing the dataframe of putative primers.}

\item{...}{Other args to match the generic.}
}
\description{
Print the results of snp_cds_primers().
}
