% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.classified_mutations}
\alias{print.classified_mutations}
\title{Print the results of classify_snps().}
\usage{
\method{print}{classified_mutations}(x, ...)
}
\arguments{
\item{x}{List containing some fun stats of variants observed.}

\item{...}{Other args to match the generic.}
}
\description{
Print the results of classify_snps().
}
