% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.classifier_evaluation}
\alias{print.classifier_evaluation}
\title{Print a summary of the result from self_evaluate_model().}
\usage{
\method{print}{classifier_evaluation}(x)
}
\arguments{
\item{x}{list of results from self_evaluate_model().}
}
\description{
Print a summary of the result from self_evaluate_model().
}
