% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.clusterprofiler_result}
\alias{print.clusterprofiler_result}
\title{Print a summary of a clusterprofiler over representation search.}
\usage{
\method{print}{clusterprofiler_result}(x)
}
\arguments{
\item{x}{list of results from clusterprofiler.}
}
\description{
Print a summary of a clusterprofiler over representation search.
}
