% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{print.combined_de_upset}
\alias{print.combined_de_upset}
\title{Print a summary from combine_de_upset}
\usage{
\method{print}{combined_de_upset}(x, ...)
}
\arguments{
\item{x}{List produced by combined_de_upset}

\item{...}{Other args for the generic.}
}
\description{
Print a summary from combine_de_upset
}
