% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{print.combined_table}
\alias{print.combined_table}
\title{Print a single combined DE result.}
\usage{
\method{print}{combined_table}(x, ...)
}
\arguments{
\item{x}{Data table of combined differential expression results.}

\item{...}{Other args to match the generic.}
}
\description{
Print a single combined DE result.
}
