% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.cross_table_comparison}
\alias{print.cross_table_comparison}
\title{Print a representation of compare_de_tables().
Note I think I want to have that function return slightly different types
depending on how the function call was set up.}
\usage{
\method{print}{cross_table_comparison}(x, ...)
}
\arguments{
\item{x}{List provided by plot_linear_scatter() containing
correlations, plots, linear model.}

\item{...}{Other args to match the generic.}
}
\description{
Print a representation of compare_de_tables().
Note I think I want to have that function return slightly different types
depending on how the function call was set up.
}
