% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.density_plot}
\alias{print.density_plot}
\title{Print a summary of the result from plot_density().}
\usage{
\method{print}{density_plot}(x)
}
\arguments{
\item{x}{The density plot and associated data in question.}
}
\description{
Print a summary of the result from plot_density().
}
