% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.density_primers}
\alias{print.density_primers}
\title{Print a summary of putative PCR primers based on variant density.}
\usage{
\method{print}{density_primers}(x, ...)
}
\arguments{
\item{x}{List from snp_density_primers()}

\item{...}{Other args for the generic.}
}
\description{
Print a summary of putative PCR primers based on variant density.
}
