% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.deseq_lrt}
\alias{print.deseq_lrt}
\title{Print the result of running deseq_lrt().}
\usage{
\method{print}{deseq_lrt}(x)
}
\arguments{
\item{x}{List containing the DESeq2 result, the associated table,
clusters from degPatterns, list of associated genes, and dataframes
of the most significant genes.}
}
\description{
Print the result of running deseq_lrt().
}
