% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.distance_heatmap}
\alias{print.distance_heatmap}
\title{Print the result of plot_disheat().}
\usage{
\method{print}{distance_heatmap}(x, ...)
}
\arguments{
\item{x}{List containing the distances observed and a recorded heatmap.3().}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of plot_disheat().
}
