% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.dream_pairwise}
\alias{print.dream_pairwise}
\title{Print a summary of the result from dream_pairwise().}
\usage{
\method{print}{dream_pairwise}(x, ...)
}
\arguments{
\item{x}{List from dream_pairwise().}

\item{...}{Other args for the generic.}
}
\description{
Print a summary of the result from dream_pairwise().
}
