% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.expt}
\alias{print.expt}
\title{Modified print function for an expt.}
\usage{
\method{print}{expt}(x)
}
\arguments{
\item{x}{List from create_expt containing the expressionSet,
annotation data, batches, conditions, colors, libsizes, etc.}
}
\description{
I am trying to understand how R collates functions.
}
