% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{print.gene_boxplot}
\alias{print.gene_boxplot}
\title{Print a result from plot_boxplot().}
\usage{
\method{print}{gene_boxplot}(x, ...)
}
\arguments{
\item{x}{List with the topn plot and summary table.}

\item{...}{Other args to match the generic.}
}
\description{
Print a result from plot_boxplot().
}
