% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.goseq_result}
\alias{print.goseq_result}
\title{Print a goseq over representation search.}
\usage{
\method{print}{goseq_result}(x, ...)
}
\arguments{
\item{x}{List containing the various goseq results, plots,
significant subsets, enrichResult coercions, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print a goseq over representation search.
}
