% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.gsva_result}
\alias{print.gsva_result}
\title{Print a summary of a gsva category search.}
\usage{
\method{print}{gsva_result}(x)
}
\arguments{
\item{x}{list of results from gprofiler.}
}
\description{
Print a summary of a gsva category search.
}
