% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.gsva_sig}
\alias{print.gsva_sig}
\title{Print a summary of gsva categories deemed 'significant'.}
\usage{
\method{print}{gsva_sig}(x)
}
\arguments{
\item{x}{list of results from gprofiler.}
}
\description{
Print a summary of gsva categories deemed 'significant'.
}
