% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.R
\name{print.hpgltools::all_gprofiler}
\alias{print.hpgltools::all_gprofiler}
\title{Print the result of all_gprofiler()}
\usage{
\method{print}{`hpgltools::all_gprofiler`}(x, ...)
}
\arguments{
\item{x}{List composed of simple_gprofiler() results for every
up/down set provided by extract_significant_genes().}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of all_gprofiler()
}
