% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{print.hpgltools::calculate_aucc}
\alias{print.hpgltools::calculate_aucc}
\title{Print the result from calculate_aucc().}
\usage{
\method{print}{`hpgltools::calculate_aucc`}(x, ...)
}
\arguments{
\item{x}{List containing the AUCC, correlation, and a plot
describing the AUCC.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from calculate_aucc().
}
