% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{print.hpgltools::classify_variants}
\alias{print.hpgltools::classify_variants}
\title{Print the results of classify_snps().}
\usage{
\method{print}{`hpgltools::classify_variants`}(x, ...)
}
\arguments{
\item{x}{List containing some fun stats of variants observed.}

\item{...}{Other args to match the generic.}
}
\description{
Print the results of classify_snps().
}
