% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{print.hpgltools::combine_de_tables}
\alias{print.hpgltools::combine_de_tables}
\title{Print a combined differential expression analysis.}
\usage{
\method{print}{`hpgltools::combine_de_tables`}(x, ...)
}
\arguments{
\item{x}{List containing the dataframes for each contrast, the
various plots, the set of wanted contrasts, models used, and
summaries of the data.}

\item{...}{Other args to match the generic.}
}
\description{
Print a combined differential expression analysis.
}
