% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{print.hpgltools::create_partitions}
\alias{print.hpgltools::create_partitions}
\title{Print something useful about the result of create_partitions()}
\usage{
\method{print}{`hpgltools::create_partitions`}(x, ...)
}
\arguments{
\item{x}{List containing the n sets of partitioned data test/train.}

\item{...}{Other args to match the generic.}
}
\description{
Print something useful about the result of create_partitions()
}
