% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_varpart.R
\name{print.hpgltools::dream_pairwise}
\alias{print.hpgltools::dream_pairwise}
\title{Print a summary of the result from dream_pairwise().}
\usage{
\method{print}{`hpgltools::dream_pairwise`}(x, ...)
}
\arguments{
\item{x}{List from dream_pairwise().}

\item{...}{Other args for the generic.}
}
\description{
Print a summary of the result from dream_pairwise().
}
