% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{print.hpgltools::gather_preprocessing_metadata}
\alias{print.hpgltools::gather_preprocessing_metadata}
\title{Print the result from gather_preprocessing_metadata().}
\usage{
\method{print}{`hpgltools::gather_preprocessing_metadata`}(x, ...)
}
\arguments{
\item{x}{List composed of the xlsx output file, new columns added
to it, and copies of the metadata before/after modification.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from gather_preprocessing_metadata().
}
