% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{print.hpgltools::get_abundant_genes}
\alias{print.hpgltools::get_abundant_genes}
\title{Print a summary of a set of abundant genes.}
\usage{
\method{print}{`hpgltools::get_abundant_genes`}(x, ...)
}
\arguments{
\item{x}{Abundant gene list comprising an element 'high' and 'low'
for the most and least abundant genes observed.}

\item{...}{Other args to match the generic.}
}
\description{
Print a summary of a set of abundant genes.
}
