% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.R
\name{print.hpgltools::get_sig_gsva_categories}
\alias{print.hpgltools::get_sig_gsva_categories}
\title{Print gsva categories deemed 'significant'.}
\usage{
\method{print}{`hpgltools::get_sig_gsva_categories`}(x, ...)
}
\arguments{
\item{x}{List of scored GSVA results, including some plots,
likelihood tables, subsets of significant categories, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print gsva categories deemed 'significant'.
}
