% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{print.hpgltools::get_snp_sets}
\alias{print.hpgltools::get_snp_sets}
\title{Print the result of get_snp_sets().}
\usage{
\method{print}{`hpgltools::get_snp_sets`}(x, ...)
}
\arguments{
\item{x}{List containing the cross references of variants by
factor, the set of observed variants, the possible combinations of
the factor, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of get_snp_sets().
}
