% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.R
\name{print.hpgltools::graph_metrics}
\alias{print.hpgltools::graph_metrics}
\title{Print a message about the results from graph_metrics().}
\usage{
\method{print}{`hpgltools::graph_metrics`}(x, ...)
}
\arguments{
\item{x}{List containing a large number of plots and some tables.}

\item{...}{Other args to match the generic.}
}
\description{
Print a message about the results from graph_metrics().
}
