% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{print.hpgltools::intersect_significant}
\alias{print.hpgltools::intersect_significant}
\title{Print the intersection of significant genes from multiple analyses.}
\usage{
\method{print}{`hpgltools::intersect_significant`}(x, ...)
}
\arguments{
\item{x}{List containing some venn diagrams, summaries of
intersections, subsets of the intersections, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print the intersection of significant genes from multiple analyses.
}
