% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{print.hpgltools::load_biomart_go}
\alias{print.hpgltools::load_biomart_go}
\title{Print the results of load_biomart_go().}
\usage{
\method{print}{`hpgltools::load_biomart_go`}(x, ...)
}
\arguments{
\item{x}{List from load_biomart_go() containing the table of data,
mart used, host used, name of the mart, and attributes.}

\item{...}{Other args to match the generic.}
}
\description{
Print the results of load_biomart_go().
}
