% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{print.hpgltools::map_keepers}
\alias{print.hpgltools::map_keepers}
\title{Print a set of mapped keepers from combine_de_tables()}
\usage{
\method{print}{`hpgltools::map_keepers`}(x, ...)
}
\arguments{
\item{x}{List full of kept information.}

\item{...}{Other args to match the generic.}
}
\description{
Print a set of mapped keepers from combine_de_tables()
}
