% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_noiseq.R
\name{print.hpgltools::noiseq_pairwise}
\alias{print.hpgltools::noiseq_pairwise}
\title{Print the results from noiseq_pairwise().}
\usage{
\method{print}{`hpgltools::noiseq_pairwise`}(x, ...)
}
\arguments{
\item{x}{result to print.}

\item{...}{arbitrary arguments.}
}
\description{
Print the results from noiseq_pairwise().
}
