% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{print.hpgltools::pattern_count_genome}
\alias{print.hpgltools::pattern_count_genome}
\title{Print some information about a pattern counted genome}
\usage{
\method{print}{`hpgltools::pattern_count_genome`}(x, ...)
}
\arguments{
\item{x}{Dataframe containing how many instances of the pattern
were observed in every gene.}

\item{...}{Other args to match the generic.}
}
\description{
Print some information about a pattern counted genome
}
