% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{print.hpgltools::plot_corheat}
\alias{print.hpgltools::plot_corheat}
\title{Print the result of plot_corheat().}
\usage{
\method{print}{`hpgltools::plot_corheat`}(x, ...)
}
\arguments{
\item{x}{List containing the correlations observed and a recorded heatmap.3().}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of plot_corheat().
}
