% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{print.hpgltools::plot_density}
\alias{print.hpgltools::plot_density}
\title{Print the result from plot_density().}
\usage{
\method{print}{`hpgltools::plot_density`}(x, ...)
}
\arguments{
\item{x}{List containing the plot, summary of the
conditions/batches/samples, and the melted table of reads/gene.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from plot_density().
}
