% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{print.hpgltools::plot_disheat}
\alias{print.hpgltools::plot_disheat}
\title{Print the result of plot_disheat().}
\usage{
\method{print}{`hpgltools::plot_disheat`}(x, ...)
}
\arguments{
\item{x}{List containing the distances observed and a recorded heatmap.3().}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of plot_disheat().
}
