% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{print.hpgltools::plot_ggcoverage}
\alias{print.hpgltools::plot_ggcoverage}
\title{Print a gg coverage plot}
\usage{
\method{print}{`hpgltools::plot_ggcoverage`}(x, ...)
}
\arguments{
\item{x}{List containing some information and the plot of interest.}

\item{...}{Other args for the generic.}
}
\description{
FIXME: Add something useful to this
}
