% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.R
\name{print.hpgltools::plot_legend}
\alias{print.hpgltools::plot_legend}
\title{Print a legend of an expressionset.}
\usage{
\method{print}{`hpgltools::plot_legend`}(x, ...)
}
\arguments{
\item{x}{List containing the condition factor, colors used, and plot.}

\item{...}{Other args to match the generic.}
}
\description{
Print a legend of an expressionset.
}
