% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{print.hpgltools::plot_libsize_prepost}
\alias{print.hpgltools::plot_libsize_prepost}
\title{Print a representation of the pre vs. post filtered data.}
\usage{
\method{print}{`hpgltools::plot_libsize_prepost`}(x, ...)
}
\arguments{
\item{x}{List containing the information before/after filtering,
the plots, and summary information.}

\item{...}{Other args to match the generic.}
}
\description{
Print a representation of the pre vs. post filtered data.
}
