% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{print.hpgltools::plot_linear_scatter}
\alias{print.hpgltools::plot_linear_scatter}
\title{Print a linear scatter plot without all the fluff.}
\usage{
\method{print}{`hpgltools::plot_linear_scatter`}(x, ...)
}
\arguments{
\item{x}{List containing the result of plot_linear_scatter()}

\item{...}{Other args for the generic.}
}
\description{
Print a linear scatter plot without all the fluff.
}
