% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{print.hpgltools::plot_nonzero}
\alias{print.hpgltools::plot_nonzero}
\title{Print a nonzero plot.}
\usage{
\method{print}{`hpgltools::plot_nonzero`}(x, ...)
}
\arguments{
\item{x}{List containing the plot and table describing the data.}

\item{...}{Other args to match the generic.}
}
\description{
Print a nonzero plot.
}
