% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{print.hpgltools::plot_pca}
\alias{print.hpgltools::plot_pca}
\title{Print the result from one of the various dimension reductions.}
\usage{
\method{print}{`hpgltools::plot_pca`}(x, ...)
}
\arguments{
\item{x}{List comprised of the residuals, variance summary, tables,
the PCA-esque plot, experimental design, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from one of the various dimension reductions.
}
