% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{print.hpgltools::plot_quantreads}
\alias{print.hpgltools::plot_quantreads}
\title{Print the library sizes from an experiment.}
\usage{
\method{print}{`hpgltools::plot_quantreads`}(x, ...)
}
\arguments{
\item{x}{List containing a summary of the library sizes, the plot,
and table.}

\item{...}{Other args to match the generic.}
}
\description{
Print the library sizes from an experiment.
}
