% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{print.hpgltools::plot_sm}
\alias{print.hpgltools::plot_sm}
\title{Print the result of plot_sm()}
\usage{
\method{print}{`hpgltools::plot_sm`}(x, ...)
}
\arguments{
\item{x}{List containing the pairwise distances/correlations, median/mean values,
quartiles, and the standard median plot.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of plot_sm()
}
