% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_varpartition.R
\name{print.hpgltools::replit_varpart_percent}
\alias{print.hpgltools::replit_varpart_percent}
\title{Print the result of a reordered variance partition analysis.}
\usage{
\method{print}{`hpgltools::replit_varpart_percent`}(x, ...)
}
\arguments{
\item{x}{List of a resorted variance partition analysis and its plot.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of a reordered variance partition analysis.
}
