% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{print.hpgltools::self_evaluate_model}
\alias{print.hpgltools::self_evaluate_model}
\title{Print the result from self_evaluate_model().}
\usage{
\method{print}{`hpgltools::self_evaluate_model`}(x, ...)
}
\arguments{
\item{x}{List showing AUC/ROC curves of the test performed, summary
thereof, the confusion matrix, and vector of incorrectly called samples.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from self_evaluate_model().
}
