% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{print.hpgltools::simple_clusterprofiler}
\alias{print.hpgltools::simple_clusterprofiler}
\title{Print a clusterprofiler over representation search.}
\usage{
\method{print}{`hpgltools::simple_clusterprofiler`}(x, ...)
}
\arguments{
\item{x}{Monstrous list of the various results, including but not
limited to plots, go-gene mappings, enrichmed, kegg, david, GO
analyses.}

\item{...}{Other args to match the generic.}
}
\description{
Print a clusterprofiler over representation search.
}
