% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gostats.R
\name{print.hpgltools::simple_gostats}
\alias{print.hpgltools::simple_gostats}
\title{Print a gostats over representation search.}
\usage{
\method{print}{`hpgltools::simple_gostats`}(x, ...)
}
\arguments{
\item{x}{List containing the various gostats results, plots,
significant subsets, enrichResult coercions, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print a gostats over representation search.
}
