% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.R
\name{print.hpgltools::simple_gsva}
\alias{print.hpgltools::simple_gsva}
\title{Print a gsva category search.}
\usage{
\method{print}{`hpgltools::simple_gsva`}(x, ...)
}
\arguments{
\item{x}{List containing signature annotations, the result from
GSVA, a modified SE, the signatures used, and method.}

\item{...}{Other args to match the generic.}
}
\description{
Print a gsva category search.
}
