% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{print.hpgltools::simple_proper}
\alias{print.hpgltools::simple_proper}
\title{Print the result from simple_proper().}
\usage{
\method{print}{`hpgltools::simple_proper`}(x, ...)
}
\arguments{
\item{x}{List including the various plots from PROPER, the
associated tables, simulation options, and example text for a
paper/grant.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from simple_proper().
}
