% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.R
\name{print.hpgltools::simple_topgo}
\alias{print.hpgltools::simple_topgo}
\title{Print a topgo over representation search.}
\usage{
\method{print}{`hpgltools::simple_topgo`}(x, ...)
}
\arguments{
\item{x}{List of the various over/under representation analyses
provided by topGO, the associated plots, and coerced enrichResults.}

\item{...}{Other args to match the generic.}
}
\description{
Print a topgo over representation search.
}
