% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_varpartition.R
\name{print.hpgltools::simple_varpart}
\alias{print.hpgltools::simple_varpart}
\title{Print variance partition results.}
\usage{
\method{print}{`hpgltools::simple_varpart`}(x, ...)
}
\arguments{
\item{x}{List of results from variancePartition including the model
information, percent/partition plots, dataframes of the
fitted/sorted data by variance, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print variance partition results.
}
