% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{print.hpgltools::snps_intersections}
\alias{print.hpgltools::snps_intersections}
\title{Print some information about the result of snp_intersections().}
\usage{
\method{print}{`hpgltools::snps_intersections`}(x, ...)
}
\arguments{
\item{x}{List containing a datatable of intersections, summaries by
chromosome and gene.}

\item{...}{Other args to match the generic.}
}
\description{
Print some information about the result of snp_intersections().
}
