% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{print.hpgltools::snps_vs_genes}
\alias{print.hpgltools::snps_vs_genes}
\title{Print the result of snps_vs_genes().}
\usage{
\method{print}{`hpgltools::snps_vs_genes`}(x, ...)
}
\arguments{
\item{x}{List containing granges of variants, variants observed by
chromosome, gene, and summaries of the result.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of snps_vs_genes().
}
