% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{print.hpgltools::write_xlsx}
\alias{print.hpgltools::write_xlsx}
\title{Print the result from write_xlsx.}
\usage{
\method{print}{`hpgltools::write_xlsx`}(x, ...)
}
\arguments{
\item{x}{List containing some information about the xlsx file.}

\item{...}{Other args for the generic.}
}
\description{
Print the result from write_xlsx.
}
