% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.legend_plot}
\alias{print.legend_plot}
\title{Print a legend of an expressionset.}
\usage{
\method{print}{legend_plot}(x)
}
\arguments{
\item{x}{list of results from plot_legend().}
}
\description{
Print a legend of an expressionset.
}
