% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.linear_scatter}
\alias{print.linear_scatter}
\title{Print a linear scatter plot without all the fluff.}
\usage{
\method{print}{linear_scatter}(x, ...)
}
\arguments{
\item{x}{List containing the result of plot_linear_scatter()}

\item{...}{Other args for the generic.}
}
\description{
Print a linear scatter plot without all the fluff.
}
