% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.mapped_keepers}
\alias{print.mapped_keepers}
\title{Print a set of mapped keepers from combine_de_tables()}
\usage{
\method{print}{mapped_keepers}(x)
}
\arguments{
\item{x}{List full of kept information.}
}
\description{
Print a set of mapped keepers from combine_de_tables()
}
