% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.meta_sankey}
\alias{print.meta_sankey}
\title{Print a summary of a metadata sankey plot.}
\usage{
\method{print}{meta_sankey}(x)
}
\arguments{
\item{x}{list of results from plot_meta_sankey().}
}
\description{
Print a summary of a metadata sankey plot.
}
