% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.nonzero_plot}
\alias{print.nonzero_plot}
\title{Print a summary of a nonzero plot.}
\usage{
\method{print}{nonzero_plot}(x)
}
\arguments{
\item{x}{list of results from plot_libsize().}
}
\description{
Print a summary of a nonzero plot.
}
