% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.R
\name{print.orgdb_annotations}
\alias{print.orgdb_annotations}
\title{Avoid annoyingly large print results of orgdb data.}
\usage{
\method{print}{orgdb_annotations}(x, ...)
}
\arguments{
\item{x}{Result from load_orgdb_annotations().}

\item{...}{pass along args}
}
\description{
Avoid annoyingly large print results of orgdb data.
}
