% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.partitioned_data}
\alias{print.partitioned_data}
\title{Print something useful about the result of create_partitions()}
\usage{
\method{print}{partitioned_data}(x)
}
\arguments{
\item{x}{List containing the n sets of partitioned data test/train.}
}
\description{
Print something useful about the result of create_partitions()
}
