% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.pca_result}
\alias{print.pca_result}
\title{Print the result from one of the various dimension reductions.}
\usage{
\method{print}{pca_result}(x)
}
\arguments{
\item{x}{List comprised of the residuals, variance summary, tables,
the PCA-esque plot, experimental design, etc.}
}
\description{
Print the result from one of the various dimension reductions.
}
