% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.prepost_filter}
\alias{print.prepost_filter}
\title{Print a representation of the pre vs. post filtered data.}
\usage{
\method{print}{prepost_filter}(x)
}
\arguments{
\item{x}{List containing the information before/after filtering,
the plots, and summary information.}
}
\description{
Print a representation of the pre vs. post filtered data.
}
