% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.preprocessing_metadata}
\alias{print.preprocessing_metadata}
\title{Print a summary of the result from gather_preprocessing_metadata().}
\usage{
\method{print}{preprocessing_metadata}(x)
}
\arguments{
\item{x}{Result from gather_preprocessing_metadata()}
}
\description{
Print a summary of the result from gather_preprocessing_metadata().
}
