% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.proper_estimate}
\alias{print.proper_estimate}
\title{Print the result from simple_proper().}
\usage{
\method{print}{proper_estimate}(x, ...)
}
\arguments{
\item{x}{List including the various plots from PROPER, the
associated tables, simulation options, and example text for a
paper/grant.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from simple_proper().
}
