% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.reordered_varpart}
\alias{print.reordered_varpart}
\title{Print the result of a reordered variance partition analysis.}
\usage{
\method{print}{reordered_varpart}(x)
}
\arguments{
\item{x}{List of a resorted variance partition analysis and its plot.}
}
\description{
Print the result of a reordered variance partition analysis.
}
