% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.sig_genes}
\alias{print.sig_genes}
\title{Print some significantly differentially expressed genes.}
\usage{
\method{print}{sig_genes}(x)
}
\arguments{
\item{x}{List containing the parameters used, gene subset tables,
plots, xlsx output file, etc.}
}
\description{
Print some significantly differentially expressed genes.
}
