% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.sig_intersect}
\alias{print.sig_intersect}
\title{Print the intersection of significant genes from multiple analyses.}
\usage{
\method{print}{sig_intersect}(x)
}
\arguments{
\item{x}{List containing some venn diagrams, summaries of
intersections, subsets of the intersections, etc.}
}
\description{
Print the intersection of significant genes from multiple analyses.
}
