% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.snp_intersections}
\alias{print.snp_intersections}
\title{Print some information about the result of snp_intersections().}
\usage{
\method{print}{snp_intersections}(x)
}
\arguments{
\item{x}{list of results from snp_intersections().}
}
\description{
Print some information about the result of snp_intersections().
}
