% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.snp_sets}
\alias{print.snp_sets}
\title{Print the result of get_snp_sets().}
\usage{
\method{print}{snp_sets}(x)
}
\arguments{
\item{x}{list of results from get_snp_sets().}
}
\description{
Print the result of get_snp_sets().
}
