% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.snps_genes}
\alias{print.snps_genes}
\title{Print the result of snps_vs_genes().}
\usage{
\method{print}{snps_genes}(x)
}
\arguments{
\item{x}{List containing granges of variants, variants observed by
chromosome, gene, and summaries of the result.}
}
\description{
Print the result of snps_vs_genes().
}
