% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.standardmedian_plot}
\alias{print.standardmedian_plot}
\title{Print the result of plot_sm()}
\usage{
\method{print}{standardmedian_plot}(x, ...)
}
\arguments{
\item{x}{List containing the pairwise distances/correlations, median/mean values,
quartiles, and the standard median plot.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of plot_sm()
}
