% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{print.topgo_result}
\alias{print.topgo_result}
\title{Print a summary of a topgo over representation search.}
\usage{
\method{print}{topgo_result}(x)
}
\arguments{
\item{x}{list of results from topgo.}
}
\description{
Print a summary of a topgo over representation search.
}
