% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.topn_plot}
\alias{print.topn_plot}
\title{Print a result from plot_topn().}
\usage{
\method{print}{topn_plot}(x)
}
\arguments{
\item{x}{list of results from plot_topn().}
}
\description{
Print a result from plot_topn().
}
