% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.varpart}
\alias{print.varpart}
\title{Print variance partition results.}
\usage{
\method{print}{varpart}(x, ...)
}
\arguments{
\item{x}{List of results from variancePartition including the model
information, percent/partition plots, dataframes of the
fitted/sorted data by variance, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print variance partition results.
}
