% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.written_expt}
\alias{print.written_expt}
\title{Print the result from write_expt.}
\usage{
\method{print}{written_expt}(x)
}
\arguments{
\item{x}{List containing all the many plots, the dataframes, etc.}
}
\description{
Print the result from write_expt.
}
