% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{print.written_se}
\alias{print.written_se}
\title{Print the result from write_se.}
\usage{
\method{print}{written_se}(x, ...)
}
\arguments{
\item{x}{List containing all the many plots, the dataframes, etc.}

\item{...}{Other args to match the generic.}
}
\description{
Print the result from write_se.
}
