% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.written_xlsx}
\alias{print.written_xlsx}
\title{Print the result from write_xlsx.}
\usage{
\method{print}{written_xlsx}(x)
}
\arguments{
\item{x}{List containing some information about the xlsx file.}
}
\description{
Print the result from write_xlsx.
}
