% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{print_ups_downs}
\alias{print_ups_downs}
\title{Reprint the output from extract_significant_genes().}
\usage{
print_ups_downs(
  upsdowns,
  wb,
  excel_basename,
  according = "limma",
  summary_count = 1,
  ma = FALSE,
  fancy = FALSE
)
}
\arguments{
\item{upsdowns}{Output from extract_significant_genes().}

\item{wb}{Workbook object to use for writing, or start a new one.}

\item{excel_basename}{Used when including plots in the xlsx sheet.}

\item{according}{Use limma, deseq, or edger for defining 'significant'.}

\item{summary_count}{For spacing sequential tables one after another.}

\item{ma}{Include ma plots?}

\item{fancy}{Print fancy plots with the xlsx file?x}
}
\value{
Return from write_xlsx.
}
\description{
I found myself needing to reprint these excel sheets because I
added some new information. This shortcuts that process for me.
}
\seealso{
[combine_de_tables]
}
