% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{proper_plots}
\alias{proper_plots}
\title{Given the result of simple_proper, plot the various results.}
\usage{
proper_plots(powers, result, add_coverage = TRUE, x_intercept = 1)
}
\arguments{
\item{powers}{The result from comparePowers()}

\item{add_coverage}{Add a line of the actual coverage in an experiment.}

\item{x_intercept}{Add a line showing the significance deemed interesting.}
}
\description{
Given the result of simple_proper, plot the various results.
}
