% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{proper_run_deseq2}
\alias{proper_run_deseq2}
\title{Stealing the hiddent function run.deseq2 from PROPER.}
\usage{
proper_run_deseq2(dat)
}
\arguments{
\item{dat}{Input data.}
}
\description{
Stealing the hiddent function run.deseq2 from PROPER.
}
