% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{proper_run_dss}
\alias{proper_run_dss}
\title{DSS is no longer maintained, just call edgeR.}
\usage{
proper_run_dss(dat)
}
\arguments{
\item{dat}{Input data.}
}
\description{
DSS is no longer maintained, just call edgeR.
}
