% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{proper_run_edger}
\alias{proper_run_edger}
\title{Also stealing a copy of run.edgeR}
\usage{
proper_run_edger(dat)
}
\arguments{
\item{dat}{Input data}
}
\description{
Also stealing a copy of run.edgeR
}
