% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{proportions_by_factors}
\alias{proportions_by_factors}
\title{Extract the proportions of each group/sample in a scd.}
\usage{
proportions_by_factors(
  scd,
  group_factor = "res0p1_clusters",
  sample_factor = "gexcells"
)
}
\arguments{
\item{scd}{Seurat single cell dataset.}

\item{group_factor}{Set of groups to examine.}

\item{sample_factor}{Column defining the samples.}
}
\description{
Extract the proportions of each group/sample in a scd.
}
