% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{prune_network}
\alias{prune_network}
\title{Exclude nodes from a network which are not well connected.}
\usage{
prune_network(network, min_weight = 0.4, min_connectivity = 1)
}
\arguments{
\item{network}{input network to prune.}

\item{min_weight}{Minimum acceptable weight.}

\item{min_connectivity}{Minimum number of nodes to which to be connected.}
}
\value{
A hopefully smaller, but not too small network.
}
\description{
Exclude nodes from a network which are not well connected.
}
