% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{pull_git_commit}
\alias{pull_git_commit}
\title{Reset the chosen git repository to a chosen commit.}
\usage{
pull_git_commit(gitdir = "~/hpgltools", commit = NULL)
}
\arguments{
\item{gitdir}{Desired repository, defaulting to my hpgltools copy.}

\item{commit}{commit ID to which to reset.}
}
\description{
Reset the chosen git repository to a chosen commit.
}
