% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{pval_plot}
\alias{pval_plot}
\title{Make a pvalue plot from a df of IDs, scores, and p-values}
\usage{
pval_plot(df, ontology = "MF")
}
\arguments{
\item{df}{some data from topgo/goseq/clusterprofiler}
}
\value{
a plot!
}
\description{
Make a pvalue plot from a df of IDs, scores, and p-values
}
\seealso{
\code{\link{goseq}}
}

