% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{random_ontology}
\alias{random_ontology}
\title{Perform a simple_ontology() on some random data.}
\usage{
random_ontology(input, method = "goseq", n = 200, ...)
}
\arguments{
\item{input}{Some input data}

\item{method}{goseq, clusterp, topgo, gostats, gprofiler.}

\item{n}{how many 'genes' to analyse?}

\item{...}{Arguments passed to the method.}
}
\value{
An ontology result
}
\description{
At the very least, the result should be less significant than the actual data!
}
\seealso{
[simple_goseq()] [simple_clusterprofiler()] [simple_topgo()] [simple_gostats()]
}
