% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071_de_plots.r
\name{rank_order_scatter}
\alias{rank_order_scatter}
\title{Plot the rank order of the data in two tables against each other.}
\usage{
rank_order_scatter(
  first,
  second = NULL,
  first_type = "limma",
  second_type = "limma",
  first_table = 1,
  alpha = 0.5,
  second_table = 2,
  first_column = "logFC",
  second_column = "logFC",
  first_p_col = "adj.P.Val",
  second_p_col = "adj.P.Val",
  p_limit = 0.05,
  both_color = "red",
  first_color = "green",
  second_color = "blue",
  no_color = "black"
)
}
\arguments{
\item{first}{First table of values.}

\item{second}{Second table of values, if null it will use the first.}

\item{first_type}{Assuming this is from all_pairwise(), use this method.}

\item{second_type}{Ibid.}

\item{first_table}{Again, assuming all_pairwise(), use this to choose the
table to extract.}

\item{alpha}{How see-through to make the dots?}

\item{second_table}{Ibid.}

\item{first_column}{What column to use to rank-order from the first table?}

\item{second_column}{What column to use to rank-order from the second table?}

\item{first_p_col}{Use this column for pretty colors from the first table.}

\item{second_p_col}{Use this column for pretty colors from the second table.}

\item{p_limit}{A p-value limit for coloring dots.}

\item{both_color}{If both columns are 'significant', use this color.}

\item{first_color}{If only the first column is 'significant', this color.}

\item{second_color}{If the second column is 'significant', this color.}

\item{no_color}{If neither column is 'significant', then this color.}
}
\value{
a list with a plot and a couple summary statistics.
}
\description{
Steve Christensen has some neat plots showing the relationship between two
tables.  I though they were super-cool, so I co-opted the idea in this
function.
}
