% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{rbind_summary_rows}
\alias{rbind_summary_rows}
\title{Append rows containing summary() information.}
\usage{
rbind_summary_rows(df)
}
\arguments{
\item{df}{Starter df}
}
\value{
the original df with a couple of new rows at the bottom.
}
\description{
Append rows containing summary() information.
}
