% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{read_counts_expt}
\alias{read_counts_expt}
\title{Read a bunch of count tables and create a usable data frame from them.}
\usage{
read_counts_expt(
  ids,
  files,
  header = FALSE,
  include_summary_rows = FALSE,
  all.x = TRUE,
  all.y = FALSE,
  merge_type = "merge",
  suffix = NULL,
  countdir = NULL,
  ...
)
}
\arguments{
\item{ids}{List of experimental ids.}

\item{files}{List of files to read.}

\item{header}{Whether or not the count tables include a header row.}

\item{include_summary_rows}{Whether HTSeq summary rows should be included.}

\item{all.x}{When merging (as opposed to join), choose the x data column.}

\item{all.y}{When merging (as opposed to join), choose the y data column.}

\item{merge_type}{Choose one, merge or join.}

\item{suffix}{Optional suffix to add to the filenames when reading them.}

\item{countdir}{Optional count directory to read from.}

\item{...}{More options for happy time!}
}
\value{
Data frame of count tables.
}
\description{
It is worth noting that this function has some logic intended for the elsayed
lab's data storage structure. It shouldn't interfere with other usages, but
it attempts to take into account different ways the data might be stored.
}
\details{
Used primarily in create_expt()
This is responsible for reading count tables given a list of filenames.  It
tries to take into account upper/lowercase filenames and uses data.table to
speed things along.
}
\examples{
\dontrun{
 count_tables <- hpgl_read_files(as.character(sample_ids), as.character(count_filenames))
}
}
\seealso{
[data.table] [create_expt()] [tximport]
}
