% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{read_snp_columns}
\alias{read_snp_columns}
\title{Read the output from bcfutils into a count-table-esque}
\usage{
read_snp_columns(samples, input_dir = "preprocessing/outputs",
  file_suffix = "_parsed_ratio.txt")
}
\arguments{
\item{samples}{Sample names to read}

\item{input_dir}{Directory from which to read them.}

\item{file_suffix}{The suffix of my output files.}
}
\value{
A big honking data table.
}
\description{
I put all my bcfutils output files into one directory, so hunt them down and
read them into a data table.
}
