% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{read_thermo_xlsx}
\alias{read_thermo_xlsx}
\title{Parse the difficult thermo fisher xlsx file.}
\usage{
read_thermo_xlsx(xlsx_file, test_row = NULL)
}
\arguments{
\item{xlsx_file}{The input xlsx file}

\item{test_row}{A single row in the xlsx file to use for testing, as I have
not yet seen two of these accursed files which had the same headers.}
}
\value{
List containing the protein names, group data, protein dataframe,
  and peptide dataframe.
}
\description{
The Thermo(TM) workflow has as its default a fascinatingly horrible excel
output.  This function parses that into a series of data frames.
}
