% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{recolor_points}
\alias{recolor_points}
\title{Quick point-recolorizer given an existing plot, df, list of rownames to
recolor, and a color.}
\usage{
recolor_points(plot, df, ids, color = "red", ...)
}
\arguments{
\item{plot}{Geom_point based plot}

\item{df}{Data frame used to create the plot}

\item{ids}{Set of ids which must be in the rownames of df to recolor}

\item{color}{Chosen color for the new points.}

\item{...}{Extra arguments are passed to arglist.}
}
\value{
prettier plot.
}
\description{
This function should make it easy to color a family of genes in any of the
point plots.
}
