% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{record_seurat_samples}
\alias{record_seurat_samples}
\title{Add into the miscellaneous SCD slot a dataframe with some summary stats.}
\usage{
record_seurat_samples(
  scd,
  type = "num_cells",
  pattern = NULL,
  column_name = NULL,
  column_prefix = NULL,
  verbose = TRUE,
  group = "Idents",
  assay = "RNA"
)
}
\arguments{
\item{scd}{Single Cell Dataset to query.}

\item{type}{Type of column to add to the metadata df, named for the column in the
Cell-annotation table to query.}

\item{pattern}{Pattern used for regex-based queries.}

\item{column_name}{Name for the new column.}

\item{column_prefix}{Prefix added to the new column.}

\item{verbose}{Print the summaries to screen?}
}
\value{
Give back the SCD with some new information.
}
\description{
There are some simple summaries which are nice to have on hand
regarding the number of RNAs, cells, rProteins, rmito observed.
This function collects them and drops them into a dataframe within
the slot 'misc' of the SCD.  I may also print to screen some
pretty skims of the results.
}
