% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{remove_batch_effect}
\alias{remove_batch_effect}
\title{## comments, the first separateis the title, the second the}
\usage{
remove_batch_effect(normalized_counts, model)
}
\description{
## comments, the first separateis the title, the second the
}

