% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.R
\name{remove_ggplot_geom}
\alias{remove_ggplot_geom}
\title{Remove an arbitrary geometry from a ggplot2 object}
\usage{
remove_ggplot_geom(ggplot2_object, geom_type = "GeomTextRepel")
}
\arguments{
\item{ggplot2_object}{Input ggplot2 plot.}

\item{geom_type}{Exclude elements of this geometry type from the
plot.}
}
\description{
Amalie asked about removing labels from a plot, I did not want to
regenerate the plot and so did a search and found the following
function which excludes a geometry from a ggplot2 object.
}
