% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{rename_vennerable_intersections}
\alias{rename_vennerable_intersections}
\title{Give the intersection list produced by vennerable easier-to-read names.}
\usage{
rename_vennerable_intersections(venn, lst)
}
\arguments{
\item{venn}{Vennerable data structure.}

\item{lst}{List of names which are legible.}
}
\description{
Give the intersection list produced by vennerable easier-to-read names.
}
