% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{renderme}
\alias{renderme}
\title{Add a little logic to rmarkdown::render to date the final outputs as per a
request from Najib.}
\usage{
renderme(file, format = "html_document", overwrite = TRUE)
}
\arguments{
\item{file}{Rmd file to render.}

\item{format}{Chosen file format.}

\item{overwrite}{Overwrite an existing file?}
}
\value{
Final filename including the prefix rundate.
}
\description{
Add a little logic to rmarkdown::render to date the final outputs as per a
request from Najib.
}
\seealso{
[rmarkdown]
}
