% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/135_model_varpartition.r
\name{replot_varpart_percent}
\alias{replot_varpart_percent}
\title{A shortcut for replotting the percent plots from variancePartition.}
\usage{
replot_varpart_percent(
  varpart_output,
  n = 30,
  column = NULL,
  decreasing = TRUE
)
}
\arguments{
\item{varpart_output}{List returned by varpart()}

\item{n}{How many genes to plot.}

\item{column}{The df column to use for sorting.}

\item{decreasing}{high->low or vice versa?}
}
\value{
The percent variance bar plots from variancePartition!
}
\description{
In case I wish to look at different numbers of genes from variancePartition
and/or different columns to sort from.
}
\seealso{
\pkg{variancePartition}
 \code{\link[variancePartition]{plotPercentBars}}
}
