% Generated by roxygen2 (4.0.1): do not edit by hand
\name{require.auto}
\alias{require.auto}
\title{Automatic loading and/or installing of packages.}
\usage{
require.auto(lib)
}
\arguments{
\item{lib}{string name of a library}
}
\value{
NULL currently
}
\description{
Automatic loading and/or installing of packages.
}
\examples{
## require.auto("ggplot2")
}
\seealso{
\code{\link{biocLite}} and \code{\link{install.packages}}
}

