% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{rowData,ExpressionSet-method}
\alias{rowData,ExpressionSet-method}
\title{A getter of the gene information from an ExpressionSet, synonymous with fData().}
\usage{
\S4method{rowData}{ExpressionSet}(x, withDimnames = TRUE, ...)
}
\description{
A getter of the gene information from an ExpressionSet, synonymous with fData().
}
