% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{rowData,expt-method}
\alias{rowData,expt-method}
\title{A getter of the gene information from an expt, synonymous with fData().}
\usage{
\S4method{rowData}{expt}(x, withDimnames = TRUE, ...)
}
\description{
A getter of the gene information from an expt, synonymous with fData().
}
