% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{rowData<-,expt-method}
\alias{rowData<-,expt-method}
\title{A setter to put the gene information into an expt.}
\usage{
\S4method{rowData}{expt}(x, ...) <- value
}
\arguments{
\item{x}{Input}

\item{...}{them too!}

\item{value}{New annotations to put into the input}
}
\description{
A setter to put the gene information into an expt.
}
