% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{rowData}
\alias{rowData}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
rowData(x, use.names = TRUE, ...)
}
\arguments{
\item{x}{Input object}

\item{use.names}{new value}

\item{...}{extra args}
}
\description{
message("I am from SummarzedExperiment and am explicitly imported, wtf.")
}
