% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.R
\name{rowmax_filter_counts}
\alias{rowmax_filter_counts}
\title{Filter low-count genes from a data set only using a simple maximum-count threshold.}
\usage{
rowmax_filter_counts(count_table, threshold = 2)
}
\arguments{
\item{count_table}{Data frame of (pseudo)counts by sample.}

\item{threshold}{Lower threshold of counts for each gene.}
}
\value{
Dataframe of counts without the low-count genes.
}
\description{
Filter low-count genes from a data set only using a simple maximum-count threshold.
}
\examples{
\dontrun{
 filtered_table <- rowmax_filter_counts(count_table)
}
}
\seealso{
[edgeR]
}
