% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{run.DESeq2}
\alias{run.DESeq2}
\title{PROPER uses this function privately, but should not.}
\usage{
run.DESeq2(dat)
}
\arguments{
\item{dat}{Input data}
}
\description{
PROPER uses this function privately, but should not.
}
