% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{run.DSS}
\alias{run.DSS}
\title{PROPER uses this privately but really make it available I think.}
\usage{
run.DSS(dat)
}
\description{
PROPER uses this privately but really make it available I think.
}
