% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150_proteomics.r
\name{s2s_all_filters}
\alias{s2s_all_filters}
\title{Gather together the various SWATH2stats filters into one place.}
\usage{
s2s_all_filters(
  s2s_exp,
  column = "proteinname",
  pep_column = "fullpeptidename",
  fft = 0.7,
  plot = FALSE,
  target_fdr = 0.02,
  upper_fdr = 0.05,
  mscore = 0.01,
  percentage = 0.75,
  remove_decoys = TRUE,
  max_peptides = 15,
  min_peptides = 2,
  do_mscore = TRUE,
  do_freqobs = TRUE,
  do_fdr = TRUE,
  do_proteotypic = TRUE,
  do_peptide = TRUE,
  do_max = TRUE,
  do_min = TRUE,
  ...
)
}
\arguments{
\item{s2s_exp}{SWHAT2stats result from the sample_annotation()
function. (s2s_exp stands for: SWATH2stats experiment)}

\item{column}{What column in the data contains the protein name?}

\item{pep_column}{What column in the data contains the peptide name (not
currently used, but it should be.)}

\item{fft}{Ratio of false negatives to true positives, used by
assess_by_fdr() and similar functions.}

\item{plot}{Print plots of the various rates by sample?}

\item{target_fdr}{When invoking mscore4assayfdr, choose an mscore which
corresponds to this false discovery date.}

\item{upper_fdr}{Used by filter_mscore_fdr() to choose the minimum threshold
of identification confidence.}

\item{mscore}{Mscore cutoff for the mscore filter.}

\item{percentage}{Cutoff for the mscore_freqobs filter.}

\item{remove_decoys}{Get rid of decoys in the final filter, if they were not
already removed.}

\item{max_peptides}{A maximum number of peptides filter.}

\item{min_peptides}{A minimum number of peptides filter.}

\item{do_mscore}{Perform the mscore filter? SWATH2stats::filter_mscore()}

\item{do_freqobs}{Perform the mscore_freqobs filter?
SWATH2stats::filter_mscore_freqobs()}

\item{do_fdr}{Perform the fdr filter? SWATH2stats::filter_mscore_fdr()}

\item{do_proteotypic}{Perform the proteotypic filter?
SWATH2stats::filter_proteotypic_peptides()}

\item{do_peptide}{Perform the single-peptide filter?
SWATH2stats::filter_all_peptides()}

\item{do_max}{Perform the maximum peptide filter?
SWATH2stats::filter_max_peptides()}

\item{do_min}{Perform the minimum peptide filter?
SWATH2stats::filter_min_peptides()}

\item{...}{Other arguments passed down to the filters.}
}
\value{
Smaller SWATH2stats data set.
}
\description{
There are quite a few filters available in SWATH2stats.  Reading the
documentation, it seems at least possible, if not appropriate, to use them
together when filtering DIA data before passing it to MSstats/etc.  This
function attempts to formalize and simplify that process.
}
