% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{s_p}
\alias{s_p}
\title{Silence, peasant!}
\usage{
s_p(code)
}
\arguments{
\item{code}{Some code to shut up.}
}
\value{
List of the output log, message log, and result of the expression.
}
\description{
Some libraries/functions just won't shut up.  Ergo, silence, peasant!
This function uses 2 invocations of capture.output and a try(silent=TRUE) to capture the strings
of the outputs from the given expression in 'output', and the messages in 'message'.  The result
of the expression goes into 'result.'  If there is an error in the expression, it is returned as
a try-error object which may therefore be inspected as needed.
}

