% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{sampleNames,expt-method}
\alias{sampleNames,expt-method}
\title{A getter to get the samples names from an expt.}
\usage{
\S4method{sampleNames}{expt}(object)
}
\arguments{
\item{object}{Input}
}
\description{
A getter to get the samples names from an expt.
}
