% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{sampleNames<-,expt,ANY-method}
\alias{sampleNames<-,expt,ANY-method}
\title{A setter to put the samples names into an expt.}
\usage{
\S4method{sampleNames}{expt,ANY}(object) <- value
}
\description{
A setter to put the samples names into an expt.
}
