% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{sampleNames<-,SummarizedExperiment,ANY-method}
\alias{sampleNames<-,SummarizedExperiment,ANY-method}
\title{A setter to put the samples names into a SummarizedExperiment.}
\usage{
\S4method{sampleNames}{SummarizedExperiment,ANY}(object) <- value
}
\description{
A setter to put the samples names into a SummarizedExperiment.
}
