% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\docType{methods}
\name{sampleNames<-}
\alias{sampleNames<-}
\alias{sampleNames<-,}
\alias{sampleNames<--methods}
\title{Extend Biobase::sampleNames<- to handle expt objects.}
\usage{
\S4method{sampleNames}{expt,ANY}(object) <- value
}
\arguments{
\item{object}{The expt object from which to extract the expressionset.}
}
\description{
Extend Biobase::sampleNames<- to handle expt objects.
}
