% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{samtools_snp_coverage}
\alias{samtools_snp_coverage}
\title{Use Rsamtools to read alignments and get snp coverage.}
\usage{
samtools_snp_coverage(expt, type = "counts",
  input_dir = "preprocessing/outputs", tolower = TRUE,
  bam_suffix = ".bam")
}
\arguments{
\item{expt}{Expressionset to analyze}

\item{type}{counts or percent?}

\item{input_dir}{Directory containing the samtools results.}

\item{tolower}{lowercase the sample names?}

\item{bam_suffix}{In case the data came from sam.}
}
\value{
It is so slow I no longer know if it works.
}
\description{
This is horrifyingly slow.
}
