% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_annotations,SummarizedExperiment-method}
\alias{sanitize_annotations,SummarizedExperiment-method}
\title{Given an expressionset, sanitize the gene information data.}
\usage{
\S4method{sanitize_annotations}{SummarizedExperiment}(
  input,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL,
  numeric = FALSE
)
}
\arguments{
\item{input}{Input expressionset.}

\item{columns}{Set of columns to sanitize, otherwise all of them.}

\item{na_value}{Fill in NA with this.}

\item{lower}{sanitize capitalization.}

\item{punct}{Remove punctuation?}

\item{factorize}{Convert columns to factors?  When set to 'heuristic'
this tries out as.factor and sees if the number of levels is silly.}

\item{max_levels}{The definition of 'silly' above.}

\item{spaces}{Allow spaces in the data?}

\item{numbers}{Sanitize number formats (e.g. 1.000.000,0 vs. 1,000,000.0)}

\item{numeric}{Set columns to numeric when possible?}
}
\description{
Given an expressionset, sanitize the gene information data.
}
