% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sanitize_annotations}
\alias{sanitize_annotations}
\title{Synonym for sanitize_expt_fData()}
\usage{
sanitize_annotations(...)
}
\arguments{
\item{...}{Passed along}
}
\description{
Synonym for sanitize_expt_fData()
}
