% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{sanitize_expt}
\alias{sanitize_expt}
\title{Get rid of characters which will mess up contrast making and such before
playing with an expt.}
\usage{
sanitize_expt(expt, keep_underscore = TRUE, factors = c("condition", "batch"))
}
\arguments{
\item{expt}{An expt object to clean.}

\item{keep_underscore}{Sanitize underscores too?}

\item{factors}{Specific factors to check.}
}
\description{
Get rid of characters which will mess up contrast making and such before
playing with an expt.
}
