% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sanitize_expt_pData}
\alias{sanitize_expt_pData}
\title{Adding an alias to sanitize_metadata until I decide how I want to name this.}
\usage{
sanitize_expt_pData(
  expt,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL,
  numeric = FALSE
)
}
\arguments{
\item{expt}{Input expressionset.}

\item{columns}{Set of columns to sanitize, otherwise all of them.}

\item{na_value}{Fill in NA with this.}

\item{lower}{sanitize capitalization.}

\item{punct}{Remove punctuation?}

\item{factorize}{Convert columns to factors?  When set to 'heuristic'
this tries out as.factor and sees if the number of levels is silly.}

\item{max_levels}{The definition of 'silly' above.}

\item{spaces}{Allow spaces in the data?}

\item{numbers}{Sanitize number formats (e.g. 1.000.000,0 vs. 1,000,000.0)}

\item{numeric}{Set columns to numeric when possible?}
}
\description{
Adding an alias to sanitize_metadata until I decide how I want to name this.
}
