% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_metadata,DFrame-method}
\alias{sanitize_metadata,DFrame-method}
\title{Sanitize a DFrame}
\usage{
\S4method{sanitize_metadata}{DFrame}(
  input,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL,
  numeric = FALSE
)
}
\arguments{
\item{input}{Input datastructure}
}
\description{
Sanitize a DFrame
}
