% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{sanitize_metadata,expt-method}
\alias{sanitize_metadata,expt-method}
\title{Metadata sanitizers for an expt}
\usage{
\S4method{sanitize_metadata}{expt}(
  meta,
  columns = NULL,
  na_string = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL
)
}
\description{
Metadata sanitizers for an expt
}
