% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{sanitize_model}
\alias{sanitize_model}
\title{Ensure an experimental model is safe to use}
\usage{
sanitize_model(
  model,
  keep_underscore = FALSE,
  exclude_strings = c("condition", "batch")
)
}
\arguments{
\item{model}{input model}

\item{keep_underscore}{I previously dropped all punctuation including underscores.}

\item{exclude_strings}{simplify the strings for these factors a little.}
}
\description{
Ensure an experimental model is safe to use
}
