% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{sanitize_number_encoding}
\alias{sanitize_number_encoding}
\title{Ensure that we handle numbers encoded as '4.012.321,10' are
properly (from the perspective of R using my encoding system)
interpreted as 'four million twelve thousand three hundred
twenty-one and one tenth.'}
\usage{
sanitize_number_encoding(numbers, df = NULL)
}
\arguments{
\item{numbers}{Column of numbers.}

\item{optional}{df}
}
\description{
Ensure that we handle numbers encoded as '4.012.321,10' are
properly (from the perspective of R using my encoding system)
interpreted as 'four million twelve thousand three hundred
twenty-one and one tenth.'
}
