% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{sanitize_percent}
\alias{sanitize_percent}
\title{Sanitize unreliable presentation of percent values from excel.}
\usage{
sanitize_percent(numbers, df = NULL)
}
\arguments{
\item{numbers}{Either a vector of excel crap, or a column
name/number.}

\item{df}{When provided, a data frame from which to extract the
numbers.}
}
\value{
Either the numbers or dataframe with the sanitized information.
}
\description{
In a recent sample sheet, we had some percentage values which were
'0.5', '5%', ' 6%' and a few other weirdo things. This function
should sanitize such shenanigans.
}
