% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{sanitize_se}
\alias{sanitize_se}
\title{Get rid of characters which will mess up contrast making and such before
playing with an se.}
\usage{
sanitize_se(se, keep_underscore = TRUE, factors = c("condition", "batch"))
}
\arguments{
\item{se}{An se object to clean.}

\item{keep_underscore}{Sanitize underscores too?}

\item{factors}{Specific factors to check.}
}
\description{
Get rid of characters which will mess up contrast making and such before
playing with an se.
}
