% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{saveme}
\alias{saveme}
\title{\code{saveme()}  Make a backup rdata file for future reference}
\usage{
saveme(directory = "savefiles", backups = 4)
}
\arguments{
\item{backups}{default=10  how many revisions?

I often use R over a sshfs connection, sometimes with significant latency, and
I want to be able to save/load my R sessions relatively quickly.
Thus this function uses pxz to compress the R session maximally and relatively fast.
This assumes you have pxz installed and >= 4 CPUs.}

\item{dir}{the directory to save the Rdata file.}
}
\value{
the command used to save the global environment
}
\description{
\code{saveme()}  Make a backup rdata file for future reference
}
\examples{
\dontrun{
saveme()
}
}
\seealso{
\link{save} \link{pipe}
}

