% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{sb_annot}
\alias{sb_annot}
\title{Portion of the raw trinotate 'annotation' output for Solanum betaceum.}
\format{
An object of class \code{character} of length 1.
}
\usage{
sb_annot
}
\description{
The full S.betaceum trinotate annotation is quite large, so I just pulled a
portion as an example for this package.
}
\keyword{datasets}
