% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{sb_data}
\alias{sb_data}
\title{Portion of the RNASeq results from Solanum betaceum.}
\format{
An object of class \code{character} of length 1.
}
\usage{
sb_data
}
\description{
I had the opportunity to work the Sandra Correia, she was awesome.  She was
seeking to learn about differences among embryogenic cells in the Tree
tomato.  I therefore got to learn first-hand a tiny portion of what is meant
when one says 'plant genetics are hard.'  I had it far easier than Sandra.  I
just used Trinity to make some de-novo transcriptomes and attempt to provide
some metrics about which ones are real and really different across conditions
in her experiment.  Her work was many thousands of times more difficult.
}
\keyword{datasets}
