% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptides.R
\name{score_amino_acids}
\alias{score_amino_acids}
\title{Given a set of nucleotides/amino acids, use the Peptides
package to create a matrix of scores describing them.}
\usage{
score_amino_acids(file, sanitize_names = TRUE)
}
\arguments{
\item{file}{Input fasta file of peptide sequences.}

\item{sanitize_names}{Sanitize the fasta file IDs?}
}
\description{
For the moment, I am going to just use an amino acid fasta file.
I think I would like to be able to take an expressionset, which
would obviously only work for non-spliced organisms -- unless the
annotations include the introns...
}
