% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnseq.r
\name{score_mhess}
\alias{score_mhess}
\title{A scoring function for the mh_ess TNSeq method.}
\usage{
score_mhess(expt, ess_column = "essm1")
}
\arguments{
\item{expt}{Input expressionset with a metadata column with the ess output files.}

\item{ess_column}{Metadata column containing the mh_ess output files.}
}
\value{
List containing the scores along with the genes which have changed using it.
}
\description{
I dunno, I might delete this function, I am not sure if it will ever get use.
}
