% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{score_to_meaning}
\alias{score_to_meaning}
\title{Given a bam file, print out the score meanings}
\usage{
score_to_meaning(score)
}
\arguments{
\item{score}{Observed score in a bam alignment.}
}
\value{
vector of the meanings associated with the bitwise score.
}
\description{
I suspect there other implementations of this elsewhere, but I
wanted to write one for myself in the hope that I would remember
more of the score meanings.
}
