% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{self_evaluate_model}
\alias{self_evaluate_model}
\title{Create a confusion matrix and ROC of a model against its training data. (and test data
if the annotations are known)}
\usage{
self_evaluate_model(predictions, datasets, which_partition = 1, type = "train")
}
\arguments{
\item{predictions}{Model created by train()}

\item{datasets}{Set of training/testing partitions along with
associated metadata annotations.}

\item{which_partition}{Choose a paritiont to evaluate}

\item{type}{Use the training or testing data?}
}
\description{
This assumes a set of partitions from create_partitions() which
keeps the training metadata alongside the matrix of model
variables.  When available, that function also keeps the known
annotations of the testing data.  Given those annotations and the
model created/tested from them, this runs confusionMatrix and ROC,
collects the results, and provides them as a list.
}
