% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{semantic_copynumber_extract}
\alias{semantic_copynumber_extract}
\title{Extract multicopy genes from up/down gene expression lists.}
\usage{
semantic_copynumber_extract(...)
}
\arguments{
\item{...}{Arguments for semantic_copynumber_filter()}
}
\description{
The function semantic_copynumber_filter() is the inverse of this.
}
\details{
Currently untested, used for Trypanosome analyses primarily, thus the default
strings.
}
