% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{semantic_filter,expt-method}
\alias{semantic_filter,expt-method}
\title{Filter an expt}
\usage{
\S4method{semantic_filter}{expt}(
  input,
  invert = FALSE,
  topn = NULL,
  semantic = c("mucin", "sialidase", "RHS", "MASP", "DGF", "GP63"),
  semantic_column = "description"
)
}
\arguments{
\item{input}{Input expt}

\item{invert}{Flip the logic, e.g. keep these elements.}

\item{topn}{Keep the top-n genes}

\item{semantic}{Strings to search to define genes to remove.}

\item{semantic_column}{Annotation column containing text to search.}
}
\description{
Filter an expt
}
