% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{sequence_attributes}
\alias{sequence_attributes}
\title{Gather some simple sequence attributes}
\usage{
sequence_attributes(fasta, gff = NULL, type = "gene", key = "locus_tag")
}
\arguments{
\item{fasta}{a fasta genome}

\item{gff}{an optional gff of annotations (if not provided it will just ask the whole genome.}

\item{type}{the column to get frmo the gff file}

\item{key}{what type of entry of the gff file to key from?}

\item{pattern}{what pattern to search for?  This was used for tnseq and TA is the mariner insertion point.}
}
\value{
num_pattern a data frame of names and numbers.
}
\description{
Gather some simple sequence attributes
}
\examples{
\dontrun{
num_pattern = sequence_attributes('mgas_5005.fasta', 'mgas_5005.gff')
}
}
\seealso{
\pkg{Biostrings} \pkg{Rsamtools} \link[Biostrings]{letterFrequency} \link[Rsamtools]{FaFile}
}

