% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{sequential_variants}
\alias{sequential_variants}
\title{Search a set of variants for ones which are relatively sequential.}
\usage{
sequential_variants(
  snp_sets,
  conditions = NULL,
  minimum = 3,
  maximum_separation = 3,
  one_away_file = "one_away.csv",
  two_away_file = "two_away.csv",
  doubles_files = "doubles.csv",
  singles_file = "singles.csv"
)
}
\arguments{
\item{snp_sets}{Result from get_snp_sets() containing the variants
with respect to known conditions.}

\item{conditions}{Set of conditions to search against.}

\item{minimum}{Minimum number of variants required for a candiate.}

\item{maximum_separation}{How far apart from each other are these
>={minimum} variants allowed to be?}
}
\description{
One potential way to screen strains is to use PCR primers which
should(not) anneal due to variants with respect to the genome.
This function seeks to find variants which are clustered
sufficiently close to each other that this is possible.
}
