% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_annotation,SummarizedExperiment-method}
\alias{set_annotation,SummarizedExperiment-method}
\title{Get the annotation pkg from a SummarizedExperiment}
\usage{
\S4method{set_annotation}{SummarizedExperiment}(exp, annot, ...)
}
\arguments{
\item{exp}{Input SE}

\item{annot}{Annotation package name}

\item{...}{The rest of the arguments to set_annotation()}
}
\description{
Get the annotation pkg from a SummarizedExperiment
}
