% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_assay,SummarizedExperiment-method}
\alias{set_assay,SummarizedExperiment-method}
\title{Set the asay of a Summarized Experiment}
\usage{
\S4method{set_assay}{SummarizedExperiment}(exp, values)
}
\arguments{
\item{exp}{Input SummarizedExperiment}

\item{values}{New assay values.}
}
\description{
This is a shortcut so I can easily pass assay along in a pipeline.
}
