% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_assay}
\alias{set_assay}
\title{A generic assay setter with a reminder if I pass it something unknown.}
\usage{
set_assay(exp, value, ...)
}
\arguments{
\item{exp}{Input data structure.}

\item{value}{New value}

\item{...}{Extra arguments.}
}
\description{
A generic assay setter with a reminder if I pass it something unknown.
}
