% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_batches,SummarizedExperiment-method}
\alias{set_batches,SummarizedExperiment-method}
\title{Get the batches from a SummarizedExperiment}
\usage{
\S4method{set_batches}{SummarizedExperiment}(exp, ...)
}
\arguments{
\item{exp}{Input SE}

\item{...}{The rest of the arguments to set_se_batches()}
}
\description{
Get the batches from a SummarizedExperiment
}
