% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_batches}
\alias{set_batches}
\title{Create a generic set_batches function in case I pass something unknown.}
\usage{
set_batches(exp, ...)
}
\arguments{
\item{exp}{Input data structure.}

\item{...}{Arguments passed along.}
}
\description{
I like to think in terms of conditions and batches.  This function
is intended to set the latter for a dataset.
}
