% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_colors,data.frame,NULL-method}
\alias{set_colors,data.frame,NULL-method}
\title{Set the colors of a dataframe with no provided colors.}
\usage{
\S4method{set_colors}{data.frame,NULL}(
  exp,
  colors = NULL,
  color_column = "color",
  change_by = "condition",
  chosen_palette = "Dark2",
  ...
)
}
\arguments{
\item{exp}{Datastructure to modify.}

\item{colors}{Vector/List/etc to explicitly define desired colors.}

\item{color_column}{metadta column containing colors.}

\item{change_by}{Metadata column to use to define colors.}

\item{chosen_palette}{ColorBrewer palette used when choosing colors.}

\item{...}{Args passed to the various setters.}
}
\value{
Modified data structure.
}
\description{
Set the colors of a dataframe with no provided colors.
}
