% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_colors,SummarizedExperiment-method}
\alias{set_colors,SummarizedExperiment-method}
\title{Set colors for a SE.}
\usage{
\S4method{set_colors}{SummarizedExperiment}(exp, ...)
}
\arguments{
\item{exp}{Input se.}

\item{...}{Arguments passed along to set_se_colors().}
}
\description{
Set colors for a SE.
}
