% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_conditions,expt-method}
\alias{set_conditions,expt-method}
\title{This method should go away soon.}
\usage{
\S4method{set_conditions}{expt}(
  exp,
  fact = NULL,
  ids = NULL,
  prefix = NULL,
  null_cell = "null",
  colors = TRUE,
  ...
)
}
\description{
This method should go away soon.
}
