% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_conditions,SummarizedExperiment-method}
\alias{set_conditions,SummarizedExperiment-method}
\title{Set conditions to a se.}
\usage{
\S4method{set_conditions}{SummarizedExperiment}(
  exp,
  fact = NULL,
  ids = NULL,
  prefix = NULL,
  handle_na = "unknown",
  null_cell = "null",
  colors = TRUE,
  ...
)
}
\arguments{
\item{exp}{Input se}

\item{fact}{Factor of conditions}

\item{ids}{Set of ids to change.}

\item{prefix}{Prefix of each sample name}

\item{null_cell}{If a cell is null, what to change it to?}

\item{colors}{Set the colors as well?}

\item{...}{Arbitrary arguments.}
}
\description{
Set conditions to a se.
}
