% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_conditions}
\alias{set_conditions}
\title{Generic prefix condition setter.}
\usage{
set_conditions(exp, ...)
}
\arguments{
\item{exp}{Input datastructure.}

\item{...}{Arguments passed along to methods.}
}
\description{
Generic prefix condition setter.
}
