% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_batch}
\alias{set_expt_batch}
\title{Change the batches of an expt}
\usage{
set_expt_batch(expt, fact, ids = NULL, ...)
}
\arguments{
\item{expt}{Expt to modify}

\item{factor}{Batches to replace}
}
\value{
expt Send back the expt with some new metadata
#' \dontrun{
 expt = set_expt_batch(big_expt, factor=c(some,stuff,here))")
}
}
\description{
When exploring differential analyses, it might be useful to play with the conditions/batches of
the experiment.  Use this to make that easier.
}

