% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_expt_batches,SummarizedExperiment-method}
\alias{set_expt_batches,SummarizedExperiment-method}
\title{set the batches of a SE}
\usage{
\S4method{set_expt_batches}{SummarizedExperiment}(expt, fact, ids = NULL, ...)
}
\arguments{
\item{expt}{Input summarized experiment.}

\item{fact}{Factor to use.}

\item{ids}{Or specific IDs}

\item{...}{Other arguments.}
}
\description{
set the batches of a SE
}
