% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_colors}
\alias{set_expt_colors}
\title{Change the colors of an expt}
\usage{
set_expt_colors(expt, colors = TRUE, chosen_palette = "Dark2")
}
\arguments{
\item{expt}{Expt to modify}

\item{colors}{colors to replace}

\item{chosen_palette}{I usually use Dark2 as the RColorBrewer palette.}
}
\value{
expt Send back the expt with some new metadata
}
\description{
When exploring differential analyses, it might be useful to play with the conditions/batches of
the experiment.  Use this to make that easier.
}

