% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_condition}
\alias{set_expt_condition}
\title{Change the condition of an expt}
\usage{
set_expt_condition(expt, fact, ids = NULL, ...)
}
\arguments{
\item{expt}{Expt to modify}

\item{ids}{Specific sample IDs to change.}

\item{factor}{Conditions to replace}

\item{colors}{Reset the set of colors (Give a factor if you want to choose your own).}
}
\value{
expt Send back the expt with some new metadata
}
\description{
When exploring differential analyses, it might be useful to play with the conditions/batches of
the experiment.  Use this to make that easier.
}
\examples{
\dontrun{
 expt = set_expt_condition(big_expt, factor=c(some,stuff,here))
}
}

