% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{set_expt_conditions}
\alias{set_expt_conditions}
\title{Change the condition of an expt}
\usage{
set_expt_conditions(
  expt,
  fact = NULL,
  ids = NULL,
  prefix = NULL,
  null_cell = "null",
  colors = TRUE,
  ...
)
}
\arguments{
\item{expt}{Expt to modify}

\item{fact}{Conditions to replace}

\item{ids}{Specific sample IDs to change.}

\item{null_cell}{How to fill elements of the design which are null?}

\item{...}{Extra arguments are given to arglist.}
}
\value{
expt Send back the expt with some new metadata
}
\description{
When exploring differential analyses, it might be useful to play with the
conditions/batches of the experiment.  Use this to make that easier.
}
\examples{
\dontrun{
 expt = set_expt_conditions(big_expt, factor = c(some,stuff,here))
}
}
\seealso{
[set_expt_batches()] [create_expt()]
}
