% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{set_expt_genename_column}
\alias{set_expt_genename_column}
\title{Switch the gene names of an expressionset using a column from fData.}
\usage{
set_expt_genename_column(expt, new_column)
}
\arguments{
\item{expt}{Current expressionSet.}

\item{new_column}{Column from the gene annotations containing the
new gene IDs.}
}
\value{
The expressionset with swapped out IDs.
}
\description{
I am not sure if set_expt_genenames() is smart enough to check for
missing values.  It definitely handles duplicates.
}
