% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_genenames}
\alias{set_expt_genenames}
\title{Change the gene names of an expt.}
\usage{
set_expt_genenames(expt, ids = NULL, ...)
}
\arguments{
\item{expt}{Expt to modify}

\item{ids}{Specific sample IDs to change.}

\item{...}{Extra arguments are given to arglist.}
}
\value{
expt Send back the expt with some new metadata
}
\description{
I want to change all the gene names of a big expressionset to the
ortholog groups.  But I want to also continue using my expts.
Ergo this little function.
}
\examples{
\dontrun{
 expt = set_expt_conditions(big_expt, factor = c(some,stuff,here))
}
}
\seealso{
[set_expt_conditions()] [create_expt()]
}
