% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_samplenames}
\alias{set_expt_samplenames}
\title{Change the sample names of an expt.}
\usage{
set_expt_samplenames(expt, newnames)
}
\arguments{
\item{expt}{Expt to modify}

\item{newnames}{New names, currently only a character vector.}
}
\value{
expt Send back the expt with some new metadata
}
\description{
Sometimes one does not like the hpgl identifiers, so provide a way to change them on-the-fly.
}
\examples{
\dontrun{
 expt = set_expt_samplenames(expt, c("a","b","c","d","e","f"))
}
}
\seealso{
\code{\link{set_expt_condition}} \code{\link{set_expt_batch}}
}
