% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_factors}
\alias{set_factors}
\title{Set arbitrary experimental design factors.}
\usage{
set_factors(
  se,
  condition = NULL,
  batch = NULL,
  ids = NULL,
  table = "metadata",
  class = "factor",
  columns = NULL,
  ...
)
}
\arguments{
\item{se}{Input datastructure}

\item{condition}{condition factor}

\item{batch}{batch factor}

\item{ids}{Set of ids to change}

\item{table}{Optionally change gene annotations}

\item{class}{Set the column(s) to this class}

\item{columns}{columns to change}

\item{...}{Extra args to pass along.}
}
\description{
I do not use this function very often, perhaps just remove it.
}
