% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_genenames}
\alias{set_genenames}
\title{Generic function to change gene names of a datastructure}
\usage{
set_genenames(exp, ...)
}
\arguments{
\item{exp}{Input datastructure}

\item{...}{Extra arguments passed along.}
}
\description{
Generic function to change gene names of a datastructure
}
