% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_se_batches}
\alias{set_se_batches}
\title{Set the batches for a summarized experiment.}
\usage{
set_se_batches(se, fact, ids = NULL)
}
\arguments{
\item{se}{Input se}

\item{fact}{factor of new batches}

\item{ids}{specific IDs to change}
}
\description{
Set the batches for a summarized experiment.
}
