% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_se_colors}
\alias{set_se_colors}
\title{Set the colors of a summarized experiment.}
\usage{
set_se_colors(
  se,
  colors = TRUE,
  chosen_palette = "Dark2",
  change_by = "condition"
)
}
\arguments{
\item{se}{Input se}

\item{colors}{Set of colors to add.}

\item{chosen_palette}{If colors is TRUE, use this palette to set colors.}

\item{change_by}{Use this factor to set the colors.}
}
\description{
Set the colors of a summarized experiment.
}
