% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{set_se_conditions}
\alias{set_se_conditions}
\title{Set conditions to a se.}
\usage{
set_se_conditions(
  se,
  fact = NULL,
  ids = NULL,
  prefix = NULL,
  null_cell = "null",
  colors = TRUE,
  ...
)
}
\arguments{
\item{se}{Input se}

\item{fact}{Factor of conditions}

\item{ids}{Set of ids to change.}

\item{prefix}{Prefix of each sample name}

\item{null_cell}{If a cell is null, what to change it to?}

\item{colors}{Set the colors as well?}

\item{...}{Arbitrary arguments.}
}
\description{
Set conditions to a se.
}
