% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.r
\name{sig_ontologies}
\alias{sig_ontologies}
\title{Take the result from extract_significant_genes() and perform ontology searches.}
\usage{
sig_ontologies(significant_result, excel_prefix = "excel/sig_ontologies",
  search_by = "deseq", excel_suffix = ".xlsx", type = "gprofiler",
  ...)
}
\arguments{
\item{significant_result}{Result from extract_siggenes()}

\item{excel_prefix}{How to start the output filenames?}

\item{search_by}{Use the definition of 'significant' from which program?}

\item{excel_suffix}{How to end the excel filenames?}

\item{type}{Which specific ontology search to use?}

\item{...}{Arguments passed to the various simple_ontology() function.}
}
\value{
A list of the up/down results of the ontology searches.
}
\description{
It can be annoying/confusing to extract individual sets of 'significant' genes from a
differential expression analysis.  This function should make that process easier.
}
