% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{simple_clusterprofiler}
\alias{simple_clusterprofiler}
\title{Perform the array of analyses in the 2016-04 version of clusterProfiler}
\usage{
simple_clusterprofiler(
  sig_genes,
  de_table = NULL,
  orgdb = "org.Hs.eg.db",
  todo = NULL,
  orgdb_from = NULL,
  orgdb_to = "ENTREZID",
  go_level = 3,
  pcutoff = 0.05,
  organism = "human",
  qcutoff = 0.1,
  fc_column = "logFC",
  second_fc_column = "deseq_logfc",
  internal = FALSE,
  updown = "up",
  permutations = 1000,
  min_groupsize = 5,
  max_groupsize = 500,
  kegg_prefix = NULL,
  kegg_organism = NULL,
  categories = 12,
  excel = NULL,
  david_id = "ENTREZ_GENE_ID",
  padj_type = "BH",
  david_user = "abelew@umd.edu",
  mesh_category = "C",
  mesh_dbname = "gendoo",
  msigdb_category = "C2",
  msig_db = NULL,
  kegg_universe = NULL,
  reactome_organism = NULL,
  mesh_db = NULL,
  ah_data = NULL,
  signature_data = NULL,
  signature_df = NULL,
  de_table_namedf = NULL,
  sig_genes_namedf = NULL
)
}
\arguments{
\item{sig_genes}{Dataframe of genes deemed 'significant.'}

\item{de_table}{Dataframe of all genes in the analysis, primarily for GSEA.}

\item{orgdb}{Name of the orgDb used for gathering annotation data.}

\item{orgdb_from}{Name of a key in the orgdb used to cross reference to entrez IDs.}

\item{orgdb_to}{List of keys to grab from the orgdb for cross referencing
ontologies.}

\item{go_level}{How deep into the ontology tree should this dive for over
expressed categories.}

\item{pcutoff}{P-value cutoff for 'significant' analyses.}

\item{organism}{String name of the organism.}

\item{qcutoff}{Q-value cutoff for 'significant' analyses.}

\item{fc_column}{When extracting vectors of all genes, what column should be used?}

\item{second_fc_column}{When extracting vectors of all genes, what column
should be tried the second time around?}

\item{internal}{I dunno}

\item{updown}{Include the less than expected ontologies?}

\item{permutations}{How many permutations for GSEA-ish analyses?}

\item{min_groupsize}{Minimum size of an ontology before it is included.}

\item{max_groupsize}{Ignore groups which are too big.}

\item{kegg_prefix}{Many KEGG ids need a prefix before they will cross reference.}

\item{kegg_organism}{Choose the 3 letter KEGG organism name here.}

\item{categories}{How many categories should be plotted in bar/dot plots?}

\item{excel}{Print the results to an excel file?}

\item{david_id}{Which column to use for cross-referencing to DAVID?}

\item{padj_type}{Use this FDR}

\item{david_user}{Default registered username to use.}

\item{mesh_category}{Use this category for MESH.}

\item{mesh_dbname}{Use this MESH sub-database.}

\item{msigdb_category}{Use this mSigDB sub-database.}

\item{msig_db}{Use this database file for the msigdb data.}

\item{do_gsea}{Perform gsea searches?}

\item{do_david}{Attempt to use the DAVID database for a search?}

\item{do_kegg}{Perform kegg search?}

\item{do_reactome}{what it says on the tin.}

\item{do_dose}{Attempt disease ontology search.}

\item{do_mesh}{Attempt MESH search.}

\item{do_msigdb}{Attempt mSigDB search.}
}
\value{
a list
}
\description{
The new version of clusterProfiler has a bunch of new toys.  However, it is
more stringent in terms of input in that it now explicitly expects to receive
annotation data in terms of a orgdb object.  This is mostly advantageous, but
will probably cause some changes in the other ontology functions in the near
future.  This function is an initial pass at making something similar to my
previous 'simple_clusterprofiler()' but using these new toys.
}
\examples{
\dontrun{
 holyasscrackers <- simple_clusterprofiler(gene_list, all_genes, "org.Dm.eg.db")
}
}
\seealso{
[clusterProfiler] [AnnotationDbi] [KEGGREST]
}
