% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_limma.r
\name{simple_comparison}
\alias{simple_comparison}
\title{Perform a simple experimental/control comparison.}
\usage{
simple_comparison(subset, workbook = "simple_comparison.xls",
  sheet = "simple_comparison", basename = NA, batch = TRUE,
  combat = FALSE, combat_noscale = TRUE, pvalue_cutoff = 0.05,
  logfc_cutoff = 0.6, tooltip_data = NULL, ...)
}
\arguments{
\item{subset}{Experimental subset with two conditions to compare.}

\item{workbook}{Excel workbook to which to write.}

\item{sheet}{Excel worksheet to which to write.}

\item{basename}{Url to which to send click evens in clicky volcano/ma plots.}

\item{batch}{Whether or not to include batch in limma's model.}

\item{combat}{Whether or not to use combatMod().}

\item{combat_noscale}{Whether or not to include combat_noscale (makes combat a little less heavy-handed).}

\item{pvalue_cutoff}{P-value definition of 'significant.'}

\item{logfc_cutoff}{Fold-change cutoff of significance. 0.6 on the low end and therefore 1.6 on the high.}

\item{tooltip_data}{Text descriptions of genes if one wants google graphs.}

\item{...}{More parameters!}
}
\value{
A list containing the following pieces:
  amean_histogram = a histogram of the mean values between the two conditions
  coef_amean_cor = a correlation test between the mean values and coefficients (this should be a p-value of 1)
  coefficient_scatter = a scatter plot of condition 2 on the y axis and condition 1 on x
  coefficient_x = a histogram of the x axis
  coefficient_y = a histogram of the y axis
  coefficient_both = a histogram of both
  coefficient_lm = a description of the line described by y=slope(y/x)+b where
  coefficient_lmsummary = the r-squared and such information for the linear model
  coefficient_weights = the weights against the linear model, higher weights mean closer to the line
  comparisons = the result from eBayes()
  contrasts = the result from contrasts.fit()
  contrast_histogram = a histogram of the coefficients
  downsignificant = a subset from toptable() of the 'down-regulated' genes (< 1 Z from the mean)
  fit = the result from lmFit(voom_result)
  ma_plot = an ma plot using the voom$E data and p-values
  psignificant = a subset from toptable() of all genes with p-values <= pvalue_cutoff
  pvalue_histogram = a histogram of all the p-values
  table = everything from toptable()
  upsignificant = a subset from toptable() of 'up-regulated' genes (> 1 Z from the mean)
  volcano_plot = a volcano plot of x/y
  voom_data = the result from calling voom()
  voom_plot = a plot from voom(), redunant with voom_data
}
\description{
This is a function written primarily to provide examples for how to use limma.
It does the following:
1.  Makes a model matrix using condition/batch
2.  Optionally uses sva's combat (from cbcbSEQ)  3.  Runs voom/lmfit
4.  Sets the first element of the design to "changed" and the second to "control".
5.  Performs a makeContrasts() of changed - control.  6.  Fits them
7.  Makes histograms of the two elements of the contrast, cor.tests() them,
makes a histogram of the p-values, ma-plot, volcano-plot, writes out the results in
an excel sheet, pulls the up/down significant and p-value significant (maybe this should be
replaced with write_limma()? 8.  And returns a list containining these data and plots.
Currently this assumes that a variant of toptable was used which gives adjusted p-values.  This
is not always the case and I should check for that, but I have not yet.
}
\examples{
\dontrun{
 model = model.matrix(~ 0 + subset$conditions)
 simple_comparison(subset, model)
}
}
\seealso{
\link{hpgl_gvis_ma_plot} \link[limma]{toptable}
\link[limma]{voom} \link{hpgl_voom}
\link[limma]{lmFit} \link[limma]{makeContrasts} \link[limma]{contrasts.fit}
}

