% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{simple_cp_enricher,ANY,ANY,GeneSetCollection-method}
\alias{simple_cp_enricher,ANY,ANY,GeneSetCollection-method}
\title{Invoke simple_cp_enricher when the input database is a GeneSetCollection.}
\usage{
\S4method{simple_cp_enricher}{ANY,ANY,GeneSetCollection}(
  sig_genes,
  de_table,
  db,
  current_id = "ENSEMBL",
  needed_id = "SYMBOL",
  org = "org.Hs.eg.db"
)
}
\arguments{
\item{sig_genes}{Set of 'significant' genes as a table.}

\item{de_table}{All genes from the original analysis.}

\item{db}{Dataframe of GO->ID matching the gene names of sig_genes to GO
categories.}

\item{current_id}{Starting ID of the genes.}

\item{needed_id}{ID type to coerce the data to.}

\item{org}{DBI to do the conversion.}
}
\value{
Table of 'enriched' categories.
}
\description{
Invoke simple_cp_enricher when the input database is a GeneSetCollection.
}
