% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{simple_cprofiler}
\alias{simple_cprofiler}
\title{I cannot be trusted to type 'cluster'}
\usage{
simple_cprofiler(...)
}
\arguments{
\item{...}{Passed to simple_clusterprofiler()}
}
\description{
I cannot be trusted to type 'cluster'
}
