% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.r
\name{simple_dorothea}
\alias{simple_dorothea}
\title{Invoke dorothea in an attempt to hunt down cool TFs.}
\usage{
simple_dorothea(
  expt,
  gene_column = "ensembl_gene_id",
  hgnc_column = "hgnc_symbol",
  transform = "log2",
  conf = c("A", "B", "C"),
  dorothea_options = NULL,
  lfc = 1,
  p = 0.05,
  species = "hsapiens"
)
}
\arguments{
\item{expt}{Expressionset}

\item{gene_column}{Column in fData containing the gene IDs.}

\item{hgnc_column}{fData column containing the HGNC symbols as used
by dorothea/viper}

\item{transform}{Explicitly set the scale to log2 (TODO: improve
this)}

\item{conf}{Vector of confidence scores to filter the data.}

\item{dorothea_options}{Optional configuration list.}

\item{species}{Either human or mouse.}
}
\value{
list containing some information from dorothea and limma.
}
\description{
dorothea: https://github.com/saezlab/dorothea appears to provide
experimentally verified mappings from genes->transcription
factors as well as a set of functions which allow one to pass it an
expressionset/matrix of counts(log scale) and get back scores by
tf.  This function is an attempt to smooth it out and prod it for
usability.
}
