% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{simple_filter_counts}
\alias{simple_filter_counts}
\title{Filter low-count genes from a data set only using a simple threshold and number of samples.}
\usage{
simple_filter_counts(count_table, threshold = 2)
}
\arguments{
\item{count_table}{Data frame of (pseudo)counts by sample.}

\item{threshold}{Lower threshold of counts for each gene.}
}
\value{
Dataframe of counts without the low-count genes.
}
\description{
This was a function written by Kwame Okrah and perhaps also Laura Dillon to remove low-count
genes.  It drops genes based on a threshold and number of samples.
}
\examples{
\dontrun{
 filtered_table <- simple_filter_counts(count_table)
}
}
\seealso{
[edgeR]
}
