% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.r
\name{simple_gadem}
\alias{simple_gadem}
\title{run the rGADEM suite}
\usage{
simple_gadem()
}
\description{
This is another function I started but never had cause to finish
for the test sequences it works though
}

