% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.r
\name{simple_goseq}
\alias{simple_goseq}
\title{Perform a simplified goseq analysis.}
\usage{
simple_goseq(de_genes, all_genes = NULL, lengths = NULL, goids = NULL,
  doplot = TRUE, adjust = 0.1, pvalue = 0.1, qvalue = 0.1,
  goseq_method = "Wallenius", padjust_method = "BH", species = NULL,
  length_db = "ensGene", gff = NULL, ...)
}
\arguments{
\item{de_genes}{Data frame of differentially expressed genes, containing IDs etc.}

\item{all_genes}{Universe of possible genes.}

\item{lengths}{Length of each gene with an ID in de_genes.}

\item{goids}{List of ontology accessions to gene accessions.}

\item{doplot}{Include pwf plots?}

\item{adjust}{Minimum adjusted pvalue for 'significant.'}

\item{pvalue}{Minimum pvalue for 'significant.'}

\item{qvalue}{Minimum qvalue for 'significant.'}

\item{goseq_method}{Statistical test for goseq to use.}

\item{padjust_method}{Which method to use to adjust the pvalues.}

\item{species}{Optionally choose a species from supportedOrganisms().}

\item{length_db}{Source of gene lengths?}

\item{gff}{gff file source of gene lengths.}

\item{...}{Extra parameters which I do not recall}
}
\value{
Big list including:
  the pwd:pwf function,
  alldata:the godata dataframe,
  pvalue_histogram:p-value histograms,
  godata_interesting:the ontology information of the enhanced groups,
  term_table:the goterms with some information about them,
  mf_subset:a plot of the MF enhanced groups,
  mfp_plot:the pvalues of the MF group,
  bp_subset:a plot of the BP enhanced groups,
  bpp_plot,
  cc_subset,
  and ccp_plot
}
\description{
goseq can be pretty difficult to get set up for non-supported organisms.  This attempts to make
that process a bit simpler as well as give some standard outputs which should be similar to those
returned by clusterprofiler/topgo/gostats/gprofiler.
}
\seealso{
\pkg{goseq} \link[goseq]{goseq} \link[goseq]{nullp}
}

