% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.r
\name{simple_gprofiler}
\alias{simple_gprofiler}
\title{Run searches against the web service g:Profiler.}
\usage{
simple_gprofiler(de_genes, species = "hsapiens", first_col = "logFC",
  second_col = "limma_logfc", do_go = TRUE, do_kegg = TRUE,
  do_reactome = TRUE, do_mi = TRUE, do_tf = TRUE, do_corum = TRUE,
  do_hp = TRUE)
}
\arguments{
\item{de_genes}{guess!}

\item{species}{an organism supported by gprofiler}

\item{first_col}{where to search for the order of 'significant' first}

\item{second_col}{if that fails, try some where else.}
}
\value{
a list of results for go, kegg, reactome, and a few more.
}
\description{
Thank you Ginger for showing me your thesis, gProfiler is pretty cool!
}

