% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.r
\name{simple_gprofiler2}
\alias{simple_gprofiler2}
\title{Run searches against the web service g:Profiler.}
\usage{
simple_gprofiler2(
  sig_genes,
  species = "hsapiens",
  convert = TRUE,
  first_col = "logFC",
  second_col = "limma_logfc",
  do_go = TRUE,
  do_kegg = TRUE,
  do_reactome = TRUE,
  do_mi = TRUE,
  do_tf = TRUE,
  do_corum = TRUE,
  do_hp = TRUE,
  do_hpa = TRUE,
  do_wp = TRUE,
  significant = FALSE,
  exclude_iea = FALSE,
  do_under = FALSE,
  evcodes = TRUE,
  threshold = 0.05,
  adjp = "fdr",
  domain_scope = "annotated",
  bg = NULL,
  pseudo_gsea = TRUE,
  id_col = "row.names",
  excel = NULL
)
}
\arguments{
\item{sig_genes}{Guess!  The set of differentially expressed/interesting
genes.}

\item{species}{Organism supported by gprofiler.}

\item{convert}{Use gProfileR's conversion utility?}

\item{first_col}{First place used to define the order of 'significant'.}

\item{second_col}{If that fails, try a second column.}

\item{do_go}{Perform GO search?}

\item{do_kegg}{Perform KEGG search?}

\item{do_reactome}{Perform reactome search?}

\item{do_mi}{Do miRNA search?}

\item{do_tf}{Search for transcription factors?}

\item{do_corum}{Do corum search?}

\item{do_hp}{Do the hp search?}

\item{do_hpa}{Do the hpa search?}

\item{do_wp}{Do the wp search?}

\item{significant}{Only return the statistically significant hits?}

\item{exclude_iea}{Passed directly to gprofiler2.}

\item{do_under}{Perform under-representation search?}

\item{evcodes}{Get the set of evcodes in the data?  This makes it take
longer.}

\item{threshold}{p-value 'significance' threshold.}

\item{adjp}{Method to adjust p-values.}

\item{domain_scope}{Passed to gprofiler2.}

\item{bg}{Background genes.}

\item{pseudo_gsea}{Is the data in a ranked order by significance?}

\item{id_col}{Which column in the table should be used for gene ID
crossreferencing?  gProfiler uses Ensembl ids.  So if you have a table of
entrez or whatever, translate it!}

\item{excel}{Print the results to an excel file?}
}
\value{
a list of results for go, kegg, reactome, and a few more.
}
\description{
This is the beginning of a reimplementation to use gprofiler2.  However,
AFAICT gprofiler2 does not yet actually work for anything other than their GO
data.
}
\examples{
\dontrun{
 gprofiler_is_nice_and_easy <- simple_gprofiler(genes, species='mmusculus')
}
}
\seealso{
\pkg{gProfiler}
}
