% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.R
\name{simple_gprofiler_old}
\alias{simple_gprofiler_old}
\title{Run searches against the web service g:Profiler.}
\usage{
simple_gprofiler_old(
  sig_genes,
  species = "hsapiens",
  convert = TRUE,
  first_col = "logFC",
  second_col = "limma_logfc",
  do_go = TRUE,
  do_kegg = TRUE,
  do_reactome = TRUE,
  do_mi = TRUE,
  do_tf = TRUE,
  do_corum = TRUE,
  do_hp = TRUE,
  significant = TRUE,
  pseudo_gsea = TRUE,
  id_col = "row.names",
  excel = NULL
)
}
\arguments{
\item{sig_genes}{Guess!  The set of differentially expressed/interesting
genes.}

\item{species}{Organism supported by gprofiler.}

\item{convert}{Use gProfileR's conversion utility?}

\item{first_col}{First place used to define the order of 'significant'.}

\item{second_col}{If that fails, try a second column.}

\item{do_go}{Perform GO search?}

\item{do_kegg}{Perform KEGG search?}

\item{do_reactome}{Perform reactome search?}

\item{do_mi}{Do miRNA search?}

\item{do_tf}{Search for transcription factors?}

\item{do_corum}{Do corum search?}

\item{do_hp}{Do the hp search?}

\item{significant}{Only return the statistically significant hits?}

\item{pseudo_gsea}{Is the data in a ranked order by significance?}

\item{id_col}{Which column in the table should be used for gene ID
crossreferencing?  gProfiler uses Ensembl ids.  So if you have a table of
entrez or whatever, translate it!}

\item{excel}{Print the results to an excel file?}
}
\value{
List of results for go, kegg, reactome, and a few more.
}
\description{
Thank you Ginger for showing me your thesis, gProfiler is pretty cool!
}
\examples{
\dontrun{
 gprofiler_is_nice_and_easy <- simple_gprofiler(genes, species='mmusculus')
}
}
\seealso{
[gProfiler]
}
