% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.R
\name{simple_gsva,expt-method}
\alias{simple_gsva,expt-method}
\title{Invoke simple_gsva using an expt as input.}
\usage{
\S4method{simple_gsva}{expt}(
  eset,
  signatures = "c2BroadSets",
  data_pkg = "GSVAdata",
  signature_category = "c2",
  cores = NULL,
  current_id = "ENSEMBL",
  required_id = "ENTREZID",
  id_source = "orgdb",
  min_catsize = 5,
  orgdb = "org.Hs.eg.db",
  method = "ssgsea",
  kcdf = NULL,
  ranking = FALSE,
  msig_db = NULL,
  wanted_meta = "all",
  mx_diff = TRUE,
  verbose = FALSE,
  id_type = "entrez"
)
}
\arguments{
\item{eset}{Expt object to be analyzed.}

\item{signatures}{Provide an alternate set of signatures (GeneSetCollections)}

\item{data_pkg}{What package contains the requisite dataset?}

\item{signature_category}{Specify a subset category to extract from the signatures database.}

\item{cores}{How many CPUs to use?}

\item{current_id}{Where did the IDs of the genes come from?}

\item{required_id}{gsva (I assume) always requires ENTREZ IDs, but just in
case this is a parameter.}

\item{id_source}{How to find the IDs (DBI or fData).}

\item{min_catsize}{Minimum category size to consider interesting (passed to gsva()).}

\item{orgdb}{What is the data source for the rownames()?}

\item{method}{Which gsva method to use? Changed this from gsva to ssgsea
because it was throwing segmentation faults.}

\item{kcdf}{Options for the gsva methods.}

\item{ranking}{another gsva option.}

\item{msig_db}{File contining msigdb annotations.}

\item{wanted_meta}{Desired metadata elements from the mxig_xml file.}

\item{mx_diff}{Passed to gsva(), I do not remember what it does.}

\item{verbose}{Print some information while running?}

\item{id_type}{Specify the ID type when loading the signature database.}
}
\description{
Invoke simple_gsva using an expt as input.
}
