% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.R
\name{simple_gsva}
\alias{simple_gsva}
\title{Provide some defaults and guidance when attempting to use gsva.}
\usage{
simple_gsva(
  expt,
  signatures = "c2BroadSets",
  data_pkg = "GSVAdata",
  signature_category = "c2",
  cores = NULL,
  current_id = "ENSEMBL",
  required_id = "ENTREZID",
  min_catsize = 5,
  orgdb = "org.Hs.eg.db",
  method = "ssgsea",
  kcdf = NULL,
  ranking = FALSE,
  msig_db = NULL,
  wanted_meta = "all",
  mx_diff = TRUE,
  verbose = FALSE,
  id_type = "entrez"
)
}
\arguments{
\item{expt}{Expt object to be analyzed.}

\item{signatures}{Provide an alternate set of signatures (GeneSetCollections)}

\item{data_pkg}{What package contains the requisite dataset?}

\item{signature_category}{Specify a subset category to extract from the signatures database.}

\item{cores}{How many CPUs to use?}

\item{current_id}{Where did the IDs of the genes come from?}

\item{required_id}{gsva (I assume) always requires ENTREZ IDs, but just in
case this is a parameter.}

\item{min_catsize}{Minimum category size to consider interesting (passed to gsva()).}

\item{orgdb}{What is the data source for the rownames()?}

\item{method}{Which gsva method to use? Changed this from gsva to ssgsea
because it was throwing segmentation faults.}

\item{kcdf}{Options for the gsva methods.}

\item{ranking}{another gsva option.}

\item{msig_db}{File contining msigdb annotations.}

\item{wanted_meta}{Desired metadata elements from the mxig_xml file.}

\item{mx_diff}{Passed to gsva(), I do not remember what it does.}

\item{verbose}{Print some information while running?}

\item{id_type}{Specify the ID type when loading the signature database.}
}
\value{
List containing three elements: first a modified expressionset using
 the result of gsva in place of the original expression data; second the
 result from gsva, and third a data frame of the annotation data for the
 gene sets in the expressionset.  This seems a bit redundant, perhaps I
 should revisit it?
}
\description{
gsva seems to hold a tremendous amount of potential.  Unfortunately, it is
somewhat opaque and its requirements are difficult to pin down.  This
function will hopefully provide some of the requisite defaults and do some
sanity checking to make it more likely that a gsva analysis will succeed.
}
\seealso{
[GSEABase] [load_gmt_signatures()] [create_expt()] [GSVA]
}
