% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999_mlseq.r
\name{simple_mlseq}
\alias{simple_mlseq}
\title{Use MLSeq to seek important genes given an experimental factor and an expressionSet.}
\usage{
simple_mlseq(
  expt,
  comparison = "condition",
  number_by_var = 100,
  ceiling_factor = 1/3,
  training_number = 2,
  training_repeats = 10,
  training_method = "repeatedcv",
  classify_method = "svmRadial",
  classify_preprocess = "deseq-rlog",
  reference_factor = NULL,
  ...
)
}
\arguments{
\item{expt}{Input expressionset.}

\item{comparison}{Metadata column from the experimental design for the search.}

\item{number_by_var}{Take the top-n most variant genes. Use all genes if null.}

\item{ceiling_factor}{Define how many columns(experimental samples) to take
when sampling the expressionset for training vs. testing data.}

\item{training_number}{Iterations when training.}

\item{training_repeats}{Also iterations when training... (in other words, I dunno).}

\item{training_method}{which caret method to train?}

\item{classify_method}{which caret method to classify the data?}

\item{classify_preprocess}{Which mlseq method to preprocess/normalize the data?}

\item{reference_factor}{What factor in the experimental metadata contains the reference?}

\item{...}{Extra arguments}
}
\description{
MLSeq provides interfaces to the various machine learning methodologies from
caret in the context of RNASeq data.  It furthermore provides bridge methods
which provide links from the normalization methods from limma/edgeR/DESeq2 to
the various ML methods in caret.
}
