% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{simple_motifRG}
\alias{simple_motifRG}
\title{Run motifRG on a fasta file.}
\usage{
simple_motifRG(
  input_fasta,
  control_fasta,
  maximum = 3,
  title = "Motifs of XXX",
  prefix = "motif",
  genome = "BSgenome.Hsapiens.UCSC.hg19"
)
}
\arguments{
\item{input_fasta}{Input file.}

\item{control_fasta}{control file.}

\item{maximum}{3}

\item{title}{Output image title.}

\item{prefix}{Prefix for the output files.}

\item{genome}{Package containing the full genome.}
}
\description{
Run motifRG on a fasta file.
}
\seealso{
[motifRG]
}
