% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_kegg.r
\name{simple_pathview}
\alias{simple_pathview}
\title{Print some data onto KEGG pathways.}
\usage{
simple_pathview(
  path_data,
  indir = "pathview_in",
  outdir = "pathview",
  pathway = "all",
  species = "lma",
  from_list = NULL,
  to_list = NULL,
  suffix = "_colored",
  filenames = "id",
  fc_column = "limma_logfc",
  format = "png",
  verbose = TRUE
)
}
\arguments{
\item{path_data}{Some differentially expressed genes.}

\item{indir}{Directory into which the unmodified kegg images will be
downloaded (or already exist).}

\item{outdir}{Directory which will contain the colored images.}

\item{pathway}{Perform the coloring for a specific pathway?}

\item{species}{Kegg identifier for the species of interest.}

\item{from_list}{Regex to help in renaming KEGG categories/gene names from
one format to another.}

\item{to_list}{Regex to help in renaming KEGG categories/gene names from one
format to another.}

\item{suffix}{Add a suffix to the completed, colored files.}

\item{filenames}{Name the final files by id or name?}

\item{fc_column}{What is the name of the fold-change column to extract?}

\item{format}{Format of the resulting images, I think only png really works
well.}

\item{verbose}{When on, this function is quite chatty.}
}
\value{
A list of some information for every KEGG pathway
  downloaded/examined.  This information includes:
 a. The filename of the final image for each pathway.
 b. The number of genes which were found in each pathway image.
 c. The number of genes in the 'up' category
 d. The number of genes in the 'down' category
}
\description{
KEGGREST and pathview provide neat functions for coloring molecular pathways
with arbitrary data. Unfortunately they are somewhat evil to use.  This
attempts to alleviate that.
}
\examples{
\dontrun{
 thy_el_comp2_path = hpgl_pathview(thy_el_comp2_kegg, species = "spz", indir = "pathview_in",
                                   outdir = "kegg_thy_el_comp2", string_from = "_Spy",
                                   string_to = "_Spy_", filenames = "pathname")
}
}
\seealso{
\pkg{Ramigo} \pkg{pathview}
}
