% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{simple_proper}
\alias{simple_proper}
\title{Invoke PROPER and replace its default data set with data of interest.}
\usage{
simple_proper(
  de_tables,
  p = 0.05,
  experiment = "cheung",
  nsims = 20,
  reps = c(3, 5, 7, 10),
  de_method = "edger",
  alpha_type = "fdr",
  alpha = 0.1,
  stratify = "expr",
  target = "lfc",
  mean_or_median = "mean",
  filter = "none",
  delta = 1,
  add_coverage = TRUE,
  target_power = 0.8,
  mean_gene_length = 2000,
  nt_per_read = 200,
  describe_samples = 5
)
}
\arguments{
\item{de_tables}{A set of differential expression results, presumably from
EdgeR or DESeq2.}

\item{p}{Cutoff}

\item{experiment}{The default data set in PROPER is entitled 'cheung'.}

\item{nsims}{Number of simulations to perform.}

\item{reps}{Simulate these number of experimental replicates.}

\item{de_method}{There are a couple choices here for tools which are pretty
old, my version of this only accepts deseq or edger.}

\item{alpha_type}{I assume p-adjust type.}

\item{alpha}{Accepted fdr rate.}

\item{stratify}{There are a few options here, I don't fully understand them.}

\item{target}{Cutoff.}

\item{mean_or_median}{Use mean or median values?}

\item{filter}{Apply a filter?}

\item{delta}{Not epsilon! (E.g. I forget what this does).}
}
\value{
List containin the various tables and plots returned by PROPER.
}
\description{
Recent reviewers of Najib's grants have taken an increased interest in
knowing the statistical power of the various experiments.  He queried
Dr. Corrada-Bravo who suggested PROPER.  I spent some time looking through it
and, with some revervations, modified its workflow to (at least in theory) be
able to examine any dataset.  The workflow in question is particularly odd
and warrants further discussion/analysis.  This function is a modified
version of 'default_proper()' above and invokes PROPER after re-formatting a
given dataset in the way expected by PROPER.
}
\seealso{
[PROPER] DOI:10.1093/bioinformatics/btu640
}
