% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{simple_topgo,character-method}
\alias{simple_topgo,character-method}
\title{Coerce simple_topgo to accept a vector of gene IDs instead of a real dataframe of significance.}
\usage{
\S4method{simple_topgo}{character}(
  sig_genes,
  goid_map = "id2go.map",
  go_db = NULL,
  pvals = NULL,
  limitby = "fisher",
  limit = 0.1,
  signodes = 100,
  sigforall = TRUE,
  numchar = 300,
  selector = "topDiffGenes",
  pval_column = "deseq_adjp",
  overwrite = FALSE,
  densities = FALSE,
  pval_plots = TRUE,
  excel = NULL,
  ...
)
}
\description{
Doing this voids the topgo warantee.
}
