% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{simple_topgo}
\alias{simple_topgo}
\title{Perform a simplified topgo analysis}
\usage{
simple_topgo(de_genes, goid_map = "id2go.map", goids_df = NULL,
  pvals = NULL, limitby = "fisher", limit = 0.1, signodes = 100,
  sigforall = TRUE, numchar = 300, selector = "topDiffGenes",
  pval_column = "adj.P.Val", overwrite = FALSE, densities = FALSE,
  pval_plots = TRUE, ...)
}
\arguments{
\item{de_genes}{a data frame of differentially expressed genes, containing IDs and whatever other columns}

\item{goid_map}{a file containing mappings of genes to goids in the format expected by topgo}

\item{goids_df}{a data frame of the goids which may be used to make the goid_map}

\item{pvals}{a set of pvalues in the DE data which may be used to improve the topgo results}

\item{limitby}{test to index the results by}

\item{limit}{ontology pvalue to use as the lower limit}

\item{signodes}{I don't remember right now}

\item{sigforall}{provide the significance for all nodes?}

\item{numchar}{character limit for the table of results}

\item{selector}{a function name for choosing genes to include}

\item{pval_column}{column from which to acquire scores}

\item{overwrite}{yeah I do not remember this one either}

\item{densities}{the densities, yeah, the densities}

\item{pval_plots}{include pvalue plots of the results a la clusterprofiler}

\item{...}{other options which I do not remember right now}
}
\value{
a big list including the various outputs from topgo
}
\description{
This will attempt to make it easier to run topgo on a set of genes.
}

