% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_varpartition.r
\name{simple_varpart}
\alias{simple_varpart}
\title{I think varpart() is a bad name for a function, so I will rename it.}
\usage{
simple_varpart(...)
}
\arguments{
\item{...}{Arguments for varpart.}
}
\description{
I think varpart() is a bad name for a function, so I will rename it.
}
