% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{simple_xcell}
\alias{simple_xcell}
\title{Invoke xCell and pretty-ify the result.}
\usage{
simple_xcell(expt, label_size = NULL, col_margin = 6,
  row_margin = 12, ...)
}
\arguments{
\item{expt}{Expressionset to query.}

\item{label_size}{How large to make labels when printing the final heatmap.}

\item{col_margin}{Used by par() when printing the final heatmap.}

\item{row_margin}{Ibid.}

\item{...}{Extra arguments when normalizing the data for use with xCell.}
}
\value{
Small list providing the output from xCell, the set of signatures,
and heatmap.
}
\description{
I initially thought xCell might prove the best tool/method for exploring cell
deconvolution.  I slowly figured out its limitations, but still think it
seems pretty nifty for its use case.  Thus this function is intended to make
invoking it easier/faster.
}
