% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.R
\name{single_topgo_tree}
\alias{single_topgo_tree}
\title{Collapse the logic for collecting topgo trees into one little function.}
\usage{
single_topgo_tree(
  tg,
  score_column = "mf_fisher",
  node_data = "fmf_godata",
  score_limit = 0.1,
  sigforall = TRUE
)
}
\description{
Collapse the logic for collecting topgo trees into one little function.
}
