% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{skim_seurat_metadata}
\alias{skim_seurat_metadata}
\title{Use skimr to make pretty summaries of cell-data.}
\usage{
skim_seurat_metadata(
  sample_meta,
  obj_meta,
  meta_query = "nCount_RNA",
  group_column = NULL,
  summary_query = "numeric.mean",
  column_name = NULL,
  column_prefix = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sample_meta}{df of the known samples by name.}

\item{obj_meta}{The 'meta.data' slot of a SCD}

\item{meta_query}{Column to query.}

\item{group_column}{Column used to group the cells.}

\item{summary_query}{Which of the various data produced by skimr should be extracted?}

\item{column_name}{Add the new column with this name.}

\item{column_prefix}{And this prefix.}

\item{verbose}{Print the pretty skimr table?}
}
\value{
df with some new meta(meta?)data.
}
\description{
I think I want to expand this to handle RNA summaries as well.
}
