% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{slide_de_threshold}
\alias{slide_de_threshold}
\title{make a slideable threshold for cutting off significant genes in a DE table.}
\usage{
slide_de_threshold(
  de_table,
  contrast = 1,
  lfc = 1,
  p = 0.05,
  according_to = "deseq"
)
}
\arguments{
\item{de_table}{Table of DE results.}

\item{lfc}{LogFC starting cutoff.}

\item{p}{Starting pvalue cutoff.}

\item{according_to}{Which method to deem the arbiter of truth?}

\item{control}{Which contrast to examine?}
}
\description{
make a slideable threshold for cutting off significant genes in a DE table.
}
