% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{sm}
\alias{sm}
\title{Silence, m...}
\usage{
sm(...)
}
\arguments{
\item{...}{Some code to shut up.}
}
\value{
Whatever the code would have returned.
}
\description{
Some libraries/functions just won't shut up.  Ergo, silence, peasant!
This is a simpler silence peasant.
}

