% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{snp_add_file}
\alias{snp_add_file}
\title{Add a new snp table to a set of comparisons for clustering.}
\usage{
snp_add_file(sample, input_dir = "preprocessing/outputs",
  file_suffix = "_parsed_ratio.txt")
}
\description{
This is used by expt_snp to read input files and relatively quickly merge them.
}

