% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{snp_by_chr}
\alias{snp_by_chr}
\title{The real worker.  This extracts positions for a single chromosome and puts
them into a parallelizable data structure.}
\usage{
snp_by_chr(medians, chr_name = "01", limit = 1)
}
\arguments{
\item{medians}{A set of medians by position to look through}

\item{chr_name}{Chromosome name to search}

\item{limit}{Minimum number of median hits/position to count as a snp.}
}
\value{
A list of variant positions where each element is one chromosome.
}
\description{
The real worker.  This extracts positions for a single chromosome and puts
them into a parallelizable data structure.
}
