% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{snp_density_primers}
\alias{snp_density_primers}
\title{Create a density function given a variant output and some metadata
In this case I will take it as the result of:
  snp_count <- count_expt_snps(pheno, annot_column = "bcftable")}
\usage{
snp_density_primers(
  snp_count,
  pdata_column = "condition",
  condition = "z2.3",
  cutoff = 20,
  bin_width = 600,
  divide = FALSE,
  topn = 400,
  target_temp = 53,
  max_primer_length = 50,
  bsgenome = "BSGenome.Leishmania.panamensis.MHOMCOL81L13.v52",
  gff = "reference/lpanamensis_col_v46.gff",
  feature_type = "protein_coding_gene",
  feature_start = "start",
  feature_end = "end",
  feature_strand = "strand",
  feature_chr = "seqnames",
  feature_type_column = "type",
  feature_id = "ID",
  feature_name = "description",
  truncate = TRUE
)
}
\description{
Create a density function given a variant output and some metadata
In this case I will take it as the result of:
  snp_count <- count_expt_snps(pheno, annot_column = "bcftable")
}
