% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{snp_subset_genes}
\alias{snp_subset_genes}
\title{Look for only the variant positions in a subset of genes.}
\usage{
snp_subset_genes(
  expt,
  snp_expt,
  start_col = "start",
  end_col = "end",
  expt_name_col = "chromosome",
  snp_name_col = "chromosome",
  snp_start_col = "position",
  expt_gid_column = "gid",
  genes = c("LPAL13_120010900", "LPAL13_340013000", "LPAL13_000054100",
    "LPAL13_140006100", "LPAL13_180018500", "LPAL13_320022300")
)
}
\arguments{
\item{expt}{Initial expressionset.}

\item{snp_expt}{Variant position expressionset.}

\item{start_col}{Metadata column with the start positions for each gene.}

\item{end_col}{Metadata column with the end of the genes.}

\item{expt_name_col}{Metadata column with the chromosome names.}

\item{snp_name_col}{Ditto for the snp_expressionset.}

\item{snp_start_col}{Metadata column containing the variant positions.}

\item{expt_gid_column}{ID column for the genes.}

\item{genes}{Set of genes to cross reference.}
}
\value{
New expressionset with only the variants for the genes of interest.
}
\description{
This was written in response to a query from Nancy and Maria Adelaida who
wanted to look only at the variant positions in a few specific genes.
}
\seealso{
[GenomicRanges::makeGRangesFromDataFrame()] [IRanges::subsetByOverlaps()]
}
