% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/160_variants.r
\name{snps_intersections}
\alias{snps_intersections}
\title{Cross reference observed variants against the transcriptome annotation.}
\usage{
snps_intersections(expt, snp_result, chr_column = "seqnames")
}
\arguments{
\item{expt}{The original expressionset.  This provides the annotation data.}

\item{snp_result}{The result from get_snp_sets or count_expt_snps.}

\item{chr_column}{Column in the annotation with the chromosome names.}
}
\value{
List containing the set of intersections in the conditions contained
  in snp_result, the summary of numbers of variants per chromosome, and
  summary of numbers per gene.
}
\description{
This function should provide counts of how many variant positions were
observed with respect to each chromosome and with respect to each annotated
sequence (currently this is limited to CDS, but that is negotiable).
}
\examples{
 \dontrun{
 expt <- create_expt(metadata, gene_information)
 snp_expt <- count_expt_snps(expt)
 snp_result <- get_snp_sets(snp_expt)
 intersections <- snps_vs_intersections(expt, snp_result)
}
}
\seealso{
\code{\link{snps_vs_genes}}
}
