% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{snps_vs_genes}
\alias{snps_vs_genes}
\title{Make a summary of the observed snps/gene}
\usage{
snps_vs_genes(expt, snp_result, start_col = "start", end_col = "end")
}
\arguments{
\item{expt}{The original expressionset}

\item{snp_result}{The result from get_snp_sets()}

\item{start_col}{Which column provides the start of each gene?}

\item{end_col}{and the end column of each gene?}
}
\value{
a fun list with some information by gene.
}
\description{
Make a summary of the observed snps/gene
}
