% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{sp}
\alias{sp}
\title{Some ggplot2 stats functions have not yet implemented the new dropped_aes
flag, and it is driving me nuts.  Hopefully this will make it less frustrating for me.}
\usage{
sp(...)
}
\description{
Some ggplot2 stats functions have not yet implemented the new dropped_aes
flag, and it is driving me nuts.  Hopefully this will make it less frustrating for me.
}
