% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{state,expt-method}
\alias{state,expt-method}
\title{Extract the state of an expt vis a vis normalization.}
\usage{
\S4method{state}{expt}(expt)
}
\description{
Extract the state of an expt vis a vis normalization.
}
