% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{state<-,expt-method}
\alias{state<-,expt-method}
\title{Put the current state into an expt.}
\usage{
\S4method{state}{expt}(input) <- value
}
\arguments{
\item{input}{Input expt}

\item{value}{New state.}
}
\description{
Put the current state into an expt.
}
