% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{state<-}
\alias{state<-}
\title{Set the state of the data in an expt.}
\usage{
state(expt) <- value
}
\arguments{
\item{expt}{Experiment requiring a state update.}

\item{value}{New state!}
}
\description{
Set the state of the data in an expt.
}
