% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{state}
\alias{state}
\title{Get the state from an expt.}
\usage{
state(expt)

state(expt)
}
\arguments{
\item{expt}{Experiment containing the state.}
}
\value{
List with the methods used to modify the data (if any).
}
\description{
Get the state from an expt.

Get the state of the data in an expt.
}
