% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{steal_salmon_tx_ids}
\alias{steal_salmon_tx_ids}
\title{Steal transcript IDs from the first count table.}
\usage{
steal_salmon_tx_ids(
  meta,
  annotations,
  meta_column = "salmon_count_table",
  annot_gene_column = "ensembl_gene_id",
  annot_tx_column = "ensembl_transcript_id",
  keep_unique = TRUE
)
}
\arguments{
\item{meta}{Input metadata containing the salmon count table names.}

\item{annotations}{Extant set of gene annotations, likely from biomart.}

\item{meta_column}{metadata column with the filenames.}

\item{annot_gene_column}{Column of annotations with the gene IDs.}

\item{annot_tx_column}{Column of annotations with the transcript IDs.}

\item{keep_unique}{Drop the potential duplicate GIDs?}
}
\value{
List containing modified annotations for the genes, transcripts,
 and the map between them.
}
\description{
Steal transcript IDs from the first count table.
}
