% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{[.expt}
\alias{[.expt}
\title{Subset an expt}
\usage{
\method{[}{expt}(expt, i, j, ...)
}
\arguments{
\item{expt}{expt to subset}

\item{i}{Set of genes to keep}

\item{j}{Set of samples to keep.}

\item{...}{Parameters to pass to subset_genes/subset_expt.}
}
\description{
Working on improving my understanding of how R sets up functions by type.
}
