% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[[,expt,ANY,missing-method}
\alias{[[,expt,ANY,missing-method}
\title{Simplifying subset on metadata when i is anything and j is missing.}
\usage{
\S4method{[[}{expt,ANY,missing}(x, i, j, ...)
}
\arguments{
\item{x}{an expt}

\item{i}{Column to extract}

\item{j}{not sure}

\item{...}{extra arguments.}
}
\description{
Simplifying subset on metadata when i is anything and j is missing.
}
