% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_subset.R
\name{[[,expt,character,ANY-method}
\alias{[[,expt,character,ANY-method}
\title{Simplifying subset on metadata.}
\usage{
\S4method{[[}{expt,character,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{an expt}

\item{i}{Column to extract}
}
\description{
Simplifying subset on metadata.
}
