% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{subset_genes_se}
\alias{subset_genes_se}
\title{FIXME: Redo this with S4. Subset a SE bu genes.}
\usage{
subset_genes_se(
  se,
  column = "txtype",
  method = "remove",
  ids = NULL,
  warning_cutoff = 90,
  meta_column = NULL,
  patterns = c("snRNA", "tRNA", "rRNA"),
  ...
)
}
\arguments{
\item{se}{SummarizedExperiment from which to remove some genes..}

\item{column}{fData column to use for subsetting.}

\item{method}{Either remove explicit rows, or keep them.}

\item{ids}{Specific IDs to exclude.}

\item{warning_cutoff}{Print the sample IDs for anything which has less than this percent left.}

\item{meta_column}{Save the amount of data lost to this metadata column when not null.}

\item{patterns}{Character list of patterns to remove/keep}

\item{...}{Extra arguments are passed to arglist, currently unused.#'}
}
\description{
FIXME: Redo this with S4. Subset a SE bu genes.
}
