% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{subset_se}
\alias{subset_se}
\title{Use an expression to subset a summarized experiment.}
\usage{
subset_se(
  se,
  subset = NULL,
  ids = NULL,
  min_replicates = NULL,
  nonzero = NULL,
  coverage = NULL,
  fact = "condition",
  print_excluded = TRUE
)
}
\arguments{
\item{se}{Input se.}

\item{subset}{expression to use to subset on the metadata.}

\item{ids}{Optional vector of sample IDs.}

\item{nonzero}{A number of nonzero genes to use instead.}

\item{coverage}{A minimum coverage to use instead.}

\item{print_excluded}{Print the sampleIDs excluded by this subset.}
}
\description{
I like just passing an expression string to get subsets.  Note, as written this
only subsets samples, not genes.  subset_se_genes handles that, I should spend a
minute and clean that up.
}
