% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{subtract_de_results}
\alias{subtract_de_results}
\title{On very rare occasions, one might want to directly compare logFC values}
\usage{
subtract_de_results(
  first_table,
  second_table,
  first_lfc = "deseq_logfc",
  second_lfc = "deseq_logfc",
  first_p = "deseq_adjp",
  second_p = "deseq_adjp",
  first_name = "first",
  second_name = "second",
  excel = NULL
)
}
\arguments{
\item{first_table}{First DE table}

\item{second_table}{The second DE table}

\item{first_lfc}{logFC column from the first table}

\item{second_lfc}{and the second table.}

\item{first_p}{p-value column from the first table.}

\item{second_p}{and the second...}

\item{first_name}{Name of the first table.}

\item{second_name}{and the second...}

\item{excel}{Write the result here}
}
\description{
On very rare occasions, one might want to directly compare logFC values
}
