% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{subtract_expt}
\alias{subtract_expt}
\title{Try a very literal subtraction}
\usage{
subtract_expt(
  expt,
  new_meta,
  sample_column = "sample",
  convert_state = "cpm",
  transform_state = "raw",
  handle_negative = "zero",
  savefile = "subtracted.rda",
  ...
)
}
\arguments{
\item{expt}{Input expressionset.}

\item{new_meta}{dataframe containing the new metadata.}

\item{sample_column}{Column in the sample sheet to use to acquire the sample IDs given the
subtractions.}

\item{convert_state}{Expected state of the input data vis a vis conversion (rpkm/cpm).}

\item{transform_state}{Expected state of the input data vis a vis transformation (log/linear).}

\item{handle_negative}{Set negative subtracted values to zero?}

\item{savefile}{Save the new expt data to this file.}

\item{...}{Parameters to pass to normalize_expt()}
}
\value{
New expt
}
\description{
Try a very literal subtraction
}
