% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_expt.r
\name{sum_eupath_exon_counts}
\alias{sum_eupath_exon_counts}
\title{I want an easy way to sum counts in eupathdb-derived data sets.
These have a few things which should make this relatively easy.
Notably: The gene IDs look like: "exon_ID-1 exon_ID-2 exon_ID-3"
Therefore we should be able to quickly merge these.}
\usage{
sum_eupath_exon_counts(counts)
}
\arguments{
\item{counts}{Matrix/df/dt of count data.}
}
\value{
The same data type but with the exons summed.
}
\description{
I want an easy way to sum counts in eupathdb-derived data sets.
These have a few things which should make this relatively easy.
Notably: The gene IDs look like: "exon_ID-1 exon_ID-2 exon_ID-3"
Therefore we should be able to quickly merge these.
}
