% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{sum_exons}
\alias{sum_exons}
\title{Given a data frame of exon counts and annotation information, sum the exons.}
\usage{
sum_exons(data, gff = NULL, annotdf = NULL, parent = "Parent",
  child = "row.names")
}
\arguments{
\item{data}{a count table by exon}

\item{gff}{a gff filename}

\item{annotdf}{a dataframe of annotations (probably from gff2df)}

\item{parent}{a column from the annotations with the gene names}

\item{child}{a column from the annotations with the exon names}
}
\value{
a list of 2 data frames, counts and lengths by summed exons
}
\description{
This function will merge a count table to an annotation table by the child column.
It will then sum all rows of exons by parent gene and sum the widths of the exons.
Finally it will return a list containing a df of gene lengths and summed counts.
}
\examples{
\dontrun{
summed <- sum_exons(counts, gff='reference/xenopus_laevis.gff.xz')
}
}
\seealso{
\pkg{rtracklayer}
}

