% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{summarize_scd_clusters}
\alias{summarize_scd_clusters}
\title{Summarize scores across observed clusters in a scd.}
\usage{
summarize_scd_clusters(
  scd,
  fx = "mean",
  column_prefix = "descartes",
  column_range = NULL,
  cluster_column = "cluster_sample",
  real_column_names = NULL,
  abbreviate = TRUE,
  min_mean = NULL
)
}
\arguments{
\item{scd}{Input dataset.}

\item{fx}{Function to summarize, this may change.}

\item{column_prefix}{Prefix for the scores of interest.}

\item{column_range}{Explicitly set the range of interested columns.}

\item{cluster_column}{The column containing the information about
cluster occupancy.}

\item{real_column_names}{The original columns get names like bob1
to bobn, this can be used to make them more informative.}

\item{abbreviate}{When using mSigDB information, the category names
are exceedingly long with often a consistent prefix.}

\item{min_mean}{Currently unused, but intended to filter out gsc
which are not observed to any significant degree.}
}
\description{
Currently this assumes the set of outputs produced by Seurat's
AddModuleScore() for a gsc.  It summarizes those scores for each
cluster and gives back the mean, sd, and z.
}
