% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{summarize_ups_downs}
\alias{summarize_ups_downs}
\title{Create a summary table of the ranges of fold-change values of potential interest.}
\usage{
summarize_ups_downs(ups, downs)
}
\description{
The columns have names with explicit lfc values, but the numbers which get put in them
may represent any arbitrary cutoff employed by the caller.
}
