% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{sv_fstatistics}
\alias{sv_fstatistics}
\title{Calculate f-statistics between metadata factors and surrogate estimates.}
\usage{
sv_fstatistics(
  expt,
  num_surrogates = NULL,
  filter = TRUE,
  norm = "raw",
  convert = "cpm",
  transform = "log2",
  batch = "svaseq",
  sv_df = NULL,
  queries = c("typeofcells", "visitnumber", "donor"),
  ...
)
}
\arguments{
\item{expt}{Input expressionset, redo everything to use SE, stupid.}

\item{num_surrogates}{Specificy the number of surrogates or let it choose.}

\item{filter}{Pre-filter the data?}

\item{norm}{Pre-normalize?}

\item{convert}{Pre-convert?}

\item{transform}{Pre-transform?}

\item{batch}{Use this method}

\item{sv_df}{Or provide your own set of SVs}

\item{queries}{List of metadata factors to query.}

\item{...}{Used to make compatible with pc_fstatistics and to pass stuff to normalize().}
}
\description{
This is taken directly from Theresa's TMRC work and is her idea.  I
mainly want to be able to use it on a few datasets without risking
typeographical or logical errors.  In addition, I would like to be
able to play with things like the number of surrogates and/or other
methods of estimating them.  In addition, I have some f-statistics
of PCs vs metadata in the function 'pca_information().' which I
think is likely complementary to her work (which makes sense,
Hector was her professor before she joined us, and Hector suggested
the PC idea to me).
}
