% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{sva_modify_pvalues}
\alias{sva_modify_pvalues}
\title{Use sva's f.pvalue to adjust p-values for data adjusted by combat.}
\usage{
sva_modify_pvalues(results)
}
\arguments{
\item{results}{Table of differential expression results.}
}
\description{
This is from section 5 of the sva manual:  "Adjusting for surrogate
values using the f.pvalue function." The following chunk of code is longer
and more complex than I would like. This is because f.pvalue() assumes a
pairwise comparison of a data set containing only two experimental
factors. As a way to provide an example of _how_ to calculate
appropriately corrected p-values for surrogate factor adjusted models,
this is great; but when dealing with actual data, it falls a bit short.
}
\seealso{
[sva]
}
