% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{synchronize_expt}
\alias{synchronize_expt}
\title{Synchronize the extra elements of an expt with a new expressionset.}
\usage{
synchronize_expt(expt, previous = NULL, ...)
}
\arguments{
\item{expt}{Modified/new expt}

\item{previous}{Optional previous state to use as a template.}

\item{...}{Parameters used to fill in other optional slots.}
}
\description{
Synchronize the extra elements of an expt with a new expressionset.
}
