% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{table_style}
\alias{table_style}
\title{Set the xlsx table style}
\format{An object of class \code{character} of length 1.}
\usage{
table_style
}
\description{
Set the xlsx table style
}
\keyword{datasets}
