% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{tabularp}
\alias{tabularp}
\title{A predicate to check for the various tabular formats.}
\usage{
tabularp(x)
}
\arguments{
\item{x}{Datum to check.}
}
\description{
I might want an argument to recast the input to a specific datatype.
}
