% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{take_from_ah}
\alias{take_from_ah}
\title{Create an orgdb from an taxonID}
\usage{
take_from_ah(ahid = NULL, title = NULL, species = NULL, type = "OrgDb")
}
\arguments{
\item{taxid}{TaxonID from AnnotationHub}
}
\value{
An Orgdb instance
}
\description{
This function is a bit more fragile than I would like.  I am not completely sold
on AnnotationHub yet.
}
\examples{
\dontrun{
 orgdbi <- mytaxIdToOrgDb(taxid)
}
}
\seealso{
\pkg{AnnotationHub} \pkg{S4Vectors}
}
