% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{tar_meta_column,character-method}
\alias{tar_meta_column,character-method}
\title{Make a tarball using a metadata column given an xlsx file.}
\usage{
\S4method{tar_meta_column}{character}(
  meta,
  column = "hisatcounttable",
  output = NULL,
  compression = "xz"
)
}
\description{
Make a tarball using a metadata column given an xlsx file.
}
