% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{tar_meta_column,summarizedExperiment-method}
\alias{tar_meta_column,summarizedExperiment-method}
\title{Tar a column from a SE}
\usage{
\S4method{tar_meta_column}{summarizedExperiment}(
  meta,
  column = "hisatcounttable",
  output = NULL,
  compression = "xz"
)
}
\arguments{
\item{meta}{dataframe of the good stuff.}

\item{column}{Column containing filenames to archive.}

\item{output}{Output prefix for the tarball's name.}

\item{compression}{Actually, this might be a mistake, I think utils::tar takes 'gzip', not 'gz'?}
}
\description{
Tar a column from a SE
}
