% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{tar_meta_column}
\alias{tar_meta_column}
\title{Make an archive using a column from the metadata.}
\usage{
tar_meta_column(
  meta,
  column = "hisatcounttable",
  output = NULL,
  compression = "xz"
)
}
\arguments{
\item{meta}{dataframe of the good stuff.}

\item{column}{Column containing filenames to archive.}

\item{output}{Output prefix for the tarball's name.}

\item{compression}{Actually, this might be a mistake, I think utils::tar takes 'gzip', not 'gz'?}
}
\description{
I am hoping this will be useful for either backing up count tables or making
containerized versions of analyses.
}
