% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{tempfile}
\alias{tempfile}
\title{A cheater redefinition of tempfile.}
\usage{
tempfile(pattern = "file", tmpdir = tempdir(), fileext = "")
}
\arguments{
\item{pattern}{starting string of each tempfile.}

\item{tmpdir}{Location to put the file.}

\item{fileext}{suffix.}
}
\description{
I found this at:
https://stackoverflow.com/questions/5262332/parallel-processing-and-temporary-files
and was intrigued.  I did not think to overwrite the tempfile definition.
Something in me says this is a terrible idea.
The same page suggests creating all the tempfile names _before_
beginning the parallel operations.  I think this might be the way
to go; however I do not know how that will affect the tempfile
names produced by knitr when it is making its images.
}
