% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{test_design_model_rank,expt-method}
\alias{test_design_model_rank,expt-method}
\title{Pass an expt to test_design_model_rank.}
\usage{
\S4method{test_design_model_rank}{expt}(design, fstring = "~ condition + batch")
}
\arguments{
\item{design}{In this instance, an expressionset.}

\item{fstring}{Formula string to query.}
}
\description{
Pass an expt to test_design_model_rank.
}
