% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{test_design_model_rank}
\alias{test_design_model_rank}
\title{Check on the rank of an experimental design}
\usage{
test_design_model_rank(design, fstring = "~ condition + batch")
}
\arguments{
\item{design}{Dataframe of metadata}

\item{fstring}{Formula to test.}
}
\value{
List containing some information about degrees of freedom
}
\description{
Check on the rank of an experimental design
}
