% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pca.r
\name{test_pca_methods}
\alias{test_pca_methods}
\title{The pcaMethods package has some interesting methods to test PCA methodologies.}
\usage{
test_pca_methods(data, design = NULL, plot_colors = NULL,
  plot_labels = NULL, scale = "none", center = TRUE, plot_title = NULL,
  plot_size = 5, size_column = NULL, ...)
}
\arguments{
\item{data}{expt to poke}

\item{design}{Experimental design}

\item{plot_colors}{Colors to use when plotting}

\item{plot_labels}{Labels for the plots.}

\item{scale}{Scale them?}

\item{center}{Center them?}

\item{plot_title}{Title them?}

\item{plot_size}{Size of the sigils.}

\item{size_column}{A factor to size the sigils.}

\item{...}{arglist}
}
\description{
I made this function to be able to test out some of them.
}
