% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{tmpmd5file}
\alias{tmpmd5file}
\title{A hopefully more robust version of tempfile().}
\usage{
tmpmd5file(pattern = "", suffix = "", digits = 6, body = NULL, fileext = "")
}
\arguments{
\item{pattern}{Filename prefix.}

\item{suffix}{Filename suffix.}

\item{digits}{Currently I use Sys.time() with this number of digits.}

\item{body}{No implemented, intended to use other sources of digest()}

\item{fileext}{Filename extension as per tempfile().}
}
\value{
md5 based tempfilename.
}
\description{
A hopefully more robust version of tempfile().
}
