% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology.R
\name{topDiffGenes}
\alias{topDiffGenes}
\title{A very simple selector of strong scoring genes (by p-value)}
\usage{
topDiffGenes(allScore)
}
\arguments{
\item{allScore}{The scores of the genes}
}
\description{
This function was provided in the topGO documentation, but not defined.
It was copied/pasted here.  I have ideas for including up/down expression
but have so far deemed them not needed because I am feeding topGO
already explicit lists of genes which are up/down/whatever.
But it still is likely to be useful to be able to further subset the data.
}

