% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{topgo_pval_plot}
\alias{topgo_pval_plot}
\title{Make a pvalue plot from topgo data}
\usage{
topgo_pval_plot(topgo, wrapped_width = 20, cutoff = 0.1, n = 12,
  type = "fisher")
}
\arguments{
\item{topgo}{some data from topgo!}

\item{wrapped_width}{default=20  maximum width of the text names}

\item{cutoff}{default=0.1  p-value cutoff for the plots}

\item{n}{default=12  maximum number of ontologies to include}

\item{type}{default='fisher'  type of score to use}
}
\value{
a list of MF/BP/CC pvalue plots
}
\description{
Make a pvalue plot from topgo data
}
\seealso{
\pkg{topgo} \code{goseq}
}

