% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.r
\name{translate_ids_querymany}
\alias{translate_ids_querymany}
\title{Use mygene's queryMany to translate gene ID types}
\usage{
translate_ids_querymany(queries, from = "ensembl", to = "entrez",
  species = "human")
}
\arguments{
\item{queries}{Gene IDs to translate.}

\item{from}{Database to translate IDs from.}

\item{to}{Database to translate IDs into.}

\item{species}{Human readable species for translation (Eg. 'human' instead of 'hsapiens'.)}
}
\value{
Df of translated IDs/accessions
}
\description{
Juggling between entrez, ensembl, etc can be quite a hassel.  This hopes to make it easier.
}
\examples{
\dontrun{
 data <- translate_ids_querymany(genes)
}
}
\seealso{
\link[mygene]{queryMany}
}

