% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tritrypdb.r
\name{tritryp_downloads}
\alias{tritryp_downloads}
\title{Download the various data files from http://tritrypdb.org/}
\usage{
tritryp_downloads(version = "27", species = "lmajor", strain = "friedlin",
  dl_dir = "organdb/tritryp", quiet = TRUE)
}
\arguments{
\item{version}{What version of the tritrypdb to use?}

\item{species}{Human readable species to use.}

\item{strain}{Strain of the given species to download.}

\item{dl_dir}{Directory into which to download the various files.}

\item{quiet}{Print download progress?}
}
\description{
The tritrypdb nicely makes their downloads standardized!
}

