% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{tryCatch.W.E}
\alias{tryCatch.W.E}
\title{tryCatch both warnings and errors}
\usage{
tryCatch.W.E(expr)
}
\arguments{
\item{expr}{an expression to try}
}
\value{
a list with 'value' and 'warning', where
 'value' may be an error caught.
}
\description{
We want to catch *and* save both errors and warnings, and in the case of
a warning, also keep the computed result.
}
\details{
This was taken from:http://r.789695.n4.nabble.com/How-to-catch-both-warnings-and-errors-td3073597.html
and http://tolstoy.newcastle.edu.au/R/help/04/06/0217.html
}
\author{
Martin Maechler
}

