% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{u_plot}
\alias{u_plot}
\title{u_plot()  Plot the rank order svd$u elements to get a view of how much the first genes contribute to the total variance by PC.}
\usage{
u_plot(plotted_us)
}
\arguments{
\item{plotted_us}{a list of svd$u elements}
}
\value{
a recordPlot() plot showing the first 3 PCs by rank-order svd$u.
}
\description{
u_plot()  Plot the rank order svd$u elements to get a view of how much the first genes contribute to the total variance by PC.
}

