% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_app.R
\docType{data}
\name{ui}
\alias{ui}
\title{Set up the Main siny page.}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
ui
}
\description{
Set up the Main siny page.
}
\keyword{datasets}
