% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{unAsIs}
\alias{unAsIs}
\title{Remove the AsIs attribute from some data structure.}
\usage{
unAsIs(stuff)
}
\arguments{
\item{stuff}{The data from which to remove the AsIs classification.}
}
\description{
Notably, when using some gene ontology libraries, the returned data
structures include information which is set to type 'AsIs' which turns out to
be more than slightly difficult to work with.
}
