% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_estimation.R
\name{update.RNAseq.SimOptions.2grp}
\alias{update.RNAseq.SimOptions.2grp}
\title{Use my cheater infix ::: to handle some unexported stuff from PROPER.}
\usage{
\method{update}{RNAseq.SimOptions.2grp}(sim.opts)
}
\description{
Use my cheater infix ::: to handle some unexported stuff from PROPER.
}
