% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{upsetr_sig}
\alias{upsetr_sig}
\title{Use UpSetR to compare significant gene lists.}
\usage{
upsetr_sig(
  sig,
  according_to = "deseq",
  contrasts = NULL,
  up = TRUE,
  down = TRUE,
  both = FALSE,
  scale = 2,
  ...
)
}
\arguments{
\item{sig}{datastructure of significantly DE genes.}

\item{according_to}{Choose your favorite method.}

\item{contrasts}{Choose a specific contrast(s)}

\item{up}{Make a plot of the up genes?}

\item{down}{Make a plot of the down genes?}

\item{both}{Make a plot of the up+down genes?}

\item{scale}{Make the numbers larger and easier to read?}

\item{...}{Other parameters to pass to upset().}
}
\description{
Use UpSetR to compare significant gene lists.
}
