% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpart.r
\name{varpart}
\alias{varpart}
\title{Use variancePartition to gather some plots about a troubling dataset}
\usage{
varpart(expt, factors = c("condition", "batch"), cpus = 6, genes = 20)
}
\arguments{
\item{expt}{Some data}

\item{factors}{Character list of columns in the experiment design to query}
}
\value{
partitions  List of plots and variance data frames
}
\description{
variancePartition is the newest toy introduced by Hector
}

