% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_varpartition.r
\name{varpart_summaries}
\alias{varpart_summaries}
\title{Attempt to use variancePartition's fitVarPartModel() function.}
\usage{
varpart_summaries(expt, factors = c("condition", "batch"), cpus = 6)
}
\arguments{
\item{expt}{Input expressionset.}

\item{factors}{Set of factors to query}

\item{cpus}{Number of cpus to use in doParallel.}
}
\value{
Summaries of the new model,  in theory this would be a nicely
  batch-corrected data set.
}
\description{
Note the word 'attempt'.  This function is so ungodly slow that it probably
will never be used.
}
\seealso{
\pkg{variancePartition}
}
