% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\docType{data}
\name{verbose}
\alias{verbose}
\title{Set a default verbosity, for now this just queries if this is an interactive session.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
verbose
}
\description{
Set a default verbosity, for now this just queries if this is an interactive session.
}
\keyword{datasets}
