% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{wgcna_network}
\alias{wgcna_network}
\title{Just putting down an example from Alejandro's work and
https://bioinformaticsworkbook.org/tutorials/wgcna.html#gsc.tab=0
so that I have someplace to remember the general path wgcna takes and
as a starting point to explore further.}
\usage{
wgcna_network(expt)
}
\description{
Just putting down an example from Alejandro's work and
https://bioinformaticsworkbook.org/tutorials/wgcna.html#gsc.tab=0
so that I have someplace to remember the general path wgcna takes and
as a starting point to explore further.
}
