% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{what_happened}
\alias{what_happened}
\title{Print a string describing what happened to this data.}
\usage{
what_happened(expt = NULL, transform = "raw", convert = "raw",
  norm = "raw", filter = "raw", batch = "raw")
}
\description{
Print a string describing what happened to this data.
}

