% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{write_cds_entries}
\alias{write_cds_entries}
\title{Extract CDS sequences from a genome and set of annotations.}
\usage{
write_cds_entries(
  genome,
  annot,
  ids = NULL,
  output = "all_cds.fasta",
  strand_column = "strand",
  chr_column = "chromosome",
  start_column = "start",
  end_column = "end",
  name_column = "rownames",
  name_prefix = "lpanamensis_mcol"
)
}
\arguments{
\item{genome}{BSGenome containing the raw sequence.}

\item{annot}{Annotation dataframe.}

\item{ids}{Set of annotations to write, write them all if null.}

\item{output}{Fasta file to write.}

\item{strand_column}{Column name with the strand information.}

\item{chr_column}{Column name with the chromosomes.}

\item{start_column}{Column with the start positions.}

\item{end_column}{Column with the end positions.}

\item{name_column}{Names of the CDS}

\item{name_prefix}{Prefix to add to the entries.}
}
\description{
Given a BSGenome and some annotations, write out the CDS entries.
}
