% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{write_classifier_summary}
\alias{write_classifier_summary}
\title{Write out the results of classify_n_times().}
\usage{
write_classifier_summary(result, excel = "ML_summary.xlsx", name = NULL)
}
\arguments{
\item{result}{Ibid.}

\item{excel}{Output excel file}

\item{name}{Name of the sheet to write.}
}
\description{
Write out the results of classify_n_times().
}
