% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{write_combined_legend}
\alias{write_combined_legend}
\title{Write the legend of an excel file for combine_de_tables()}
\usage{
write_combined_legend(
  wb,
  excel_basename,
  plot_dim,
  apr,
  includes,
  padj_type,
  fancy = FALSE
)
}
\arguments{
\item{wb}{Workbook to write}

\item{excel_basename}{Where to write it}

\item{plot_dim}{Default plot size.}

\item{apr}{The all_pairwise() result.}

\item{includes}{List of booleans defining which methods to examine.}

\item{padj_type}{P-adjustment employed.}

\item{fancy}{Write fancy plots with the xlsx file?}
}
\description{
Write the legend of an excel file for combine_de_tables()
}
