% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{write_de_table}
\alias{write_de_table}
\title{Writes out the results of a single pairwise comparison.}
\usage{
write_de_table(data, type = "limma", ...)
}
\arguments{
\item{data}{Output from results().}

\item{adjust}{Pvalue adjustment chosen.}

\item{n}{Number of entries to report, 0 says do them all.}

\item{coef}{Which coefficients/contrasts to report, NULL says do them all.}

\item{excel}{Write an excel workbook?}

\item{annot_df}{Optional data frame including annotation information to include with the tables.}
}
\value{
List of data frames comprising the toptable output for each coefficient, I also added a
    qvalue entry to these toptable() outputs.
}
\description{
However, this will do a couple of things to make one's life easier:
1.  Make a list of the output, one element for each comparison of the contrast matrix
2.  Write out the results() output for them in separate .csv files and/or sheets in excel
3.  Since I have been using qvalues a lot for other stuff, add a column for them.
}
\details{
Tested in test_24deseq.R
Rewritten in 2016-12 looking to simplify combine_de_tables().  That function is far too big,
This should become a template for that.
}
\examples{
\dontrun{
 finished_comparison = eBayes(deseq_output)
 data_list = write_deseq(finished_comparison, workbook="excel/deseq_output.xls")
}
}
\seealso{
\link[deseq]{toptable} \link{write_xls}
}

