% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{write_density_primers}
\alias{write_density_primers}
\title{Write out a set of primers for testing.}
\usage{
write_density_primers(
  density_primers,
  prefix = "pf",
  column = "fwd_primer",
  fasta = "forward_primers.fasta"
)
}
\arguments{
\item{density_primers}{List containing a series of putative sequencing/PCR primers.}

\item{prefix}{Sequence name prefix, 'pf' meaning 'primer forward'.}

\item{column}{Column from the dataframe of putative primers.}

\item{fasta}{Output filename.}
}
\value{
DNAStringSet of the primers with side effect of written fasta file.
}
\description{
Write out a set of primers for testing.
}
