% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_deseq.r
\name{write_deseq}
\alias{write_deseq}
\title{Writes out the results of a deseq search using write_de_table()}
\usage{
write_deseq(data, ...)
}
\arguments{
\item{data}{Output from deseq_pairwise()}

\item{...}{Options for writing the xlsx file.}
}
\description{
Looking to provide a single interface for writing tables from deseq and friends.
}
\details{
Tested in test_24deseq.R
}
\examples{
\dontrun{
 finished_comparison = deseq_pairwise(expressionset)
 data_list = write_deseq(finished_comparison)
}
}
\seealso{
\link[deseq]{toptable} \link{write_xls}
}

