% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_varpart.R
\name{write_dream}
\alias{write_dream}
\title{Nearly a copy of write_limma().}
\usage{
write_dream(data, ...)
}
\arguments{
\item{data}{Output from limma_pairwise()}

\item{...}{Options for writing the xlsx file.}
}
\description{
This will add a couple of columns to the output table which are
specific to variancePartition's dream
}
\examples{
\dontrun{
 finished_comparison = limma_pairwise(expressionset)
 data_list = write_limma(finished_comparison)
}
}
\seealso{
[write_de_table()]
}
