% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_edger.r
\name{write_edger}
\alias{write_edger}
\title{Writes out the results of a edger search using topTags()}
\usage{
write_edger(data, adjust = "fdr", n = 0, coef = NULL,
  workbook = "excel/edger.xls", excel = FALSE, csv = FALSE,
  annot_df = NULL)
}
\arguments{
\item{data}{Output from topTags().}

\item{adjust}{Pvalue adjustment chosen.}

\item{n}{Number of entries to report, 0 says do them all.}

\item{coef}{Which coefficients/contrasts to report, NULL says do them all.}

\item{workbook}{Excel filename into which to write the data.}

\item{excel}{Write an excel workbook?}

\item{csv}{Write out csv files of the tables?}

\item{annot_df}{Optional data frame including annotation information to include with the tables.}
}
\value{
List of data frames comprising the toptable output for each coefficient, I also added a
    qvalue entry to these toptable() outputs.
}
\description{
However, this will do a couple of things to make one's life easier:
1.  Make a list of the output, one element for each comparison of the contrast matrix
2.  Write out the results() output for them in separate .csv files and/or sheets in excel
3.  Since I have been using qvalues a lot for other stuff, add a column for them.
}
\examples{
\dontrun{
 finished_comparison = topTags(edger_output)
 data_list = write_edger(finished_comparison, workbook="excel/edger_output.xls")
}
}
\seealso{
\link[edger]{toptable} \link{write_xls}
}

