% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.R
\name{write_enricher_data}
\alias{write_enricher_data}
\title{Write the results of an arbitrary enricher() result.}
\usage{
write_enricher_data(
  enricher_result,
  excel = "excel/enricher.xlsx",
  add_trees = TRUE,
  order_by = "qvalue",
  pval = 0.1,
  add_plots = TRUE,
  height = 15,
  width = 10,
  decreasing = FALSE,
  primary_key = 1,
  ...
)
}
\arguments{
\item{enricher_result}{Return from  all_enricher()}

\item{excel}{xlsx output filename.}

\item{add_trees}{Add tree plots}

\item{order_by}{Row order by this column}

\item{pval}{P cutoff}

\item{add_plots}{Add some plots to the xlsx output.}

\item{height}{Plot height in the xlsx file.}

\item{width}{Plot width in the xlsx file.}

\item{decreasing}{Reverse order of the results?}

\item{primary_key}{Use this column as the primary key.}

\item{...}{Arbitrary options passed along.}
}
\description{
Write the results of an arbitrary enricher() result.
}
