% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{write_expt}
\alias{write_expt}
\title{Write out a summarizedExperiment when using deprecated function names.}
\usage{
write_expt(...)
}
\description{
Write out a summarizedExperiment when using deprecated function names.
}
