% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{write_go_openxlsx}
\alias{write_go_openxlsx}
\title{write_go_openxlsx()  Write gene ontology tables for excel using openxlsx}
\usage{
write_go_openxlsx(lst, file)
}
\arguments{
\item{goseq}{The goseq result from simple_goseq()}

\item{cluster}{The result from simple_clusterprofiler()}

\item{topgo}{Guess}

\item{gostats}{Yep, ditto}

\item{go_file}{default='excel/merged_go'  the file to save the results.}

\item{n}{default=30  the number of ontology categories to include in each table.}
}
\value{
a set of excel sheet/coordinates
}
\description{
I have found a few tools which purportedly read/write excel files.
This implementation uses openxlsx.
}

