% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{write_go_xls}
\alias{write_go_xls}
\title{write_go_xls()  Write gene ontology tables for excel}
\usage{
write_go_xls(goseq, cluster, topgo, gostats, file = "excel/merged_go",
  dated = TRUE, n = 30, overwritefile = TRUE)
}
\arguments{
\item{goseq}{The goseq result from simple_goseq()}

\item{cluster}{The result from simple_clusterprofiler()}

\item{topgo}{Guess}

\item{gostats}{Yep, ditto}

\item{file}{default='excel/merged_go'  the file to save the results.}

\item{dated}{default=TRUE  date the excel file}

\item{n}{default=30  the number of ontology categories to include in each table.}

\item{overwritefile}{default=TRUE  overwrite an existing excel file}
}
\value{
the list of ontology information
}
\description{
Combine the results from goseq, cluster profiler, topgo, and gostats and drop them into excel
Hopefully with a relatively consistent look.
}

