% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{write_go_xlsx}
\alias{write_go_xlsx}
\title{write_go_xlsx()  Write gene ontology tables for excel using xlsx}
\usage{
write_go_xlsx(lst, file)
}
\arguments{
\item{goseq}{The goseq result from simple_goseq()}

\item{cluster}{The result from simple_clusterprofiler()}

\item{topgo}{Guess}

\item{gostats}{Yep, ditto}

\item{go_file}{default='excel/merged_go'  the file to save the results.}

\item{n}{default=30  the number of ontology categories to include in each table.}
}
\description{
I have found a few tools which purportedly read/write excel files.
This implementation uses xlsx.
}

