% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.r
\name{write_goseq_data}
\alias{write_goseq_data}
\title{Make a pretty table of goseq data in excel.}
\usage{
write_goseq_data(goseq, excel = "excel/goseq.xlsx", wb = NULL,
  add_trees = TRUE, pval = 0.1, add_plots = TRUE, height = 15,
  width = 10, ...)
}
\arguments{
\item{goseq}{A set of results from simple_goseq().}

\item{excel}{An excel file to which to write some pretty results.}

\item{wb}{Workbook object to write to.}

\item{add_trees}{Include topgoish ontology trees?}

\item{pval}{Choose a cutoff for reporting by p-value.}

\item{add_plots}{Include some pvalue plots in the excel output?}

\item{height}{Height of included plots.}

\item{width}{and their width.}

\item{...}{Extra arguments are passed to arglist.}
}
\value{
The result from openxlsx
}
\description{
It is my intention to make a function like this for each ontology tool in my repetoire
}

