% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.r
\name{write_gostats_data}
\alias{write_gostats_data}
\title{Make a pretty table of gostats data in excel.}
\usage{
write_gostats_data(gostats_result, excel = "excel/gostats.xlsx", wb = NULL,
  add_trees = TRUE, order_by = "qvalue", pval = 0.1, add_plots = TRUE,
  height = 15, width = 10, decreasing = FALSE, ...)
}
\arguments{
\item{gostats_result}{A set of results from simple_gostats().}

\item{excel}{An excel file to which to write some pretty results.}

\item{wb}{Workbook object to write to.}

\item{add_trees}{Include topgoish ontology trees?}

\item{order_by}{Which column to order the data by?}

\item{pval}{Choose a cutoff for reporting by p-value.}

\item{add_plots}{Include some pvalue plots in the excel output?}

\item{height}{Height of included plots.}

\item{width}{and their width.}

\item{decreasing}{Which order?}

\item{...}{Extra arguments are passed to arglist.}
}
\value{
The result from openxlsx in a prettyified xlsx file.
}
\description{
It is my intention to make a function like this for each ontology tool in my repetoire
}
\seealso{
\pkg{openxlsx} \pkg{gostats}
}
