% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.r
\name{write_gprofiler_data}
\alias{write_gprofiler_data}
\title{Write some excel results from a gprofiler search.}
\usage{
write_gprofiler_data(gprofiler_result, wb = NULL,
  excel = "excel/gprofiler_result.xlsx", add_plots = TRUE, ...)
}
\arguments{
\item{gprofiler_result}{The result from simple_gprofiler().}

\item{wb}{Optional workbook object, if you wish to append to an existing workbook.}

\item{excel}{Excel file to which to write.}

\item{add_plots}{Add some pvalue plots?}
}
\description{
Gprofiler is pretty awesome.  This function will attempt to write its results to an excel file.
}

