% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r, R/testfile.r
\name{write_gsva}
\alias{write_gsva}
\title{Write out my various attempts at making sense of gsva.}
\usage{
write_gsva(retlist, excel, plot_dim = 6)

write_gsva(retlist, excel, plot_dim = 6)
}
\arguments{
\item{retlist}{Result from running get_sig_gsva}

\item{excel}{Excel file to write}

\item{plot_dim}{Plot dimensions, likely needs adjustment.}
}
\description{
While I am trying to make sense of gsva, I will use this function to write
out the results I get so I can pass them to Najib/Maria Adelaida/Theresa to
see if I am making sense.

While I am trying to make sense of gsva, I will use this function to write
out the results I get so I can pass them to Najib/Maria Adelaida/Theresa to
see if I am making sense.
}
