% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_noiseq.R
\name{write_noiseq}
\alias{write_noiseq}
\title{Passes noiseq result to write_de_table.}
\usage{
write_noiseq(data, ...)
}
\arguments{
\item{data}{noiseq_pairwise() result.}

\item{...}{arbitrary arguments.}
}
\description{
Passes noiseq result to write_de_table.
}
