% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{write_normalized_se}
\alias{write_normalized_se}
\title{Write out an xlsx file of just normalized expression data.}
\usage{
write_normalized_se(
  se,
  excel = "excel/pretty_counts.xlsx",
  norm = "raw",
  convert = "cpm",
  transform = "log2",
  batch = "raw",
  filter = "cbcb",
  color_na = "#DD0000",
  merge_order = "counts_first",
  ...
)
}
\arguments{
\item{se}{Input se}

\item{excel}{Output xlsx file.}

\item{norm}{Normalization method to employ.}

\item{convert}{Conversion to apply.}

\item{transform}{Transformation to apply.}

\item{batch}{Use this batch/sv method.}

\item{filter}{Filter the data using this function.}

\item{color_na}{Highlight problematic cells with red (or not).}

\item{merge_order}{Put the counts or annotations first?}
}
\description{
As per a recent request from Marina.
}
