% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{write_sample_design}
\alias{write_sample_design}
\title{Put the metadata at the end of combined_de_tables()}
\usage{
write_sample_design(wb, apr)
}
\arguments{
\item{wb}{workbook object.}

\item{apr}{Pairwise result.}
}
\description{
For the moment this is a stupidly short function.  I am betting we will
elaborate on this over time.
}
