% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/072_de_xlsx.r
\name{write_sig_legend}
\alias{write_sig_legend}
\title{Internal function to write a legend for significant gene tables.}
\usage{
write_sig_legend(excel)
}
\arguments{
\item{excel}{xlsx file to which to write.}
}
\description{
Internal function to write a legend for significant gene tables.
}
