% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{write_snps}
\alias{write_snps}
\title{Write a matrix of variants in an alignment-esque format.}
\usage{
write_snps(expt, output_file = "funky.aln")
}
\arguments{
\item{expt}{variant expressionset.}

\item{output_file}{File to write, presumably to be passed to something like phyML.}
}
\description{
Write a matrix of variants in an alignment-esque format.
}
