% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_splicing.r
\name{write_suppa_table}
\alias{write_suppa_table}
\title{Take a set of results from suppa and attempt to write it to a pretty xlsx file.}
\usage{
write_suppa_table(
  table,
  annotations = NULL,
  by_table = "gene_name",
  by_annot = "ensembl_gene_id",
  columns = "default",
  excel = "excel/suppa_table.xlsx"
)
}
\arguments{
\item{table}{Result table from suppa.}

\item{annotations}{Set of annotation data to include with the suppa result.}

\item{by_table}{Use this column to merge the annotations and data tables from
the perspective of the data table.}

\item{by_annot}{Use this column to merge the annotations and data tables
from the perspective of the annotations.}

\item{columns}{Choose a subset of columns to include, or leave the defaults.}

\item{excel}{Provide an excel file to write.}
}
\value{
Data frame of the merged data.
}
\description{
Suppa provides a tremendous amount of output, this attempts to standardize
those results and print them to an excel sheet.
}
\examples{
 \dontrun{
 prettier_table <- write_suppa_table(suppa_result_file,
                                     annotations = gene_info,
                                     excel = "excel/pretty_suppa_table.xlsx")
}
}
\seealso{
[write_xlsx()]
}
