% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{write_svpc_fstats}
\alias{write_svpc_fstats}
\title{Write an xlsx file of SV/PC f-statistics}
\usage{
write_svpc_fstats(input, excel = "excel/svpc_fstats.xlsx")
}
\arguments{
\item{input}{Result from svpc_fstats()}

\item{excel}{Output excel file.}
}
\description{
Write an xlsx file of SV/PC f-statistics
}
