% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.r
\name{write_topgo_data}
\alias{write_topgo_data}
\title{Make a pretty table of topgo data in excel.}
\usage{
write_topgo_data(topgo_result, excel = "excel/topgo.xlsx", wb = NULL,
  order_by = "fisher", decreasing = FALSE, pval = 0.1,
  add_plots = TRUE, height = 15, width = 10, ...)
}
\arguments{
\item{topgo_result}{A set of results from simple_topgo().}

\item{excel}{An excel file to which to write some pretty results.}

\item{wb}{Workbook object to write to.}

\item{order_by}{Which column to order the results by?}

\item{decreasing}{In forward or reverse order?}

\item{pval}{Choose a cutoff for reporting by p-value.}

\item{add_plots}{Include some pvalue plots in the excel output?}

\item{height}{Height of included plots.}

\item{width}{and their width.}

\item{...}{Extra arguments are passed to arglist.}
}
\value{
The result from openxlsx in a prettyified xlsx file.
}
\description{
It is my intention to make a function like this for each ontology tool in my
repetoire
}
\seealso{
\pkg{openxlsx} \pkg{topgo}
}
