% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{write_xls}
\alias{write_xls}
\title{Write a dataframe to an excel spreadsheet sheet.}
\usage{
write_xls(data = NULL, sheet = "first", file = "excel/workbook.xls",
  rowname = "rownames", overwrite = FALSE)
}
\arguments{
\item{data}{a dataframe of information}

\item{sheet}{the name of an excel sheet in a workbook.}

\item{file}{an excel workbook to which to write.  Defaults to "excel/workbook.xls"}

\item{rowname}{include rownames?  Defalts to no.}
}
\value{
NULL, on the say it creates a workbook if necessary,
creates a sheet, and writes the data to it.
}
\description{
Write a dataframe to an excel spreadsheet sheet.
}
\examples{
## write_xls(dataframe, "hpgl_data")
## Sometimes it is a good idea to go in and delete the workbook and
## re-create it if this is used heavily, because it will get crufty.
}
\seealso{
\code{\link{loadWorkbook}}, \code{\link{createSheet}},
\code{\link{writeWorksheet}}, \code{\link{saveWorkbook}}
}

