% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{write_xlsx,list,ANY,ANY,ANY-method}
\alias{write_xlsx,list,ANY,ANY,ANY-method}
\title{Pass a list to write_xlsx.}
\usage{
\S4method{write_xlsx}{list,ANY,ANY,ANY}(
  data = NULL,
  wb = NULL,
  sheet = "first",
  excel = NULL,
  rownames = TRUE,
  start_row = 1,
  start_col = 1,
  title = NULL,
  float_format = "0.000",
  data_table = TRUE,
  freeze_first_row = TRUE,
  freeze_first_column = TRUE,
  date_format = "yyyy-mm-dd",
  column_width = "heuristic",
  ...
)
}
\description{
Pass a list to write_xlsx.
}
