% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.r
\name{xlsx_plot_png}
\alias{xlsx_plot_png}
\title{An attempt to improve the behaivor of openxlsx's plot inserter.}
\usage{
xlsx_plot_png(
  a_plot,
  wb = NULL,
  sheet = 1,
  width = 6,
  height = 6,
  res = 90,
  plotname = "plot",
  savedir = "saved_plots",
  fancy_type = "pdf",
  start_row = 1,
  start_col = 1,
  file_type = "png",
  units = "in",
  ...
)
}
\arguments{
\item{a_plot}{The plot provided}

\item{wb}{Workbook to which to write.}

\item{sheet}{Name or number of the sheet to which to add the plot.}

\item{width}{Plot width in the sheet.}

\item{height}{Plot height in the sheet.}

\item{res}{Resolution of the png image inserted into the sheet.}

\item{plotname}{Prefix of the pdf file created.}

\item{savedir}{Directory to which to save pdf copies of the plots.}

\item{fancy_type}{Plot publication quality images in this format.}

\item{start_row}{Row on which to place the plot in the sheet.}

\item{start_col}{Column on which to place the plot in the sheet.}

\item{file_type}{Currently this only does pngs, but perhaps I will parameterize this.}

\item{units}{Units for the png plotter.}

\item{...}{Extra arguments are passed to arglist (Primarily for vennerable
plots which are odd)}
}
\value{
List containing the result of the tryCatch{} used to invoke the plot prints.
}
\description{
The functions provided by openxlsx for adding plots to xlsx files are quite
nice, but they can be a little annoying.  This attempt to catch some corner cases
and potentially save an extra svg-version of each plot inserted.
}
\examples{
 \dontrun{
  fun_plot <- plot_pca(stuff)$plot
  df <- some_data_frame
  wb <- write_xlsx(df, excel = "funkytown.xlsx")$workbook
  try_results <- xlsx_plot_png(fun_plot, wb = wb)
}
}
\seealso{
[openxlsx::insertImage()]
}
