% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{xref_regions}
\alias{xref_regions}
\title{If I were smart I would use an I/GRanges for this.}
\usage{
xref_regions(
  sequence_df,
  gff,
  bin_width = 600,
  feature_type = "protein_coding_gene",
  feature_start = "start",
  feature_end = "end",
  feature_strand = "strand",
  feature_chr = "seqnames",
  feature_type_column = "type",
  feature_id = "ID",
  feature_name = "description",
  name_type = NULL,
  desc_column = "description"
)
}
\arguments{
\item{sequence_df}{dataframe of sequence regions of interest.}

\item{gff}{gff annotations against which to hunt.}

\item{bin_width}{size of the regions of interest (e.g. the amplicon size)}

\item{feature_type}{What feature type to hunt for?}

\item{feature_start}{Column containing the starts.}

\item{feature_end}{Column containing the ends.}

\item{feature_strand}{Column containing strand information.}

\item{feature_chr}{Column containing the chromosome names.}

\item{feature_type_column}{Column containing the feature types.}

\item{feature_id}{Column with the IDs (coming from the gff tags).}

\item{feature_name}{Column with the descriptive name.}

\item{name_type}{I dont remember, I think this allows one to limit the search to a feature type.}

\item{desc_column}{Use this column to extract full-text gene descriptions.}
}
\description{
But I was asked to get the closest feature if it is not inside one.
I am not sure how to do that with a ranges. Sadly, I think it will
be easier for me to just iterate over the sequence_df and query
each feature on that chromosome/scaffold.
}
