% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{ymxb_print}
\alias{ymxb_print}
\title{Print a model as y = mx + b just like in grade school!}
\usage{
ymxb_print(model)
}
\arguments{
\item{model}{Model to print from glm/lm/robustbase.}
}
\value{
a string representation of that model.
}
\description{
Because, why not!?
}
