% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{all_cprofiler}
\alias{all_cprofiler}
\title{Run simple_clusterprofiler on every table from extract_significant_genes()}
\usage{
all_cprofiler(
  sig,
  tables,
  according_to = "deseq",
  together = FALSE,
  orgdb = "org.Hs.eg.db",
  orgdb_from = NULL,
  orgdb_to = "ENTREZID",
  go_level = 3,
  pcutoff = 0.05,
  qcutoff = 0.2,
  fc_column = "logFC",
  second_fc_column = "deseq_logfc",
  internal = FALSE,
  updown = "up",
  permutations = 1000,
  min_groupsize = 5,
  kegg_prefix = NULL,
  kegg_organism = NULL,
  do_gsea = TRUE,
  categories = 12,
  do_david = FALSE,
  do_kegg = TRUE,
  padj_type = "BH",
  plot_type = "all",
  do_reactome = TRUE,
  excel = "excel/all_cp.xlsx",
  ...
)
}
\arguments{
\item{sig}{Result from extract_significant_genes}

\item{tables}{Result from combine_de_tables}

\item{according_to}{Use this result type for the clusterprofiler searches.}

\item{together}{Concatenate the up/down genes into one set?}

\item{plot_type}{Choose a plot method as the default.}

\item{...}{Arguments to pass to simple_clusterprofiler().}
}
\description{
Run simple_clusterprofiler on every table from extract_significant_genes()
}
