% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{all_cprofiler}
\alias{all_cprofiler}
\title{Run simple_clusterprofiler on every table from extract_significant_genes()}
\usage{
all_cprofiler(
  sig,
  tables,
  according_to = "deseq",
  together = FALSE,
  plot_type = "dotplot",
  excel = "all_cp.xlsx",
  ...
)
}
\arguments{
\item{sig}{Result from extract_significant_genes}

\item{tables}{Result from combine_de_tables}

\item{according_to}{Use this result type for the clusterprofiler searches.}

\item{together}{Concatenate the up/down genes into one set?}

\item{plot_type}{Choose a plot method as the default.}

\item{...}{Arguments to pass to simple_clusterprofiler().}
}
\description{
Run simple_clusterprofiler on every table from extract_significant_genes()
}
