% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{all_ontology_searches}
\alias{all_ontology_searches}
\title{all_ontology_searches()  Perform ontology searches of the output from limma.}
\usage{
all_ontology_searches(de_out, gene_lengths = NULL, goids = NULL, n = NULL,
  z = NULL, fc = NULL, overwrite = FALSE,
  goid_map = "reference/go/id2go.map", gff_file = NULL, goids_df = NULL,
  do_goseq = TRUE, do_cluster = TRUE, do_topgo = TRUE,
  do_gostats = TRUE, do_trees = FALSE, workbook = "excel/ontology.xls",
  csv = TRUE, excel = FALSE)
}
\arguments{
\item{de_out}{a list of topTables comprising limma/deseq/edger outputs.}

\item{gene_lengths}{default=NULL  a data frame of gene lengths for goseq.}

\item{goids}{default=NULL  a data frame of goids and genes.}

\item{n}{default=NULL  a number of genes at the top/bottom to search.}

\item{z}{default=NULL  a number of standard deviations to search. (if this and n are null, it assumes 1z)}

\item{overwrite}{default=FALSE  a boolean of whether to overwrite the id2go mapping for goseq.}

\item{goid_map}{default='reference/go/id2go.map'  a map file used by topGO, if it does not exist then provide goids_df to make it.}

\item{gff_file}{default=NULL  a gff file containing the annotations used by gff2genetable from clusterprofiler, which I hacked to make faster.}

\item{goids_df}{default=NULL  FIXME! a dataframe of genes and goids which I am relatively certain is no longer needed and superseded by goids.}

\item{do_goseq}{default=TRUE  perform simple_goseq()?}

\item{do_cluster}{default=TRUE  perform simple_clusterprofiler()?}

\item{do_topgo}{default=TRUE  perform simple_topgo()?}

\item{do_gostats}{default=TRUE  perform simple_gostats()?}

\item{do_trees}{default=FALSE  make topGO trees from the data?}

\item{workbook}{default='excel/ontology.xls'  generate an excel workbook of the ontology data?}

\item{csv}{default=TRUE  generate csv files using excel/ontology.csv as a basename}

\item{excel}{default=FALSE  generate the excel workbook?}
}
\value{
a list of up/down ontology results from goseq/clusterprofiler/topgo/gostats, and associated trees, all optionally.
}
\description{
This passes a set of limma results to (optionally) goseq, clusterprofiler, topgo, and gostats,
collects the outputs, and provides them as a list.  This function needs a species argument,
as I recently made the simple_() functions able to automatically use the various supported organisms.
}
\examples{
## many_comparisons = limma_pairwise(expt=an_expt)
## tables = many_comparisons$limma
## this_takes_forever = limma_ontology(tables, gene_lengths=lengthdb, goids=goids_df, z=1.5, gff_file='length_db.gff')
}

