% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{all_pairwise}
\alias{all_pairwise}
\title{all_pairwise(): Wrap up limma/DESeq2/EdgeR pairwise analyses in one call.}
\usage{
all_pairwise(expt = NULL, data = NULL, conditions = NULL,
  batches = NULL, model_cond = TRUE, model_batch = FALSE,
  model_intercept = FALSE, extra_contrasts = NULL, alt_model = NULL,
  libsize = NULL)
}
\arguments{
\item{expt}{an expt class containing count tables, normalization state, etc.}

\item{conditions}{a factor of conditions in the experiment}

\item{batches}{a factor of batches in the experiment}

\item{model_cond}{Include condition in the model?  This should pretty much always be true.}

\item{model_batch}{Include batch in the model? FALSE by default, but hopefully true often.}

\item{model_intercept}{Perform a cell-means or intercept model?  FALSE by default because I understand subtraction math better.
But I have tested and get the same answer either way.}

\item{extra_contrasts}{some extra contrasts to add to the list
This can be pretty neat, lets say one has conditions A,B,C,D,E
and wants to do (C/B)/A and (E/D)/A or (E/D)/(C/B) then use this
with a string like: "c_minus_b_ctrla = (C-B)-A, e_minus_d_ctrla = (E-D)-A,
de_minus_cb = (E-D)-(C-B),"}

\item{libsize}{I've recently figured out that libsize is far more important than I previously realized.  Play with it here.}

\item{...}{The elipsis parameter is fed to write_limma() at the end.}
}
\value{
A list of limma, deseq, edger results.
}
\description{
all_pairwise(): Wrap up limma/DESeq2/EdgeR pairwise analyses in one call.
}
\examples{
## finished_comparison = eBayes(limma_output)
## data_list = write_limma(finished_comparison, workbook="excel/limma_output.xls")
}

