% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{all_pairwise}
\alias{all_pairwise}
\title{all_pairwise()  Wrap up limma/DESeq2/EdgeR pairwise analyses in one call.}
\usage{
all_pairwise(input, conditions = NULL, batches = NULL, model_cond = TRUE,
  model_batch = TRUE, model_intercept = FALSE, extra_contrasts = NULL,
  alt_model = NULL, libsize = NULL)
}
\arguments{
\item{input}{a dataframe/vector or expt class containing count tables, normalization state, etc.}

\item{conditions}{default=NULL  a factor of conditions in the experiment}

\item{batches}{default=NULL  a factor of batches in the experiment}

\item{model_cond}{default=TRUE  include condition in the model?  This is likely always true.}

\item{model_batch}{default=FALSE  include batch in the model?}

\item{model_intercept}{default=FALSE  use an intercept model instead of cell means?}

\item{extra_contrasts}{default=NULL some extra contrasts to add to the list
This can be pretty neat, lets say one has conditions A,B,C,D,E
and wants to do (C/B)/A and (E/D)/A or (E/D)/(C/B) then use this
with a string like: "c_minus_b_ctrla = (C-B)-A, e_minus_d_ctrla = (E-D)-A,
de_minus_cb = (E-D)-(C-B),"}

\item{alt_model}{default=NULL an optional alternate model to use rather than just condition/batch}

\item{libsize}{default=NULL the library size of the original data to help voom()}

\item{...}{The elipsis parameter is fed to write_limma() at the end.}
}
\value{
A list of limma, deseq, edger results.
}
\description{
all_pairwise()  Wrap up limma/DESeq2/EdgeR pairwise analyses in one call.
}
\examples{
## finished_comparison = eBayes(limma_output)
## data_list = write_limma(finished_comparison, workbook="excel/limma_output.xls")
}

