% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{annotation,expt-method}
\alias{annotation,expt-method}
\title{A series of setMethods for expts, ExpressionSets, and SummarizedExperiments.}
\usage{
\S4method{annotation}{expt}(object)
}
\description{
A series of setMethods for expts, ExpressionSets, and SummarizedExperiments.
}
