% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{annotation,expt-method}
\alias{annotation,expt-method}
\title{A series of setMethods for expts, ExpressionSets, and SummarizedExperiments.}
\usage{
\S4method{annotation}{expt}(object)
}
\arguments{
\item{object}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{value}{New value to add to the object.}

\item{x}{I absoluately do not understand why R explicitly requires
this.}

\item{i}{Nor this.}
}
\description{
A series of setMethods for expts, ExpressionSets, and SummarizedExperiments.
}
