% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{annotation<-}
\alias{annotation<-}
\title{Add a annotation mcolumn to a se.}
\usage{
annotation(se) <- value
}
\arguments{
\item{se}{Summarized Experiment to modify.}

\item{value}{vector of batches.}
}
\description{
Add a annotation mcolumn to a se.
}
