% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{annotation}
\alias{annotation}
\title{Generic get_annotations function with a reminder if I pass it something new.}
\usage{
annotation(object, ...)
}
\arguments{
\item{object}{Input data structure.}

\item{...}{Other options}
}
\description{
Generic get_annotations function with a reminder if I pass it something new.
}
