% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{assay,expt,ANY-method}
\alias{assay,expt,ANY-method}
\title{A getter to pull the assay data from an expt.}
\usage{
\S4method{assay}{expt,ANY}(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{i}{I am guessing a subsetter}

\item{withDimnames}{I do not know.}

\item{...}{Extra args!}
}
\description{
A getter to pull the assay data from an expt.
}
