% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{assay,expt,ANY-method}
\alias{assay,expt,ANY-method}
\title{A series of setMethods for expts, ExpressionSets, and SummarizedExperiments.}
\usage{
\S4method{assay}{expt,ANY}(x, i, withDimnames = TRUE, ...)
}
\description{
A series of setMethods for expts, ExpressionSets, and SummarizedExperiments.
}
