% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoloads.r
\name{autoloads_all}
\alias{autoloads_all}
\title{Automatic loading of stuff I use, I am deprecating this now.}
\usage{
autoloads_all(update = FALSE)
}
\value{
NULL currently
}
\description{
Automatic loading of stuff I use, I am deprecating this now.
}
\seealso{
\code{\link[BiocInstaller]{biocLite}} and \code{\link{install.packages}}
}

