% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoloads.R
\name{autoloads_all}
\alias{autoloads_all}
\title{Automatic loading of stuff I use}
\usage{
autoloads_all(...)
}
\value{
NULL currently
}
\description{
Automatic loading of stuff I use
}
\seealso{
\code{\link{biocLite}} and \code{\link{install.packages}}
}

