% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoloads.r
\name{autoloads_all}
\alias{autoloads_all}
\title{Automatic loading of stuff I use, I am deprecating this now.}
\usage{
autoloads_all(update = FALSE)
}
\arguments{
\item{update}{default=FALSE  update packages?}
}
\value{
NULL currently
}
\description{
Automatic loading of stuff I use, I am deprecating this now.
}
\seealso{
\link[BiocInstaller]{biocLite} \link{install.packages}
}

