% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{backup_file}
\alias{backup_file}
\title{backup_file()  Make a backup of an existing file with n revisions, like VMS!}
\usage{
backup_file(backup_file, backups = 10)
}
\arguments{
\item{backup_file}{the file to backup.}

\item{backups}{default=10  how many revisions?}
}
\description{
backup_file()  Make a backup of an existing file with n revisions, like VMS!
}

