% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{backup_file}
\alias{backup_file}
\title{Make a backup of an existing file with n revisions, like VMS!}
\usage{
backup_file(backup_file, backups = 4)
}
\arguments{
\item{backup_file}{Filename to backup.}

\item{backups}{How many revisions?}
}
\description{
Sometimes I just want to kick myself for overwriting important files and then
I remember using VMS and wish modern computers were a little more like it.
}
