% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpgltools.r
\docType{data}
\name{base_size}
\alias{base_size}
\title{The following sets the ggplot2 default text size.}
\format{An object of class \code{numeric} of length 1.}
\usage{
base_size
}
\description{
The following sets the ggplot2 default text size.
}
\keyword{datasets}
