% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_basic.R
\name{basic_pairwise}
\alias{basic_pairwise}
\title{The simplest possible differential expression method.}
\usage{
basic_pairwise(
  input = NULL,
  design = NULL,
  conditions = NULL,
  batches = NULL,
  model_cond = TRUE,
  model_intercept = FALSE,
  alt_model = NULL,
  model_batch = FALSE,
  force = FALSE,
  keepers = NULL,
  fx = "mean",
  ...
)
}
\arguments{
\item{input}{Count table by sample.}

\item{design}{Data frame of samples and conditions.}

\item{conditions}{Not currently used, but passed from all_pairwise()}

\item{batches}{Not currently used, but passed from all_pairwise()}

\item{model_cond}{Not currently used, but passed from all_pairwise()}

\item{model_intercept}{Not currently used, but passed from all_pairwise()}

\item{alt_model}{Not currently used, but passed from all_pairwise()}

\item{model_batch}{Not currently used, but passed from all_pairwise()}

\item{force}{Force as input non-normalized data?}

\item{keepers}{Set of specific contrasts to perform instead of all.}

\item{fx}{What function to use for mean/median?}

\item{...}{Extra options passed to arglist.}
}
\value{
Df of pseudo-logFC, p-values, numerators, and denominators.
}
\description{
Perform a pairwise comparison among conditions which takes
nothing into account.  It _only_ takes the conditions, a mean value/variance
among them, divides by condition, and returns the result.  No fancy
nomalizations, no statistical models, no nothing.  It should be the very
worst method possible. But, it should also provide a baseline to compare the
other tools against, they should all do better than this, always.
}
\details{
Tested in test_27de_basic.R
This function was written after the corresponding functions in de_deseq.R,
de_edger.R, and de_limma.R.  Like those, it performs the full set of pairwise
comparisons and returns a list of the results.  As mentioned above, unlike
those, it is purposefully stupid.
}
\examples{
\dontrun{
 expt <- create_expt(metadata = "sample_sheet.xlsx", gene_info = "annotations")
 basic_de <- basic_pairwise(expt)
 basic_tables <- combine_de_tables(basic_de)
}
}
\seealso{
[deseq_pairwise()] [limma_pairwise()] [edger_pairwise()] [ebseq_pairwise()]
}
