% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_basic.r
\name{basic_pairwise}
\alias{basic_pairwise}
\title{The simplest possible differential expression method.}
\usage{
basic_pairwise(input = NULL, design = NULL, force = FALSE, ...)
}
\arguments{
\item{input}{Count table by sample.}

\item{design}{Data frame of samples and conditions.}

\item{force}{Force as input non-normalized data?}

\item{...}{Extra options passed to arglist.}
}
\value{
Df of pseudo-logFC, p-values, numerators, and denominators.
}
\description{
Perform a pairwise comparison among conditions which takes
nothing into account.  It _only_ takes the conditions, a mean value/variance among
them, divides by condition, and returns the result.  No fancy nomalizations, no
statistical models, no nothing.  It should be the very worst method possible.
But, it should also provide a baseline to compare the other tools against, they should
all do better than this, always.
}
\examples{
\dontrun{
stupid_de <- basic_pairwise(expt)
}
}
\seealso{
\pkg{limma} \pkg{DESeq2} \pkg{edgeR}
}

