% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{batch_counts}
\alias{batch_counts}
\title{batch_counts() Perform different batch corrections using limma, sva, ruvg, and cbcbSEQ}
\usage{
batch_counts(count_table, design, batch = TRUE, batch1 = "batch",
  batch2 = "time", noscale = TRUE, ...)
}
\arguments{
\item{count_table}{A matrix of counts}

\item{design}{A model matrix defining the experimental conditions/batches/etc}

\item{batch}{A string describing the method to try to remove the batch effect (or FALSE to leave it alone)}

\item{batch1}{The column in the design table describing the presumed primary covariant (batch effect)}

\item{batch2}{The column in the design table describing the seconary covariant (only used by limma atm)}

\item{noscale}{Used for combatmod, remove the scaling parameter from the invocation of the modified combat.}
}
\value{
The 'batch corrected' count table and new library size.  Please remember that this library size may not be what you want for voom/limma.
}
\description{
batch_counts() Perform different batch corrections using limma, sva, ruvg, and cbcbSEQ
}

