% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{Beta.NA}
\alias{Beta.NA}
\title{Beta.NA: Perform a quick solve to gather residuals etc
This was provided by Kwame for something which I don't remember a loong time ago.}
\usage{
Beta.NA(y, X)
}
\description{
Beta.NA: Perform a quick solve to gather residuals etc
This was provided by Kwame for something which I don't remember a loong time ago.
}

