% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_install.r
\name{bioc_all}
\alias{bioc_all}
\title{Grab a copy of all bioconductor packages and install them by type}
\usage{
bioc_all(release = "3.5", mirror = "bioconductor.statistik.tu-dortmund.de",
  base = "packages", type = "software", suppress_updates = TRUE,
  suppress_auto = TRUE, force = FALSE)
}
\arguments{
\item{release}{Bioconductor release to use, should probably be adjusted to automatically find
it.}

\item{mirror}{Bioconductor mirror to use.}

\item{base}{Base directory on the mirror to download from.}

\item{type}{Type in the tree to use (software or annotation)}

\item{suppress_updates}{For BiocLite(), don't update?}

\item{suppress_auto}{For BiocLite(), don't update?}

\item{force}{Install if already installed?}
}
\value{
a number of packages installed
}
\description{
This uses jsonlite to get a copy of all bioconductor packages by name and then iterates through
them with BiocInstaller to install all of them.  It performs a sleep between each installation in
an attempt to avoid being obnoxious.  As a result, it will of a necessity take forever.
}
\examples{
\dontrun{
 go_get_some_coffee_this_will_take_a_while <- bioc_all()
}
}
\seealso{
\pkg{BiocInstaller}
}
