% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.R
\name{calculate_aucc}
\alias{calculate_aucc}
\title{Calculate the Area under the Concordance Curve.}
\usage{
calculate_aucc(
  tbl,
  tbl2 = NULL,
  px = "deseq_adjp",
  py = "edger_adjp",
  lx = "deseq_logfc",
  ly = "edger_logfc",
  cor_method = "pearson",
  topn = 0.1
)
}
\arguments{
\item{tbl}{DE table}

\item{tbl2}{Second table}

\item{px}{first set of p-values column}

\item{py}{second set}

\item{lx}{first set of logFCs column}

\item{ly}{second set}

\item{cor_method}{Method to pass to cor().}

\item{topn}{Number of genes to consider (or percentage of the
whole).}
}
\description{
This is taken verbatim from a recent paper sent to me by Julie
Cridland.
}
