% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{cbcb_batch_effect}
\alias{cbcb_batch_effect}
\title{\code{cbcb_batch_effect()}  A function suggested by Hector Corrada Bravo and Kwame Okrah for batch removal}
\usage{
cbcb_batch_effect(normalized_counts, model)
}
\arguments{
\item{normalized_counts}{a data frame of log2cpm counts}

\item{model}{a balanced experimental model containing condition and batch factors}
}
\value{
a dataframe of residuals after subtracting batch from the model
}
\description{
During a lab meeting, the following function was suggested as a quick and dirty batch removal tool
}
\examples{
\dontrun{
newdata <- cbcb_batch_effect(counts, expt_model)
}
}
\seealso{
\code{\link[limma]{voom}} \code{\link[limma]{lmFit}}
}

