% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.R
\name{cbcb_filter_counts}
\alias{cbcb_filter_counts}
\title{Filter low-count genes from a data set using cpm data and a threshold.}
\usage{
cbcb_filter_counts(count_table, threshold = 1, min_samples = 2, libsize = NULL)
}
\arguments{
\item{count_table}{Data frame of (pseudo)counts by sample.}

\item{threshold}{Lower threshold of counts for each gene.}

\item{min_samples}{Minimum number of samples.}

\item{libsize}{Table of library sizes.}
}
\value{
Dataframe of counts without the low-count genes.
}
\description{
This was a function written by Kwame Okrah and perhaps also Laura Dillon to remove low-count
genes.  It drops genes based on a cpm threshold and number of samples.
}
\examples{
\dontrun{
 filtered_table <- cbcb_filter_counts(count_table)
}
}
\seealso{
[edgeR]
}
