% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{cbcb_filter_counts}
\alias{cbcb_filter_counts}
\title{Filter low-count genes from a data set.}
\usage{
cbcb_filter_counts(count_table, threshold = 2, min_samples = 2,
  verbose = FALSE)
}
\arguments{
\item{threshold}{lower threshold of counts (default: 4)}

\item{min_samples}{minimum number of samples (default: 2)}

\item{verbose}{If set to true, prints number of genes removed / remaining}

\item{df}{input data frame of counts by sample}
}
\value{
dataframe of counts without the low-count genes
}
\description{
Filter low-count genes from a data set.
}
\examples{
## filtered_table = filter_counts(count_table)
}
\seealso{
\code{\link{log2CPM}} which this uses to decide what to keep
}

