% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{cbcb_filter_counts}
\alias{cbcb_filter_counts}
\title{cbcb_filter_counts()  Filter low-count genes from a data set.}
\usage{
cbcb_filter_counts(count_table, threshold = 2, min_samples = 2,
  verbose = FALSE)
}
\arguments{
\item{count_table}{a data frame of (pseudo)counts by sample.}

\item{threshold}{default=2  lower threshold of counts for each gene.}

\item{min_samples}{default=2   minimum number of samples}

\item{verbose}{default=FALSE  if set to true, prints number of genes removed and remaining.}
}
\value{
dataframe of counts without the low-count genes
}
\description{
This was a function written by Kwame Okrah and perhaps also Laura Dillon to remove low-count genes.  It drops genes based on a threshold and number of samples.
}
\examples{
## filtered_table = filter_counts(count_table)
}
\seealso{
\code{\link{log2CPM}} which this uses to decide what to keep
}

