% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.r
\name{cbcb_lowfilter_counts}
\alias{cbcb_lowfilter_counts}
\title{\code{cbcb_lowfilter_counts()} Filter low-count genes from a data set using cbcbSEQ::filterCounts()}
\usage{
cbcb_lowfilter_counts(count_table, thresh = 2, min_samples = 2,
  verbose = FALSE)
}
\arguments{
\item{min_samples}{minimum number of samples (default: 2)}

\item{verbose}{If set to true, prints number of genes removed / remaining}

\item{df}{input data frame of counts by sample}

\item{threshold}{lower threshold of counts (default: 4)}
}
\value{
dataframe of counts without the low-count genes
}
\description{
\code{cbcb_lowfilter_counts()} Filter low-count genes from a data set using cbcbSEQ::filterCounts()
}
\examples{
\dontrun{
 filtered_table = cbcb_lowfilter_counts(count_table)
}
}
\seealso{
\code{\link[cbcbSEQ]{log2CPM}} which this uses to decide what to keep
}

