% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{cbcb_lowfilter_counts}
\alias{cbcb_lowfilter_counts}
\title{Filter low-count genes from a data set using filterCounts() originally from cbcbSEQ.}
\usage{
cbcb_lowfilter_counts(count_table, thresh = 2, min_samples = 2,
  libsize = NULL)
}
\arguments{
\item{count_table}{Input data frame of counts by sample.}

\item{thresh}{Lower threshold of counts.}

\item{min_samples}{Minimum number of samples.}

\item{libsize}{Pre-quantile/normalization library sizes.}
}
\value{
Dataframe of counts without the low-count genes.
}
\description{
Use a threshold and minimum number of samples to drop low-count genes from a data set.
}
\examples{
\dontrun{
 filtered_table = cbcb_lowfilter_counts(count_table)
}
}
\seealso{
log2CPM which this uses to decide what to keep.
}

