% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{cbcb_lowfilter_counts}
\alias{cbcb_lowfilter_counts}
\title{Filter low-count genes from a data set using cbcbSEQ::filterCounts()}
\usage{
cbcb_lowfilter_counts(count_table, thresh = 2, min_samples = 2,
  verbose = FALSE)
}
\arguments{
\item{count_table}{input data frame of counts by sample}

\item{thresh}{lower threshold of counts (default: 4)}

\item{min_samples}{minimum number of samples (default: 2)}

\item{verbose}{If set to true, prints number of genes removed / remaining}
}
\value{
dataframe of counts without the low-count genes
}
\description{
Filter low-count genes from a data set using cbcbSEQ::filterCounts()
}
\examples{
\dontrun{
 filtered_table = cbcb_lowfilter_counts(count_table)
}
}
\seealso{
\link[cbcbSEQ]{log2CPM} which this uses to decide what to keep
}

