% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{check_eupath_species}
\alias{check_eupath_species}
\title{Search the eupathdb metadata for a given species substring.}
\usage{
check_eupath_species(species = "Leishmania major strain Friedlin",
  metadata = NULL, ...)
}
\arguments{
\item{species}{Guess}

\item{metadata}{Eupathdb metadata.}

\item{...}{Extra arguments passed to download_eupathdb_metadata()}
}
\value{
A single entry from the eupathdb metadata.
}
\description{
If the specific species is not found, look for a reasonably approximation.
stop() if nothing is found.
}
