% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{check_includes}
\alias{check_includes}
\title{Convert a vector of yes/no by DE method to a list.}
\usage{
check_includes(
  apr,
  basic = TRUE,
  deseq = TRUE,
  ebseq = TRUE,
  edger = TRUE,
  dream = TRUE,
  limma = TRUE,
  noiseq = TRUE
)
}
\arguments{
\item{apr}{The result from all_pairwise()}

\item{basic}{The user wants the basic analysis, let us see if we
can provide it here.}

\item{deseq}{The user wants DESeq2.}

\item{ebseq}{The user wants EBSeq.}

\item{edger}{The user wants EdgeR.}

\item{dream}{The user wants the variancePartition method.}

\item{limma}{The user wants limma.}

\item{noiseq}{The user wants NoiSeq.}
}
\value{
List containing TRUE/FALSE for each method desired,
 depending on if we actually have the relevant data.
}
\description{
This compiles the set of possible methods to include in an
all_pairwise() from a series of booleans into a simpler list and
checks that the elements have some data that may be used.
}
