% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{choose_dataset}
\alias{choose_dataset}
\title{Choose a suitable data set for Edger/DESeq}
\usage{
choose_dataset(input, choose_for = "limma", force = FALSE, ...)
}
\arguments{
\item{input}{Expt input.}

\item{choose_for}{One of limma, deseq, edger, or basic.  Defines the requested data state.}

\item{force}{Force non-standard data?}

\item{...}{More options for future expansion.}
}
\value{
List the data, conditions, and batches in the data.
}
\description{
The _pairwise family of functions all demand data in specific formats.
This tries to make that consistent.
}
\details{
Invoked by _pairwise().
}
\examples{
 \dontrun{
 starting_data <- create_expt(metadata)
 modified_data <- normalize_expt(starting_data, transform="log2", norm="quant")
 a_dataset <- choose_dataset(modified_data, choose_for="deseq")
 ## choose_dataset should see that log2 data is inappropriate for DESeq2 and
 ## return it to a base10 state.
}
}
\seealso{
\code{\link{choose_binom_dataset}} \code{\link{choose_limma_dataset}}
 \code{\link{choose_basic_dataset}}
}
