% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_arc}
\alias{circos_arc}
\title{Write arcs between chromosomes in circos.}
\usage{
circos_arc(df, cfgout = "circos/conf/default.conf", first_col = "chr1",
  second_col = "chr2", color = "blue", radius = 0.75, thickness = 3)
}
\arguments{
\item{df}{Dataframe with starts/ends and the floating point information.}

\item{cfgout}{Master configuration file to write.}

\item{first_col}{Name of the first chromosome.}

\item{second_col}{Name of the second chromosome.}

\item{color}{Color of the chromosomes.}

\item{radius}{Outer radius at which to add the arcs.}

\item{thickness}{Integer thickness of the arcs.}
}
\value{
The file to which the arc configuration information was written.
}
\description{
Ok, so when I said I only do 1 chromosome images, I lied.
This function tries to make writing arcs between chromosomes easier.
It too works in 3 stages,
It writes out a data file using cfgout as a basename and the data
from df in the circos arc format into
circos/data/bob_arc.txt
It then writes out a configuration plot stanza in
circos/conf/bob_arc.conf
and finally adds an include to circos/bob.conf
}
\details{
In its current implementation, this only understands two
chromosomes.  A minimal amount of logic and data organization will
address this weakness.
}
