% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{circos_check_chromosomes}
\alias{circos_check_chromosomes}
\title{Make sure I haven't mixed and matched genomes.}
\usage{
circos_check_chromosomes(
  cfg,
  df,
  annot_chr_column = "chr",
  annot_gene_column = "rownames",
  df_chr_column = "names",
  df_gene_column = NULL
)
}
\arguments{
\item{cfg}{circos configuration object.}

\item{df}{dataframe of chromsome information.}

\item{annot_chr_column}{Column containing the chromosome names.}

\item{annot_gene_column}{Column containing the gene IDs.}

\item{df_chr_column}{Column in the cfg df containing the chromosome names.}
}
\description{
This is mostly intended to stop things early if I accidently use
the wrong reference genome, but it currently does return the number
of observed chrosomes.
}
