% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{circos_heatmap}
\alias{circos_heatmap}
\title{Write tiles of arbitrary heat-mappable data in
circos.}
\usage{
circos_heatmap(df, cfgout = "circos/conf/default.conf", colname = "datum",
  chr = "chr1", colors = NULL, outer = 0.9, width = 0.08, spacing = 0)
}
\arguments{
\item{df}{a dataframe with starts/ends and the floating point information}

\item{cfgout}{The master configuration file to write.}

\item{colname}{The name of the column with the
data of interest.}

\item{chr}{the name of the chromosome (This
currently assumes a bacterial chromosome)}

\item{colors}{the color of the histogram}

\item{outer}{the floating point radius of the circle
into which to place the plus-strand data}

\item{width}{the radial width of each tile}

\item{spacing}{the radial distance between outer,inner
and inner,whatever follows.}
}
\value{
the radius after adding the histogram and the spacing.
}
\description{
This function tries to make the writing circos heatmaps easier.  Like
circos_plus_minus() and circos_hist() it works in 3 stages,
It writes out a data file using cfgout as a basename and the data
from df in the circos histogram format into
circos/data/bob_heatmap.txt
It then writes out a configuration plot stanza in
circos/conf/bob_heatmap.conf
and finally adds an include to circos/bob.conf
}

