% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_heatmap}
\alias{circos_heatmap}
\title{Write tiles of arbitrary heat-mappable data in circos.}
\usage{
circos_heatmap(
  cfg,
  df,
  colname = "logFC",
  color_mapping = 0,
  min_value = NULL,
  max_value = NULL,
  basename = "",
  colors = NULL,
  color_choice = "spectral-9-div",
  scale_log_base = 1,
  outer = 0.9,
  rules = NULL,
  width = 0.08,
  spacing = 0.02
)
}
\arguments{
\item{cfg}{Result of circos_prefix(), contains a bunch of useful material.}

\item{df}{Dataframe with starts/ends and the floating point information.}

\item{colname}{Name of the column with the data of interest.}

\item{color_mapping}{0 means no overflows for min/max, 1 means overflows
of min get a chosen color, 2 means overflows of both min/max get chosen colors.}

\item{min_value}{Minimum value for the data.}

\item{max_value}{Maximum value for the data.}

\item{basename}{Make sure the written configuration files get different names with this.}

\item{colors}{Colors of the heat map.}

\item{color_choice}{Name of the heatmap to use, I forget how this interacts with color...}

\item{scale_log_base}{Defines how the range of colors will be ranged with
respect to the values in the data.}

\item{outer}{Floating point radius of the circle into which to place the heatmap.}

\item{rules}{some extra rules?}

\item{width}{Width of each tile in the heatmap.}

\item{spacing}{Radial distance between outer, inner, and inner to whatever follows.}
}
\value{
Radius after adding the histogram and the spacing.
}
\description{
This function tries to make the writing circos heatmaps easier.  Like
circos_plus_minus() and circos_hist() it works in 3 stages,
It writes out a data file using cfgout as a basename and the data
from df in the circos histogram format into
circos/data/bob_heatmap.txt
It then writes out a configuration plot stanza in
circos/conf/bob_heatmap.conf
and finally adds an include to circos/bob.conf
}
