% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circos.R
\name{circos_hist}
\alias{circos_hist}
\title{circos_hist()  Write histograms of arbitrary floating point data in
circos.}
\usage{
circos_hist(df, cfgout = "circos/conf/default.conf", colname = "datum",
  chr = "chr1", color = "blue", fill_color = "blue", outer = 0.9,
  width = 0.08, spacing = 0)
}
\arguments{
\item{df}{a dataframe with starts/ends and the floating point information}

\item{cfgout}{default='circos/conf/default.conf'  The master
configuration file to write.}

\item{colname}{default='datum'  The name of the column with the
data of interest.}

\item{chr}{default='chr1'  the name of the chromosome (This
currently assumes a bacterial chromosome)}

\item{color}{default='blue'  the color of the histogram}

\item{fill_color}{default='blue'  guess}

\item{outer}{default=1.0  the floating point radius of the circle
into which to place the plus-strand data}

\item{width}{default=0.08  the radial width of each tile}

\item{spacing}{default=0.0  the radial distance between outer,inner
    and inner,whatever follows.}
}
\value{
the radius after adding the histogram and the spacing.
}
\description{
This function tries to make the writing of histogram data in circos
easier.  Like circos_plus_minus() it works in 3 stages,
It writes out a data file using cfgout as a basename and the data
from df in the circos histogram format into
circos/data/bob_hist.txt
It then writes out a configuration plot stanza in
circos/conf/bob_hist.conf
and finally adds an include to circos/bob.conf
}

