% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{circos_ideogram}
\alias{circos_ideogram}
\title{Create the description of chromosome markings}
\usage{
circos_ideogram(name = "default", conf_dir = "circos/conf",
  band_url = NULL)
}
\arguments{
\item{name}{the name of the configuration}

\item{conf_dir}{where does the configuration live?}

\item{band_url}{provide a url for making these imagemaps.}
}
\value{
undef
}
\description{
This function writes ideogram files for circos.
Currently it only has a single format.
}

