% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{circos_ideogram}
\alias{circos_ideogram}
\title{circos_ideogram()  Create the description of chromosome markings}
\usage{
circos_ideogram(name = "default", conf_dir = "circos/conf",
  band_url = NULL)
}
\arguments{
\item{name}{default='default'  the name of the configuration}

\item{conf_dir}{default='circos/conf'  where does the configuration live?}

\item{band_url}{default=NULL  provide a url for making these imagemaps.}
}
\value{
undef
}
\description{
This function writes ideogram files for circos.
Currently it only has a single format.
}

