% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{circos_karyotype}
\alias{circos_karyotype}
\title{Create the description of (a)chromosome(s) for circos.}
\usage{
circos_karyotype(
  cfg,
  segments = 6,
  color = "white",
  fasta = NULL,
  lengths = NULL,
  chromosomes = NULL
)
}
\arguments{
\item{cfg}{Result from circos_prefix(), contains a bunch of useful things.}

\item{segments}{How many segments to cut the chromosome into?}

\item{color}{Color segments of the chromosomal arc?}

\item{fasta}{Fasta file to use to create the karyotype.}

\item{lengths}{If no sequence file is provided, use a named numeric vector to provide them.}
}
\value{
The output filename.
}
\description{
This function tries to save me from having to get the lengths of arcs for
bacterial chromosomes manually correct, and writes them as a circos
compatible karyotype file. The outfile parameter was chosen to match the
configuration directive outlined in circos_prefix(), however that will need
to be changed in order for this to work in variable conditions.  Next time I
make one of these graphs I will do that I suspect. In addition, this
currently only understands how to write bacterial chromosomes, that will
likely be fixed when I am asked to write out a L.major karyotype. These
defaults were chosen because I have a chromosome of this length that is
correct.
}
