% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_karyotype}
\alias{circos_karyotype}
\title{Create the description of (a)chromosome(s) for circos.}
\usage{
circos_karyotype(name = "default", conf_dir = "circos/conf",
  length = NULL, chr_name = "chr1", segments = 6, color = "white",
  chr_num = 1, fasta = NULL)
}
\arguments{
\item{name}{Name of the chromosome (This currently assumes a bacterial chromosome).}

\item{conf_dir}{Where to put the circos configuration file(s).}

\item{length}{Length of the chromosome (the default is mgas5005).}

\item{chr_name}{Short name of the chromosome.}

\item{segments}{How many segments to cut the chromosome into?}

\item{color}{Color segments of the chromosomal arc?}

\item{chr_num}{Number to record for each chromosome.}

\item{fasta}{Fasta file to use to create the karyotype.}
}
\value{
The output filename.
}
\description{
This function tries to save me from having to get the lengths of arcs for
bacterial chromosomes manually correct, and writes them as a circos
compatible karyotype file. The outfile parameter was chosen to match the
configuration directive outlined in circos_prefix(), however that will need
to be changed in order for this to work in variable conditions.  Next time I
make one of these graphs I will do that I suspect. In addition, this
currently only understands how to write bacterial chromosomes, that will
likely be fixed when I am asked to write out a L.major karyotype. These
defaults were chosen because I have a chromosome of this length that is
correct.
}
