% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{circos_make}
\alias{circos_make}
\title{circos_make()  Write a simple makefile for circos.}
\usage{
circos_make(target = "", output = "circos/Makefile",
  circos = "/usr/bin/circos")
}
\arguments{
\item{target}{default=''  the make target}

\item{output}{default='circos/Makefile' the makefile}

\item{circos}{default='/usr/bin/circos'  the location of circos. (I
have a copy in home/bin/circos and use that sometimes.}
}
\value{
a kitten
}
\description{
I regenerate all my circos pictures with make(1).  This is my
makefile.
}

