% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_make}
\alias{circos_make}
\title{Write a simple makefile for circos.}
\usage{
circos_make(cfg, target = "", circos = "circos", verbose = FALSE)
}
\arguments{
\item{cfg}{Configuration from circos_prefix().}

\item{target}{Default make target.}

\item{circos}{Location of circos.  I have a copy in home/bin/circos and use that sometimes.}
}
\value{
a kitten
}
\description{
I regenerate all my circos pictures with make(1).  This is my makefile.
}
