% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_make}
\alias{circos_make}
\title{Write a simple makefile for circos.}
\usage{
circos_make(target = "", output = "circos/Makefile", circos = "circos")
}
\arguments{
\item{target}{Default make target.}

\item{output}{Makefile to write.}

\item{circos}{Location of circos.  I have a copy in home/bin/circos and use that sometimes.}
}
\value{
a kitten
}
\description{
I regenerate all my circos pictures with make(1).  This is my makefile.
}
