% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_plus_minus}
\alias{circos_plus_minus}
\title{Write tiles of bacterial ontology groups using the categories from
microbesonline.org.}
\usage{
circos_plus_minus(table, cfgout = "circos/conf/default.conf",
  chr = "chr1", outer = 1, width = 0.08, spacing = 0,
  acol = "orange", bcol = "reds-9-seq", ccol = "yellow",
  dcol = "vlpurple", ecol = "vlgreen", fcol = "dpblue",
  gcol = "vlgreen", hcol = "vlpblue", icol = "vvdpgreen",
  jcol = "dpred", kcol = "orange", lcol = "vvlorange",
  mcol = "dpgreen", ncol = "vvlpblue", ocol = "vvlgreen",
  pcol = "vvdpred", qcol = "ylgn-3-seq", rcol = "vlgrey",
  scol = "grey", tcol = "vlpurple", ucol = "greens-3-seq",
  vcol = "vlred", wcol = "vvdppurple", xcol = "black",
  ycol = "lred", zcol = "vlpblue")
}
\arguments{
\item{table}{Dataframe with starts/ends and categories.}

\item{cfgout}{Master configuration file to write.}

\item{chr}{Name of the chromosome.}

\item{outer}{Floating point radius of the circle into which to place the
plus-strand data.}

\item{width}{Radial width of each tile.}

\item{spacing}{Radial distance between outer, inner, and inner to whatever follows.}

\item{acol}{A color: RNA processing and modification.}

\item{bcol}{B color: Chromatin structure and dynamics.}

\item{ccol}{C color: Energy production conversion.}

\item{dcol}{D color: Cell cycle control, mitosis and meiosis.}

\item{ecol}{E color: Amino acid transport metabolism.}

\item{fcol}{F color: Nucleotide transport and metabolism.}

\item{gcol}{G color: Carbohydrate transport and metabolism.}

\item{hcol}{H color: Coenzyme transport and metabolism.}

\item{icol}{I color: Lipid transport and metabolism.}

\item{jcol}{J color: Translation, ribosome structure and biogenesis.}

\item{kcol}{K color: Transcription.}

\item{lcol}{L color: Replication, recombination, and repair.}

\item{mcol}{M color: Cell wall/membrane biogenesis.}

\item{ncol}{N color: Cell motility}

\item{ocol}{O color: Posttranslational modification, protein turnover, chaperones.}

\item{pcol}{P color: Inorganic ion transport and metabolism.}

\item{qcol}{Q color: Secondary metabolite biosynthesis, transport, and catabolism.}

\item{rcol}{R color: General function prediction only.}

\item{scol}{S color: Function unknown.}

\item{tcol}{T color: Signal transduction mechanisms.}

\item{ucol}{U color: Intracellular trafficking(sp?) and secretion.}

\item{vcol}{V color: Defense mechanisms.}

\item{wcol}{W color: Extracellular structures.}

\item{xcol}{X color: Not in COG.}

\item{ycol}{Y color: Nuclear structure.}

\item{zcol}{Z color: Cytoskeleton.}
}
\value{
Radius after adding the plus/minus information and the spacing between them.
}
\description{
This function tries to save me from writing out ontology
definitions and likely making mistakes.  It uses the start/ends
from the gff annotation along with the 1 letter GO-like categories
from microbesonline.org.  It then writes two data files
circos/data/bob_plus_go.txt, circos/data/bob_minus_go.txt along
with two configuration files circos/conf/bob_minus_go.conf and
circos/conf/bob_plus_go.conf and finally adds an include to
circos/bob.conf
}
