% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_plus_minus}
\alias{circos_plus_minus}
\title{Write tiles of bacterial ontology groups using the categories from microbesonline.org.}
\usage{
circos_plus_minus(table, cfgout = "circos/conf/default.conf", chr = "chr1",
  outer = 1, width = 0.08, spacing = 0)
}
\arguments{
\item{cfgout}{Master configuration file to write.}

\item{chr}{Name of the chromosome.}

\item{outer}{Floating point radius of the circle into which to place the plus-strand data.}

\item{width}{Radial width of each tile.}

\item{spacing}{Radial distance between outer, inner, and inner to whatever follows.}

\item{go_table}{Dataframe with starts/ends and categories.}
}
\value{
Radius after adding the plus/minus information and the spacing between them.
}
\description{
This function tries to save me from writing out ontology
definitions and likely making mistakes.  It uses the start/ends
from the gff annotation along with the 1 letter GO-like categories
from microbesonline.org.  It then writes two data files
circos/data/bob_plus_go.txt, circos/data/bob_minus_go.txt along
with two configuration files circos/conf/bob_minus_go.conf and
circos/conf/bob_plus_go.conf and finally adds an include to
circos/bob.conf
}
