% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circos.R
\name{circos_plus_minus}
\alias{circos_plus_minus}
\title{circos_plus_minus()  Write tiles of bacterial ontology groups using the categories from microbesonline.org}
\usage{
circos_plus_minus(go_table, cfgout = "circos/conf/default.conf",
  chr = "chr1", outer = 1, width = 0.08, spacing = 0)
}
\arguments{
\item{go_table}{a dataframe with starts/ends and categories}

\item{cfgout}{default='circos/conf/default.conf'  The master
configuration file to write.}

\item{chr}{default='chr1'  the name of the chromosome (This currently assumes a bacterial chromosome)}

\item{outer}{default=1.0  the floating point radius of the circle
into which to place the plus-strand data}

\item{width}{default=0.08  the radial width of each tile}

\item{spacing}{default=0.0  the radial distance between outer,inner
    and inner,whatever follows.}
}
\value{
the radius after adding the plus/minus information and the spacing between them.
}
\description{
This function tries to save me from writing out ontology
definitions and likely making mistakes.  It uses the start/ends
from the gff annotation along with the 1 letter GO-like categories
from microbesonline.org.  It then writes two data files
circos/data/bob_plus_go.txt, circos/data/bob_minus_go.txt along
with two configuration files circos/conf/bob_minus_go.conf and
circos/conf/bob_plus_go.conf and finally adds an include to
circos/bob.conf
}

