% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circos.R
\name{circos_prefix}
\alias{circos_prefix}
\title{circos_prefix()  Write the beginning of a circos configuration file.}
\usage{
circos_prefix(cfgout = "circos/conf/default.conf", radius = 1800)
}
\arguments{
\item{cfgout}{default='circos/conf/default.conf'  The master
configuration file to write.}

\item{radius}{default=1800  The size of the image.}
}
\value{
undef
}
\description{
A few parameters need to be set when starting circos.  This sets
some of them and gets ready for plot stanzas.
}
\details{
In its current implementation, this really assumes that there will
be no highlight stanzas and at most 1 link stanza.
chromosomes.  A minimal amount of logic and data organization will
address these weaknesses.
}

