% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_prefix}
\alias{circos_prefix}
\title{Write the beginning of a circos configuration file.}
\usage{
circos_prefix(
  name = "mgas",
  conf_dir = "circos/conf",
  radius = 1800,
  chr_units = 1000,
  band_url = NULL,
  ...
)
}
\arguments{
\item{name}{Name of the map, called with 'make name'.}

\item{conf_dir}{Directory containing the circos configuration data.}

\item{radius}{Size of the image.}

\item{chr_units}{How often to print chromosome in 'prefix' units.}

\item{band_url}{Place to imagemap link.}

\item{...}{Extra arguments passed to the tick/karyotype makers.}
}
\value{
The master configuration file name.
}
\description{
A few parameters need to be set when starting circos.  This sets
some of them and gets ready for plot stanzas.
}
\details{
In its current implementation, this really assumes that there will
be no highlight stanzas and at most 1 link stanza.
chromosomes.  A minimal amount of logic and data organization will
address these weaknesses.
}
