% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{circos_prefix}
\alias{circos_prefix}
\title{circos_prefix()  Write the beginning of a circos configuration file.}
\usage{
circos_prefix(name = "default", conf_dir = "circos/conf", radius = 1800,
  band_url = NULL)
}
\arguments{
\item{name}{default='default'  The name of the map, called with 'make name'}

\item{radius}{default=1800  The size of the image.}

\item{circos_dir}{default='circos/conf'  The directory containing the circos configuration data.}
}
\value{
undef
}
\description{
A few parameters need to be set when starting circos.  This sets
some of them and gets ready for plot stanzas.
}
\details{
In its current implementation, this really assumes that there will
be no highlight stanzas and at most 1 link stanza.
chromosomes.  A minimal amount of logic and data organization will
address these weaknesses.
}

