% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circos.R
\name{circos_suffix}
\alias{circos_suffix}
\title{circos_suffix()  Write the end of a circos master configuration.}
\usage{
circos_suffix(cfgout = "circos/conf/default.conf")
}
\arguments{
\item{cfgout}{default='circos/conf/default.conf'  The master
configuration file to write.}
}
\value{
undef
}
\description{
circos configuration files need an ending.  This writes it.
}

