% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.R
\name{circos_suffix}
\alias{circos_suffix}
\title{Write the end of a circos master configuration.}
\usage{
circos_suffix(cfg)
}
\arguments{
\item{cfg}{Result from circos_prefix()}
}
\value{
Filename of the configuration.
}
\description{
circos configuration files need an ending.  This writes it.
}
