% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos.r
\name{circos_ticks}
\alias{circos_ticks}
\title{Create the ticks for a circos plot.}
\usage{
circos_ticks(name = "default", conf_dir = "circos/conf",
  tick_separation = 2, min_label_distance = 0, label_separation = 5,
  label_offset = 5, label_size = 8, multiplier = 0.001,
  main_color = "black", main_thickness = 3, main_size = 20,
  first_size = 10, first_spacing = 1, first_color = "black",
  first_show_label = "no", first_label_size = 12, second_size = 15,
  second_spacing = 5, second_color = "black",
  second_show_label = "yes", second_label_size = 16, third_size = 18,
  third_spacing = 10, third_color = "black",
  third_show_label = "yes", third_label_size = 16,
  fourth_spacing = 100, fourth_color = "black",
  fourth_show_label = "yes", suffix = " kb", fourth_label_size = 36,
  include_first_label = TRUE, include_second_label = TRUE,
  include_third_label = TRUE, include_fourth_label = TRUE, ...)
}
\arguments{
\item{name}{Name of the configuration file to which to add the ideogram.}

\item{conf_dir}{Where does the configuration live?}

\item{tick_separation}{Top-level separation between tick marks.}

\item{min_label_distance}{distance to the edge of the plot for labels.}

\item{label_separation}{radial distance between labels.}

\item{label_offset}{The offset for the labels.}

\item{label_size}{Top-level label size.}

\item{multiplier}{When writing the position, by what factor to lower the numbers?}

\item{main_color}{Color for top-level labels?}

\item{main_thickness}{Top-level thickness of lines etc.}

\item{main_size}{Top-level size of text.}

\item{first_size}{Second level size of text.}

\item{first_spacing}{Second level spacing of ticks.}

\item{first_color}{Second-level text color.}

\item{first_show_label}{Show a label for the second level ticks?}

\item{first_label_size}{Text size for second level labels?}

\item{second_size}{Size of ticks for the third level.}

\item{second_spacing}{third-level spacing}

\item{second_color}{Text color for the third level.}

\item{second_show_label}{Give them a label?}

\item{second_label_size}{And a size.}

\item{third_size}{Now for the size of the almost-largest ticks}

\item{third_spacing}{How far apart?}

\item{third_color}{and their color}

\item{third_show_label}{give a label?}

\item{third_label_size}{and a size.}

\item{fourth_spacing}{The largest ticks!}

\item{fourth_color}{The largest color.}

\item{fourth_show_label}{Provide a label?}

\item{suffix}{String for printing chromosome distances.}

\item{fourth_label_size}{They are big!}

\item{include_first_label}{Provide the smallest labels?}

\item{include_second_label}{Second smallest labels?}

\item{include_third_label}{Second biggest labels?}

\item{include_fourth_label}{Largest labels?}

\item{...}{Extra arguments from circos_prefix().}
}
\value{
The file to which the ideogram configuration was written.
}
\description{
This function writes ticks for circos.  This has lots of options, the
defaults are all taken from the circos example documentation for a bacterial
genome.
}
