% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{clean_pkg}
\alias{clean_pkg}
\title{Cleans up illegal characters in packages generated by make_organismdbi(), make_orgdb(),
and make_txdb(). This attempts to fix some of the common problems therein.}
\usage{
clean_pkg(path, removal = "-like", replace = "", sqlite = TRUE)
}
\arguments{
\item{path}{Location for the original Db/Dbi instance.}

\item{removal}{String to remove from the instance.}

\item{replace}{What to replace removal with, when necessary.}

\item{sqlite}{Also modify the sqlite database?}
}
\value{
A hopefully cleaner OrgDb/TxDb/OrganismDbi sqlite package.
}
\description{
The primary problem this function seeks to solve is derived from the fact
that some species names in the eupathdb contain characters which are not
allowed in orgdb/txdb/organismdbi instances.  Thus this invokes a couple of
regular expressions in an attempt to make sure these generated packages are
actually installable.
}
\details{
One thing I should consider is to add some of this logic to my eupath queries
rather than perform these clunky modifications to the already-generated
packages.
}
\author{
atb
}
