% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.r
\name{cluster_trees}
\alias{cluster_trees}
\title{Take clusterprofile group data and print it on a tree as per topGO.}
\usage{
cluster_trees(de_genes, cpdata, goid_map = "id2go.map", go_db = NULL,
  score_limit = 0.2, overwrite = FALSE, selector = "topDiffGenes",
  pval_column = "adj.P.Val")
}
\arguments{
\item{de_genes}{List of genes deemed 'interesting'.}

\item{cpdata}{Data from simple_clusterprofiler().}

\item{goid_map}{Mapping file of IDs to GO ontologies.}

\item{go_db}{Dataframe of mappings used to build goid_map.}

\item{score_limit}{Scoring limit above which to ignore genes.}

\item{overwrite}{Overwrite an existing goid mapping file?}

\item{selector}{Name of a function for applying scores to the trees.}

\item{pval_column}{Name of the column in the GO table from which to extract scores.}
}
\value{
plots! Trees! oh my!
}
\description{
TopGO's ontology trees can be very illustrative.  This function shoe-horns
clusterProfiler data into the format expected by topGO and uses it to make
those trees.
}
\examples{
\dontrun{
 cluster_data <- simple_clusterprofiler(genes, stuff)
 ctrees <- cluster_trees(genes, cluster_data)
}
}
\seealso{
\pkg{Ramigo}
 \code{\link[topGO]{showSigOfNodes}}
}
