% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{cluster_trees}
\alias{cluster_trees}
\title{Make fun trees a la topgo from goseq data.}
\usage{
cluster_trees(cpdata, goid_map = "reference/go/id2go.map", goids_df = NULL,
  score_limit = 0.2, overwrite = FALSE, selector = "topDiffGenes",
  pval_column = "P.Value")
}
\arguments{
\item{godata}{data from cluster Profiler}

\item{goids}{a mapping of IDs to GO in the Ramigo expected format}

\item{sigforall}{Print significance on all nodes?}
}
\value{
plots! Trees! oh my!
}
\description{
Make fun trees a la topgo from goseq data.
}
\seealso{
\code{\link{Ramigo}}
}

