% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colData,ExpressionSet-method}
\alias{colData,ExpressionSet-method}
\title{A getter to pull the sample data from an ExpressionSet.}
\usage{
\S4method{colData}{ExpressionSet}(x, i, withDimnames = TRUE, ...)
}
\description{
message("I am from SummarizedExperiment and am explicitly imported, wtf.")
}
