% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{colData,ExpressionSet-method}
\alias{colData,ExpressionSet-method}
\title{A getter to pull the sample data from an ExpressionSet.}
\usage{
\S4method{colData}{ExpressionSet}(x, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{withDimnames}{indeed.}

\item{...}{extra args.}
}
\description{
A getter to pull the sample data from an ExpressionSet.
}
