% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{colData<-,expt,ANY-method}
\alias{colData<-,expt,ANY-method}
\title{A setter to put the sample data into an expt.}
\usage{
\S4method{colData}{expt,ANY}(x, i, withDimnames = TRUE, ...) <- value
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{i}{Subset to replace.}

\item{withDimnames}{indeed.}

\item{...}{extra args.}

\item{value}{New Sample data for the expt.}
}
\description{
A setter to put the sample data into an expt.
}
