% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colData<-,SummarizedExperiment,data.frame-method}
\alias{colData<-,SummarizedExperiment,data.frame-method}
\title{A setter to put the sample data into an ExpressionSet.}
\usage{
\S4method{colData}{SummarizedExperiment,data.frame}(x, ...) <- value
}
\arguments{
\item{x}{One of my various expressionset analogs, expt,
expressionSet, or summarizedExperiment.}

\item{...}{args for the arglist}

\item{value}{New values for the expressionset.}
}
\description{
A setter to put the sample data into an ExpressionSet.
}
