% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colData<-}
\alias{colData<-}
\title{If you mess up the NAMESPACE file, the following becomes necessary}
\usage{
colData(x, ...) <- value
}
\arguments{
\item{x}{Data structure from which to acquire column information (sample metadata).}

\item{...}{Other options.}

\item{value}{New column data to replace the existing values.}
}
\description{
message("I am from SummarizedExperiment and am explicitly imported, wtf.")
}
