% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{colors<-,expt-method}
\alias{colors<-,expt-method}
\title{A setter to put the colors into an x.}
\usage{
\S4method{colors}{expt}(x, ...) <- value
}
\arguments{
\item{x}{An x.}

\item{...}{Extra args}

\item{value}{List of new colors.}
}
\description{
A setter to put the colors into an x.
}
