% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{colors<-,SummarizedExperiment-method}
\alias{colors<-,SummarizedExperiment-method}
\title{A setter to put the colors into a SummarizedExperiment.}
\usage{
\S4method{colors}{SummarizedExperiment}(expt, lst) <- value
}
\arguments{
\item{expt}{A SummarizedExperiment.}
}
\description{
A setter to put the colors into a SummarizedExperiment.
}
