% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R, R/zzz_methods.R
\name{colors<-}
\alias{colors<-}
\title{Set the colors to an expt.}
\usage{
colors(expt, ...) <- value

colors(expt, ...) <- value
}
\arguments{
\item{expt}{Expt to modify.}

\item{...}{Colors!}

\item{value}{List of colors.}
}
\value{
The expt with new colors associated with each sample.
}
\description{
Set the colors to an expt.

Set the colors for an expt.
}
