% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{colors<-}
\alias{colors<-}
\title{Set the colors to an expt.}
\usage{
colors(expt, ...) <- value
}
\arguments{
\item{expt}{Expt to which to add colors.}

\item{lst}{List of colors.}
}
\description{
Set the colors to an expt.
}
