% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{colors}
\alias{colors}
\title{Get the colors from an expt.}
\usage{
colors(expt)
}
\arguments{
\item{expt}{One of my slightly modified ExpressionSets.}
}
\value{
List with the methods used to modify the data (if any).
}
\description{
Get the colors from an expt.
}
