% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{colors}
\alias{colors}
\title{colors provided to hpgltools for plotting}
\format{a \code{data.frame} instance}
\source{
the hash package
}
\usage{
color_hash
}
\description{
A hash of color to sample names, this is
redundant and either this or color_hash should be removed
}
\author{
atb
}

