% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{combine_de_tables}
\alias{combine_de_tables}
\title{Combine portions of deseq/limma/edger table output.}
\usage{
combine_de_tables(all_pairwise_result, extra_annot = NULL, excel = NULL,
  excel_title = "Table SXXX: Combined Differential Expression of YYY",
  excel_sheet = "combined_DE", keepers = "all", include_basic = TRUE,
  add_plots = TRUE, plot_dim = 6)
}
\arguments{
\item{all_pairwise_result}{Output from all_pairwise().}

\item{extra_annot}{Add some annotation information?}

\item{excel}{Filename for the excel workbook, or null if not printed.}

\item{excel_title}{Title for the excel sheet(s).  If it has the
string 'YYY', that will be replaced by the contrast name.}

\item{excel_sheet}{Name the excel sheet.}

\item{keepers}{List of reformatted table names to explicitly keep
certain contrasts in specific orders and orientations.}

\item{include_basic}{Include my stupid basic logFC tables?}

\item{add_plots}{Add plots to the end of the sheets with expression values?}

\item{plot_dim}{Number of inches squared for the plot if added.}
}
\value{
Table combining limma/edger/deseq outputs.
}
\description{
This hopefully makes it easy to compare the outputs from
limma/DESeq2/EdgeR on a table-by-table basis.
}
\examples{
\dontrun{
pretty = combine_de_tables(big_result, table='t12_vs_t0')
}
}
\seealso{
\code{\link{all_pairwise}}
}

