% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{combine_de_tables}
\alias{combine_de_tables}
\title{Combine portions of deseq/limma/edger table output.}
\usage{
combine_de_tables(all_pairwise_result, extra_annot = NULL, csv = NULL,
  excel = NULL,
  excel_title = "Table SXXX: Combined Differential Expression of YYY",
  excel_sheet = "combined_DE", keepers = "all", excludes = NULL,
  adjp = TRUE, include_basic = TRUE, add_plots = TRUE, plot_dim = 6,
  compare_plots = TRUE)
}
\arguments{
\item{all_pairwise_result}{Output from all_pairwise().}

\item{extra_annot}{Add some annotation information?}

\item{csv}{On some computers (Edson!) printing to excel runs the machine oom for big data sets.}

\item{excel}{Filename for the excel workbook, or null if not printed.}

\item{excel_title}{Title for the excel sheet(s).  If it has the
string 'YYY', that will be replaced by the contrast name.}

\item{excel_sheet}{Name the excel sheet.}

\item{keepers}{List of reformatted table names to explicitly keep
certain contrasts in specific orders and orientations.}

\item{excludes}{List of columns and patterns to use for excluding genes.}

\item{include_basic}{Include my stupid basic logFC tables?}

\item{add_plots}{Add plots to the end of the sheets with expression values?}

\item{plot_dim}{Number of inches squared for the plot if added.}

\item{compare_plots}{In an attempt to save memory when printing to excel, make it possible to
exclude comparison plots in the summary sheet.}
}
\value{
Table combining limma/edger/deseq outputs.
}
\description{
This hopefully makes it easy to compare the outputs from
limma/DESeq2/EdgeR on a table-by-table basis.
}
\examples{
\dontrun{
pretty = combine_de_tables(big_result, table='t12_vs_t0')
pretty = combine_de_tables(big_result, table='t12_vs_t0', keepers=list("avsb" = c("a","b")))
pretty = combine_de_tables(big_result, table='t12_vs_t0', keepers=list("avsb" = c("a","b")),
                           excludes=list("description" = c("sno","rRNA")))
}
}
\seealso{
\code{\link{all_pairwise}}
}

