% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{combine_de_tables}
\alias{combine_de_tables}
\title{Combine portions of deseq/limma/edger table output.}
\usage{
combine_de_tables(
  apr,
  extra_annot = NULL,
  keepers = "all",
  excludes = NULL,
  adjp = TRUE,
  include_limma = TRUE,
  excel = NULL,
  include_deseq = TRUE,
  include_edger = TRUE,
  include_ebseq = TRUE,
  include_basic = TRUE,
  rownames = TRUE,
  add_plots = TRUE,
  loess = FALSE,
  plot_dim = 6,
  compare_plots = TRUE,
  padj_type = "ihw",
  fancy = FALSE,
  lfc_cutoff = 1,
  p_cutoff = 0.05,
  de_types = c("limma", "deseq", "edger"),
  excel_title = "Table SXXX: Combined Differential Expression of YYY",
  rda = NULL,
  start_worksheet = "S02",
  format_sig = 4,
  ...
)
}
\arguments{
\item{apr}{Output from all_pairwise().}

\item{extra_annot}{Add some annotation information?}

\item{keepers}{List of reformatted table names to explicitly keep
certain contrasts in specific orders and orientations.}

\item{excludes}{List of columns and patterns to use for excluding genes.}

\item{adjp}{Perhaps you do not want the adjusted p-values for plotting?}

\item{include_limma}{Include limma analyses in the table?}

\item{excel}{Filename for the excel workbook, or null if not printed.}

\item{include_deseq}{Include deseq analyses in the table?}

\item{include_edger}{Include edger analyses in the table?}

\item{include_ebseq}{Include ebseq analyses in the table?}

\item{include_basic}{Include my stupid basic logFC tables?}

\item{rownames}{Add rownames to the xlsx printed table?}

\item{add_plots}{Add plots to the end of the sheets with expression values?}

\item{loess}{Add time intensive loess estimation to plots?}

\item{plot_dim}{Number of inches squared for the plot if added.}

\item{compare_plots}{Add some plots comparing the results.}

\item{padj_type}{Add a consistent p adjustment of this type.}

\item{fancy}{Save a set of fancy plots along with the xlsx file?}

\item{lfc_cutoff}{In this context, only used for plotting volcano/MA plots.}

\item{p_cutoff}{In this context, used for volcano/MA plots.}

\item{de_types}{Used for plotting pvalue/logFC cutoffs.}

\item{excel_title}{Title for the excel sheet(s).  If it has the
string 'YYY', that will be replaced by the contrast name.}

\item{rda}{Write a rda file of the results.}

\item{start_worksheet}{This will now increment worksheet titles from this point forward.}

\item{...}{Arguments passed to significance and abundance tables.}
}
\value{
Table combining limma/edger/deseq outputs.
}
\description{
This hopefully makes it easy to compare the outputs from
limma/DESeq2/EdgeR on a table-by-table basis.
}
\examples{
\dontrun{
 expt <- create_expt(metadata="some_metadata.xlsx", gene_info=funkytown)
 big_result <- all_pairwise(expt, model_batch=FALSE)
 pretty <- combine_de_tables(big_result, table='t12_vs_t0')
 pretty <- combine_de_tables(big_result, table='t12_vs_t0',
                             keepers=list("avsb"=c("a","b")))
 pretty <- combine_de_tables(big_result, table='t12_vs_t0',
                             keepers=list("avsb"=c("a","b")),
                             excludes=list("description"=c("sno","rRNA")))
}
}
\seealso{
[all_pairwise()] [extract_significant_genes()]
}
