% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.r
\name{combine_de_tables}
\alias{combine_de_tables}
\title{combine_de_tables()  Combine portions of deseq/limma/edger table output}
\usage{
combine_de_tables(all_pairwise_result, annot_df = NULL, excel = NULL,
  excel_title = "Table SXXX: Combined Differential Expression of YYY",
  excel_sheet = "combined_DE", keepers = NULL)
}
\arguments{
\item{all_pairwise_result}{the output from all_pairwise()}

\item{keepers}{default NULL  a list of reformatted table names to explicitly keep certain contrasts in specific orders}

\item{table}{default='wt_vs_mut'  the name of a table comparison performed by deseq/limma/edger.}
}
\value{
a table combinine limma/edger/deseq outputs.
}
\description{
This hopefully makes it easy to compare the outputs from limma/DESeq2/EdgeR on a table-by-table basis.
}
\examples{
## pretty = combine_de_tables(big_result, table='t12_vs_t0')
}
\seealso{
\code{\link{all_pairwise}}
}

