% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{combine_expts}
\alias{combine_expts}
\title{Take two expressionsets and smoosh them together.}
\usage{
combine_expts(expt1, expt2, condition = "condition", batch = "batch")
}
\arguments{
\item{expt1}{First expt object.}

\item{expt2}{Second expt object.}

\item{condition}{Column with which to reset the conditions.}

\item{batch}{Column with which to reset the batches.}
}
\value{
Larger expt.
}
\description{
Because of the extra sugar I added to expressionSets, the combine() function
needs a little help when combining expts.  Notably, the information from
tximport needs some help.
}
