% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{combine_expts}
\alias{combine_expts}
\title{Take two expressionsets and smoosh them together.}
\usage{
combine_expts(
  expt1,
  expt2,
  condition = "condition",
  all_x = TRUE,
  all_y = TRUE,
  batch = "batch",
  merge_meta = TRUE
)
}
\arguments{
\item{expt1}{First expt object.}

\item{expt2}{Second expt object.}

\item{condition}{Column with which to reset the conditions.}

\item{all_x}{Keep all of the first expt's annotations/counts if there are mismatches?}

\item{all_y}{Keep all the second expt's annotations/counts if there are mismatches?}

\item{batch}{Column with which to reset the batches.}

\item{merge_meta}{Merge the metadata when they mismatch?  This should perhaps default to TRUE.}
}
\value{
Larger expt.
}
\description{
Because of the extra sugar I added to expressionSets, the combine() function
needs a little help when combining expts.  Notably, the information from
tximport needs some help.
}
\examples{
 \dontrun{
  ## I am trying to get rid of all my dontrun sections, but I don't have two
  ## expressionsets to combine.
  expt1 <- create_expt(first_meta)
  expt2 <- create_expt(second_meta)
  combined <- combine_expts(expt1, expt2, merge_meta = TRUE)
}
}
\seealso{
[set_expt_batches()] [set_expt_conditions()] [set_expt_colors()]
 [set_expt_genenames()] [set_expt_samplenames()] [subset_expt()] [create_expt()]
}
