% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{combine_extracted_plots}
\alias{combine_extracted_plots}
\title{Gather data required to make MA/Volcano plots for pairwise comparisons.}
\usage{
combine_extracted_plots(
  name,
  combined,
  denominator,
  numerator,
  plot_inputs,
  plot_basic = TRUE,
  plot_deseq = TRUE,
  plot_edger = TRUE,
  plot_limma = TRUE,
  plot_ebseq = FALSE,
  loess = FALSE,
  logfc = 1,
  pval = 0.05,
  found_table = NULL,
  p_type = "all",
  plot_colors = NULL,
  fancy = FALSE,
  do_inverse = FALSE,
  invert_colors = FALSE,
  z = 1.5,
  alpha = 0.4,
  z_lines = FALSE,
  label = 10,
  label_column = "hgncsymbol"
)
}
\arguments{
\item{name}{Name of the table to plot.}

\item{combined}{Modified pairwise result, containing the various DE methods.}

\item{denominator}{Name of the denominator coefficient.}

\item{numerator}{Name of the numerator coefficient.}

\item{plot_inputs}{The individual outputs from limma etc.}

\item{loess}{Add a loess estimation?}

\item{logfc}{For Volcano/MA plot lines.}

\item{found_table}{The table name actually used.}

\item{p_type}{Use this/these methods' p-value for determining significance.}

\item{do_inverse}{Flip the numerator/denominator?}

\item{include_basic}{Add basic data?}

\item{include_deseq}{Add deseq data?}

\item{include_edger}{Add edger data?}

\item{include_limma}{Add limma data?}

\item{include_ebseq}{Add ebseq data?}

\item{p}{For Volcano/MA plot lines.}
}
\description{
It should be possible to significantly simplify the arguments passed to this
function, but I have thus far focused only on getting it to work with the
newly split-apart combine_de_tables() functions.
}
