% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{combine_extracted_plots}
\alias{combine_extracted_plots}
\title{Gather data required to make MA/Volcano plots for pairwise comparisons.}
\usage{
combine_extracted_plots(
  name,
  combined,
  denominator,
  numerator,
  plot_inputs,
  include_basic = TRUE,
  include_deseq = TRUE,
  include_edger = TRUE,
  include_limma = TRUE,
  include_ebseq = FALSE,
  loess = FALSE,
  logfc = 1,
  p = 0.05,
  do_inverse = FALSE,
  found_table = NULL
)
}
\arguments{
\item{name}{Name of the table to plot.}

\item{combined}{Modified pairwise result, containing the various DE methods.}

\item{denominator}{Name of the denominator coefficient.}

\item{numerator}{Name of the numerator coefficient.}

\item{plot_inputs}{The individual outputs from limma etc.}

\item{include_basic}{Add basic data?}

\item{include_deseq}{Add deseq data?}

\item{include_edger}{Add edger data?}

\item{include_limma}{Add limma data?}

\item{include_ebseq}{Add ebseq data?}

\item{loess}{Add a loess estimation?}

\item{logfc}{For Volcano/MA plot lines.}

\item{p}{For Volcano/MA plot lines.}

\item{do_inverse}{Flip the numerator/denominator?}

\item{found_table}{The table name actually used.}
}
\description{
It should be possible to significantly simplify the arguments passed to this
function, but I have thus far focused only on getting it to work with the
newly split-apart combine_de_tables() functions.
}
