% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{combine_extracted_plots}
\alias{combine_extracted_plots}
\title{Gather data required to make MA/Volcano plots for pairwise comparisons.}
\usage{
combine_extracted_plots(
  name,
  combined,
  denominator,
  numerator,
  plot_inputs,
  plot_basic = TRUE,
  plot_deseq = TRUE,
  plot_edger = TRUE,
  plot_limma = TRUE,
  plot_ebseq = FALSE,
  plot_noiseq = FALSE,
  loess = FALSE,
  logfc = 1,
  pval = 0.05,
  found_table = NULL,
  p_type = "all",
  plot_colors = NULL,
  fancy = FALSE,
  adjp = TRUE,
  do_inverse = FALSE,
  invert_colors = FALSE,
  z = 1.5,
  alpha = 0.4,
  z_lines = FALSE,
  label = 10,
  label_column = "hgncsymbol"
)
}
\arguments{
\item{name}{Name of the table to plot.}

\item{combined}{Modified pairwise result, containing the various DE methods.}

\item{denominator}{Name of the denominator coefficient.}

\item{numerator}{Name of the numerator coefficient.}

\item{plot_inputs}{The individual outputs from limma etc.}

\item{plot_basic}{Add basic data?}

\item{plot_deseq}{Add deseq data?}

\item{plot_edger}{Add edger data?}

\item{plot_limma}{Add limma data?}

\item{plot_ebseq}{Add ebseq data?}

\item{loess}{Add a loess estimation?}

\item{logfc}{For Volcano/MA plot lines.}

\item{pval}{For Volcano/MA plot lines.}

\item{found_table}{The table name actually used.}

\item{p_type}{Use this/these methods' p-value for determining
significance.}

\item{plot_colors}{Use these colors for numerators/denominators.}

\item{fancy}{Include fancy pdf/svg versions of plots for publication?}

\item{do_inverse}{Flip the numerator/denominator?}

\item{invert_colors}{Conversely, keep the values the same, but flip
the colors.  I think these invert parameters are not needed anymore.}

\item{z}{Use a z-score cutoff for coefficient plots.}

\item{alpha}{Add some transparency to the plots.}

\item{z_lines}{Add lines for zscore cutoffs?}

\item{label}{Label this number of the top genes.}

\item{label_column}{Label the top genes with this column.}
}
\description{
It should be possible to significantly simplify the arguments passed to this
function, but I have thus far focused only on getting it to work with the
newly split-apart combine_de_tables() functions.
}
