% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{compare_go_searches}
\alias{compare_go_searches}
\title{compare_go_searches()  Compare the results from different ontology tools}
\usage{
compare_go_searches(goseq = NULL, cluster = NULL, topgo = NULL,
  gostats = NULL)
}
\arguments{
\item{goseq}{The goseq result from simple_goseq()}

\item{cluster}{The result from simple_clusterprofiler()}

\item{topgo}{Guess}

\item{gostats}{Yep, ditto}

\item{go_file}{default='excel/merged_go'  the file to save the results.}

\item{n}{default=30  the number of ontology categories to include in each table.}
}
\description{
Combine the results from goseq, cluster profiler, topgo, and gostats; poke at them
with a stick and see what happens.
The general idea is to pull the p-value data from each tool and contrast that to the
set of all possibile ontologies.  This allows one to do a correlation coefficient
between them.  In addition, take the 1-pvalue for each ontology for each tool.
Thus for strong p-values the score will be near 1 and so we can sum the scores
for all the tools.  Since topgo has 4 tools, the total possible is 7 if everything
has a p-value equal to 0.
}

