% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{compare_logfc_plots}
\alias{compare_logfc_plots}
\title{Compare logFC values from limma and friends}
\usage{
compare_logfc_plots(combined_tables)
}
\arguments{
\item{combined_tables}{The combined tables from limma et al.}
}
\value{
Some plots
}
\description{
There are some peculiar discrepencies among these tools, what is up with that?
}
\details{
Invoked by combine_de_tables() in order to compare the results.
}
\seealso{
\code{\link{plot_linear_scatter}}
}
