% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tables.r
\name{concatenate_runs}
\alias{concatenate_runs}
\title{Sum the reads/gene for multiple sequencing runs of a single condition/batch.}
\usage{
concatenate_runs(expt, column = "replicate")
}
\arguments{
\item{expt}{Experiment class containing the requisite metadata and count tables.}

\item{column}{Column of the design matrix used to specify which samples are replicates.}
}
\value{
Expt with the concatenated counts, new design matrix, batches, conditions, etc.
}
\description{
On occasion we have multiple technical replicates of a sequencing run.  This can use a column in
the experimental design to identify those replicates and sum the counts into a single column in
the count tables.
}
\examples{
\dontrun{
 compressed = concatenate_runs(expt)
}
}
\seealso{
\pkg{Biobase}
}

