% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{conditions<-,expt-method}
\alias{conditions<-,expt-method}
\title{Add experimental conditions to an expt.}
\usage{
\S4method{conditions}{expt}(exp, ...) <- value
}
\arguments{
\item{...}{extra args}

\item{object}{Output expt}

\item{value}{vector of new conditions}
}
\description{
Add experimental conditions to an expt.
}
