% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{conditions<-,expt-method}
\alias{conditions<-,expt-method}
\title{Add experimental conditions to an expt.}
\usage{
\S4method{conditions}{expt}(exp, ...) <- value
}
\arguments{
\item{exp}{Output expt}

\item{...}{extra args}

\item{value}{The new value (currently in ... I think)}
}
\description{
Add experimental conditions to an expt.
}
