% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_convert.R
\name{convert_counts,SummarizedExperiment-method}
\alias{convert_counts,SummarizedExperiment-method}
\title{Convert a summarized experiment.}
\usage{
\S4method{convert_counts}{SummarizedExperiment}(
  count_table,
  method = "raw",
  annotations = NULL,
  length_column = "width",
  na_to_zero = TRUE,
  ...
)
}
\description{
Convert a summarized experiment.
}
