% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{convert_counts}
\alias{convert_counts}
\title{convert_counts(): Perform a cpm/rpkm/whatever transformation of a count table.}
\usage{
convert_counts(count_table, convert = "raw", annotations = NULL, ...)
}
\arguments{
\item{count_table}{A matrix of count data}

\item{convert='raw'}{A type of conversion to perform: edgecpm/cpm/rpkm/cp_seq_m}

\item{annotations=NULL}{a set of gff annotations are needed if using rpkm so we can get gene lengths.
I should probably tell it to also handle a simple df/vector/list of gene lengths, but I haven't.

cp_seq_m is a cpm conversion of the data followed by a rp-ish
conversion which normalizes by the number of the given oligo.  By
default this oligo is 'TA' because it was used for tnseq which
should be normalized by the number of possible transposition sites
by mariner.  It could, however, be used to normalize by the number
of methionines, for example -- if one wanted to do such a thing.}
}
\value{
dataframe of cpm/rpkm/whatever(counts)
}
\description{
convert_counts(): Perform a cpm/rpkm/whatever transformation of a count table.
}
\examples{
## converted_table = convert_counts(count_table, convert='edgecpm')
}

