% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{convert_ids}
\alias{convert_ids}
\title{Change gene IDs to the format expected by gsva using an orgdb.}
\usage{
convert_ids(ids, from = "ENSEMBL", to = "ENTREZID", orgdb = "org.Hs.eg.db")
}
\arguments{
\item{ids}{Vector of IDS to modify.}

\item{from}{Change from this format.}

\item{to}{Change to this format.}

\item{orgdb}{Using this orgdb instance.}
}
\value{
New vector of ENTREZ IDs.
}
\description{
Though it is possible to use gsva without ENTREZ IDs, it is not trivial.
This function attempts to ensure that the IDs in one's expressionset are
therefore entrez IDs. It is possible that this function is at least partially
redundant with other functions in this package and should be replaced.
}
\seealso{
[AnnotationDbi]
}
