% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell.R
\name{count_clonotype_by_cluster}
\alias{count_clonotype_by_cluster}
\title{Add a df of clonotype observations by cell cluster to the @misc slot of the Seurat data.}
\usage{
count_clonotype_by_cluster(
  scd,
  column = "res0p2_clusters",
  clono_column = "raw_clonotype_id",
  add_sum = TRUE
)
}
\description{
Add a df of clonotype observations by cell cluster to the @misc slot of the Seurat data.
}
