% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{count_expt_snps}
\alias{count_expt_snps}
\title{Gather snp information for an expt}
\usage{
count_expt_snps(expt, type = "counts", input_dir = "preprocessing/outputs",
  tolower = TRUE)
}
\arguments{
\item{expt}{an expressionset from which to extract information.}

\item{input_dir}{Directory to scan for snps output files.}

\item{tolower}{Lowercase stuff like 'HPGL'?}

\item{file_suffix}{What to add on the end of the files for the resulting output.}

\item{bam_suffix}{How do we find the bam files?}
}
\value{
A new expt object
}
\description{
I have some initial code for working with snps, but it seems that it will be getting more use, so
make it testable etc.
}
