% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{create_combined_table}
\alias{create_combined_table}
\title{Given a limma, edger, and deseq table, combine them}
\usage{
create_combined_table(li, ed, de, ba, table_name, annot_df = NULL,
  inverse = FALSE, include_basic = TRUE, fc_cutoff = 1, p_cutoff = 0.05)
}
\arguments{
\item{li}{a limma output}

\item{ed}{a edger output}

\item{de}{a deseq output}

\item{ba}{a basic output}

\item{table_name}{name of the table to merge}

\item{annot_df}{add some annotation information}

\item{inverse}{invert the fold changes}

\item{include_basic}{include the basic table?}

\item{fc_cutoff}{logfoldchange cutoff}

\item{p_cutoff}{a pvalue cutoff}
}
\description{
Given a limma, edger, and deseq table, combine them
}

