% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{create_combined_table}
\alias{create_combined_table}
\title{Given a limma, edger, and deseq table, combine them into one.}
\usage{
create_combined_table(li, ed, de, ba, table_name, annot_df = NULL,
  inverse = FALSE, adjp = TRUE, include_basic = TRUE, fc_cutoff = 1,
  p_cutoff = 0.05, excludes = NULL)
}
\arguments{
\item{li}{Limma output table.}

\item{ed}{Edger output table.}

\item{de}{Deseq2 output table.}

\item{ba}{Basic output table.}

\item{table_name}{Name of the table to merge.}

\item{annot_df}{Add some annotation information?}

\item{inverse}{Invert the fold changes?}

\item{include_basic}{Include the basic table?}

\item{fc_cutoff}{Preferred logfoldchange cutoff.}

\item{p_cutoff}{Preferred pvalue cutoff.}
}
\value{
List containing a) Dataframe containing the merged
    limma/edger/deseq/basic tables, and b) A summary of how many
    genes were observed as up/down by output table.
}
\description{
This combines the outputs from the various differential expression
tools and formalizes some column names to make them a little more
consistent.
}

