% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_tables.R
\name{create_experiment}
\alias{create_experiment}
\title{Wrap bioconductor's expressionset to include some other extraneous
information.}
\usage{
create_experiment(file = NULL, color_hash, suffix = ".count.gz",
  header = FALSE, genes = NULL, by_type = FALSE, by_sample = FALSE,
  include_type = "all", include_gff = NULL, count_dataframe = NULL,
  meta_dataframe = NULL, sep = ",", ...)
}
\arguments{
\item{file}{a comma separated file describing the samples with
information like condition,batch,count_filename,etc}

\item{color_hash}{a hash which describes how to color the samples}
}
\value{
experiment an expressionset
}
\description{
Wrap bioconductor's expressionset to include some other extraneous
information.
}
\examples{
## new_experiment = create_experiment("some_csv_file.csv", color_hash)
## Dude, you need to remember that this depends on an existing data structure of
## gene annotations.
}
\seealso{
\code{\link{pData}}, \code{\link{fData}},
\code{\link{exprs}}, \code{\link{hpgl_read_files}},
\code{\link{as.list.hash}}
}

