% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_tables.R
\name{create_experiment}
\alias{create_experiment}
\title{create_experiment()  Wrap bioconductor's expressionset to include some other extraneous
information.}
\usage{
create_experiment(file = NULL, color_hash, suffix = ".count.gz",
  header = FALSE, gene_info = NULL, by_type = FALSE, by_sample = FALSE,
  include_type = "all", include_gff = NULL, count_dataframe = NULL,
  meta_dataframe = NULL, sep = ",", ...)
}
\arguments{
\item{file}{default=NULL  a comma separated file describing the samples with
information like condition,batch,count_filename,etc.}

\item{color_hash}{a hash which describes how to color the samples}

\item{suffix}{default='.count.gz'  when looking for the count tables in processed_data
look for this suffix on the end of the files.}

\item{header}{default=FALSE  Does the csv metadata file have a header?}

\item{by_type}{default=FALSE  when looking for count tables, are they organized by type?}

\item{by_sample}{default=FALSE  or by sample?  I do all mine by sample, but others do by type...}

\item{include_type}{default='all'  I have usually assumed that all gff annotations should be used,
but that is not always true, this allows one to limit.}

\item{include_gff}{default=NULL  A gff file to help in sorting which features to keep}

\item{count_dataframe}{default=NULL  If one does not wish to read the count tables from processed_data/
they may instead be fed here}

\item{meta_dataframe}{default=NULL  an optional dataframe containing the metadata rather than a file}

\item{sep}{default=','  some people prefer their csv files as tab or semicolon separated.}

\item{genes}{default=NULL  annotation information describing the rows of the data set, usually
this comes from a call to import.gff()}
}
\value{
experiment an expressionset
}
\description{
create_experiment()  Wrap bioconductor's expressionset to include some other extraneous
information.
}
\examples{
## new_experiment = create_experiment("some_csv_file.csv", color_hash)
}
\seealso{
\code{\link{pData}}, \code{\link{fData}},
\code{\link{exprs}}, \code{\link{hpgl_read_files}},
\code{\link{as.list.hash}}
}

