% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count_tables.R
\name{create_expt}
\alias{create_expt}
\title{Wrap bioconductor's expressionset to include some other extraneous
information.  This simply calls create_experiment and then does
expt_subset for everything}
\usage{
create_expt(file, color_hash = NULL, suffix = ".count.gz", header = FALSE,
  genes = NULL, by_type = FALSE, by_sample = FALSE, sep = ",",
  include_type = "all", include_gff = NULL, count_dataframe = NULL,
  savefile = "expt", low_files = FALSE, ...)
}
\arguments{
\item{file}{a comma separated file describing the samples with
information like condition,batch,count_filename,etc}

\item{color_hash}{a hash which describes how to color the samples}
}
\value{
experiment an expressionset
}
\description{
Wrap bioconductor's expressionset to include some other extraneous
information.  This simply calls create_experiment and then does
expt_subset for everything
}
\examples{
## new_experiment = create_experiment("some_csv_file.csv", color_hash)
## Dude, you need to remember that this depends on an existing data structure of
## gene annotations.
}
\seealso{
\code{\link{pData}}, \code{\link{fData}},
\code{\link{exprs}}, \code{\link{hpgl_read_files}},
\code{\link{as.list.hash}}
}

