% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_tables.R
\name{create_expt}
\alias{create_expt}
\title{create_expt()  Wrap bioconductor's expressionset to include some other extraneous
information.  This simply calls create_experiment and then does
expt_subset for everything}
\usage{
create_expt(file = NULL, color_hash = NULL, suffix = ".count.gz",
  header = FALSE, gene_info = NULL, by_type = FALSE, by_sample = FALSE,
  sep = ",", include_type = "all", include_gff = NULL,
  count_dataframe = NULL, meta_dataframe = NULL, savefile = "expt",
  low_files = FALSE, ...)
}
\arguments{
\item{file}{a comma separated file describing the samples with
information like condition,batch,count_filename,etc}

\item{color_hash}{default=NULL  a hash which describes how to color the samples,
it will generate its own colors using colorBrewer}

\item{suffix}{default='.count.gz'  when looking for the count tables in processed_data
look for this suffix on the end of the files.}

\item{header}{default=FALSE  Does the csv metadata file have a header?}

\item{by_type}{default=FALSE  when looking for count tables, are they organized by type?}

\item{by_sample}{default=FALSE  or by sample?  I do all mine by sample, but others do by type...}

\item{sep}{default=','  some people prefer their csv files as tab or semicolon separated.}

\item{include_type}{default='all'  I have usually assumed that all gff annotations should be used,
but that is not always true, this allows one to limit.}

\item{include_gff}{default=NULL  A gff file to help in sorting which features to keep}

\item{count_dataframe}{default=NULL  If one does not wish to read the count tables from processed_data/
they may instead be fed here}

\item{meta_dataframe}{default=NULL  an optional dataframe containing the metadata rather than a file}

\item{savefile}{default='expt'  an Rdata filename prefix for saving the data of the resulting expt.}

\item{low_files}{default=FALSE  whether or not to explicitly lowercase the filenames when searching in processed_data/
this is relevant because the ceph object storage by default lowercases filenames.

It is worth noting that this function has a lot of logic used to
find the count tables in the local filesystem.  This logic has been
superceded by simply adding a field to the .csv file called
'file'.  create_expt() will then just read that filename, it may be
a full pathname or local to the cwd of the project.}

\item{genes}{default=NULL  annotation information describing the rows of the data set, usually
this comes from a call to import.gff()}
}
\value{
experiment an expressionset
}
\description{
create_expt()  Wrap bioconductor's expressionset to include some other extraneous
information.  This simply calls create_experiment and then does
expt_subset for everything
}
\examples{
## new_experiment = create_experiment("some_csv_file.csv", color_hash)
## Remember that this depends on an existing data structure of gene annotations.
}
\seealso{
\code{\link{pData}}, \code{\link{fData}},
\code{\link{exprs}}, \code{\link{hpgl_read_files}},
\code{\link{as.list.hash}}
}

