% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{create_expt}
\alias{create_expt}
\title{Wrap bioconductor's expressionset to include some other extraneous
information.}
\usage{
create_expt(metadata = NULL, gene_info = NULL,
  count_dataframe = NULL, sample_colors = NULL, title = NULL,
  notes = NULL, include_type = "all", include_gff = NULL,
  file_column = "file", savefile = "expt", low_files = FALSE, ...)
}
\arguments{
\item{metadata}{Comma separated file (or excel) describing the samples with
information like condition, batch, count_filename, etc.}

\item{gene_info}{Annotation information describing the rows of the data set,
this often comes from a call to import.gff() or biomart or organismdbi.}

\item{count_dataframe}{If one does not wish to read the count tables from the
filesystem, they may instead be fed as a data frame here.}

\item{sample_colors}{List of colors by condition, if not provided it will
generate its own colors using colorBrewer.}

\item{title}{Provide a title for the expt?}

\item{notes}{Additional notes?}

\item{include_type}{I have usually assumed that all gff annotations should be
used, but that is not always true, this allows one to limit to a specific
annotation type.}

\item{include_gff}{Gff file to help in sorting which features to keep.}

\item{file_column}{Column to use in a gene information dataframe for}

\item{savefile}{Rdata filename prefix for saving the data of the resulting
expt.}

\item{low_files}{Explicitly lowercase the filenames when searching the
filesystem?}

\item{...}{More parameters are fun!}
}
\value{
experiment an expressionset
}
\description{
It is worth noting that this function has a lot of logic used to
find the count tables in the local filesystem.  This logic has been
superceded by simply adding a field to the .csv file called
'file'.  create_expt() will then just read that filename, it may be
a full pathname or local to the cwd of the project.
}
\examples{
\dontrun{
 new_experiment <- create_expt("some_csv_file.csv", gene_info=gene_df)
 ## Remember that this depends on an existing data structure of gene annotations.
}
}
\seealso{
\pkg{Biobase}
 \code{\link[Biobase]{pData}} \code{\link[Biobase]{fData}}
  \code{\link[Biobase]{exprs}} \code{\link{read_counts_expt}}
}
