% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.r
\name{de_venn}
\alias{de_venn}
\title{Create venn diagrams describing how well deseq/limma/edger agree.}
\usage{
de_venn(table, adjp = FALSE, p = 0.05, lfc = 0, ...)
}
\arguments{
\item{table}{Which table to query?}

\item{adjp}{Use adjusted p-values}

\item{p}{p-value cutoff, I forget what for right now.}

\item{lfc}{What fold-change cutoff to include?}

\item{...}{More arguments are passed to arglist.}
}
\value{
A list of venn plots
}
\description{
The sets of genes provided by limma and friends would ideally always agree,
but they do not. Use this to see out how much the (dis)agree.
}
\examples{
\dontrun{
 bunchovenns <- de_venn(pairwise_result)
}
}
\seealso{
\pkg{venneuler} \pkg{Vennerable}
}
