% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/de_deseq.r
\name{deseq2_pairwise}
\alias{deseq2_pairwise}
\title{deseq2_pairwise()  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions using DESeq2.}
\usage{
deseq2_pairwise(input, conditions = NULL, batches = NULL,
  model_cond = TRUE, model_batch = FALSE, annot_df = NULL, ...)
}
\arguments{
\item{input}{A dataframe/vector or expt class containing data, normalization state, etc.}

\item{conditions}{default=NULL  A factor of conditions in the experiment}

\item{batches}{default=NULL A factor of batches in the experiment}

\item{model_cond}{default=TRUE  Have condition in the experimental model?}

\item{model_batch}{default=FALSE  Have batch in the experimental model?}

\item{annot_df}{default=NULL  Include some annotation information in the results?}

\item{...}{triple dots!}
}
\value{
A list including the following information:
  run = the return from calling DESeq()
  denominators = list of denominators in the contrasts
  numerators = list of the numerators in the contrasts
  conditions = the list of conditions in the experiment
  coefficients = list of coefficients making the contrasts
  all_tables = list of DE tables
}
\description{
deseq2_pairwise()  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions using DESeq2.
}
\examples{
\dontrun{
pretend = deseq2_pairwise(data, conditions, batches)
}
}
\seealso{
\pkg{DESeq2} \code{\link[DESeq2]{results}} \code{\link[DESeq2]{estimateSizeFactors}}
         \code{\link[DESeq2]{estimateDispersions}} \code{\link[DESeq2]{nbinomWaldTest}}
}

