% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_deseq.r
\name{deseq2_pairwise}
\alias{deseq2_pairwise}
\title{Set up model matrices contrasts and do pairwise comparisons of all conditions using DESeq2.}
\usage{
deseq2_pairwise(input, conditions = NULL, batches = NULL,
  model_cond = TRUE, alt_model = NULL, extra_contrasts = NULL,
  model_batch = TRUE, annot_df = NULL, force = FALSE, ...)
}
\arguments{
\item{input}{Dataframe/vector or expt class containing data, normalization state, etc.}

\item{conditions}{Factor of conditions in the experiment.}

\item{batches}{Factor of batches in the experiment.}

\item{model_cond}{Is condition in the experimental model?}

\item{alt_model}{Provide an arbitrary model here.}

\item{extra_contrasts}{Provide extra contrasts here.}

\item{model_batch}{Is batch in the experimental model?}

\item{annot_df}{Include some annotation information in the results?}

\item{force}{Force deseq to accept data which likely violates its assumptions.}

\item{...}{triple dots!  Options are passed to arglist.}
}
\value{
List including the following information:
  run = the return from calling DESeq()
  denominators = list of denominators in the contrasts
  numerators = list of the numerators in the contrasts
  conditions = the list of conditions in the experiment
  coefficients = list of coefficients making the contrasts
  all_tables = list of DE tables
}
\description{
Invoking DESeq2 is confusing, this should help.
}
\examples{
\dontrun{
pretend = deseq2_pairwise(data, conditions, batches)
}
}
\seealso{
\pkg{DESeq2} \code{\link[DESeq2]{results}} \code{\link[DESeq2]{estimateSizeFactors}}
         \code{\link[DESeq2]{estimateDispersions}} \code{\link[DESeq2]{nbinomWaldTest}}
}

