% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{deseq2_pairwise}
\alias{deseq2_pairwise}
\title{deseq2_pairwise()  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions using DESeq2.}
\usage{
deseq2_pairwise(input, conditions = NULL, batches = NULL, excel = FALSE,
  csv = TRUE, annot_df = NULL, workbook = "excel/deseq.xls", ...)
}
\arguments{
\item{input}{a dataframe/vector or  expt class containing data, normalization state, etc.}

\item{conditions}{default=NULL  a factor of conditions in the experiment}

\item{batches}{default=NULL a factor of batches in the experiment}
}
\value{
A list including the following information:
  run = the return from calling DESeq()
  result = the return from calling results()
  result_mle = the return from calling results() with the MLE parameter.
  results = the return when a deseq result class is cast as a dataframe.
}
\description{
deseq2_pairwise()  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions using DESeq2.
}
\examples{
## pretend = deseq2_pairwise(data, conditions, batches)
}
\seealso{
\code{\link{topTags}} \code{\link{glmLRT}} \code{\link{makeContrasts}}
}

