% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_deseq.R
\name{deseq_lrt}
\alias{deseq_lrt}
\title{Bring together some of the likelihood ratio test analyses.}
\usage{
deseq_lrt(
  expt,
  interactor_column = "visitnumber",
  interest_column = "clinicaloutcome",
  transform = "rlog",
  factors = NULL,
  cutoff = 0.05,
  minc = 3,
  interaction = TRUE
)
}
\arguments{
\item{expt}{Input expressionset}

\item{interactor_column}{Potentially interacting metadata}

\item{interest_column}{Essentially the condition in other analyses.}

\item{transform}{DESeq2 transformation applied (vst or rlog).}

\item{factors}{Other factors of interest}

\item{cutoff}{Significance cutoff}

\item{minc}{Minimum number of elements for a group}

\item{interaction}{Use an interaction model?}
}
\description{
This function hopes to wrap up some of the ideas/methods for LRT.
}
