% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_deseq.r
\name{deseq_ma}
\alias{deseq_ma}
\title{Make a MA plot of some deseq output with pretty colors and shapes}
\usage{
deseq_ma(output, table = NULL, p_col = "qvalue", fc_col = "logFC",
  expr_col = "logExpr", fc = 1, pval_cutoff = 0.05)
}
\arguments{
\item{output}{The result from all_pairwise(), which should be changed to handle other invocations too.}

\item{table}{Result from deseq to use, left alone it chooses the first.}

\item{p_col}{Column to use for p-value data.}

\item{fc_col}{Column for logFC data.}

\item{expr_col}{Column for the average data.}

\item{fc}{Fold-change cutoff for significant.}

\item{pval_cutoff}{p-value cutoff ibid.}
}
\value{
a plot!
}
\description{
Yay pretty colors and shapes!
}
\examples{
 \dontrun{
  prettyplot <- limma_ma(all_aprwise) ## [sic, I'm witty! and can speel]
}
}
\seealso{
\link{plot_ma_de}
}

