% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{deseq_pairwise}
\alias{deseq_pairwise}
\title{deseq_pairwise()  Because I can't be trusted to remember '2'}
\usage{
deseq_pairwise(...)
}
\arguments{
\item{look}{at deseq2_pairwise}
}
\value{
stuff from deseq2_pairwise
}
\description{
This calls deseq2_pairwise(...) because I am determined to forget typing deseq2
}

