% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{disjunct_pvalues}
\alias{disjunct_pvalues}
\title{Test for infected/control/beads -- a placebo effect?}
\usage{
disjunct_pvalues(contrast_fit, cellmeans_fit, conj_contrasts, disj_contrast)
}
\arguments{
\item{contrast_fit}{The result of lmFit.}

\item{cellmeans_fit}{The result of a cellmeans fit.}

\item{conj_contrasts}{The result from the makeContrasts of the first set.}

\item{disj_contrast}{The result of the makeContrasts of the second set.}
}
\description{
The goal is therefore to find responses different than beads
The null hypothesis is (H0): (infected == uninfected) || (infected == beads)
The alt hypothesis is (HA): (infected != uninfected) && (infected != beads)
}
