% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_csv_search}
\alias{dispatch_csv_search}
\title{Pull some information from a csv/tsv file.}
\usage{
dispatch_csv_search(
  meta,
  column,
  input_file_spec,
  type = "csv",
  which = "first",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{meta}{Input metadata}

\item{column}{Column to yank from}

\item{input_file_spec}{Input file specification to hunt down the
file of interest.}

\item{type}{csv or tsv?}

\item{which}{Take the first entry, or some subset.}

\item{verbose}{Print diagnostic information while running?}

\item{...}{Other arguments for glue.}
}
\description{
This function is a bit more generic than the others, but it grabs from a
column of a csv/tsv file.
}
