% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_fasta_lengths}
\alias{dispatch_fasta_lengths}
\title{Get the lengths of sequences from a fasta file.}
\usage{
dispatch_fasta_lengths(
  meta,
  input_file_spec,
  verbose = verbose,
  basedir = "preprocessing"
)
}
\arguments{
\item{meta}{Input metadata}

\item{input_file_spec}{Input file specification to hunt down the
file of interest.}

\item{verbose}{Print diagnostic information while running?}
}
\description{
Get the lengths of sequences from a fasta file.
}
