% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_filename_search}
\alias{dispatch_filename_search}
\title{Pull out the filename matching an input spec}
\usage{
dispatch_filename_search(
  meta,
  input_file_spec,
  verbose = verbose,
  species = "*",
  type = "genome",
  subtype = "gene",
  tag = "ID",
  basedir = "preprocessing",
  new_spec = NULL
)
}
\arguments{
\item{meta}{Input metadata}

\item{input_file_spec}{Input file specification to hunt down the
file of interest.}

\item{verbose}{Print diagnostic information while running?}

\item{species}{Specify a species to search for, or '*' for anything.}

\item{type}{Some likely filename searches may be for genome
vs. rRNA vs other feature types.}

\item{subtype}{Specific annotation type to seek.}

\item{tag}{GFF tag from which to seek IDs.}

\item{basedir}{Root directory containing the files/logs of metadata.}

\item{new_spec}{New specification resulting from this.}
}
\description{
This is useful for putting the count table name into a metadata file.
}
