% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_filename_search}
\alias{dispatch_filename_search}
\title{Pull out the filename matching an input spec}
\usage{
dispatch_filename_search(
  meta,
  input_file_spec,
  verbose = verbose,
  species = "*",
  type = "genome",
  basedir = "preprocessing"
)
}
\arguments{
\item{meta}{Input metadata}

\item{input_file_spec}{Input file specification to hunt down the
file of interest.}

\item{verbose}{Print diagnostic information while running?}

\item{basedir}{Root directory containing the files/logs of metadata.}
}
\description{
This is useful for putting the count table name into a metadata file.
}
