% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{dispatch_metadata_ratio}
\alias{dispatch_metadata_ratio}
\title{Given two metadata columns, print a ratio.}
\usage{
dispatch_metadata_ratio(
  meta,
  numerator_column = NULL,
  denominator_column = NULL,
  digits = 3,
  verbose = FALSE
)
}
\arguments{
\item{meta}{metadata, contains the column names!}

\item{numerator_column}{what it says on the tin.}

\item{denominator_column}{what it says on the tin.}

\item{digits}{Number of significant digits to keep in the output.}

\item{verbose}{unsed for the moment.}
}
\description{
Given two metadata columns, print a ratio.
}
