% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dispatch_sum_column}
\alias{dispatch_sum_column}
\title{Note: This is not taking some important things into account yet.}
\usage{
dispatch_sum_column(
  meta,
  input_file_spec,
  verbose = verbose,
  species = "*",
  basedir = "preprocessing",
  type = "genome",
  subtype = "gene",
  new_spec = NULL,
  tag = "ID"
)
}
\description{
Note: This is not taking some important things into account yet.
}
