% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.r
\name{divide_seq}
\alias{divide_seq}
\title{\code{divide_seq()}  Express a data frame of counts as reads per pattern per
million(library).}
\usage{
divide_seq(counts, pattern = "TA", fasta = "testme.fasta",
  gff = "testme.gff", entry_type = "gene")
}
\arguments{
\item{counts}{read count matrix}

\item{pattern}{pattern to search against.  Defaults to 'TA'}

\item{fasta}{a fasta genome to search}

\item{gff}{the gff set of annotations to define start/ends of genes.}

\item{entry_type}{which type of gff entry to search against.  Defaults to 'gene'.}
}
\value{
The 'RPseqM' counts
}
\description{
\code{divide_seq()}  Express a data frame of counts as reads per pattern per
million(library).
}
\examples{
\dontrun{
cptam <- divide_seq(cont_table, fasta="mgas_5005.fasta.xz", gff="mgas_5005.gff.xz")
}
}
\seealso{
\code{\link[Rsamtools]{FaFile}} \code{\link[edgeR]{rpkm}}
}

