% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_convert.r
\name{divide_seq}
\alias{divide_seq}
\title{Express a data frame of counts as reads per pattern per
million(library).}
\usage{
divide_seq(counts, pattern = "TA", fasta = "testme.fasta",
  gff = "testme.gff", entry_type = "gene", ...)
}
\arguments{
\item{counts}{Read count matrix.}

\item{pattern}{Pattern to search against.  Defaults to 'TA'.}

\item{fasta}{Fasta genome to search.}

\item{gff}{Gff set of annotations to define start/ends of genes.}

\item{entry_type}{Type of gff entry to search against.  Defaults to 'gene'.}

\item{...}{Options I might pass from other functions are dropped into arglist.}
}
\value{
The 'RPseqM' counts
}
\description{
This uses a sequence pattern rather than length to normalize sequence.  It is essentially rpkm
but fancy pants.
}
\examples{
\dontrun{
cptam <- divide_seq(cont_table, fasta="mgas_5005.fasta.xz", gff="mgas_5005.gff.xz")
}
}
\seealso{
\code{\link[Rsamtools]{FaFile}} \code{\link[edgeR]{rpkm}}
}

