% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{download_eupathdb_metadata}
\alias{download_eupathdb_metadata}
\title{Returns metadata for all eupathdb organisms.}
\usage{
download_eupathdb_metadata()
}
\value{
Dataframe with lots of rows for the various species in eupathdb.
}
\description{
Returns metadata for all eupathdb organisms.
}
\author{
Keith Hughitt
}
