% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{download_eupathdb_metadata}
\alias{download_eupathdb_metadata}
\title{Returns metadata for all eupathdb organisms.}
\usage{
download_eupathdb_metadata(overwrite = FALSE, dir = ".")
}
\arguments{
\item{overwrite}{Overwrite existing data?}

\item{dir}{Where to put the json.}
}
\value{
Dataframe with lots of rows for the various species in eupathdb.
}
\description{
Returns metadata for all eupathdb organisms.
}
\author{
Keith Hughitt
}
