% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_genbank.r
\name{download_gbk}
\alias{download_gbk}
\title{A genbank accession downloader scurrilously stolen from ape}
\usage{
download_gbk(accessions, write = TRUE)
}
\arguments{
\item{write}{Write the files?  Otherwise return a list of the strings}

\item{accession}{An accession!}
}
\value{
A list containing the number of files downloaded and the character strings actually acquired
}
\description{
This takes and downloads genbank accessions
}

