% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_genbank.R
\name{download_gbk}
\alias{download_gbk}
\title{A genbank accession downloader scurrilously stolen from ape.}
\usage{
download_gbk(accessions = "AE009949", write = TRUE)
}
\arguments{
\item{accessions}{An accession -- actually a set of them.}

\item{write}{Write the files?  Otherwise return a list of the strings}
}
\value{
A list containing the number of files downloaded and the character
  strings acquired.
}
\description{
This takes and downloads genbank accessions.
}
\details{
Tested in test_40ann_biomartgenbank.R
In this function I stole the same functionality from the ape package and set
a few defaults so that it hopefully fails less often.
}
\examples{
 written <- download_gbk(accessions = "AE009949")
 written$written_file
}
\seealso{
[ape]
}
\author{
The ape authors with some modifications by atb.
}
