% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.R
\name{download_microbesonline_files}
\alias{download_microbesonline_files}
\title{Download the various file formats from microbesoline.}
\usage{
download_microbesonline_files(id = "160490", type = "gbk")
}
\arguments{
\item{id}{Species ID to query.}

\item{type}{File type(s) to download, if left null it will grab the genbank,
tab, protein fasta, transcript fasta, and genome.}
}
\value{
List describing the files downloaded and their locations.
}
\description{
Microbesonline provides an interesting set of file formats to download.  Each
format proves useful under one condition or another, ergo this defaults to
iterating through them all and getting every file.
}
\examples{
## On those relatively rare occasions when the internet is being
## cranky, these tests will most certainly fail.

## Get the numeric ID from microbesonline for a species of interest.
## When microbesonline is under maintenance/etc this fails.
ecoli_id <- try(get_microbesonline_taxid(species = "S88"))
ecoli_id

if (! "try-error" \%in\% class(ecoli_id)) {
  ## Download a genbank/fasta/etc file for that species.
  ecoli_files <- try(download_microbesonline_files(id = ecoli_id, type = "gbk"))
  ecoli_files
  if (! "try-error" \%in\% class(ecoli_files)) {
    removed <- file.remove(ecoli_files[["gbk"]])
  }

  ## Download a species annotations to a dataframe
  ecoli_annot <- try(load_microbesonline_annotations(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_annot)) {
    dim(ecoli_annot)
    colnames(ecoli_annot)
  }

  ## Get gene ontology data from microbesonline
  ecoli_go <- try(load_microbesonline_go(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_go)) {
    dim(ecoli_go)
  }
}
}
