% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_ebseq.r
\name{ebseq_pairwise}
\alias{ebseq_pairwise}
\title{Set up model matrices contrasts and do pairwise comparisons of all conditions using EBSeq.}
\usage{
ebseq_pairwise(input = NULL, patterns = NULL, ng_vector = NULL,
  rounds = 10, target_fdr = 0.05, method = "pairwise_subset",
  norm = "median", ...)
}
\arguments{
\item{input}{Dataframe/vector or expt class containing data, normalization state, etc.}

\item{patterns}{Set of expression patterns to query.}

\item{ng_vector}{I think this is for isoform quantification, but am not yet
certain.}

\item{rounds}{Number of iterations for doing the multi-test}

\item{target_fdr}{Definition of 'significant'}

\item{norm}{Normalization method to use.}

\item{...}{Extra arguments currently unused.}

\item{conditions}{Factor of conditions in the experiment.}
}
\description{
Invoking EBSeq is confusing, this should help.
}
