% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_ebseq.R
\name{ebseq_two}
\alias{ebseq_two}
\title{The primary function used in my EBSeq implementation.}
\usage{
ebseq_two(
  pair_data,
  factors,
  numerator = 2,
  denominator = 1,
  fast = TRUE,
  ng_vector = NULL,
  rounds = 20,
  Alpha = NULL,
  Beta = NULL,
  Qtrm = 1,
  QtrmCut = 0,
  step1 = 1e-06,
  step2 = 0.01,
  thre = log(2),
  sthre = 0,
  filter = 10,
  stopthre = 1e-04,
  target_fdr = 0.05,
  norm = "median",
  force = FALSE
)
}
\arguments{
\item{pair_data}{Matrix containing the samples comprising two experimental
factors of interest.}

\item{numerator}{Which factor has the numerator in the data.}

\item{denominator}{Which factor has the denominator in the data.}

\item{ng_vector}{Passed to ebseq.}

\item{rounds}{Passed to ebseq.}

\item{target_fdr}{Passed to ebseq.}

\item{norm}{Normalization method of ebseq to apply.}

\item{force}{Force inappropriate data into ebseq?}

\item{conditions}{Factor of conditions in the data.}
}
\value{
EBSeq result table with some extra formatting.
}
\description{
Most of the time, my invocation of ebseq will fall into this function.
}
\seealso{
[ebseq_pairwise()]
}
