% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_edger.r
\name{edger_pairwise}
\alias{edger_pairwise}
\title{Set up a model matrix and set of contrasts to do pairwise comparisons using EdgeR.}
\usage{
edger_pairwise(input, conditions = NULL, batches = NULL,
  model_cond = TRUE, model_batch = TRUE, model_intercept = FALSE,
  alt_model = NULL, extra_contrasts = NULL, annot_df = NULL,
  force = FALSE, ...)
}
\arguments{
\item{input}{Dataframe/vector or expt class containing data, normalization state, etc.}

\item{conditions}{Factor of conditions in the experiment.}

\item{batches}{Factor of batches in the experiment.}

\item{model_cond}{Include condition in the experimental model?}

\item{model_batch}{Include batch in the model?  In most cases this is a good thing(tm).}

\item{model_intercept}{Use cell means or intercept?}

\item{alt_model}{Alternate experimental model to use?}

\item{extra_contrasts}{Add some extra contrasts to add to the list of pairwise contrasts.
This can be pretty neat, lets say one has conditions A,B,C,D,E
and wants to do (C/B)/A and (E/D)/A or (E/D)/(C/B) then use this
with a string like: "c_vs_b_ctrla = (C-B)-A, e_vs_d_ctrla = (E-D)-A,
de_vs_cb = (E-D)-(C-B),"}

\item{annot_df}{Annotation information to the data tables?}

\item{force}{Force edgeR to accept inputs which it should not have to deal with.}

\item{...}{The elipsis parameter is fed to write_edger() at the end.}
}
\value{
List including the following information:
  contrasts = The string representation of the contrasts performed.
  lrt = A list of the results from calling glmLRT(), one for each contrast.
  contrast_list = The list of each call to makeContrasts()
  I do this to avoid running into the limit on # of contrasts addressable by topTags()
  all_tables = a list of tables for the contrasts performed.
}
\description{
This function performs the set of possible pairwise comparisons using EdgeR.
}
\examples{
\dontrun{
 pretend = edger_pairwise(data, conditions, batches)
}
}
\seealso{
\pkg{edgeR} \code{\link[edgeR]{topTags}} \code{\link[edgeR]{glmLRT}}
  \code{\link{make_pairwise_contrasts}} \code{\link[edgeR]{DGEList}}
  \code{\link[edgeR]{calcNormFactors}} \code{\link[edgeR]{estimateTagwiseDisp}}
  \code{\link[edgeR]{estimateCommonDisp}} \code{\link[edgeR]{estimateGLMCommonDisp}}
  \code{\link[edgeR]{estimateGLMTrendedDisp}} \code{\link[edgeR]{glmFit}}
}

