% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{edger_pairwise}
\alias{edger_pairwise}
\title{edger_pairwise():  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions/batches.  In this case, there
must be a balanced set of batches for each condition.}
\usage{
edger_pairwise(expt = NULL, data = NULL, conditions = NULL,
  batches = NULL, extra_contrasts = NULL, ...)
}
\arguments{
\item{conditions}{a factor of conditions in the experiment}

\item{batches}{a factor of batches in the experiment}

\item{extra_contrasts}{some extra contrasts to add to the list
This can be pretty neat, lets say one has conditions A,B,C,D,E
and wants to do (C/B)/A and (E/D)/A or (E/D)/(C/B) then use this
with a string like: "c_minus_b_ctrla = (C-B)-A, e_minus_d_ctrla = (E-D)-A,
de_minus_cb = (E-D)-(C-B),"}

\item{...}{The elipsis parameter is fed to write_limma() at the end.}
}
\value{
A list including the following information:
  macb = the mashing together of condition/batch so you can look at it
  macb_model = The result of calling model.matrix(~0 + macb)
  macb_fit =  The result of calling lmFit(data, macb_model)
  voom_result = The result from voom()
  voom_design = The design from voom (redundant from voom_result, but convenient)
  macb_table = A table of the number of times each condition/batch pairing happens
  cond_table = A table of the number of times each condition appears (the denominator for the identities)
  batch_table = How many times each batch appears
  identities = The list of strings defining each condition by itself
  all_pairwise = The list of strings defining all the pairwise contrasts
  contrast_string = The string making up the makeContrasts() call
  pairwise_fits = The result from calling contrasts.fit()
  pairwise_comparisons = The result from eBayes()
  edger_result = The result from calling write_limma()
}
\description{
edger_pairwise():  Set up a model matrix and set of contrasts to do
a pairwise comparison of all conditions/batches.  In this case, there
must be a balanced set of batches for each condition.
}
\examples{
## pretend = edger_pairwise(data, conditions, batches)
}
\seealso{
\code{\link{write_limma}}
}

