% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{exclude_genes_expt}
\alias{exclude_genes_expt}
\title{Exclude some genes given a pattern match}
\usage{
exclude_genes_expt(
  expt,
  column = "txtype",
  method = "remove",
  ids = NULL,
  patterns = c("snRNA", "tRNA", "rRNA"),
  ...
)
}
\arguments{
\item{expt}{Expressionset containing expt object.}

\item{column}{fData column to use for subsetting.}

\item{method}{Either remove explicit rows, or keep them.}

\item{ids}{Specific IDs to exclude.}

\item{patterns}{Character list of patterns to remove/keep}

\item{...}{Extra arguments are passed to arglist, currently unused.}
}
\value{
A smaller expt
}
\description{
Because I am too lazy to remember that expressionsets use matrix subsets for
gene and sample.  Also those methods lead to shenanigans when I want to know
what happened to the data over the course of the subset.
}
\examples{
 \dontrun{
  all_expt <- create_expt(metadata)
  ## This assumes a column in the metadata named 'txtype' containing the
  ## information telling us what type of transcript each gene is.
  no_ribosomes <- exclude_genes_expt(all_expt, column="txtype",
                                     patterns=c("snRNA", "tRNA", "rRNA"))
  i_hate_these_genes <- exclude_genes_expt(all_expt, ids=c("gene1", "gene2"))
  only_ribosomes <- exclude_genes_expt(all_expt, method="keep")
}
}
\seealso{
\code{\link{create_expt}}
}
