% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{exprs,expt-method}
\alias{exprs,expt-method}
\title{A getter to pull the expression data from an expt.}
\usage{
\S4method{exprs}{expt}(object)
}
\arguments{
\item{object}{An expt.}
}
\description{
A getter to pull the expression data from an expt.
}
