% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\docType{methods}
\name{exprs}
\alias{exprs}
\alias{exprs,}
\alias{exprs-methods}
\title{Extend Biobase::exprs to handle expt ojects.}
\arguments{
\item{object}{The expt object from which to extract the expressionset.}
}
\description{
Extend Biobase::exprs to handle expt ojects.
}
