% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{expt_snp}
\alias{expt_snp}
\title{Gather snp information for an expt}
\usage{
expt_snp(expt, input_dir = "preprocessing/outputs",
  file_suffix = "_parsed_ratio.txt", bam_suffix = "_accepted_paired.bam",
  tolower = TRUE)
}
\arguments{
\item{expt}{an expressionset from which to extract information.}
}
\value{
some stuff
}
\description{
I have some initial code for working with snps, but it seems that it will be getting more use, so
make it testable etc.
}

