% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_tables.r
\name{expt_subset}
\alias{expt_subset}
\title{\code{expt_subset()}  Extract a subset of samples following some rule(s) from an
experiment class}
\usage{
expt_subset(expt, subset = NULL)
}
\arguments{
\item{expt}{an expt which is a home-grown class containing an
expressionSet, design, colors, etc.}

\item{subset}{a valid R expression which defines a subset of the
design to keep.}
}
\value{
metadata an expt class which contains the smaller set of
data
}
\description{
\code{expt_subset()}  Extract a subset of samples following some rule(s) from an
experiment class
}
\examples{
\dontrun{
 smaller_expt = expt_subset(big_expt, "condition=='control'")
 all_expt = expt_subset(expressionset, "")  ## extracts everything
}
}
\seealso{
\pkg{Biobase} \link[Biobase]{pData}
\link[Biobase]{exprs} \link[Biobase]{fData}
}

