% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expt_subset}
\alias{expt_subset}
\title{Extract a subset of samples following some rule(s) from an
experiment class}
\usage{
expt_subset(expt, subset)
}
\arguments{
\item{expt}{an expt which is a home-grown class containing an
expressionSet, design, colors, etc.}

\item{subset}{a valid R expression which defines a subset of the
design to keep.}
}
\value{
metadata an expt class which contains the smaller set of
data
}
\description{
Extract a subset of samples following some rule(s) from an
experiment class
}
\examples{
## smaller_expt = expt_subset(big_expt, "condition=='control'")
## all_expt = expt_subset(expressionset, "")  ## extracts everything
}
\seealso{
\code{\link{ExpressionSet}}, \code{\link{pData}},
\code{\link{exprs}}, and \code{\link{fData}}
}

