% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tables.r
\name{expt_subset}
\alias{expt_subset}
\title{Extract a subset of samples following some rule(s) from an
experiment class.}
\usage{
expt_subset(expt, subset = NULL)
}
\arguments{
\item{expt}{Expt chosen to extract a subset of data.}

\item{subset}{Valid R expression which defines a subset of the design to keep.}
}
\value{
metadata Expt class which contains the smaller set of data.
}
\description{
Sometimes an experiment has too many parts to work with conveniently, this operation allows one
to break it into smaller pieces.
}
\examples{
\dontrun{
 smaller_expt = expt_subset(big_expt, "condition=='control'")
 all_expt = expt_subset(expressionset, "")  ## extracts everything
}
}
\seealso{
\pkg{Biobase} \link[Biobase]{pData}
\link[Biobase]{exprs} \link[Biobase]{fData}
}

