% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{extract_de_plots}
\alias{extract_de_plots}
\title{Make a MA plot of some limma output with pretty colors and shapes.}
\usage{
extract_de_plots(
  pairwise,
  combined = NULL,
  type = NULL,
  invert = FALSE,
  invert_colors = c(),
  numerator = NULL,
  denominator = NULL,
  alpha = 0.4,
  z = 1.5,
  n = NULL,
  logfc = 1,
  pval = 0.05,
  adjp = TRUE,
  found_table = NULL,
  p_type = "adj",
  color_high = NULL,
  color_low = NULL,
  loess = FALSE,
  z_lines = FALSE,
  label = 10,
  label_column = "hgncsymbol"
)
}
\arguments{
\item{pairwise}{The result from all_pairwise(), which should be changed to
handle other invocations too.}

\item{combined}{Result from one of the combine_de_table functions.}

\item{type}{Type of table to use: deseq, edger, limma, basic.}

\item{invert}{Invert the plot?}

\item{invert_colors}{vector of new colors.}

\item{numerator}{Use this factor as the numerator.}

\item{denominator}{Use this factor as the denominator.}

\item{alpha}{Use this transparency.}

\item{z}{z-score cutoff for coefficient significance.}

\item{n}{Choose the top/bottom-n by logFC.}

\item{logfc}{What logFC to use for the MA plot horizontal lines.}

\item{pval}{Cutoff to define 'significant' by p-value.}

\item{adjp}{Use adjusted p-value?}

\item{found_table}{Result from edger to use, left alone it chooses the first.}

\item{p_type}{Adjusted or raw pvalues?}

\item{color_high}{Color to use for the 'high' genes.}

\item{color_low}{Color to use for the 'low' genes.}

\item{loess}{Add a loess estimator to the coefficient plot?}

\item{z_lines}{Add the z-score lines?}

\item{label}{Label this number of top-diff genes.}

\item{label_columns}{Use this column for labelling genes.}
}
\value{
a plot!
}
\description{
Yay pretty colors and shapes!
This function should be reworked following my rewrite of combine_de_tables().
It is certainly possible to make the logic here much simpler now.
}
\examples{
\dontrun{
 prettyplot <- edger_ma(all_aprwise) ## [sic, I'm witty! and can speel]
}
}
\seealso{
[plot_ma_de()] [plot_volcano_de()]
}
