% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.r
\name{extract_de_plots}
\alias{extract_de_plots}
\title{Make a MA plot of some limma output with pretty colors and shapes.}
\usage{
extract_de_plots(
  pairwise,
  type = "edger",
  table = NULL,
  logfc = 1,
  p_type = "adj",
  p = 0.05,
  invert = FALSE,
  ...
)
}
\arguments{
\item{pairwise}{The result from all_pairwise(), which should be changed to
handle other invocations too.}

\item{type}{Type of table to use: deseq, edger, limma, basic.}

\item{table}{Result from edger to use, left alone it chooses the first.}

\item{logfc}{What logFC to use for the MA plot horizontal lines.}

\item{p_type}{Adjusted or raw pvalues?}

\item{p}{Cutoff to define 'significant' by p-value.}

\item{invert}{Invert the plot?}

\item{...}{Extra arguments are passed to arglist.}
}
\value{
a plot!
}
\description{
Yay pretty colors and shapes!
This function should be reworked following my rewrite of combine_de_tables().
It is certainly possible to make the logic here much simpler now.
}
\examples{
\dontrun{
 prettyplot <- edger_ma(all_aprwise) ## [sic, I'm witty! and can speel]
}
}
\seealso{
[plot_ma_de()] [plot_volcano_de()]
}
