% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{extract_eupath_orthologs}
\alias{extract_eupath_orthologs}
\title{Given 2 species names from the eupathdb, make orthology tables betwixt them.}
\usage{
extract_eupath_orthologs(db, master = "GID", query_species = NULL,
  id_column = "ORTHOLOG_ID", org_column = "ORGANISM",
  url_column = "ORTHOLOG_GROUP", count_column = "ORTHOLOG_COUNT",
  print_speciesnames = FALSE)
}
\arguments{
\item{db}{Species name (subset) from one eupath database.}

\item{master}{Primary keytype to use for indexing the various tables.}

\item{query_species}{A list of exact species names to search for.  If uncertain
about them, add print_speciesnames=TRUE and be ready for a big blob of
text.  If left null, then it will pull all species.}

\item{id_column}{What column in the database provides the set of ortholog IDs?}

\item{org_column}{What column provides the species name?}

\item{url_column}{What column provides the orthomcl group ID?}

\item{count_column}{Name of the column with the count of species represented.}

\item{print_speciesnames}{Dump the species names for diagnostics?}
}
\value{
A big table of orthoMCL families, the columns are:
 \enumerate{
  \item  GID: The gene ID
  \item  ORTHOLOG_ID: The gene ID of the associated ortholog.
  \item  ORTHOLOG_SPECIES: The species of the associated ortholog.
  \item  ORTHOLOG_URL: The OrthoMCL group ID's URL.
  \item  ORTHOLOG_COUNT: The number of all genes from all species represented in
  this group.
  \item  ORTHOLOG_GROUP: The family ID
  \item  QUERIES_IN_GROUP: How many of the query species are represented in this
  group?
  \item  GROUP_REPRESENTATION: ORTHOLOG_COUNT / the number of possible species.
 }
}
\description{
The eupathdb provides such a tremendous wealth of information.  For me
though, it is difficult sometimes to boil it down into just the bits of
comparison I want for 1 species or between 2 species.  A singularly common
question I am asked is: "What are the most similar genes between species x
and y among these two arbitrary parasites?"  There are lots of ways to poke
at this question: run BLAST/fasta36, use biomart, query the ortholog tables
from the eupathdb, etc.  However, in all these cases, it is not trivial to
ask the next question:  What about: a:b and b:a?
This function attempts to address that for the case of two eupath species
from the same domain. (tritrypdb/fungidb/etc.)  It does however assume that
the sqlite package has been installed locally, if not it suggests you run the
make_organismdbi function in order to do that.
}
