% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{extract_go}
\alias{extract_go}
\title{Extract a set of geneID to GOID mappings from a suitable data source.}
\usage{
extract_go(db, metadf = NULL, keytype = "ENTREZID")
}
\arguments{
\item{db}{Data source containing mapping information.}

\item{metadf}{Data frame containing extant information.}

\item{keytype}{Keytype used for querying}
}
\value{
Dataframe of 2 columns: geneID and goID.
}
\description{
Like extract_lengths above, this is primarily intended to read gene ID and GO ID mappings from a
OrgDb/OrganismDbi object.
}
\seealso{
\pkg{AnnotationDbi}
}
