% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{extract_keepers}
\alias{extract_keepers}
\title{When a list of 'keeper' contrasts is specified, extract it from the data.}
\usage{
extract_keepers(
  extracted,
  keepers,
  table_names,
  all_coefficients,
  limma,
  edger,
  ebseq,
  deseq,
  basic,
  noiseq,
  adjp,
  annot_df,
  include_deseq,
  include_edger,
  include_ebseq,
  include_limma,
  include_basic,
  include_noiseq,
  excludes,
  padj_type,
  fancy = FALSE,
  loess = FALSE,
  lfc_cutoff = 1,
  p_cutoff = 0.05,
  sheet_prefix = NULL,
  sheet_number = NULL,
  format_sig = 4,
  plot_colors = plot_colors,
  z = 1.5,
  alpha = 0.4,
  z_lines = FALSE,
  label = 10,
  label_column = "hgncsymbol"
)
}
\arguments{
\item{extracted}{Tables extracted from the all_pairwise data.}

\item{keepers}{In this case, one may assume either NULL or 'all'.}

\item{table_names}{The set of tables produced by all_pairwise().}

\item{all_coefficients}{The set of all experimental conditions in the
experimental metadata.}

\item{limma}{The limma data from all_pairwise().}

\item{edger}{The edger data from all_pairwise().}

\item{ebseq}{The ebseq data from all_pairwise().}

\item{deseq}{The deseq data from all_pairwise().}

\item{basic}{The basic data from all_pairwise().}

\item{adjp}{Pull out the adjusted p-values from the data?}

\item{annot_df}{What annotations should be added to the table?}

\item{include_deseq}{Whether or not to include the deseq data.}

\item{include_edger}{Whether or not to include the edger data.}

\item{include_ebseq}{Whether or not to include the ebseq data.}

\item{include_limma}{Whether or not to include the limma data.}

\item{include_basic}{Whether or not to include the basic data.}

\item{excludes}{Set of genes to exclude.}

\item{padj_type}{Choose a specific p adjustment.}

\item{fancy}{Include larger pdf/svg plots with the xlsx output?}

\item{loess}{Add a loess to plots?}

\item{lfc_cutoff}{Passed for volcano/MA plots and defining 'significant'}

\item{p_cutoff}{Passed for volcano/MA plots and defining 'significant'}

\item{sheet_prefix}{Prefix for this worksheet id.}

\item{sheet_number}{Which sheet is this?}

\item{format_sig}{Number of significant digits for stuff like
pvalues.}

\item{plot_colors}{Define what colors should be used for
'up'/'down'}

\item{z}{Define significantly away from the identity line in a
coefficient plot.}

\item{alpha}{Use this alpha transparency for plots.}

\item{z_lines}{Include lines denoting significant z-scores?}

\item{label}{When not NULL, label this many genes.}

\item{label_column}{Try using this column for labeling genes.}
}
\description{
This is the most interesting of the extract_keeper functions.  It must check
that the numerators and denominators match the desired contrast and flip the
signs in the logFCs when appropriate.
}
