% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{extract_keepers_lst}
\alias{extract_keepers_lst}
\title{When a list of 'keeper' contrasts is specified, extract it from the data.}
\usage{
extract_keepers_lst(
  extracted,
  keepers,
  table_names,
  all_coefficients,
  limma,
  edger,
  ebseq,
  deseq,
  basic,
  adjp,
  annot_df,
  include_deseq,
  include_edger,
  include_ebseq,
  include_limma,
  include_basic,
  excludes,
  padj_type
)
}
\arguments{
\item{keepers}{In this case, one may assume either NULL or 'all'.}

\item{table_names}{The set of tables produced by all_pairwise().}

\item{all_coefficients}{The set of all experimental conditions in the
experimental metadata.}

\item{limma}{The limma data from all_pairwise().}

\item{edger}{The edger data from all_pairwise().}

\item{ebseq}{The ebseq data from all_pairwise().}

\item{deseq}{The deseq data from all_pairwise().}

\item{basic}{The basic data from all_pairwise().}

\item{adjp}{Pull out the adjusted p-values from the data?}

\item{annot_df}{What annotations should be added to the table?}

\item{include_deseq}{Whether or not to include the deseq data.}

\item{include_edger}{Whether or not to include the edger data.}

\item{include_ebseq}{Whether or not to include the ebseq data.}

\item{include_limma}{Whether or not to include the limma data.}

\item{include_basic}{Whether or not to include the basic data.}

\item{excludes}{Set of genes to exclude.}

\item{padj_type}{Choose a specific p adjustment.}

\item{sheet_count}{What sheet is being written?}

\item{numerators}{Set of numerators}

\item{denominators}{Set of denominators}

\item{combo}{empty list}
}
\description{
This is the most interesting of the extract_keeper functions.  It must check
that the numerators and denominators match the desired contrast and flip the
signs in the logFCs when appropriate.
}
