% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{extract_linear_regression}
\alias{extract_linear_regression}
\title{Gather the models and perform forest plots to look at various regression analyses.}
\usage{
extract_linear_regression(
  meta,
  query = "condition",
  multivariable = TRUE,
  intercept = FALSE,
  factors = NULL,
  excel = NULL
)
}
\arguments{
\item{meta}{Experimental design}

\item{query}{Factor to query against}

\item{multivariable}{If set to FALSE, this will iterate over every factor individually.}

\item{intercept}{Set an intercept? (unused)}

\item{factors}{Set of factors to query}

\item{excel}{output xlsx file.}
}
\description{
Gather the models and perform forest plots to look at various regression analyses.
}
