% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{extract_metadata}
\alias{extract_metadata}
\title{Pull metadata from a table (xlsx/xls/csv/whatever)}
\usage{
extract_metadata(metadata, ...)
}
\description{
Pull metadata from a table (xlsx/xls/csv/whatever)
}
