% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Pull metadata from a table (xlsx/xls/csv/whatever)}
\usage{
extract_metadata(metadata, id_column = "sampleid", fill = NULL, ...)
}
\arguments{
\item{metadata}{file or df of metadata}

\item{id_column}{Column in the metadat containing the sample names.}

\item{...}{Arguments to pass to the child functions (read_csv etc).}
}
\value{
Metadata dataframe hopefully cleaned up to not be obnoxious.
}
\description{
I find that when I acquire metadata from a paper or collaborator, annoyingly
often there are many special characters or other shenanigans in the column
names.  This function performs some simple sanitizations.  In addition, if I
give it a filename it calls my generic 'read_metadata()' function before
sanitizing.
}
\examples{
 \dontrun{
  sanitized <- extract_metadata("some_random_supplemental.xls")
  saniclean <- extract_metadata(some_goofy_df)
}
}
