% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{extract_mzxml_data}
\alias{extract_mzxml_data}
\title{Read a bunch of mzXML files to acquire their metadata.}
\usage{
extract_mzxml_data(metadata, write_windows = TRUE, parallel = TRUE, ...)
}
\arguments{
\item{metadata}{Data frame describing the samples, including the mzXML
filenames.}

\item{write_windows}{Write out SWATH window frames.}

\item{...}{Extra arguments, presumably color palettes and column names and
stuff like that.}
}
\value{
metadata!#'
}
\description{
I have had difficulties getting the full set of correct parameters for a
DDA/DIA experiment.  After some poking, I eventually found most of these
required prameters in the mzXML raw files.  Ergo, this function uses them.
}
