% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.r
\name{extract_significant_genes}
\alias{extract_significant_genes}
\title{extract_significant_genes()  Pull the highly up/down genes in combined tables}
\usage{
extract_significant_genes(combined, according_to = "limma", fc = 1,
  p = 0.05, z = NULL, n = NULL,
  sig_table = "excel/significant_genes.xlsx")
}
\arguments{
\item{combined}{the output from combine_de_tables()}

\item{according_to}{default='limma'  one may use the deseq, edger, limma, or meta data.}

\item{fc}{default=1.0  a log fold change to define 'significant'}

\item{p}{default=0.05  a (adjusted)p-value to define 'significant'}

\item{z}{default=NULL  a z-score to define 'significant'}

\item{n}{default=NULL  a set of top/bottom-n genes}

\item{sig_table}{default="excel/significant_genes.xlsx"  an excel file to write}
}
\value{
a set of up-genes, down-genes, and numbers therein
}
\description{
Given the output from combine_de_tables(), extract the fun genes.
}
\seealso{
\code{\link{combine_de_tables}}
}

