% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{extract_significant_genes}
\alias{extract_significant_genes}
\title{Pull the highly up/down genes in combined tables}
\usage{
extract_significant_genes(combined, according_to = "limma", fc = 1,
  p = 0.05, z = NULL, n = NULL, excel = "excel/significant_genes.xlsx")
}
\arguments{
\item{combined}{the output from combine_de_tables()}

\item{according_to}{one may use the deseq, edger, limma, or meta data.}

\item{fc}{a log fold change to define 'significant'}

\item{p}{a (adjusted)p-value to define 'significant'}

\item{z}{a z-score to define 'significant'}

\item{n}{a set of top/bottom-n genes}

\item{excel}{an excel file to write}
}
\value{
a set of up-genes, down-genes, and numbers therein
}
\description{
Given the output from combine_de_tables(), extract the fun genes.
}
\seealso{
\code{\link{combine_de_tables}}
}

