% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{extract_significant_genes}
\alias{extract_significant_genes}
\title{Extract the sets of genes which are significantly up/down regulated
from the combined tables.}
\usage{
extract_significant_genes(combined, according_to = "limma", fc = 1,
  p = 0.05, z = NULL, n = NULL, excel = "excel/significant_genes.xlsx")
}
\arguments{
\item{combined}{Output from combine_de_tables().}

\item{according_to}{What tool(s) decide 'significant?'  One may use
the deseq, edger, limma, basic, meta, or all.}

\item{fc}{Log fold change to define 'significant'.}

\item{p}{(Adjusted)p-value to define 'significant'.}

\item{z}{Z-score to define 'significant'.}

\item{n}{Take the top/bottom-n genes.}

\item{excel}{Write the results to this excel file, or NULL.}
}
\value{
The set of up-genes, down-genes, and numbers therein.
}
\description{
Given the output from combine_de_tables(), extract the genes in
which we have the greatest likely interest, either because they
have the largest fold changes, lowest p-values, fall outside a
z-score, or are at the top/bottom of the ranked list.
}
\seealso{
\code{\link{combine_de_tables}}
}

