% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{extract_significant_genes}
\alias{extract_significant_genes}
\title{Extract the sets of genes which are significantly up/down regulated
from the combined tables.}
\usage{
extract_significant_genes(
  combined,
  according_to = "all",
  lfc = 1,
  p = 0.05,
  sig_bar = TRUE,
  z = NULL,
  n = NULL,
  min_mean_exprs = NULL,
  exprs_column = NULL,
  top_percent = NULL,
  p_type = "adj",
  invert_barplots = FALSE,
  excel = NULL,
  fc_column = NULL,
  p_column = NULL,
  siglfc_cutoffs = c(0, 1, 2),
  column_suffix = TRUE,
  gmt = FALSE,
  category = "category",
  fancy = FALSE,
  phenotype_name = "phenotype",
  set_name = "set",
  current_id = "ENSEMBL",
  required_id = "ENTREZID",
  min_gmt_genes = 10,
  ...
)
}
\arguments{
\item{combined}{Output from combine_de_tables().}

\item{according_to}{What tool(s) decide 'significant?'  One may use
the deseq, edger, limma, basic, meta, or all.}

\item{lfc}{Log fold change to define 'significant'.}

\item{p}{(Adjusted)p-value to define 'significant'.}

\item{sig_bar}{Add bar plots describing various cutoffs of 'significant'?}

\item{z}{Z-score to define 'significant'.}

\item{n}{Take the top/bottom-n genes.}

\item{min_mean_exprs}{Add a minimum expression value.}

\item{exprs_column}{Use this column to define expression.}

\item{top_percent}{Use a percentage to get the top-n genes.}

\item{p_type}{use an adjusted p-value?}

\item{invert_barplots}{Invert the significance barplots as per Najib's request?}

\item{excel}{Write the results to this excel file, or NULL.}

\item{fc_column}{Column in the DE data containing the foldchange values.}

\item{p_column}{Column in the DE data containing the pvalues.}

\item{siglfc_cutoffs}{Set of cutoffs used to define levels of 'significant.'}

\item{column_suffix}{Used to help determine which columns are used to find significant
genes via logfc/p-value.}

\item{gmt}{Write a gmt file using this result?}

\item{category}{When writing gmt files, set the category here.}

\item{fancy}{Write fancy plots with the xlsx file?}

\item{phenotype_name}{When writing gmt files, set the phenotype flag here.}

\item{set_name}{When writing gmt files, assign the set here.}

\item{current_id}{Choose the current ID type for an output gmt file.}

\item{required_id}{Choose the desired ID type for an output gmt file.}

\item{min_gmt_genes}{Define the minimum number of genes in a gene set for writing a gmt file.}

\item{...}{Arguments passed into arglist.}
}
\value{
The set of up-genes, down-genes, and numbers therein.
}
\description{
Given the output from combine_de_tables(), extract the genes in
which we have the greatest likely interest, either because they
have the largest fold changes, lowest p-values, fall outside a
z-score, or are at the top/bottom of the ranked list.
}
\seealso{
\code{\link{combine_de_tables}}
}
