% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{extract_significant_genes}
\alias{extract_significant_genes}
\title{Extract the sets of genes which are significantly up/down regulated
from the combined tables.}
\usage{
extract_significant_genes(
  combined,
  according_to = "all",
  lfc = 1,
  p = 0.05,
  sig_bar = TRUE,
  z = NULL,
  n = NULL,
  top_percent = NULL,
  ma = TRUE,
  p_type = "adj",
  invert_barplots = FALSE,
  excel = NULL,
  siglfc_cutoffs = c(0, 1, 2),
  ...
)
}
\arguments{
\item{combined}{Output from combine_de_tables().}

\item{according_to}{What tool(s) decide 'significant?'  One may use
the deseq, edger, limma, basic, meta, or all.}

\item{lfc}{Log fold change to define 'significant'.}

\item{p}{(Adjusted)p-value to define 'significant'.}

\item{sig_bar}{Add bar plots describing various cutoffs of 'significant'?}

\item{z}{Z-score to define 'significant'.}

\item{n}{Take the top/bottom-n genes.}

\item{top_percent}{Use a percentage to get the top-n genes.}

\item{ma}{Add ma plots to the sheets of 'up' genes?}

\item{p_type}{use an adjusted p-value?}

\item{invert_barplots}{Invert the significance barplots as per Najib's request?}

\item{excel}{Write the results to this excel file, or NULL.}

\item{siglfc_cutoffs}{Set of cutoffs used to define levels of 'significant.'}

\item{...}{Arguments passed into arglist.}
}
\value{
The set of up-genes, down-genes, and numbers therein.
}
\description{
Given the output from combine_de_tables(), extract the genes in
which we have the greatest likely interest, either because they
have the largest fold changes, lowest p-values, fall outside a
z-score, or are at the top/bottom of the ranked list.
}
\seealso{
\code{\link{combine_de_tables}}
}
