% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{factor_rsquared}
\alias{factor_rsquared}
\title{Collect the r^2 values from a linear model fitting between a singular
value decomposition and factor}
\usage{
factor_rsquared(svd_v, factor)
}
\arguments{
\item{svd_v}{The V' V = I portion of a fast.svd call}

\item{factor}{a factor describing the original data}
}
\value{
The r^2 values of the linear model as a %
}
\description{
Collect the r^2 values from a linear model fitting between a singular
value decomposition and factor
}
\seealso{
\code{\link{fast.svd}}
}

