% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{factor_rsquared}
\alias{factor_rsquared}
\title{factor_rsquared()  Collect the r^2 values from a linear model fitting between a singular
value decomposition and factor.}
\usage{
factor_rsquared(svd_v, factor)
}
\arguments{
\item{svd_v}{the V' V = I portion of a fast.svd call.}

\item{factor}{an experimental factor from the original data.}
}
\value{
The r^2 values of the linear model as a percentage.
}
\description{
factor_rsquared()  Collect the r^2 values from a linear model fitting between a singular
value decomposition and factor.
}
\seealso{
\code{\link{fast.svd}}
}

