% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{fData,expt-method}
\alias{fData,expt-method}
\title{A getter to pull the gene annotation data from an expt.}
\usage{
\S4method{fData}{expt}(object)
}
\arguments{
\item{object}{An expt.}
}
\description{
A getter to pull the gene annotation data from an expt.
}
