% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\docType{methods}
\name{fData}
\alias{fData}
\title{Extend Biobase::fData to handle expt objects.}
\usage{
\S4method{fData}{expt}(object)
}
\arguments{
\item{object}{An expt from which to extract the expressionset.}
}
\description{
Extend Biobase::fData to handle expt objects.
}
