% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{fData<-,expt,ANY-method}
\alias{fData<-,expt,ANY-method}
\title{A setter to put the gene annotation data into an expt.}
\usage{
\S4method{fData}{expt,ANY}(object) <- value
}
\arguments{
\item{expt}{An expt.}
}
\description{
A setter to put the gene annotation data into an expt.
}
