% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{fData<-,SummarizedExperiment,ANY-method}
\alias{fData<-,SummarizedExperiment,ANY-method}
\title{A setter to put the gene annotation data into a SummarizedExperiment.}
\usage{
\S4method{fData}{SummarizedExperiment,ANY}(object) <- value
}
\arguments{
\item{object}{A SummarizedExperiment.}

\item{value}{New annotations for the se.}
}
\description{
A setter to put the gene annotation data into a SummarizedExperiment.
}
