% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filter_counts}
\alias{filter_counts}
\title{Filter low-count genes from a data set.}
\usage{
filter_counts(counts, lib.size = NULL, thresh = 4, minSamples = 2)
}
\arguments{
\item{df}{input data frame of counts by sample}

\item{lib.size}{optional list of library sizes}

\item{thresh}{lower threshold of counts (4 by default)}

\item{minSamples}{minimum number of samples}
}
\value{
dataframe of counts without the low-count genes
}
\description{
Filter low-count genes from a data set.
}
\examples{
## filtered_table = filter_counts(count_table)
}
\seealso{
\code{\link{log2CPM}} which this uses to decide what to keep
}

