% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{filter_counts}
\alias{filter_counts}
\title{Call various count filters.}
\usage{
filter_counts(
  count_table,
  method = "cbcb",
  p = 0.01,
  A = 1,
  k = 1,
  cv_min = 0.01,
  cv_max = 1000,
  thresh = 2,
  min_samples = 2,
  ...
)
}
\arguments{
\item{count_table}{Some counts to filter.}

\item{method}{Filtering method to apply (cbcb, pofa, kofa, cv right now).}

\item{p}{Used by genefilter's pofa().}

\item{A}{Also for pofa().}

\item{k}{Used by genefilter's kofa().}

\item{cv_min}{Used by genefilter's cv().}

\item{cv_max}{Also used by cv().}

\item{thresh}{Minimum threshold across samples for cbcb.}

\item{min_samples}{Minimum number of samples for cbcb.}

\item{...}{More options might be needed, especially if I fold cv/p/etc into ...}
}
\value{
Data frame of filtered counts.
}
\description{
This calls the various filtering functions in genefilter along with
suggestions made in our lab meetings; defaulting to the threshold based
filter suggested by Hector.
}
\examples{
\dontrun{
 new <- filter_counts(old)
}
}
\seealso{
\pkg{genefilter}
}
