% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{find_working_dataset}
\alias{find_working_dataset}
\title{Search a mart for a usable dataset.}
\usage{
find_working_dataset(mart, trydataset, species)
}
\arguments{
\item{mart}{Biomart instance to poke at in an attempt to find a dataset.}

\item{trydataset}{Dataset to attempt to query.}

\item{species}{Species at the mart for which to search.}
}
\description{
Search a mart for a usable dataset.
}
