% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{find_working_mart}
\alias{find_working_mart}
\title{Find a functional biomart instance.}
\usage{
find_working_mart(
  default_hosts = c("useast.ensembl.org", "www.ensembl.org", "asia.ensembl.org"),
  trymart = "ENSEMBL_MART_ENSEMBL",
  archive = FALSE,
  year = NULL,
  month = NULL
)
}
\arguments{
\item{default_hosts}{List of biomart mirrors to try.}

\item{trymart}{Specific mart to query.}

\item{archive}{Try an archive server instead of a mirror?  If this is a
character, it will assume it is a specific archive hostname.}

\item{year}{Choose specific year(s) for the archive servers?}

\item{month}{Choose specific month(s) for the archive servers?}
}
\value{
Either a mart instance or NULL if no love was forthcoming.
}
\description{
In my experience, the various biomart mirrors are not varyingly likely to be
functional at any given time.  In addition, I often find it useful to use an
archive instance rather than the most recent ensembl instance.  This function
therefore iterates over the various mirrors; or if archive = TRUE it will try a
series of archive servers from 1, 2, and 3 years ago.
}
\examples{
## Provide examples of the biomart annotation functions.

## Search for a biomart instance which will accept queries.
archive_mart <- try(find_working_mart(archive = TRUE, year = 2023, month = 02))
class(archive_mart)
archive_mart

## I should probably change this function to accept the mart from
## find_working_mart()
hs_annot <- try(load_biomart_annotations(year = 2023, month = 02))
class(hs_annot)

## I added a S3 print method to try to avoid being made sad if I
## print the result from load_annotations.
hs_annot

## If one wishes to check out all of the possible data types for a
## single gene, try this:
## However, this function is painfully slow and likely annoying to the
## ensembl servers.
## single_gene <- get_biomart_example_gene()

## Grab gene ontology data for a species:
hs_go <- try(load_biomart_go(year = 2023, month = 02))
hs_go

## The following example is fragile I think because multiple
## ensembl instances must be responding properly to work
## If it fails; check that the examples for biomaRt::getLDS are
## functioning -- most likely they are failing due to an upstream
## error.
hs_mm_ortho <- try(load_biomart_orthologs(gene_ids = "ENSMUSG00000000001",
                                          year = 2023, month = 02,
                                          first_species = "mmusculus",
                                          second_species = "scerevisiae"))

## Create salmon-like ID.version strings using ensembl annotations.
tx_gene_map <- try(make_tx_gene_map(as.data.frame(hs_annot)))
head(tx_gene_map)
}
\seealso{
[biomaRt::useMart()] [biomaRt::listMarts()]
}
