% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{gather_genes}
\alias{gather_genes}
\title{gather_genes()  Given a set of goseq data from simple_goseq(), make a list of genes represented in each ontology.}
\usage{
gather_genes(goseq_data, ontology = "MF", pval = 0.05)
}
\arguments{
\item{goseq_data}{a list of goseq specific results as generated by simple_goseq()}

\item{pval}{default=0.05  a maximum accepted pvalue to include in the list of categories to cross reference.}

\item{ont}{default='MF'  an ontology to search}
}
\value{
a data frame of categories/genes.
}
\description{
This function uses the GO2ALLEG data structure to reverse map ontology categories to a list of genes represented.
It therefore assumes that the GO2ALLEG.rda data structure has been deposited in pwd().  This in turn may be generated
by clusterProfilers buildGOmap() function if it doesn't exist.  For some species it may also be auto-generated.
With little work this can be made much more generic, and it probably should.
}
\examples{
## data = simple_goseq(de_genes=limma_output, lengths=annotation_df, goids=goids_df)
## genes_in_cats = gather_genes(data, ont='BP')
}
\seealso{
\code{\link{simple_goseq}}, \code{\link{buildGOmap}},
}

