% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_genbank.r
\name{gbk_annotations}
\alias{gbk_annotations}
\title{Extract some useful information from a gbk imported as a txDb.}
\usage{
gbk_annotations(gbr)
}
\arguments{
\item{gbr}{TxDb object to poke at.}
}
\value{
Granges data
}
\description{
Maybe this should get pulled into the previous function?
}
\details{
Tested in test_40ann_biomartgenbank.R
This function should provide a quick reminder of how to use the AnnotationDbi select function
if it does nothing else.  It also (hopefully helpfully) returns a granges object containing
the essential information one might want for printing out a gff or whatever.
}
\examples{
\dontrun{
 annotations <- gbk_annotations("saureus_txdb")
}
}
\seealso{
\pkg{AnnotationDbi} \pkg{GenomeInfoDb} \pkg{GenomicFeatures}
 \code{\link[AnnotationDbi]{select}}
}

