% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/015_annotation_genbank.r
\name{gbk_annotations}
\alias{gbk_annotations}
\title{Extract some useful information from a gbk imported as a txDb.}
\usage{
gbk_annotations(gbr)
}
\arguments{
\item{gbr}{TxDb object to poke at.}
}
\value{
Granges data
}
\description{
This function no longer really stands on its own, but is more accessible
from load_genbank_annotations().
}
\details{
Tested in test_40ann_biomartgenbank.R
This function should provide a quick reminder of how to use the AnnotationDbi
select function if it does nothing else.  It also (hopefully helpfully)
returns a granges object containing the essential information one might want
for printing out a gff or whatever.

I should revisit this function and improve the generated ranges objects to
have better metadata columns via the mcols() function.  For examples of some
useful tasks one can do here, check out snp.r.
}
\seealso{
\pkg{AnnotationDbi} \pkg{GenomeInfoDb} \pkg{GenomicFeatures}
 \code{\link[AnnotationDbi]{select}}
}
