% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{genefilter_cv_counts}
\alias{genefilter_cv_counts}
\title{Filter genes from a dataset outside a range of variance.}
\usage{
genefilter_cv_counts(count_table, cv_min = 0.01, cv_max = 1000)
}
\arguments{
\item{count_table}{Input data frame of counts by sample.}

\item{cv_min}{Minimum coefficient of variance.}

\item{cv_max}{Maximum coefficient of variance.}
}
\value{
Dataframe of counts without the high/low variance genes.
}
\description{
This function from genefilter removes genes surpassing a variance cutoff.  It is not therefore a
low-count filter per se.
}
\examples{
\dontrun{
 filtered_table = genefilter_kofa_counts(count_table)
}
}
\seealso{
\pkg{genefilter}
 \code{\link[genefilter]{kOverA}}
}
