% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{genefilter_cv_counts}
\alias{genefilter_cv_counts}
\title{Filter genes from a dataset outside a range of variance}
\usage{
genefilter_cv_counts(count_table, cv_min = 0.01, cv_max = 1000,
  verbose = FALSE)
}
\arguments{
\item{cv_min}{a minimum coefficient of variance}

\item{cv_max}{guess}

\item{verbose}{If set to true, prints number of genes removed / remaining}

\item{counts}{input data frame of counts by sample}
}
\value{
dataframe of counts without the low-count genes
}
\description{
Filter genes from a dataset outside a range of variance
}
\examples{
## filtered_table = genefilter_kofa_counts(count_table)
}
\seealso{
\code{\link{genefilter}} \code{\link{kOverA}} which this uses to decide what to keep
}

