% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{genefilter_kofa_counts}
\alias{genefilter_kofa_counts}
\title{Filter low-count genes from a data set using genefilter's kOverA().}
\usage{
genefilter_kofa_counts(count_table, k = 1, A = 1)
}
\arguments{
\item{count_table}{Input data frame of counts by sample.}

\item{k}{Minimum number of samples to have >A counts.}

\item{A}{Minimum number of counts for each gene's sample in kOverA().}
}
\value{
Dataframe of counts without the low-count genes.
}
\description{
This is the most similar to the function suggested by Hector I think.
}
\examples{
\dontrun{
 filtered_table = genefilter_kofa_counts(count_table)
}
}
\seealso{
\pkg{genefilter}
 \code{\link[genefilter]{kOverA}}
}
