% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{genefilter_kofa_counts}
\alias{genefilter_kofa_counts}
\title{Filter low-count genes from a data set using genefilter's kOverA()}
\usage{
genefilter_kofa_counts(count_table, k = 1, A = 1)
}
\arguments{
\item{count_table}{input data frame of counts by sample}

\item{k}{a minimum number of samples to have >A counts}

\item{A}{the minimum number of counts for each gene's sample in kOverA()}
}
\value{
dataframe of counts without the low-count genes
}
\description{
Filter low-count genes from a data set using genefilter's kOverA()
}
\examples{
\dontrun{
 filtered_table = genefilter_kofa_counts(count_table)
}
}
\seealso{
\pkg{genefilter} \code{\link[genefilter]{kOverA}} which this uses to decide what to keep
}

