% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{genefilter_kofa_counts}
\alias{genefilter_kofa_counts}
\title{Filter low-count genes from a data set using genefilter's kOverA()}
\usage{
genefilter_kofa_counts(count_table, k = 1, A = 1, verbose = FALSE)
}
\arguments{
\item{k}{a minimum number of samples to have >A counts}

\item{A}{the minimum number of counts for each gene's sample in kOverA()}

\item{verbose}{If set to true, prints number of genes removed / remaining}

\item{counts}{input data frame of counts by sample}
}
\value{
dataframe of counts without the low-count genes
}
\description{
Filter low-count genes from a data set using genefilter's kOverA()
}
\examples{
## filtered_table = genefilter_kofa_counts(count_table)
}
\seealso{
\code{\link{genefilter}} \code{\link{kOverA}} which this uses to decide what to keep
}

