% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.r
\name{get_biomart_annotations}
\alias{get_biomart_annotations}
\title{Extract annotation information from biomart without having to remember the stupid biomart parameters}
\usage{
get_biomart_annotations(species = "hsapiens", overwrite = FALSE,
  do_save = FALSE, host = "dec2015.archive.ensembl.org",
  trymart = "ENSEMBL_MART_ENSEMBL")
}
\arguments{
\item{species}{currently only hsapiens}

\item{overwrite}{overwrite a savefile if it exists}

\item{do_save}{create a savefile of the annotations}

\item{host}{ensembl hostname to use}

\item{trymart}{biomart has become a circular dependency, this makes me sad, now to list the marts, you need to have a mart loaded...}
}
\description{
Extract annotation information from biomart without having to remember the stupid biomart parameters
}
\examples{
\dontrun{
 tt = get_biomart_annotations()
}
}

