% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{get_biomart_annotations}
\alias{get_biomart_annotations}
\title{Extract annotation information from biomart without having to remember the stupid biomart parameters}
\usage{
get_biomart_annotations(species = "hsapiens", overwrite = FALSE)
}
\arguments{
\item{species}{currently only hsapiens}
}
\description{
Extract annotation information from biomart without having to remember the stupid biomart parameters
}
\examples{
\dontrun{
 tt = get_biomarg_annotations()
}
}

