% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_colors,SummarizedExperiment-method}
\alias{get_colors,SummarizedExperiment-method}
\title{Get colors from a SE.}
\usage{
\S4method{get_colors}{SummarizedExperiment}(exp, ...)
}
\arguments{
\item{exp}{Input se}

\item{...}{Arguments passed along}
}
\description{
Get colors from a SE.
}
