% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_colors_by_condition}
\alias{get_colors_by_condition}
\title{Get a list of colors by condition instead of by sample id.}
\usage{
get_colors_by_condition(exp, ...)
}
\arguments{
\item{exp}{Input data structure}

\item{...}{Arguments to pass along.}
}
\description{
Get a list of colors by condition instead of by sample id.
}
