% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{get_de_table,hpgltools::combine_de_tables-method}
\alias{get_de_table,hpgltools::combine_de_tables-method}
\title{Extract tables from the result of combine_de_tables.}
\usage{
\S4method{get_de_table}{hpgltools::combine_de_tables}(apr, table, type = "deseq", combined = FALSE)
}
\arguments{
\item{apr}{Input data, defaulting to the result from all_pairwise.}

\item{table}{Table name to extract.}

\item{type}{Choose the result from this DE method.}

\item{combined}{Combine all the result types.}
}
\value{
Single table from the pairwise result.
}
\description{
I am getting some strange unescaped backslashes in the roxygen created
documentation files for some functions, including this one.  I do
not think there are any backslash escaped characters anywhere in
this portion of my code.
}
