% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{get_de_table}
\alias{get_de_table}
\title{Extract a DE table from some data structure.}
\usage{
get_de_table(apr, table, type = "deseq", combined = FALSE)
}
\arguments{
\item{apr}{Input data, defaulting to the result from all_pairwise.}

\item{table}{Table name to extract}

\item{type}{Choose the result from this DE method.}

\item{combined}{Combine all the result types.}
}
\value{
Single table from the pairwise result.
}
\description{
I am getting some strange unescaped backslashes in the roxygen-created
documentation files for some functions, including this one.
}
