% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_testing.R
\name{get_degrees}
\alias{get_degrees}
\title{Use a vector of factors and design to count up the levels.}
\usage{
get_degrees(design, fctrs)
}
\description{
Use a vector of factors and design to count up the levels.
}
