% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{get_eupath_ortholog_table}
\alias{get_eupath_ortholog_table}
\title{Returns a mapping of gene ID to ortholog genes}
\usage{
get_eupath_ortholog_table(provider = "TriTrypDB",
  species = "Leishmania major", overwrite = FALSE, dir = ".")
}
\arguments{
\item{species}{Full name of organism, as used by EuPathDB APIs}

\item{overwrite}{Overwrite existing data?}

\item{dir}{Where to place the data.}

\item{data_provider}{Name of data provider to query (e.g. 'TriTrypDB')}
}
\value{
Dataframe with ....
}
\description{
Returns a mapping of gene ID to ortholog genes
}
\author{
Keith Hughitt
}
