% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{get_eupath_pathway_table}
\alias{get_eupath_pathway_table}
\title{2017/07/08: IN DEVELOPMENT}
\usage{
get_eupath_pathway_table(data_provider, organism)
}
\arguments{
\item{data_provider}{Name of data provider to query (e.g. 'TriTrypDB')}

\item{organism}{Full name of organism, as used by EuPathDB APIs}
}
\value{
Dataframe with gene/pathway mapping
}
\description{
Returns a mapping of gene ID to metabolic pathways (KEGG, LeishCyc, etc.)
}
\author{
Keith Hughitt
}
