% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{get_genelengths}
\alias{get_genelengths}
\title{get_genelengths() Extract gene lengths from a gff file}
\usage{
get_genelengths(gff, ID = "Note")
}
\arguments{
\item{gff}{The file to extract}

\item{ID}{Note the field to cross reference against to extract the genes}
}
\description{
get_genelengths() Extract gene lengths from a gff file
}

