% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_helpers_misc.r
\name{get_git_commit}
\alias{get_git_commit}
\title{Get the current git commit for hpgltools}
\usage{
get_git_commit(gitdir = "~/hpgltools")
}
\arguments{
\item{gitdir}{Directory containing the git repository.}
}
\description{
One might reasonably ask about this function: "Why?"  I invoke this function
at the end of my various knitr documents so that if necessary I can do a
> git reset <commit id> and get back to the exact state of my code.
}
