% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{get_gsvadb_names}
\alias{get_gsvadb_names}
\title{Extract the GeneSets corresponding to the provided name(s).}
\usage{
get_gsvadb_names(sig_data, requests = NULL)
}
\arguments{
\item{sig_data}{The pile of GeneSets, probably from GSVAdata.}

\item{requests}{Character list of sources to keep.}
}
\value{
Whatever GeneSets remain.
}
\description{
Many of the likely GSCs contain far more gene sets than one actually wants to
deal with.  This will subset them according to a the desired 'requests'.
}
