% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{get_inter_txdb}
\alias{get_inter_txdb}
\title{Given an input gff, create intron/intergenic gff files.}
\usage{
get_inter_txdb(
  input,
  intron_gff = "introns.gff",
  intergenic_gff = "intergenic.gff"
)
}
\description{
I was poking around for ways to mix and match gff files and found this:
https://support.bioconductor.org/p/66003/
}
