% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.R
\name{get_microbesonline_taxid}
\alias{get_microbesonline_taxid}
\title{Extract microbesonline taxon IDs without having to click on the weird boxes
at the top of the website.}
\usage{
get_microbesonline_taxid(species = "Acyrthosiphon pisum virus")
}
\arguments{
\item{species}{String to search the set of microbesonline taxa.}
}
\value{
NULL or 1 or more taxon ids.
}
\description{
This should simplify getting material from microbesonline.
}
\examples{
## On those relatively rare occasions when the internet is being
## cranky, these tests will most certainly fail.

## Get the numeric ID from microbesonline for a species of interest.
## When microbesonline is under maintenance/etc this fails.
ecoli_id <- try(get_microbesonline_taxid(species = "S88"))
ecoli_id

if (! "try-error" \%in\% class(ecoli_id)) {
  ## Download a genbank/fasta/etc file for that species.
  ecoli_files <- try(download_microbesonline_files(id = ecoli_id, type = "gbk"))
  ecoli_files
  if (! "try-error" \%in\% class(ecoli_files)) {
    removed <- file.remove(ecoli_files[["gbk"]])
  }

  ## Download a species annotations to a dataframe
  ecoli_annot <- try(load_microbesonline_annotations(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_annot)) {
    dim(ecoli_annot)
    colnames(ecoli_annot)
  }

  ## Get gene ontology data from microbesonline
  ecoli_go <- try(load_microbesonline_go(id = ecoli_id))
  if (! "try-error" \%in\% class(ecoli_go)) {
    dim(ecoli_go)
  }
}
}
\seealso{
[xml2]
}
