% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{get_msigdb_metadata}
\alias{get_msigdb_metadata}
\title{Create a metadata dataframe of msigdb data, this hopefully will be usable to
fill the fData slot of a gsva returned expressionset.}
\usage{
get_msigdb_metadata(
  gsva_result = NULL,
  msig_xml = "msigdb_v6.2.xml",
  wanted_meta = c("ORGANISM", "DESCRIPTION_BRIEF", "AUTHORS", "PMID")
)
}
\arguments{
\item{gsva_result}{Some data from GSVA to modify.}

\item{msig_xml}{msig XML file downloaded from broad.}

\item{wanted_meta}{Choose metadata columns of interest.}
}
\value{
list containing 2 data frames: all metadata from broad, and the set
 matching the sig_data GeneSets.
}
\description{
Create a metadata dataframe of msigdb data, this hopefully will be usable to
fill the fData slot of a gsva returned expressionset.
}
\seealso{
[xml2] [rvest]
}
