% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r, R/testfile.r
\name{get_msigdb_metadata}
\alias{get_msigdb_metadata}
\title{Create a metadata dataframe of msigdb data, this hopefully will be usable to
fill the fData slot of a gsva returned expressionset.}
\usage{
get_msigdb_metadata(
  sig_data = NULL,
  msig_xml = "msigdb_v6.2.xml",
  gsva_result = NULL
)

get_msigdb_metadata(
  sig_data = NULL,
  msig_xml = "msigdb_v6.2.xml",
  gsva_result = NULL
)
}
\arguments{
\item{sig_data}{GeneSetCollection from the broad msigdb.}

\item{msig_xml}{msig XML file downloaded from broad.}

\item{gsva_result}{Some data from GSVA to modify.}
}
\value{
list containing 2 data frames: all metadata from broad, and the set
 matching the sig_data GeneSets.

list containing 2 data frames: all metadata from broad, and the set
 matching the sig_data GeneSets.
}
\description{
Create a metadata dataframe of msigdb data, this hopefully will be usable to
fill the fData slot of a gsva returned expressionset.

Create a metadata dataframe of msigdb data, this hopefully will be usable to
fill the fData slot of a gsva returned expressionset.
}
