% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{get_se_colors}
\alias{get_se_colors}
\title{Get the colors from a summarized experiment.}
\usage{
get_se_colors(se, keep_underscore = TRUE)
}
\arguments{
\item{se}{Input se}

\item{keep_underscore}{Sanitize the columns for underscores?}
}
\description{
Get the colors from a summarized experiment.
}
