% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{get_sig_genes}
\alias{get_sig_genes}
\title{get_sig_genes()  Get a set of up/down genes using the top/bottom n or >/< z scores away from the median.}
\usage{
get_sig_genes(table, n = NULL, z = NULL, fc = NULL, column = "logFC",
  fold = "plusminus")
}
\arguments{
\item{table}{a table from limma/edger/deseq.}

\item{n}{default=NULL  a rank-order top/bottom number of genes to take.}

\item{z}{default=NULL  a number of z-scores >/< the median to take.}

\item{fc}{default=NULL  a number of fold-changes to take}

\item{column}{default='logFC'  a column to use to distinguish top/bottom}

\item{fold}{default='plusminus'  an identifier reminding how to get the bottom portion of a fold-change (plusminus says to get the negative of the positive, otherwise 1/positive is taken).}
}
\value{
a list of up/down genes
}
\description{
get_sig_genes()  Get a set of up/down genes using the top/bottom n or >/< z scores away from the median.
}

