% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{get_snp_sets}
\alias{get_snp_sets}
\title{Create all possible sets of variants by sample (types).}
\usage{
get_snp_sets(snp_expt, factor = "pathogenstrain", limit = 1,
  do_save = FALSE, savefile = "variants")
}
\arguments{
\item{snp_expt}{The result of count_expt_snps()}

\item{factor}{Experimental factor to use for cutting and splicing the data.}

\item{limit}{Minimum median number of hits / factor to define a position as
a hit.}

\item{do_save}{Save the result?}

\item{savefile}{Prefix for a savefile if one chooses to save the result.}
}
\value{
A funky list by chromosome containing:  'medians', the median number
of hits / position by sample type; 'possibilities', the;
'intersections', the groupings as detected by Vennerable;
'chr_data', the raw data; 'set_names', a character list of the actual names
of the groupings; 'invert_names', the opposite of set_names which is to say
the names of groups which do \emph{not} include samples x,y,z; 'density', a list
of snp densities with respect to chromosomes.  Note that this last one is
approximate as I just calculate with the largest chromosome position
number, not the explicit number of nucleotides in the chromosome.
}
\description{
I like this function.  It generates an exhaustive catalog of the snps by
chromosome for all the various categories as defined by factor.
}
