% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_ontology_topgo.r
\name{getEdgeWeights}
\alias{getEdgeWeights}
\title{Plot the ontology DAG.}
\usage{
getEdgeWeights(graph)
}
\arguments{
\item{graph}{Graph from topGO}
}
\value{
Weights!
}
\description{
This function was stolen from topgo in order to figure out where it was failing.
}
