% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{gff2df}
\alias{gff2df}
\title{Try to make import.gff a little more robust
I acquire (hopefully) valid gff3 files from various sources:
yeastgenome.org, microbesonline, tritrypdb, ucsc, ncbi.
To my eyes, they all look like reasonably good gff3 files, but
some of them must be loaded with import.gff2, import.gff3, etc.
That is super annoying.
Also, I pretty much always just do as.data.frame() when I get something
valid from rtracklayer, so this does that for me, I have another function
which returns the iranges etc.}
\usage{
gff2df(gff, type = NULL)
}
\arguments{
\item{gff}{a gff filename}

\item{type}{subset the gff file for entries of a specific type}
}
\value{
a df!
}
\description{
This function wraps import.gff/import.gff3/import.gff2 calls in try()
Because sometimes those functions fail in unpredictable ways.
}
\examples{
\dontrun{
funkytown <- gff2df('reference/gff/saccharomyces_cerevsiae.gff.xz')
}
}
\seealso{
\pkg{rtracklayer} \link[rtracklayer]{import.gff} \link[rtracklayer]{import.gff2} \link[rtracklayer]{import.gff3}
}

