% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_gff.R
\name{gff2gr}
\alias{gff2gr}
\title{Rewrite a gff file as a granges with full seqinfo if possible.}
\usage{
gff2gr(gff, type = NULL, type_column = "type")
}
\arguments{
\item{gff}{Input gff file.}

\item{type}{Feature type to extract.}

\item{type_column}{Tag from the gff file to use when extracting the type.}
}
\value{
GRanges represenatation of the gff portion of interest.
}
\description{
Rewrite a gff file as a granges with full seqinfo if possible.
}
\examples{
## Load some annotations from a gff file
example_gff <- system.file("share", "gas.gff", package = "hpgldata")

gas_iranges <- gff2irange(example_gff)
colnames(as.data.frame(gas_iranges))

gas_gr <- gff2gr(example_gff)
head(as.data.frame(gas_gr))

gas_df <- load_gff_annotations(example_gff, id_col = "locus_tag")
dim(gas_df)
}
