% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{gff2irange}
\alias{gff2irange}
\title{Try to make import.gff a little more robust}
\usage{
gff2irange(gff, type = NULL)
}
\arguments{
\item{gff}{a gff filename}

\item{type}{a subset to extract

Essentially gff2df() above, but returns data suitable for getSet()}
}
\value{
an iranges! (useful for getSeq())
}
\description{
Try to make import.gff a little more robust
}
\examples{
\dontrun{
library(BSgenome.Tcruzi.clbrener.all)
tc_clb_all <- BSgenome.Tcruzi.clbrener.all
cds_ranges <- gff2irange('reference/gff/tcruzi_clbrener.gff.xz', type='CDS')
cds_sequences <- Biostrings::getSeq(tc_clb_all, cds_ranges)
}
}
\seealso{
\pkg{rtracklayer} \link{gff2df} \link[Biostrings]{getSeq}
}

