% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gff2irange}
\alias{gff2irange}
\title{gff2irange()  Try to make import.gff a little more robust}
\usage{
gff2irange(gff)
}
\arguments{
\item{gff}{a gff filename

This function wraps import.gff/import.gff3/import.gff2 calls in try()
Because sometimes those functions fail in unpredictable ways.}
}
\value{
an iranges! (useful for getSeq())
}
\description{
gff2irange()  Try to make import.gff a little more robust
}

