% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/123_plot_misc.r
\name{ggplotly_url}
\alias{ggplotly_url}
\title{Add a little logic to ggplotly to simplify adding clicky link.}
\usage{
ggplotly_url(
  plot,
  filename,
  id_column = "id",
  title = NULL,
  url_data = NULL,
  url_column = "url",
  tooltip = "all"
)
}
\arguments{
\item{plot}{Plot generated via ggplot2.}

\item{filename}{filename to save the output html plot.}

\item{id_column}{Column containing the gene IDs.}

\item{title}{Provide a title for the generated html file.}

\item{url_data}{Either a glue() string or column of urls.}

\item{url_column}{Name of the column in the ggplot data containing the URLs.}

\item{tooltip}{Passed to ggplotly().}
}
\value{
plotly with clicky links.
}
\description{
There are some other ease of life improvements I have in a few of my plotly
invocations which I should add here.
}
