% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{godef}
\alias{godef}
\title{Get a go long-form definition from an id}
\usage{
godef(go)
}
\arguments{
\item{id}{A go ID -- this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Get a go long-form definition from an id
}
\examples{
## godef("GO:0032432")
## > GO:0032432
## > "An assembly of actin filaments that are on the same axis but may be oriented with the same or opposite polarities and may be packed with different levels of tightness."
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

