% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{golev}
\alias{golev}
\title{Get a go level approximation from an ID.}
\usage{
golev(go, verbose = FALSE)
}
\arguments{
\item{go}{GO id, this may be a character or list (assuming the elements are goids).}

\item{verbose}{print some information as it recurses.}
}
\value{
Set of numbers corresponding to approximate tree positions of the GO ids.
}
\description{
Sometimes it is useful to know how far up/down the ontology tree a given id resides.  This
attmepts to answer that question.
}
\examples{
\dontrun{
 golev("GO:0032559")
 ## > 3
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}

