% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{golev}
\alias{golev}
\title{Get a go level approximation from an ID}
\usage{
golev(go, verbose = FALSE)
}
\arguments{
\item{id}{A go ID -- this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Get a go level approximation from an ID
}
\examples{
## golev("GO:0032559")
## > 3
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

