% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{golev}
\alias{golev}
\title{Get a go level approximation from an ID.}
\usage{
golev(go, verbose = FALSE)
}
\arguments{
\item{go}{a go ID, this may be a character or list (assuming the elements are goids).}

\item{verbose}{print some information as it recurses.}
}
\value{
Some text
}
\description{
Get a go level approximation from an ID.
}
\examples{
\dontrun{
 golev("GO:0032559")
 ## > 3
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}

