% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{golevel}
\alias{golevel}
\title{Get a go level approximation from a set of IDs
This just wraps golev() in mapply.}
\usage{
golevel(go)
}
\arguments{
\item{id}{a character list of IDs}
}
\value{
Some text
}
\description{
Get a go level approximation from a set of IDs
This just wraps golev() in mapply.
}
\examples{
## golevel(c("GO:0032559", "GO:0000001")
## > 3 4
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

