% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_ontology_shared.r
\name{golevel}
\alias{golevel}
\title{Get a go level approximation from a set of IDs.}
\usage{
golevel(go = c("GO:0032559", "GO:0000001"))
}
\arguments{
\item{go}{Character list of IDs.}
}
\value{
Set pf approximate levels within the onlogy.
}
\description{
This just wraps golev() in mapply.
}
\examples{
\dontrun{
 golevel(c("GO:0032559", "GO:0000001")
 ## > 3 4
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}
