% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{golevel_df}
\alias{golevel_df}
\title{golevel_df()  Extract a dataframe of golevels using getGOLevel() from clusterProfiler.}
\usage{
golevel_df(ont = "MF", savefile = "ontlevel.rda")
}
\arguments{
\item{ont}{default='MF'  the ontology to recurse.}

\item{savefile}{default='ontlevel.rda'  a file to save the results for future lookups.}
}
\value{
golevels  a dataframe of goids<->highest level
}
\description{
This function is way faster than my previous iterative golevel function.
That is not to say it is very fast, so it saves the result to ontlevel.rda for future lookups.
}

