% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{golevel_df}
\alias{golevel_df}
\title{Extract a dataframe of golevels using getGOLevel() from clusterProfiler.}
\usage{
golevel_df(ont = "MF", savefile = "ontlevel.rda")
}
\arguments{
\item{ont}{the ontology to recurse.}

\item{savefile}{a file to save the results for future lookups.}
}
\value{
golevels  a dataframe of goids<->highest level
}
\description{
This function is way faster than my previous iterative golevel function.
That is not to say it is very fast, so it saves the result to ontlevel.rda
for future lookups.
}
\seealso{
\pkg{clusterProfiler}
}
