% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{goont}
\alias{goont}
\title{Get a go ontology name from an ID.}
\usage{
goont(go)
}
\arguments{
\item{go}{a go ID, this may be a character or list (assuming the elements are goids).}
}
\value{
Some text
}
\description{
Get a go ontology name from an ID.
}
\examples{
\dontrun{
 goont(c("GO:0032432", "GO:0032433"))
## > GO:0032432 GO:0032433
## > "CC" "CC"
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}

