% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{goont}
\alias{goont}
\title{Get a go ontology name from an ID.}
\usage{
goont(go = c("GO:0032432", "GO:0032433"))
}
\arguments{
\item{go}{GO id, this may be a character or list (assuming the elements are goids).}
}
\value{
The set of ontology IDs associated with the GO ids, thus 'MF' or 'BP' or 'CC'.
}
\description{
Get a go ontology name from an ID.
}
\examples{
\dontrun{
 goont(c("GO:0032432", "GO:0032433"))
 ## > GO:0032432 GO:0032433
 ## > "CC" "CC"
}
}
\seealso{
[AnnotationDbi] [GO.db]
}
