% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{gosec}
\alias{gosec}
\title{Get a go secondary ID from an id}
\usage{
gosec(go)
}
\arguments{
\item{id}{A go ID -- this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Unfortunately, GOTERM's returns for secondary IDs are not consistent, so this function
has to have a whole bunch of logic to handle the various outputs.
}
\examples{
## gosec("GO:0032432")
## > GO:0032432
## > "GO:0000141" "GO:0030482"
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

