% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.R
\name{goseq_pval_plots}
\alias{goseq_pval_plots}
\title{Make a pvalue plot from goseq data}
\usage{
goseq_pval_plots(goterms, wrapped_width = 20, cutoff = 0.1, n = 10,
  mincat = 10, level = NULL)
}
\arguments{
\item{goterms}{some data from goseq!}

\item{wrapped_width}{20 the number of characters before wrapping to help legibility}

\item{cutoff}{pvalue cutoff for the plot}

\item{n}{10 how many groups to include}
}
\value{
plots!
}
\description{
Make a pvalue plot from goseq data
}
\seealso{
\code{\link{goseq}} \code{\link{clusterProfiler}} \code{\link{pval_plot}}
}

