% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_goseq.R
\name{goseq_pval_plots}
\alias{goseq_pval_plots}
\title{Make a pvalue plot from goseq data}
\usage{
goseq_pval_plots(goterms, wrapped_width = 20, cutoff = 0.1, n = 10)
}
\arguments{
\item{topgo_data}{some data from goseq!}
}
\value{
a plot!
}
\description{
Make a pvalue plot from goseq data
}
\seealso{
\code{\link{goseq}}
}

