% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.r
\name{goseq_pval_plots}
\alias{goseq_pval_plots}
\title{Make a pvalue plot from goseq data}
\usage{
goseq_pval_plots(goterms, wrapped_width = 20, cutoff = 0.1, n = 10,
  mincat = 10, level = NULL)
}
\arguments{
\item{goterms}{some data from goseq!}

\item{wrapped_width}{default=20 the number of characters before wrapping to help legibility}

\item{cutoff}{default=0.1  pvalue cutoff for the plot}

\item{n}{default=10   how many groups to include}

\item{mincat}{default=10  minimum size of the category}

\item{level}{default=NULL  levels of the ontology tree to use}
}
\value{
plots!
}
\description{
Make a pvalue plot from goseq data
}
\seealso{
\link[goseq]{goseq} \pkg{clusterProfiler} \code{\link{pval_plot}}
}

