% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq-test.r, R/ontology_goseq.r
\name{goseq_table}
\alias{goseq_table}
\title{Enhance the goseq table of gene ontology information.}
\usage{
goseq_table(df, file = NULL)

goseq_table(df, file = NULL)
}
\arguments{
\item{df}{Dataframe of ontology information.  This is intended to be the output from goseq
including information like numbers/category, GOids, etc.  It requires a column 'category'
which contains: GO:000001 and such.}

\item{file}{Csv file to which to write the table.}

\item{df}{Dataframe of ontology information.  This is intended to be the output from goseq
including information like numbers/category, GOids, etc.  It requires a column 'category'
which contains: GO:000001 and such.}

\item{file}{Csv file to which to write the table.}
}
\value{
Ontology table with annotation information included.

Ontology table with annotation information included.
}
\description{
While goseq has some nice functionality, the table of outputs it provides is somewhat lacking.
This attempts to increase that with some extra helpful data like ontology categories,
definitions, etc.

While goseq has some nice functionality, the table of outputs it provides is somewhat lacking.
This attempts to increase that with some extra helpful data like ontology categories,
definitions, etc.
}
\examples{
\dontrun{
 annotated_go = goseq_table(go_ids)
 head(annotated_go, n=1)
## >        category numDEInCat numInCat over_represented_pvalue
## > 571  GO:0006364          9       26            4.655108e-08
## >      under_represented_pvalue       qvalue ontology
## > 571                 1.0000000 6.731286e-05       BP
## >                                term
## > 571                 rRNA processing
## >                               synonym
## > 571        "35S primary transcript processing, GO:0006365"
## >        secondary    definition
## > 571    GO:0006365   Any process involved in the conversion of a primary ribosomal
##          RNA (rRNA) transcript into one or more mature rRNA molecules.
}
\dontrun{
 annotated_go = goseq_table(go_ids)
 head(annotated_go, n=1)
## >        category numDEInCat numInCat over_represented_pvalue
## > 571  GO:0006364          9       26            4.655108e-08
## >      under_represented_pvalue       qvalue ontology
## > 571                 1.0000000 6.731286e-05       BP
## >                                term
## > 571                 rRNA processing
## >                               synonym
## > 571        "35S primary transcript processing, GO:0006365"
## >        secondary    definition
## > 571    GO:0006365   Any process involved in the conversion of a primary ribosomal
##          RNA (rRNA) transcript into one or more mature rRNA molecules.
}
}
\seealso{
\pkg{goseq}

\pkg{goseq}
}

