% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.r
\name{goseq_trees}
\alias{goseq_trees}
\title{Make fun trees a la topgo from goseq data.}
\usage{
goseq_trees(de_genes, godata, goid_map = "reference/go/id2go.map",
  score_limit = 0.01, goids_df = NULL, overwrite = FALSE,
  selector = "topDiffGenes", pval_column = "adj.P.Val")
}
\arguments{
\item{de_genes}{some differentially expressed genes}

\item{godata}{data from goseq}

\item{goid_map}{default='reference/go/id2go.map'  file to save go id mapping}

\item{score_limit}{default=0.01  score limit for the coloring}

\item{goids_df}{default=NULL  a mapping of IDs to GO in the Ramigo expected format}

\item{overwrite}{default=FALSE  overwrite the trees}

\item{selector}{default='topDiffGenes'  a function for choosing genes}

\item{pval_column}{default='adj.P.Val' column to acquire pvalues}
}
\value{
a plot!
}
\description{
Make fun trees a la topgo from goseq data.
}
\seealso{
\pkg{Ramigo}
}

