% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{goseq_trees}
\alias{goseq_trees}
\title{Make fun trees a la topgo from goseq data.}
\usage{
goseq_trees(de_genes, godata, goid_map = "reference/go/id2go.map",
  score_limit = 0.01, goids_df = NULL, overwrite = FALSE,
  selector = "topDiffGenes", pval_column = "adj.P.Val")
}
\arguments{
\item{de_genes}{some differentially expressed genes}

\item{godata}{data from goseq}

\item{goids}{a mapping of IDs to GO in the Ramigo expected format}

\item{sigforall}{Print significance on all nodes?}
}
\value{
a plot!
}
\description{
Make fun trees a la topgo from goseq data.
}
\seealso{
\code{\link{Ramigo}}
}

