% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{gostats_kegg}
\alias{gostats_kegg}
\title{Use gostats() against kegg pathways.}
\usage{
gostats_kegg(organism = "Homo sapiens", pathdb = "org.Hs.egPATH",
  godb = "org.Hs.egGO")
}
\arguments{
\item{organism}{The organism used to make the KEGG frame, human readable no taxonomic.}

\item{pathdb}{Name of the pathway database for this organism.}

\item{godb}{Name of the ontology database for this organism.}
}
\value{
Results from hyperGTest using the KEGG pathways.
}
\description{
This sets up a GSEABase analysis using KEGG pathways rather than gene ontologies.
Does this even work?  I don't think I have ever tested it yet.
oh, it sort of does, maybe if I export it I will rembmer it.
}

