% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gostats.r
\name{gostats_pval_plots}
\alias{gostats_pval_plots}
\title{Make a pvalue plot similar to that from clusterprofiler from gostats data}
\usage{
gostats_pval_plots(gs_result, wrapped_width = 20, cutoff = 0.1, n = 12,
  group_minsize = 5)
}
\arguments{
\item{wrapped_width}{how big to make the text so that it is legible}

\item{cutoff}{0.1 what is the maximum pvalue allowed}

\item{n}{10 how many groups to include in the plot}

\item{mf_over}{molecular function data overrepresented}

\item{bp_over}{biological process data overrepresented}

\item{cc_over}{cellular component data overrepresented}

\item{mf_under}{molecular function data underrepresented}

\item{bp_under}{biological process data underrepresented}

\item{cc_under}{cellular component data underrepresented}
}
\value{
plots!
}
\description{
clusterprofiler provides beautiful plots describing significantly overrepresented categories.
This function attempts to expand the repetoire of data available to them to include data from gostats.
}
\details{
The pval_plot function upon which this is based now has a bunch of new helpers now
that I understand how the ontology trees work better, this should take advantage of that, but
currently does not.
}
\seealso{
\code{\link{clusterProfiler}} \code{\link{pval_plot}}
}

