% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gostats.r
\name{gostats_pval_plots}
\alias{gostats_pval_plots}
\title{Make a pvalue plot similar to that from clusterprofiler from gostats data}
\usage{
gostats_pval_plots(gs_result, wrapped_width = 20, cutoff = 0.1, n = 12,
  group_minsize = 5)
}
\arguments{
\item{gs_result}{ontology search results}

\item{wrapped_width}{default=20  how big to make the text so that it is legible}

\item{cutoff}{default=0.1  what is the maximum pvalue allowed}

\item{n}{default=12  how many groups to include in the plot}

\item{group_minsize}{default=5  minimum group size before inclusion}
}
\value{
plots!
}
\description{
clusterprofiler provides beautiful plots describing significantly overrepresented categories.
This function attempts to expand the repetoire of data available to them to include data from gostats.
}
\details{
The pval_plot function upon which this is based now has a bunch of new helpers now
that I understand how the ontology trees work better, this should take advantage of that, but
currently does not.
}
\seealso{
\pkg{clusterProfiler} \link{pval_plot}
}

