% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.r
\name{gostats_trees}
\alias{gostats_trees}
\title{Take gostats data and print it on a tree as topGO does.}
\usage{
gostats_trees(
  gostats_result,
  goid_map = "id2go.map",
  score_limit = 0.01,
  overwrite = FALSE,
  selector = "topDiffGenes",
  pval_column = "adj.P.Val"
)
}
\arguments{
\item{gostats_result}{Return from simple_gostats().}

\item{goid_map}{Mapping of IDs to GO in the Ramigo expected format.}

\item{score_limit}{Maximum score to include as 'significant'.}

\item{overwrite}{Overwrite the goid_map?}

\item{selector}{Function to choose differentially expressed genes in the data.}

\item{pval_column}{Column in the data to be used to extract pvalue scores.}
}
\value{
plots! Trees! oh my!
}
\description{
This shoehorns gostats data into a format acceptable by topgo and uses it to
print pretty ontology trees showing the over represented ontologies.
}
\seealso{
\pkg{topGO} \pkg{gostats}
}
