% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_gostats.R
\name{gostats_trees}
\alias{gostats_trees}
\title{Make fun trees a la topgo from goseq data.}
\usage{
gostats_trees(de_genes, mf_over, bp_over, cc_over, mf_under, bp_under, cc_under,
  goid_map = "reference/go/id2go.map", score_limit = 0.01,
  goids_df = NULL, overwrite = FALSE, selector = "topDiffGenes",
  pval_column = "adj.P.Val")
}
\arguments{
\item{de_genes}{some differentially expressed genes}

\item{goid_map}{a mapping of IDs to GO in the Ramigo expected format}

\item{score_limit}{maximum score to include as 'significant'}

\item{goids_df}{a dataframe of available goids (used to generate goid_map)}

\item{overwrite}{overwrite the goid_map?}

\item{selector}{a function to choose differentially expressed genes in the data}

\item{pval_column}{a column in the data to be used to extract pvalue scores}

\item{mf_over/bp_over/cc_over/mf_under/bp_under/cc_under}{over/under expression data}
}
\value{
plots! Trees! oh my!
}
\description{
Make fun trees a la topgo from goseq data.
}
\seealso{
\code{\link{topGO}}
}

