% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{gosyn}
\alias{gosyn}
\title{Get a go synonym from an ID.}
\usage{
gosyn(go = "GO:0000001")
}
\arguments{
\item{go}{GO id, this may be a character or list(assuming the elements are goids).}
}
\value{
Some text providing the synonyms for the given id(s).
}
\description{
I think I will need to do similar parsing of the output for this function as
per gosec() In some cases this also returns stuff like c("some text",
"GO:someID") versus "some other text"  versus NULL versus NA.
This function just goes a mapply(gosn, go).
}
\examples{
\dontrun{
 text =  gosyn("GO:0000001")
 text
 ## > GO:000001
 ## > "mitochondrial inheritance"
}
}
\seealso{
[AnnotationDbi] [GO.db]
}
