% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{gosyn}
\alias{gosyn}
\title{gosyn()  Get a go synonym from an ID.}
\usage{
gosyn(go)
}
\arguments{
\item{go}{a go ID, this may be a character or list(assuming the elements are goids).}
}
\value{
Some text
}
\description{
I think I will need to do similar parsing of the output for this function as per gosec()
In some cases this also returns stuff like c("some text", "GO:someID")
versus "some other text"  versus NULL versus NA
}
\details{
This function just goes a mapply(gosn, go).
}
\examples{
## text =  gosyn("GO:0000001")
## text
## > GO:000001
## > "mitochondrial inheritance"
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

