% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{goterm}
\alias{goterm}
\title{Get a go term from ID}
\usage{
goterm(go = "GO:0032559")
}
\arguments{
\item{id}{A go ID -- this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Get a go term from ID
}
\examples{
## goterm("GO:0032559")
## > GO:0032559
## > "adenyl ribonucleotide binding"
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

