% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_ontology_shared.r
\name{goterm}
\alias{goterm}
\title{Get a go term from ID.}
\usage{
goterm(go = "GO:0032559")
}
\arguments{
\item{go}{GO id or a list thereof, this may be a character or list(assuming
the elements, not names, are goids).}
}
\value{
Some text containing the terms associated with GO id(s).
}
\description{
Get a go term from ID.
}
\examples{
\dontrun{
 goterm("GO:0032559")
 ## > GO:0032559
 ## > "adenyl ribonucleotide binding"
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}
