% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{gotest}
\alias{gotest}
\title{Test GO ids to see if they are useful.}
\usage{
gotest(go)
}
\arguments{
\item{go}{go IDs as characters.}
}
\value{
Some text
}
\description{
This just wraps gotst in mapply.
}
\examples{
\dontrun{
 gotest("GO:0032559")
## > 1
 gotest("GO:0923429034823904")
## > 0
}
}
\seealso{
\pkg{GOTermsAnnDbBimap}
}

