% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology.R
\name{gotst}
\alias{gotst}
\title{Test a GO id to see if it is useful by my arbitrary definition of 'useful'}
\usage{
gotst(go)
}
\arguments{
\item{id}{A go ID -- this may be a character or list(assuming the elements, not names, are goids)}
}
\value{
Some text
}
\description{
Test a GO id to see if it is useful by my arbitrary definition of 'useful'
}
\examples{
## gotest("GO:0032559")
## > 1
## gotest("GO:0923429034823904")
## > 0
}
\seealso{
\code{\link{GOTERM}}, \code{\link{GO.db}},
}

