% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpgltools.r
\name{\%:::\%}
\alias{\%:::\%}
\title{R CMD check is super annoying about :::.}
\usage{
pkg \%:::\% fun
}
\arguments{
\item{pkg}{on the left hand side}

\item{fun}{on the right hand side}
}
\description{
In a fit of pique, I did a google search to see if anyone else has been
annoyed in the same was as I.  I was in no way surprised to see that Yihui
Xie was, and in his email to r-devel in 2013 he proposed a game of
hide-and-seek; a game which I am repeating here.
}
\details{
This just implements ::: as an infix operator that will not trip check.
}
