% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{graph_metrics}
\alias{graph_metrics}
\title{Make lots of graphs!}
\usage{
graph_metrics(expt, cormethod = "pearson", distmethod = "euclidean",
  title_suffix = NULL, qq = NULL, ma = NULL, ...)
}
\arguments{
\item{expt}{an expt to process}

\item{cormethod}{the correlation test for heatmaps.}

\item{distmethod}{define the distance metric for heatmaps.}

\item{title_suffix}{text to add to the titles of the plots.}

\item{qq}{include qq plots}

\item{ma}{include pairwise ma plots}

\item{...}{extra parameters optionally fed to the various plots}
}
\value{
a loooong list of plots including the following:
  nonzero = a ggplot2 plot of the non-zero genes vs library size
  libsize = a ggplot2 bar plot of the library sizes
  boxplot = a ggplot2 boxplot of the raw data
  corheat = a recordPlot()ed pairwise correlation heatmap of the raw data
  smc = a recordPlot()ed view of the standard median pairwise correlation of the raw data
  disheat = a recordPlot()ed pairwise euclidean distance heatmap of the raw data
  smd = a recordPlot()ed view of the standard median pairwise distance of the raw data
  pcaplot = a recordPlot()ed PCA plot of the raw samples
  pcatable = a table describing the relative contribution of condition/batch of the raw data
  pcares =  a table describing the relative contribution of condition/batch of the raw data
  pcavar = a table describing the variance of the raw data
  qq = a recordPlotted() view comparing the quantile/quantiles between the mean of all data and every raw sample
  density = a ggplot2 view of the density of each raw sample (this is complementary but more fun than a boxplot)
}
\description{
Plot out a set of metrics describing the state of an experiment
including library sizes, # non-zero genes, heatmaps, boxplots,
density plots, pca plots, standard median distance/correlation, and
qq plots.
}
\examples{
\dontrun{
toomany_plots <- graph_metrics(expt)
toomany_plots$pcaplot
norm <- normalize_expt(expt, convert="cpm", batch=TRUE, filter_low=TRUE, transform="log2", norm="rle")
holy_asscrackers <- graph_metrics(norm, qq=TRUE, ma=TRUE)
## good luck, you are going to be waiting a while for the ma plots to print!
}
}
\seealso{
\pkg{Biobase} \pkg{ggplot2} \pkg{grDevices} \pkg{gplots}
\link[Biobase]{exprs} \link{hpgl_norm} \link{hpgl_nonzero} \link{hpgl_libsize}
\link{hpgl_boxplot} \link{hpgl_corheat} \link{hpgl_smc} \link{hpgl_disheat}
\link{hpgl_smd} \link{hpgl_pca} \link{hpgl_qq_all} \link{hpgl_pairwise_ma}
}

