% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{graph_metrics}
\alias{graph_metrics}
\title{graph_metrics() Make lots of graphs!}
\usage{
graph_metrics(expt, cormethod = "pearson", distmethod = "euclidean",
  title_suffix = NULL, scale = "raw", sink = FALSE, ...)
}
\arguments{
\item{cormethod}{define the correlation test for heatmaps.
Defaults to pearson (Available: pearson, spearman, kendal, robust)}

\item{distmethod}{define the distance metric for heatmaps.
Defaults to euclidean (Lots are available, I don't understand them.)}

\item{...}{extra parameters optionally fed to the various plots}

\item{data}{an expt/df/exprs experiment containing the data, design, and colors}

\item{norm_type}{normalization strategy for the data.  Defaults to
quantile.}

\item{filter}{whether to log2/10 filter the data.  Defaults to
log2.}

\item{out_type}{whether to cpm/rpkm the data.  Defaults to cpm.}

\item{filter_low}{whether to low-count filter the data.  Defaults
to TRUE.}
}
\value{
a loooong list of plots including the following:
  nonzero = a ggplot2 plot of the non-zero genes vs library size
  libsize = a ggplot2 bar plot of the library sizes
  boxplot = a ggplot2 boxplot of the raw data
  corheat = a recordPlot()ed pairwise correlation heatmap of the raw data
  smc = a recordPlot()ed view of the standard median pairwise correlation of the raw data
  disheat = a recordPlot()ed pairwise euclidean distance heatmap of the raw data
  smd = a recordPlot()ed view of the standard median pairwise distance of the raw data
  pcaplot = a recordPlot()ed PCA plot of the raw samples
  pcatable = a table describing the relative contribution of condition/batch of the raw data
  pcares =  a table describing the relative contribution of condition/batch of the raw data
  pcavar = a table describing the variance of the raw data
  qq = a recordPlotted() view comparing the quantile/quantiles between the mean of all data and every raw sample
  density = a ggplot2 view of the density of each raw sample (this is complementary but more fun than a boxplot)
}
\description{
Plot out a set of metrics describing the state of an experiment
including library sizes, # non-zero genes, heatmaps, boxplots,
density plots, pca plots, standard median distance/correlation, and
qq plots.
}
\examples{
## toomany_plots = graph_metrics(expt)
## testnorm = graph_metrics(expt, norm_type="tmm", filter="log2", out_type="rpkm", cormethod="robust")
## haha sucker, you are going to be waiting a while!
}
\seealso{
\code{\link{exprs}}, \code{\link{hpgl_norm}},
\code{\link{graph_nonzero}}, \code{\link{hpgl_libsize}},
\code{\link{hpgl_boxplot}}, \code{\link{hpgl_corheat}},
\code{\link{hpgl_smc}}, \code{\link{hpgl_disheat}},
\code{\link{hpgl_smd}}, \code{\link{hpgl_pca}},
\code{\link{replayPlot}}, \code{\link{recordPlot}}
}

