% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{graph_metrics}
\alias{graph_metrics}
\title{graph_metrics()  Make lots of graphs!}
\usage{
graph_metrics(expt, cormethod = "pearson", distmethod = "euclidean",
  title_suffix = NULL, scale = "raw", sink = FALSE, ...)
}
\arguments{
\item{expt}{an expt to process}

\item{cormethod}{default='pearson'  the correlation test for heatmaps.}

\item{distmethod}{default='euclidean'  define the distance metric for heatmaps.}

\item{title_suffix}{default=NULL  text to add to the titles of the plots.}

\item{scale}{default='raw'  scale for the axes, sometimes useful to set as log2.}

\item{sink}{default=FALSE  add in the kitchen sink?  This includes all pairwise ma plots and qq plots and takes potentially forever.}

\item{...}{extra parameters optionally fed to the various plots}
}
\value{
a loooong list of plots including the following:
  nonzero = a ggplot2 plot of the non-zero genes vs library size
  libsize = a ggplot2 bar plot of the library sizes
  boxplot = a ggplot2 boxplot of the raw data
  corheat = a recordPlot()ed pairwise correlation heatmap of the raw data
  smc = a recordPlot()ed view of the standard median pairwise correlation of the raw data
  disheat = a recordPlot()ed pairwise euclidean distance heatmap of the raw data
  smd = a recordPlot()ed view of the standard median pairwise distance of the raw data
  pcaplot = a recordPlot()ed PCA plot of the raw samples
  pcatable = a table describing the relative contribution of condition/batch of the raw data
  pcares =  a table describing the relative contribution of condition/batch of the raw data
  pcavar = a table describing the variance of the raw data
  qq = a recordPlotted() view comparing the quantile/quantiles between the mean of all data and every raw sample
  density = a ggplot2 view of the density of each raw sample (this is complementary but more fun than a boxplot)
}
\description{
Plot out a set of metrics describing the state of an experiment
including library sizes, # non-zero genes, heatmaps, boxplots,
density plots, pca plots, standard median distance/correlation, and
qq plots.
}
\examples{
## toomany_plots = graph_metrics(expt)
## testnorm = graph_metrics(expt, norm_type="tmm", filter="log2", out_type="rpkm", cormethod="robust")
## haha sucker, you are going to be waiting a while!
}
\seealso{
\code{\link{exprs}}, \code{\link{hpgl_norm}},
\code{\link{graph_nonzero}}, \code{\link{hpgl_libsize}},
\code{\link{hpgl_boxplot}}, \code{\link{hpgl_corheat}},
\code{\link{hpgl_smc}}, \code{\link{hpgl_disheat}},
\code{\link{hpgl_smd}}, \code{\link{hpgl_pca}},
\code{\link{replayPlot}}, \code{\link{recordPlot}}
}

