% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.r
\name{graph_metrics}
\alias{graph_metrics}
\title{Make lots of graphs!}
\usage{
graph_metrics(expt, cormethod = "pearson", distmethod = "euclidean",
  title_suffix = NULL, qq = FALSE, ma = FALSE, gene_heat = FALSE,
  ...)
}
\arguments{
\item{expt}{an expt to process}

\item{cormethod}{the correlation test for heatmaps.}

\item{distmethod}{define the distance metric for heatmaps.}

\item{title_suffix}{text to add to the titles of the plots.}

\item{qq}{include qq plots?}

\item{ma}{include pairwise ma plots?}

\item{gene_heat}{Include a heatmap of the gene expression data?}

\item{...}{extra parameters optionally fed to the various plots}
}
\value{
a loooong list of plots including the following:
\enumerate{
  \item nonzero = a ggplot2 plot of the non-zero genes vs library size
  \item libsize = a ggplot2 bar plot of the library sizes
  \item boxplot = a ggplot2 boxplot of the raw data
  \item corheat = a recordPlot()ed pairwise correlation heatmap of the raw data
  \item smc = a recordPlot()ed view of the standard median pairwise correlation of the raw data
  \item disheat = a recordPlot()ed pairwise euclidean distance heatmap of the raw data
  \item smd = a recordPlot()ed view of the standard median pairwise distance of the raw data
  \item pcaplot = a recordPlot()ed PCA plot of the raw samples
  \item pcatable = a table describing the relative contribution of condition/batch of the raw data
  \item pcares =  a table describing the relative contribution of condition/batch of the raw data
  \item pcavar = a table describing the variance of the raw data
  \item qq = a recordPlotted() view comparing the quantile/quantiles between
     the mean of all data and every raw sample
  \item density = a ggplot2 view of the density of each raw sample (this is
     complementary but more fun than a boxplot)
}
}
\description{
Plot out a set of metrics describing the state of an experiment
including library sizes, # non-zero genes, heatmaps, boxplots,
density plots, pca plots, standard median distance/correlation, and
qq plots.
}
\examples{
\dontrun{
 toomany_plots <- graph_metrics(expt)
 toomany_plots$pcaplot
 norm <- normalize_expt(expt, convert="cpm", batch=TRUE, filter_low=TRUE,
                        transform="log2", norm="rle")
 holy_asscrackers <- graph_metrics(norm, qq=TRUE, ma=TRUE)
}
}
\seealso{
\pkg{Biobase} \pkg{ggplot2} \pkg{grDevices} \pkg{gplots}
  \code{\link[Biobase]{exprs}} \code{\link{hpgl_norm}}
  \code{\link{plot_nonzero}} \code{\link{plot_libsize}}
  \code{\link{plot_boxplot}} \code{\link{plot_corheat}} \code{\link{plot_sm}}
  \code{\link{plot_disheat}} \code{\link{plot_pca}} \code{\link{plot_qq_all}}
  \code{\link{plot_pairwise_ma}}
}
