% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{graph_nonzero}
\alias{graph_nonzero}
\title{Make a ggplot graph of the number of non-zero genes by sample.  Made by Ramzi Temanni.}
\usage{
graph_nonzero(df = NULL, design = NULL, colors = NULL, expt = NULL,
  title = NULL, ...)
}
\arguments{
\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}

\item{colors}{a color scheme}

\item{expt}{an expt set of samples}
}
\value{
a ggplot2 plot of the number of non-zero genes with respect to each library's CPM
}
\description{
Make a ggplot graph of the number of non-zero genes by sample.  Made by Ramzi Temanni.
}
\examples{
## nonzero_plot = graph_nonzero(expt=expt)
## nonzero_plot  ## ooo pretty
}
\seealso{
\code{\link{geom_point}}, \code{\link{geom_dl}}
}

