% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r, R/testfile.r
\name{gsva_likelihoods}
\alias{gsva_likelihoods}
\title{Score the results from gsva().}
\usage{
gsva_likelihoods(
  gsva_result,
  score = NULL,
  category = NULL,
  factor = NULL,
  sample = NULL,
  factor_column = "condition",
  method = "mean",
  label_size = NULL,
  col_margin = 6,
  row_margin = 12,
  cutoff = 0.05
)

gsva_likelihoods(
  gsva_result,
  score = NULL,
  category = NULL,
  factor = NULL,
  sample = NULL,
  factor_column = "condition",
  method = "mean",
  label_size = NULL,
  col_margin = 6,
  row_margin = 12,
  cutoff = 0.05
)
}
\arguments{
\item{gsva_result}{Input result from simple_gsva()}

\item{score}{What type of scoring to perform, against a value, column, row?}

\item{category}{What category to use as baseline?}

\item{factor}{Which experimental factor to compare against?}

\item{sample}{Which sample to compare against?}

\item{factor_column}{When comparing against an experimental factor, which design
column to use to find it?}

\item{method}{mean or median when when bringing together values?}
}
\value{
The scores according to the provided category, factor, sample, or
 score(s).

The scores according to the provided category, factor, sample, or
 score(s).
}
\description{
Yeah, this is a bit meta, but the scores from gsva seem a bit meaningless to
me, so I decided to look at the distribution of observed scores in some of my
data; I quickly realized that they follow a nicely normal distribution.
Therefore, I thought to calculate some scores of gsva() using that
information.

Yeah, this is a bit meta, but the scores from gsva seem a bit meaningless to
me, so I decided to look at the distribution of observed scores in some of my
data; I quickly realized that they follow a nicely normal distribution.
Therefore, I thought to calculate some scores of gsva() using that
information.
}
\details{
The nicest thing in this, I think, is that it provides its scoring metric(s)
according to a few different possibilities, including:
  * the mean of samples found in an experimental factor
  * All provided scores against the distribution of observed scores as
    z-scores.
  * A single score against all scores.
  * Rows (gene sets) against the set of all gene sets.

The nicest thing in this, I think, is that it provides its scoring metric(s)
according to a few different possibilities, including:
  * the mean of samples found in an experimental factor
  * All provided scores against the distribution of observed scores as
    z-scores.
  * A single score against all scores.
  * Rows (gene sets) against the set of all gene sets.
}
