% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{guess_num_surrogates}
\alias{guess_num_surrogates}
\title{Use methods from sva/RUVseq to guesstimate the number of surrogates.}
\usage{
guess_num_surrogates(
  design,
  linear_mtrx,
  log2_mtrx,
  model_svs = "svaseq",
  conditional_fstring = "~ condition",
  guess_type = "be"
)
}
\arguments{
\item{design}{Dataframe describing the experimental design}

\item{linear_mtrx}{linear scale matrix of the input data}

\item{log2_mtrx}{Log2 scale matrix of the input data}

\item{model_svs}{Method used to model the surrogate variables.}

\item{conditional_fstring}{Formula string describing the variable
of interest.}

\item{guess_type}{Method to guess at the number of surrogate variables.}
}
\description{
Use methods from sva/RUVseq to guesstimate the number of surrogates.
}
