% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040_annotation_orgdb.r
\name{guess_orgdb_keytype}
\alias{guess_orgdb_keytype}
\title{Iterate over keytypes looking for matches against a set of IDs.}
\usage{
guess_orgdb_keytype(ids, orgdb)
}
\arguments{
\item{ids}{Set of gene IDs to seek.}

\item{orgdb}{Orgdb instance to iterate through.}
}
\value{
Likely keytype which provides the desired IDs.
}
\description{
Sometimes, one does not know what the correct keytype is for a given set of
IDs.  This will hopefully find them.
}
\examples{
 ids <- c("Dm.9", "Dm.2294", "Dm.4971")
 keytype_guess <- guess_orgdb_keytype(ids, orgdb)
 keytype_guess
}
