% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap.3}
\alias{heatmap.3}
\title{slight modification of heatmap.2}
\usage{
heatmap.3(x, Rowv = TRUE, Colv = if (symm) "Rowv" else TRUE,
  distfun = dist, hclustfun = hclust, dendrogram = c("both", "row",
  "column", "none"), reorderfun = function(d, w) reorder(d, w),
  symm = FALSE, scale = c("none", "row", "column"), na.rm = TRUE,
  revC = identical(Colv, "Rowv"), add.expr, breaks, symbreaks = min(x < 0,
  na.rm = TRUE) || scale != "none", col = "heat.colors", colsep, rowsep,
  sepcolor = "white", sepwidth = c(0.05, 0.05), cellnote, notecex = 1,
  notecol = "cyan", na.color = par("bg"), trace = c("column", "row",
  "both", "none"), tracecol = "cyan", hline = median(breaks),
  vline = median(breaks), linecol = tracecol, margins = c(5, 5),
  ColSideColors, RowSideColors, cexRow = 0.2 + 1/log10(nr), cexCol = 0.2 +
  1/log10(nc), labRow = NULL, labCol = NULL, srtRow = NULL,
  srtCol = NULL, adjRow = c(0, NA), adjCol = c(NA, 0), offsetRow = 0.5,
  offsetCol = 0.5, key = TRUE, keysize = 1.5,
  density.info = c("histogram", "density", "none"), denscol = tracecol,
  symkey = min(x < 0, na.rm = TRUE) || symbreaks, densadj = 0.25,
  key.title = NULL, key.xlab = NULL, key.ylab = NULL,
  key.xtickfun = NULL, key.ytickfun = NULL, key.par = list(),
  main = NULL, xlab = NULL, ylab = NULL, lmat = NULL, lhei = NULL,
  lwid = NULL, extrafun = NULL, linewidth = 1, ...)
}
\arguments{
\item{holy}{crap so many parameters, same as heatmap.2}
}
\value{
a heatmap!
}
\description{
slight modification of heatmap.2
}
\seealso{
\code{\link{heatmap.2}}
}

