% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{hpgl_bcv_plot}
\alias{hpgl_bcv_plot}
\title{Steal edgeR's plotBCV() and make it a ggplot2.}
\usage{
hpgl_bcv_plot(data)
}
\arguments{
\item{data}{A dataframe/expt/exprs with count data}
}
\value{
a plot! of the BCV a la ggplot2.
}
\description{
This was written primarily to understand what that function is doing in edgeR.
}
\examples{
\dontrun{
bcv <- hpgl_bcv_plot(expt)
summary(bcv$data)
bcv$plot
}
}
\seealso{
\pkg{edgeR} \link[edgeR]{plotBCV}
}

