% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_boxplot}
\alias{hpgl_boxplot}
\title{\code{hpgl_boxplot()}  Make a ggplot boxplot of a set of samples.}
\usage{
hpgl_boxplot(data, colors = NULL, names = NULL, title = NULL,
  scale = NULL, ...)
}
\arguments{
\item{data}{an expt or data frame set of samples.}

\item{colors}{default=NULL  a color scheme, if not provided will make its own.}

\item{names}{default=NULL  a nicer version of the sample names.}

\item{title}{default=NULL  A title!}

\item{scale}{default='raw'  whether to log scale the y-axis.}

\item{...}{more parameters are fun}
}
\value{
a ggplot2 boxplot of the samples.  Each boxplot
contains the following information: a centered line describing the
median value of counts of all genes in the sample, a box around the
line describing the inner-quartiles around the median (quartiles 2
and 3 for those who are counting), a vertical line above/below the
box which shows 1.5x the inner quartile range (a common metric of
the non-outliers), and single dots for each gene which is outside
that range.  A single dot is transparent.
}
\description{
\code{hpgl_boxplot()}  Make a ggplot boxplot of a set of samples.
}
\examples{
\dontrun{
 a_boxplot <- hpgl_boxplot(expt)
 a_boxplot  ## ooo pretty boxplot look at the lines
}
}
\seealso{
\pkg{ggplot2} \pkg{reshape2} \link[ggplot2]{geom_boxplot}
\link[reshape2]{melt} \link[ggplot2]{scale_x_discrete}
}

