% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{hpgl_boxplot}
\alias{hpgl_boxplot}
\title{Make a ggplot boxplot of a set of samples.}
\usage{
hpgl_boxplot(data, colors = NULL, names = NULL, title = NULL,
  scale = NULL, ...)
}
\arguments{
\item{data}{Expt or data frame set of samples.}

\item{colors}{Color scheme, if not provided will make its own.}

\item{names}{Another version of the sample names for printing.}

\item{title}{A title!}

\item{scale}{Whether to log scale the y-axis.}

\item{...}{More parameters are more fun!}
}
\value{
Ggplot2 boxplot of the samples.  Each boxplot
contains the following information: a centered line describing the
median value of counts of all genes in the sample, a box around the
line describing the inner-quartiles around the median (quartiles 2
and 3 for those who are counting), a vertical line above/below the
box which shows 1.5x the inner quartile range (a common metric of
the non-outliers), and single dots for each gene which is outside
that range.  A single dot is transparent.
}
\description{
Boxplots and density plots provide complementary views of data distributions.  The general idea
is that if the box for one sample is significantly shifted from the others, then it is likely an
outlier in the same way a density plot shifted is an outlier.
}
\examples{
\dontrun{
 a_boxplot <- hpgl_boxplot(expt)
 a_boxplot  ## ooo pretty boxplot look at the lines
}
}
\seealso{
\pkg{ggplot2} \pkg{reshape2} \link[ggplot2]{geom_boxplot}
\link[reshape2]{melt} \link[ggplot2]{scale_x_discrete}
}

