% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.r
\name{hpgl_combatMod}
\alias{hpgl_combatMod}
\title{A modified version of comBatMod.}
\usage{
hpgl_combatMod(dat, batch, mod, noScale = TRUE, prior.plots = FALSE, ...)
}
\arguments{
\item{dat}{Df to modify.}

\item{batch}{Factor of batches.}

\item{mod}{Factor of conditions.}

\item{noScale}{The normal 'scale' option squishes the data too much, so this defaults to TRUE.}

\item{prior.plots}{Print out prior plots?}

\item{...}{Extra options are passed to arglist}
}
\value{
Df of batch corrected data
}
\description{
This is a hack of Kwame Okrah's combatMod to make it not fail on corner-cases.
This was mostly copy/pasted from https://github.com/kokrah/cbcbSEQ/blob/master/R/transform.R
}
\examples{
\dontrun{
 df_new = hpgl_combatMod(df, batches, model)
}
}
\seealso{
\pkg{sva}
 \code{\link[sva]{ComBat}}
}
