% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{hpgl_combatMod}
\alias{hpgl_combatMod}
\title{Use a modified version of combat on some data
This is a hack of Kwame's combatMod to make it not fail on corner-cases.}
\usage{
hpgl_combatMod(dat, batch, mod, noScale = TRUE, prior.plots = FALSE)
}
\arguments{
\item{dat}{a df to modify}

\item{batch}{a factor of batches}

\item{mod}{a factor of conditions}

\item{noScale}{the normal 'scale' option squishes the data too much, so this defaults to TRUE}

\item{prior.plots}{print out prior plots? FALSE}
}
\value{
a df of batch corrected data
}
\description{
Use a modified version of combat on some data
This is a hack of Kwame's combatMod to make it not fail on corner-cases.
}
\examples{
## df_new = hpgl_combatMod(df, batches, model)
}
\seealso{
\code{\link{sva}}, \code{\link{combat}},
}

