% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{hpgl_cor}
\alias{hpgl_cor}
\title{hpgl_cor()  Wrap cor() to include robust correlations.}
\usage{
hpgl_cor(df, method = "pearson", ...)
}
\arguments{
\item{df}{a data frame to test.}

\item{method}{default='pearson'  correlation method to use. Includes pearson, spearman, kendal, robust.}

\item{...}{other options to pass to stats::cor()}
}
\value{
correlation some fun correlation statistics
}
\description{
hpgl_cor()  Wrap cor() to include robust correlations.
}
\examples{
## hpgl_cor(df=df)
## hpgl_cor(df=df, method="robust")
}
\seealso{
\code{\link{cor}}, \code{\link{cov}}, \code{\link{covRob}}
}

