% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{hpgl_cor}
\alias{hpgl_cor}
\title{Wrap cor() to include robust correlations}
\usage{
hpgl_cor(df = NULL, method = "pearson", ...)
}
\arguments{
\item{df}{a data frame to test}

\item{method}{Correlation method to use.  Defaults to pearson.
Includes pearson, spearman, kendal, robust.}
}
\value{
correlation some fun correlation statistics
}
\description{
Wrap cor() to include robust correlations
}
\examples{
## hpgl_cor(df=df)
## hpgl_cor(df=df, method="robust")
}
\seealso{
\code{\link{cor}}, \code{\link{cov}}, \code{\link{covRob}}
}

