% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{hpgl_corheat}
\alias{hpgl_corheat}
\title{Make a heatmap.3 description of the correlation between samples.}
\usage{
hpgl_corheat(data, colors = NULL, design = NULL, method = "pearson",
  names = NULL, row = "batch", title = NULL, ...)
}
\arguments{
\item{data}{a dataframe, expt, or expressionset to work with.}

\item{colors}{a color scheme.}

\item{design}{a design matrix.}

\item{method}{correlation statistic to use.}

\item{names}{alternate names to use.}

\item{row}{what to place on the row of the map, batches or conditions?}

\item{title}{a title for the plot.}

\item{...}{more options are wonderful}
}
\value{
corheat_plot a gplots heatmap describing how the samples
pairwise correlate with one another.
}
\description{
Make a heatmap.3 description of the correlation between samples.
}
\examples{
## corheat_plot = hpgl_corheat(expt=expt, method="robust")
## corheat_plot
}
\seealso{
\link{hpgl_cor} \link[RColorBrewer]{brewer.pal}
\link[grDevices]{recordPlot}
}

