% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_corheat}
\alias{hpgl_corheat}
\title{hpgl_corheat()  Make a heatmap.3 description of the correlation between samples.}
\usage{
hpgl_corheat(data, colors = NULL, design = NULL, method = "pearson",
  names = NULL, row = "batch", title = NULL, ...)
}
\arguments{
\item{data}{a dataframe, expt, or expressionset to work with.}

\item{colors}{default=NULL  a color scheme.}

\item{design}{default=NULL  a design matrix.}

\item{method}{default='pearson'   correlation statistic to use.}

\item{names}{default=NULL  alternate names to use.}

\item{row}{default='batch'  what to place on the row of the map, batches or conditions?}

\item{title}{default=NULL  a title for the plot.}
}
\value{
corheat_plot a gplots heatmap describing how the samples
pairwise correlate with one another.
}
\description{
hpgl_corheat()  Make a heatmap.3 description of the correlation between samples.
}
\examples{
## corheat_plot = hpgl_corheat(expt=expt, method="robust")
## corheat_plot
}
\seealso{
\code{\link{hpgl_cor}}, \code{\link{brewer.pal}},
\code{\link{heatmap.2}}, \code{\link{recordPlot}}
}

