% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_corheat}
\alias{hpgl_corheat}
\title{Make a heatmap.3 description of the correlation between samples.}
\usage{
hpgl_corheat(df = NULL, colors = NULL, design = NULL, expt = NULL,
  method = "pearson", names = NULL, row = "batch", title = NULL, ...)
}
\arguments{
\item{df}{alternately a data frame which must be accompanied by}

\item{colors}{a color scheme}

\item{design}{a design matrix and}

\item{expt}{an expt set of samples}

\item{method}{correlation statistic to use.  Defaults to pearson.}
}
\value{
corheat_plot a gplots heatmap describing how the samples
pairwise correlate with one another.
}
\description{
Make a heatmap.3 description of the correlation between samples.
}
\examples{
## corheat_plot = hpgl_corheat(expt=expt, method="robust")
## corheat_plot
}
\seealso{
\code{\link{hpgl_cor}}, \code{\link{brewer.pal}},
\code{\link{heatmap.2}}, \code{\link{recordPlot}}
}

