% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_density}
\alias{hpgl_density}
\title{Density plots!}
\usage{
hpgl_density(data, colors = NULL, sample_names = NULL,
  position = "identity", fill = NULL, title = NULL, scale = NULL)
}
\arguments{
\item{data}{an expt, expressionset, or data frame.}

\item{colors}{a color scheme to use.}

\item{position}{how to place the lines, either let them overlap (identity), or stack them.}

\item{fill}{fill the distributions?  This might make the plot unreasonably colorful.}

\item{title}{a title for the plot.}

\item{scale}{plot on the log scale?}

\item{names}{names of the samples.}
}
\value{
a density plot!
}
\description{
Density plots!
}
\examples{
\dontrun{
funkytown <- hpgl_density(data)
}
}
\seealso{
\pkg{ggplot2} \link[ggplot2]{geom_density}
}

