% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_density_plot}
\alias{hpgl_density_plot}
\title{Density plots!}
\usage{
hpgl_density_plot(data, colors = NULL, names = NULL,
  position = "identity", fill = NULL, title = NULL, log = FALSE)
}
\arguments{
\item{names}{names of the samples}

\item{position}{stack or identity}

\item{fill}{NULL, but otherwise it fills in the colors for the rainbow coalition!}

\item{df}{a data frame to look at}

\item{expt}{or an expt class!}
}
\value{
a density plot!
}
\description{
Density plots!
}

