% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{hpgl_disheat}
\alias{hpgl_disheat}
\title{Make a heatmap.3 description of the similarity (euclildean distance) between samples.}
\usage{
hpgl_disheat(data, colors = NULL, design = NULL, method = "euclidean",
  names = NULL, row = "batch", title = NULL, ...)
}
\arguments{
\item{data}{a dataframe, expt, or expressionset to work with.}

\item{colors}{a color scheme.}

\item{design}{a design matrix.}

\item{method}{distance metric to use.}

\item{names}{alternate names to use.}

\item{row}{what to place on the row of the map, batches or conditions?}

\item{title}{a title for the plot.}

\item{...}{more parameters}
}
\value{
a recordPlot() heatmap describing the distance between samples.
}
\description{
Make a heatmap.3 description of the similarity (euclildean distance) between samples.
}
\examples{
\dontrun{
 disheat_plot = hpgl_disheat(expt=expt, method="euclidean")
 disheat_plot
}
}
\seealso{
\link[RColorBrewer]{brewer.pal} \link[gplots]{heatmap.2} \link[grDevices]{recordPlot}
}

