% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_disheat}
\alias{hpgl_disheat}
\title{Make a heatmap.3 description of the similarity (euclildean distance) between samples.}
\usage{
hpgl_disheat(data, colors = NULL, design = NULL, method = "euclidean",
  names = NULL, row = "batch", title = NULL, ...)
}
\arguments{
\item{colors}{a color scheme}

\item{design}{a design matrix and}

\item{method}{correlation statistic to use.  Defaults to euclidean.}

\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}
}
\value{
a recordPlot() heatmap describing the distance between samples.
}
\description{
Make a heatmap.3 description of the similarity (euclildean distance) between samples.
}
\examples{
## disheat_plot = hpgl_disheat(expt=expt, method="euclidean")
## disheat_plot
}
\seealso{
\code{\link{brewer.pal}},
\code{\link{heatmap.2}}, \code{\link{recordPlot}}
}

