% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{hpgl_disheat}
\alias{hpgl_disheat}
\title{Make a heatmap.3 description of the distances (euclidean by default) between samples.}
\usage{
hpgl_disheat(data, colors = NULL, design = NULL, method = "euclidean",
  names = NULL, title = NULL, ...)
}
\arguments{
\item{data}{Dataframe, expt, or expressionset to work with.}

\item{colors}{Color scheme (not needed if an expt is provided).}

\item{design}{Design matrix (not needed if an expt is provided).}

\item{method}{Distance metric to use.}

\item{names}{Alternate names to use for the samples.}

\item{title}{Title for the plot.}

\item{...}{More parameters!}
}
\value{
a recordPlot() heatmap describing the distance between samples.
}
\description{
Given a set of count tables and design, this will calculate the pairwise distances and plot
them as a heatmap.  It attempts to standardize the inputs and eventual output.
}
\examples{
\dontrun{
 disheat_plot = hpgl_disheat(expt=expt, method="euclidean")
 disheat_plot
}
}
\seealso{
\link[RColorBrewer]{brewer.pal} \link[gplots]{heatmap.2} \link[grDevices]{recordPlot}
}

