% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_disheat}
\alias{hpgl_disheat}
\title{hpgl_disheat()  Make a heatmap.3 description of the similarity (euclildean distance) between samples.}
\usage{
hpgl_disheat(data, colors = NULL, design = NULL, method = "euclidean",
  names = NULL, row = "batch", title = NULL, ...)
}
\arguments{
\item{data}{a dataframe, expt, or expressionset to work with.}

\item{colors}{default=NULL  a color scheme.}

\item{design}{default=NULL  a design matrix.}

\item{method}{default='euclidean'   distance metric to use.}

\item{names}{default=NULL  alternate names to use.}

\item{row}{default='batch'  what to place on the row of the map, batches or conditions?}

\item{title}{default=NULL  a title for the plot.}
}
\value{
a recordPlot() heatmap describing the distance between samples.
}
\description{
hpgl_disheat()  Make a heatmap.3 description of the similarity (euclildean distance) between samples.
}
\examples{
## disheat_plot = hpgl_disheat(expt=expt, method="euclidean")
## disheat_plot
}
\seealso{
\code{\link{brewer.pal}},
\code{\link{heatmap.2}}, \code{\link{recordPlot}}
}

