% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_dist_scatter}
\alias{hpgl_dist_scatter}
\title{Make a pretty scatter plot between two sets of numbers with a
cheesy distance metric and some statistics of the two sets.}
\usage{
hpgl_dist_scatter(df, tooltip_data = NULL, gvis_filename = NULL, size = 2)
}
\arguments{
\item{df}{a dataframe likely containing two columns}

\item{tooltip_data}{a df of tooltip information for gvis
graphs.}

\item{gvis_filename}{a filename to write a fancy html graph.
Defaults to NULL in which case the following parameter isn't needed.}

\item{size}{size of the dots}
}
\value{
a ggplot2 scatter plot.  This plot provides a "bird's eye"
view of two data sets.  This plot assumes the two data structures
are not correlated, and so it calculates the median/mad of each
axis and uses these to calculate a stupid, home-grown distance
metric away from both medians.  This distance metric is used to
color dots which are presumed the therefore be interesting because
they are far from 'normal.'  This will make a fun clicky googleVis
graph if requested.
}
\description{
The distance metric should be codified and made more intelligent.
Currently it creates a dataframe of distances which are absolute
distances from each axis, multiplied by each other, summed by axis,
then normalized against the maximum.
}
\examples{
## hpgl_dist_scatter(lotsofnumbers_intwo_columns, tooltip_data=tooltip_dataframe, gvis_filename="html/fun_scatterplot.html")
}
\seealso{
\pkg{ggplot2} \link{hpgl_gvis_scatter} \link[ggplot2]{geom_point}
\link{hpgl_linear_scatter}
}

