% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_enrich.internal}
\alias{hpgl_enrich.internal}
\title{A minor hack in the clusterProfiler function 'enrich.internal'}
\usage{
hpgl_enrich.internal(gene, organism, pvalueCutoff = 1, pAdjustMethod = "BH",
  ont, minGSSize = 2, qvalueCutoff = 0.2, readable = FALSE,
  universe = NULL)
}
\arguments{
\item{gene}{some differentially expressed genes}

\item{organism}{again will cause this to pull ensG if recognized by clusteprofiler}

\item{pvalueCutoff}{a pvalue cutoff}

\item{pAdjustMethod}{p adjust method}

\item{ont}{mf bp or cc}

\item{minGSSize}{a minimum gs size}

\item{qvalueCutoff}{maximum q value}

\item{readable}{set the readable flag for dose}

\item{universe}{a universe to use}
}
\value{
some clusterProfiler data
}
\description{
A minor hack in the clusterProfiler function 'enrich.internal'
}
\seealso{
\pkg{clusterProfiler}
}

