% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_enrich.internal}
\alias{hpgl_enrich.internal}
\title{A minor hack in the clusterProfiler function 'enrich.internal'}
\usage{
hpgl_enrich.internal(gene, organism, pvalueCutoff = 1, pAdjustMethod = "BH",
  ont, minGSSize = 2, qvalueCutoff = 0.2, readable = FALSE,
  universe = NULL)
}
\arguments{
\item{gene}{some differentially expressed genes}

\item{organism}{by default 'human'}

\item{ont}{by default 'MF'}
}
\value{
some clusterProfiler data
}
\description{
A minor hack in the clusterProfiler function 'enrich.internal'
}
\seealso{
\code{\link{clusterProfiler}}
}

