% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_enrich.internal}
\alias{hpgl_enrich.internal}
\title{A minor hack in the clusterProfiler function 'enrich.internal'.}
\usage{
hpgl_enrich.internal(gene, organism, pvalueCutoff = 1, pAdjustMethod = "BH",
  ont, minGSSize = 2, qvalueCutoff = 0.2, readable = FALSE,
  universe = NULL)
}
\arguments{
\item{gene}{Differentially expressed genes.}

\item{organism}{Pull ensembl annotations if this is a supported species.}

\item{pvalueCutoff}{P-value cutoff.}

\item{pAdjustMethod}{P-adjust method.}

\item{ont}{Molecular function, Biological process, or Cellular component?}

\item{minGSSize}{Minimum gs size?}

\item{qvalueCutoff}{Maximum allowed q-value.}

\item{readable}{Set the readable flag for the DOSE object?}

\item{universe}{Universe of genes to score significance against.}
}
\value{
Some clusterProfiler data.
}
\description{
I do not remember any longer why, but enrichGO errors out in ways which do not always make sense,
this was written to alleviate that problem.  I believe I sent a diff to the clusterProfiler
author but did not hear back and so added this function.
}
\seealso{
\pkg{clusterProfiler}
}

