% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_enrichGO}
\alias{hpgl_enrichGO}
\title{A minor hack in the clusterProfiler function 'enrichGO'}
\usage{
hpgl_enrichGO(gene, organism = "human", ont = "MF", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", universe, qvalueCutoff = 0.2, minGSSize = 2,
  readable = FALSE)
}
\arguments{
\item{gene}{some differentially expressed genes}

\item{organism}{if used will cause this to pull the ensG annotations}

\item{ont}{mf bp or cc}

\item{pvalueCutoff}{pvalue cutoff}

\item{pAdjustMethod}{p-value adjustment}

\item{universe}{the gene universe}

\item{qvalueCutoff}{maximum qvalue before adding}

\item{minGSSize}{smallest group size}

\item{readable}{readable tag on the object}
}
\value{
some clusterProfiler data
}
\description{
A minor hack in the clusterProfiler function 'enrichGO'
}
\seealso{
\pkg{clusterProfiler}
}

