% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_enrichGO}
\alias{hpgl_enrichGO}
\title{A minor hack in the clusterProfiler function 'enrichGO'}
\usage{
hpgl_enrichGO(gene, organism = "human", ont = "MF", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", universe, qvalueCutoff = 0.2, minGSSize = 2,
  readable = FALSE)
}
\arguments{
\item{gene}{some differentially expressed genes}

\item{organism}{by default 'human'}

\item{ont}{by default 'MF'}
}
\value{
some clusterProfiler data
}
\description{
A minor hack in the clusterProfiler function 'enrichGO'
}
\seealso{
\code{\link{clusterProfiler}}
}

