% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{hpgl_Gff2GeneTable}
\alias{hpgl_Gff2GeneTable}
\title{A copy and paste of clusterProfiler's readGff}
\usage{
hpgl_Gff2GeneTable(gffFile, compress = TRUE, split = "=")
}
\description{
A copy and paste of clusterProfiler's readGff
}

