% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology.R
\name{hpgl_GOplot}
\alias{hpgl_GOplot}
\title{A minor hack in the topGO GOplot function}
\usage{
hpgl_GOplot(dag, sigNodes, dag.name = "GO terms", edgeTypes = T,
  nodeShape.type = c("box", "circle", "ellipse", "plaintext")[3],
  genNodes = NULL, wantedNodes = NULL, showEdges = T, useFullNames = T,
  oldSigNodes = NULL, nodeInfo = nodeInfo, maxchars = 30)
}
\description{
A minor hack in the topGO GOplot function
}

