% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{hpgl_GOplot}
\alias{hpgl_GOplot}
\title{A minor hack in the topGO GOplot function}
\usage{
hpgl_GOplot(dag, sigNodes, dag.name = "GO terms", edgeTypes = TRUE,
  nodeShape.type = c("box", "circle", "ellipse", "plaintext")[3],
  genNodes = NULL, wantedNodes = NULL, showEdges = TRUE,
  useFullNames = TRUE, oldSigNodes = NULL, nodeInfo = nodeInfo,
  maxchars = 30)
}
\description{
A minor hack in the topGO GOplot function
}

