% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{hpgl_GOplot}
\alias{hpgl_GOplot}
\title{A minor hack of the topGO GOplot function.}
\usage{
hpgl_GOplot(dag, sigNodes, dag.name = "GO terms", edgeTypes = TRUE,
  nodeShape.type = c("box", "circle", "ellipse", "plaintext")[3],
  genNodes = NULL, wantedNodes = NULL, showEdges = TRUE,
  useFullNames = TRUE, oldSigNodes = NULL, nodeInfo = NULL,
  maxchars = 30)
}
\arguments{
\item{dag}{DAG tree of ontologies.}

\item{sigNodes}{Set of significant ontologies (with p-values).}

\item{dag.name}{Name for the graph.}

\item{edgeTypes}{Types of the edges for graphviz.}

\item{nodeShape.type}{Shapes on the tree.}

\item{genNodes}{Generate the nodes?}

\item{wantedNodes}{Subset of the ontologies to plot.}

\item{showEdges}{Show the arrows?}

\item{useFullNames}{Full names of the ontologies (they can get long).}

\item{oldSigNodes}{I dunno.}

\item{nodeInfo}{Hmm.}

\item{maxchars}{Maximum characters per line inside the shapes.}
}
\value{
Topgo plot!
}
\description{
This allows me to change the line widths from the default.
}
\seealso{
\pkg{topGO}
}
