% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{hpgl_GOplot}
\alias{hpgl_GOplot}
\title{hpgl_GOplot() A minor hack of the topGO GOplot function
This allows me to change the line widths from the default.}
\usage{
hpgl_GOplot(dag, sigNodes, dag.name = "GO terms", edgeTypes = TRUE,
  nodeShape.type = c("box", "circle", "ellipse", "plaintext")[3],
  genNodes = NULL, wantedNodes = NULL, showEdges = TRUE,
  useFullNames = TRUE, oldSigNodes = NULL, nodeInfo = NULL,
  maxchars = 30)
}
\arguments{
\item{dag}{The DAG tree of ontologies}

\item{sigNodes}{The set of significant ontologies (with p-values)}

\item{dag.name}{default='GO terms'  A name for the graph}

\item{edgeTypes}{default=TRUE  Set the types of the edges for graphviz}

\item{nodeShape.type}{default=c(box, circle, ellipse, plaintext)  The shapes on the tree}

\item{genNodes}{default=NULL  Generate the nodes?}

\item{wantedNodes}{default=NULL  A subset of the ontologies to plot}

\item{showEdges}{default=TRUE  Show the arrows?}

\item{useFullNames}{default=TRUE  Full names of the ontologies (they can get long)}

\item{oldSigNodes}{default=NULL  I dunno}

\item{nodeInfo}{default=nodeInfo   Hmm}

\item{maxchars}{default=30  Maximum characters per line inside the shapes}
}
\value{
a topgo plot
}
\description{
hpgl_GOplot() A minor hack of the topGO GOplot function
This allows me to change the line widths from the default.
}

