% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_ontology_topgo.r
\name{hpgl_GroupDensity}
\alias{hpgl_GroupDensity}
\title{A hack of topGO's groupDensity()}
\usage{
hpgl_GroupDensity(object, whichGO, ranks = TRUE, rm.one = FALSE)
}
\arguments{
\item{object}{TopGO enrichment object.}

\item{whichGO}{Individual ontology group to compare against.}

\item{ranks}{Rank order the set of ontologies?}

\item{rm.one}{Remove pvalue=1 groups?}
}
\value{
plot of group densities.
}
\description{
This just adds a couple wrappers to avoid errors in groupDensity.
}
