% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvis_plots.r
\name{hpgl_gvis_ma_plot}
\alias{hpgl_gvis_ma_plot}
\title{Make an html version of an MA plot.}
\usage{
hpgl_gvis_ma_plot(counts, degenes, tooltip_data = NULL,
  filename = "html/gvis_ma_plot.html", base_url = "", ...)
}
\arguments{
\item{counts}{df of linear-modelling, normalized counts by sample-type,
which is to say the output from voom/voomMod/hpgl_voom().}

\item{degenes}{df from toptable or its friends containing p-values.}

\item{tooltip_data}{a df of tooltip information.}

\item{filename}{a filename to write a fancy html graph.}

\item{base_url}{a string with a basename used for generating URLs for clicking dots on the graph.}

\item{...}{more options are more options}
}
\value{
NULL, but along the way an html file is generated which
contains a googleVis MA plot.  See hpgl_ma_plot() for details.
}
\description{
Make an html version of an MA plot.
}
\examples{
\dontrun{
 hpgl_gvis_ma_plot(voomed_data, toptable_data, filename="html/fun_ma_plot.html", base_url="http://yeastgenome.org/accession?")
}
}
\seealso{
\link{hpgl_ma_plot}
}

