% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvis_plots.r
\name{hpgl_gvis_scatter}
\alias{hpgl_gvis_scatter}
\title{Make an html version of a scatter plot.}
\usage{
hpgl_gvis_scatter(df, tooltip_data = NULL,
  filename = "html/gvis_scatter.html", base_url = "", trendline = NULL)
}
\arguments{
\item{df}{df of two columns to compare}

\item{tooltip_data}{a df of tooltip information for gvis graphs.}

\item{filename}{a filename to write a fancy html graph.}

\item{base_url}{a url to send click events which will be suffixed with the gene name}

\item{trendline}{add a trendline?}
}
\value{
NULL, but along the way an html file is generated which
contains a googleVis scatter plot.  See hpgl_scatter_plot() for details.
}
\description{
Make an html version of a scatter plot.
}
\examples{
\dontrun{
 hpgl_gvis_scatter(a_dataframe_twocolumns, filename="html/fun_scatter_plot.html", base_url="http://yeastgenome.org/accession?")
}
}
\seealso{
\link[googleVis]{gvisScatterChart}
}

