% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{hpgl_heatmap}
\alias{hpgl_heatmap}
\title{Make a heatmap.3 plots, does the work for hpgl_disheat and hpgl_corheat.}
\usage{
hpgl_heatmap(data, colors = NULL, design = NULL, method = "pearson",
  names = NULL, type = "correlation", row = "batch", title = NULL, ...)
}
\arguments{
\item{data}{a dataframe, expt, or expressionset to work with.}

\item{colors}{a color scheme.}

\item{design}{a design matrix.}

\item{method}{distance or correlation metric to use.}

\item{names}{alternate names to use.}

\item{type}{correlation or distance or sample}

\item{row}{what to place on the row of the map, batches or conditions?}

\item{title}{a title for the plot.}

\item{...}{I like elipses}
}
\value{
a recordPlot() heatmap describing the distance between samples.
}
\description{
Make a heatmap.3 plots, does the work for hpgl_disheat and hpgl_corheat.
}
\seealso{
\link[RColorBrewer]{brewer.pal} \link[grDevices]{recordPlot}
}

