% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_heatmap}
\alias{hpgl_heatmap}
\title{hpgl_heatmap()  Make a heatmap.3 plots, does the work for hpgl_disheat and hpgl_corheat.}
\usage{
hpgl_heatmap(data, colors = NULL, design = NULL, method = "pearson",
  names = NULL, type = "correlation", row = "batch", title = NULL, ...)
}
\arguments{
\item{data}{a dataframe, expt, or expressionset to work with.}

\item{colors}{default=NULL  a color scheme.}

\item{design}{default=NULL  a design matrix.}

\item{method}{default='pearson'   distance or correlation metric to use.}

\item{names}{default=NULL  alternate names to use.}

\item{row}{default='batch'  what to place on the row of the map, batches or conditions?}

\item{title}{default=NULL  a title for the plot.}
}
\value{
a recordPlot() heatmap describing the distance between samples.
}
\description{
hpgl_heatmap()  Make a heatmap.3 plots, does the work for hpgl_disheat and hpgl_corheat.
}
\seealso{
\code{\link{brewer.pal}},
\code{\link{heatmap.2}}, \code{\link{recordPlot}}
}

