% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_histogram}
\alias{hpgl_histogram}
\title{Make a pretty histogram of something.}
\usage{
hpgl_histogram(df, binwidth = NULL, log = FALSE, bins = 500,
  verbose = FALSE, fillcolor = "darkgrey", color = "black")
}
\arguments{
\item{df}{a dataframe of lots of pretty numbers.}

\item{binwidth}{width of the bins for the histogram.}

\item{log}{replot on the log scale?}

\item{bins}{bins for the histogram}

\item{verbose}{be verbose?}

\item{fillcolor}{change the fill colors of the plotted elements.}

\item{color}{change the color of the lines of the plotted elements.}
}
\value{
a ggplot histogram
}
\description{
Make a pretty histogram of something.
}
\examples{
\dontrun{
 kittytime = hpgl_histogram(df)
}
}
\seealso{
\link[ggplot2]{geom_histogram} \link[ggplot2]{geom_density}
}

