% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_histogram}
\alias{hpgl_histogram}
\title{Make a pretty histogram of something}
\usage{
hpgl_histogram(df, binwidth = NULL, log = FALSE, bins = 500,
  verbose = FALSE, fillcolor = "darkgrey", color = "black")
}
\arguments{
\item{df}{a dataframe of lots of pretty numbers}
}
\value{
a ggplot histogram
}
\description{
Make a pretty histogram of something
}
\examples{
## kittytime = hpgl_histogram(df)
}
\seealso{
\code{\link{geom_histogram}}, \code{\link{geom_density}},
}

