% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_libsize}
\alias{hpgl_libsize}
\title{Make a ggplot graph of library sizes.}
\usage{
hpgl_libsize(data, colors = NULL, scale = TRUE, names = NULL,
  title = NULL, text = TRUE, ...)
}
\arguments{
\item{colors}{a color scheme}

\item{scale}{whether or not to log10 the y-axis}

\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{design}{a design matrix and}
}
\value{
a ggplot2 bar plot of every sample's size
}
\description{
Make a ggplot graph of library sizes.
}
\examples{
## libsize_plot = hpgl_libsize(expt=expt)
## libsize_plot  ## ooo pretty bargraph
}
\seealso{
\code{\link{geom_bar}}, \code{\link{geom_text}},
\code{\link{prettyNum}}, \code{\link{scale_y_log10}}
}

