% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_linear_scatter}
\alias{hpgl_linear_scatter}
\title{Make a pretty scatter plot between two sets of numbers with a
linear model superimposed and some supporting statistics.}
\usage{
hpgl_linear_scatter(df, tooltip_data = NULL, gvis_filename = NULL,
  cormethod = "pearson", size = 2, verbose = FALSE, loess = FALSE,
  identity = FALSE, gvis_trendline = NULL, first = NULL, second = NULL,
  base_url = NULL, pretty_colors = TRUE)
}
\arguments{
\item{df}{a dataframe likely containing two columns}

\item{tooltip_data}{a df of tooltip information for gvis
graphs.}

\item{gvis_filename}{a filename to write a fancy html graph.}

\item{cormethod}{what type of correlation to check?}

\item{size}{size of the dots on the plot.}

\item{verbose}{be verbose?}

\item{loess}{add a loess estimation?}

\item{identity}{add the identity line?}

\item{gvis_trendline}{add a trendline to the gvis plot?  There are a couple possible types, I think linear is the most common.}

\item{first}{first column to plot}

\item{second}{second column to plot}

\item{base_url}{a base url to add to the plot}

\item{pretty_colors}{colors!}
}
\value{
a list including a ggplot2 scatter plot and some
histograms.  This plot provides a "bird's eye"
view of two data sets.  This plot assumes a (potential) linear
correlation between the data, so it calculates the correlation
between them.  It then calculates and plots a robust linear model
of the data using an 'SMDM' estimator (which I don't remember how
to describe, just that the document I was reading said it is good).
The median/mad of each axis is calculated and plotted as well.  The
distance from the linear model is finally used to color the dots on
the plot.  Histograms of each axis are plotted separately and then
together under a single cdf to allow tests of distribution
similarity.  This will make a fun clicky googleVis graph if
requested.
}
\description{
Make a pretty scatter plot between two sets of numbers with a
linear model superimposed and some supporting statistics.
}
\examples{
\dontrun{
 hpgl_linear_scatter(lotsofnumbers_intwo_columns, tooltip_data=tooltip_dataframe, gvis_filename="html/fun_scatterplot.html")
}
}
\seealso{
\link[robust]{lmRob} \link[stats]{weights} \link{hpgl_histogram}
}

