% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{hpgl_log2cpm}
\alias{hpgl_log2cpm}
\title{\code{hpgl_log2cpm()}  Converts count matrix to log2 counts-per-million reads.}
\usage{
hpgl_log2cpm(counts, lib.size = NULL)
}
\arguments{
\item{counts}{read count matrix}
}
\value{
log2-CPM read count matrix
}
\description{
Based on the method used by limma as described in the Law et al. (2014) voom
paper.
}
\examples{
\dontrun{
l2cpm <- hpgl_log2cpm(counts)
}
}
\seealso{
\pkg{cbcbSEQ} \pkg{edgeR}
}

