% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_ma_plot}
\alias{hpgl_ma_plot}
\title{hpgl_ma_plot()  Make a pretty MA plot from the output of voom/limma/eBayes/toptable.}
\usage{
hpgl_ma_plot(counts, de_genes, adjpval_cutoff = 0.05, alpha = 0.6,
  size = 2, tooltip_data = NULL, gvis_filename = NULL, ...)
}
\arguments{
\item{counts}{df of linear-modelling, normalized counts by sample-type,
which is to say the output from voom/voomMod/hpgl_voom().}

\item{de_genes}{df from toptable or its friends containing p-values.}

\item{adjpval_cutoff}{default=0.05  a cutoff defining significant from not.}

\item{alpha}{default=0.6  how transparent to make the dots.}

\item{size}{default=2  how big are the dots?}

\item{tooltip_data}{default=NULL  a df of tooltip information for gvis
graphs.}

\item{gvis_filename}{default=NULL  a filename to write a fancy html graph.}

\item{...}{more poptions por pou}
}
\value{
a ggplot2 MA scatter plot.  This is defined as the rowmeans
of the normalized counts by type across all sample types on the
x-axis, and the log fold change between conditions on the y-axis.
Dots are colored depending on if they are 'significant.'  This will
make a fun clicky googleVis graph if requested.
}
\description{
hpgl_ma_plot()  Make a pretty MA plot from the output of voom/limma/eBayes/toptable.
}
\examples{
## hpgl_ma_plot(voomed_data, toptable_data, gvis_filename="html/fun_ma_plot.html")
## Currently this assumes that a variant of toptable was used which
## gives adjusted p-values.  This is not always the case and I should
## check for that, but I have not yet.
}
\seealso{
\link{hpgl_gvis_ma_plot} \link[limma]{toptable}
\link[limma]{voom} \link[cbcbSEQ]{voomMod} \link{hpgl_voom}
\link[limma]{lmFit} \link[limma]{makeContrasts}
\link[limma]{contrasts.fit}
}

