% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_multihistogram}
\alias{hpgl_multihistogram}
\title{Make a pretty histogram of multiple datasets}
\usage{
hpgl_multihistogram(data, log = FALSE, binwidth = NULL, bins = NULL,
  verbose = FALSE)
}
\arguments{
\item{df}{a dataframe of lots of pretty numbers, this also accepts lists.}
}
\value{
a ggplot histogram comparing multiple data sets
Along the way this generates pairwise t tests of the columns of
data.
}
\description{
Make a pretty histogram of multiple datasets
}
\examples{
## kittytime = hpgl_multihistogram(df)
}
\seealso{
\code{\link{pairwise.t.test}}, \code{\link{ddply}},
\code{\link{rbind}}
}

