% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.r
\name{hpgl_multiplot}
\alias{hpgl_multiplot}
\title{multiplot()  Make a grid of plots.}
\usage{
hpgl_multiplot(plots, file, cols = NULL, layout = NULL)
}
\arguments{
\item{plots}{a list of plots}

\item{file}{a file to write to}

\item{cols}{default=NULL  the number of columns in the grid}

\item{layout}{default=NULL set the layout specifically}
}
\value{
a multiplot!
}
\description{
multiplot()  Make a grid of plots.
}

