% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{hpgl_norm}
\alias{hpgl_norm}
\title{\code{hpgl_norm()} Normalize a dataframe/expt, express it, and/or transform it}
\usage{
hpgl_norm(data, design = NULL, transform = "raw", norm = "raw",
  convert = "raw", batch = "raw", batch1 = "batch", batch2 = NULL,
  filter_low = FALSE, annotations = NULL, entry_type = "gene",
  fasta = NULL, verbose = FALSE, thresh = 2, min_samples = 2,
  noscale = TRUE, p = 0.01, A = 1, k = 1, cv_min = 0.01,
  cv_max = 1000, ...)
}
\arguments{
\item{transform}{defines whether to log(2|10) transform the
data. Defaults to raw.}

\item{norm}{specify the normalization strategy.  Defaults to
raw.  This makes use of DESeq/EdgeR to provide: RLE, upperquartile,
size-factor, or tmm normalization.  I tend to like quantile, but there are
definitely corner-case scenarios for all strategies.}

\item{convert}{defines the output type which may be raw, cpm,
rpkm, or cp_seq_m.  Defaults to raw.}

\item{filter_low}{choose whether to low-count filter the data.
Defaults to true.}

\item{annotations}{is used for rpkm or sequence normalizations to
extract the lengths of sequences for normalization}

\item{expt=expt}{an expt class containing all the necessary
metadata}

\item{df=df}{alternately a dataframe of counts may be used}

\item{design=design}{but a design dataframe must come with it}
}
\value{
edgeR's DGEList expression of a count table.  This seems to
me to be the easiest to deal with.
}
\description{
\code{hpgl_norm()} Normalize a dataframe/expt, express it, and/or transform it
}
\examples{
\dontrun{
df_raw = hpgl_norm(expt=expt)  ## Only performs low-count filtering
df_raw = hpgl_norm(df=a_df, design=a_design) ## Same, but using a df
df_ql2rpkm = hpgl_norm(expt=expt, norm='quant', transform='log2', convert='rpkm')  ## Quantile, log2, rpkm
count_table = df_ql2rpkm$counts
}
}
\seealso{
\code{\link[edgeR]{cpm}}, \code{\link[edgeR]{rpkm}},
\code{\link{hpgl_rpkm}}, \code{\link[cbcbSEQ]{filterCounts}},
\code{\link[DESeq]{DESeqDataSetFromMatrix}},
\code{\link[DESeq]{estimateSizeFactors}}, \code{\link[DESeq]{DGEList}},
\code{\link[edgeR]{calcNormFactors}}
}

