% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{hpgl_norm}
\alias{hpgl_norm}
\title{Normalize a dataframe/expt, express it, and/or transform it}
\usage{
hpgl_norm(df = NULL, expt = NULL, design = NULL, transform = "raw",
  norm = "raw", convert = "raw", batch = "raw", batch1 = "batch",
  batch2 = NULL, filter_low = TRUE, annotations = NULL, verbose = FALSE,
  thresh = 2, min_samples = 2, ...)
}
\arguments{
\item{transform}{defines whether to log(2|10) transform the
data. Defaults to raw.}

\item{norm}{specify the normalization strategy.  Defaults to
raw.  This makes use of DESeq/EdgeR to provide: RLE, upperquartile,
size-factor, or tmm normalization.  I tend to like quantile, but there are
definitely corner-case scenarios for all strategies.}

\item{convert}{defines the output type which may be raw, cpm,
rpkm, or cp_seq_m.  Defaults to raw.}

\item{filter_low}{choose whether to low-count filter the data.
Defaults to true.}

\item{annotations}{is used for rpkm or sequence normalizations to
extract the lengths of sequences for normalization}

\item{expt=expt}{an expt class containing all the necessary
metadata}

\item{df=df}{alternately a dataframe of counts may be used}

\item{design=design}{but a design dataframe must come with it}
}
\value{
edgeR's DGEList expression of a count table.  This seems to
me to be the easiest to deal with.
}
\description{
Normalize a dataframe/expt, express it, and/or transform it
}
\examples{
## df_raw = hpgl_norm(expt=expt)  ## Only performs low-count filtering
## df_raw = hpgl_norm(df=a_df, design=a_design) ## Same, but using a df
## df_ql2rpkm = hpgl_norm(expt=expt, norm_type='quant', filter='log2', out_type='rpkm'  ## Quantile, log2, rpkm
## count_table = df_ql2rpkm$counts
}
\seealso{
\code{\link{cpm}}, \code{\link{rpkm}},
\code{\link{hpgl_rpkm}}, \code{\link{filterCounts}},
\code{\link{DESeqDataSetFromMatrix}},
\code{\link{estimateSizeFactors}}, \code{\link{DGEList}},
\code{\link{qNorm}}, \code{\link{calcNormFactors}}
}

