% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{hpgl_norm}
\alias{hpgl_norm}
\title{Normalize a dataframe/expt, express it, and/or transform it}
\usage{
hpgl_norm(data, ...)
}
\arguments{
\item{data}{Some data as a df/expt/whatever.}

\item{...}{I should put all those other options here}
}
\value{
edgeR's DGEList expression of a count table.  This seems to
 me to be the easiest to deal with.
}
\description{
There are many possible options to this function.  Refer to normalize_expt()
for a more complete list.
}
\examples{
\dontrun{
 df_raw = hpgl_norm(expt = expt)  ## Only performs low-count filtering
 df_raw = hpgl_norm(df = a_df, design = a_design) ## Same, but using a df
 df_ql2rpkm = hpgl_norm(expt = expt, norm='quant', transform='log2',
                        convert='rpkm')  ## Quantile, log2, rpkm
 count_table = df_ql2rpkm$counts
}
}
\seealso{
[edgeR] [DESeq2] [edgeR::cpm()] [filter_counts()] [batch_counts()]
 [convert_counts()] [transform_counts()]
}
