% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_pairwise_ma}
\alias{hpgl_pairwise_ma}
\title{hpgl_pairwise_ma() Plot all pairwise MA plots in an experiment}
\usage{
hpgl_pairwise_ma(expt = NULL, df = NULL, log = NULL, ...)
}
\arguments{
\item{expt}{an expt expressionset superclass or...}

\item{df}{a data frame containing the data to explore.}

\item{log}{is the data in log format? (it will likely figure it out
if not)}
}
\value{
a list of affy::maplots
}
\description{
Use affy's ma.plot() on every pair of columns in a data set to help
diagnose problematic samples.
}
\examples{
## ma_plots = hpgl_pairwise_ma(expt=some_expt)
}
\seealso{
\code{\link{ma.plot}}
}

