% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{hpgl_pathview}
\alias{hpgl_pathview}
\title{Print some data onto KEGG pathways}
\usage{
hpgl_pathview(path_data, indir = "pathview_in", outdir = "pathview",
  pathway = "all", species = "lma", string_from = "LmjF",
  string_to = "LMJF", suffix = "_colored", second_from = NULL,
  second_to = NULL, verbose = FALSE, filenames = "id")
}
\arguments{
\item{indir}{A directory into which the unmodified kegg images will be downloaded (or already exist).  Defaults to 'pathview_in'}

\item{outdir}{A directory which will contain the colored images.}

\item{pathway}{Perform the coloring for a specific pathway?  Defaults to 'all'}

\item{species}{The kegg identifier for the species of interest.}

\item{filenames}{Whether to give filenames as the kegg ID or pathway name.  Defaults to 'id'.}

\item{de_genes}{some differentially expressed genes}
}
\value{
A list of some information for every KEGG pathway downloaded/examined.  This information includes:
  a. The filename of the final image for each pathway.
  b. The number of genes which were found in each pathway image.
  c. The number of genes in the 'up' category
  d. The number of genes in the 'down' category
}
\description{
Print some data onto KEGG pathways
}
\examples{
## thy_el_comp2_path = hpgl_pathview(thy_el_comp2_kegg, species="spz", indir="pathview_in", outdir="kegg_thy_el_comp2", string_from="_Spy", string_to="_Spy_", filenames="pathname")
}
\seealso{
\code{\link{Ramigo}}
}

