% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{hpgl_pathview}
\alias{hpgl_pathview}
\title{Print some data onto KEGG pathways.}
\usage{
hpgl_pathview(path_data, indir = "pathview_in", outdir = "pathview",
  pathway = "all", species = "lma", string_from = "LmjF",
  string_to = "LMJF", suffix = "_colored", second_from = NULL,
  second_to = NULL, filenames = "id", fc_column = "limma_logfc",
  format = "png")
}
\arguments{
\item{path_data}{Some differentially expressed genes.}

\item{indir}{Directory into which the unmodified kegg images will be downloaded (or already exist).}

\item{outdir}{Directory which will contain the colored images.}

\item{pathway}{Perform the coloring for a specific pathway?}

\item{species}{Kegg identifier for the species of interest.}

\item{string_from}{Regex to help in renaming KEGG categories/gene names from one format to another.}

\item{string_to}{Regex to help in renaming KEGG categories/gene names from one format to another.}

\item{suffix}{Add a suffix to the completed, colored files.}

\item{second_from}{Sometimes just one regex is not enough!}

\item{second_to}{Sometimes just one regex is not enough!}

\item{filenames}{Name the final files by id or name?}
}
\value{
A list of some information for every KEGG pathway downloaded/examined.  This information includes:
  a. The filename of the final image for each pathway.
  b. The number of genes which were found in each pathway image.
  c. The number of genes in the 'up' category
  d. The number of genes in the 'down' category
}
\description{
KEGGREST and pathview provide neat functions for coloring molecular pathways with arbitrary data.
Unfortunately they are somewhat evil to use.  This attempts to alleviate that.
}
\examples{
\dontrun{
 thy_el_comp2_path = hpgl_pathview(thy_el_comp2_kegg, species="spz", indir="pathview_in",
                                   outdir="kegg_thy_el_comp2", string_from="_Spy",
                                   string_to="_Spy_", filenames="pathname")
}
}
\seealso{
\pkg{Ramigo} \pkg{pathview}
}

