% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{hpgl_pca}
\alias{hpgl_pca}
\title{Make a ggplot PCA plot describing the samples' clustering}
\usage{
hpgl_pca(data, colors = NULL, design = NULL, title = NULL,
  labels = NULL, size = 3, ...)
}
\arguments{
\item{colors}{a color scheme}

\item{design}{a design matrix and}

\item{labels}{add labels?  Also, what type?  FALSE, "default", or "fancy"}

\item{expt}{an expt set of samples}

\item{df}{alternately a data frame which must be accompanied by}

\item{method}{a correlation method to use.  Defaults to pearson.}

\item{names}{use pretty names for the samples?}
}
\value{
a list containing the following:
  plot = ggplot2 pca_plot describing the principle component analysis of the samples.
  table = a table of the PCA plot data
  res = a table of the PCA res data
  variance = a table of the PCA plot variance
This makes use of cbcbSEQ and prints the table of variance by component.
}
\description{
Make a ggplot PCA plot describing the samples' clustering
}
\examples{
## pca_plot = hpgl_pca(expt=expt)
## pca_plot
}
\seealso{
\code{\link{makeSVD}}, \code{\link{pcRes}},
\code{\link{geom_dl}}
}

