% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_qq_all}
\alias{hpgl_qq_all}
\title{quantile/quantile comparison of all samples (in this case the mean of all samples, and each sample)}
\usage{
hpgl_qq_all(df = NULL, expt = NULL, verbose = FALSE, labels = "short")
}
\arguments{
\item{df}{the counts}

\item{expt}{or the expt class}

\item{verbose}{be chatty while running?}

\item{against}{either the mean of all samples, or a sample dataset generated on the fly}

\item{norm}{normalize the data? -- this is no longer needed I think}
}
\value{
a list containing:
  logs = a recordPlot() of the pairwise log qq plots
  ratios = a recordPlot() of the pairwise ratio qq plots
  means = a table of the median values of all the summaries of the qq plots
}
\description{
quantile/quantile comparison of all samples (in this case the mean of all samples, and each sample)
}

