% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_qq_all}
\alias{hpgl_qq_all}
\title{hpgl_qq_all()  quantile/quantile comparison of all samples (in this case the mean of all samples, and each sample)}
\usage{
hpgl_qq_all(data, verbose = FALSE, labels = "short")
}
\arguments{
\item{data}{an expressionset, expt, or dataframe of samples.}

\item{verbose}{default=FALSE  be chatty while running?}

\item{labels}{default='short'  what kind of labels to print?}
}
\value{
a list containing:
  logs = a recordPlot() of the pairwise log qq plots
  ratios = a recordPlot() of the pairwise ratio qq plots
  means = a table of the median values of all the summaries of the qq plots
}
\description{
hpgl_qq_all()  quantile/quantile comparison of all samples (in this case the mean of all samples, and each sample)
}

