% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_qq_all_pairwise}
\alias{hpgl_qq_all_pairwise}
\title{Perform qq plots of every column against every other column of a dataset}
\usage{
hpgl_qq_all_pairwise(df = NULL, expt = NULL, verbose = FALSE)
}
\arguments{
\item{df}{the data}

\item{expt}{or an expt class}
}
\value{
a list containing the recordPlot() output of the ratios, logs, and means among samples
}
\description{
Perform qq plots of every column against every other column of a dataset
}

