% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{hpgl_qq_all_pairwise}
\alias{hpgl_qq_all_pairwise}
\title{Perform qq plots of every column against every other column of a dataset.}
\usage{
hpgl_qq_all_pairwise(data)
}
\arguments{
\item{data}{Dataframe to perform pairwise qqplots with.}
}
\value{
List containing the recordPlot() output of the ratios, logs, and means among samples.
}
\description{
This function is stupid, don't use it.  It makes more sense to just use hpgl_qq, however I am not
quite read to delete this function yet.
}

