% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{hpgl_qq_all_pairwise}
\alias{hpgl_qq_all_pairwise}
\title{Perform qq plots of every column against every other column of a dataset.
This function is stupid, don't use it.}
\usage{
hpgl_qq_all_pairwise(data, verbose = FALSE)
}
\arguments{
\item{data}{the data}

\item{verbose}{talky talky}
}
\value{
a list containing the recordPlot() output of the ratios, logs, and means among samples
}
\description{
Perform qq plots of every column against every other column of a dataset.
This function is stupid, don't use it.
}

