% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_qq_all_pairwise}
\alias{hpgl_qq_all_pairwise}
\title{hpgl_qq_all_pairwise()  Perform qq plots of every column against every other column of a dataset.
This function is stupid, don't use it.}
\usage{
hpgl_qq_all_pairwise(df = NULL, expt = NULL, verbose = FALSE)
}
\arguments{
\item{df}{the data}

\item{expt}{or an expt class}
}
\value{
a list containing the recordPlot() output of the ratios, logs, and means among samples
}
\description{
hpgl_qq_all_pairwise()  Perform qq plots of every column against every other column of a dataset.
This function is stupid, don't use it.
}

