% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{hpgl_qq_plot}
\alias{hpgl_qq_plot}
\title{Perform a qqplot between two columns of a matrix.}
\usage{
hpgl_qq_plot(data, x = 1, y = 2, labels = TRUE)
}
\arguments{
\item{data}{data frame/expt/expressionset.}

\item{x}{the first column.}

\item{y}{the second column.}

\item{labels}{include the lables?}
}
\value{
a list of the logs, ratios, and mean between the plots as ggplots.
}
\description{
Perform a qqplot between two columns of a matrix.
}

