% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_qq_plot}
\alias{hpgl_qq_plot}
\title{Perform a qqplot between two columns of a matrix}
\usage{
hpgl_qq_plot(df = NULL, expt = NULL, x = 1, y = 2, labels = TRUE)
}
\arguments{
\item{x}{the first column}

\item{y}{the second column}

\item{labels}{whether to include the labels}

\item{a}{data frame}

\item{or}{an expt!}
}
\value{
a list of the logs, ratios, and mean between the plots as ggplots.
}
\description{
Perform a qqplot between two columns of a matrix
}

