% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_norm.r
\name{hpgl_qshrink}
\alias{hpgl_qshrink}
\title{A hacked copy of Kwame's qsmooth/qstats code.}
\usage{
hpgl_qshrink(
  data = NULL,
  groups = NULL,
  refType = "mean",
  groupLoc = "mean",
  window = 99,
  groupCol = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{Count table to modify}

\item{groups}{Factor of the experimental conditions}

\item{refType}{Method for grouping conditions}

\item{groupLoc}{Method for grouping groups}

\item{window}{Window, for looking!}

\item{groupCol}{Column to define conditions}

\item{plot}{Plot the quantiles?}

\item{...}{More options}
}
\value{
New data frame of normalized counts
}
\description{
I made a couple small changes to Kwame's qstats() function to make
it not fail when on corner-cases.  I sent him a diff, but haven't
checked to see if it was useful yet.
}
\examples{
\dontrun{
 df <- hpgl_qshrink(data)
}
}
\seealso{
\pkg{qsmooth}
}
