% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{hpgl_qshrink}
\alias{hpgl_qshrink}
\title{hpgl_qstats(): A hacked copy of Kwame's qsmooth/qstats code}
\usage{
hpgl_qshrink(exprs = NULL, groups = NULL, refType = "mean",
  groupLoc = "mean", window = 99, verbose = FALSE, groupCol = NULL,
  plot = TRUE, ...)
}
\description{
I made a couple small changes to Kwame's qstats() function to make
it not fail when on corner-cases.  I sent him a diff, but haven't
checked to see if it was useful yet.
}

