% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{hpgl_qstats}
\alias{hpgl_qstats}
\title{hpgl_qstats(): A hacked copy of Kwame's qsmooth/qstats code}
\usage{
hpgl_qstats(exprs, groups, refType = "mean", groupLoc = "mean",
  window = 99)
}
\description{
I made a couple small changes to Kwame's qstats() function to make
it not fail when on corner-cases.  I sent him a diff, but haven't
checked to see if it was useful yet.
}

