% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_norm.r
\name{hpgl_qstats}
\alias{hpgl_qstats}
\title{A hacked copy of Kwame's qsmooth/qstats code.}
\usage{
hpgl_qstats(data, groups, refType = "mean", groupLoc = "mean",
  window = 99)
}
\arguments{
\item{data}{Initial count data}

\item{groups}{Experimental conditions as a factor.}

\item{refType}{Method to separate groups, mean or median.}

\item{groupLoc}{I don't remember what this is for.}

\item{window}{Window for basking!}
}
\value{
Some new data.
}
\description{
I made a couple small changes to Kwame's qstats() function to make
it not fail when on corner-cases.  I sent him a diff, but haven't
checked to see if it was useful yet.
}
\examples{
\dontrun{
qstatted <- hpgl_qstats(data, conditions)
}
}

