% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{hpgl_rpkm}
\alias{hpgl_rpkm}
\title{Express a data frame of counts as reads per kilobase(gene) per
million(library).}
\usage{
hpgl_rpkm(df, annotations = gene_annotations)
}
\arguments{
\item{df}{a data frame of counts, alternately an edgeR DGEList}

\item{annotations}{containing gene lengths, defaulting to
'gene_annotations'}
}
\value{
rpkm_df a data frame of counts expressed as rpkm
}
\description{
Express a data frame of counts as reads per kilobase(gene) per
million(library).
}
\examples{
## rpkm_df = hpgl_rpkm(df, annotations=gene_annotations)
}
\seealso{
\code{\link{edgeR}} and \code{\link{cpm}}
}

