% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_convert.r
\name{hpgl_rpkm}
\alias{hpgl_rpkm}
\title{Reads/(kilobase(gene) * million reads)}
\usage{
hpgl_rpkm(count_table, ...)
}
\arguments{
\item{count_table}{Data frame of counts, alternately an edgeR DGEList.}

\item{...}{extra options including annotations for defining gene lengths.}
}
\value{
Data frame of counts expressed as rpkm.
}
\description{
Express a data frame of counts as reads per kilobase(gene) per
million(library). This function wraps EdgeR's rpkm in an attempt to make sure
that the required gene lengths get sent along.
}
\examples{
\dontrun{
 rpkm_df = hpgl_rpkm(df, annotations = gene_annotations)
}
}
\seealso{
\pkg{edgeR}
 \code{\link[edgeR]{cpm}} \code{\link[edgeR]{rpkm}}
}
