% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.r
\name{hpgl_rpkm}
\alias{hpgl_rpkm}
\title{\code{hpgl_rpkm()}  Reads/(kilobase(gene) * million reads)}
\usage{
hpgl_rpkm(df, annotations = get0("gene_annotations"))
}
\arguments{
\item{df}{a data frame of counts, alternately an edgeR DGEList}

\item{annotations}{containing gene lengths, defaulting to
'gene_annotations'}
}
\value{
rpkm_df a data frame of counts expressed as rpkm
}
\description{
Express a data frame of counts as reads per kilobase(gene) per
million(library).
}
\details{
This function wraps EdgeR's rpkm in an attempt to make sure that
the required gene lengths get sent along.
}
\examples{
\dontrun{
rpkm_df = hpgl_rpkm(df, annotations=gene_annotations)
}
}
\seealso{
\pkg{edgeR} and \code{\link[edgeR]{cpm}} \code{\link[edgeR]{rpkm}}
}

