% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_sample_heatmap}
\alias{hpgl_sample_heatmap}
\title{Make a heatmap.3 description of the similarity of the genes among samples.}
\usage{
hpgl_sample_heatmap(df = NULL, colors = NULL, design = NULL,
  expt = NULL, method = "pearson", names = NULL, type = "correlation",
  row = "batch", title = NULL, ...)
}
\arguments{
\item{df}{alternately a data frame which must be accompanied by}

\item{colors}{a color scheme}

\item{design}{a design matrix and}

\item{expt}{an expt set of samples}

\item{method}{correlation statistic to use.  Defaults to euclidean.}
}
\value{
a recordPlot() heatmap describing the samples.
}
\description{
Make a heatmap.3 description of the similarity of the genes among samples.
}
\seealso{
\code{\link{brewer.pal}},
\code{\link{heatmap.3}}, \code{\link{recordPlot}}
}

