% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_sample_heatmap}
\alias{hpgl_sample_heatmap}
\title{hpgl_sample_heatmap()  Make a heatmap.3 description of the similarity of the genes among samples.}
\usage{
hpgl_sample_heatmap(data, colors = NULL, design = NULL, names = NULL,
  title = NULL, Rowv = FALSE, ...)
}
\arguments{
\item{data}{an expt/expressionset/dataframe set of samples}

\item{colors}{default=NULL  a color scheme}

\item{design}{default=NULL  a design matrix}

\item{names}{default=NULL  add names?}

\item{title}{default=NULL  title of the plot.}
}
\value{
a recordPlot() heatmap describing the samples.
}
\description{
hpgl_sample_heatmap()  Make a heatmap.3 description of the similarity of the genes among samples.
}
\seealso{
\code{\link{brewer.pal}},
\code{\link{heatmap.3}}, \code{\link{recordPlot}}
}

