% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_scatter}
\alias{hpgl_scatter}
\title{Make a pretty scatter plot between two sets of numbers}
\usage{
hpgl_scatter(df, tooltip_data = NULL, color = "black",
  gvis_filename = NULL, size = 3)
}
\arguments{
\item{df}{a dataframe likely containing two columns}

\item{tooltip_data}{a df of tooltip information for gvis
graphs. NULL by default.}

\item{gvis_filename}{a filename to write a fancy html graph.
Defaults to NULL in which case the following parameter isn't needed.}
}
\value{
a ggplot2 scatter plot.
}
\description{
Make a pretty scatter plot between two sets of numbers
}
\examples{
## hpgl_scatter(lotsofnumbers_intwo_columns, tooltip_data=tooltip_dataframe, gvis_filename="html/fun_scatterplot.html")
}
\seealso{
\code{\link{hpgl_gvis_scatter}}, \code{\link{geom_scatter}},
\code{\link{hpgl_linear_scatter}}
}

