% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{hpgl_scatter}
\alias{hpgl_scatter}
\title{Make a pretty scatter plot between two sets of numbers.}
\usage{
hpgl_scatter(df, tooltip_data = NULL, color = "black",
  gvis_filename = NULL, size = 2)
}
\arguments{
\item{df}{a dataframe likely containing two columns}

\item{tooltip_data}{a df of tooltip information for gvis}

\item{color}{color of the dots on the graph.}

\item{gvis_filename}{a filename to write a fancy html graph.}

\item{size}{the size of the dots on the graph.}
}
\value{
a ggplot2 scatter plot.
}
\description{
Make a pretty scatter plot between two sets of numbers.
}
\examples{
\dontrun{
hpgl_scatter(lotsofnumbers_intwo_columns, tooltip_data=tooltip_dataframe,
             gvis_filename="html/fun_scatterplot.html")
}
}
\seealso{
\link{hpgl_gvis_scatter} \link[ggplot2]{geom_point}
\link{hpgl_linear_scatter}
}

