% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_smc}
\alias{hpgl_smc}
\title{hpgl_smc()  Make an R plot of the standard median correlation among samples.}
\usage{
hpgl_smc(data, colors = NULL, method = "pearson", names = NULL,
  title = NULL, ...)
}
\arguments{
\item{data}{an expt, expressionset, or data frame.}

\item{colors}{default=NULL  a color scheme}

\item{method}{default='pearson'  a correlation method to use.}

\item{names}{default=NULL  use pretty names for the samples?}

\item{title}{default=NULL  title for the graph.}
}
\value{
a recordPlot() of the standard median pairwise correlation
among the samples.  This will also write to an
open device.  The resulting plot measures the median correlation of
each sample among its peers.  It notes 1.5* the interquartile range
among the samples and makes a horizontal line at that correlation
coefficient.  Any sample which falls below this line is considered
for removal because it is much less similar to all of its peers.
}
\description{
This was written by a mix of Kwame Okrah <kokrah at gmail dot com>, Laura
Dillon <dillonl at umd dot edu>, and Hector Corrada Bravo <hcorrada at umd dot edu>
}
\examples{
## smc_plot = hpgl_smc(expt=expt)
}
\seealso{
\code{\link{hpgl_cor}}, \code{\link{matrixStats::rowMedians}},
\code{\link{quantile}}, \code{\link{diff}}, \code{\link{recordPlot}}
}

