% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.r
\name{hpgl_smd}
\alias{hpgl_smd}
\title{Make an R plot of the standard median distance among samples.}
\usage{
hpgl_smd(data, colors = NULL, names = NULL, method = "euclidean",
  title = NULL, ...)
}
\arguments{
\item{data}{an expt/expressionset/data frame of samples.}

\item{colors}{a color scheme}

\item{names}{use pretty names for the samples?}

\item{method}{a distance metric to use.}

\item{title}{title for the graph.}

\item{...}{parameters make me happy}
}
\value{
smd_plot a recordPlot of plot.  This will also write to an
open device.  This plot takes the median distance of each sample
with all of its peers.  It then calculates 1.5* the interquartile
range of distances.  Any sample which has a median distance greater
than this is considered for removal.
}
\description{
Make an R plot of the standard median distance among samples.
}
\examples{
\dontrun{
 smd_plot = hpgl_smd(expt=expt)
}
}
\seealso{
\code{\link{dist}}, \code{\link{quantile}},
\code{\link{diff}}, \code{\link{recordPlot}}
}

