% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{hpgl_volcano_plot}
\alias{hpgl_volcano_plot}
\title{Make a pretty Volcano plot!}
\usage{
hpgl_volcano_plot(toptable_data, tooltip_data = NULL, gvis_filename = NULL,
  fc_cutoff = 0.8, p_cutoff = 0.05, size = 2, alpha = 0.6, ...)
}
\arguments{
\item{toptable_data}{a dataframe from limma's toptable which
includes log(fold change) and an adjusted p-value.}

\item{tooltip_data}{a df of tooltip information for gvis
graphs. NULL by default.}

\item{gvis_filename}{a filename to write a fancy html graph.
Defaults to NULL in which case the following parameter isn't needed.}

\item{fc_cutoff}{a cutoff defining the minimum/maximum fold change
for interesting.  This is log, so I went with +/- 0.8 mostly
arbitrarily as the default.}

\item{p_cutoff}{a cutoff defining significant from not.
Defaults to 0.05.}

\item{size}{how big are the dots?  Defaults to 2.}

\item{alpha}{how transparent to make the dots.  Defaults to 0.6.}
}
\value{
a ggplot2 MA scatter plot.  This is defined as the
-log10(p-value) with respect to log(fold change).  The cutoff
values are delineated with lines and mark the boundaries between
'significant' and not.  This will make a fun clicky googleVis graph
if requested.
}
\description{
Make a pretty Volcano plot!
}
\examples{
## hpgl_volcano_plot(toptable_data, gvis_filename="html/fun_ma_plot.html")
## Currently this assumes that a variant of toptable was used which
## gives adjusted p-values.  This is not always the case and I should
## check for that, but I have not yet.
}
\seealso{
\code{\link{hpgl_gvis_ma_plot}}, \code{\link{toptable}},
\code{\link{voom}}, \code{\link{voomMod}}, \code{\link{hpgl_voom}},
\code{\link{lmFit}}, \code{\link{makeContrasts}},
\code{\link{contrasts.fit}}
}

