% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{hpgl_voom}
\alias{hpgl_voom}
\title{hpgl_voom():  A slight modification of limma's voom() function.
Estimate mean-variance relationship between samples and generate
'observational-level weights' in preparation for linear modelling
RNAseq data.  This particular implementation was primarily scabbed
from cbcbSEQ, but changes the mean-variance plot slightly and
attempts to handle corner cases where the sample design is
confounded by setting the coefficient to 1 for those samples rather
than throwing an unhelpful error.  Also, the Elist output gets a
'plot' slot which contains the plot rather than just printing it.}
\usage{
hpgl_voom(dataframe, model, libsize = NULL, stupid = FALSE,
  logged = FALSE, converted = FALSE)
}
\arguments{
\item{dataframe}{a dataframe of sample counts which have been
normalized and log transformed}

\item{model}{an experimental model defining batches/conditions/etc}

\item{libsize}{the size of the libraries (usually provided by
edgeR).  NULL by default.}

\item{stupid}{An TRUE/FALSE of whether or not to cheat when the
resulting matrix is not solvable. FALSE by default.}
}
\value{
an EList containing the following information:
  E = The normalized data
  weights = The weights of said data
  design = The resulting design
  lib.size = The size in pseudocounts of the library
  plot = A ggplot of the mean/variance trend with a blue loess fit and red trend fit
}
\description{
hpgl_voom():  A slight modification of limma's voom() function.
Estimate mean-variance relationship between samples and generate
'observational-level weights' in preparation for linear modelling
RNAseq data.  This particular implementation was primarily scabbed
from cbcbSEQ, but changes the mean-variance plot slightly and
attempts to handle corner cases where the sample design is
confounded by setting the coefficient to 1 for those samples rather
than throwing an unhelpful error.  Also, the Elist output gets a
'plot' slot which contains the plot rather than just printing it.
}
\examples{
## funkytown = hpgl_voom(samples, model)
}
\seealso{
\code{\link{voom}}, \code{\link{voomMod}}, \code{\link{lmFit}}
}

