% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_limma.r
\name{hpgl_voom}
\alias{hpgl_voom}
\title{A slight modification of limma's voom().}
\usage{
hpgl_voom(dataframe, model = NULL, libsize = NULL,
  normalize.method = "none", span = 0.5, stupid = FALSE,
  logged = FALSE, converted = FALSE, ...)
}
\arguments{
\item{dataframe}{Dataframe of sample counts which have been normalized and
log transformed.}

\item{model}{Experimental model defining batches/conditions/etc.}

\item{libsize}{Size of the libraries (usually provided by edgeR).}

\item{normalize.method}{Normalization method used in voom().}

\item{span}{The span used in voom().}

\item{stupid}{Cheat when the resulting matrix is not solvable?}

\item{logged}{Is the input data is known to be logged?}

\item{converted}{Is the input data is known to be cpm converted?}

\item{...}{Extra arguments are passed to arglist.}
}
\value{
EList containing the following information:
E = The normalized data
weights = The weights of said data
design = The resulting design
lib.size = The size in pseudocounts of the library
plot = A ggplot of the mean/variance trend with a blue loess fit and red trend fit
}
\description{
Estimate mean-variance relationship between samples and generate
'observational-level weights' in preparation for linear modeling RNAseq data.
This particular implementation was primarily scabbed from cbcbSEQ, but
changes the mean-variance plot slightly and attempts to handle corner cases
where the sample design is confounded by setting the coefficient to 1 for
those samples rather than throwing an unhelpful error.  Also, the Elist
output gets a 'plot' slot which contains the plot rather than just printing
it.
}
\examples{
\dontrun{
 funkytown = hpgl_voom(samples, model)
}
}
\seealso{
\pkg{limma} \pkg{ggplot2}
}
