% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.r
\name{init_xlsx}
\alias{init_xlsx}
\title{Initialize an xlsx file with a little bit of logic to make sure there are no
annoying downstream errors.}
\usage{
init_xlsx(excel = "excel/something.xlsx")
}
\arguments{
\item{excel}{Excel file to create.}
}
\description{
Initialize an xlsx file with a little bit of logic to make sure there are no
annoying downstream errors.
}
