% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{intersect_significant}
\alias{intersect_significant}
\title{Find the sets of intersecting significant genes}
\usage{
intersect_significant(combined, lfc = 1, p = 0.05, padding_rows = 2,
  z = NULL, p_type = "adj", selectors = c("limma", "deseq", "edger"),
  order = "inverse", excel = "excel/intersect_significant.xlsx", ...)
}
\arguments{
\item{combined}{A result from combine_de_tables().}

\item{lfc}{Define significant via fold-change.}

\item{p}{Or p-value.}

\item{padding_rows}{How much space to put between groups of data?}

\item{z}{Use a z-score filter?}

\item{p_type}{Use normal or adjusted p-values.}

\item{excel}{An optional excel workbook to which to write.}

\item{first}{First method to include.}

\item{second}{Second method to include.}

\item{third}{Third method to include, or null.}

\item{fourth}{You get the idea, but at this point it will only include the full intersection.}

\item{fifth}{Ditto.}

\item{sixth}{Unlikely, but not completely impossible.}

\item{extra_annot}{Provide an extra set of annotation columns?}
}
\description{
Use extract_significant_genes() to find the points of agreement between limma/deseq/edger.
}
