% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{intersect_significant}
\alias{intersect_significant}
\title{Find the sets of intersecting significant genes}
\usage{
intersect_significant(
  combined,
  lfc = 1,
  p = 0.05,
  padding_rows = 2,
  z = NULL,
  p_type = "adj",
  selectors = c("limma", "deseq", "edger"),
  order = "inverse",
  excel = "excel/intersect_significant.xlsx",
  ...
)
}
\arguments{
\item{combined}{Result from combine_de_tables().}

\item{lfc}{Define significant via fold-change.}

\item{p}{Or p-value.}

\item{padding_rows}{How much space to put between groups of data?}

\item{z}{Use a z-score filter?}

\item{p_type}{Use normal or adjusted p-values.}

\item{selectors}{List of methods to intersect.}

\item{order}{When set to the default 'inverse', go from the set with the most
least intersection to the most. E.g. Start with abc,bc,ac,c,ab,b,a as
opposed to a,b,ab,c,ac,bc,abc.}

\item{excel}{An optional excel workbook to which to write.}

\item{...}{Extra arguments for extract_significant_genes() and friends.}
}
\value{
List containing the intersections between the various DE methods for
  both the up and down sets of genes.  It should also provide some venn
  diagrams showing the degree of similarity between the methods.
}
\description{
Use extract_significant_genes() to find the points of agreement between
limma/deseq/edger.
}
\examples{
 \dontrun{
 expt <- create_expt(metadata="some_metadata.xlsx", gene_info=funkytown)
 big_result <- all_pairwise(expt, model_batch=FALSE)
 pretty <- combine_de_tables(big_result, excel="excel/combined_expt.xlsx")
 intersect <- intersect_significant(pretty, excel="excel/intersecting_genes.xlsx")
}
}
