% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{kegg_get_orgn}
\alias{kegg_get_orgn}
\title{Search KEGG identifiers for a given species name.}
\usage{
kegg_get_orgn(species = "Leishmania", short = TRUE)
}
\arguments{
\item{species}{Search string (Something like 'Homo sapiens').}

\item{short}{Only pull the orgid?}
}
\value{
Data frame of possible KEGG identifier codes, genome ID numbers, species, and
    phylogenetic classifications.
}
\description{
KEGG identifiers do not always make sense.  For example, how am I supposed to remember that
Leishmania major is lmj?  This takes in a human readable string and finds the KEGG identifiers
that match it.
}
\examples{
\dontrun{
 fun = kegg_get_orgn('Canis')
## >     Tid     orgid      species                   phylogeny
## >  17 T01007   cfa Canis familiaris (dog) Eukaryotes;Animals;Vertebrates;Mammals
}
}
\seealso{
\pkg{RCurl}
}

