% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_kegg.r
\name{kegg_vector_to_df}
\alias{kegg_vector_to_df}
\title{Convert a potentially non-unique vector from kegg into a normalized data
frame.}
\usage{
kegg_vector_to_df(vector, final_colname = "first", flatten = TRUE)
}
\arguments{
\item{vector}{Information from KEGGREST}

\item{final_colname}{Column name for the new information}

\item{flatten}{Flatten nested data?}
}
\value{
A normalized data frame of gene IDs to whatever.
}
\description{
This function seeks to reformat data from KEGGREST into something which is
rather easier to use.
}
\details{
This could probably benefit from a tidyr-ish revisitation.
}
\seealso{
[KEGGREST] [load_kegg_annotations()]
}
