% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_limma.r
\name{limma_coefficient_scatter}
\alias{limma_coefficient_scatter}
\title{Plot out 2 coefficients with respect to one another from limma.}
\usage{
limma_coefficient_scatter(output, toptable = NULL, x = 1, y = 2,
  gvis_filename = NULL, gvis_trendline = TRUE, z = 1.5,
  tooltip_data = NULL, base_url = NULL, color_low = "#7B9F35",
  color_high = "#DD0000", ...)
}
\arguments{
\item{output}{Set of pairwise comparisons provided by limma_pairwise().}

\item{toptable}{Use this to get up/downs and color them on the scatter plot.}

\item{x}{Name or number of the x-axis coefficient column to extract.}

\item{y}{Name or number of the y-axis coefficient column to extract}

\item{gvis_filename}{Filename for plotting gvis interactive graphs of the data.}

\item{gvis_trendline}{Add a trendline to the gvis plot?}

\item{z}{How far from the median to color the plot red and green.}

\item{tooltip_data}{Dataframe of gene annotations to be used in the gvis plot.}

\item{base_url}{Basename for gvis plots.}

\item{color_low}{Color for the ups.}

\item{color_high}{Color for the downs.}

\item{...}{More parameters to make you happy!}
}
\value{
Ggplot2 plot showing the relationship between the two coefficients.
}
\description{
It can be nice to see a plot of two coefficients from a limma comparison with respect to one
another. This hopefully makes that easy.
}
\examples{
\dontrun{
 pretty = coefficient_scatter(limma_data, x="wt", y="mut")
}
}
\seealso{
\link{plot_linear_scatter} \link{limma_pairwise}
}

