% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{limma_scatter}
\alias{limma_scatter}
\title{Plot arbitrary data from limma}
\usage{
limma_scatter(all_pairwise_result, first_table = 1, first_column = "logFC",
  second_table = 2, second_column = "logFC", type = "linear_scatter", ...)
}
\arguments{
\item{all_pairwise_result}{The result from calling balanced_pairwise()}

\item{type}{A type of scatter plot (linear model, distance, vanilla)}

\item{...}{so that you may feed it the gvis/tooltip information to make clicky graphs if so desired.}

\item{first_name}{A table inside all_pairwise_result$limma_result}

\item{first_type}{A column within the chosen table}

\item{second_name}{Another table inside all_pairwise_result$limma_result}

\item{second_type}{A column to compare against}
}
\value{
a hpgl_linear_scatter() set of plots comparing the chosen columns
If you forget to specify tables to compare, it will try the first vs the second.
}
\description{
Plot arbitrary data from limma
}
\examples{
## compare_logFC = limma_scatter(all_pairwise, first_table="wild_type", second_column="mutant", first_table="AveExpr", second_column="AveExpr")
## compare_B = limma_scatter(all_pairwise, first_column="B", second_column="B")
}
\seealso{
\code{\link{hpgl_linear_scatter}}, \code{\link{topTable}},
}

