% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{limma_subset}
\alias{limma_subset}
\title{limma_subset():  A quick and dirty way to pull the top/bottom genes from toptable()}
\usage{
limma_subset(table, n = NULL, z = NULL)
}
\arguments{
\item{n}{A number of genes to keep}

\item{z}{A number of z-scores from the mean}

\item{df}{The original data from limma}
}
\value{
a dataframe subset from toptable
}
\description{
limma_subset():  A quick and dirty way to pull the top/bottom genes from toptable()
}
\examples{
## subset = limma_subset(df, n=400)
## subset = limma_subset(df, z=1.5)
}
\seealso{
\code{\link{limma}}
}

