% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_shared.r
\name{load_annotations}
\alias{load_annotations}
\title{Use one of the load_*_annotations() functions to gather annotation data.}
\usage{
load_annotations(type = NULL, ...)
}
\arguments{
\item{type}{Explicitly state the type of annotation data to load.  If not
provided, try to figure it out automagically.}

\item{...}{Arguments passed to the other load_*_annotations().}
}
\value{
Some annotations, hopefully.
}
\description{
We should be able to have an agnostic annotation loader which can take some
standard arguments and figure out where to gather data on its own.
}
\examples{
 gff_annotations <- load_annotations(type = "gff", gff = example_gff)
 dim(gff_annotations)
}
\seealso{
[load_biomart_annotations()] [load_gff_annotations()] [load_genbank_annotations()]
 [load_kegg_annotations()] [load_trinotate_annotations()] [load_microbesonline_annotations()]
 [load_uniprot_annotations()]
}
