% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{load_biomart_go}
\alias{load_biomart_go}
\title{Extract gene ontology information from biomart.}
\usage{
load_biomart_go(
  species = "hsapiens",
  overwrite = FALSE,
  do_save = TRUE,
  host = NULL,
  trymart = "ENSEMBL_MART_ENSEMBL",
  archive = TRUE,
  default_hosts = c("useast.ensembl.org", "uswest.ensembl.org", "www.ensembl.org",
    "asia.ensembl.org"),
  year = NULL,
  month = NULL,
  drop_haplotypes = TRUE,
  trydataset = NULL,
  dl_rows = c("ensembl_gene_id", "go_accession"),
  dl_rowsv2 = c("ensembl_gene_id", "go_id")
)
}
\arguments{
\item{species}{Species to query.}

\item{overwrite}{Overwrite existing savefile?}

\item{do_save}{Create a savefile of the annotations? (if not false, then a filename.)}

\item{host}{Ensembl hostname to use.}

\item{trymart}{Biomart has become a circular dependency, this makes me sad,
now to list the marts, you need to have a mart loaded.}

\item{archive}{Try an archive server instead of a mirror?  If this is a
character, it will assume it is a specific archive hostname.}

\item{default_hosts}{List of biomart mirrors to try.}

\item{year}{Choose specific year(s) for the archive servers?}

\item{month}{Choose specific month(s) for the archive servers?}

\item{drop_haplotypes}{Some chromosomes have stupid names because they are
from non-standard haplotypes and they should go away.  Setting this to
false stops that.}

\item{dl_rows}{List of rows from the final biomart object to download.}

\item{dl_rowsv2}{A second list of potential rows.}

\item{secondtry}{The newer mart name.}
}
\value{
List containing the following:  data frame of ontology data, a copy
 of the biomart instance for further querying, the host queried, the biomart
 queried, a vector providing the attributes queried, and the ensembl dataset
 queried.
}
\description{
I perceive that every time I go to acquire annotation data from biomart, they
have changed something important and made it more difficult for me to find
what I want. I recently found the *.archive.ensembl.org, and so this function
uses that to try to keep things predictable, if not consistent.
}
\details{
Tested in test_40ann_biomart.R
This function makes a couple of attempts to pick up the correct tables from
biomart.  It is worth noting that it uses the archive.ensembl host(s) because
of changes in table organization after December 2015 as well as an attempt to
keep the annotation sets relatively consistent.
}
\examples{
 hs_biomart_ontology <-load_biomart_go()
 summary(hs_biomart_ontology)
 dim(hs_biomart_ontology$go)
}
\seealso{
[biomaRt::listMarts()] [biomaRt::useDatasets()] [biomaRt::getBM()]
}
