% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.r
\name{load_biomart_go}
\alias{load_biomart_go}
\title{Extract gene ontology information from biomart.}
\usage{
load_biomart_go(species = "hsapiens", overwrite = FALSE,
  do_save = TRUE, host = "dec2015.archive.ensembl.org",
  trymart = "ENSEMBL_MART_ENSEMBL", secondtry = "_gene",
  dl_rows = c("ensembl_gene_id", "go_accession"),
  dl_rowsv2 = c("ensembl_gene_id", "go_id"))
}
\arguments{
\item{species}{Species to query.}

\item{overwrite}{Overwrite existing savefile?}

\item{do_save}{Create a savefile of the annotations? (if not false, then a filename.)}

\item{host}{Ensembl hostname to use.}

\item{trymart}{Default mart to try, newer marts use a different notation.}

\item{secondtry}{The newer mart name.}

\item{dl_rows}{List of rows from the final biomart object to download.}

\item{dl_rowsv2}{A second list of potential rows.}
}
\value{
List containing the following:  data frame of ontology data, a copy
of the biomart instance for further querying, the host queried, the biomart
queried, a vector providing the attributes queried, and the ensembl dataset
queried.
}
\description{
I perceive that every time I go to acquire annotation data from biomart, they
have changed something important and made it more difficult for me to find
what I want. I recently found the *.archive.ensembl.org, and so this function
uses that to try to keep things predictable, if not consistent.
}
\details{
Tested in test_40ann_biomart.R
This function makes a couple of attempts to pick up the correct tables from
biomart.  It is worth noting that it uses the archive.ensembl host(s) because
of changes in table organization after December 2015 as well as an attempt to
keep the annotation sets relatively consistent.
}
\examples{
\dontrun{
 tt = get_biomart_ontologies()
}
}
\seealso{
\pkg{biomaRt}
\code{\link[biomaRt]{listMarts}} \code{\link[biomaRt]{useDataset}}
\code{\link[biomaRt]{getBM}}
}
\author{
atb
}
