% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_genbank.R
\name{load_genbank_annotations}
\alias{load_genbank_annotations}
\title{Given a genbank accession, make a txDb object along with sequences, etc.}
\usage{
load_genbank_annotations(
  accession = "AE009949",
  db = "nucleotide",
  file = NULL,
  sequence = TRUE,
  type = "CDS",
  savetxdb = FALSE,
  restez_db = "/sw/local/genbank/current"
)
}
\arguments{
\item{accession}{Accession to download and import.}

\item{db}{Entrez Database to query}

\item{file}{Use a file instead of downloading the accession?}

\item{sequence}{Download the sequence with the annotations?}

\item{type}{Extract entries of this type.}

\item{savetxdb}{Attempt saving a txdb object?}

\item{restez_db}{Location of local genbank used by restez.}
}
\value{
List containing a txDb, sequences, and some other stuff which I
 haven't yet finalized.
}
\description{
Let us admit it, sometimes biomart is a pain.  It also does not have easily
accessible data for microbes.  Genbank does!
}
\details{
Tested in test_40ann_biomartgenbank.R and test_70expt_spyogenes.R
This primarily sets some defaults for the genbankr service in order to
facilitate downloading genomes from genbank and dumping them into a local
txdb instance.
}
\examples{
## 202601 eutils seems to be having some trouble.

test_accession <- "AE009948"
gbk_file <- try(download_gbk(accessions = test_accession))
if (! "try-error" \%in\% class(gbk_file)) {
  removed <- file.remove(gbk_file[["written_file"]])
}

sagalacticae_genbank_annot <- load_genbank_annotations(accession = test_accession, type = "CDS")
if (! "try-error" \%in\% class(sagalacticae_genbank_annot)) {
  dim(sagalacticae_genbank_annot)
}
}
\seealso{
[Biostrings] [GenomicFeatures] [genbankr::import()] [genbankr::readGenBank()]
}
