% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_genbank.r
\name{load_genbank_annotations}
\alias{load_genbank_annotations}
\title{Given a genbank accession, make a txDb object along with sequences, etc.}
\usage{
load_genbank_annotations(accession = "AE009949", reread = TRUE,
  savetxdb = FALSE)
}
\arguments{
\item{accession}{Accession to download and import}

\item{reread}{Re-read (download) the file from genbank}

\item{savetxdb}{Attempt saving a txdb object?}
}
\value{
List containing a txDb, sequences, and some other stuff which I
  haven't yet finalized.
}
\description{
Let us admit it, sometimes biomart is a pain.  It also does not have easily
accessible data for microbes.  Genbank does!
}
\details{
Tested in test_40ann_biomartgenbank.R and test_70expt_spyogenes.R
This primarily sets some defaults for the genbankr service in order to
facilitate downloading genomes from genbank and dumping them into a local
txdb instance.
}
\examples{
\dontrun{
 txdb_result <- load_genbank_annotations(accession="AE009948", savetxdb=TRUE)
}
}
\seealso{
\pkg{genbankr} \pkg{rentrez}
 \code{\link[genbankr]{import}}
}
