% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{load_gmt_signatures}
\alias{load_gmt_signatures}
\title{Load signatures from either a gmt file, xml file, or directly from the
GSVAdata data set in R.}
\usage{
load_gmt_signatures(
  signatures = "c2BroadSets",
  data_pkg = "GSVAdata",
  signature_category = "c2"
)
}
\arguments{
\item{signatures}{Either the filename downloaded or the variable's name as
found in the environment created by data_pkg.}

\item{data_pkg}{Used when signatures is not a filename to load a data
package, presumably GSVAdata.}

\item{signature_category}{Probably not needed unless you download a signature
file containing lots of different categories.}
}
\value{
signature dataset which may be used by gsva()
}
\description{
There are a bunch of places from which to acquire signature data.  This
function attempts to provide a single place to load them.  The easiest way to
get up to date signatures is to download them from msigdb and set the
signatures parameter to the downloaded filename.
}
