% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{load_go_terms}
\alias{load_go_terms}
\title{Retrieve GO terms associated with a set of genes.}
\usage{
load_go_terms(orgdb, gene_ids, keytype = "ENSEMBL")
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Identifiers of the genes to retrieve annotations.}

\item{keytype}{the mysterious keytype returns yet again to haunt my dreams}
}
\value{
Data frame of gene IDs, go terms, and names.
}
\description{
AnnotationDbi provides a reasonably complete set of GO mappings between gene ID and
ontologies.  This will extract that table for a given set of gene IDs.
}
\examples{
\dontrun{
go_terms <- load_go_terms(org, c("a","b"))
}
}
\seealso{
\link[AnnotationDbi]{select}
}

