% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{load_go_terms}
\alias{load_go_terms}
\title{Retrieve GO terms associated with a set of genes}
\usage{
load_go_terms(orgdb, gene_ids, keytype = "ENSEMBL")
}
\arguments{
\item{orgdb}{An OrganismDb instance}

\item{gene_ids}{Identifiers of the genes to retrieve annotations for}
}
\value{
a data frame of gene IDs, go terms, and names.
}
\description{
Retrieve GO terms associated with a set of genes
}

