% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{load_host_annotations}
\alias{load_host_annotations}
\title{Load organism annotation data (mouse/human)}
\usage{
load_host_annotations(orgdb, gene_ids, keytype = "ENSEMBL",
  fields = c("TXCHROM", "GENENAME", "TXSTRAND", "TXSTART", "TXEND"),
  biomart_dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{orgdb}{An OrganismDb instance}

\item{gene_ids}{Identifiers of the genes to retrieve annotations for}

\item{fields}{Columns to include in the output}

\item{biomart_dataset}{Name of the biomaRt dataset to query for gene type}
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the OrganismDbi interface.
}

