% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_kegg.r
\name{load_kegg_annotations}
\alias{load_kegg_annotations}
\title{Create a data frame of pathways to gene IDs from KEGGREST}
\usage{
load_kegg_annotations(species = "ecoli", abbreviation = NULL,
  flatten = TRUE)
}
\arguments{
\item{species}{String to use to query KEGG abbreviation.}

\item{abbreviation}{If you already know the abbreviation, use it.}

\item{flatten}{Flatten nested tables?}
}
\value{
dataframe with rows of KEGG gene IDs and columns of NCBI gene IDs and KEGG paths.
}
\description{
Create a data frame of pathways to gene IDs from KEGGREST
}
