% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{load_kegg_mapping}
\alias{load_kegg_mapping}
\title{Creates a gene/KEGG mapping dataframe.}
\usage{
load_kegg_mapping(orgdb, gene_ids, keytype = "ENSEMBL")
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Identifiers of the genes to retrieve annotations.}

\item{keytype}{the keytype, damn I really need to read this code}
}
\value{
Df of kegg mappings
}
\description{
In much the same way AnnotationDbi provides GO data, it also provides KEGG data.
}
\examples{
\dontrun{
kegg_data <- load_kegg_mapping(org, c("a","b"))
}
}
\seealso{
\link[AnnotationDbi]{select}
}

