% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.r
\name{load_microbesonline_annotations}
\alias{load_microbesonline_annotations}
\title{Skip the db and download all the text annotations for a given species.}
\usage{
load_microbesonline_annotations(species = NULL, id = NULL)
}
\arguments{
\item{species}{Microbesonline species.}

\item{id}{Microbesonline ID to query.}
}
\value{
Dataframe containing the annotation information.
}
\description{
The microbesonline publicly available mysqldb is rather more complex than I
prefer.  This skips that process and just grabs a tsv copy of everything and
loads it into a dataframe.  I have not yet figured out how to so-easily query
microbesonline for species IDs, thus one will have to manually query the
database to find species of interest.
}
\details{
Tested in test_70expt_spyogenes.R
There is so much awesome information in microbesonline, but damn is it
annoying to download. This function makes that rather easier, or so I hope at
least.
}
\examples{
 pa14_microbesonline_annot <- load_microbesonline_annotations(species = "PA14")
 colnames(pa14_microbesonline_annot)
}
\seealso{
\pkg{rvest}
}
