% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.r
\name{load_microbesonline_go}
\alias{load_microbesonline_go}
\title{Extract the set of GO categories by microbesonline locus}
\usage{
load_microbesonline_go(id = "160490", id_column = "name",
  data_column = "GO", name = NULL)
}
\arguments{
\item{id}{Which species to query.}

\item{id_column}{This no longer uses MySQL, so which column from the html table to pull?}

\item{data_column}{Similar to above, there are lots of places from which one might
extract the data.}

\item{name}{Allowing for non-specific searches by species name.}
}
\value{
data frame of GO terms from www.microbesonline.org
}
\description{
The microbesonline is such a fantastic resource, it is a bit of a shame that it is such a pain
to query.
}
\details{
Tested in test_42ann_microbes.R
I am not 100% certain that this is giving me the full correct set of gene ontology accessions.
At the very least, it does return a large number of them, which is a start.
}
\examples{
\dontrun{
 go_df <- get_loci_go(id="160490")
}
}
