% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040_annotation_orgdb.r
\name{load_orgdb_annotations}
\alias{load_orgdb_annotations}
\title{Load organism annotation data from an orgdb sqlite package.}
\usage{
load_orgdb_annotations(
  orgdb = NULL,
  gene_ids = NULL,
  include_go = FALSE,
  keytype = "ensembl",
  strand_column = "cdsstrand",
  start_column = "cdsstart",
  end_column = "cdsend",
  chromosome_column = "cdschrom",
  type_column = "gene_type",
  name_column = "cdsname",
  fields = NULL,
  sum_exon_widths = FALSE
)
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Search for a specific set of genes?}

\item{include_go}{Ask the Dbi for gene ontology information?}

\item{keytype}{mmm the key type used?}

\item{strand_column}{There are a few fields I want to gather by default:
start, end, strand, chromosome, type, and name; but these do not
necessarily have consistent names, use this column for the chromosome
strand.}

\item{start_column}{Use this column for the gene start.}

\item{end_column}{Use this column for the gene end.}

\item{chromosome_column}{Use this column to identify the chromosome.}

\item{type_column}{Use this column to identify the gene type.}

\item{name_column}{Use this column to identify the gene name.}

\item{fields}{Columns included in the output.}

\item{sum_exon_widths}{Perform a sum of the exons in the data set?}
}
\value{
Table of geneids, chromosomes, descriptions, strands, types, and lengths.
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the AnnotationDbi interface.
}
\details{
Tested in test_45ann_organdb.R
This defaults to a few fields which I have found most useful, but the brave
or pathological can pass it 'all'.
}
\examples{
 hs_orgdb_annot <- load_orgdb_annotations()
 summary(hs_orgdb_annot$genes)
}
\seealso{
\pkg{AnnotationDbi} \pkg{GenomicFeatures} \pkg{BiocGenerics}
 \code{\link[AnnotationDbi]{columns}} \code{\link[AnnotationDbi]{keytypes}}
 \code{\link[AnnotationDbi]{select}} \code{\link[GenomicFeatures]{exonsBy}}
}
