% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{load_orgdb_go}
\alias{load_orgdb_go}
\title{Retrieve GO terms associated with a set of genes.}
\usage{
load_orgdb_go(
  orgdb = NULL,
  gene_ids = NULL,
  keytype = "ensembl",
  columns = c("go", "goall", "goid")
)
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Identifiers of the genes to retrieve annotations.}

\item{keytype}{The mysterious keytype returns yet again to haunt my dreams.}

\item{columns}{The set of columns to request.}
}
\value{
Data frame of gene IDs, go terms, and names.
}
\description{
AnnotationDbi provides a reasonably complete set of GO mappings between gene
ID and ontologies.  This will extract that table for a given set of gene
IDs.
}
\details{
Tested in test_45ann_organdb.R
This is a nice way to extract GO data primarily because the Orgdb data sets
are extremely fast and flexible, thus by changing the keytype argument, one
may use a lot of different ID types and still score some useful ontology data.
}
\examples{
 drosophila_orgdb_go <- load_orgdb_go(orgdb="org.Dm.eg.db")
 head(drosophila_orgdb_go)
}
\seealso{
\pkg{AnnotationDbi} \pkg{GO.db} \pkg{magrittr}
 \code{\link[AnnotationDbi]{select}}
}
\author{
I think Keith provided the initial implementation of this, but atb
 messed with it pretty extensively.
}
