% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{load_parasite_annotations}
\alias{load_parasite_annotations}
\title{I see no reason to have load_host_annotations and load_parasite_annotations.}
\usage{
load_parasite_annotations(...)
}
\arguments{
\item{...}{Arguments to be passed to load_annotations.}
}
\description{
Thus I am making them both into aliases to load_annotations.
}

