% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/045_annotation_txt.r
\name{load_trinotate_annotations}
\alias{load_trinotate_annotations}
\title{Read a csv file from trinotate and make an annotation data frame.}
\usage{
load_trinotate_annotations(trinotate = "reference/trinotate.csv")
}
\arguments{
\item{trinotate}{CSV of trinotate annotation data.}
}
\value{
Dataframe of fun data.
}
\description{
Trinotate performs some neat sequence searches in order to seek out likely
annotations for the trinity contigs.  The resulting csv file is encoded in a
peculiar fashion, so this function attempts to make it easier to read and put
them into a format usable in an expressionset.
}
\examples{
 trinotate <- system.file("sb/trinotate_head.csv.xz", package="hpgltools")
 a_few_trinotate <- load_trinotate_annotations(trinotate=trinotate)
 dim(a_few_trinotate)
}
