% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_txt.r
\name{load_trinotate_annotations}
\alias{load_trinotate_annotations}
\title{Read a csv file and make an annotation data frame.}
\usage{
load_trinotate_annotations(trinotate = "reference/trinotate.csv")
}
\arguments{
\item{trinotate}{CSV of trinotate annotation data.}
}
\value{
Dataframe of fun data.
}
\description{
Yay!
}
