% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_txt.R
\name{load_trinotate_go}
\alias{load_trinotate_go}
\title{Read a csv file from trinotate and extract ontology data from it.}
\usage{
load_trinotate_go(
  trinotate = "reference/trinotate.csv",
  blast2go_column = "gene_ontology_BLASTX",
  pfam_column = "gene_ontology_pfam",
  length_column = "transcript",
  fill = 1500,
  collapse = TRUE,
  id_column = "#gene_id"
)
}
\arguments{
\item{trinotate}{CSV of trinotate annotation data.}

\item{blast2go_column}{Column name containing BLAST2GO data.}

\item{pfam_column}{Column containing data from pfam searches.}

\item{length_column}{Column containing the gene lengths.}

\item{fill}{Cheat and fill in an arbitrary value for gene lengths if all else fails.}

\item{collapse}{Collapse isforms to genes?}

\item{id_column}{Column containing the gene IDs.}
}
\value{
List of the extracted GO data, a table of it, length data, and the
 resulting length table.
}
\description{
Trinotate performs some neat sequence searches in order to seek out likely
annotations for the trinity contigs.  This function extracts ontology data
from it.  Keep in mind that this data is primarily from Blast2GO.
}
\examples{
example_txt <- system.file("share", "sb", "trinotate_head.csv.xz", package = "hpgldata")

## Create a dataframe from the peculiarly formatted tsv/csv from trinotate.
trinotate_df <- load_trinotate_annotations(example_txt)
head(trinotate_df)

## Extract the gene ontology information from trinotate.
trinotate_go <- load_trinotate_go(example_txt,
                                  blast2go_column = "gene_ontology_blast")
head(trinotate_go[["go_table"]])
}
\seealso{
[load_trinotate_annotations()]
}
