% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_uniprot.R
\name{load_uniprot_annotations}
\alias{load_uniprot_annotations}
\title{Download the txt uniprot data for a given accession/species.}
\usage{
load_uniprot_annotations(
  accession = NULL,
  species = "H37Rv",
  taxonomy = NULL,
  all = FALSE,
  first = FALSE
)
}
\arguments{
\item{accession}{Which accession to grab?}

\item{species}{Or perhaps species?}

\item{taxonomy}{Query for a specific taxonomy ID rather than species/accession?}

\item{all}{If there are more than 1 hit, grab them all?}

\item{first}{Or perhaps just grab the first hit?}
}
\value{
A filename/accession tuple.
}
\description{
Uniprot is an astonishing resource, but man is it a pain to use.  Hopefully
this function will help.  It takes either a uniprot accession, taxonomy ID,
or species name and does its best to find the appropriate uniprot data.  This
is therefore primarily used by load_uniprot_annotations().
}
\examples{
 uniprot_sc_downloaded <- download_uniprot_proteome(species = "Saccharomyces cerevisiae S288c")
 uniprot_sc_downloaded$filename
 uniprot_sc_downloaded$species
}
\seealso{
[xml2] [rvest]
}
