% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_uniprot.r
\name{load_uniprot_go}
\alias{load_uniprot_go}
\title{Extract ontology information from a uniprot dataframe.}
\usage{
load_uniprot_go(input)
}
\arguments{
\item{input}{uniprot filename or dataframe.}
}
\value{
Ontology dataframe
}
\description{
Extract ontology information from a uniprot dataframe.
}
\examples{
 sc_uniprot_go <- load_uniprot_go(sc_uniprot_annot)
 head(sc_uniprot_go)
}
\seealso{
[load_uniprot_annotations()] [stringr] [tidyr]
}
