% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{loadme}
\alias{loadme}
\title{Load a backup rdata file}
\usage{
loadme(dir = "savefiles")
}
\arguments{
\item{dir}{Directory containing the RData.rda.xz file.}
}
\value{
a bigger global environment
}
\description{
I often use R over a sshfs connection, sometimes with significant latency, and I want to be able
to save/load my R sessions relatively quickly. Thus this function uses my backup directory to
load its R environment.
}
\examples{
\dontrun{
loadme()
saveme()
}
}
\seealso{
\link{load} \link{save}
}

