% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_filter.r
\name{lowfilter_counts}
\alias{lowfilter_counts}
\title{Call various low-count filters.}
\usage{
lowfilter_counts(count_table, type = "cbcb", p = 0.01, A = 1, k = 1,
  cv_min = 0.01, cv_max = 1000, thresh = 2, min_samples = 2, ...)
}
\arguments{
\item{count_table}{Some counts to filter.}

\item{type}{Filtering method to apply (cbcb, pofa, kofa, cv right now).}

\item{p}{Used by genefilter's pofa().}

\item{A}{Also for pofa().}

\item{k}{Used by genefilter's kofa().}

\item{cv_min}{Used by genefilter's cv().}

\item{cv_max}{Also used by cv().}

\item{thresh}{Minimum threshold across samples for cbcb.}

\item{min_samples}{Minimum number of samples for cbcb.}
}
\value{
Data frame of lowfiltered counts.
}
\description{
This calls the various filtering functions in genefilter along with suggestions made in our lab
meetings; defaulting to the threshold based filter suggested by Hector.
}
\examples{
\dontrun{
new <- lowfilter_counts(old)
}
}
\seealso{
\pkg{genefilter}
}

