% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{lowfilter_counts}
\alias{lowfilter_counts}
\title{lowfilter_counts(): A caller for different low-count filters}
\usage{
lowfilter_counts(count_table, type = "cbcb", p = 0.01, A = 1, k = 1,
  cv_min = 0.01, cv_max = 1000, thresh = 2, min_samples = 2)
}
\description{
lowfilter_counts(): A caller for different low-count filters
}

