% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_organismdbi}
\alias{make_eupath_organismdbi}
\title{Create an organismDbi instance for an eupathdb organism.}
\usage{
make_eupath_organismdbi(species = "Leishmania major strain Friedlin",
  dir = ".", overwrite = FALSE, kegg_abbreviation = NULL,
  db_version = NULL, parallel = TRUE)
}
\arguments{
\item{species}{A species in the eupathDb metadata.}

\item{dir}{Directory in which to build the packages.}
}
\value{
The result of attempting to install the organismDbi package.
}
\description{
Create an organismDbi instance for an eupathdb organism.
}
\author{
Keith Hughitt
}
