% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_organismdbi}
\alias{make_eupath_organismdbi}
\title{Create an organismDbi instance for an eupathdb organism.}
\usage{
make_eupath_organismdbi(species = "Leishmania major strain Friedlin",
  entry = NULL, dir = "eupathdb", reinstall = FALSE, metadata = NULL,
  kegg_abbreviation = NULL, exclude_join = "ENTREZID", ...)
}
\arguments{
\item{species}{A species in the eupathDb metadata.}

\item{entry}{A row from the eupathdb metadataframe.}

\item{dir}{Directory in which to build the packages.}

\item{reinstall}{Overwrite existing data files?}

\item{metadata}{Use a pre-existing metadata table, or download a new one.}

\item{kegg_abbreviation}{For when we cannot automagically find the kegg species id.}

\item{exclude_join}{I had a harebrained idea to automatically set up the
joins between columns of GO.db/reactome.db/orgdb/txdb objects.  This
variable is intended to exclude columns with common IDs that might
multi-match spuriously -- I think in the end I killed the idea though,
perhaps this should be removed or resurrected.}

\item{...}{Extra arguments when downloading metadata.}
}
\value{
The result of attempting to install the organismDbi package.
}
\description{
Create an organismDbi instance for an eupathdb organism.
}
\author{
Keith Hughitt
}
