% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_orgdb}
\alias{make_eupath_orgdb}
\title{Create an orgdb SQLite database from the tables in eupathdb.}
\usage{
make_eupath_orgdb(species = NULL, entry = NULL, dir = "eupathdb",
  version = NULL, kegg_abbreviation = NULL, reinstall = FALSE,
  metadata = NULL, ...)
}
\arguments{
\item{species}{A specific species ID to query}

\item{entry}{If not provided, then species will get this, it contains all the information.}

\item{dir}{Where to put all the various temporary files.}

\item{version}{Which version of the eupathdb to use for creating this package?}

\item{kegg_abbreviation}{If known, provide the kegg abbreviation.}

\item{reinstall}{Re-install an already existing orgdb?}

\item{metadata}{Use an existing metadata table to get the entry?}

\item{...}{Extra parameters when searching for metadata}
}
\value{
Currently only the name of the installed package.  This should
probably change.
}
\description{
This function has passed through multiple iterations as the preferred
method(s) for accessing data in the eupathdb has changed.  It currently uses
my empirically defined set of queries against the eupathdb webservices.  As a
result, I have made some admittedly bizarre choices when creating the
queries.  Check through eupath_webservices.r for some amusing examples of how
I have gotten around the idiosyncrasies in the eupathdb.
}
\author{
Keith Hughitt with significant modifications by atb.
}
