% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_orgdb}
\alias{make_eupath_orgdb}
\title{Generate OrgDb for EuPathDB organism}
\usage{
make_eupath_orgdb(species = NULL, entry = NULL, dir = "eupathdb",
  kegg_abbreviation = NULL, reinstall = FALSE, metadata = NULL, ...)
}
\arguments{
\item{species}{Single species to work on.}

\item{entry}{One dimensional dataframe with organism metadata}

\item{dir}{Where to put the data}

\item{kegg_abbreviation}{A kegg abbreviation when it cannot be automatically determined.}

\item{reinstall}{Overwrite an existing installed package?}

\item{metadata}{Large metadata dataframe.}

\item{...}{Extra arguments when downloading metadata.}
}
\value{
OrgDb instance package name.
}
\description{
This is probably the most difficult function in this family, as orgdb instances are rather
more complex than the other annotation data (at least for the eupathdb).  It gathers the
KEGG/GO/gene/etc data and bundles it into a single sqlite instance and does some
rudimentary sanity checks on the result.
}
\details{
FIXME: The eupathdb family is moving away from the json-based queries.  The
get_something() family of functions upon which this process depends is
therefore not likely to continue to work over time.  It needs therefore to be
migrated to queries similar to what is performed in get_eupath_text().
}
\author{
Keith Hughitt
}
