% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_orgdb}
\alias{make_eupath_orgdb}
\title{Generate OrgDb for EuPathDB organism}
\usage{
make_eupath_orgdb(entry, output_dir, kegg_abbreviation = NULL)
}
\arguments{
\item{entry}{One dimensional dataframe with organism metadata}
}
\value{
OrgDb instance
}
\description{
Generate OrgDb for EuPathDB organism
}
\author{
Keith Hughitt
}
