% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_orgdb}
\alias{make_eupath_orgdb}
\title{Create an orgdb SQLite database from the tables in eupathdb.}
\usage{
make_eupath_orgdb(species = NULL, entry = NULL, dir = "eupathdb",
  kegg_abbreviation = NULL, reinstall = FALSE, metadata = NULL, ...)
}
\arguments{
\item{species}{A specific species ID to query}

\item{entry}{If not provided, then species will get this, it contains all the information.}

\item{dir}{Where to put all the various temporary files.}

\item{kegg_abbreviation}{If known, provide the kegg abbreviation.}

\item{reinstall}{Re-install an already existing orgdb?}

\item{metadata}{Use an existing metadata table to get the entry?}

\item{...}{Extra parameters when searching for metadata}
}
\value{
Currently only the name of the installed package.  This should probably change.
}
\description{
This now uses the new POST version of the eupathdb.  Theoretically it is better,
I am not yet convinced, but the QUERY version of the eupathdb apparently will not
be supported over time.
}
