% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{make_eupath_txdb}
\alias{make_eupath_txdb}
\title{Generate TxDb for EuPathDB organism}
\usage{
make_eupath_txdb(entry, dir = ".")
}
\arguments{
\item{entry}{One dimensional dataframe with organism metadata.}

\item{dir}{Base directory for building the package.}
}
\value{
TxDb instance
}
\description{
Generate TxDb for EuPathDB organism
}
\author{
atb
}
