% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{make_exampledata}
\alias{make_exampledata}
\title{A small hack of limma's exampleData()
function to allow for arbitrary data set sizes.}
\usage{
make_exampledata(ngenes = 1000, columns = 5)
}
\arguments{
\item{ngenes}{how many genes in the fictional data set.}

\item{columns}{how many samples in this data set.}
}
\value{
a matrix of pretend counts
}
\description{
A small hack of limma's exampleData()
function to allow for arbitrary data set sizes.
}
\examples{
\dontrun{
 pretend = make_exampledata()
}
}
\seealso{
\pkg{limma}
}

