% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{make_exampledata}
\alias{make_exampledata}
\title{make_exampledata():  A small hack of limma's exampleData()
function to allow for arbitrary data set sizes.}
\usage{
make_exampledata(ngenes = 1000, columns = 5)
}
\arguments{
\item{ngenes}{how many genes in the fictional data set.  1000 by default.}

\item{columns}{how many samples in this data set.  5 by default.}
}
\value{
a matrix of pretend counts
}
\description{
make_exampledata():  A small hack of limma's exampleData()
function to allow for arbitrary data set sizes.
}
\seealso{
\code{\link{makeExampleData}}
}

