% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_expt.r
\name{make_exampledata}
\alias{make_exampledata}
\title{Small hack of limma's exampleData() to allow for arbitrary data set
sizes.}
\usage{
make_exampledata(ngenes = 1000, columns = 5)
}
\arguments{
\item{ngenes}{How many genes in the fictional data set?}

\item{columns}{How many samples in this data set?}
}
\value{
Matrix of pretend counts.
}
\description{
exampleData has a set number of genes/samples it creates. This
relaxes that restriction.
}
\examples{
\dontrun{
 pretend = make_exampledata()
}
}
\seealso{
\pkg{limma} \pkg{stats} \pkg{DESeq}
}
