% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets.R
\name{make_gsc_from_ids}
\alias{make_gsc_from_ids}
\title{Create a gene set collection from a set of arbitrary IDs.}
\usage{
make_gsc_from_ids(
  first_ids,
  second_ids = NULL,
  annotation_name = "org.Hs.eg.db",
  researcher_name = "elsayed",
  study_name = "macrophage",
  category_name = "infection",
  phenotype_name = NULL,
  identifier_type = "entrez",
  organism = NULL,
  pair_names = "up",
  current_id = "ENSEMBL",
  required_id = "ENTREZID"
)
}
\arguments{
\item{first_ids}{The required IDs for a single set.}

\item{second_ids}{Potentially null optionally used for a second, presumably
contrasting set.}

\item{annotation_name}{Orgdb annotation, used to translate IDs to the required type.}

\item{researcher_name}{Prefix of the name for the generated set(s).}

\item{study_name}{Second element in the name of the generated set(s).}

\item{category_name}{Third element in the name of the generated set(s).}

\item{phenotype_name}{Optional phenotype data for the generated set(s).}

\item{pair_names}{The suffix of the generated set(s).}

\item{current_id}{What type of ID is the data currently using?}

\item{required_id}{What type of ID should the use?}
}
\value{
Small list comprised of the created gene set collection(s).
}
\description{
This function attempts to simplify the creation of a gsva compatible
GeneSet.  Some important caveats when working with gsva, notably the gene IDs
we use are not usually compatible with the gene IDs used by gsva, thus the
primary logic in this function is intended to bridge these IDs.
}
\seealso{
[GSEABase]
}
