% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_gsva.r
\name{make_gsc_from_pairwise}
\alias{make_gsc_from_pairwise}
\title{Given a pairwise result, make a gene set collection.}
\usage{
make_gsc_from_pairwise(
  pairwise,
  according_to = "deseq",
  orgdb = "org.Hs.eg.db",
  pair_names = c("ups", "downs"),
  category_name = "infection",
  phenotype_name = "parasite",
  set_name = "elsayed_macrophage",
  color = TRUE,
  current_id = "ENSEMBL",
  required_id = "ENTREZID",
  ...
)
}
\arguments{
\item{pairwise}{A pairwise result, or combined de result, or extracted genes.}

\item{according_to}{When getting significant genes, use this method.}

\item{orgdb}{Annotation dataset.}

\item{pair_names}{Describe the contrasts of the GSC: up vs. down, high vs. low, etc.}

\item{category_name}{What category does the GSC describe?}

\item{phenotype_name}{When making color sets, use this phenotype name.}

\item{set_name}{A name for the created gene set.}

\item{color}{Make a colorSet?}

\item{current_id}{Usually we use ensembl IDs, but that does not _need_ to be the case.}

\item{required_id}{gsva uses entrezids by default.}

\item{...}{Extra arguments for extract_significant_genes().}
}
\value{
List containing 3 GSCs, one containing both the ups/downs called
  'colored', one of the ups, and one of the downs.
}
\description{
If I want to play with gsva and friends, then I need GeneSetCollections!  To
that end, this function uses extract_significant_genes() in order to gather
sets of genes deemed 'significant'.  It then passes these sets to
make_gsc_from_ids().
}
