% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{make_id2gomap}
\alias{make_id2gomap}
\title{Make a go mapping from IDs in a format suitable for topGO.}
\usage{
make_id2gomap(
  goid_map = "reference/go/id2go.map",
  go_db = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{goid_map}{TopGO mapping file.}

\item{go_db}{If there is no goid_map, create it with this data frame.}

\item{overwrite}{Rewrite the mapping file?}
}
\value{
Summary of the new goid table.
}
\description{
When using a non-supported organism, one must write out mappings in the
format expected by topgo.  This handles that process and gives a summary of
the new table.
}
\seealso{
[topGO]
}
