% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_kegg.r
\name{make_kegg_df}
\alias{make_kegg_df}
\title{Use pathfindR to get a dataframe of KEGG IDs.}
\usage{
make_kegg_df(org_code)
}
\arguments{
\item{org_code}{Organism code from KEGG.}
}
\value{
Dataframe of gene IDs to KEGG IDs.
}
\description{
The various KEGG conversion methods from KEGGREST appear to only work
for a small subset of species now.  This uses a differnt query format to get
a less flexible version of the same information. But at least it works.
}
