% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_limma.r
\name{make_limma_tables}
\alias{make_limma_tables}
\title{Writes out the results of a limma search using toptable().}
\usage{
make_limma_tables(fit = NULL, adjust = "BH", n = 0, coef = NULL,
  annot_df = NULL, intercept = FALSE)
}
\arguments{
\item{fit}{Result from lmFit()/eBayes()}

\item{adjust}{Pvalue adjustment chosen.}

\item{n}{Number of entries to report, 0 says do them all.}

\item{coef}{Which coefficients/contrasts to report, NULL says do them all.}

\item{annot_df}{Optional data frame including annotation information to
include with the tables.}

\item{intercept}{Intercept model?}
}
\value{
List of data frames comprising the toptable output for each
  coefficient, I also added a qvalue entry to these toptable() outputs.
}
\description{
However, this will do a couple of things to make one's life easier:
1.  Make a list of the output, one element for each comparison of the contrast matrix
2.  Write out the toptable() output in separate .csv files and/or sheets in excel
3.  Since I have been using qvalues a lot for other stuff, add a column for them.
}
\examples{
\dontrun{
 finished_comparison = eBayes(limma_output)
 table = make_limma_tables(finished_comparison, adjust="fdr")
}
}
\seealso{
\pkg{limma} \pkg{qvalue}
 \code{\link{write_xls}} \code{\link[limma]{topTable}}
}
