% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{make_ntile_factor}
\alias{make_ntile_factor}
\title{Make an arbitrary factor out of a numeric vector}
\usage{
make_ntile_factor(numeric_vector, n)
}
\arguments{
\item{numeric_vector}{Vector of numbers!}

\item{n}{numer of levels for the resulting factor.}
}
\value{
Factor with levels from q1 to q4.
}
\description{
Make an arbitrary factor out of a numeric vector
}
