% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{make_pairwise_contrasts}
\alias{make_pairwise_contrasts}
\title{Run makeContrasts() with all pairwise comparisons.}
\usage{
make_pairwise_contrasts(model, conditions, do_identities = FALSE,
  do_pairwise = TRUE, extra_contrasts = NULL)
}
\arguments{
\item{model}{Describe the conditions/batches/etc in the experiment.}

\item{conditions}{Factor of conditions in the experiment.}

\item{do_identities}{Include all the identity strings? Limma can
use this information while edgeR can not.}

\item{do_pairwise}{Include all pairwise strings? This shouldn't
need to be set to FALSE, but just in case.}

\item{extra_contrasts}{Optional string of extra contrasts to include.}
}
\value{
List including the following information:
 \enumerate{
  \item all_pairwise_contrasts = the result from makeContrasts(...)
  \item identities = the string identifying each condition alone
  \item all_pairwise = the string identifying each pairwise comparison alone
  \item contrast_string = the string passed to R to call makeContrasts(...)
  \item names = the names given to the identities/contrasts
 }
}
\description{
In order to have uniformly consistent pairwise contrasts, I decided
to avoid potential human erors(sic) by having a function generate
all contrasts.
}
\details{
Invoked by the _pairwise() functions.
}
\examples{
\dontrun{
 pretend <- make_pairwise_contrasts(model, conditions)
}
}
\seealso{
\pkg{limma}
 \code{\link[limma]{makeContrasts}}
}
