% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{make_pombe_expt}
\alias{make_pombe_expt}
\title{Create a Schizosaccharomyces cerevisiae expt.}
\usage{
make_pombe_expt(annotation = TRUE, host = "nov2020-fungi.ensembl.org")
}
\arguments{
\item{annotation}{Add annotation data?}

\item{host}{ensembl host to query.}
}
\value{
Expressionset/expt of fission.
}
\description{
This just saves some annoying typing if one wishes to make a standard
expressionset superclass out of the publicly available fission data set.
}
\examples{

## Check if some input data is loaded; if not, load it!
pombe_se <- get0("pombe_se")
if (is.null(pombe_se)) {
  pombe_se <- make_pombe_se()
}
summary(pombe_se)
}
\seealso{
[fission] [create_expt()]
}
