% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{make_pombe_expt}
\alias{make_pombe_expt}
\title{Create a Schizosaccharomyces cerevisiae expt.}
\usage{
make_pombe_expt(annotation = TRUE)
}
\arguments{
\item{annotation}{Add annotation data?}
}
\description{
Create a Schizosaccharomyces cerevisiae expt.
}
