% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{make_pombe_expt}
\alias{make_pombe_expt}
\title{Create a Schizosaccharomyces cerevisiae expt.}
\usage{
make_pombe_expt(annotation = TRUE)
}
\arguments{
\item{annotation}{Add annotation data?}
}
\value{
Expressionset/expt of fission.
}
\description{
This just saves some annoying typing if one wishes to make a standard
expressionset superclass out of the publicly available fission data set.
}
\seealso{
[fission] [create_expt()]
}
