% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{make_pombe_se}
\alias{make_pombe_se}
\title{Create a default summarized experiment using the publicly available fission dataset.}
\usage{
make_pombe_se(annotation = TRUE, host = "nov2020-fungi.ensembl.org")
}
\arguments{
\item{annotation}{Include annotations?}

\item{host}{ensembl host to query}
}
\description{
Create a default summarized experiment using the publicly available fission dataset.
}
