% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{make_pombe_se}
\alias{make_pombe_se}
\title{Analagous function to make_pombe_expt()}
\usage{
make_pombe_se(annotation = TRUE)
}
\arguments{
\item{annotation}{Include annotations?}
}
\description{
Analagous function to make_pombe_expt()
}
