% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{make_report}
\alias{make_report}
\title{make_report()  Make a knitr report with some defaults set}
\usage{
make_report(name = "report", type = "pdf")
}
\arguments{
\item{type}{default='pdf'  html/pdf/fancy html reports?}
}
\value{
a dated report file
}
\description{
make_report()  Make a knitr report with some defaults set
}

