% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{make_tooltips}
\alias{make_tooltips}
\title{\code{make_tooltips()}  Create a simple df from gff which contains tooltip usable
information for gVis graphs. The tooltip column is also a handy proxy for
anontations information when it would otherwise be too troublesome.}
\usage{
make_tooltips(annotations, desc_col = "description")
}
\arguments{
\item{gff}{or annotations: Either a gff file or annotation data frame (which likely came from a gff file.)}
}
\value{
a df of tooltip information or name of a gff file
}
\description{
\code{make_tooltips()}  Create a simple df from gff which contains tooltip usable
information for gVis graphs. The tooltip column is also a handy proxy for
anontations information when it would otherwise be too troublesome.
}
\examples{
\dontrun{
tooltips <- make_tooltips('reference/gff/saccharomyces_cerevisiae.gff.gz')
}
}
\seealso{
\pkg{googleVis} \link{gff2df}
}

