% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_gff.r
\name{make_tooltips}
\alias{make_tooltips}
\title{Create a simple df from a gff which contains tooltips usable in googleVis graphs.}
\usage{
make_tooltips(annotations, desc_col = "description", type = "gene",
  id_col = "ID", ...)
}
\arguments{
\item{annotations}{Either a gff file or annotation data frame (which likely came from a gff file.).}

\item{desc_col}{Gff column from which to gather data.}

\item{type}{Gff type to use as the master key.}

\item{id_col}{Which annotation column to cross reference against?}

\item{...}{Extra arguments dropped into arglist.}
}
\value{
Df of tooltip information or name of a gff file.
}
\description{
The tooltip column is a handy proxy for more thorough anontations information when it would
otherwise be too troublesome to acquire.
}
\examples{
\dontrun{
 tooltips <- make_tooltips('reference/gff/saccharomyces_cerevisiae.gff.gz')
}
}
\seealso{
\pkg{googleVis}
 \code{\link{load_gff_annotations}}
}
