% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_biomart.R
\name{make_tx_gene_map}
\alias{make_tx_gene_map}
\title{I keep messing up the creation of the salmon trancript to gene map.}
\usage{
make_tx_gene_map(
  annotations,
  gene_column = "ensembl_gene_id",
  transcript_column = "ensembl_transcript_id",
  tx_version_column = "transcript_version",
  new_column = "salmon_transcript"
)
}
\arguments{
\item{annotations}{Annoation database to merge.}

\item{gene_column}{Column containing the gene IDs.}

\item{transcript_column}{Column containing the transcript IDs.}

\item{tx_version_column}{Salmon uses tx version numbers, find them here.}

\item{new_column}{Add the new combined IDs here.}
}
\description{
Maybe this will help.  I have a smarter but much slower method in
the tmrc3 data which first creates an expressionset without
annotations then cross references the rownames against combinations
of columns in the annotations to figure out the correct pairing.
This helps when I have a combined transcriptome and get confused.
}
\details{
This probably doesn't belong in this file.
}
\examples{
## Provide examples of the biomart annotation functions.

## Search for a biomart instance which will accept queries.
archive_mart <- try(find_working_mart(archive = TRUE, year = 2023, month = 02))
class(archive_mart)
archive_mart

## I should probably change this function to accept the mart from
## find_working_mart()
hs_annot <- try(load_biomart_annotations(year = 2023, month = 02))
class(hs_annot)

## I keep forgetting that load_biomart_annotations() returns a list
## so I made a method to coerce it for me.
dim(as.data.frame(hs_annot))

## I also added a S3 print method to try to avoid being made sad if I
## print the result from load_annotations.
hs_annot

## If one wishes to check out all of the possible data types for a
## single gene, try this:
## However, this function is painfully slow and likely annoying to the
## ensembl servers.
## single_gene <- get_biomart_example_gene()

## Grab gene ontology data for a species:
hs_go <- try(load_biomart_go(year = 2023, month = 02))
hs_go

## The following example is fragile I think because multiple
## ensembl instances must be responding properly to work
## If it fails; check that the examples for biomaRt::getLDS are
## functioning -- most likely they are failing due to an upstream
## error.
hs_mm_ortho <- try(load_biomart_orthologs(gene_ids = "ENSMUSG00000000001",
                                          year = 2023, month = 02,
                                          first_species = "mmusculus",
                                          second_species = "scerevisiae"))

## Create salmon-like ID.version strings using ensembl annotations.
tx_gene_map <- try(make_tx_gene_map(as.data.frame(hs_annot)))
head(tx_gene_map)
}
