% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{makeSVD}
\alias{makeSVD}
\title{make_SVD() is a function scabbed from Hector and Kwame's cbcbSEQ
It just does fast.svd of a matrix against its rowMeans().}
\usage{
makeSVD(x)
}
\arguments{
\item{data}{A data frame to decompose}
}
\value{
a list containing the s,v,u from fast.svd
}
\description{
make_SVD() is a function scabbed from Hector and Kwame's cbcbSEQ
It just does fast.svd of a matrix against its rowMeans().
}
\examples{
## svd = makeSVD(data)
}
\seealso{
\code{\link{fast.svd}}
}

