% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{makeSVD}
\alias{makeSVD}
\title{\code{make_SVD()} is a function scabbed from Hector and Kwame's cbcbSEQ
It just does fast.svd of a matrix against its rowMeans().}
\usage{
makeSVD(data)
}
\arguments{
\item{data}{A data frame to decompose}
}
\value{
a list containing the s,v,u from fast.svd
}
\description{
\code{make_SVD()} is a function scabbed from Hector and Kwame's cbcbSEQ
It just does fast.svd of a matrix against its rowMeans().
}
\examples{
\dontrun{
 svd = makeSVD(data)
}
}
\seealso{
\pkg{corpcor} \code{\link[corpcor]{fast.svd}}
}

