% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{map_orgdb_ids}
\alias{map_orgdb_ids}
\title{Map AnnotationDbi keys from one column to another.}
\usage{
map_orgdb_ids(orgdb, gene_ids = NULL, mapto = "ensembl", keytype = "geneid")
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Gene identifiers for retrieving annotations.}

\item{mapto}{Key to map the IDs against.}

\item{keytype}{Choose a keytype, this will yell if it doesn't like your choice.}
}
\value{
a table of gene information
}
\description{
Given a couple of keytypes, this provides a quick mapping across them.  I
might have an alternate version of this hiding in the gsva code, which
requires ENTREZIDs.  In the mean time, this creates a dataframe of the mapped
columns for a given set of gene ids using the in a sqlite instance.
}
\examples{
 dm_unigene_to_ensembl <- map_orgdb_ids("org.Dm.eg.db", mapto = "ensembl", keytype = "unigene")
 head(dm_unigene_to_ensembl)
}
\seealso{
[AnnotationDbi]
}
\author{
Keith Hughitt with changes by atb.
}
