% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_microbesonline.r
\name{mdesc_table}
\alias{mdesc_table}
\title{Get the description of a microbesonline genomics table}
\usage{
mdesc_table(table = "Locus2Go")
}
\arguments{
\item{table}{Choose a table to query.}
}
\value{
Data frame describing the relevant table
}
\description{
This at least in theory is only used by get_microbesonline,  but if one needs a quick and dirty SQL query
it might prove useful.
}
\examples{
\dontrun{
 description <- mdesc_table(table="Locus2Go")
}
}
\seealso{
\pkg{DBI}
 \code{\link[DBI]{dbSendQuery}} \code{\link[DBI]{fetch}}
}
