% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tables.r
\name{median_by_factor}
\alias{median_by_factor}
\title{Create a data frame of the medians of rows by a given factor in the data.}
\usage{
median_by_factor(data, fact = "condition")
}
\arguments{
\item{data}{Data frame, presumably of counts.}

\item{fact}{Factor describing the columns in the data.}
}
\value{
Data frame of the medians.
}
\description{
This assumes of course that (like expressionsets) there are separate columns for each replicate
of the conditions.  This will just iterate through the levels of a factor describing the columns,
extract them, calculate the median, and add that as a new column in a separate data frame.
}
\details{
Used in write_expt() as well as a few random collaborations.
}
\examples{
\dontrun{
 compressed = hpgltools:::median_by_factor(data, experiment$condition)
}
}

