% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hpgltools.r
\name{mesg}
\alias{mesg}
\title{message() but with a verbose flag.}
\usage{
mesg(..., verbosity = NULL)
}
\arguments{
\item{...}{parameters for message()}

\item{verbosity}{actually print the message?}
}
\description{
message() but with a verbose flag.
}
