% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{my_identifyAUBlocks}
\alias{my_identifyAUBlocks}
\title{copy/paste the function from SeqTools and figure out where it falls on its ass.}
\usage{
my_identifyAUBlocks(x, min.length = 20, p.to.start = 0.8, p.to.end = 0.55)
}
\arguments{
\item{x}{Sequence object}

\item{min.length}{I dunno.}

\item{p.to.start}{P to start of course}

\item{p.to.end}{The p to end -- wtf who makes names like this?}
}
\value{
a list of IRanges which contain a bunch of As and Us.
}
\description{
Yeah, I do not remember what I changed in this function.
}
