% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/180_power_estimation.r
\name{my_runsims}
\alias{my_runsims}
\title{A version of PROPER:::runsims which is (hopefully) a little more robust.}
\usage{
my_runsims(
  Nreps = c(3, 5, 7, 10),
  Nreps2,
  nsims = 100,
  sim.opts,
  DEmethod = c("edgeR", "DSS", "DESeq", "DESeq2"),
  verbose = TRUE
)
}
\arguments{
\item{Nreps}{Vector of numbers of replicates to simulate.}

\item{Nreps2}{Second vector of replicates.}

\item{nsims}{How many simulations to perform?}

\item{sim.opts}{Options provided in a list which include information about the expression,
numbers of genes, logFC values, etc.}

\item{DEmethod}{I suggest using only either edgeR or DESeq2.}

\item{verbose}{Print some information along the way?}
}
\description{
When I was testing PROPER, it fell down mysteriously on a few occasions.  The
source ended up being in runsims(), ergo this function.
}
