% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/109_ontology_kegg.r
\name{myretrieveKGML}
\alias{myretrieveKGML}
\title{A couple functions from KEGGgraph that have broken}
\usage{
myretrieveKGML(
  pathway,
  organism,
  destfile,
  silent = TRUE,
  hostname = "http://www.kegg.jp",
  ...
)
}
\arguments{
\item{pathway}{The path to query.}

\item{organism}{Which organism to query?}

\item{destfile}{File to which to download.}

\item{silent}{Send stdout and stderr to dev null?}

\item{hostname}{Host to download from (this is what is broken.)}

\item{...}{Arglist!}
}
\description{
Some material in KEGGREST is borken.
}
