% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_kegg.r
\name{myretrieveKGML}
\alias{myretrieveKGML}
\title{A couple functions from KEGGgraph that have broken}
\usage{
myretrieveKGML(pathwayid, organism, destfile, method = "wget",
  hostname = "http://www.kegg.jp", ...)
}
\arguments{
\item{pathwayid}{The path to query.}

\item{organism}{Which organism to query?}

\item{destfile}{File to which to download.}

\item{method}{Which download method to use?}

\item{hostname}{Host to download from (this is what is broken.)}

\item{...}{Arglist!}
}
\description{
Some material in KEGGREST is borken.
}
