% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_noiseq.R
\name{noiseq_pairwise}
\alias{noiseq_pairwise}
\title{Perform pairwise comparisons using noiseq.}
\usage{
noiseq_pairwise(
  input = NULL,
  conditions = NULL,
  batches = NULL,
  model_cond = TRUE,
  model_batch = TRUE,
  annot_df = NULL,
  ...
)
}
\arguments{
\item{input}{Expressionset to compare.}

\item{conditions}{Set of conditions to query}

\item{batches}{known batches in the data, or a surrogate estimator.}

\item{model_cond}{Add condition to the model?}

\item{model_batch}{Add batch to the model, noiseq has its own combat-like method,
so maybe not necessary?}

\item{annot_df}{Extra annotations.}

\item{...}{Extra arguments.}
}
\value{
List similar to deseq_pairwise/edger_pairwise/etc.
}
\description{
Perform pairwise comparisons using noiseq.
}
