% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{normalize,expt-method}
\alias{normalize,expt-method}
\title{Normalization of an expt}
\usage{
\S4method{normalize}{expt}(object, ...)
}
\arguments{
\item{object}{Data structure to normalize}

\item{...}{Other options.}
}
\description{
The primary definition for normalize resides in BiocGenerics.
}
