% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{normalize}
\alias{normalize}
\title{Every time I think I have a handle on R dispatch I get some BS like this.}
\usage{
normalize(object, ...)
}
\arguments{
\item{object}{Input to normalize}

\item{...}{other arguments passed along to BiocGenerics::normalize().}
}
\description{
Every time I think I have a handle on R dispatch I get some BS like this.
}
