% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{normalize}
\alias{normalize}
\title{I think this is already defined in BiocGenerics}
\usage{
normalize(object, ...)
}
\arguments{
\item{object}{thing to nromalize}

\item{...}{args passed along.}
}
\description{
But for reasons passing all understanding, it is giving me grief.
}
