% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shared.R
\name{normalize}
\alias{normalize}
\title{Simplified and ideally improved normalization function}
\usage{
normalize(expt, todo = list())
}
\arguments{
\item{expt}{Input data}

\item{todo}{List of tasks to perform.}
}
\description{
This function is ideally should provide a simpler and more capable
version of normalize_expt. I also want to move everything to using
summarizedExperiments and this simpler method provides an
opportunity.
}
