% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_norm.r
\name{normalize_counts}
\alias{normalize_counts}
\title{Perform a simple normalization of a count table}
\usage{
normalize_counts(data, design = NULL, norm = "raw")
}
\arguments{
\item{data}{A matrix of count data}

\item{design}{A dataframe describing the experimental design
(conditions/batches/etc)}

\item{norm}{A normalization to perform:
'sf|quant|qsmooth|tmm|upperquartile|tmm|rle'
I keep wishy-washing on whether design is a required argument.}
}
\value{
dataframe of normalized(counts)
}
\description{
Perform a simple normalization of a count table
}
\examples{
\dontrun{
norm_table = normalize_counts(count_table, design=design, norm='qsmooth')
}
}
\seealso{
\pkg{edgeR} \pkg{limma} \pkg{DESeq2}
}

