% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_counts}
\alias{normalize_counts}
\title{normalize_counts(): Perform a simple normalization of a count table}
\usage{
normalize_counts(data, design = NULL, norm = "raw")
}
\arguments{
\item{design}{A dataframe describing the experimental design
(conditions/batches/etc)}

\item{count_table}{A matrix of count data}

\item{norm='raw'}{A normalization to perform:
'sf|quant|qsmooth|tmm|upperquartile|tmm|rle'
I keep wishy-washing on whether design is a required argument.}
}
\value{
dataframe of normalized(counts)
}
\description{
normalize_counts(): Perform a simple normalization of a count table
}
\examples{
## norm_table = normalize_counts(count_table, design=design, norm='qsmooth')
}

