% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{normalize_expt}
\alias{normalize_expt}
\title{Replace the data of an expt with normalized data.}
\usage{
normalize_expt(expt, transform = "raw", norm = "raw", convert = "raw",
  batch = "raw", filter_low = FALSE, annotations = NULL, fasta = NULL,
  entry_type = "gene", use_original = FALSE, batch1 = "batch",
  batch2 = NULL, thresh = 2, min_samples = 2, p = 0.01, A = 1,
  k = 1, cv_min = 0.01, cv_max = 1000, ...)
}
\arguments{
\item{expt}{The original expt}

\item{transform}{The transformation desired (raw, log2, log, log10)}

\item{norm}{How to normalize the data (raw, quant, sf, upperquartile, tmm, rle)}

\item{convert}{Conversion to perform (raw, cpm, rpkm, cp_seq_m)}

\item{batch}{Batch effect removal tool to use (limma sva fsva ruv etc)}

\item{filter_low}{Filter out low sequences (cbcb, pofa, kofa, others?)}

\item{annotations}{used for rpkm, a df}

\item{fasta}{fasta file for cp_seq_m counting of oligos}

\item{entry_type}{for getting genelengths by feature type (rpkm or cp_seq_m)}

\item{use_original}{whether to use the backup data in the expt class}

\item{batch1}{experimental factor to extract first}

\item{batch2}{a second factor to remove (only with limma's removebatcheffect())}

\item{thresh}{for cbcb_lowfilter}

\item{min_samples}{for cbcb_lowfilter}

\item{p}{for genefilter's pofa}

\item{A}{for genefilter's pofa}

\item{k}{for genefilter's kofa}

\item{cv_min}{for genefilter's cv()}

\item{cv_max}{for genefilter's cv()}

\item{...}{more options}
}
\value{
a new expt object with normalized data and the original data saved as 'original_expressionset'
}
\description{
Replace the data of an expt with normalized data.
}
\examples{
\dontrun{
normed <- normalize_expt(exp, transform='log2', norm='rle', convert='cpm',
                         batch='raw', filter_low='pofa')
normed_batch <- normalize_expt(exp, transform='log2', norm='rle', convert='cpm',
                               batch='sva', filter_low='pofa')
}
}
\seealso{
\pkg{genefilter} \pkg{cbcbSEQ} \pkg{limma} \pkg{sva} \pkg{edgeR} \pkg{DESeq2}
}

