% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_expt}
\alias{normalize_expt}
\title{Replace the data of an expt with normalized data}
\usage{
normalize_expt(expt, transform = "raw", norm = "raw", convert = "raw",
  batch = "raw", filter_low = FALSE, annotations = NULL, fasta = NULL,
  entry_type = "gene", verbose = FALSE, use_original = FALSE,
  batch1 = "batch", batch2 = NULL, thresh = 2, min_samples = 2,
  p = 0.01, A = 1, k = 1, cv_min = 0.01, cv_max = 1000, ...)
}
\arguments{
\item{expt=expt}{The original expt}

\item{transform="raw"}{The transformation desired (raw, log2, log, log10)}

\item{norm="raw"}{How to normalize the data (raw, quant, sf, upperquartile, tmm, rle)}

\item{convert="raw"}{Conversion to perform (raw, cpm, rpkm, cp_seq_m)}

\item{filter_low=FALSE}{Filter out low sequences (cbcb, pofa, kofa, others?)}

\item{annotations=NULL}{used for rpkm, a df}
}
\value{
a new expt object with normalized data and the original data saved as 'original_expressionset'
}
\description{
Replace the data of an expt with normalized data
}

