% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_expt}
\alias{normalize_expt}
\title{Replace the data of an expt with normalized data}
\usage{
normalize_expt(expt, transform = "raw", norm = "raw", convert = "raw",
  batch = "raw", filter_low = FALSE, annotations = NULL,
  verbose = FALSE, use_original = FALSE, batch1 = "batch",
  batch2 = NULL, thresh = 2, min_samples = 2, p = 0.01, A = 1,
  k = 1, cv_min = 0.01, cv_max = 1000, ...)
}
\arguments{
\item{expt=expt}{The original expt}

\item{transform="log2"}{The transformation desired}
}
\value{
a new expt object with normalized data and the original data saved as 'original_expressionset'
}
\description{
Replace the data of an expt with normalized data
}

