% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_expt}
\alias{normalize_expt}
\title{Replace the data of an expt with normalized data}
\usage{
normalize_expt(expt, transform = "log2", norm = "quant", convert = "cpm",
  filter_low = TRUE, annotations = NULL, verbose = FALSE,
  use_original = TRUE, thresh = 2, min_samples = 2, batch = NULL,
  batch1 = "batch", batch2 = NULL, ...)
}
\arguments{
\item{expt=expt}{The original expt}

\item{transform="log2"}{The transformation desired}
}
\value{
a new expt object with normalized data and the original data saved as 'original_expressionset'
}
\description{
Replace the data of an expt with normalized data
}

