% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_orgdb.r
\name{orgdb_idmap}
\alias{orgdb_idmap}
\title{Load organism annotation data (mouse/human).}
\usage{
orgdb_idmap(orgdb, gene_ids = NULL, mapto = c("ensembl"),
  keytype = "geneid")
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Gene identifiers for retrieving annotations.}

\item{mapto}{Key to map the IDs against.}

\item{keytype}{Choose a keytype, this will yell if it doesn't like your choice.}
}
\value{
a table of gene information
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the OrganismDbi interface.
}
\examples{
\dontrun{
 host <- load_host_annotations(org, c("a","b"))
}
}
\seealso{
\pkg{AnnotationDbi}
 \code{\link[AnnotationDbi]{select}} \code{\link[AnnotationDbi]{keytypes}}
}

