% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tritrypdb.r
\name{parse_gene_info_table}
\alias{parse_gene_info_table}
\title{TriTrypDB gene information table parser}
\usage{
parse_gene_info_table(file, verbose = FALSE)
}
\arguments{
\item{verbose}{Whether or not to enable verbose output.}

\item{filepath}{Location of TriTrypDB gene information table.}
}
\value{
Returns a dataframe of gene info.
}
\description{
An example input file is the T. brucei Lister427 gene information table
available at:
http://tritrypdb.org/common/downloads/Current_Release/TbruceiLister427/txt/data/TriTrypDB-5.0_TbruceiLister427Gene.txt
}
\author{
Keith Hughitt
}

