% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pca.r
\name{pca_highscores}
\alias{pca_highscores}
\title{Get the highest/lowest scoring genes for every principle component.}
\usage{
pca_highscores(expt, n = 20, cor = TRUE, vs = "means", logged = TRUE)
}
\arguments{
\item{n}{the number of genes to extract.}

\item{df}{a dataframe of (pseudo)counts}

\item{conditions}{a factor or character of conditions in the experiment.}

\item{batches}{a factor or character of batches in the experiment.}
}
\value{
a list including the princomp biplot, histogram, and tables
 of top/bottom n scored genes with their scores by component.
}
\description{
This function uses princomp to acquire a principle component biplot
for some data and extracts a dataframe of the top n genes for each
component by score.
}
\examples{
\dontrun{
 information <- pca_highscores(df=df, conditions=cond, batches=bat)
 information$pca_bitplot  ## oo pretty
}
}
\seealso{
\pkg{stats}
 \code{\link[stats]{princomp}}
}
