% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.r
\name{pca_highscores}
\alias{pca_highscores}
\title{Get the highest/lowest scoring genes for every principle component.}
\usage{
pca_highscores(expt, n = 20, cor = TRUE, vs = "means", logged = TRUE)
}
\arguments{
\item{expt}{Experiment to poke.}

\item{n}{Number of genes to extract.}

\item{cor}{Perform correlations?}

\item{vs}{Do a mean or median when getting ready to perform the pca?}

\item{logged}{Check for the log state of the data and adjust as deemed necessary?}
}
\value{
a list including the princomp biplot, histogram, and tables
 of top/bottom n scored genes with their scores by component.
}
\description{
This function uses princomp to acquire a principle component biplot
for some data and extracts a dataframe of the top n genes for each
component by score.
}
\examples{
\dontrun{
 information <- pca_highscores(df=df, conditions=cond, batches=bat)
 information$pca_bitplot  ## oo pretty
}
}
\seealso{
\pkg{stats}
 \code{\link[stats]{princomp}}
}
