% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{pca_information}
\alias{pca_information}
\title{Calculate some information useful for generating PCA plots}
\usage{
pca_information(expt = NULL, df = NULL, design = NULL,
  factors = c("condition", "batch"), num_components = NULL,
  plot_pcas = FALSE, labels = "fancy")
}
\arguments{
\item{df}{The data to analyze (usually exprs(somedataset))}

\item{design}{A dataframe describing the experimental design, containing columns with
useful information like the conditions, batches, number of cells, whatever...}

\item{factors}{A character list of columns from the design matrix which will be queried
for R^2 against the fast.svd of the data.  This defaults to c("condition","batch"), which
is in all of my experimental designs thus far.}

\item{components}{A number of principle components to compare the design factors against.
  If left null, it will query the same number of components as factors asked for.}
}
\value{
a list of fun pca information:
  svd_u/d/v: The u/d/v parameters from fast.svd
  rsquared_table: A table of the rsquared values between each factor and principle component
  pca_variance: A table of the pca variances
  pca_data: Coordinates for a pca plot
  pca_cor: A table of the correlations between the factors and principle components
  anova_fstats: the sum of the residuals with the factor vs without (manually calculated)
  anova_f: The result from performing anova(withfactor, withoutfactor), the F slot
  anova_p: The p-value calculated from the anova() call
  anova_sums: The RSS value from the above anova() call
  cor_heatmap: A heatmap from recordPlot() describing pca_cor.
}
\description{
Calculate some information useful for generating PCA plots
}
\examples{
## pca_info = pca_information(exprs(some_expt$expressionset), some_design, "all")
## pca_info
}
\seealso{
\code{\link{fast.svd}}, \code{\link{lm}}
}

