% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{pData<-,expt,ANY-method}
\alias{pData<-,expt,ANY-method}
\title{A setter to put the experimental metadata into an expt.}
\usage{
\S4method{pData}{expt,ANY}(object) <- value
}
\arguments{
\item{object}{An expt.}

\item{value}{New metadata.}
}
\description{
A setter to put the experimental metadata into an expt.
}
