% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/096_plot_point.r
\name{plot_bcv}
\alias{plot_bcv}
\title{Steal edgeR's plotBCV() and make it a ggplot2.}
\usage{
plot_bcv(data)
}
\arguments{
\item{data}{A dataframe/expt/exprs with count data}
}
\value{
a plot! of the BCV a la ggplot2.
}
\description{
This was written primarily to understand what that function is doing in edgeR.
}
\examples{
\dontrun{
 bcv <- plot_bcv(expt)
 summary(bcv$data)
 bcv$plot
}
}
\seealso{
\pkg{edgeR}
 \code{\link[edgeR]{plotBCV}}
}
