% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_boxplot,data.frame-method}
\alias{plot_boxplot,data.frame-method}
\title{Make a ggplot boxplot of a set of samples.}
\usage{
\S4method{plot_boxplot}{data.frame}(
  data,
  colors = NULL,
  plot_title = NULL,
  order = NULL,
  violin = FALSE,
  scale = NULL,
  sample_names = NULL,
  label_chars = 10,
  ...
)
}
\arguments{
\item{data}{Expt or data frame set of samples.}

\item{colors}{Color scheme, if not provided will make its own.}

\item{plot_title}{A title!}

\item{order}{Set the order of boxen.}

\item{violin}{Print this as a violin rather than a just box/whiskers?}

\item{scale}{Whether to log scale the y-axis.}

\item{sample_names}{Another version of the sample names for printing.}

\item{label_chars}{Maximum number of characters for abbreviating sample names.}

\item{...}{More parameters are more fun!}
}
\value{
Ggplot2 boxplot of the samples.  Each boxplot
contains the following information: a centered line describing the
median value of counts of all genes in the sample, a box around the
line describing the inner-quartiles around the median (quartiles 2
and 3 for those who are counting), a vertical line above/below the
box which shows 1.5x the inner quartile range (a common metric of
the non-outliers), and single dots for each gene which is outside
that range.  A single dot is transparent.
}
\description{
Boxplots and density plots provide complementary views of data distributions.
The general idea is that if the box for one sample is significantly shifted
from the others, then it is likely an outlier in the same way a density plot
shifted is an outlier.
}
\examples{
if (is.null(pombe_se)) {
  source(system.file("examples", "create_se.R", package = "hpgltools"))
}

## Most distribution plots are particularly informative with the raw data.

## Make a boxplot of the SE
a_boxplot <- plot_boxplot(pombe_se)
a_boxplot

## The analagous density plot of the samples
a_density <- plot_density(pombe_se)
a_density

pombe_subset <- pombe_se[, c(1, 2, 3, 4)]
## QQ plots comparing samples to each other
qq <- plot_qq_all(pombe_subset)
qq[["logs"]]

## Show the percentage of all reads comprised by the top-n genes.
## This might be smarter if moved to the log scale?
topn_genes <- plot_topn(pombe_subset)
topn_genes

## Create a violin of the coefficient of variance on a per-condition basis.
varcoef <- plot_variance_coefficients(pombe_se)
varcoef
}
\seealso{
[ggplot2]
}
