% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_density}
\alias{plot_density}
\title{Create a density plot, showing the distribution of each column of data.}
\usage{
plot_density(
  data,
  colors = NULL,
  colors_by = "condition",
  design = NULL,
  direct = NULL,
  fill = NULL,
  label_chars = 10,
  plot_title = NULL,
  position = "identity",
  sample_names = NULL,
  scale = NULL,
  ...
)
}
\arguments{
\item{data}{Expt, expressionset, or data frame.}

\item{colors}{Color scheme to use.}

\item{colors_by}{Factor for coloring the lines.}

\item{design}{Experimental design.}

\item{direct}{Use direct.labels for labeling the plot?}

\item{fill}{Fill the distributions?  This might make the plot unreasonably colorful.}

\item{label_chars}{Maximum number of characters in sample names before abbreviation.}

\item{plot_title}{Title for the plot.}

\item{position}{How to place the lines, either let them overlap (identity), or stack them.}

\item{sample_names}{Names of the samples.}

\item{scale}{Plot on the log scale?}

\item{...}{sometimes extra arguments might come from graph_metrics()}
}
\value{
ggplot2 density plot!
}
\description{
Density plots and boxplots are cousins and provide very similar views of data
distributions. Some people like one, some the other.  I think they are both
colorful and fun!
}
\examples{
\dontrun{
 funkytown <- plot_density(data)
}
}
\seealso{
[ggplot2]
}
