% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{plot_density}
\alias{plot_density}
\title{Create a density plot, showing the distribution of each column of data.}
\usage{
plot_density(data, colors = NULL, sample_names = NULL,
  position = "identity", fill = NULL, title = NULL, scale = NULL,
  colors_by = "condition")
}
\arguments{
\item{data}{Expt, expressionset, or data frame.}

\item{colors}{Color scheme to use.}

\item{sample_names}{Names of the samples.}

\item{position}{How to place the lines, either let them overlap (identity), or stack them.}

\item{fill}{Fill the distributions?  This might make the plot unreasonably colorful.}

\item{title}{Title for the plot.}

\item{scale}{Plot on the log scale?}

\item{colors_by}{Factor for coloring the lines}
}
\value{
Ggplot2 density plot!
}
\description{
Density plots and boxplots are cousins and provide very similar views of data distributions.
Some people like one, some the other.  I think they are both colorful and fun!
}
\examples{
\dontrun{
funkytown <- plot_density(data)
}
}
\seealso{
\pkg{ggplot2} \link[ggplot2]{geom_density}
}

