% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.r
\name{plot_epitrochoid}
\alias{plot_epitrochoid}
\title{Make epitrochoid plots!}
\usage{
plot_epitrochoid(radius_a = 7, radius_b = 2, dist_b = 6,
  revolutions = 7, increments = 6480)
}
\description{
7, 2, 6, 7 should give a pretty result.
}

