% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnseq.r
\name{plot_essentiality}
\alias{plot_essentiality}
\title{Plot the essentiality of a library as per DeJesus et al.}
\usage{
plot_essentiality(file)
}
\arguments{
\item{file}{a file created using the perl script 'essentiality_tas.pl'}
}
\value{
A couple of plots
}
\description{
This provides a plot of the essentiality metrics 'zbar' with respect to gene.
}
\seealso{
\pkg{ggplot2}
}
