% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.r
\name{plot_goseq_pval}
\alias{plot_goseq_pval}
\title{Make a pvalue plot from goseq data.}
\usage{
plot_goseq_pval(goterms, wrapped_width = 30, cutoff = 0.1, n = 30,
  mincat = 5, level = NULL)
}
\arguments{
\item{goterms}{Some data from goseq!}

\item{wrapped_width}{Number of characters before wrapping to help legibility.}

\item{cutoff}{Pvalue cutoff for the plot.}

\item{n}{How many groups to include?}

\item{mincat}{Minimum size of the category for inclusion.}

\item{level}{Levels of the ontology tree to use.}
}
\value{
Plots!
}
\description{
With minor changes, it is possible to push the goseq results into a clusterProfiler-ish pvalue
plot.  This handles those changes and returns the ggplot results.
}
\seealso{
\pkg{goseq} \pkg{clusterProfiler}
 \code{\link[goseq]{goseq}} \code{\link{plot_ontpval}}
}
