% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.r
\name{plot_gprofiler_pval}
\alias{plot_gprofiler_pval}
\title{Make a pvalue plot from gprofiler data.}
\usage{
plot_gprofiler_pval(gp_result, wrapped_width = 30, cutoff = 0.1, n = 30,
  group_minsize = 5, scorer = "recall", ...)
}
\arguments{
\item{gp_result}{Some data from gProfiler.}

\item{wrapped_width}{Maximum width of the text names.}

\item{cutoff}{P-value cutoff for the plots.}

\item{n}{Maximum number of ontologies to include.}

\item{group_minsize}{Minimum ontology group size to include.}

\item{scorer}{Which column to use for scoring the data.}

\item{...}{Options I might pass from other functions are dropped into arglist.}
}
\value{
List of MF/BP/CC pvalue plots.
}
\description{
The p-value plots from clusterProfiler are pretty, this sets the gprofiler data into a format
suitable for plotting in that fashion and returns the resulting plots of significant ontologies.
}
\seealso{
\pkg{topgo} \pkg{clusterProfiler}
}
