% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gvis.r
\name{plot_gvis_ma}
\alias{plot_gvis_ma}
\title{Make an html version of an MA plot: M(log ratio of conditions) / A(mean average).}
\usage{
plot_gvis_ma(
  df,
  tooltip_data = NULL,
  p = 0.05,
  logfc = 1,
  p_col = "AdjPVal",
  fc_col = "logfc",
  avg_col = "AvgExp",
  filename = "html/gvis_ma_plot.html",
  base_url = "",
  ...
)
}
\arguments{
\item{df}{Data frame of counts which have been normalized counts by
sample-type, which is to say the output from voom/voomMod/hpgl_voom().}

\item{tooltip_data}{Df of tooltip information (gene names, etc).}

\item{p}{P-value cutoff}

\item{logfc}{Logfc cutoff}

\item{p_col}{Column in the data containing the p-values.}

\item{fc_col}{Column in the data containing the fold-changes.}

\item{avg_col}{Column in the data containing the average expression values.}

\item{filename}{Filename to write a fancy html graph.}

\item{base_url}{String with a basename used for generating URLs for clicking
dots on the graph.}

\item{...}{more options are more options!}
}
\value{
NULL, but along the way an html file is generated which contains a
  googleVis MA plot. See plot_de_ma() for details.
}
\description{
A fun snippet from wikipedia: "In many microarray gene expression
experiments, an underlying assumption is that most of the genes would not see
any change in their expression therefore the majority of the points on the
y-axis (M) would be located at 0, since Log(1) is 0. If this is not the case,
then a normalization method such as LOESS should be applied to the data
before statistical analysis. If the median line is not straight, the data
should be normalized.
}
\examples{
\dontrun{
 plot_gvis_ma(df, filename="html/fun_ma_plot.html",
              base_url="http://yeastgenome.org/accession?")
}
}
\seealso{
\pkg{googleVis}
 \code{\link{plot_ma_de}}
}
