% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{plot_heatplus}
\alias{plot_heatplus}
\title{Potential replacement for heatmap.2 based plots.}
\usage{
plot_heatplus(expt, type = "correlation", method = "pearson",
  annot_columns = "batch", annot_rows = "condition", cutoff = 1,
  cluster_colors = NULL, scale = "none", cluster_width = 2,
  cluster_function = NULL, heatmap_colors = NULL)
}
\arguments{
\item{fundata}{A data frame to plot.}
}
\description{
Heatplus is an interesting tool, I have a few examples of using it and intend to include them here.
}
