% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misc.r
\name{plot_hypotrochoid}
\alias{plot_hypotrochoid}
\title{Make hypotrochoid plots!}
\usage{
plot_hypotrochoid(radius_a = 3, radius_b = 7, dist_b = 1,
  revolutions = 7, increments = 6480)
}
\arguments{
\item{radius_a}{Radius of the major circle}

\item{radius_b}{And the smaller circle.}

\item{dist_b}{between b and the drawing point.}

\item{revolutions}{How many times to revolve through the spirograph.}

\item{increments}{How many dots to lay down while writing.}
}
\description{
3,7,1 should give the classic 7 leaf clover
}
