% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics.r
\name{plot_intensity_mz}
\alias{plot_intensity_mz}
\title{Plot the peak intensities with respect to m/z}
\usage{
plot_intensity_mz(mzxml_data, loess = FALSE, alpha = 0.5, x_scale = NULL,
  y_scale = NULL, ...)
}
\arguments{
\item{mzxml_data}{The data structure from extract_mzxml or whatever it is.}

\item{loess}{Do a loess smoothing from which to extract a function
describing the data?  This is terribly slow, and in the data I have
examined so far, not very helpful, so it is FALSE by default.}

\item{...}{Extra arguments for the downstream functions.}
}
\value{
ggplot2 goodness.
}
\description{
I want to have a pretty plot of peak intensities and m/z.
}
