% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.R
\name{plot_linear_scatter}
\alias{plot_linear_scatter}
\title{Make a scatter plot between two groups with a linear model superimposed and
some supporting statistics.}
\usage{
plot_linear_scatter(
  df,
  cormethod = "pearson",
  size = 2,
  loess = FALSE,
  xcol = NULL,
  ycol = NULL,
  text_col = NULL,
  logfc = 2,
  identity = FALSE,
  z = 1.5,
  z_lines = FALSE,
  first = NULL,
  second = NULL,
  base_url = NULL,
  pretty_colors = TRUE,
  xlab = NULL,
  ylab = NULL,
  color_high = NULL,
  color_low = NULL,
  alpha = 0.4,
  ...
)
}
\arguments{
\item{df}{Dataframe likely containing two columns.}

\item{cormethod}{What type of correlation to check?}

\item{size}{Size of the dots on the plot.}

\item{loess}{Add a loess estimation?}

\item{identity}{Add the identity line?}

\item{z_lines}{Include lines defining the z-score boundaries.}

\item{first}{First column to plot.}

\item{second}{Second column to plot.}

\item{base_url}{Base url to add to the plot.}

\item{pretty_colors}{Colors!}

\item{xlab}{Alternate x-axis label.}

\item{ylab}{Alternate x-axis label.}

\item{color_high}{Chosen color for points significantly above the mean.}

\item{color_low}{Chosen color for points significantly below the mean.}

\item{alpha}{Choose an alpha channel to define how see-through the dots are.}

\item{...}{Extra args likely used for choosing significant genes.}
}
\value{
List including a ggplot2 scatter plot and some histograms.  This plot
 provides a "bird's eye" view of two data sets.  This plot assumes a
 (potential) linear correlation between the data, so it calculates the
 correlation between them.  It then calculates and plots a robust linear
 model of the data using an 'SMDM' estimator (which I don't remember how to
 describe, just that the document I was reading said it is good).  The
 median/mad of each axis is calculated and plotted as well.  The distance
 from the linear model is finally used to color the dots on the plot.
 Histograms of each axis are plotted separately and then together under a
 single cdf to allow tests of distribution similarity.  This will make a fun
 clicky googleVis graph if requested.
}
\description{
Make a scatter plot between two groups with a linear model superimposed and
some supporting statistics.
}
\examples{
\dontrun{
 plot_linear_scatter(lotsofnumbers_intwo_columns)
}
}
\seealso{
[robust] [stats] [ggplot2] [robust::lmRob] [stats::weights] [plot_histogram()]
}
