% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shared.R
\name{plot_multiplot}
\alias{plot_multiplot}
\title{Make a grid of plots.}
\usage{
plot_multiplot(plots, file, cols = NULL, layout = NULL)
}
\arguments{
\item{plots}{List of plots}

\item{file}{File to write to}

\item{cols}{Number of columns in the grid}

\item{layout}{Set the layout specifically}
}
\value{
a multiplot!
}
\description{
Make a grid of plots.
}
