% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{plot_nonzero}
\alias{plot_nonzero}
\title{Make a ggplot graph of the number of non-zero genes by sample.}
\usage{
plot_nonzero(data, design = NULL, colors = NULL, plot_labels = NULL,
  title = NULL, ...)
}
\arguments{
\item{data}{Expt, expressionset, or dataframe.}

\item{design}{Eesign matrix.}

\item{colors}{Color scheme.}

\item{plot_labels}{How do you want to label the graph? 'fancy' will use directlabels() to try to match
the labels with the positions without overlapping anything else will just stick them on a 45'
offset next to the graphed point.}

\item{title}{Add a title?}

\item{...}{rawr!}
}
\value{
a ggplot2 plot of the number of non-zero genes with respect to each library's CPM.
}
\description{
This puts the number of genes with > 0 hits on the y-axis and CPM on the x-axis. Made by Ramzi
Temanni <temanni at umd dot edu>.
}
\examples{
\dontrun{
 nonzero_plot = plot_nonzero(expt=expt)
 nonzero_plot  ## ooo pretty
}
}
\seealso{
\pkg{ggplot2}
\code{\link[ggplot2]{geom_point}} \code{\link[directlabels]{geom_dl}}
}
