% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.r
\name{plot_ontpval}
\alias{plot_ontpval}
\title{Make a pvalue plot from a df of IDs, scores, and p-values.}
\usage{
plot_ontpval(df, ontology = "MF", fontsize = 14, numerator = NULL,
  denominator = NULL)
}
\arguments{
\item{df}{Some data from topgo/goseq/clusterprofiler.}

\item{ontology}{Ontology to plot (MF,BP,CC).}

\item{fontsize}{Fiddling with the font size may make some plots more readable.}

\item{numerator}{Column used for printing a ratio of genes/category.}

\item{denominator}{Column used for printing a ratio of genes/category.}
}
\value{
Ggplot2 plot of pvalues vs. ontology.
}
\description{
This function seeks to make generating pretty pvalue plots as shown by
clusterprofiler easier.
}
\seealso{
\pkg{goseq} \pkg{ggplot2}
 \code{\link[goseq]{goseq}}
}
