% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{plot_pairwise_ma}
\alias{plot_pairwise_ma}
\title{Plot all pairwise MA plots in an experiment.}
\usage{
plot_pairwise_ma(data, log = NULL, ...)
}
\arguments{
\item{data}{Expt expressionset or data frame.}

\item{log}{Is the data in log format?}

\item{...}{Options are good and passed to arglist().}
}
\value{
List of affy::maplots
}
\description{
Use affy's ma.plot() on every pair of columns in a data set to help diagnose problematic
samples.
}
\examples{
\dontrun{
 ma_plots = plot_pairwise_ma(expt=some_expt)
}
}
\seealso{
\pkg{affy}
\code{\link[affy]{ma.plot}}
}
