% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.r
\name{plot_pcfactor}
\alias{plot_pcfactor}
\title{make a dotplot of some categorised factors and a set of principle components.}
\usage{
plot_pcfactor(pc_df, expt, exp_factor = "condition", component = "PC1")
}
\arguments{
\item{pc_df}{Df of principle components.}

\item{expt}{Expt containing counts, metadata, etc.}

\item{exp_factor}{Experimental factor to compare against.}

\item{component}{Which principal component to compare against?}
}
\value{
Plot of principle component vs factors in the data
}
\description{
This should make a quick df of the factors and PCs and plot them.
}
\examples{
\dontrun{
 estimate_vs_pcs <- plot_pcfactor(pcs, times)
}
}
\seealso{
[ggplot2]
}
