% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.r
\name{plot_pcs}
\alias{plot_pcs}
\title{A quick and dirty PCA plotter of arbitrary components against one another.}
\usage{
plot_pcs(pca_data, first = "PC1", second = "PC2", variances = NULL,
  design = NULL, plot_title = TRUE, plot_labels = NULL,
  x_label = NULL, y_label = NULL, plot_size = 5, plot_alpha = NULL,
  size_column = NULL, rug = TRUE, cis = c(0.95, 0.9), ...)
}
\arguments{
\item{pca_data}{Dataframe of principle components PC1 .. PCN with any other arbitrary information.}

\item{first}{Principle component PCx to put on the x axis.}

\item{second}{Principle component PCy to put on the y axis.}

\item{variances}{List of the percent variance explained by each component.}

\item{design}{Experimental design with condition batch factors.}

\item{plot_title}{Title for the plot.}

\item{plot_labels}{Parameter for the labels on the plot.}

\item{plot_size}{Size of the dots on the plot}

\item{size_column}{Experimental factor to use for sizing the glyphs}

\item{rug}{Include the rugs on the sides of the plot?}

\item{cis}{What (if any) confidence intervals to include.}

\item{...}{Extra arguments dropped into arglist}
}
\value{
gplot2 PCA plot
}
\description{
A quick and dirty PCA plotter of arbitrary components against one another.
}
\examples{
\dontrun{
 pca_plot = plot_pcs(pca_data, first="PC2", second="PC4", design=expt$design)
}
}
\seealso{
\pkg{ggplot2}
 \code{\link[directlabels]{geom_dl}}
}
