% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pcs}
\alias{plot_pcs}
\title{A quick and dirty PCA plotter of arbitrary components against one another.}
\usage{
plot_pcs(data, first = "PC1", second = "PC2", variances = NULL,
  design = NULL, title = NULL, labels = NULL)
}
\arguments{
\item{data}{A dataframe of principle components PC1 .. PCN with any other arbitrary information.}

\item{first}{Principle component PCx to put on the x axis}

\item{second}{Principle component PCy to put on the y axis}

\item{variances}{A list of the pct. variance explained by each component}

\item{design}{The experimental design with condition/batch}

\item{title}{Title for the plot}

\item{labels}{Whether or not one wants fancy labels for the conditions}
}
\value{
a ggplot2 PCA plot
}
\description{
A quick and dirty PCA plotter of arbitrary components against one another.
}
\examples{
## pca_plot = plot_pcs(pca_data, first="PC2", second="PC4", design=expt$design)
}
\seealso{
\code{\link{ggplot2}}, \code{\link{geom_dl}}
}

