% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{plot_pcs}
\alias{plot_pcs}
\title{A quick and dirty PCA plotter of arbitrary components against one another.}
\usage{
plot_pcs(data, first = "PC1", second = "PC2", variances = NULL,
  design = NULL, plot_title = NULL, plot_labels = NULL, size = 5)
}
\arguments{
\item{data}{a dataframe of principle components PC1 .. PCN with any other arbitrary information.}

\item{first}{principle component PCx to put on the x axis.}

\item{second}{principle component PCy to put on the y axis.}

\item{variances}{a list of the percent variance explained by each component.}

\item{design}{the experimental design with condition batch factors.}

\item{plot_title}{a title for the plot.}

\item{plot_labels}{a parameter for the labels on the plot.}
}
\value{
a ggplot2 PCA plot
}
\description{
A quick and dirty PCA plotter of arbitrary components against one another.
}
\examples{
\dontrun{
 pca_plot = plot_pcs(pca_data, first="PC2", second="PC4", design=expt$design)
}
}
\seealso{
\pkg{ggplot2} \code{\link[directlabels]{geom_dl}}
}

