% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proteomics.r
\name{plot_pyprophet_counts}
\alias{plot_pyprophet_counts}
\title{Count some aspect(s) of the pyprophet data and plot them.}
\usage{
plot_pyprophet_counts(
  pyprophet_data,
  type = "count",
  keep_real = TRUE,
  keep_decoys = TRUE,
  expt_names = NULL,
  label_chars = 10,
  title = NULL,
  scale = NULL,
  ...
)
}
\arguments{
\item{pyprophet_data}{List containing the pyprophet results.}

\item{type}{What to count/plot?}

\item{keep_real}{Do we keep the real data when plotting the data? (perhaps
we only want the decoys)}

\item{keep_decoys}{Do we keep the decoys when plotting the data?}

\item{expt_names}{Names for the x-axis of the plot.}

\item{label_chars}{Maximum number of characters before abbreviating sample names.}

\item{title}{Title the plot?}

\item{scale}{Put the data on a specific scale?}

\item{...}{Further arguments, presumably for colors or some such.}
}
\value{
Boxplot describing the desired column from the data.
}
\description{
This function is mostly redundant with the plot_mzxml_boxplot above.
Unfortunately, the two data types are subtly different enough that I felt it
not worth while to generalize the functions.
}
