% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_qq_all}
\alias{plot_qq_all}
\title{Quantile/quantile comparison of the mean of all samples vs. each sample.}
\usage{
plot_qq_all(data, design = NULL, colors = NULL, labels = "short", ...)
}
\arguments{
\item{data}{Expressionset, exp, or dataframe of samples.}

\item{design}{Experimental design}

\item{colors}{Input colors}

\item{labels}{What kind of labels to print?}

\item{...}{Arguments passed presumably from graph_metrics().}
}
\value{
List containing:
 logs = a recordPlot() of the pairwise log qq plots.
 ratios = a recordPlot() of the pairwise ratio qq plots.
 means = a table of the median values of all the summaries of the qq plots.
}
\description{
This allows one to visualize all individual data columns against the mean of
all columns of data in order to see if any one is significantly different
than the cloud.
}
\examples{
if (is.null(pombe_se)) {
  source(system.file("examples", "create_se.R", package = "hpgltools"))
}

## Most distribution plots are particularly informative with the raw data.

## Make a boxplot of the SE
a_boxplot <- plot_boxplot(pombe_se)
a_boxplot

## The analagous density plot of the samples
a_density <- plot_density(pombe_se)
a_density

pombe_subset <- pombe_se[, c(1, 2, 3, 4)]
## QQ plots comparing samples to each other
qq <- plot_qq_all(pombe_subset)
qq[["logs"]]

## Show the percentage of all reads comprised by the top-n genes.
## This might be smarter if moved to the log scale?
topn_genes <- plot_topn(pombe_subset)
topn_genes

## Create a violin of the coefficient of variance on a per-condition basis.
varcoef <- plot_variance_coefficients(pombe_se)
varcoef
}
\seealso{
[Biobase]
}
