% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_sample_heatmap,ExpressionSet-method}
\alias{plot_sample_heatmap,ExpressionSet-method}
\title{Plot a sample heatmap of an ExpressionSet.}
\usage{
\S4method{plot_sample_heatmap}{ExpressionSet}(
  data,
  colors = NULL,
  design = NULL,
  heatmap_colors = NULL,
  expt_names = NULL,
  dendrogram = "column",
  row_label = NA,
  plot_title = NULL,
  Rowv = TRUE,
  Colv = TRUE,
  label_chars = 10,
  filter = TRUE,
  ...
)
}
\description{
Plot a sample heatmap of an ExpressionSet.
}
