% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_sample_heatmap}
\alias{plot_sample_heatmap}
\title{Make a heatmap.3 description of the similarity of the genes among samples.}
\usage{
plot_sample_heatmap(
  data,
  colors = NULL,
  design = NULL,
  heatmap_colors = NULL,
  expt_names = NULL,
  dendrogram = "column",
  row_label = NA,
  plot_title = NULL,
  Rowv = TRUE,
  Colv = TRUE,
  label_chars = 10,
  filter = TRUE,
  ...
)
}
\arguments{
\item{data}{Expt/expressionset/dataframe set of samples.}

\item{colors}{Color scheme of the samples (not needed if input is an expt).}

\item{design}{Design matrix describing the experiment (gotten for free if an expt).}

\item{expt_names}{Alternate samples names.}

\item{dendrogram}{Where to put dendrograms?}

\item{row_label}{Passed through to heatmap.2.}

\item{plot_title}{Title of the plot!}

\item{Rowv}{Reorder the rows by expression?}

\item{Colv}{Reorder the columns by expression?}

\item{label_chars}{Maximum number of characters before abbreviating sample names.}

\item{filter}{Filter the data before performing this plot?}

\item{...}{More parameters for a good time!}
}
\value{
a recordPlot() heatmap describing the samples.
}
\description{
Sometimes you just want to see how the genes of an experiment are related to
each other.  This can handle that.  These heatmap functions should probably
be replaced with neatmaps or heatplus or whatever it is, as the annotation
dataframes in them are pretty awesome.
}
\seealso{
[gplots::heatmap.2()]
}
