% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_significant_bar}
\alias{plot_significant_bar}
\title{Plot significant genes by contrast with different colors for significance levels.}
\usage{
plot_significant_bar(
  ups,
  downs,
  maximum = NULL,
  text = TRUE,
  color_list = c("lightcyan", "lightskyblue", "dodgerblue", "plum1", "orchid",
    "purple4"),
  color_names = c("a_up_inner", "b_up_middle", "c_up_outer", "a_down_inner",
    "b_down_middle", "c_down_outer")
)
}
\arguments{
\item{ups}{Set of up-regulated genes.}

\item{downs}{Set of down-regulated genes.}

\item{maximum}{Maximum/minimum number of genes to display.}

\item{text}{Add text at the ends of the bars describing the number of genes >/< 0 fc.}

\item{color_list}{Set of colors to use for the bars.}

\item{color_names}{Categories associated with aforementioned colors.}
}
\value{
weird significance bar plots
}
\description{
This is my attempt to recapitulate some plots made in Laura and Najib's mbio
paper.  The goal of the plot is to show a few ranges of significance as
differently colored and stacked bars.  The colors are nice because Najib and
Laura chose them.
}
\seealso{
[ggplot2] [extract_significant_genes()]
}
