% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_splicing.r
\name{plot_suppa}
\alias{plot_suppa}
\title{Given some psi and tpm data, make a pretty plot!}
\usage{
plot_suppa(dpsi, tpm, events = NULL, psi = NULL, sig_threshold = 0.05)
}
\description{
This should take either a dataframe or filename for the psi data from suppa,
along with the same for the average log tpm data (acquired from suppa
diffSplice with --save_tpm_events)
}
