% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_splicing.r
\name{plot_suppa}
\alias{plot_suppa}
\title{Given some psi and tpm data, make a pretty plot!}
\usage{
plot_suppa(
  dpsi,
  tpm,
  events = NULL,
  psi = NULL,
  sig_threshold = 0.05,
  label_type = NULL,
  alpha = 0.7
)
}
\arguments{
\item{dpsi}{Table provided by suppa containing all the metrics.}

\item{tpm}{Table provided by suppa containing all the tpm values.}

\item{events}{List of event types to include.}

\item{psi}{Limit the set of included events by psi value?}

\item{sig_threshold}{Use this significance threshold.}

\item{label_type}{Choose a type of event to label.}

\item{alpha}{How see-through should the points be in the plot?}
}
\value{
List containing the plot and some of the requisite data.
}
\description{
This should take either a dataframe or filename for the psi data from suppa,
along with the same for the average log tpm data (acquired from suppa
diffSplice with --save_tpm_events)
}
\examples{
 \dontrun{
 suppa_plot <- plot_suppa(dpsi_file, tmp_file)
}
}
\seealso{
[plot_rmats()]
}
