% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/093_plot_dotplot.r
\name{plot_svfactor}
\alias{plot_svfactor}
\title{Make a dotplot of some categorised factors and a set of SVs (for other factors).}
\usage{
plot_svfactor(
  expt,
  svest,
  sv = 1,
  chosen_factor = "batch",
  factor_type = "factor"
)
}
\arguments{
\item{expt}{Experiment from which to acquire the design, counts, etc.}

\item{svest}{Set of surrogate variable estimations from sva/svg
or batch estimates.}

\item{sv}{Which surrogate to plot?}

\item{chosen_factor}{Factor to compare against.}

\item{factor_type}{This may be a factor or range, it is intended to plot
a scatterplot if it is a range, a dotplot if a factor.}
}
\value{
surrogate variable plot as per Leek's work
}
\description{
This should make a quick df of the factors and surrogates and plot them.
}
\examples{
\dontrun{
 estimate_vs_snps <- plot_svfactor(start, surrogate_estimate, "snpcategory")
}
}
\seealso{
\pkg{ggplot2}
}
