% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.r
\name{plot_topgo_densities}
\alias{plot_topgo_densities}
\title{Plot the density of categories vs. the possibilities of all categories.}
\usage{
plot_topgo_densities(godata, table)
}
\arguments{
\item{godata}{Result from topgo.}

\item{table}{Table of genes.}
}
\value{
density plot as per topgo
}
\description{
This can make a large number of plots.
}
\seealso{
\pkg{topGO}
}
