% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.r
\name{plot_topn}
\alias{plot_topn}
\title{Plot the representation of the top-n genes in the total counts / sample.}
\usage{
plot_topn(data, title = NULL, direct = TRUE, num = 100, ...)
}
\arguments{
\item{data}{Dataframe to perform pairwise qqplots with.}

\item{title}{A title for the plot.}

\item{num}{The N in top-n genes, if null, do them all.}

\item{...}{Extra arguments, currently unused.}
}
\value{
List containing the ggplot2
}
\description{
One question we might ask is: how much do the most abundant genes in a
samples comprise the entire sample?  This plot attempts to provide a visual
hint toward answering this question.  It does so by rank-ordering all the
genes in every sample and dividing their counts by the total number of reads
in that sample.  It then smooths the points to provide the resulting trend.
The steeper the resulting line, the more over-represented these top-n genes
are.  I suspect, but haven't tried yet, that the inflection point of the
resulting curve is also a useful diagnostic in this question.
}
