% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plots.R
\name{plot_topn_gsea,all_cprofiler-method}
\alias{plot_topn_gsea,all_cprofiler-method}
\title{Plot topn GSEA results given the result from all_cprofiler}
\usage{
\S4method{plot_topn_gsea}{all_cprofiler}(gse, topn = 20, id = NULL, add_score = TRUE)
}
\description{
Plot topn GSEA results given the result from all_cprofiler
}
