% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{plot_topn_gsea,clusterprofiler_result-method}
\alias{plot_topn_gsea,clusterprofiler_result-method}
\title{Plot topn GSEA results given the result from simple_clusterprofiler}
\usage{
\S4method{plot_topn_gsea}{clusterprofiler_result}(gse, topn = 20, id = NULL, add_score = TRUE)
}
\description{
Plot topn GSEA results given the result from simple_clusterprofiler
}
