% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/115_dimension_reduction.r
\name{plot_tsne}
\alias{plot_tsne}
\title{Shortcut to plot_pca(pc_method="tsne")}
\usage{
plot_tsne(...)
}
\arguments{
\item{...}{Arguments for plot_pca()}
}
\description{
Shortcut to plot_pca(pc_method="tsne")
}
