% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tsne.r
\name{plot_tsne}
\alias{plot_tsne}
\title{Make a ggplot TSNE plot describing the samples' clustering.}
\usage{
plot_tsne(data, design = NULL, plot_colors = NULL, seed = 1,
  chosen_features = NULL, number_features = NULL, plot_labels = NULL,
  perplexity = NULL, min_variance = 0.001, plot_title = NULL,
  plot_size = 5, size_column = NULL, components = 2, iterations = 1000,
  theta = 0.3, pca = TRUE, component_x = 1, component_y = 2, ...)
}
\arguments{
\item{data}{an expt set of samples.}

\item{design}{a design matrix and.}

\item{plot_colors}{a color scheme.}

\item{seed}{A seed for Rtsne}

\item{chosen_features}{Use these features?}

\item{number_features}{And this number.}

\item{plot_labels}{add labels?  Also, what type?  FALSE, "default", or "fancy".}

\item{perplexity}{I am perplexed.}

\item{min_variance}{Only include genes with more than this variance.}

\item{plot_title}{a title for the plot.}

\item{plot_size}{size for the glyphs on the plot.}

\item{size_column}{use an experimental factor to size the glyphs of the plot}

\item{components}{Look for n components.}

\item{iterations}{Perform n iterations of tsne.}

\item{theta}{Yay greek!}

\item{pca}{Seed with an initial pca plot?}

\item{component_x}{Which component goes on the x-axis?}

\item{component_y}{And which goes on the y-axis?}

\item{...}{arglist from elipsis!}
}
\value{
a list containing the following:
\enumerate{
 \item  plot = a plot
}
}
\description{
Make a ggplot TSNE plot describing the samples' clustering.
}
\examples{
\dontrun{
 tsne_plot <- plot_tsne(expt=expt)
 tsne_plot
}
}
\seealso{
\pkg{directlabels}
 \code{\link[directlabels]{geom_dl}} \code{\link{plot_pcs}}
}
