% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_variance_coefficients}
\alias{plot_variance_coefficients}
\title{Look at the (biological)coefficient of variation/quartile coefficient of dispersion
with respect to an experimental factor.}
\usage{
plot_variance_coefficients(
  data,
  x_axis = "condition",
  colors = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{Expressionset/epxt to poke at.}

\item{x_axis}{Factor in the experimental design we may use to group the data
and calculate the dispersion metrics.}

\item{colors}{Set of colors to use when making the violins}

\item{title}{Optional title to include with the plot.}

\item{...}{Extra arguments to pass along.}
}
\value{
List of plots showing the coefficients vs. genes along with the data.
}
\description{
I want to look at the (B)CV of some data with respect to condition/batch/whatever.
This function should make that possible, with some important caveats.  The
most appropriate metric  is actually the biological coefficient of variation
as calculated by DESeq2/EdgeR; but the metrics I am currently taking are the
simpler and less appropriate CV(sd/mean) and QCD(q3-q1/q3+q1).
}
