% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{plot_volcano_condition_de,hpgltools::combine_de_tables-method}
\alias{plot_volcano_condition_de,hpgltools::combine_de_tables-method}
\title{Plot a volcano plot using the result from combine_de_tables}
\usage{
\S4method{plot_volcano_condition_de}{hpgltools::combine_de_tables}(
  input,
  table_name,
  alpha = 0.5,
  fc_col = "logFC",
  fc_name = "log2 fold change",
  line_color = "black",
  line_position = "bottom",
  logfc = 1,
  p_col = "adj.P.Val",
  p_name = "-log10 p-value",
  pval = 0.05,
  shapes_by_state = FALSE,
  stroke = TRUE,
  fill = TRUE,
  color_high = NULL,
  color_low = NULL,
  size = 2,
  size_by = NULL,
  size_categories = NULL,
  invert = FALSE,
  label = NULL,
  label_type = "text",
  label_column = "hgnc_symbol",
  label_size = 6,
  nudge_x = 0,
  nudge_y = 0,
  ...
)
}
\arguments{
\item{input}{Table of DE values, likely from combine_de_tables().}

\item{table_name}{Name the table!}

\item{alpha}{Make see-through.}

\item{fc_col}{Column containing the fold-change values.}

\item{fc_name}{Axis label.}

\item{line_color}{Color for the demarcation lines.}

\item{line_position}{Put the lines above or below the dots.}

\item{logfc}{Demarcation line for fold-change significance.}

\item{p_col}{Column containing the significance information.}

\item{p_name}{Axis label for the significance.}

\item{pval}{Demarcation for (in)significance.}

\item{shapes_by_state}{Change point shapes according to their states?}

\item{stroke}{Add a stroke to the dots.}

\item{fill}{Fill the dots.}

\item{color_high}{Color for the ups.}

\item{color_low}{and the downs.}

\item{size}{Point size}

\item{size_by}{Use the provided column to change the size of each dot according
to the range of values in this factor.}

\item{size_categories}{Define the number of categories used to resize the dots with size_by.}

\item{invert}{Flip the plot?}

\item{label}{Label some points?}

\item{label_type}{What kind of label to apply.}

\item{label_column}{Using this column in the data.}

\item{label_size}{Use this font size for the labels on the plot.}

\item{nudge_x}{Nudge the label on the x axis.}

\item{nudge_y}{Nudge the label on the y axis.}

\item{...}{Arbitrary args.}
}
\description{
I therefore took a modified copy of her implementation and added it here.
}
