% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{plot_volcano_condition_de}
\alias{plot_volcano_condition_de}
\title{Theresa's volcano plots are objectively nicer because they are colored by condition.}
\usage{
plot_volcano_condition_de(
  input,
  table_name,
  alpha = 0.5,
  fc_col = "logFC",
  fc_name = "log2 fold change",
  line_color = "black",
  line_position = "bottom",
  logfc = 1,
  p_col = "adj.P.Val",
  p_name = "-log10 p-value",
  pval = 0.05,
  shapes_by_state = FALSE,
  color_high = "darkred",
  color_low = "darkblue",
  size = 2,
  invert = FALSE,
  label = NULL,
  label_column = "hgncsymbol",
  label_size = 6
)
}
\arguments{
\item{input}{Table of DE values, likely from combine_de_tables().}

\item{table_name}{Name the table!}

\item{alpha}{Make see-through.}

\item{fc_col}{Column containing the fold-change values.}

\item{fc_name}{Axis label.}

\item{line_color}{Color for the demarcation lines.}

\item{line_position}{Put the lines above or below the dots.}

\item{logfc}{Demarcation line for fold-change significance.}

\item{p_col}{Column containing the significance information.}

\item{p_name}{Axis label for the significance.}

\item{pval}{Demarcation for (in)significance.}

\item{shapes_by_state}{Change point shapes according to their states?}

\item{color_high}{Color for the ups.}

\item{color_low}{and the downs.}

\item{size}{Point size}

\item{invert}{Flip the plot?}

\item{label}{Label some points?}

\item{label_column}{Using this column in the data.}
}
\description{
I therefore took a modified copy of her implementation and added it here.
}
