% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point.r
\name{plot_volcano_de}
\alias{plot_volcano_de}
\title{Make a pretty Volcano plot!}
\usage{
plot_volcano_de(table, tooltip_data = NULL, gvis_filename = NULL,
  logfc_cutoff = 1, pval_cutoff = 0.05, size = 2, alpha = 0.6,
  fc_col = "logFC", p_col = "adj.P.Val", fc_name = "log2 fold change",
  p_name = "-log10 p-value", line_color = "black", color_by = "p",
  line_position = "bottom", shapes_by_state = TRUE, color_list = c(`FALSE`
  = "darkred", `TRUE` = "darkblue"), ...)
}
\arguments{
\item{table}{Dataframe from limma's toptable which includes log(fold change) and an
adjusted p-value.}

\item{tooltip_data}{Df of tooltip information for gvis.}

\item{gvis_filename}{Filename to write a fancy html graph.}

\item{size}{How big are the dots?}

\item{alpha}{How transparent to make the dots.}

\item{...}{I love parameters!}

\item{p_cutoff}{Cutoff defining significant from not.}

\item{fc_cutoff}{Cutoff defining the minimum/maximum fold change for interesting.  This is log,
so I went with +/- 0.8 mostly arbitrarily as the default.}

\item{xaxis_column}{Column from the data to use on the x axis (logFC)}

\item{yaxis_column}{Column from the data to use on the y axis (p-value)}
}
\value{
Ggplot2 volcano scatter plot.  This is defined as the -log10(p-value) with respect to
    log(fold change).  The cutoff values are delineated with lines and mark the boundaries
    between 'significant' and not.  This will make a fun clicky googleVis graph if requested.
}
\description{
Volcano plots and MA plots provide quick an easy methods to view the set of (in)significantly
differentially expressed genes.  In the case of a volcano plot, it places the -log10 of the
p-value estimate on the y-axis and the fold-change between conditions on the x-axis.  Here is a
neat snippet from wikipedia: "The concept of volcano plot can be generalized to other
applications, where the x-axis is related to a measure of the strength of a statistical signal,
and y-axis is related to a measure of the statistical significance of the signal."
}
\examples{
\dontrun{
 plot_volcano_de(table, gvis_filename="html/fun_ma_plot.html")
 ## Currently this assumes that a variant of toptable was used which
 ## gives adjusted p-values.  This is not always the case and I should
 ## check for that, but I have not yet.
}
}
\seealso{
\pkg{limma}
 \code{\link{plot_gvis_ma}} \code{\link[limma]{toptable}}
 \code{\link[limma]{voom}} \code{\link{hpgl_voom}} \code{\link[limma]{lmFit}}
 \code{\link[limma]{makeContrasts}} \code{\link[limma]{contrasts.fit}}
}
