% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{pp}
\alias{pp}
\title{png() shortcut}
\usage{
pp(file)
}
\arguments{
\item{a}{filename to write}
}
\value{
a png with height=width=9 inches and a high resolution
}
\description{
I hate remembering my options for png()
}

