% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{pp}
\alias{pp}
\title{png() shortcut}
\usage{
pp(file, width = 9, height = 9, res = 180, ...)
}
\arguments{
\item{file}{a filename to write}

\item{width}{How wide?}

\item{height}{How high?}

\item{res}{The chosen resolution.}
}
\value{
a png/svg/eps/ps/pdf with height=width=9 inches and a high resolution
}
\description{
I hate remembering my options for png()
}
