% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.aucc_info}
\alias{print.aucc_info}
\title{Print a summary of the result from calculate_aucc().}
\usage{
\method{print}{aucc_info}(x)
}
\arguments{
\item{x}{list of results from calculate_aucc().}
}
\description{
Print a summary of the result from calculate_aucc().
}
