% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.classifier_evaluation}
\alias{print.classifier_evaluation}
\title{Print the result from self_evaluate_model().}
\usage{
\method{print}{classifier_evaluation}(x)
}
\arguments{
\item{x}{List showing AUC/ROC curves of the test performed, summary
thereof, the confusion matrix, and vector of incorrectly called samples.}
}
\description{
Print the result from self_evaluate_model().
}
