% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.clusterprofiler_result}
\alias{print.clusterprofiler_result}
\title{Print a clusterprofiler over representation search.}
\usage{
\method{print}{clusterprofiler_result}(x, ...)
}
\arguments{
\item{x}{Monstrous list of the various results, including but not
limited to plots, go-gene mappings, enrichmed, kegg, david, GO
analyses.}

\item{...}{Other args to match the generic.}
}
\description{
Print a clusterprofiler over representation search.
}
