% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{print.combined_de}
\alias{print.combined_de}
\title{Print a summary of a combined differential expression analysis.}
\usage{
\method{print}{combined_de}(x)
}
\arguments{
\item{x}{Set of DE results in one big pile.}
}
\description{
Print a summary of a combined differential expression analysis.
}
