% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{print.combined_table}
\alias{print.combined_table}
\title{Print a summary of a single combined DE result.}
\usage{
\method{print}{combined_table}(x)
}
\arguments{
\item{x}{Big table and some summaries.}
}
\description{
Print a summary of a single combined DE result.
}
