% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.correlation_heatmap}
\alias{print.correlation_heatmap}
\title{Print the result of plot_corheat().}
\usage{
\method{print}{correlation_heatmap}(x, ...)
}
\arguments{
\item{x}{List containing the correlations observed and a recorded heatmap.3().}

\item{...}{Other args to match the generic.}
}
\description{
Print the result of plot_corheat().
}
