% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.density_plot}
\alias{print.density_plot}
\title{Print the result from plot_density().}
\usage{
\method{print}{density_plot}(x)
}
\arguments{
\item{x}{List containing the plot, summary of the
conditions/batches/samples, and the melted table of reads/gene.}
}
\description{
Print the result from plot_density().
}
