% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{print.goseq_result}
\alias{print.goseq_result}
\title{Print a summary of a goseq over representation search.}
\usage{
\method{print}{goseq_result}(x)
}
\arguments{
\item{x}{list of results from goseq.}
}
\description{
Print a summary of a goseq over representation search.
}
