% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{print.gostats_result}
\alias{print.gostats_result}
\title{Print a summary of a gostats over representation search.}
\usage{
\method{print}{gostats_result}(x)
}
\arguments{
\item{x}{list of results from gostats.}
}
\description{
Print a summary of a gostats over representation search.
}
