% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{print.gprofiler_result}
\alias{print.gprofiler_result}
\title{Print a summary of a gprofiler over representation search.}
\usage{
\method{print}{gprofiler_result}(x)
}
\arguments{
\item{x}{list of results from gprofiler.}
}
\description{
Print a summary of a gprofiler over representation search.
}
