% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.gsva_result}
\alias{print.gsva_result}
\title{Print a gsva category search.}
\usage{
\method{print}{gsva_result}(x)
}
\arguments{
\item{x}{List containing signature annotations, the result from
GSVA, a modified expressionset, the signatures used, and method.}
}
\description{
Print a gsva category search.
}
