% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.gsva_sig}
\alias{print.gsva_sig}
\title{Print gsva categories deemed 'significant'.}
\usage{
\method{print}{gsva_sig}(x)
}
\arguments{
\item{x}{List of scored GSVA results, including some plots,
likelihood tables, subsets of significant categories, etc.}
}
\description{
Print gsva categories deemed 'significant'.
}
