% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.libsize_plot}
\alias{print.libsize_plot}
\title{Print a summary of the library sizes from an experiment.}
\usage{
\method{print}{libsize_plot}(x)
}
\arguments{
\item{x}{list of results from plot_libsize().}
}
\description{
Print a summary of the library sizes from an experiment.
}
