% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.nonzero_plot}
\alias{print.nonzero_plot}
\title{Print a nonzero plot.}
\usage{
\method{print}{nonzero_plot}(x, ...)
}
\arguments{
\item{x}{List containing the plot and table describing the data.}

\item{...}{Other args to match the generic.}
}
\description{
Print a nonzero plot.
}
