% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.pca_result}
\alias{print.pca_result}
\title{Print a summary of the result from one of the various dimension reductions.}
\usage{
\method{print}{pca_result}(x)
}
\arguments{
\item{x}{Result from plot_pca()}
}
\description{
Print a summary of the result from one of the various dimension reductions.
}
