% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.proper_estimate}
\alias{print.proper_estimate}
\title{Print a summary of the result from simple_proper().}
\usage{
\method{print}{proper_estimate}(x)
}
\arguments{
\item{x}{list of results from simple_proper().}
}
\description{
Print a summary of the result from simple_proper().
}
