% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.sig_genes}
\alias{print.sig_genes}
\title{Print a summary of some significantly differentially expressed genes.}
\usage{
\method{print}{sig_genes}(x)
}
\arguments{
\item{x}{Significant gene data structure.}
}
\description{
Print a summary of some significantly differentially expressed genes.
}
