% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.sig_intersect}
\alias{print.sig_intersect}
\title{Print a summary of a set of intersection significant genes from multiple analyses.}
\usage{
\method{print}{sig_intersect}(x)
}
\arguments{
\item{x}{Intersection gene lists.}
}
\description{
Print a summary of a set of intersection significant genes from multiple analyses.
}
