% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.topgo_result}
\alias{print.topgo_result}
\title{Print a topgo over representation search.}
\usage{
\method{print}{topgo_result}(x, ...)
}
\arguments{
\item{x}{List of the various over/under representation analyses
provided by topGO, the associated plots, and coerced enrichResults.}

\item{...}{Other args to match the generic.}
}
\description{
Print a topgo over representation search.
}
