% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.varpart}
\alias{print.varpart}
\title{Print a summary of variance partition results.}
\usage{
\method{print}{varpart}(x)
}
\arguments{
\item{x}{list of results from simple_varpart().}
}
\description{
Print a summary of variance partition results.
}
