% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{print_ups_downs}
\alias{print_ups_downs}
\title{Reprint the output from extract_significant_genes()}
\usage{
print_ups_downs(upsdowns, sig_table = "excel/significant_genes.xlsx")
}
\arguments{
\item{upsdowns}{the output from extract_significant_genes()}

\item{sig_table}{table to write to}
}
\value{
the return from write_xls
}
\description{
I found myself needing to reprint these excel sheets because I added some new information.
This shortcuts that process for me.
}
\seealso{
\code{\link{combine_de_tables}}
}

