% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{print_ups_downs}
\alias{print_ups_downs}
\title{Reprint the output from extract_significant_genes().}
\usage{
print_ups_downs(upsdowns, wb = NULL, excel = "excel/significant_genes.xlsx",
  csv = NULL, according = "limma", summary_count = 1, ma = FALSE)
}
\arguments{
\item{upsdowns}{Output from extract_significant_genes().}

\item{wb}{Workbook object to use for writing, or start a new one.}

\item{excel}{Filename for writing the data.}

\item{csv}{Write a csv instead/also?}

\item{according}{Use limma, deseq, or edger for defining 'significant'.}

\item{summary_count}{For spacing sequential tables one after another.}

\item{ma}{Include ma plots?}
}
\value{
Return from write_xls.
}
\description{
I found myself needing to reprint these excel sheets because I
added some new information. This shortcuts that process for me.
}
\seealso{
\code{\link{combine_de_tables}}
}

