% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{pval_plot}
\alias{pval_plot}
\title{Make a pvalue plot from a df of IDs, scores, and p-values.}
\usage{
pval_plot(df, ontology = "MF")
}
\arguments{
\item{df}{Some data from topgo/goseq/clusterprofiler.}

\item{ontology}{Ontology to plot (MF,BP,CC).}
}
\value{
Ggplot2 plot of pvalues vs. ontology.
}
\description{
This function seeks to make generating pretty pvalue plots as shown by clusterprofiler easier.
}
\seealso{
\link[goseq]{goseq} \pkg{ggplot2}
}

