% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{pval_plot}
\alias{pval_plot}
\title{pval_plot()  Make a pvalue plot from a df of IDs, scores, and p-values.}
\usage{
pval_plot(df, ontology = "MF")
}
\arguments{
\item{df}{some data from topgo/goseq/clusterprofiler.}

\item{ontology}{default='MF'  an ontology to plot (MF,BP,CC).}
}
\value{
a plot!
}
\description{
This function seeks to make generating pretty pvalue plots as shown by clusterprofiler easier.
}
\seealso{
\code{\link{goseq}}
}

