% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{query_eupathdb}
\alias{query_eupathdb}
\title{Queries one of the EuPathDB APIs using a GET request and returns a dataframe
representation of the result.}
\usage{
query_eupathdb(data_provider, organism, query_args,
  wadl = "GeneQuestions/GenesByTaxon", format = "json",
  workaround = FALSE)
}
\arguments{
\item{data_provider}{Name of data provider to query (e.g. 'TriTrypDB')}

\item{organism}{Full name of organism, as used by EuPathDB APIs}

\item{query_args}{String of additional query arguments}

\item{wadl}{String specifying API service to be queried}

\item{format}{String specifying API response type (currently only 'json'
is supported)}
}
\value{
list containing response from API request.

More information
----------------
1. http://tritrypdb.org/tritrypdb/serviceList.jsp
}
\description{
Queries one of the EuPathDB APIs using a GET request and returns a dataframe
representation of the result.
}
\author{
Keith Hughitt
}
