% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tables.r
\name{read_metadata}
\alias{read_metadata}
\title{Given a table of meta data, read it in for use by create_expt()}
\usage{
read_metadata(file, header = FALSE, sep = ",")
}
\arguments{
\item{file}{a csv/xls file to read}

\item{header}{does the table have a header (usually for csv)}

\item{sep}{separator for csv files}
}
\value{
a df of metadata
}
\description{
Given a table of meta data, read it in for use by create_expt()
}

