% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tables.r
\name{read_metadata}
\alias{read_metadata}
\title{Given a table of meta data, read it in for use by create_expt().}
\usage{
read_metadata(file, ...)
}
\arguments{
\item{file}{Csv/xls file to read.}

\item{...}{Arguments for arglist, used by sep, header and similar read.csv/read.table parameters.}
}
\value{
Df of metadata.
}
\description{
Reads an experimental design in a few different formats in preparation for creating an expt.
}

