% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/160_variants.r
\name{read_snp_columns}
\alias{read_snp_columns}
\title{Read the output from bcfutils into a count-table-esque}
\usage{
read_snp_columns(samples, file_lst, column = "diff_count")
}
\arguments{
\item{samples}{Sample names to read.}

\item{file_lst}{Set of files to read.}

\item{column}{Column from the bcf file to read.}
}
\value{
A big honking data table.
}
\description{
Previously, I put all my bcfutils output files into one directory.  This
function would iterate through every file in that directory and add the
contents as columns to this growing data table.  Now it works by accepting a
list of filenames (presumably kept in the metadata for the experiment) and
reading them into the data table.  It is worth noting that it can accept
either a column name or index -- which when you think about it is pretty much
always true, but in this context is particularly interesting since I changed
the names of all the columns when I rewrote this functionality.
}
