% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{read_snp_columns}
\alias{read_snp_columns}
\title{Read the output from bcfutils into a count-table-esque}
\usage{
read_snp_columns(samples, file_lst, column = "diff_count")
}
\arguments{
\item{samples}{Sample names to read.}

\item{file_lst}{Set of files to read.}
}
\value{
A big honking data table.
}
\description{
I put all my bcfutils output files into one directory, so hunt them down and
read them into a data table.
}
