% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_install.r
\name{require.auto}
\alias{require.auto}
\title{Automatic loading and/or installing of packages.}
\usage{
require.auto(lib, update = FALSE)
}
\arguments{
\item{lib}{String name of a library to check/install.}

\item{update}{Update packages?}
}
\value{
0 or 1, whether a package was installed or not.
}
\description{
Load a library, install it first if necessary.
}
\details{
This was taken from:
http://sbamin.com/2012/11/05/tips-for-working-in-r-automatically-install-missing-package/
}
\examples{
\dontrun{
 require.auto("ggplot2")
}
}
\seealso{
\pkg{BiocInstaller}
 \code{\link[BiocInstaller]{biocLite}} \code{\link{install.packages}}
}
