% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoloads.R
\name{require.auto}
\alias{require.auto}
\title{##  Paragraphs are separated by singlelines
Automatic loading and/or installing of packages.}
\usage{
require.auto(lib, github_path = NULL, verbose = FALSE, update = FALSE)
}
\arguments{
\item{lib}{string name of a library}
}
\value{
NULL currently
}
\description{
\code{require.auto} loads a library, and installs it first if necessary.
}
\details{
This was taken from:
http://sbamin.com/2012/11/05/tips-for-working-in-r-automatically-install-missing-package/
}
\examples{
## require.auto("ggplot2")
}
\seealso{
\code{\link{biocLite}} and \code{\link{install.packages}}
}

