% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoloads.r
\name{require.auto}
\alias{require.auto}
\title{Automatic loading and/or installing of packages.}
\usage{
require.auto(lib, github_path = NULL, verbose = FALSE, update = FALSE)
}
\arguments{
\item{lib}{string name of a library}

\item{github_path}{an optional github username/path.}

\item{verbose}{print some information while loading.}

\item{update}{update packages?}
}
\value{
NULL currently
}
\description{
Load a library, install it first if necessary.
}
\details{
This was taken from:
http://sbamin.com/2012/11/05/tips-for-working-in-r-automatically-install-missing-package/
}
\examples{
\dontrun{
require.auto("ggplot2")
}
}
\seealso{
\link[BiocInstaller]{biocLite} \link{install.packages}
}

