% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_eupathdb.r
\name{retrieve_eupathdb_attributes}
\alias{retrieve_eupathdb_attributes}
\title{Queries one of the EuPathDB APIs for gene data}
\usage{
retrieve_eupathdb_attributes(provider = "TriTrypDB",
  species = "Leishmania major", table = "GOTerms",
  wadl = "GeneQuestions/GenesByTaxonGene", format = "json",
  overwrite = FALSE, dir = "eupathdb")
}
\arguments{
\item{provider}{Name of data provider to query (e.g. 'TriTrypDB')}

\item{species}{Full name of organism, as used by EuPathDB APIs}

\item{table}{Name of the particular table to be retrieved (e.g.
'GoTerms')}

\item{wadl}{String specifying API service to be queried}

\item{format}{String specifying API response type (currently only 'json'
is supported)}

\item{overwrite}{Overwrite existing data?}

\item{dir}{Where to place the downloaded information?}
}
\value{
list containing response from API request.
}
\description{
Note that as of version 30, EuPathDB no longer supports table queries for
genes, and as such this method must be used instead. Support for tables
queries is likely to be returned in future versions of EuPathDB.
}
\author{
Keith Hughitt
}
