% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{rex}
\alias{rex}
\title{Resets the display and xauthority variables to the new computer I am using so
that plot() works.}
\usage{
rex(display = ":0")
}
\arguments{
\item{display}{DISPLAY variable to use,  if NULL it looks in ~/.displays/$(host).last}
}
\description{
Resets the display and xauthority variables to the new computer I am using so
that plot() works.
}
