% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{rowData,ExpressionSet-method}
\alias{rowData,ExpressionSet-method}
\title{A getter of the gene information from an ExpressionSet, synonymous with fData().}
\usage{
\S4method{rowData}{ExpressionSet}(x, use.names = TRUE, ...)
}
\arguments{
\item{x}{Input}

\item{...}{them too!}

\item{withDimnames}{yes}
}
\description{
A getter of the gene information from an ExpressionSet, synonymous with fData().
}
