% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{rowData,expt-method}
\alias{rowData,expt-method}
\title{A getter of the gene information from an expt, synonymous with fData().}
\usage{
\S4method{rowData}{expt}(x, use.names = TRUE, ...)
}
\arguments{
\item{x}{Input}

\item{use.names}{Use those names...}

\item{...}{them too!}
}
\description{
A getter of the gene information from an expt, synonymous with fData().
}
