% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{rowData<-,expt-method}
\alias{rowData<-,expt-method}
\title{A setter to put the gene information into an expt.}
\usage{
\S4method{rowData}{expt}(x, i, withDimnames = TRUE, ...) <- value
}
\description{
A setter to put the gene information into an expt.
}
