% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{sampleNames<-,SummarizedExperiment,character-method}
\alias{sampleNames<-,SummarizedExperiment,character-method}
\title{A setter to put the samples names into a SummarizedExperiment.}
\usage{
\S4method{sampleNames}{SummarizedExperiment,character}(object) <- value
}
\arguments{
\item{object}{Input}

\item{value}{new names.}
}
\description{
A setter to put the samples names into a SummarizedExperiment.
}
