% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{sampleNames,SummarizedExperiment-method}
\alias{sampleNames,SummarizedExperiment-method}
\title{A getter to get the samples names from a SummarizedExperiment.}
\usage{
\S4method{sampleNames}{SummarizedExperiment}(object)
}
\description{
A getter to get the samples names from a SummarizedExperiment.
}
