% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_annotations}
\alias{sanitize_annotations}
\title{Annotation sanitizers}
\usage{
sanitize_annotations(input, ...)
}
\arguments{
\item{input}{Input dataset to sanitize.}

\item{...}{Arguments passed through.}
}
\description{
Annotation sanitizers
}
