% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{sanitize_expt}
\alias{sanitize_expt}
\title{Get rid of characters which will mess up contrast making and such before
playing with an expt.}
\usage{
sanitize_expt(expt)
}
\arguments{
\item{expt}{An expt object to clean.}
}
\description{
Get rid of characters which will mess up contrast making and such before
playing with an expt.
}
