% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sanitize_expt_metadata}
\alias{sanitize_expt_metadata}
\title{Adding an alias to sanitize_metadata until I decide how I want to name this.}
\usage{
sanitize_expt_metadata(...)
}
\arguments{
\item{...}{Arguments for sanitize_metadata().}
}
\description{
Adding an alias to sanitize_metadata until I decide how I want to name this.
}
