% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sanitize_expt_pData}
\alias{sanitize_expt_pData}
\title{Adding an alias to sanitize_metadata until I decide how I want to name this.}
\usage{
sanitize_expt_pData(
  expt,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL,
  numeric = FALSE
)
}
\arguments{
\item{...}{Arguments for sanitize_metadata().}
}
\description{
Adding an alias to sanitize_metadata until I decide how I want to name this.
}
