% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_metadata,character-method}
\alias{sanitize_metadata,character-method}
\title{Sanitize a character vector (e.g. a single column/row)}
\usage{
\S4method{sanitize_metadata}{character}(
  input,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL,
  numeric = FALSE
)
}
\arguments{
\item{input}{Input datastructure}
}
\description{
I should probably revisit this, for the moment it just takes the
vector, coerces it to a single-column dataframe, and invokes
sanitize on it.
}
