% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_metadata,DFrame-method}
\alias{sanitize_metadata,DFrame-method}
\title{Sanitize a DFrame}
\usage{
\S4method{sanitize_metadata}{DFrame}(
  input,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL,
  numeric = FALSE
)
}
\arguments{
\item{input}{Input metadata}

\item{columns}{Set of columns to check, if left NULL, all columns
will be molested.}

\item{na_value}{Fill NA values with a string.}

\item{lower}{Set everything to lowercase?}

\item{punct}{Remove punctuation?}

\item{factorize}{Set some columns to factors?  If set to a vector
of length >=1, then set all of the provided columns to factors.
When set to 'heuristic', set any columns with <= max_levels
different elements to factors.}

\item{max_levels}{When heuristically setting factors, use this as
the heuristic, when NULL it is the number of samples / 6}

\item{spaces}{Remove any spaces in this column?}

\item{numbers}{Sanitize numbers by adding a prefix character to them?}

\item{numeric}{Recast the values as numeric when possible?}
}
\description{
I wrote this function after spending a couple of hours confused
because one cell in my metadata said 'cure ' instead of 'cure' and
I could not figure out why chaos reigned in my analyses.  There is
a sister to this somewhere else which checks that the expected
levels of a metadata factor are consistent; this is because in
another analysis we essentially had a cell which said 'cyre' and a
similar data explosion occurred.
}
