% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_metadata,ExpressionSet-method}
\alias{sanitize_metadata,ExpressionSet-method}
\title{Sanitize an expressionset}
\usage{
\S4method{sanitize_metadata}{ExpressionSet}(
  input,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  spaces = FALSE,
  numbers = NULL
)
}
\arguments{
\item{input}{Input datastructure}
}
\description{
Sanitize an expressionset
}
