% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_metadata,ExpressionSet-method}
\alias{sanitize_metadata,ExpressionSet-method}
\title{Given an expressionset, sanitize pData columns of interest.}
\usage{
\S4method{sanitize_metadata}{ExpressionSet}(
  input,
  columns = NULL,
  na_string = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  spaces = FALSE,
  numbers = NULL
)
}
\arguments{
\item{input}{Input metadata}

\item{columns}{Set of columns to check, if left NULL, all columns
will be molested.}

\item{na_string}{Fill NA values with a string.}

\item{lower}{Set everything to lowercase?}

\item{punct}{Remove punctuation?}

\item{spaces}{Remove any spaces in this column?}

\item{numbers}{Sanitize numbers by adding a prefix character to them?}
}
\description{
I wrote this function after spending a couple of hours confused
because one cell in my metadata said 'cure ' instead of 'cure' and
I could not figure out why chaos reigned in my analyses.  There is
a sister to this somewhere else which checks that the expected
levels of a metadata factor are consistent; this is because in
another analysis we essentially had a cell which said 'cyre' and a
similar data explosion occurred.
}
