% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_metadata,SummarizedExperiment-method}
\alias{sanitize_metadata,SummarizedExperiment-method}
\title{Metadata sanitizers for an expt}
\usage{
\S4method{sanitize_metadata}{SummarizedExperiment}(
  input,
  columns = NULL,
  na_value = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  factorize = "heuristic",
  max_levels = NULL,
  spaces = FALSE,
  numbers = NULL,
  numeric = FALSE
)
}
\description{
Metadata sanitizers for an expt
}
