% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{sanitize_metadata,SummarizedExperiment-method}
\alias{sanitize_metadata,SummarizedExperiment-method}
\title{Metadata sanitizers for an expressionset}
\usage{
\S4method{sanitize_metadata}{SummarizedExperiment}(
  meta,
  columns = NULL,
  na_string = "notapplicable",
  lower = TRUE,
  punct = TRUE,
  spaces = FALSE,
  numbers = NULL
)
}
\description{
Metadata sanitizers for an expressionset
}
