% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{saveme}
\alias{saveme}
\title{saveme()  Make a backup rdata file for future reference}
\usage{
saveme(directory = "savefiles", backups = 4)
}
\arguments{
\item{backups}{default=10  how many revisions?

I often use R over a sshfs connection, sometimes with significant latency, and
I want to be able to save/load my R sessions relatively quickly.
Thus this function uses pxz to compress the R session maximally and relatively fast.}

\item{dir}{the directory to save the Rdata file.}
}
\description{
saveme()  Make a backup rdata file for future reference
}

