% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.r
\name{saveme}
\alias{saveme}
\title{Make a backup rdata file for future reference}
\usage{
saveme(directory = "savefiles", backups = 4, filename = "Rdata.rda.xz")
}
\arguments{
\item{directory}{Directory to save the Rdata file.}

\item{backups}{How many revisions?}

\item{filename}{Choose a filename.}
}
\value{
Command string used to save the global environment.
}
\description{
I often use R over a sshfs connection, sometimes with significant latency, and
I want to be able to save/load my R sessions relatively quickly.
Thus this function uses pxz to compress the R session maximally and relatively fast.
This assumes you have pxz installed and >= 4 CPUs.
}
\examples{
\dontrun{
 saveme()
}
}
\seealso{
\link{save} \link{pipe}
}

