% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{semantic_copynumber_extract}
\alias{semantic_copynumber_extract}
\title{Extract multicopy genes from up/down gene expression lists.}
\usage{
semantic_copynumber_extract(de_list, min_copies = 2, semantic = c("mucin",
  "sialidase", "RHS", "MASP", "DGF", "GP63"), semantic_column = "1.tooltip")
}
\arguments{
\item{de_list}{List of sets of genes deemed significantly
up/down with a column expressing approximate count numbers.}

\item{min_copies}{Keep only those genes with >= n putative
copies.}

\item{semantic}{Set of strings with gene names to exclude.}

\item{semantic_column}{Column in the DE table used to find the
semantic strings for removal.}
}
\value{
Smaller list of up/down genes.
}
\description{
The function semantic_copynumber_filter() is the inverse of this.
}
\details{
Currently untested, used for Trypanosome analyses primarily, thus the default strings.
}

