% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_shared.r
\name{semantic_copynumber_filter}
\alias{semantic_copynumber_filter}
\title{Remove multicopy genes from up/down gene expression lists.}
\usage{
semantic_copynumber_filter(
  input,
  max_copies = 2,
  use_files = FALSE,
  invert = TRUE,
  semantic = c("mucin", "sialidase", "RHS", "MASP", "DGF", "GP63"),
  semantic_column = "1.tooltip"
)
}
\arguments{
\item{input}{List of sets of genes deemed significantly
up/down with a column expressing approximate count numbers.}

\item{max_copies}{Keep only those genes with <= n putative
copies.}

\item{use_files}{Use a set of sequence alignments to define the copy numbers?}

\item{invert}{Keep these genes rather than drop them?}

\item{semantic}{Set of strings with gene names to exclude.}

\item{semantic_column}{Column in the DE table used to find the
semantic strings for removal.}
}
\value{
Smaller list of up/down genes.
}
\description{
In our parasite data, there are a few gene types which are
consistently obnoxious.  Multi-gene families primarily where the
coding sequences are divergent, but the UTRs nearly identical.  For
these genes, our sequence based removal methods fail and so this
just excludes them by name.
}
\details{
Currently untested, used for Trypanosome analyses primarily, thus the default
strings.
}
\examples{
 \dontrun{
  pruned <- semantic_copynumber_filter(table, semantic=c("ribosomal"))
  ## Get rid of all genes with 'ribosomal' in the annotations.
}
}
\seealso{
\code{\link{semantic_copynumber_extract}}
}
