% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{semantic_expt_filter}
\alias{semantic_expt_filter}
\title{Filter a dataset by gene when using deprecated function names.}
\usage{
semantic_expt_filter(...)
}
\description{
Filter a dataset by gene when using deprecated function names.
}
