% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{semantic_expt_filter}
\alias{semantic_expt_filter}
\title{Remove/keep specifically named genes from an expt.}
\usage{
semantic_expt_filter(
  input,
  invert = FALSE,
  topn = NULL,
  semantic = c("mucin", "sialidase", "RHS", "MASP", "DGF", "GP63"),
  semantic_column = "description"
)
}
\arguments{
\item{input}{Expt to filter.}

\item{invert}{The default is to remove the genes with the semantic strings.
Keep them when inverted.}

\item{topn}{Take the topn most abundant genes rather than a text based heuristic.}

\item{semantic}{Character list of strings to search for in the annotation
data.}

\item{semantic_column}{Column in the annotations to search.}
}
\value{
A presumably smaller expt.
}
\description{
I find subsetting weirdly confusing.  Hopefully this function will allow one
to include/exclude specific genes/families based on string comparisons.
}
