% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{semantic_filter,expt-method}
\alias{semantic_filter,expt-method}
\title{Filter an expt}
\usage{
\S4method{semantic_filter}{expt}(
  input,
  invert = FALSE,
  topn = NULL,
  semantic = c("mucin", "sialidase", "RHS", "MASP", "DGF", "GP63"),
  semantic_column = "description"
)
}
\description{
Filter an expt
}
