% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequence_attributes,character-method}
\alias{sequence_attributes,character-method}
\title{Invoke sequence_attributes() when provided a fasta filename or name of BSgenome.}
\usage{
\S4method{sequence_attributes}{character}(fasta, gff = NULL, type = "gene", key = NULL)
}
\arguments{
\item{fasta}{String describing a bsgenome or fasta file.}
}
\description{
Invoke sequence_attributes() when provided a fasta filename or name of BSgenome.
}
