% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{sequence_attributes}
\alias{sequence_attributes}
\title{Gather some simple sequence attributes.}
\usage{
sequence_attributes(fasta, gff = NULL, type = "gene", key = NULL)
}
\arguments{
\item{fasta}{Genome encoded as a fasta file.}

\item{gff}{Optional gff of annotations (if not provided it will just ask the
whole genome).}

\item{type}{Column of the gff file to use.}

\item{key}{What type of entry of the gff file to key from?}
}
\value{
List of data frames containing gc/at/gt/ac contents.
}
\description{
This extends the logic of the pattern searching in pattern_count_genome() to
search on some other attributes.
}
\examples{
 pa_data <- get_paeruginosa_data()
 pa_fasta <- pa_data[["fasta"]]
 pa_gff <- pa_data[["gff"]]
 pa_attribs <- sequence_attributes(pa_fasta, gff = pa_gff)
 head(pa_attribs)
}
\seealso{
[Biostrings] [Rsamtools]
}
