% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_batches,expt-method}
\alias{set_batches,expt-method}
\title{The following method should disappear soon.}
\usage{
\S4method{set_batches}{expt}(exp, fact, ids = NULL)
}
\arguments{
\item{exp}{Input expt}

\item{fact}{Factor used to set the batches}

\item{ids}{IDs to set}
}
\description{
The following method should disappear soon.
}
