% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_colors,data.frame,factor-method}
\alias{set_colors,data.frame,factor-method}
\title{I want to move the logic of these various functions to their simplest
accessors...
Thus, use this as a starting point and then create methods for
each of the pieces here.}
\usage{
\S4method{set_colors}{data.frame,factor}(
  exp,
  colors = NULL,
  color_column = "color",
  change_by = "condition",
  chosen_palette = "Dark2",
  ...
)
}
\arguments{
\item{exp}{Datastructure to modify.}

\item{color_column}{metadta column containing colors.}

\item{change_by}{Metadata column to use to define colors.}

\item{chosen_palette}{ColorBrewer palette used when choosing colors.}

\item{...}{Args passed to the various setters.}
}
\value{
Modified data structure.
}
\description{
I want to move the logic of these various functions to their simplest
accessors...
Thus, use this as a starting point and then create methods for
each of the pieces here.
}
