% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_colors}
\alias{set_colors}
\title{Create a simple generic for setting colors in expressionsets/SEs/etc.}
\usage{
set_colors(exp, ...)
}
\arguments{
\item{exp}{Datastructure to modify.}

\item{...}{Args passed to the various setters.}
}
\value{
Modified data structure.
}
\description{
Create a simple generic for setting colors in expressionsets/SEs/etc.
}
