% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_conditions,expt-method}
\alias{set_conditions,expt-method}
\title{This method should go away soon.}
\usage{
\S4method{set_conditions}{expt}(
  exp,
  fact = NULL,
  ids = NULL,
  prefix = NULL,
  null_cell = "null",
  colors = TRUE,
  ...
)
}
\arguments{
\item{exp}{Input expt}

\item{fact}{metadata factor to use.}

\item{ids}{Specific IDs to change.}

\item{prefix}{add a prefix to the conditions.}

\item{null_cell}{What to do with missing data?}

\item{colors}{Set the colors as well?}

\item{...}{Arguments passed along.}
}
\description{
Set conditions of an expt
}
