% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_colors}
\alias{set_expt_colors}
\title{Change the colors of an expt}
\usage{
set_expt_colors(expt, colors = TRUE, chosen_palette = "Dark2",
  change_by = "condition")
}
\arguments{
\item{expt}{Expt to modify}

\item{colors}{colors to replace}

\item{chosen_palette}{I usually use Dark2 as the RColorBrewer palette.}

\item{change_by}{Assuming a list is passed, cross reference by condition or sample?}
}
\value{
expt Send back the expt with some new metadata
}
\description{
When exploring differential analyses, it might be useful to play with the conditions/batches of
the experiment.  Use this to make that easier.
}
\examples{
\dontrun{
unique(esmer_expt$design$conditions)
chosen_colors <- list(
   "cl14_epi" = "#FF8D59",
   "clbr_epi" = "#962F00",
   "cl14_tryp" = "#D06D7F",
   "clbr_tryp" = "#A4011F",
   "cl14_late" = "#6BD35E",
   "clbr_late" = "#1E7712",
   "cl14_mid" = "#7280FF",
   "clbr_mid" = "#000D7E")
esmer_expt <- set_expt_colors(expt=esmer_expt, colors=chosen_colors)
}
}

