% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_condition}
\alias{set_expt_condition}
\title{Change the condition of an expt}
\usage{
set_expt_condition(expt, fact = NULL, ids = NULL, ...)
}
\arguments{
\item{expt}{Expt to modify}

\item{fact}{Conditions to replace}

\item{ids}{Specific sample IDs to change.}

\item{...}{Extra arguments are given to arglist.}
}
\value{
expt Send back the expt with some new metadata
}
\description{
When exploring differential analyses, it might be useful to play with the conditions/batches of
the experiment.  Use this to make that easier.
}
\examples{
\dontrun{
 expt = set_expt_condition(big_expt, factor=c(some,stuff,here))
}
}
\seealso{
\code{\link{set_expt_batch}} \code{\link{create_expt}}
}
