% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{set_expt_factors}
\alias{set_expt_factors}
\title{Change the factors (condition and batch) of an expt}
\usage{
set_expt_factors(expt, condition = NULL, batch = NULL, ids = NULL, ...)
}
\arguments{
\item{expt}{Expt to modify}

\item{condition}{New condition factor}

\item{batch}{New batch factor}

\item{ids}{Specific sample IDs to change.}

\item{...}{Arguments passed along (likely colors)}
}
\value{
expt Send back the expt with some new metadata
}
\description{
When exploring differential analyses, it might be useful to play with the conditions/batches of
the experiment.  Use this to make that easier.
}
\examples{
\dontrun{
 expt = set_expt_factors(big_expt, condition="column", batch="another_column")
}
}
\seealso{
\code{\link{set_expt_condition}} \code{\link{set_expt_batch}}
}
