% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_genenames,SummarizedExperiment-method}
\alias{set_genenames,SummarizedExperiment-method}
\title{Set the genenames of a SE}
\usage{
\S4method{set_genenames}{SummarizedExperiment}(exp, ids = NULL, column = NULL, ...)
}
\description{
Set the genenames of a SE
}
