% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{significant_barplots}
\alias{significant_barplots}
\title{Given the set of significant genes from combine_de_tables(), provide a view
of how many are significant up/down.}
\usage{
significant_barplots(
  combined,
  lfc_cutoffs = c(0, 1, 2),
  invert = FALSE,
  p = 0.05,
  z = NULL,
  p_type = "adj",
  according_to = "all",
  order = NULL,
  maximum = NULL,
  ...
)
}
\arguments{
\item{combined}{Result from combine_de_tables and/or extract_significant_genes().}

\item{lfc_cutoffs}{Choose 3 fold changes to define the queries.  0, 1, 2
mean greater/less than 0 followed by 2 fold and 4 fold cutoffs.}

\item{invert}{Reverse the order of contrasts for readability?}

\item{p}{Chosen p-value cutoff.}

\item{z}{Choose instead a z-score cutoff.}

\item{p_type}{Adjusted or not?}

\item{according_to}{limma, deseq, edger, basic, or all of the above.}

\item{order}{Choose a specific order for the plots.}

\item{maximum}{Set a specific limit on the number of genes on the x-axis.}

\item{...}{More arguments are passed to arglist.}
}
\value{
list containing the significance bar plots and some information to
 hopefully help interpret them.
}
\description{
These plots are pretty annoying, and I am certain that this function is not
well written, but it provides a series of bar plots which show the number of
genes/contrast which are up and down given a set of fold changes and
p-value.
}
\examples{
\dontrun{
 expt <- create_expt(metadata = "some_metadata.xlsx", gene_info = annotations)
 pairwise_result <- all_pairwise(expt)
 combined_result <- combine_de_tables(pairwise_result)
 ## Damn I wish I were smrt enough to make this elegant, but I cannot.
 barplots <- significant_barplots(combined_result)
}
}
