% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{simple_clusterprofiler}
\alias{simple_clusterprofiler}
\title{Perform the array of analyses in the 2016-04 version of clusterProfiler}
\usage{
simple_clusterprofiler(sig_genes, all_genes, orgdb = "org.Dm.eg.db",
  orgdb_from = "FLYBASE", orgdb_to = c("ENSEMBL", "SYMBOL", "ENTREZID"),
  go_level = 3, pcutoff = 0.05, qcutoff = 0.1, fc_column = "logFC",
  permutations = 100, min_groupsize = 5, kegg_prefix = "Dmel_",
  kegg_organism = "dme", kegg_id_column = "FLYBASECG", categories = 12)
}
\arguments{
\item{sig_genes}{Dataframe of genes deemed 'significant.'}

\item{all_genes}{Dataframe of all genes in the analysis, primarily for gse analyses.}

\item{orgdb}{Name of the orgDb used for gathering annotation data.}

\item{orgdb_from}{Name of a key in the orgdb used to cross reference to entrez IDs.}

\item{orgdb_to}{List of keys to grab from the orgdb for cross referencing ontologies.}

\item{go_level}{How deep into the ontology tree should this dive for over expressed categories.}

\item{pcutoff}{P-value cutoff for 'significant' analyses.}

\item{qcutoff}{Q-value cutoff for 'significant' analyses.}

\item{fc_column}{When extracting vectors of all genes, what column should be used?}

\item{permutations}{How many permutations for GSEA-ish analyses?}

\item{min_groupsize}{What is the minimum ontology group's size?}

\item{kegg_prefix}{Many KEGG ids need a prefix before they will cross reference.}

\item{kegg_organism}{Choose the 3 letter KEGG organism name here.}

\item{kegg_id_column}{Column in the orgdb to use for cross referencing to KEGG.}

\item{categories}{How many categories should be plotted in bar/dot plots?}
}
\description{
The new version of clusterProfiler has a bunch of new toys.  However, it is more stringent in
terms of input in that it now explicitly expects to receive annotation data in terms of a orgdb
object.  This is mostly advantageous, but will probably cause some changes in the other ontology
functions in the near future.  This function is an initial pass at making something similar to my
previous 'simple_clusterprofiler()' but using these new toys.
}

