% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{simple_clusterprofiler}
\alias{simple_clusterprofiler}
\title{Perform a simplified clusterProfiler analysis}
\usage{
simple_clusterprofiler(de_genes, goids = NULL, golevel = 4, pcutoff = 0.1,
  qcutoff = 1, fold_changes = NULL, include_cnetplots = FALSE,
  showcategory = 12, universe = NULL, organism = "lm", gff = NULL,
  wrapped_width = 20, method = "Wallenius", padjust = "BH", ...)
}
\arguments{
\item{de_genes}{a data frame of differentially expressed genes, containing IDs and whatever other columns}

\item{goids}{a file containing mappings of genes to goids in the format expected by topgo}

\item{golevel}{a relative level in the tree for printing p-value plots, higher is more specific}

\item{pcutoff}{a p-value cutoff}

\item{qcutoff}{a q-value cutoff}

\item{fold_changes}{a df of fold changes for the DE genes}

\item{include_cnetplots}{the cnetplots are often stupid and can be left behind}

\item{showcategory}{how many categories to show in p-value plots}

\item{universe}{universe to use}

\item{organism}{name of the species to use}

\item{gff}{gff file to generate the universe}

\item{wrapped_width}{width of ontology names in the pvalue plots}

\item{method}{pvalue calculation method}

\item{padjust}{a method for adjusting the p-values}

\item{...}{more options!}
}
\value{
a big list including the following:
  mf_interesting: A table of the interesting molecular function groups
  bp_interesting: A table of the interesting biological process groups
  cc_interesting: A table of the interesting cellular component groups
  mf_pvals: A histogram of the molecular function p-values
  bp_pvals: Ditto, biological process
  cc_pvals: And cellular component...
  mf_enriched: A table of the enriched molecular function groups by adjusted p-value.
  bp_enriched: yep, you guessed it
  cc_enriched: cellular component, too
  mf_all/bp_all/cc_all: A table of all go categories observed (mf/bp/cc respectively)
  mfp_plot/bpp_plot/ccp_plot: ggplot2 p-value bar plots describing the over represented groups
  mf_cnetplot/bp_cnetplot/cc_cnetplot: clusterProfiler cnetplots
  mf_group_barplot/bp_group_barplot/cc_group_barplot: The group barplots from clusterProfiler
}
\description{
Perform a simplified clusterProfiler analysis
}
\examples{
\dontrun{
 up_cluster = simple_clusterprofiler(mga2_ll_thy_top, goids=goids, gff="reference/genome/gas.gff")
## > Some chattery while it runs
## tail(head(up_cluster$bp_interesting, n=10), n=1)
## > ID ont GeneRatio BgRatio     pvalue   p.adjust    qvalue
## > 10 GO:0009311  BP     5/195 10/1262 0.01089364 0.01089364 0.1272835
## >   geneID Count
## >   10 M5005_Spy1632/M5005_Spy1637/M5005_Spy1635/M5005_Spy1636/M5005_Spy1638     5
## >   Description
## >   10 oligosaccharide metabolic process
}
}

