% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.r
\name{simple_clusterprofiler}
\alias{simple_clusterprofiler}
\title{Perform the array of analyses in the 2016-04 version of clusterProfiler}
\usage{
simple_clusterprofiler(
  sig_genes,
  de_table = NULL,
  orgdb = "org.Dm.eg.db",
  orgdb_from = NULL,
  orgdb_to = "ENTREZID",
  go_level = 3,
  pcutoff = 0.05,
  qcutoff = 0.1,
  fc_column = "logFC",
  second_fc_column = "limma_logfc",
  updown = "up",
  permutations = 100,
  min_groupsize = 5,
  kegg_prefix = NULL,
  kegg_organism = NULL,
  do_gsea = TRUE,
  categories = 12,
  excel = NULL,
  do_david = FALSE,
  david_id = "ENTREZ_GENE_ID",
  david_user = "unknown@unknown.org"
)
}
\arguments{
\item{sig_genes}{Dataframe of genes deemed 'significant.'}

\item{de_table}{Dataframe of all genes in the analysis, primarily for gse
analyses.}

\item{orgdb}{Name of the orgDb used for gathering annotation data.}

\item{orgdb_from}{Name of a key in the orgdb used to cross reference to entrez IDs.}

\item{orgdb_to}{List of keys to grab from the orgdb for cross referencing
ontologies.}

\item{go_level}{How deep into the ontology tree should this dive for over
expressed categories.}

\item{pcutoff}{P-value cutoff for 'significant' analyses.}

\item{qcutoff}{Q-value cutoff for 'significant' analyses.}

\item{fc_column}{When extracting vectors of all genes, what column should be used?}

\item{second_fc_column}{When extracting vectors of all genes, what column
should be tried the second time around?}

\item{updown}{Include the less than expected ontologies?}

\item{permutations}{How many permutations for GSEA-ish analyses?}

\item{min_groupsize}{Minimum size of an ontology before it is included.}

\item{kegg_prefix}{Many KEGG ids need a prefix before they will cross reference.}

\item{kegg_organism}{Choose the 3 letter KEGG organism name here.}

\item{do_gsea}{Perform gsea searches?}

\item{categories}{How many categories should be plotted in bar/dot plots?}

\item{excel}{Print the results to an excel file?}

\item{do_david}{Attempt to use the DAVID database for a search?}

\item{david_id}{Which column to use for cross-referencing to DAVID?}

\item{david_user}{Default registered username to use.}
}
\value{
a list
}
\description{
The new version of clusterProfiler has a bunch of new toys.  However, it is
more stringent in terms of input in that it now explicitly expects to receive
annotation data in terms of a orgdb object.  This is mostly advantageous, but
will probably cause some changes in the other ontology functions in the near
future.  This function is an initial pass at making something similar to my
previous 'simple_clusterprofiler()' but using these new toys.
}
\examples{
\dontrun{
 holyasscrackers <- simple_clusterprofiler(gene_list, all_genes, "org.Dm.eg.db")
}
}
\seealso{
\pkg{clusterProfiler}
}
