% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_clusterprofiler.R
\name{simple_cp_enricher}
\alias{simple_cp_enricher}
\title{Generic enrichment using clusterProfiler.}
\usage{
simple_cp_enricher(sig_genes, de_table, go_db = NULL)
}
\arguments{
\item{sig_genes}{Set of 'significant' genes as a table.}

\item{de_table}{All genes from the original analysis.}

\item{go_db}{Dataframe of GO->ID matching the gene names of sig_genes to GO
categories.}
}
\value{
Table of 'enriched' categories.
}
\description{
culsterProfiler::enricher provides a quick and easy enrichment analysis given
a set of siginficant' genes and a data frame which connects each gene to a
category.
}
