% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/280_motif.r
\name{simple_gadem}
\alias{simple_gadem}
\title{run the rGADEM suite}
\usage{
simple_gadem(
  inputfile,
  genome = "BSgenome.Hsapiens.UCSC.hs19",
  p = 0.1,
  e = 0,
  ...
)
}
\arguments{
\item{inputfile}{Fasta or bed file containing sequences to search.}

\item{genome}{BSgenome to read.}

\item{...}{Parameters for plotting the gadem result.}
}
\value{
A list containing slots for plots, the stdout output from gadem, the
  gadem result, set of occurences of motif, and the returned set of motifs.
}
\description{
This should provide a set of rGADEM results given an input file of sequences
and a genome.
}
