% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.r
\name{simple_goseq}
\alias{simple_goseq}
\title{simple_goseq() Perform a simplified goseq analysis}
\usage{
simple_goseq(de_genes, all_genes = NULL, lengths = NULL, goids = NULL,
  doplot = TRUE, adjust = 0.1, pvalue = 0.1, qvalue = 0.1,
  goseq_method = "Wallenius", padjust_method = "BH", species = NULL,
  length_db = "ensGene", gff = NULL, ...)
}
\arguments{
\item{de_genes}{a data frame of differentially expressed genes, containing IDs and whatever other columns}

\item{all_genes}{the universe of possible genes}

\item{lengths}{the length of each gene with an ID in de_genes}

\item{goids}{a list of ontology accessions to gene accessions}

\item{doplot}{include pwf plots}

\item{adjust}{minimum adjusted pvalue}

\item{pvalue}{minimum pvalue}

\item{qvalue}{minimum qvalue}

\item{goseq_method}{testing used by goseq}

\item{padjust_method}{which method to adjust the pvalues}

\item{species}{optionally choose a species from supportedOrganisms()}

\item{length_db}{Source of gene lengths}

\item{gff}{gff file source of gene lengths}

\item{...}{extra parameters which I do not recall}
}
\value{
a big list including:
  the pwd:pwf function,
  alldata:the godata dataframe,
  pvalue_histogram:p-value histograms,
  godata_interesting:the ontology information of the enhanced groups,
  term_table:the goterms with some information about them,
  mf_subset:a plot of the MF enhanced groups,
  mfp_plot:the pvalues of the MF group,
  bp_subset:a plot of the BP enhanced groups,
  bpp_plot,
  cc_subset,
  and ccp_plot
}
\description{
simple_goseq() Perform a simplified goseq analysis
}
\seealso{
\pkg{goseq} \link[goseq]{goseq} \link[goseq]{nullp}
}

