% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_goseq.R
\name{simple_goseq}
\alias{simple_goseq}
\title{Perform a simplified goseq analysis}
\usage{
simple_goseq(de_genes, all_genes = NULL, lengths = NULL, goids = NULL,
  adjust = 0.1, pvalue = 0.1, qvalue = 0.1, goseq_method = "Wallenius",
  padjust_method = "BH", species = NULL, length_db = "ensGene")
}
\arguments{
\item{de_genes}{a data frame of differentially expressed genes, containing IDs and whatever other columns}

\item{lengths}{the length of each gene with an ID in de_genes}

\item{goids}{a list of ontology accessions to gene accessions}
}
\value{
a big list including:
  the pwd:pwf function,
  alldata:the godata dataframe,
  pvalue_histogram:p-value histograms,
  godata_interesting:the ontology information of the enhanced groups,
  term_table:the goterms with some information about them,
  mf_subset:a plot of the MF enhanced groups,
  mfp_plot:the pvalues of the MF group,
  bp_subset:a plot of the BP enhanced groups,
  bpp_plot,
  cc_subset,
  and ccp_plot
}
\description{
Perform a simplified goseq analysis
}
\seealso{
\code{\link{goseq}} and \code{\link{clusterProfiler}}
}

