% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.r
\name{simple_goseq}
\alias{simple_goseq}
\title{Perform a simplified goseq analysis.}
\usage{
simple_goseq(
  sig_genes,
  go_db = NULL,
  length_db = NULL,
  doplot = TRUE,
  adjust = 0.1,
  pvalue = 0.1,
  plot_title = NULL,
  length_keytype = "transcripts",
  go_keytype = "entrezid",
  goseq_method = "Wallenius",
  padjust_method = "BH",
  bioc_length_db = "ensGene",
  expand_categories = TRUE,
  excel = NULL,
  ...
)
}
\arguments{
\item{sig_genes}{Data frame of differentially expressed genes, containing IDs etc.}

\item{go_db}{Database of go to gene mappings (OrgDb/OrganismDb)}

\item{length_db}{Database of gene lengths (gff/TxDb)}

\item{doplot}{Include pwf plots?}

\item{adjust}{Minimum adjusted pvalue for 'significant.'}

\item{pvalue}{Minimum pvalue for 'significant.'}

\item{length_keytype}{Keytype to provide to extract lengths}

\item{go_keytype}{Keytype to provide to extract go IDs}

\item{goseq_method}{Statistical test for goseq to use.}

\item{padjust_method}{Which method to use to adjust the pvalues.}

\item{bioc_length_db}{Source of gene lengths?}

\item{expand_categories}{Expand the GO categories to make the results more readable?}

\item{excel}{Print the results to an excel file?}

\item{...}{Extra parameters which I do not recall}
}
\value{
Big list including:
 the pwd:pwf function,
 alldata:the godata dataframe,
 pvalue_histogram:p-value histograms,
 godata_interesting:the ontology information of the enhanced groups,
 term_table:the goterms with some information about them,
 mf_subset:a plot of the MF enhanced groups,
 mfp_plot:the pvalues of the MF group,
 bp_subset:a plot of the BP enhanced groups,
 bpp_plot,
 cc_subset,
 and ccp_plot
}
\description{
goseq can be pretty difficult to get set up for non-supported organisms.
This attempts to make that process a bit simpler as well as give some
standard outputs which should be similar to those returned by
clusterprofiler/topgo/gostats/gprofiler.
}
\examples{
\dontrun{
 lotsotables <- simple_goseq(gene_list, godb, lengthdb)
}
}
\seealso{
[goseq] [GO.db] [GenomicFeatures] [stats::p.adjust()]
}
