% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gostats.r
\name{simple_gostats}
\alias{simple_gostats}
\title{A simplification function for gostats, in the same vein as those written for clusterProfiler, goseq, and topGO.}
\usage{
simple_gostats(de_genes, gff, goids, universe_merge = "ID",
  second_merge_try = "locus_tag", organism = "fun", pcutoff = 0.1,
  direction = "over", conditional = FALSE, categorysize = NULL,
  gff_type = "CDS", ...)
}
\arguments{
\item{de_genes}{The set of differentially expressed genes in the limma format as before}

\item{gff}{The annotation information for this genome}

\item{goids}{The set of GOids, as before in the format ID/GO}
}
\value{
dunno yet
}
\description{
GOstats has a couple interesting peculiarities:  Chief among them: the gene IDs must be integers.
As a result, I am going to have this function take a gff file in order to get the go ids and
gene ids on the same page.
}
\seealso{
\code{\link{GOstats}}
}

