% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gostats.r
\name{simple_gostats}
\alias{simple_gostats}
\title{Simplification function for gostats, in the same vein as those written for clusterProfiler,
goseq, and topGO.}
\usage{
simple_gostats(sig_genes, gff, goids_df, universe_merge = "id",
  second_merge_try = "locus_tag", species = "fun", pcutoff = 0.1,
  direction = "over", conditional = FALSE, categorysize = NULL,
  gff_type = "cds", excel = NULL, ...)
}
\arguments{
\item{sig_genes}{Input list of differentially expressed genes.}

\item{gff}{Annotation information for this genome.}

\item{goids_df}{Set of GOids, as before in the format ID/GO.}

\item{universe_merge}{Column from which to create the universe of genes.}

\item{second_merge_try}{If the first universe merge fails, try this.}

\item{species}{Genbank organism to use.}

\item{pcutoff}{Pvalue cutoff for deciding significant.}

\item{direction}{Under or over represented categories.}

\item{conditional}{Perform a conditional search?}

\item{categorysize}{Category size below which to not include groups.}

\item{gff_type}{Gff column to use for creating the universe.}

\item{...}{More parameters!}
}
\value{
List of returns from GSEABase, Category, etc.
}
\description{
GOstats has a couple interesting peculiarities:  Chief among them: the gene IDs must be
integers. As a result, I am going to have this function take a gff file in order to get the go
ids and gene ids on the same page.
}
\examples{
\dontrun{
 knickerbockers <- simple_gostats(sig_genes, gff_file, goids)
}
}
\seealso{
\pkg{GSEABase} \pkg{Category}
}
