% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.R
\name{simple_gprofiler}
\alias{simple_gprofiler}
\title{Redirect users to simple_gprofiler2}
\usage{
simple_gprofiler(...)
}
\arguments{
\item{...}{Arguments passed to simple_gprofiler2()}
}
\description{
Redirect users to simple_gprofiler2
}
