% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_gprofiler.r
\name{simple_gprofiler}
\alias{simple_gprofiler}
\title{Run searches against the web service g:Profiler}
\usage{
simple_gprofiler(gene_list, species = "hsapiens", first_col = "logFC",
  second_col = "limma_logfc")
}
\arguments{
\item{gene_list}{guess!}

\item{first_col}{where to search for the order of 'significant' first}

\item{second_col}{if that fails, try some where else.}

\item{organism}{an organism supported by gprofiler}
}
\value{
a list of results for go, kegg, reactome, and a few more.
}
\description{
Run searches against the web service g:Profiler
}

