% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.R
\name{simple_topgo}
\alias{simple_topgo}
\title{Perform a simplified topgo analysis}
\usage{
simple_topgo(de_genes, goid_map = "reference/go/id2go.map", goids_df = NULL,
  pvals = NULL, limitby = "fisher", limit = 0.1, signodes = 100,
  sigforall = TRUE, numchar = 300, selector = "topDiffGenes",
  overwrite = FALSE, densities = FALSE)
}
\arguments{
\item{de_genes}{a data frame of differentially expressed genes, containing IDs and whatever other columns}

\item{goid_map}{a file containing mappings of genes to goids in the format expected by topgo}
}
\value{
a big list including the various outputs from topgo
}
\description{
Perform a simplified topgo analysis
}

