% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_varpartition.R
\name{simple_varpart}
\alias{simple_varpart}
\title{Use variancePartition to try and understand where the variance lies in a data set.}
\usage{
simple_varpart(
  input,
  fstring = "~ condition + batch",
  do_fit = FALSE,
  cor_gene = 1,
  cpus = NULL,
  genes = 40,
  parallel = TRUE,
  strict_filter = TRUE,
  modify_input = TRUE
)
}
\arguments{
\item{input}{Some data}

\item{fstring}{Formula string describing the factors to query.}

\item{do_fit}{Perform a fitting using variancePartition?}

\item{cor_gene}{Provide a set of genes to look at the correlations, defaults
to the first gene.}

\item{cpus}{Number cpus to use}

\item{genes}{Number of genes to count.}

\item{parallel}{Use doParallel?}

\item{strict_filter}{Perform a strict filtering of the results via median_by_factor and dropping
any genes with a 0.}

\item{modify_input}{Add annotation columns with the variance/factor?}
}
\value{
List of plots and variance data frames
}
\description{
The arguments and usage of variancePartition are a bit opaque.  This function
attempts to fill in reasonable values and simplify its invocation.
}
\seealso{
[variancePartition] DOI:10.1186/s12859-016-1323-z.
}
