% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.r
\name{simple_xcell}
\alias{simple_xcell}
\title{Invoke xCell and pretty-ify the result.}
\usage{
simple_xcell(
  expt,
  signatures = NULL,
  genes = NULL,
  spill = NULL,
  expected_types = NULL,
  label_size = NULL,
  col_margin = 6,
  row_margin = 12,
  sig_cutoff = 0.2,
  verbose = TRUE,
  cores = 4,
  ...
)
}
\arguments{
\item{expt}{Expressionset to query.}

\item{signatures}{Alternate set of signatures to use.}

\item{genes}{Subset of genes to query.}

\item{spill}{The xCell spill parameter.}

\item{expected_types}{Set of assumed types in the data.}

\item{label_size}{How large to make labels when printing the final heatmap.}

\item{col_margin}{Used by par() when printing the final heatmap.}

\item{row_margin}{Ibid.}

\item{cores}{How many CPUs to use?}

\item{...}{Extra arguments when normalizing the data for use with xCell.}
}
\value{
Small list providing the output from xCell, the set of signatures,
 and heatmap.
}
\description{
I initially thought xCell might prove the best tool/method for exploring cell
deconvolution.  I slowly figured out its limitations, but still think it
seems pretty nifty for its use case.  Thus this function is intended to make
invoking it easier/faster.
}
\seealso{
[xCell]
}
