% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{slide_de_threshold}
\alias{slide_de_threshold}
\title{make a slideable threshold for cutting off significant genes in a DE table.}
\usage{
slide_de_threshold(
  de_table,
  contrast = 1,
  lfc = 1,
  p = 0.05,
  according_to = "deseq"
)
}
\description{
make a slideable threshold for cutting off significant genes in a DE table.
}
