% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_helpers_misc.r
\name{sm}
\alias{sm}
\title{Silence}
\usage{
sm(..., wrap = TRUE)
}
\arguments{
\item{...}{Some code to shut up.}

\item{wrap}{Wrap the invocation and try again if it failed?}
}
\value{
Whatever the code would have returned.
}
\description{
Some libraries/functions just won't shut up.  Ergo, silence, peasant!
This is a simpler silence peasant.
}
