% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{snp_density_primers}
\alias{snp_density_primers}
\title{Create a density function given a variant output and some metadata}
\usage{
snp_density_primers(
  snp_count,
  pdata_column = "condition",
  condition = "z2.3",
  cutoff = 20,
  bin_width = 600,
  divide = FALSE,
  topn = 400,
  target_temp = 53,
  max_primer_length = 50,
  bsgenome = "BSGenome.Leishmania.panamensis.MHOMCOL81L13.v52",
  gff = "reference/lpanamensis_col_v46.gff",
  feature_type = "protein_coding_gene",
  feature_start = "start",
  feature_end = "end",
  feature_strand = "strand",
  feature_chr = "seqnames",
  feature_type_column = "type",
  feature_id = "ID",
  feature_name = "description",
  truncate = TRUE,
  xref_genes = TRUE,
  verbose = FALSE,
  min_contig_length = NULL,
  min_gc_prop = 0.25,
  max_nmer_run = 5
)
}
\arguments{
\item{snp_count}{Result from count_expt_snps()}

\item{pdata_column}{Metadata column containing the condition of
interest.}

\item{condition}{Chosen condition to search for variants.}

\item{cutoff}{Minimum number of variants in a region.}

\item{bin_width}{Bin size/region of genome to consider.}

\item{divide}{Normalize by bin width?}

\item{topn}{Keep only this number of candidates.}

\item{target_temp}{Try to get primers with this Tm.}

\item{max_primer_length}{Keep primers at or less than this length.}

\item{bsgenome}{Genome package containing the sequence of interest.}

\item{gff}{GFF to define regions of interest.}

\item{feature_type}{GFF feature type to search against.}

\item{feature_start}{GFF column with the starts (needed?)}

\item{feature_end}{GFF column with the ends (needed?)}

\item{feature_strand}{GFF column with strand information (needed?)}

\item{feature_chr}{GFF column with chromosome information.}

\item{feature_type_column}{GFF column with type information.}

\item{feature_id}{GFF tag with the ID information.}

\item{feature_name}{GFF tag with the names.}

\item{truncate}{Truncate the results to just the columns I think
are useful.}

\item{min_gc_prop}{Minimum GC content for a suitable primer.}
}
\description{
It is hoped that this will point out regions of a genome which
might prove useful when designing PCR primers for a specific
condition in a dataset of variants.
}
