% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{snps_vs_genes}
\alias{snps_vs_genes}
\title{Make a summary of the observed snps by gene ID.}
\usage{
snps_vs_genes(
  exp,
  snp_result,
  start_column = "start",
  end_column = "end",
  snp_name_column = "seqnames",
  observed_in = NULL,
  more_than = 0,
  chr_column = "chromosome",
  ignore_strand = TRUE
)
}
\arguments{
\item{exp}{The original expressionset.}

\item{snp_result}{The result from get_snp_sets().}

\item{start_column}{Which column provides the start of each gene?}

\item{end_column}{and the end column of each gene?}

\item{snp_name_column}{Name of the column in the metadata with the sequence names.}

\item{observed_in}{Minimum proportion of samples required before this is deemed real.}

\item{chr_column}{Name of the metadata column with the chromosome names.}

\item{ignore_strand}{Ignore strand information when returning?}
}
\value{
List with some information by gene.
}
\description{
Instead of cross referencing variant positions against experimental
condition, one might be interested in seeing what variants are observed per
gene.  This function attempts to answer that question.
}
\examples{
 \dontrun{
 exp <- create_se(metadata, gene_information)
 snp_exp <- count_snps(exp)
 snp_result <- get_snp_sets(snp_exp)
 gene_intersections <- snps_vs_genes(exp, snp_result)
}
}
\seealso{
[GenomicRanges::makeGRangesFromDataFrame()] [IRanges::subsetByOverlaps()]
 [IRanges::mergeByOverlaps()] [IRanges::countOverlaps()]
}
