% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.r
\name{snps_vs_genes}
\alias{snps_vs_genes}
\title{Make a summary of the observed snps/gene}
\usage{
snps_vs_genes(expt, snp_result)
}
\arguments{
\item{expt}{The original expressionset}

\item{snp_result}{The result from get_snp_sets()}
}
\value{
a fun list with some information by gene.
}
\description{
Make a summary of the observed snps/gene
}
