% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{snps_vs_genes_padded}
\alias{snps_vs_genes_padded}
\title{A copy of the above function with padding for species without defined UTRs}
\usage{
snps_vs_genes_padded(
  exp,
  snp_result,
  start_column = "start",
  end_column = "end",
  strand_column = "strand",
  padding = 200,
  normalize = TRUE,
  snp_name_column = "seqnames",
  chr_column = "chromosome",
  observed_in = NULL,
  ignore_strand = TRUE
)
}
\arguments{
\item{exp}{The original expressionset.}

\item{snp_result}{The result from get_snp_sets().}

\item{start_column}{Which column provides the start of each gene?}

\item{end_column}{and the end column of each gene?}

\item{strand_column}{Define strands.}

\item{padding}{Add this amount to each CDS.}

\item{normalize}{Normalize the returns to the length of the putative CDS.}

\item{snp_name_column}{Name of the column in the metadata with the sequence names.}

\item{chr_column}{Name of the metadata column with the chromosome names.}

\item{observed_in}{Print some information about how many variants were observed.}

\item{ignore_strand}{Ignore the strand information when returning?}
}
\value{
List with some information by gene.
}
\description{
A copy of the above function with padding for species without defined UTRs
}
\examples{
 \dontrun{
 exp <- create_se(metadata, gene_information)
 snp_exp <- count_snps(exp)
 snp_result <- get_snp_sets(snp_exp)
 gene_intersections <- snps_vs_genes(exp, snp_result)
}
}
\seealso{
[GenomicRanges::makeGRangesFromDataFrame()] [IRanges::subsetByOverlaps()]
 [IRanges::mergeByOverlaps()] [IRanges::countOverlaps()]
}
