% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{snps_vs_genes_padded}
\alias{snps_vs_genes_padded}
\title{A copy of the above function with padding for species without defined UTRs}
\usage{
snps_vs_genes_padded(
  expt,
  snp_result,
  start_col = "start",
  end_col = "end",
  strand_col = "strand",
  padding = 200,
  normalize = TRUE,
  snp_name_col = "seqnames",
  expt_name_col = "chromosome",
  observed_in = NULL,
  ignore_strand = TRUE
)
}
\description{
A copy of the above function with padding for species without defined UTRs
}
