% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.r
\name{spirograph}
\alias{spirograph}
\title{spirograph()  Make spirographs!  Taken (with modifications) from: http://menugget.blogspot.com/2012/12/spirograph-with-r.html#more}
\usage{
spirograph(radius_a = 1, radius_b = -4, dist_bc = -2, revolutions = 158,
  increments = 3160, center_a = list(x = 0, y = 0))
}
\arguments{
\item{increments}{default=360  The number of radial increments to be calculated per revolution}

\item{rad_a}{default=1  The radius of the primary circle.}

\item{rad_b}{default=2  The radius of the circle travelling around a.}

\item{c}{a point relative to the center of 'b' which rotates with the turning of 'b'.}

\item{bc}{the distance from the center of 'b' to 'c'.}

\item{rev_a}{default=4  The number of revolutions for 'b' to travel about 'a'.}

\item{cen_a}{default=list(x=0,y=0)  The position of the center of 'a'.

A positive value for 'B' will result in a epitrochoid, while a negative value will result in a hypotrochoid.}
}
\value{
somethign which I don't yet know.
}
\description{
spirograph()  Make spirographs!  Taken (with modifications) from: http://menugget.blogspot.com/2012/12/spirograph-with-r.html#more
}

