% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{state<-,SummarizedExperiment-method}
\alias{state<-,SummarizedExperiment-method}
\title{Put the state into a SummarizedExperiment.}
\usage{
\S4method{state}{SummarizedExperiment}(input) <- value
}
\arguments{
\item{input}{Input summarized experiment.}

\item{value}{new state.}
}
\description{
Put the state into a SummarizedExperiment.
}
