% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{state<-,SummarizedExperiment-method}
\alias{state<-,SummarizedExperiment-method}
\title{Put the state into a SummarizedExperiment.}
\usage{
\S4method{state}{SummarizedExperiment}(expt) <- value
}
\description{
Put the state into a SummarizedExperiment.
}
