% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{state<-}
\alias{state<-}
\title{Set the state of the data in an expt.}
\usage{
state(input) <- value
}
\arguments{
\item{input}{Experiment requiring a state update.}

\item{value}{New state!}
}
\description{
Set the state of the data in an expt.
}
