% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_attributes.R
\name{state,SummarizedExperiment-method}
\alias{state,SummarizedExperiment-method}
\title{Get the state from a SummarizedExperiment.}
\usage{
\S4method{state}{SummarizedExperiment}(input)
}
\arguments{
\item{input}{Input summarized experiment.}
}
\description{
Get the state from a SummarizedExperiment.
}
