% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{state}
\alias{state}
\title{Get the state of the data in an expt.}
\usage{
state(input)
}
\arguments{
\item{input}{Experiment containing the state.}
}
\description{
Get the state of the data in an expt.
}
