% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R, R/zzz_methods.R
\name{state}
\alias{state}
\title{Get the state of the data in an expt.}
\usage{
state(expt)

state(expt)
}
\arguments{
\item{expt}{One of my slightly modified ExpressionSets.}
}
\value{
List with the methods used to modify the data (if any).
}
\description{
Get the state of the data in an expt.

Get the state from an expt.
}
