% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.expt}
\alias{[.expt}
\title{Subset an expt}
\usage{
\method{[}{expt}(exp, i, j, ..., value)
}
\arguments{
\item{exp}{Dataset to subset}

\item{i}{Set of genes to keep}

\item{j}{Set of samples to keep.}

\item{...}{Parameters to pass to subset_genes/subset.}

\item{value}{New subset value}
}
\description{
Working on improving my understanding of how R sets up functions by type.
}
