% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[[,expt,ANY,ANY-method}
\alias{[[,expt,ANY,ANY-method}
\title{Simplifying subset on metadata.}
\usage{
\S4method{[[}{expt,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{an expt}

\item{i}{Column to extract}

\item{j}{not sure}

\item{...}{extra args.}
}
\description{
Simplifying subset on metadata.
}
