% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[[,expt,character,ANY-method}
\alias{[[,expt,character,ANY-method}
\title{Simplifying subset on metadata when i is a character.}
\usage{
\S4method{[[}{expt,character,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{an expt}

\item{i}{Column to extract}

\item{j}{not sure}

\item{...}{extra arguments.}
}
\description{
Simplifying subset on metadata when i is a character.
}
