% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_subset.R
\name{[[<-,expt,ANY,missing,ANY-method}
\alias{[[<-,expt,ANY,missing,ANY-method}
\title{Simplifying subset on metadata when i is anything and j is missing.}
\usage{
\S4method{[[}{expt,ANY,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{an expt}

\item{i}{Column to extract}

\item{j}{not sure}

\item{...}{extra arguments.}
}
\description{
Simplifying subset on metadata when i is anything and j is missing.
}
