% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{subset_expt,SummarizedExperiment-method}
\alias{subset_expt,SummarizedExperiment-method}
\title{Subset a SummarizedExperiment with some extra syntax.}
\usage{
\S4method{subset_expt}{SummarizedExperiment}(expt, subset = NULL, ids = NULL, nonzero = NULL, coverage = NULL)
}
\description{
Subset a SummarizedExperiment with some extra syntax.
}
