% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{subset_expt}
\alias{subset_expt}
\title{Subset a summarizedExperiment when using deprecated function names.}
\usage{
subset_expt(...)
}
\description{
Subset a summarizedExperiment when using deprecated function names.
}
