% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{subset_expt}
\alias{subset_expt}
\title{An alias to expt_subset, because it is stupid to have something start with verbs and others start with nouns.}
\usage{
subset_expt(...)
}
\arguments{
\item{...}{All arguments are passed to expt_subset.}
}
\description{
This just calls expt_subset.
}

