% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{subset_expt}
\alias{subset_expt}
\title{Extract a subset of samples following some rule(s) from an
experiment class.}
\usage{
subset_expt(expt, subset = NULL, ids = NULL, nonzero = NULL, coverage = NULL)
}
\arguments{
\item{expt}{Expt chosen to extract a subset of data.}

\item{subset}{Valid R expression which defines a subset of the design to keep.}

\item{ids}{List of sample IDs to extract.}

\item{coverage}{Request a minimum coverage/sample rather than text-based subset.}
}
\value{
metadata Expt class which contains the smaller set of data.
}
\description{
Sometimes an experiment has too many parts to work with conveniently, this
operation allows one to break it into smaller pieces.
}
\examples{
\dontrun{
 smaller_expt <- expt_subset(big_expt, "condition=='control'")
 all_expt <- expt_subset(expressionset, "")  ## extracts everything
}
}
\seealso{
[Biobase] [pData()] [exprs()] [fData()]
}
