% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{subset_ontology_search}
\alias{subset_ontology_search}
\title{subset_ontology_search()  Perform ontology searches on data subsets.}
\usage{
subset_ontology_search(changed_counts, num_cpus = NULL, doplot = FALSE, ...)
}
\description{
subset_ontology_search()  Perform ontology searches on data subsets.
}

