% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.r
\name{subset_ontology_search}
\alias{subset_ontology_search}
\title{Perform ontology searches on up/down subsets of differential expression data.}
\usage{
subset_ontology_search(changed_counts, doplot = TRUE, do_goseq = TRUE,
  do_cluster = TRUE, do_topgo = TRUE, do_gostats = TRUE,
  do_gprofiler = TRUE, according_to = "limma", ...)
}
\arguments{
\item{changed_counts}{List of changed counts as ups and downs.}

\item{doplot}{Include plots in the results?}

\item{do_goseq}{Perform goseq search?}

\item{do_cluster}{Perform clusterprofiler search?}

\item{do_topgo}{Perform topgo search?}

\item{do_gostats}{Perform gostats search?}

\item{do_gprofiler}{Do a gprofiler search?}

\item{according_to}{If results from multiple DE tools were passed, which one
defines 'significant'?}

\item{...}{Extra arguments!}
}
\value{
List of ontology search results, up and down for each contrast.
}
\description{
In the same way all_pairwise() attempts to simplify using multiple DE tools,
this function seeks to make it easier to extract subsets of differentially
expressed data and pass them to goseq, clusterProfiler, topGO, GOstats, and
gProfiler.
}
\seealso{
\pkg{goseq} \pkg{clusterProfiler} \pkg{topGO} \pkg{goStats} \pkg{gProfiler}
}
