% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_batch.R
\name{svpc_fstats}
\alias{svpc_fstats}
\title{Get the f-stats before/after messing with sva.}
\usage{
svpc_fstats(exp, ...)
}
\arguments{
\item{exp}{input}

\item{...}{Args passed to everything else.}
}
\description{
Get the f-stats before/after messing with sva.
}
