% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnseq.r
\name{tnseq_saturation}
\alias{tnseq_saturation}
\title{Make a plot and some simple numbers about tnseq saturation}
\usage{
tnseq_saturation(file)
}
\arguments{
\item{file}{a file created using the perl script 'essentiality_tas.pl'}
}
\value{
A plot and some numbers
}
\description{
This function takes as input a tab separated file from essentiality_tas.pl
This is a perl script written to read a bam alignment of tnseq reads against a genome
and count how many hits were observed on every TA in the given genome.
It furthermore has some logic to tell the difference between reads which were
observed on the forward vs. reverse strand as well as reads which appear to be on both strands
(eg. they start and end with 'TA').
}
\seealso{
\pkg{ggplot2}
}
