% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnseq.r
\name{tnseq_saturation}
\alias{tnseq_saturation}
\title{Make a plot and some simple numbers about tnseq saturation}
\usage{
tnseq_saturation(data, column = "Reads", ylimit = 100)
}
\arguments{
\item{data}{data to plot}

\item{column}{which column to use for plotting}
}
\value{
A plot and some numbers:
 \enumerate{
  \item maximum_reads = The maximum number of reads observed in a single position.
  \item hits_by_position = The full table of hits / position
  \item num_hit_table = A table of how many times every number of hits was observed.
  \item eq_0 = How many times were 0 hits observed?
  \item gt_1 = How many positions have > 1 hit?
  \item gt_2 = How many positions have > 2 hits?
  \item gt_4 = How many positions have > 4 hits?
  \item gt_8 = How many positions have > 8 hits?
  \item gt_16 = How many positions have > 16 hits?
  \item gt_32 = How many positions have > 32 hits?
  \item ratios = Character vector of the ratios of each number of hits vs. 0 hits.
  \item hit_positions = 2 column data frame of positions and the number of observed hits.
  \item hits_summary = summary(hit_positions)
  \item plot = Histogram of the number of hits observed.
 }
}
\description{
This function takes as input a tab separated file from essentiality_tas.pl
This is a perl script written to read a bam alignment of tnseq reads against
a genome and count how many hits were observed on every TA in the given
genome. It furthermore has some logic to tell the difference between reads
which were observed on the forward vs. reverse strand as well as reads which
appear to be on both strands (eg. they start and end with 'TA').
}
\examples{
 \dontrun{
 input <- "preprocessing/hpgl0837/essentiality/hpgl0837-trimmed_ca_ta-v0M1.wig"
 saturation <- tnseq_saturation(file=input)
}
}
\seealso{
\pkg{ggplot2}
}
