% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_shared.R
\name{topgo2enrich}
\alias{topgo2enrich}
\title{Convert a simple_topgo() result to an enrichResult.}
\usage{
topgo2enrich(
  retlist,
  ontology = "mf",
  pval = 0.05,
  column = "fisher",
  padjust_method = "BH"
)
}
\arguments{
\item{retlist}{result from simple_topgo()}

\item{ontology}{Ontology subtree to act upon.}

\item{pval}{Cutoff, hmm I think I need to standardize these.}

\item{column}{Table column to export.}
}
\description{
Same idea as goseq2enrich.
}
