% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ontology.R
\name{topgo_pval_plot}
\alias{topgo_pval_plot}
\title{Make a pvalue plot from topgo data}
\usage{
topgo_pval_plot(topgo, wrapped_width = 20, cutoff = 0.1, n = 12,
  type = "fisher")
}
\arguments{
\item{topgo_data}{some data from topgo!}
}
\value{
a plot!
}
\description{
Make a pvalue plot from topgo data
}
\seealso{
\code{\link{goseq}}
}

