% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ontology.R
\name{topgo_tables}
\alias{topgo_tables}
\title{Make tables out of topGO data}
\usage{
topgo_tables(result, limit = 0.01, limitby = "fisher", numchar = 300,
  orderby = "classic", ranksof = "classic")
}
\description{
Make tables out of topGO data
}

