% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{topgo_tables}
\alias{topgo_tables}
\title{Make pretty tables out of topGO data}
\usage{
topgo_tables(result, limit = 0.1, limitby = "fisher", numchar = 300,
  orderby = "classic", ranksof = "classic")
}
\arguments{
\item{result}{Topgo result.}

\item{limit}{Pvalue limit defining 'significant'.}

\item{limitby}{Type of test to perform.}

\item{numchar}{How many characters to allow in the description?}

\item{orderby}{Which of the available columns to order the table by?}

\item{ranksof}{Which of the available columns are used to rank the data?}
}
\value{
prettier tables
}
\description{
The topgo function GenTable is neat, but it needs some simplification to not be obnoxious.
}
\seealso{
\pkg{topGO}
}
