% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_topgo.r
\name{topgo_trees}
\alias{topgo_trees}
\title{Print trees from topGO}
\usage{
topgo_trees(tg, score_limit = 0.01, sigforall = TRUE,
  do_mf_fisher_tree = TRUE, do_bp_fisher_tree = TRUE,
  do_cc_fisher_tree = TRUE, do_mf_ks_tree = FALSE, do_bp_ks_tree = FALSE,
  do_cc_ks_tree = FALSE, do_mf_el_tree = FALSE, do_bp_el_tree = FALSE,
  do_cc_el_tree = FALSE, do_mf_weight_tree = FALSE,
  do_bp_weight_tree = FALSE, do_cc_weight_tree = FALSE)
}
\arguments{
\item{de_genes}{a data frame of differentially expressed genes, containing IDs and whatever other columns}

\item{goid_map}{a file containing mappings of genes to goids in the format expected by topgo}
}
\value{
a big list including the various outputs from topgo
}
\description{
Print trees from topGO
}

