% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{transform_counts}
\alias{transform_counts}
\title{transform_counts(): Perform a simple transformation of a count table (log2)}
\usage{
transform_counts(count_table, transform = "raw", converted = "raw",
  base = NULL, add = 0.5)
}
\arguments{
\item{count_table}{A matrix of count data}

\item{transform='raw'}{A type of transformation to perform: log2/log10/log}

\item{converted='raw'}{Whether or not the data has been converted.
Only important if the data was previously cpm'd because that does a +1, thus
this will avoid a double+1 on the data.}
}
\value{
dataframe of logx(counts)
}
\description{
transform_counts(): Perform a simple transformation of a count table (log2)
}
\examples{
## filtered_table = transform_counts(count_table, transform='log2', converted='cpm')
}

