% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_varpartition.r
\name{varpart}
\alias{varpart}
\title{Use variancePartition to gather some plots about a troubling dataset}
\usage{
varpart(expt, factors = c("condition", "batch"), cpus = 6, genes = 20)
}
\arguments{
\item{expt}{Some data}

\item{factors}{Character list of columns in the experiment design to query}

\item{cpus}{Number cpus to use}

\item{genes}{Number of genes to count}
}
\value{
partitions  List of plots and variance data frames
}
\description{
variancePartition is the newest toy introduced by Hector
}

