% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{what_happened}
\alias{what_happened}
\title{Print a string describing what happened to this data.}
\usage{
what_happened(
  expt = NULL,
  transform = "raw",
  convert = "raw",
  norm = "raw",
  filter = "raw",
  batch = "raw"
)
}
\arguments{
\item{expt}{The expressionset.}

\item{transform}{How was it transformed?}

\item{convert}{How was it converted?}

\item{norm}{How was it normalized?}

\item{filter}{How was it filtered?}

\item{batch}{How was it batch-corrected?}
}
\value{
An expression describing what has been done to this data.
}
\description{
Sometimes it is nice to have a string like: log2(cpm(data)) describing what
happened to the data.
}
\seealso{
[create_expt()] [normalize_expt()]
}
