% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_basic.r
\name{write_basic}
\alias{write_basic}
\title{Writes out the results of a basic search using write_de_table()}
\usage{
write_basic(data, ...)
}
\arguments{
\item{data}{Output from basic_pairwise()}

\item{...}{Options for writing the xlsx file.}
}
\description{
Looking to provide a single interface for writing tables from basic and friends.
}
\details{
Tested in test_26basic.R
}
\examples{
\dontrun{
 finished_comparison <- basic_pairwise(expressionset)
 data_list <- write_basic(finished_comparison)
}
}
\seealso{
[basic_pairwise()] [write_de_table()]
}
