% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{write_combined_legend}
\alias{write_combined_legend}
\title{Write the legend of an excel file for combine_de_tables()}
\usage{
write_combined_legend(
  wb,
  excel_basename,
  plot_dim,
  apr,
  limma,
  include_limma,
  deseq,
  include_deseq,
  edger,
  include_edger,
  ebseq,
  include_ebseq,
  basic,
  include_basic,
  padj_type
)
}
\arguments{
\item{wb}{Workbook to write}

\item{excel_basename}{Where to write it}

\item{plot_dim}{Default plot size.}

\item{apr}{The all_pairwise() result.}

\item{limma}{The limma result, which is redundant.}

\item{include_limma}{Include the limma result?}

\item{deseq}{The deseq result, which is redundant.}

\item{include_deseq}{Include the deseq result?}

\item{edger}{The edger result, which is redundant.}

\item{include_edger}{Include the edger result?}

\item{ebseq}{The ebseq result, which is redundant.}

\item{include_ebseq}{Include the ebseq result?}

\item{basic}{Basic data}

\item{include_basic}{Include the basic result?}

\item{padj_type}{P-adjustment employed.}
}
\description{
Write the legend of an excel file for combine_de_tables()
}
