% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{write_combined_summary}
\alias{write_combined_summary}
\title{Internal function to write a summary of some combined data}
\usage{
write_combined_summary(wb, excel_basename, apr, extracted, compare_plots)
}
\arguments{
\item{wb}{xlsx workbook to which to write.}

\item{excel_basename}{basename for printing plots.}

\item{apr}{a pairwise result}

\item{extracted}{table extracted from the pairwise result}

\item{compare_plots}{series of plots to print out.}
}
\description{
Internal function to write a summary of some combined data
}
