% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{write_de_table}
\alias{write_de_table}
\title{Writes out the results of a single pairwise comparison.}
\usage{
write_de_table(data, type = "limma", ...)
}
\arguments{
\item{data}{Output from results().}

\item{type}{Which DE tool to write.}

\item{...}{Parameters passed downstream, dumped into arglist and passed, notably the number
of genes (n), the coefficient column (coef)}
}
\value{
List of data frames comprising the toptable output for each coefficient, I also added a
qvalue entry to these toptable() outputs.
}
\description{
However, this will do a couple of things to make one's life easier:
\enumerate{
\item Make a list of the output, one element for each comparison of the contrast matrix.
\item Write out the results() output for them in separate sheets in excel.
\item Since I have been using qvalues a lot for other stuff, add a column for them.
}
}
\details{
Tested in test_24deseq.R
Rewritten in 2016-12 looking to simplify combine_de_tables().  That function is far too big,
This should become a template for that.
}
\examples{
\dontrun{
 finished_comparison = eBayes(deseq_output)
 data_list = write_deseq(finished_comparison, workbook="excel/deseq_output.xls")
}
}
\seealso{
\code{\link{write_xls}}
}
