% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{write_edger}
\alias{write_edger}
\title{write_edger()  Writes out the results of an EdgeR search using topTags()
However, this will do a couple of things to make one's life easier:
1.  Make a list of the output, one element for each comparison of the contrast matrix
2.  Write out the toptable() output for them in separate .csv files and/or sheets in excel
3.  Since I have been using qvalues a lot for other stuff, add a column for them.}
\usage{
write_edger(data, adjust = "fdr", n = 0, coef = NULL,
  workbook = "excel/edger.xls", excel = FALSE, csv = TRUE,
  annotation = NULL)
}
\arguments{
\item{data}{the output from EdgeR}

\item{adjust}{default='fdr'  the pvalue adjustment chosen.}

\item{n}{default=0  the number of entries to report, 0 says to do them all.}

\item{coef}{default=NULL  which coefficients/contrasts to report, NULL says do them all.}

\item{workbook}{default='excel/edger.xls'  an excel filename into which to write the data, used for csv files too.}

\item{excel}{default=FALSE  whether or not to write an excel workbook rather than csv.}

\item{csv}{default=TRUE  whether to write out the output tables as csv files.}

\item{annotation}{default=NULL  an optional data frame of annotations to include with the tables.}
}
\value{
a list of data frames comprising the toptable output for each coefficient,
   I also added a qvalue entry to these toptable() outputs.
}
\description{
write_edger()  Writes out the results of an EdgeR search using topTags()
However, this will do a couple of things to make one's life easier:
1.  Make a list of the output, one element for each comparison of the contrast matrix
2.  Write out the toptable() output for them in separate .csv files and/or sheets in excel
3.  Since I have been using qvalues a lot for other stuff, add a column for them.
}
\examples{
## finished_comparison = eBayes(limma_output)
## data_list = write_limma(finished_comparison, workbook="excel/limma_output.xls")
}
\seealso{
\code{\link{toptable}}. \code{\link{write_xls}}
}

