% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tables.r
\name{write_expt}
\alias{write_expt}
\title{Make pretty xlsx files of count data.}
\usage{
write_expt(expt, excel = "excel/pretty_counts.xlsx", norm = "quant",
  violin = FALSE, convert = "cpm", transform = "log2", batch = "sva",
  filter = "cbcb")
}
\arguments{
\item{expt}{An expressionset to print.}

\item{excel}{Filename to write.}

\item{norm}{Normalization to perform.}

\item{violin}{Include violin plots?}

\item{convert}{Conversion to perform.}

\item{transform}{Transformation used.}

\item{batch}{Batch correction applied.}

\item{filter}{Filtering method used.}
}
\value{
A big honking excel file and a list including the dataframes and images created.
}
\description{
Some folks love excel for looking at this data.  ok.
}

