% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tables.r
\name{write_expt}
\alias{write_expt}
\title{Make pretty xlsx files of count data.}
\usage{
write_expt(expt, excel = "excel/pretty_counts.xlsx", norm = "quant",
  convert = "cpm", transform = "log2", batch = "sva", filter = "cbcb")
}
\description{
Make pretty xlsx files of count data.
}

