% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.r
\name{write_expt}
\alias{write_expt}
\title{Make pretty xlsx files of count data.}
\usage{
write_expt(
  expt,
  excel = "excel/pretty_counts.xlsx",
  norm = "quant",
  violin = TRUE,
  sample_heat = TRUE,
  convert = "cpm",
  transform = "log2",
  batch = "sva",
  filter = TRUE,
  med_or_mean = "mean",
  color_na = "#DD0000",
  ...
)
}
\arguments{
\item{expt}{An expressionset to print.}

\item{excel}{Filename to write.}

\item{norm}{Normalization to perform.}

\item{violin}{Include violin plots?}

\item{sample_heat}{Include sample heatmaps?}

\item{convert}{Conversion to perform.}

\item{transform}{Transformation used.}

\item{batch}{Batch correction applied.}

\item{filter}{Filtering method used.}

\item{med_or_mean}{When printing mean by condition, one may want median.}

\item{...}{Parameters passed down to methods called here (graph_metrics, etc).}
}
\value{
A big honking excel file and a list including the dataframes and images created.
}
\description{
Some folks love excel for looking at this data.  ok.
}
\details{
Tested in test_03graph_metrics.R
This performs the following:  Writes the raw data, graphs the raw data,
normalizes the data, writes it, graphs it, and does a median-by-condition and
prints that.  I replaced the openxlsx function which writes images into xlsx
files with one which does not require an opening of a pre-existing plotter.
Instead it (optionally)opens a pdf device, prints the plot to it, opens a png
device, prints to that, and inserts the resulting png file.  Thus it
sacrifices some flexibility for a hopefully more consistent behaivor.  In
addition, one may use the pdfs as a set of images importable into illustrator
or whatever.
}
\examples{
\dontrun{
 excel_sucks <- write_expt(expt)
}
}
\seealso{
[openxlsx] [Biobase] [normalize_expt()] [graph_metrics()]
}
