% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.r
\name{write_go_xls}
\alias{write_go_xls}
\title{Write gene ontology tables for excel}
\usage{
write_go_xls(goseq, cluster, topgo, gostats, gprofiler,
  file = "excel/merged_go", dated = TRUE, n = 30, overwritefile = TRUE)
}
\arguments{
\item{goseq}{The goseq result from simple_goseq()}

\item{cluster}{The result from simple_clusterprofiler()}

\item{topgo}{Guess}

\item{gostats}{Yep, ditto}

\item{gprofiler}{woo hoo!}

\item{file}{the file to save the results.}

\item{dated}{date the excel file}

\item{n}{the number of ontology categories to include in each table.}

\item{overwritefile}{overwrite an existing excel file}
}
\value{
the list of ontology information
}
\description{
Combine the results from goseq, cluster profiler, topgo, and gostats and drop them into excel
Hopefully with a relatively consistent look.
}
\seealso{
\pkg{openxlsx} \pkg{goseq} \pkg{clusterProfiler} \pkg{goStats} \pkg{topGO} \pkg{gProfiler}
}
