% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_goseq.r
\name{write_goseq_data}
\alias{write_goseq_data}
\title{Make a pretty table of goseq data in excel.}
\usage{
write_goseq_data(goseq, file = "excel/goseq.xlsx", pval = 0.1,
  add_plots = TRUE)
}
\arguments{
\item{goseq}{A set of results from simple_goseq().}

\item{file}{An excel file to which to write some pretty results.}

\item{pval}{Choose a cutoff for reporting by p-value.}

\item{add_plots}{Include some pvalue plots in the excel output?}
}
\value{
The result from openxlsx
}
\description{
It is my intention to make a function like this for each ontology tool in my repetoire
}

