% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_xlsx.r
\name{write_gprofiler_data}
\alias{write_gprofiler_data}
\title{Write some excel results from a gprofiler search.}
\usage{
write_gprofiler_data(
  gprofiler_result,
  wb = NULL,
  excel = "excel/gprofiler_result.xlsx",
  order_by = "recall",
  add_plots = TRUE,
  height = 15,
  width = 10,
  decreasing = FALSE,
  ...
)
}
\arguments{
\item{gprofiler_result}{The result from simple_gprofiler().}

\item{wb}{Optional workbook object, if you wish to append to an existing workbook.}

\item{excel}{Excel file to which to write.}

\item{order_by}{Which column to order the data by?}

\item{add_plots}{Add some pvalue plots?}

\item{height}{Height of included plots?}

\item{width}{And their width.}

\item{decreasing}{Which order?}

\item{...}{More options, not currently used I think.}
}
\value{
A prettyified table in an xlsx document.
}
\description{
Gprofiler is pretty awesome.  This function will attempt to write its results
to an excel file.
}
\seealso{
\pkg{openxlsx} \pkg{gProfiler}
}
