% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.r
\name{write_intersect_significant}
\alias{write_intersect_significant}
\title{Attempt to find the significant shared genes between edger/deseq/limma or a subset thereof.}
\usage{
write_intersect_significant(tables, excel = "excel/significant_shared.xlsx",
  extra_annot = NULL, ...)
}
\arguments{
\item{tables}{The result from extract_significant_genes() or similar.}

\item{excel}{An excel file to write.}

\item{extra_annot}{Extra annotations to add to the tables.}

\item{...}{Extra arguments for writing the file (currently unused).}
}
\value{
a list of shared genes by table name.
}
\description{
Attempt to find the significant shared genes between edger/deseq/limma or a subset thereof.
}
