% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{write_limma}
\alias{write_limma}
\title{write_limma(): Writes out the results of a limma search using toptable()
However, this will do a couple of things to make one's life easier:
1.  Make a list of the output, one element for each comparison of the contrast matrix
2.  Write out the toptable() output for them in separate .csv files and/or sheets in excel
3.  Since I have been using qvalues a lot for other stuff, add a column for them.}
\usage{
write_limma(data = NULL, adjust = "fdr", n = 0, coef = NULL,
  workbook = "excel/limma.xls", excel = FALSE, csv = TRUE,
  annotation = NULL)
}
\arguments{
\item{data}{The output from eBayes()}

\item{adjust}{The pvalue adjustment chosen (fdr by default)}

\item{n}{The number of entries to report, defaults to 0, which says do them all}

\item{coef}{which coefficients/contrasts to report, NULL says do them all}

\item{workbook}{an excel filename into which to write the data, used for csv files too.}

\item{excel}{T/F whether or not to write an excel workbook (useful if they are too big)}
}
\value{
a list of data frames comprising the toptable output for each coefficient,
   I also added a qvalue entry to these toptable() outputs.
}
\description{
write_limma(): Writes out the results of a limma search using toptable()
However, this will do a couple of things to make one's life easier:
1.  Make a list of the output, one element for each comparison of the contrast matrix
2.  Write out the toptable() output for them in separate .csv files and/or sheets in excel
3.  Since I have been using qvalues a lot for other stuff, add a column for them.
}
\examples{
## finished_comparison = eBayes(limma_output)
## data_list = write_limma(finished_comparison, workbook="excel/limma_output.xls")
}
\seealso{
\code{\link{toptable}}. \code{\link{write_xls}}
}

