% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_limma.r
\name{write_limma}
\alias{write_limma}
\title{Writes out the results of a limma search using write_de_table()}
\usage{
write_limma(data, ...)
}
\arguments{
\item{data}{Output from limma_pairwise()}

\item{...}{Options for writing the xlsx file.}
}
\description{
Looking to provide a single interface for writing tables from limma and friends.
}
\details{
Tested in test_21limma.R
}
\examples{
\dontrun{
 finished_comparison = limma_pairwise(expressionset)
 data_list = write_limma(finished_comparison)
}
}
\seealso{
\code{\link{write_de_table}}
}
