% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_xlsx.R
\name{write_sig_legend}
\alias{write_sig_legend}
\title{Internal function to write a legend for significant gene tables.}
\usage{
write_sig_legend(wb)
}
\arguments{
\item{wb}{xlsx workbook object from openxlsx.}
}
\description{
Internal function to write a legend for significant gene tables.
}
