% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{write_xls}
\alias{write_xls}
\title{write_xls()  Write a dataframe to an excel spreadsheet sheet.}
\usage{
write_xls(data, sheet = "first", file = "excel/workbook",
  overwrite_file = TRUE, newsheet = FALSE, overwrite_sheet = TRUE,
  dated = TRUE, suffix = ".xlsx", type = "openxlsx", ...)
}
\arguments{
\item{type}{default='xlsx'  whether to use the xlsx or XLConnect package.}

\item{tripledots}{the set of arguments given to either xlsx or XLConnect}
}
\value{
an excel workbook
}
\description{
write_xls()  Write a dataframe to an excel spreadsheet sheet.
}
\examples{
## write_xls(dataframe, "hpgl_data")
}
\seealso{
\code{\link{XLConnect}}, \code{\link{xlsx}},
}

