% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_xls}
\alias{write_xls}
\title{Write a dataframe to an excel spreadsheet sheet.}
\usage{
write_xls(data, sheet, file = "excel/workbook.xls", rowname = NA)
}
\arguments{
\item{data}{a dataframe of information}

\item{sheet}{the name of an excel sheet in a workbook.}

\item{file}{an excel workbook to which to write.  Defaults to "excel/workbook.xls"}

\item{rowname}{include rownames?  Defalts to no.}
}
\value{
NULL, on the say it creates a workbook if necessary,
creates a sheet, and writes the data to it.
}
\description{
Write a dataframe to an excel spreadsheet sheet.
}
\examples{
## write_xls(dataframe, "my_data")
## re-create it if this is used heavily, because it will get crufty.
}
\seealso{
\code{\link{loadWorkbook}}, \code{\link{createSheet}},
\code{\link{writeWorksheet}}, \code{\link{saveWorkbook}}
}

