% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{write_xls_xlsx}
\alias{write_xls_xlsx}
\title{write_xls_xlsx()  Write a dataframe to an excel spreadsheet sheet using xlsx.}
\usage{
write_xls_xlsx(data, sheet = "first", file = "excel/workbook",
  header = "Data Table", rownames = TRUE, colnames = TRUE,
  overwritefile = FALSE, dated = TRUE, suffix = ".xlsx")
}
\arguments{
\item{data}{a dataframe of information.}

\item{sheet}{default='first'  the name of an excel sheet in a workbook.}

\item{file}{default='excel/workbook.xls'  an excel workbook to which to write.}

\item{header}{default='Data Table'  a header to prepend to the printed worksheet.}

\item{rownames}{default=TRUE  add the rownames to the left side of the table?}

\item{colnames}{default=TRUE  add the column names to the top of the table?}

\item{overwritefile}{default=FALSE  overwrite the xls file with this new data, or use the original?}

\item{dated}{default=TRUE  append a dated suffix to the filename.}

\item{suffix}{default='.xlsx'}
}
\value{
NULL, on the say it creates a workbook if necessary,
creates a sheet, and writes the data to it.
}
\description{
write_xls_xlsx()  Write a dataframe to an excel spreadsheet sheet using xlsx.
}
\examples{
## write_xls_xlsx(dataframe)
}
\seealso{
\code{\link{xlsx.addTable}}
}

