% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{write_xlsx,ANY,ANY,ANY,written_xlsx-method}
\alias{write_xlsx,ANY,ANY,ANY,written_xlsx-method}
\title{Write an xlsx file given the result of an existing xlsx write.}
\usage{
\S4method{write_xlsx}{ANY,ANY,ANY,written_xlsx}(
  data = NULL,
  wb = NULL,
  sheet = NULL,
  excel = NULL,
  rownames = TRUE,
  start_row = 1,
  start_col = 1,
  title = NULL,
  number_format = "0.000",
  data_table = TRUE,
  freeze_first_row = TRUE,
  freeze_first_column = TRUE,
  column_width = "heuristic",
  ...
)
}
\description{
Write an xlsx file given the result of an existing xlsx write.
}
