% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200_xlsx.r
\name{write_xlsx}
\alias{write_xlsx}
\title{Write a dataframe to an excel spreadsheet sheet.}
\usage{
write_xlsx(
  data = "undef",
  wb = NULL,
  sheet = "first",
  excel = NULL,
  rownames = TRUE,
  start_row = 1,
  start_col = 1,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame to print.}

\item{wb}{Workbook to which to write.}

\item{sheet}{Name of the sheet to write.}

\item{excel}{Filename of final excel workbook to write}

\item{rownames}{Include row names in the output?}

\item{start_row}{First row of the sheet to write. Useful if writing multiple tables.}

\item{start_col}{First column to write.}

\item{title}{Title for this xlsx table.}

\item{...}{Set of extra arguments given to openxlsx.}
}
\value{
List containing the sheet and workbook written as well as the
  bottom-right coordinates of the last row/column written to the worksheet.
}
\description{
I like to give folks data in any format they prefer, even though I sort
of hate excel.  Most people I work with use it, so therefore I do too.
This function has been through many iterations, first using XLConnect,
then xlsx, and now openxlsx.  Hopefully this will not change again.
}
\examples{
 \dontrun{
  xls_coords <- write_xlsx(dataframe, sheet="hpgl_data", excel="testing.xlsx")
  xls_coords <- write_xlsx(another_df, wb=xls_coords$workbook,
                          sheet="hpgl_data", start_row=xls_coords$end_col)
 }

}
\seealso{
\pkg{openxlsx}
}
