% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.r
\name{xref_regions}
\alias{xref_regions}
\title{If I were smart I would use an I/GRanges for this.}
\usage{
xref_regions(
  sequence_df,
  gff,
  bin_width = 600,
  feature_type = "protein_coding_gene",
  feature_start = "start",
  feature_end = "end",
  feature_strand = "strand",
  feature_chr = "seqnames",
  feature_type_column = "type",
  feature_id = "ID",
  feature_name = "description"
)
}
\description{
But I was asked to get the closest feature if it is not inside one.
I am not sure how to do that with a ranges. Sadly, I think it will
be easier for me to just iterate over the sequence_df and query
each feature on that chromosome/scaffold.
}
