% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.r
\name{ymxb_print}
\alias{ymxb_print}
\title{Print a model as y = mx + b just like in grade school!}
\usage{
ymxb_print(model)
}
\arguments{
\item{model}{a model to print from glm/lm/robustbase}
}
\description{
Print a model as y = mx + b just like in grade school!
}

